% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_S3.R
\name{add}
\alias{add}
\title{S3 generic \code{add}}
\usage{
add(x, ...)
}
\arguments{
\item{x}{An object to which data will be added.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S3 generic to add data to \code{x}.
}
