% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_general_000_newStreamProject.R
\name{addAnalyses}
\alias{addAnalyses}
\title{addAnalyses}
\usage{
addAnalyses(files = utils::choose.files(), object = NULL, replicates = NULL)
}
\arguments{
\item{files}{A list of complete paths to be added.
The default is a UI to choose the files.}

\item{object}{A \linkS4class{streamProject} object to add the files.}

\item{replicates}{Optional character vector with the same length as the number of files
to be added with the name of the analysis replicate group for each file.
The default is \code{NULL} which attempts  to group the analysis using
the name of the files. The replicate names can always be edited using the class method.}

\item{blanks}{Optional character vector with the same length as the number of files
to be added with the name of the associate blank analysis replicate group for each file.
The default is \code{NULL}.}
}
\value{
Returns the object including the added files
in the \code{samples} slot.
}
\description{
Adds file/s to a \linkS4class{streamProject} object.
}
\note{
When the has files, new files are added
below in the \code{samples}. If processed data is already present,
the new files will invalidate the structure.
Therefore, data should be processed again.
}
