% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_000_generics.R
\name{addMetadata}
\alias{addMetadata}
\title{addMetadata}
\usage{
addMetadata(object, ...)
}
\arguments{
\item{object}{A method specific class object.}

\item{...}{Other method specific arguments.}
}
\description{
Generic for adding metadata.
}
