% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Analyses.R
\name{Analyses}
\alias{Analyses}
\alias{validate_object.Analyses}
\title{Base class for database-backed Analyses}
\usage{
Analyses(projectPath = ".")

\method{validate_object}{Analyses}(x)
}
\arguments{
\item{projectPath}{Path to the project directory where this database file will be created/used.}

\item{x}{A \code{Analyses} object.}
}
\value{
An object of class \code{Analyses}.
}
\description{
Minimal base class to represent analyses stored in a DuckDB file.
Used internally to expose shared S3 helpers for DB-backed analyses classes.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate_object(Analyses)}: Validate the Analyses object.

}}
