% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_AnalysesDB.R
\name{AnalysesDB}
\alias{AnalysesDB}
\alias{validate_object.AnalysesDB}
\title{Base class for database-backed Analyses}
\usage{
AnalysesDB(db, data_type)

\method{validate_object}{AnalysesDB}(x)
}
\arguments{
\item{db}{Path to DuckDB file.}

\item{data_type}{Data type string (e.g., "MassSpec").}

\item{x}{AnalysesDB object.}
}
\value{
An object of class \code{AnalysesDB}.
}
\description{
Minimal base class to represent analyses stored in a DuckDB file.
Requires a \code{db} path and a \code{type} identifier.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate_object(AnalysesDB)}: Validate the AnalysesDB object.

}}
