% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_000_generics.R
\name{analysisInfo}
\alias{analysisInfo}
\title{analysisInfo}
\usage{
analysisInfo(obj)
}
\arguments{
\item{object}{A method specific class object.}
}
\description{
Generic for getting an analysis table.
}
