% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_000_generics.R
\name{analysisNames}
\alias{analysisNames}
\title{analysisNames}
\usage{
analysisNames(object)
}
\arguments{
\item{object}{A method specific class object.}
}
\description{
Generic for getting analysis names.
}
