% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_004_peakAnnotation.R
\name{annotateFeatures}
\alias{annotateFeatures}
\title{annotateFeatures}
\usage{
annotateFeatures(object, pat, comp, prefAdduct = "[M+H]+")
}
\arguments{
\item{object}{An \linkS4class{msData} object with peaks/features.}

\item{prefAdduct}{A character string with the preferred adduct.
Possible values are \code{"[M+H]+"} (the default) and \code{"[M-H]-"} for
positive and negative ionization, respectively.}
}
\description{
Selects the ions from the components and amends the festure data table.
}
