% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_004_peakAnnotation.R
\name{annotationSettingsDefaultCAMERA}
\alias{annotationSettingsDefaultCAMERA}
\title{annotationSettingsDefaultCAMERA}
\usage{
annotationSettingsDefaultCAMERA()
}
\value{
A \linkS4class{settings} object containing parameters for
recursive integration of peaks from analyses not represented in a given feature.
}
\description{
annotationSettingsDefaultCAMERA
}
