% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Config.R
\name{AppConfig}
\alias{AppConfig}
\title{AppConfig Class}
\usage{
AppConfig()
}
\description{
Class representing the app configuration, inheriting from \link{Config}.
}
\section{Slots}{

\describe{
\item{\code{parameters}}{A list of \link{ConfigParameter} objects.}
}}

