% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_000_generics.R
\name{as.featureGroups}
\alias{as.featureGroups}
\title{as.featureGroups}
\usage{
as.featureGroups(object, ...)
}
\arguments{
\item{object}{A method specific class object.}

\item{...}{Other method specific arguments.}
}
\description{
Generic for converting to featureGroups.
}
