% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_000_generics.R
\name{as.features}
\alias{as.features}
\title{as.features}
\usage{
as.features(object, ...)
}
\arguments{
\item{object}{A method specific class object.}

\item{...}{Other method specific arguments.}
}
\description{
Generic for converting to features.
}
