% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_AuditTrail.R
\name{AuditTrailEntry}
\alias{AuditTrailEntry}
\title{Generic Entry for Audit Trail}
\usage{
AuditTrailEntry(time_stamp, value_class, value_parent, value)
}
\arguments{
\item{time_stamp}{A POSIXct timestamp of the entry.}

\item{value_class}{A character string representing the class of the entry.}

\item{value_parent}{A character string representing the parent class of the entry.}

\item{value}{A character string representing the value of the entry.}
}
\description{
The \link{AuditTrailEntry} class is used as an element in the
\link{AuditTrail} entries.
}
