% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_000_generics.R
\name{blankReplicateNames<-}
\alias{blankReplicateNames<-}
\title{blankReplicateNames<-}
\usage{
blankReplicateNames(object) <- value
}
\arguments{
\item{object}{A method specific class object.}

\item{value}{A method specific vector with names.}
}
\description{
Generic for setting blank replicate names.
}
