% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_000_generics.R
\name{blankReplicateNames}
\alias{blankReplicateNames}
\title{blankReplicateNames}
\usage{
blankReplicateNames(object)
}
\arguments{
\item{object}{A method specific class object.}
}
\description{
Generic for getting blank replicate names.
}
