% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_007_filterFeatures.R
\name{blankThresholdFeatures}
\alias{blankThresholdFeatures}
\title{blankThresholdFeatures}
\usage{
blankThresholdFeatures(obj, value = 3)
}
\arguments{
\item{obj}{An \linkS4class{msFeatures} object.}

\item{value}{A numerical value to multiply the assigned blank intensity.}
}
\description{
blankThresholdFeatures
}
