% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_R6_msData.R
\name{build_feature_groups_table_from_patRoon}
\alias{build_feature_groups_table_from_patRoon}
\title{build_feature_groups_table_from_patRoon}
\usage{
build_feature_groups_table_from_patRoon(pat, features, self)
}
\arguments{
\item{pat}{A \linkS4class{featureGroups} object from the
package \pkg{patRoon}.}

\item{self}{An `msData` object. When applied within the R6, the self object.}
}
\value{
A \linkS4class{data.table} with the feature groups.
}
\description{
build_feature_groups_table_from_patRoon
}
