% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_R6_msData.R
\name{build_features_table_from_patRoon}
\alias{build_features_table_from_patRoon}
\title{build_features_table_from_patRoon}
\usage{
build_features_table_from_patRoon(pat, self)
}
\arguments{
\item{pat}{An object with class \code{features} or \code{featureGroups} from the
package \pkg{patRoon}.}

\item{self}{An \code{msData} object. When applied within the R6, the self object.}
}
\value{
A list of with a features \linkS4class{data.table} for each analysis.
}
\description{
build_features_table_from_patRoon
}
