% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_002_peakGrouping.R
\name{buildFeatures}
\alias{buildFeatures}
\title{buildFeatures}
\usage{
buildFeatures(object, pat)
}
\arguments{
\item{object}{An \linkS4class{msData} object.}

\item{pat}{A \linkS4class{featureGroups} object from the package \pkg{patRoon}}
}
\value{
An \linkS4class{msData} object with updated features slot.
}
\description{
buildFeatures
}
