% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_001_peakPicking.R
\name{buildPeaksTable}
\alias{buildPeaksTable}
\title{buildPeaksTable}
\usage{
buildPeaksTable(object, pat)
}
\arguments{
\item{object}{A \linkS4class{msData} or \linkS4class{msAnalysis} object.}

\item{pat}{A \linkS4class{features} or \linkS4class{featureGroups} object
from the package \pkg{patRoon}.}
}
\value{
A \linkS4class{data.table} containing information of peaks found
in each sample.
}
\description{
buildPeaksTable
}
