% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Cache.R
\name{Cache}
\alias{Cache}
\alias{save_cache.Cache}
\alias{load_cache.Cache}
\alias{get_cache_info.Cache}
\alias{clear_cache.Cache}
\alias{size.Cache}
\alias{query_db.Cache}
\alias{list_db_tables.Cache}
\alias{get_db_table_info.Cache}
\title{Cache S3 Class}
\usage{
Cache(projectPath = ".")

\method{save_cache}{Cache}(x, name, hash, description, data, ...)

\method{load_cache}{Cache}(x, name, hash, ...)

\method{get_cache_info}{Cache}(x, ...)

\method{clear_cache}{Cache}(x, value = NULL, ...)

\method{size}{Cache}(x, unit = "auto", ...)

\method{query_db}{Cache}(x, sql, params = NULL)

\method{list_db_tables}{Cache}(x)

\method{get_db_table_info}{Cache}(x, tableName)
}
\arguments{
\item{projectPath}{Path to the project directory where this database file will be created/used.}

\item{x}{Cache object}

\item{name}{Name of cache entry (not used, for consistency)}

\item{hash}{Hash string for cache table}

\item{description}{Description of cache entry}

\item{data}{Any R object to cache (will be serialized)}

\item{value}{Character vector of cache entry names to delete. If NULL (default), deletes entire cache database.}

\item{unit}{Optional unit for size, possible values: "bytes", "KB", "MB", "GB" or "auto" (default).}

\item{sql}{SQL query string}

\item{params}{Query parameters (optional)}

\item{tableName}{Table name}
}
\value{
An object of class \code{Cache}.

Cached R object (deserialized)

Data.frame with cache metadata

Integer
}
\description{
S3 class to manage \code{Cache.duckdb} for caching and loading data.
}
\section{Methods (by generic)}{
\itemize{
\item \code{save_cache(Cache)}: Save data to cache table.

\item \code{load_cache(Cache)}: Load data from cache table.

\item \code{get_cache_info(Cache)}: Get cache information table.

\item \code{clear_cache(Cache)}: Reset cache (delete all cache entries or specific entries by name)

\item \code{size(Cache)}: Get size of Cache.duckdb file in bytes.

\item \code{query_db(Cache)}: Internal: execute a query on the DB.

\item \code{list_db_tables(Cache)}: Internal: list tables in the DB.

\item \code{get_db_table_info(Cache)}: Internal: get table info from the DB.

}}
