% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_DB_CacheManager.R
\name{CacheManager}
\alias{CacheManager}
\alias{save_cache.CacheManager}
\alias{load_cache.CacheManager}
\alias{get_cache_info.CacheManager}
\alias{clear_cache.CacheManager}
\alias{size.CacheManager}
\alias{query_db.CacheManager}
\alias{list_db_tables.CacheManager}
\alias{get_db_table_info.CacheManager}
\title{CacheManager S3 Class}
\usage{
CacheManager(db)

\method{save_cache}{CacheManager}(x, name, hash, description, data, ...)

\method{load_cache}{CacheManager}(x, name, hash, ...)

\method{get_cache_info}{CacheManager}(x, ...)

\method{clear_cache}{CacheManager}(x, ...)

\method{size}{CacheManager}(x, unit = "auto", ...)

\method{query_db}{CacheManager}(x, sql, params = NULL)

\method{list_db_tables}{CacheManager}(x)

\method{get_db_table_info}{CacheManager}(x, tableName)
}
\arguments{
\item{db}{Path to Cache.duckdb file}

\item{x}{CacheManager object}

\item{name}{Name of cache entry (not used, for consistency)}

\item{hash}{Hash string for cache table}

\item{description}{Description of cache entry}

\item{data}{Any R object to cache (will be serialized)}

\item{unit}{Optional unit for size, possible values: "bytes", "KB", "MB", "GB" or "auto" (default).}

\item{sql}{SQL query string}

\item{params}{Query parameters (optional)}

\item{tableName}{Table name}
}
\value{
An object of class 'CacheManager'

Cached R object (deserialized)

Data.frame with cache metadata

Integer
}
\description{
S3 class to manage Cache.duckdb file for caching and loading data
}
\section{Methods (by generic)}{
\itemize{
\item \code{save_cache(CacheManager)}: Save data to cache table.

\item \code{load_cache(CacheManager)}: Load data from cache table.

\item \code{get_cache_info(CacheManager)}: Get cache information table.

\item \code{clear_cache(CacheManager)}: Reset cache (delete all cache entries)

\item \code{size(CacheManager)}: Get size of Cache.duckdb file in bytes.

\item \code{query_db(CacheManager)}: Internal: execute a query on the DB.

\item \code{list_db_tables(CacheManager)}: Internal: list tables in the DB.

\item \code{get_db_table_info(CacheManager)}: Internal: get table info from the DB.

}}
