% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_005_peakQuality.R
\name{calculateFeaturesMetadata}
\alias{calculateFeaturesMetadata}
\title{calculateFeaturesMetadata}
\usage{
calculateFeaturesMetadata(object, targets = NULL)
}
\arguments{
\item{object}{An \linkS4class{ntsData} object containing features.}

\item{targets}{Optional character vector with features \emph{id}/s
for calculating metadata for the specified features.}
}
\value{
An \linkS4class{msData} object containing features and peaks
with amended quality indicators.
}
\description{
Function to calculate feature quality indicators
using the \pkg{MetaClean} package via the \pkg{patRoon} interface.
See \code{?MetaClean} for further details.
}
