% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ms_file_functions.R
\name{check_compatible_ms_formats_for_conversion}
\alias{check_compatible_ms_formats_for_conversion}
\title{check_compatible_ms_formats_for_conversion}
\usage{
check_compatible_ms_formats_for_conversion()
}
\value{
A \code{data.frame} with the formats compatible for conversion.
}
\description{
List of possible vendor formats to be converted to
\code{mzML} or \code{mzXML}.
}
\references{
\insertRef{proteo01}{streamFind}

\insertRef{proteo02}{streamFind}
}
