% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ms_checkOverlapRanges.R
\name{checkOverlapRanges}
\alias{checkOverlapRanges}
\title{checkOverlapRanges}
\usage{
checkOverlapRanges(vals, ranges)
}
\arguments{
\item{vals}{A numeric vector.}

\item{ranges}{A two columns data.frame with min and max ranges.}
}
\value{
A logical vector with \code{TRUE} for values within the ranges.
}
\description{
Helper function to check overlap ranges in a data.frame for
a given numeric vector.
}
