% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_clear_cache.R
\name{clear_cache.character}
\alias{clear_cache.character}
\title{clear_cache}
\usage{
\method{clear_cache}{character}(x, file = getOption("StreamFind_cache_path"))
}
\arguments{
\item{x}{A character vector with the names of the cache categories to clear. An integer
vector with the indices of the categories to clear can alternatively be given to remove
categories. If \code{NULL} (the default), the entire cache is cleared. Use the method
\code{get_cache_info} to get the cached categories.}

\item{file}{A character string specifying the cache file to use. Default is "cache.sqlite".}
}
\description{
Clears cached data using the approach as in the patRoon package.
}
\references{
\insertRef{patroon01}{StreamFind}

\insertRef{patroon02}{StreamFind}
}
