% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_clear_cache.R
\name{clear_cache}
\alias{clear_cache}
\title{clear_cache}
\usage{
clear_cache(what = NULL)
}
\description{
Clears cached data using the approach as in the patRoon package.
}
\references{
\insertRef{patroon01}{StreamFind}

\insertRef{patroon02}{StreamFind}
}
