% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ms_data_access.R
\name{clusterMSn}
\alias{clusterMSn}
\title{clusterMsn}
\usage{
clusterMSn(
  ids,
  msnList,
  clusteringMethod,
  clusteringUnit,
  clusteringWindow,
  mergeVoltages,
  mergeBy,
  targets
)
}
\arguments{
\item{ids}{...}

\item{msnList}{...}

\item{clusteringMethod}{...}

\item{clusteringUnit}{...}

\item{clusteringWindow}{...}

\item{mergeVoltages}{...}

\item{mergeBy}{...}

\item{targets}{...}
}
\value{
A data table with clustered MSn data for given targets.
}
\description{
Function to cluster MSn data.
}
