% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_R6_MassSpecData.R
\name{combine_MassSpecData}
\alias{combine_MassSpecData}
\title{Function to combine MassSpecData class objects.}
\usage{
combine_MassSpecData(combineFeatureLists = TRUE, ...)
}
\arguments{
\item{combineFeatureLists}{Logical, set to TRUE to combine feature lists.}

\item{...}{\emph{MassSpecData} class object.}
}
\value{
A \emph{MassSpecData} class object.
}
\description{
Function to combine MassSpecData class objects.
}
