% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_MassSpecEngine.R
\name{combine_MassSpecEngine}
\alias{combine_MassSpecEngine}
\title{Function to combine MassSpecEngine class objects.}
\usage{
combine_MassSpecEngine(combineFeatureLists = TRUE, ...)
}
\arguments{
\item{combineFeatureLists}{Logical, set to TRUE to combine feature lists.}

\item{...}{\emph{MassSpecEngine} class object.}
}
\value{
A \emph{MassSpecEngine} class object.
}
\description{
Function to combine MassSpecEngine class objects.
}
