% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ms_convert_files.R
\name{compatibleFileFormatsForConversion}
\alias{compatibleFileFormatsForConversion}
\title{compatibleFileFormatsForConversion}
\usage{
compatibleFileFormatsForConversion()
}
\value{
A data.table with the formats compatible for conversion.
}
\description{
compatibleFileFormatsForConversion
}
\references{
\insertRef{proteo01}{streamFind}

\insertRef{proteo02}{streamFind}
}
