% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Config.R
\name{ConfigDurationNotifications}
\alias{ConfigDurationNotifications}
\alias{validate_object.ConfigDurationNotifications}
\title{Configuration Parameter for Duration of Notifications}
\usage{
ConfigDurationNotifications(value = 10)

\method{validate_object}{ConfigDurationNotifications}(x)
}
\arguments{
\item{value}{Duration in seconds for pop-up notifications.}

\item{x}{A \code{ConfigDurationNotifications} object.}
}
\description{
Class representing a configuration for the duration of pop-up notifications in the
app, inhiberiting from \link{ConfigParameter}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate_object(ConfigDurationNotifications)}: Validates a \code{ConfigDurationNotifications} object,
returning NULL if valid.

}}
