% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Config.R
\name{ConfigExtraRoots}
\alias{ConfigExtraRoots}
\title{ConfigExtraRoots Class}
\usage{
ConfigExtraRoots(value = "")
}
\arguments{
\item{value}{Character string representing extra root directories for file selection.}
}
\description{
Class representing a configuration for extra root directories for file selection
in the app, inheriting from \link{ConfigParameter}.
}
