% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Config.R
\name{ConfigParameter}
\alias{ConfigParameter}
\title{Generic Configuration Parameter}
\usage{
ConfigParameter(name = NA_character_, description = NA_character_)
}
\arguments{
\item{name}{Name of the parameter.}

\item{description}{Description of the parameter.}
}
\description{
Class representing a configuration parameter.
}
