% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_convert_trim_files.R
\name{convert_ms_files}
\alias{convert_ms_files}
\title{Function to convert vendor MS files to \emph{mzML} or \emph{mzXML}}
\usage{
convert_ms_files(
  files = NULL,
  outputFormat = "mzML",
  outputPath = NULL,
  optList = NULL
)
}
\arguments{
\item{files}{A vector with file full path/s for conversion.}

\item{outputFormat}{A character string with the format of the output file.
Possible values are "mzML" and "mzXML".}

\item{outputPath}{The directory to place the output file. When \code{NULL}
the directory of the original file is used.}

\item{optList}{A list of further options to be passed to the respective
conversion tool. See documentation in
\url{https://proteowizard.sourceforge.io/tools/msconvert.html}
for possible commands.}
}
\description{
Function to convert vendor MS files to \emph{.mzML} or \emph{.mzXML}
using the command line interface of
\href{https://proteowizard.sourceforge.io/download.html}{MSConvert}
from \href{https://proteowizard.sourceforge.io/}{ProteoWizard}.
}
\note{
When giving paths make sure that single backslash is changed to
double backslash or to single forward slash.
}
\examples{
## The full file path
#> file = E:/ms_file.d

## Options to centroid MS1 and MS2 data in file
#> optList <- list(filter = "peakPicking vendor msLevel=1-2")

## Convert to mzML
#> convert_ms_files(file, outfile = "mzML", optList = optList)


}
\references{
\insertRef{proteo01}{streamfind}

\insertRef{proteo02}{streamfind}
}
