% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ms_convert_files.R
\name{convertFiles}
\alias{convertFiles}
\title{convertFiles}
\usage{
convertFiles(
  files = NULL,
  type = "ms",
  outfile = "mzML",
  outdir = NULL,
  opts_list = NULL
)
}
\arguments{
\item{files}{A vector with file full paths for conversion.}

\item{type}{A character string indicating the type of files for conversion.
Possible values are "ms".}

\item{outfile}{A character string with the format of the output file.
Possible values are "mzML" and "mzXML".}

\item{outdir}{The directory to place the output file. When \code{NULL}
the directory of the original file is used.}

\item{opts_list}{A list of further options to be passed to the respective
conversion tool. See \url{https://proteowizard.sourceforge.io/tools/msconvert.html}
for commands available for conversion of MS files with
\href{https://proteowizard.sourceforge.io/download.html}{msConvert}.}
}
\description{
Function to convert vendor MS file formats to mzML/mzXML.
}
\references{
\insertRef{proteo01}{streamFind}

\insertRef{proteo02}{streamFind}
}
