% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_R6_CoreEngine.R
\name{CoreEngine}
\alias{CoreEngine}
\title{\strong{CoreEngine} R6 class and methods}
\description{
The \emph{CoreEngine} R6 class is a basic data processor with generic methods for handling data.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{help}}{(\code{list()})\cr
List of function elements to access specific reference help pages.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CoreEngine-new}{\code{CoreEngine$new()}}
\item \href{#method-CoreEngine-print}{\code{CoreEngine$print()}}
\item \href{#method-CoreEngine-print_headers}{\code{CoreEngine$print_headers()}}
\item \href{#method-CoreEngine-print_analyses}{\code{CoreEngine$print_analyses()}}
\item \href{#method-CoreEngine-print_workflow}{\code{CoreEngine$print_workflow()}}
\item \href{#method-CoreEngine-get_headers}{\code{CoreEngine$get_headers()}}
\item \href{#method-CoreEngine-get_history}{\code{CoreEngine$get_history()}}
\item \href{#method-CoreEngine-get_analyses}{\code{CoreEngine$get_analyses()}}
\item \href{#method-CoreEngine-get_number_analyses}{\code{CoreEngine$get_number_analyses()}}
\item \href{#method-CoreEngine-get_overview}{\code{CoreEngine$get_overview()}}
\item \href{#method-CoreEngine-get_analysis_names}{\code{CoreEngine$get_analysis_names()}}
\item \href{#method-CoreEngine-get_replicate_names}{\code{CoreEngine$get_replicate_names()}}
\item \href{#method-CoreEngine-get_blank_names}{\code{CoreEngine$get_blank_names()}}
\item \href{#method-CoreEngine-get_files}{\code{CoreEngine$get_files()}}
\item \href{#method-CoreEngine-get_formats}{\code{CoreEngine$get_formats()}}
\item \href{#method-CoreEngine-get_types}{\code{CoreEngine$get_types()}}
\item \href{#method-CoreEngine-get_workflow_overview}{\code{CoreEngine$get_workflow_overview()}}
\item \href{#method-CoreEngine-add_headers}{\code{CoreEngine$add_headers()}}
\item \href{#method-CoreEngine-add_settings}{\code{CoreEngine$add_settings()}}
\item \href{#method-CoreEngine-add_analyses}{\code{CoreEngine$add_analyses()}}
\item \href{#method-CoreEngine-add_replicate_names}{\code{CoreEngine$add_replicate_names()}}
\item \href{#method-CoreEngine-add_blank_names}{\code{CoreEngine$add_blank_names()}}
\item \href{#method-CoreEngine-add_metadata}{\code{CoreEngine$add_metadata()}}
\item \href{#method-CoreEngine-add_modules_data}{\code{CoreEngine$add_modules_data()}}
\item \href{#method-CoreEngine-remove_headers}{\code{CoreEngine$remove_headers()}}
\item \href{#method-CoreEngine-remove_settings}{\code{CoreEngine$remove_settings()}}
\item \href{#method-CoreEngine-remove_analyses}{\code{CoreEngine$remove_analyses()}}
\item \href{#method-CoreEngine-subset_analyses}{\code{CoreEngine$subset_analyses()}}
\item \href{#method-CoreEngine-has_analyses}{\code{CoreEngine$has_analyses()}}
\item \href{#method-CoreEngine-has_settings}{\code{CoreEngine$has_settings()}}
\item \href{#method-CoreEngine-run_workflow}{\code{CoreEngine$run_workflow()}}
\item \href{#method-CoreEngine-save_headers}{\code{CoreEngine$save_headers()}}
\item \href{#method-CoreEngine-save_settings}{\code{CoreEngine$save_settings()}}
\item \href{#method-CoreEngine-save_analyses}{\code{CoreEngine$save_analyses()}}
\item \href{#method-CoreEngine-save}{\code{CoreEngine$save()}}
\item \href{#method-CoreEngine-import_headers}{\code{CoreEngine$import_headers()}}
\item \href{#method-CoreEngine-import_settings}{\code{CoreEngine$import_settings()}}
\item \href{#method-CoreEngine-import_analyses}{\code{CoreEngine$import_analyses()}}
\item \href{#method-CoreEngine-import}{\code{CoreEngine$import()}}
\item \href{#method-CoreEngine-processing_methods}{\code{CoreEngine$processing_methods()}}
\item \href{#method-CoreEngine-clone}{\code{CoreEngine$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreEngine-new"></a>}}
\if{latex}{\out{\hypertarget{method-CoreEngine-new}{}}}
\subsection{Method \code{new()}}{
Creates an R6 \emph{CoreEngine} class object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreEngine$new(headers = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{headers}}{A named list with headers or a ProjectHeaders S3 class object.
Each list element must have length one. There is no type restrictions
except for name (must be type character length 1), author (must be type
character length 1), path (must be type character length 1) and date (must
be class POSIXct or class POSIXt length 1). See more information in
\code{?ProjectHeaders}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreEngine-print"></a>}}
\if{latex}{\out{\hypertarget{method-CoreEngine-print}{}}}
\subsection{Method \code{print()}}{
Prints a summary to the console.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreEngine$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreEngine-print_headers"></a>}}
\if{latex}{\out{\hypertarget{method-CoreEngine-print_headers}{}}}
\subsection{Method \code{print_headers()}}{
Prints the headers.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreEngine$print_headers()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreEngine-print_analyses"></a>}}
\if{latex}{\out{\hypertarget{method-CoreEngine-print_analyses}{}}}
\subsection{Method \code{print_analyses()}}{
Prints the analyses present.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreEngine$print_analyses()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreEngine-print_workflow"></a>}}
\if{latex}{\out{\hypertarget{method-CoreEngine-print_workflow}{}}}
\subsection{Method \code{print_workflow()}}{
Prints all processing methods present by order.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreEngine$print_workflow()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreEngine-get_headers"></a>}}
\if{latex}{\out{\hypertarget{method-CoreEngine-get_headers}{}}}
\subsection{Method \code{get_headers()}}{
Gets the headers.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreEngine$get_headers(value = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A character vector with the name/s of the header elements.
When \code{NULL} (the default), the entire headers list is returned.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The headers list or the header elements as defined by \code{value}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreEngine-get_history"></a>}}
\if{latex}{\out{\hypertarget{method-CoreEngine-get_history}{}}}
\subsection{Method \code{get_history()}}{
Gets the object history.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreEngine$get_history()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The history list of processing steps applied.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreEngine-get_analyses"></a>}}
\if{latex}{\out{\hypertarget{method-CoreEngine-get_analyses}{}}}
\subsection{Method \code{get_analyses()}}{
Gets analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreEngine$get_analyses(analyses = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name of the analyses.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The list of analyses or the analyses as defined by \code{analyses}
argument.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreEngine-get_number_analyses"></a>}}
\if{latex}{\out{\hypertarget{method-CoreEngine-get_number_analyses}{}}}
\subsection{Method \code{get_number_analyses()}}{
Gets the number of analyses present.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreEngine$get_number_analyses()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreEngine-get_overview"></a>}}
\if{latex}{\out{\hypertarget{method-CoreEngine-get_overview}{}}}
\subsection{Method \code{get_overview()}}{
Gets an overview data.table with all the analysis names,
replicates, associated blank replicates, and full file paths.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreEngine$get_overview()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data.table with columns analysis, replicate, blank and file.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreEngine-get_analysis_names"></a>}}
\if{latex}{\out{\hypertarget{method-CoreEngine-get_analysis_names}{}}}
\subsection{Method \code{get_analysis_names()}}{
Gets the analysis names.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreEngine$get_analysis_names(analyses = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name of the analyses.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreEngine-get_replicate_names"></a>}}
\if{latex}{\out{\hypertarget{method-CoreEngine-get_replicate_names}{}}}
\subsection{Method \code{get_replicate_names()}}{
Gets the analysis replicate names.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreEngine$get_replicate_names(analyses = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name of the analyses.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreEngine-get_blank_names"></a>}}
\if{latex}{\out{\hypertarget{method-CoreEngine-get_blank_names}{}}}
\subsection{Method \code{get_blank_names()}}{
Gets the analysis blank replicate names.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreEngine$get_blank_names(analyses = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name of the analyses.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreEngine-get_files"></a>}}
\if{latex}{\out{\hypertarget{method-CoreEngine-get_files}{}}}
\subsection{Method \code{get_files()}}{
Gets the full file paths of each analysis.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreEngine$get_files(analyses = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name of the analyses.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreEngine-get_formats"></a>}}
\if{latex}{\out{\hypertarget{method-CoreEngine-get_formats}{}}}
\subsection{Method \code{get_formats()}}{
Gets the file format of each analysis.'
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreEngine$get_formats(analyses = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name of the analyses.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreEngine-get_types"></a>}}
\if{latex}{\out{\hypertarget{method-CoreEngine-get_types}{}}}
\subsection{Method \code{get_types()}}{
Gets the type of each analysis.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreEngine$get_types(analyses = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name of the analyses.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreEngine-get_workflow_overview"></a>}}
\if{latex}{\out{\hypertarget{method-CoreEngine-get_workflow_overview}{}}}
\subsection{Method \code{get_workflow_overview()}}{
A data.table with the overview of all processing methods present.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreEngine$get_workflow_overview()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreEngine-add_headers"></a>}}
\if{latex}{\out{\hypertarget{method-CoreEngine-add_headers}{}}}
\subsection{Method \code{add_headers()}}{
Adds headers. If an argument or element "name" is given,
it must be type character. If an argument or element path is given, it
must be type character and exist. If an argument or element date is
given, it must be class POSIXct or POSIXt. If given date is character,
conversion to class POSIXct or POSIXt is attempted.
See \code{?ProjectHeaders} for more information.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreEngine$add_headers(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments to be added as headers. Note that all given arguments
must be of length one. There is no type limitations. The name of the argument
becomes the name of the header entry. Alternatively, a single list argument
with the headers (all length one) can be given.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreEngine-add_settings"></a>}}
\if{latex}{\out{\hypertarget{method-CoreEngine-add_settings}{}}}
\subsection{Method \code{add_settings()}}{
Adds processing settings.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreEngine$add_settings(settings = NULL, replace = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{settings}}{A named list of ProcessingSettings S3 class objects or a
single ProcessingSettings S3 class object. The list names should match
the call name of each ProcessingSettings object. Alternatively, a named
list with call name, algorithm and parameters to be transformed and added
as ProcessingSettings S3 class object.}

\item{\code{replace}}{Logical of length one. When \code{TRUE}, existing settings are
replaced by the new settings with the same call name, except settings for
methods that can be applied more than once.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreEngine-add_analyses"></a>}}
\if{latex}{\out{\hypertarget{method-CoreEngine-add_analyses}{}}}
\subsection{Method \code{add_analyses()}}{
Adds analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreEngine$add_analyses(analyses = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{An S3 class object or a list with
analyses as elements.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreEngine-add_replicate_names"></a>}}
\if{latex}{\out{\hypertarget{method-CoreEngine-add_replicate_names}{}}}
\subsection{Method \code{add_replicate_names()}}{
Adds or redefines the analysis replicate names.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreEngine$add_replicate_names(value = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A character vector with the analysis replicate names.
Must be of the same length as the number of analyses.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreEngine-add_blank_names"></a>}}
\if{latex}{\out{\hypertarget{method-CoreEngine-add_blank_names}{}}}
\subsection{Method \code{add_blank_names()}}{
Adds or redefines the analysis blank replicate names.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreEngine$add_blank_names(value = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A character vector with the analysis blank replicate names.
Must be of the same length as the number of analyses.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreEngine-add_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-CoreEngine-add_metadata}{}}}
\subsection{Method \code{add_metadata()}}{
Adds metadata to analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreEngine$add_metadata(value = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A data.frame or data.table with metadata for the analyses.
The data.frame or data.table must have an analysis column and the same
number of rows as the number of analyses in the engine. Metadata is added
using any extra columns of the data.frame or data.table.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreEngine-add_modules_data"></a>}}
\if{latex}{\out{\hypertarget{method-CoreEngine-add_modules_data}{}}}
\subsection{Method \code{add_modules_data()}}{
Adds data from modules to the engine.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreEngine$add_modules_data(value = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A named list with data from modules.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreEngine-remove_headers"></a>}}
\if{latex}{\out{\hypertarget{method-CoreEngine-remove_headers}{}}}
\subsection{Method \code{remove_headers()}}{
Removes headers entries. Note that the name, path and date
headers cannot be removed only changed.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreEngine$remove_headers(value = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A character vector with the name/s of the elements in headers
to be removed.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreEngine-remove_settings"></a>}}
\if{latex}{\out{\hypertarget{method-CoreEngine-remove_settings}{}}}
\subsection{Method \code{remove_settings()}}{
Removes settings.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreEngine$remove_settings(call = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{call}}{A string or a vector of strings with the name/s of the
processing method/s to be removed. Alternatively, an integer vector
with the index/es of the settings to be removed. When \code{call} is
\code{NULL} all settings are removed.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreEngine-remove_analyses"></a>}}
\if{latex}{\out{\hypertarget{method-CoreEngine-remove_analyses}{}}}
\subsection{Method \code{remove_analyses()}}{
Removes analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreEngine$remove_analyses(analyses = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name of the analyses.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreEngine-subset_analyses"></a>}}
\if{latex}{\out{\hypertarget{method-CoreEngine-subset_analyses}{}}}
\subsection{Method \code{subset_analyses()}}{
Subsets on analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreEngine$subset_analyses(analyses = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name of the analyses.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new cloned object with only the analyses as
defined by the \code{analyses} argument.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreEngine-has_analyses"></a>}}
\if{latex}{\out{\hypertarget{method-CoreEngine-has_analyses}{}}}
\subsection{Method \code{has_analyses()}}{
Checks if analyses are present.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreEngine$has_analyses()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Logical value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreEngine-has_settings"></a>}}
\if{latex}{\out{\hypertarget{method-CoreEngine-has_settings}{}}}
\subsection{Method \code{has_settings()}}{
Checks if there are processing settings.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreEngine$has_settings(call = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{call}}{A string or a vector of strings with the name/s of the
processing method/s.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Logical value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreEngine-run_workflow"></a>}}
\if{latex}{\out{\hypertarget{method-CoreEngine-run_workflow}{}}}
\subsection{Method \code{run_workflow()}}{
Runs all modules represented by the added ProcessingSettings.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreEngine$run_workflow()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreEngine-save_headers"></a>}}
\if{latex}{\out{\hypertarget{method-CoreEngine-save_headers}{}}}
\subsection{Method \code{save_headers()}}{
Saves the headers list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreEngine$save_headers(format = "json", name = "headers", path = getwd())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{format}}{A string. Possible values are \code{json} and \code{rds} for saving as
JSON or RDS formats, respectively.}

\item{\code{name}}{String with the name of the file to be saved.}

\item{\code{path}}{String with the directory of the file to be saved.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Saves the headers list as the defined \code{format} in
\code{path} and returns invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreEngine-save_settings"></a>}}
\if{latex}{\out{\hypertarget{method-CoreEngine-save_settings}{}}}
\subsection{Method \code{save_settings()}}{
Saves settings.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreEngine$save_settings(
  call = NULL,
  format = "json",
  name = "settings",
  path = getwd()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{call}}{A string or a vector of strings with the name/s of the
processing method/s to be saved. When \code{call} is \code{NULL} all
settings are saved.}

\item{\code{format}}{A string. Possible values are \code{json} and \code{rds} for saving as
JSON or RDS formats, respectively.}

\item{\code{name}}{String with the name of the file to be saved.}

\item{\code{path}}{String with the directory of the file to be saved.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Saves the settings list as the defined \code{format} in
\code{path} and returns invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreEngine-save_analyses"></a>}}
\if{latex}{\out{\hypertarget{method-CoreEngine-save_analyses}{}}}
\subsection{Method \code{save_analyses()}}{
Saves analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreEngine$save_analyses(
  analyses = NULL,
  format = "json",
  name = "analyses",
  path = getwd()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name of the analyses.}

\item{\code{format}}{A string. Possible values are \code{json} and \code{rds} for saving as
JSON or RDS formats, respectively.}

\item{\code{name}}{String with the name of the file to be saved.}

\item{\code{path}}{String with the directory of the file to be saved.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Saves the list of analyses as the defined \code{format} in
\code{path} and returns invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreEngine-save"></a>}}
\if{latex}{\out{\hypertarget{method-CoreEngine-save}{}}}
\subsection{Method \code{save()}}{
Saves the private fields of the engine.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreEngine$save(format = "json", name = "EngineData", path = getwd())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{format}}{A string. Possible values are \code{json} and \code{rds} for saving as
JSON or RDS formats, respectively.}

\item{\code{name}}{String with the name of the file to be saved.}

\item{\code{path}}{String with the directory of the file to be saved.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Saves the private fields of the engine as the defined \code{format}
in the \code{path} and returns invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreEngine-import_headers"></a>}}
\if{latex}{\out{\hypertarget{method-CoreEngine-import_headers}{}}}
\subsection{Method \code{import_headers()}}{
Imports headers from a \emph{rds} or \emph{json} file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreEngine$import_headers(file = NA_character_)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{A \emph{json} or \emph{rds} file as obtained by the MassSpecData
method \code{save()}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreEngine-import_settings"></a>}}
\if{latex}{\out{\hypertarget{method-CoreEngine-import_settings}{}}}
\subsection{Method \code{import_settings()}}{
Imports processing settings from a \emph{rds} or \emph{json} file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreEngine$import_settings(file = NA_character_, replace = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{A \emph{json} or \emph{rds} file as obtained by the MassSpecData
method \code{save()}.}

\item{\code{replace}}{Logical. When \code{TRUE}, existing settings are replaced by
the new settings with the same call name.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreEngine-import_analyses"></a>}}
\if{latex}{\out{\hypertarget{method-CoreEngine-import_analyses}{}}}
\subsection{Method \code{import_analyses()}}{
Imports analyses from a \emph{rds} or \emph{json} file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreEngine$import_analyses(file = NA_character_)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{A \emph{json} or \emph{rds} file as obtained by the MassSpecData
method \code{save()}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreEngine-import"></a>}}
\if{latex}{\out{\hypertarget{method-CoreEngine-import}{}}}
\subsection{Method \code{import()}}{
Imports a \emph{CoreEngine} saved as \emph{json}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreEngine$import(file = NA_character_)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{A \emph{json} file representing a \emph{CoreEngine}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreEngine-processing_methods"></a>}}
\if{latex}{\out{\hypertarget{method-CoreEngine-processing_methods}{}}}
\subsection{Method \code{processing_methods()}}{
Possible data processing methods.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreEngine$processing_methods()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character vector with ordered possible data  processing methods.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreEngine-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CoreEngine-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreEngine$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
