% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_R6_msData.R
\name{correct_ms_parsed_json_analyses}
\alias{correct_ms_parsed_json_analyses}
\title{correct_ms_parsed_json_analyses}
\usage{
correct_ms_parsed_json_analyses(analyses = NULL)
}
\arguments{
\item{analyses}{X.}
}
\value{
X.
}
\description{
Function to correct parsed analyses from json file.
}
