% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_R6_msData.R
\name{correct_ms_parsed_json_groups}
\alias{correct_ms_parsed_json_groups}
\title{correct_ms_parsed_json_groups}
\usage{
correct_ms_parsed_json_groups(groups = NULL)
}
\arguments{
\item{groups}{X.}
}
\value{
X.
}
\description{
Function to correct parsed feature groups from json file.
}
