% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_R6_msData.R
\name{correct_parsed_json_header}
\alias{correct_parsed_json_header}
\title{correct_parsed_json_header}
\usage{
correct_parsed_json_header(header = NULL)
}
\arguments{
\item{header}{X.}
}
\value{
X.
}
\description{
Function to correct parsed header from json file.
}
