% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_R6_msData.R
\name{correct_parsed_json_headers}
\alias{correct_parsed_json_headers}
\title{correct_parsed_json_headers}
\usage{
correct_parsed_json_headers(headers = NULL)
}
\arguments{
\item{headers}{X.}
}
\value{
X.
}
\description{
Function to correct parsed headers from json file.
}
