% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_R6_msData.R
\name{correct_parsed_json_settings}
\alias{correct_parsed_json_settings}
\title{correct_parsed_json_settings}
\usage{
correct_parsed_json_settings(settings = NULL)
}
\arguments{
\item{settings}{X.}
}
\value{
X.
}
\description{
Function to correct parsed settings from json file.
}
