% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ms.R
\name{correlate_analysis_spectra}
\alias{correlate_analysis_spectra}
\title{correlate_analysis_spectra}
\usage{
correlate_analysis_spectra(
  spectra,
  splitSpectra = FALSE,
  byReplicates = FALSE,
  decimals = 2,
  minIntensity = 1000,
  method = "pearson"
)
}
\arguments{
\item{spectra}{A data.table with columns "analysis", "mz" and "intensity".
Optionally, a column named "id" or "group" can be given to split the
data.table before correlation analysis by setting the argument
\code{splitSpectra} to \code{TRUE}. Note that when both "id" and "group"
columns are present "group" is used for splitting the data.table not "id".
If a column "replicate" is present and the argument \code{byReplicates}
is set to \code{TRUE}, the correlation is performed by replicate analysis
groups.}

\item{splitSpectra}{X.}

\item{byReplicates}{X.}

\item{decimals}{X.}

\item{minIntensity}{X.}

\item{method}{X.}
}
\value{
X.
}
\description{
Function to correlate MS spectra from analyses.
}
