% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_methods_datatable.R
\name{plotEICs,data.table-method}
\alias{plotEICs,data.table-method}
\alias{plotEICs,data.table,data.table-method}
\alias{plotTICs,data.table-method}
\alias{plotTICs,data.table,data.table-method}
\alias{plotXICs,data.table-method}
\alias{plotXICs,data.table,data.table-method}
\alias{plotPeaks,data.table-method}
\alias{plotPeaks,data.table,data.table-method}
\title{plotEICs-data.table}
\usage{
\S4method{plotEICs}{data.table}(
  object,
  analyses = NULL,
  colorBy = "targets",
  legendNames = NULL,
  title = NULL,
  interactive = FALSE
)

\S4method{plotTICs}{data.table}(
  object,
  analyses = NULL,
  colorBy = "analyses",
  title = NULL,
  interactive = FALSE
)

\S4method{plotXICs}{data.table}(
  object,
  analyses = NULL,
  targets = NULL,
  legendNames = NULL,
  plotTargetMark = TRUE,
  targetsMark = NULL,
  ppmMark = 5,
  secMark = 10,
  numberRows = 1
)

\S4method{plotPeaks}{data.table}(
  object,
  pks,
  analyses = NULL,
  colorBy = "targets",
  legendNames = NULL,
  title = NULL,
  interactive = FALSE
)
}
\arguments{
\item{object}{A \link[data.table]{data.table} as produced by the method \link{EICs}.}

\item{analyses}{A numeric or character vector with the index or analysis name/s.
When \code{NULL}, all the analyses are used.}

\item{colorBy}{A length one character vector applicable to the respective method.}

\item{legendNames}{A character vector with the same length and order
as the number and order of targets to be used as plot legend.}

\item{title}{A character string with the plot tile.}

\item{interactive}{Logical value, set to \code{TRUE} to use
the \pkg{plotly} instead of \pkg{base}. The default is \code{FALSE}.}

\item{targets}{A character vector with target names.}

\item{plotTargetMark}{Logical, set to \code{TRUE} to plot a target mark.}

\item{targetsMark}{A two columns \link[data.table]{data.table} or \link{data.frame} with
\emph{m/z} and time targets. The column must be named with "mz" and "rt" for
\emph{m/z} and time values, respectively.}

\item{ppmMark}{A numeric vector of length one to define the mass deviation, in ppm,
of the target mark.}

\item{secMark}{A numeric vector of length one to define the time deviation, in seconds,
of the target mark.}

\item{numberRows}{A numeric vector of length one to define
the number of rows to grid the plots.}

\item{pks}{A table with the target peaks as obtained by the method \link{peaks}.}
}
\description{
An S4 method for plotting extracted ion chromatograms (EICs)
of data in a \link[data.table]{data.table} object obtained with the \link{EICs} method.
The \code{colorBy} argument can be be \code{"analyses"}, \code{replicates} or \code{targets}
(the default), for coloring by analyses, replicates or EICs targets (id), respectively.
The \code{legendNames} is a character vector with the same length as targets for plotting and
can be used to legend the plot. Note that, by setting \code{legendNames} the \code{colorBy}
is set to "targets" even when other colorBy is used.

Plots a total ion chromatogram (TIC) from the \link[data.table]{data.table}
obtained by the S4 method \link{TICs}. The colorBy argument can be
"analyses" or "replicates" to color the plot by analyses or by analysis replicates.

Plots three dimensional (\emph{m/z}, time and intensity)
extracted ion chromatograms (XICs) for specified \emph{m/z} and retention time pair targets
in analyses of a \link[data.table]{data.table} object as produced
by the \link{XICs} method. \code{analyses} and \code{targets} can be used to filter the XIC table.
When \code{plotTargetMark} is \code{TRUE} a target is plotted representing the deviations as defined
by the arguments \code{ppmMark} and \code{secMark} in ppm and seconds, respectively.
When ranges were given to build the XIC, exact \emph{m/z} and time targets can be specified with
the argument \code{targetsMark}. \code{targetsMark} should be a two column table named mz and rt with
exact \emph{m/z} and time targets. Note that the number of rows should be the same as the number of target
in the XIC. The number of rows to plot multiple targets can be defined by the \code{numberRows} argument.

A method for plotting EIC (\code{object}) of chromatographic peaks (\code{pks}).
The \code{colorBy} argument can be be \code{"analyses"}, \code{replicates} or \code{targets}
(the default), for coloring by analyses, replicates or target peaks (id), respectively.
The \code{legendNames} is a character vector with the same length as target peaks for plotting and
can be used to legend the plot. Note that, by setting \code{legendNames} the \code{colorBy}
is set to "targets" even when other colorBy is used.
}
