% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_internal_DataTypes.R
\name{DataTypeObjects}
\alias{DataTypeObjects}
\title{DataTypeObjects Class}
\usage{
DataTypeObjects(data_type = NULL)
}
\arguments{
\item{data_type}{Optional; specific data type to retrieve information for. If NULL, returns all data types.}
}
\description{
Constructor function for DataTypeObjects class, defining data types used in the package, associated file formats, and result classes.
}
