% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_DB_Cache.R
\name{DB_Cache}
\alias{DB_Cache}
\alias{save_cache.DB_Cache}
\alias{load_cache.DB_Cache}
\alias{get_cache_info.DB_Cache}
\alias{clear_cache.DB_Cache}
\alias{size.DB_Cache}
\alias{query_db.DB_Cache}
\alias{list_db_tables.DB_Cache}
\alias{get_db_table_info.DB_Cache}
\title{DB_Cache S3 Class}
\usage{
DB_Cache(projectPath = ".")

\method{save_cache}{DB_Cache}(x, name, hash, description, data, ...)

\method{load_cache}{DB_Cache}(x, name, hash, ...)

\method{get_cache_info}{DB_Cache}(x, ...)

\method{clear_cache}{DB_Cache}(x, value = NULL, ...)

\method{size}{DB_Cache}(x, unit = "auto", ...)

\method{query_db}{DB_Cache}(x, sql, params = NULL)

\method{list_db_tables}{DB_Cache}(x)

\method{get_db_table_info}{DB_Cache}(x, tableName)
}
\arguments{
\item{projectPath}{Path to the project directory where this database file will be created/used.}

\item{x}{DB_Cache object}

\item{name}{Name of cache entry (not used, for consistency)}

\item{hash}{Hash string for cache table}

\item{description}{Description of cache entry}

\item{data}{Any R object to cache (will be serialized)}

\item{value}{Character vector of cache entry names to delete. If NULL (default), deletes entire cache database.}

\item{unit}{Optional unit for size, possible values: "bytes", "KB", "MB", "GB" or "auto" (default).}

\item{sql}{SQL query string}

\item{params}{Query parameters (optional)}

\item{tableName}{Table name}
}
\value{
An object of class \code{DB_Cache}.

Cached R object (deserialized)

Data.frame with cache metadata

Integer
}
\description{
S3 class to manage \code{DB_Cache.duckdb} for caching and loading data.
}
\section{Methods (by generic)}{
\itemize{
\item \code{save_cache(DB_Cache)}: Save data to cache table.

\item \code{load_cache(DB_Cache)}: Load data from cache table.

\item \code{get_cache_info(DB_Cache)}: Get cache information table.

\item \code{clear_cache(DB_Cache)}: Reset cache (delete all cache entries or specific entries by name)

\item \code{size(DB_Cache)}: Get size of Cache.duckdb file in bytes.

\item \code{query_db(DB_Cache)}: Internal: execute a query on the DB.

\item \code{list_db_tables(DB_Cache)}: Internal: list tables in the DB.

\item \code{get_db_table_info(DB_Cache)}: Internal: get table info from the DB.

}}
