% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_DB_MassSpecAnalyses.R
\name{DB_MassSpecAnalyses}
\alias{DB_MassSpecAnalyses}
\alias{validate_object.DB_MassSpecAnalyses}
\alias{add_analyses.DB_MassSpecAnalyses}
\alias{remove_analyses.DB_MassSpecAnalyses}
\alias{get_analysis_names.DB_MassSpecAnalyses}
\alias{get_replicate_names.DB_MassSpecAnalyses}
\alias{get_blank_names.DB_MassSpecAnalyses}
\alias{set_replicate_names.DB_MassSpecAnalyses}
\alias{set_blank_names.DB_MassSpecAnalyses}
\alias{get_concentrations.DB_MassSpecAnalyses}
\alias{set_concentrations.DB_MassSpecAnalyses}
\alias{info.DB_MassSpecAnalyses}
\alias{show.DB_MassSpecAnalyses}
\alias{get_spectra_headers.DB_MassSpecAnalyses}
\alias{get_spectra_tic.DB_MassSpecAnalyses}
\alias{plot_spectra_tic.DB_MassSpecAnalyses}
\alias{get_spectra_bpc.DB_MassSpecAnalyses}
\alias{plot_spectra_bpc.DB_MassSpecAnalyses}
\alias{get_raw_spectra.DB_MassSpecAnalyses}
\alias{get_spectra_eic.DB_MassSpecAnalyses}
\alias{get_spectra_ms1.DB_MassSpecAnalyses}
\alias{get_spectra_ms2.DB_MassSpecAnalyses}
\alias{get_chromatograms_headers.DB_MassSpecAnalyses}
\alias{get_chromatograms.DB_MassSpecAnalyses}
\alias{plot_chromatograms.DB_MassSpecAnalyses}
\alias{query_db.DB_MassSpecAnalyses}
\alias{list_db_tables.DB_MassSpecAnalyses}
\alias{get_db_table_info.DB_MassSpecAnalyses}
\title{Database-backed Mass Spectrometry Analyses}
\usage{
DB_MassSpecAnalyses(
  db = file.path("data.sf", "DB_MassSpecAnalyses.duckdb"),
  files = NULL,
  centroid = FALSE,
  levels = c(1, 2)
)

\method{validate_object}{DB_MassSpecAnalyses}(x)

\method{add_analyses}{DB_MassSpecAnalyses}(x, files, centroid = FALSE, levels = c(1, 2))

\method{remove_analyses}{DB_MassSpecAnalyses}(x, analyses)

\method{get_analysis_names}{DB_MassSpecAnalyses}(x)

\method{get_replicate_names}{DB_MassSpecAnalyses}(x)

\method{get_blank_names}{DB_MassSpecAnalyses}(x)

\method{set_replicate_names}{DB_MassSpecAnalyses}(x, value)

\method{set_blank_names}{DB_MassSpecAnalyses}(x, value)

\method{get_concentrations}{DB_MassSpecAnalyses}(x)

\method{set_concentrations}{DB_MassSpecAnalyses}(x, value)

\method{info}{DB_MassSpecAnalyses}(x)

\method{show}{DB_MassSpecAnalyses}(x, ...)

\method{get_spectra_headers}{DB_MassSpecAnalyses}(x, analyses = NULL)

\method{get_spectra_tic}{DB_MassSpecAnalyses}(x, analyses = NULL, levels = c(1, 2), rt = NULL)

\method{plot_spectra_tic}{DB_MassSpecAnalyses}(
  x,
  analyses = NULL,
  levels = c(1, 2),
  rt = NULL,
  downsize = NULL,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  groupBy = "analysis",
  interactive = TRUE,
  colorPalette = NULL
)

\method{get_spectra_bpc}{DB_MassSpecAnalyses}(x, analyses = NULL, levels = c(1, 2), rt = NULL)

\method{plot_spectra_bpc}{DB_MassSpecAnalyses}(
  x,
  analyses = NULL,
  levels = c(1, 2),
  rt = NULL,
  downsize = NULL,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  groupBy = "analysis",
  interactive = TRUE,
  colorPalette = NULL
)

\method{get_raw_spectra}{DB_MassSpecAnalyses}(
  x,
  analyses = NULL,
  levels = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  id = NULL,
  allTraces = TRUE,
  isolationWindow = 1.3,
  minIntensityMS1 = 0,
  minIntensityMS2 = 0
)

\method{get_spectra_eic}{DB_MassSpecAnalyses}(
  x,
  analyses = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  id = NULL
)

\method{get_spectra_ms1}{DB_MassSpecAnalyses}(
  x,
  analyses = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  id = NULL,
  mzClust = 0.003,
  presence = 0.8,
  minIntensity = 1000
)

\method{get_spectra_ms2}{DB_MassSpecAnalyses}(
  x,
  analyses = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  id = NULL,
  isolationWindow = 1.3,
  mzClust = 0.005,
  presence = 0.8,
  minIntensity = 0
)

\method{get_chromatograms_headers}{DB_MassSpecAnalyses}(x, analyses = NULL)

\method{get_chromatograms}{DB_MassSpecAnalyses}(
  x,
  analyses = NULL,
  chromatograms = NULL,
  rtmin = 0,
  rtmax = 0,
  minIntensity = NULL
)

\method{plot_chromatograms}{DB_MassSpecAnalyses}(
  x,
  analyses = NULL,
  chromatograms = NULL,
  rtmin = 0,
  rtmax = 0,
  minIntensity = NULL,
  downsize = NULL,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  groupBy = "analysis",
  interactive = TRUE,
  colorPalette = NULL
)

\method{query_db}{DB_MassSpecAnalyses}(x, sql, params = NULL)

\method{list_db_tables}{DB_MassSpecAnalyses}(x)

\method{get_db_table_info}{DB_MassSpecAnalyses}(x, tableName)
}
\arguments{
\item{db}{Path to the MassSpec analyses DuckDB file. Defaults to \code{data.sf/MassSpecAnalyses.duckdb}.}

\item{files}{A character vector with full file path/s of \emph{mzML}
or \emph{mzXML} file/s or a data.frame with the columns \emph{file}, \emph{replicate}
and \emph{blank} with the full file path of \emph{mzML} or \emph{mzXML} file/s
(character), the replicate group name (character) and the associated blank
replicate group name (character). Note the file paths must have the extension
\emph{.mzML} or \emph{.mzXML}.}

\item{centroid}{Logical (length 1). Set to \code{TRUE} to centroid data when converting from vendor formats to mzML.}

\item{levels}{Integer vector with the levels of MS data.}

\item{x}{MassSpecAnalysesDB object.}

\item{analyses}{Character or numeric vector with names or indexes of analyses in the \code{Analyses}
object.}

\item{value}{Numeric vector of concentrations matching the number of analyses.}

\item{rt}{A vector with target retention time values (in seconds) or
a two columns data.table or data.frame with minimum and maximum retention
time values (in seconds).}

\item{downsize}{An integer of length one to downsize the TIC plot. The default is 1.}

\item{xLab}{X axis label for the plot.}

\item{yLab}{Y axis label for the plot.}

\item{title}{Plot title.}

\item{groupBy}{Column to use for color grouping (default: "analysis").}

\item{interactive}{Logical, use plotly if TRUE, ggplot2 if FALSE.}

\item{colorPalette}{Optional vector of colors for the plot; if not provided, an internal function is used.}

\item{mass}{A vector with target neutral mass value/s (in Da) or a two columns data.table or data.frame named \code{min}
and \code{max} with minimum and maximum neutral mass values (in Da), respectively. Alternatively, neutral mass (in Da)
and retention time (in seconds) and/or drift time values (in milliseconds) can be given as one data.table or
data.frame with columns named \code{mass} and \code{rt} and/or \code{drift}. Then, the deviations given in the \code{ppm}, \code{sec}
and \code{millisec}  arguments are used to calculate the ranges. Also works with a data.table or data.frame with
minimum and maximum values of neutral mass, retention time and drift time targets. In this case, the column names
must be \code{min}, \code{max}, \code{rtmin}, \code{rtmax}, \code{driftmin} and \code{driftmax}. Note that when mass/time ranges are given, the
\code{ppm}, \code{sec} and \code{millisec}
arguments are not used.}

\item{mz}{A vector with target \emph{m/z} value/s (in Da) or a two columns data.table or data.frame named \code{mzmin} and
\code{mzmax} with minimum and maximum \emph{m/z} values (in Da), respectively. Alternatively, \emph{m/z} (in Da) and
retention time values (in seconds) can be given as one data.table or data.frame with columns named \code{mz} and \code{rt}
and/or \code{drift}. Then, the deviations given in the \code{ppm}, \code{sec} and \code{millisec} arguments are used to
calculate the ranges. Also works with a data.table or data.frame with minimum and maximum values of \emph{m/z},
retention time and drift time targets. In this case, the column names must be \code{mzmin}, \code{mzmax}, \code{rtmin}, \code{rtmax},
\code{driftmin} and \code{driftmax}. Note that when mass/time ranges are given, the \code{ppm}, \code{sec} and \code{millisec}
arguments are not used.}

\item{mobility}{A vector with target drift time values (in milliseconds) or
a two columns data.table or data.frame with minimum and maximum drift
time values (in milliseconds).}

\item{ppm}{Numeric of length one with the mass deviation, in ppm.}

\item{sec}{Numeric of length one with the time deviation, in seconds.}

\item{millisec}{Numeric of length one with the drift time deviation,
in milliseconds.}

\item{id}{Character with the same length as \emph{m/z} and retention time
targets to be used as identifier/s. When not given, the id is built as a
combination of the \emph{m/z} and retention time ranges or values.}

\item{allTraces}{Logical, when \code{TRUE} all level 2 data is returned.
When \code{FALSE} and level has 2, only the MS2 traces of MS1 targets
are returned, using the \code{isolationWindow} to calculate the mass window.}

\item{isolationWindow}{Numeric value with the isolation window (in Da)
applied for ion isolation and fragmentation during acquisition of tandem
data (i.e., MS2 data).}

\item{minIntensityMS1}{Numeric value with the minimum intensity of level 1
data (i.e., MS1 data).}

\item{minIntensityMS2}{Numeric value with the minimum intensity of level 2
data (i.e., MS2 data).}

\item{mzClust}{Numeric (length 1) with the mass deviation threshold (in Da) to cluster mass traces. Highly dependent
on the mass resolution of the MS data.}

\item{presence}{Numeric (length 1) with the required presence ratio from 0 (i.e., doesn't need to be present in any
spectra) to 1 (i.e., must be present in all spectra) for traces during clustering of spectra.}

\item{minIntensity}{Numeric (length 1) with the minimum intensity.}

\item{chromatograms}{A character or integer vector with the ID (i.e. name) or the index of the chromatograms.}

\item{rtmin}{Numeric (length 1) with the minimum retention time.}

\item{rtmax}{Numeric (length 1) with the maximum retention time.}

\item{sql}{SQL query string to execute against the database.}

\item{params}{Query parameters to be used with the SQL query (optional).}

\item{tableName}{Name of the table to query or retrieve information about.}
}
\value{
An object of class \code{DB_MassSpecAnalyses}, extending \code{DB_Analyses}.
}
\description{
The \code{DB_MassSpecAnalyses} class stores MassSpec analyses metadata and headers in a DuckDB file.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate_object(DB_MassSpecAnalyses)}: Validate the DB_MassSpecAnalyses object (schema + type).

\item \code{add_analyses(DB_MassSpecAnalyses)}: Add analyses to the DB (append; overwrites only duplicates)

\item \code{remove_analyses(DB_MassSpecAnalyses)}: Remove analyses from the DB.

\item \code{get_analysis_names(DB_MassSpecAnalyses)}: Get analysis names.

\item \code{get_replicate_names(DB_MassSpecAnalyses)}: Get replicate names.

\item \code{get_blank_names(DB_MassSpecAnalyses)}: Get blank names.

\item \code{set_replicate_names(DB_MassSpecAnalyses)}: Set replicate names.

\item \code{set_blank_names(DB_MassSpecAnalyses)}: Set blank names.

\item \code{get_concentrations(DB_MassSpecAnalyses)}: Get concentrations.

\item \code{set_concentrations(DB_MassSpecAnalyses)}: Set concentrations.

\item \code{info(DB_MassSpecAnalyses)}: Get a summary table.

\item \code{show(DB_MassSpecAnalyses)}: Show method for DB_MassSpecAnalyses objects.

\item \code{get_spectra_headers(DB_MassSpecAnalyses)}: Fetch spectra headers for a given analysis.

\item \code{get_spectra_tic(DB_MassSpecAnalyses)}: Get the total ion current (TIC) spectra for the specified analyses.

\item \code{plot_spectra_tic(DB_MassSpecAnalyses)}: Plot total ion current (TIC) spectra for the specified analyses.

\item \code{get_spectra_bpc(DB_MassSpecAnalyses)}: Get the base peak chromatograms (BPC) spectra for the specified analyses.

\item \code{plot_spectra_bpc(DB_MassSpecAnalyses)}: Plot base peak chromatogram (BPC) spectra for the specified analyses.

\item \code{get_raw_spectra(DB_MassSpecAnalyses)}: Get raw spectra data from specified analyses, returning a \code{data.table} with the spectra data.

\item \code{get_spectra_eic(DB_MassSpecAnalyses)}: Get extracted ion chromatograms (EIC) for the specified analyses and targets.

\item \code{get_spectra_ms1(DB_MassSpecAnalyses)}: Get MS1 spectra for the specified analyses and targets.

\item \code{get_spectra_ms2(DB_MassSpecAnalyses)}: Get MS2 spectra for the specified analyses and targets.

\item \code{get_chromatograms_headers(DB_MassSpecAnalyses)}: Fetch chromatograms headers for a given analysis.

\item \code{get_chromatograms(DB_MassSpecAnalyses)}: Get chromatograms for the specified analyses and chromatogram IDs/indices.

\item \code{plot_chromatograms(DB_MassSpecAnalyses)}: Plot chromatograms for the specified analyses and chromatogram IDs/indices.

\item \code{query_db(DB_MassSpecAnalyses)}: Internal: execute a query on the DB.

\item \code{list_db_tables(DB_MassSpecAnalyses)}: Internal: list tables in the DB.

\item \code{get_db_table_info(DB_MassSpecAnalyses)}: Internal: get table info from the DB.

}}
