% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/class_DB_MassSpecMethod_AssignTransformationProducts.R
\name{DB_MassSpecMethod_AssignTransformationProducts_native}
\alias{DB_MassSpecMethod_AssignTransformationProducts_native}
\title{DB_MassSpecMethod_AssignTransformationProducts_native class}
\usage{
DB_MassSpecMethod_AssignTransformationProducts_native(
  transformation_products = NULL
)
}
\arguments{
\item{transformation_products}{A data.frame with transformation products and parent entries. Uses \code{SMILES} to
link parents and products via \code{precursor_SMILES}.}
}
\description{
Native StreamFind method to link transformation products and parents using an existing Suspects
table and assign feature groups based on matched suspect features. Writes the TransformationProducts table.
}
