% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_DB_MassSpecMethod_CreateComponents.R
\name{DB_MassSpecMethod_CreateComponents_native}
\alias{DB_MassSpecMethod_CreateComponents_native}
\title{DB_MassSpecMethod_CreateComponents_native class}
\usage{
DB_MassSpecMethod_CreateComponents_native(rtWindow = c(0, 0))
}
\arguments{
\item{rtWindow}{Numeric length-2 vector of retention time offsets (seconds) applied left/right to the FWHM window when checking overlaps. Defaults to no offset.}
}
\description{
Native StreamFind method to assign component identifiers by clustering features that overlap in retention time using their FWHM windows.
}
