% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/class_DB_MassSpecMethod_FeatureBlankSubtraction.R
\name{DB_MassSpecMethod_FeatureBlankSubtraction_native}
\alias{DB_MassSpecMethod_FeatureBlankSubtraction_native}
\title{DB_MassSpecMethod_FeatureBlankSubtraction_native class}
\usage{
DB_MassSpecMethod_FeatureBlankSubtraction_native(blankThreshold = 5)
}
\arguments{
\item{blankThreshold}{numeric(1) threshold multiplier for blank feature intensities to consider a feature as present in blanks. To be present a features must be at least blankThreshold times more intense in samples than in blanks.}
}
\description{
Native StreamFind method for blank subtraction of features.
}
