% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_DB_MassSpecMethod_FindFeatures.R
\name{DB_MassSpecMethod_FindFeatures_native}
\alias{DB_MassSpecMethod_FindFeatures_native}
\title{DB_MassSpecMethod_FindFeatures_native class}
\usage{
DB_MassSpecMethod_FindFeatures_native(
  rtWindows = data.frame(rtmin = numeric(), rtmax = numeric()),
  ppmThreshold = 15,
  noiseThreshold = 250,
  minSNR = 3,
  minTraces = 3,
  baselineWindow = 200,
  maxWidth = 100,
  baseQuantile = 0.1,
  debugMZ = 0,
  debugSpecIdx = -1
)
}
\arguments{
\item{rtWindows}{data.frame with rtmin and rtmax columns for retention time windows for data inclusion.}

\item{ppmThreshold}{integer(1) maximum allowed mass error in ppm for considering traces in a mass cluster.}

\item{noiseThreshold}{numeric(1) lowest threshold to clean data (i.e., no trace with intensity below this level is kept).}

\item{minSNR}{numeric(1) minimum signal-to-noise ratio for considering a trace and chromatographic peak.}

\item{minTraces}{numeric(1) minimum number of traces to consider a mass cluster and chromatographic peak.}

\item{baselineWindow}{numeric(1) retention time window to build a baseline in a mass cluster.}

\item{maxWidth}{numeric(1) expected maximum window for a chromatographic peak.}

\item{baseQuantile}{numeric(1) quantile to estimate the baseline in a mass cluster.}

\item{debugMZ}{numeric(1) m/z value to enable debugging for specific mass traces (0 to disable).}

\item{debugSpecIdx}{integer(1) spectrum index to enable debugging for denoising (-1 to disable).}
}
\description{
Native StreamFind method for finding features (i.e., chromatographic peaks) in liquid chromatography coupled to high resolution mass spectrometry files.
}
