% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_DB_MassSpecMethod_LoadFeaturesMS1.R
\name{DB_MassSpecMethod_LoadFeaturesMS1_native}
\alias{DB_MassSpecMethod_LoadFeaturesMS1_native}
\title{DB_MassSpecMethod_LoadFeaturesMS1_native class}
\usage{
DB_MassSpecMethod_LoadFeaturesMS1_native(
  rtWindow = c(-2, 2),
  mzWindow = c(-1, 6),
  mzClust = 0.005,
  presence = 0.8,
  minIntensity = 250,
  filtered = FALSE
)
}
\arguments{
\item{rtWindow}{Numeric length-2 vector of retention time offsets (seconds) to expand left/right feature windows. Note that left offset should be negative.}

\item{mzWindow}{Numeric length-2 vector of m/z offsets (Da) to expand left/right feature windows. Note that left offset should be negative.}

\item{mzClust}{Numeric(1) m/z tolerance used when clustering traces.}

\item{presence}{Numeric(1) minimum fraction (0-1) of scans required to keep a cluster.}

\item{minIntensity}{Numeric(1) minimum trace intensity to extract.}

\item{filtered}{Logical(1) whether to include features already marked as filtered.}
}
\description{
Native StreamFind method for loading MS1 traces into features stored in a database.
}
