% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_DB_MassSpecMethod_LoadFeaturesMS2.R
\name{DB_MassSpecMethod_LoadFeaturesMS2_native}
\alias{DB_MassSpecMethod_LoadFeaturesMS2_native}
\title{DB_MassSpecMethod_LoadFeaturesMS2_native class}
\usage{
DB_MassSpecMethod_LoadFeaturesMS2_native(
  isolationWindow = 1.3,
  mzClust = 0.005,
  presence = 0.8,
  minIntensity = 10,
  filtered = FALSE
)
}
\arguments{
\item{isolationWindow}{Numeric(1) isolation window (Da) around precursor m/z for MS2 extraction.}

\item{mzClust}{Numeric(1) m/z tolerance used when clustering traces.}

\item{presence}{Numeric(1) minimum fraction (0-1) of scans required to keep a cluster.}

\item{minIntensity}{Numeric(1) minimum trace intensity to extract.}

\item{filtered}{Logical(1) whether to include features already marked as filtered.}
}
\description{
Native StreamFind method for loading MS2 spectra into features stored in a database.
}
