% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_DB_Results.R
\name{DB_Results}
\alias{DB_Results}
\alias{validate_object.DB_Results}
\alias{query_db.DB_Results}
\alias{list_db_tables.DB_Results}
\alias{get_db_table_info.DB_Results}
\title{Base class for database-backed Results}
\usage{
DB_Results(db, dataType)

\method{validate_object}{DB_Results}(x)

\method{query_db}{DB_Results}(x, sql, params = NULL)

\method{list_db_tables}{DB_Results}(x)

\method{get_db_table_info}{DB_Results}(x, tableName)
}
\arguments{
\item{db}{Path to DuckDB file.}

\item{dataType}{Data type string (e.g., "DB_MassSpec").}

\item{x}{ResultsDB object.}

\item{sql}{SQL query string to execute against the database.}

\item{params}{Query parameters to be used with the SQL query (optional).}

\item{tableName}{Name of the table to query or retrieve information about.}
}
\value{
An object of class \code{DB_Results}.
}
\description{
Minimal base class to represent results stored in a DuckDB file.
Requires a \code{db} path and a \code{dataType} identifier.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate_object(DB_Results)}: Validate the DB_Results object.

\item \code{query_db(DB_Results)}: Internal: execute a query on the DB.

\item \code{list_db_tables(DB_Results)}: Internal: list tables in the DB.

\item \code{get_db_table_info(DB_Results)}: Internal: get table info from the DB.

}}
