% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Config.R
\name{EngineConfig}
\alias{EngineConfig}
\title{Engine Configuration}
\usage{
EngineConfig(parameters = list(ConfigCache = ConfigCache()))
}
\description{
Class representing the engine configuration, inheriting from \link{Config}.
}
