% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_EngineDataBase.R
\name{EngineDataBase}
\alias{EngineDataBase}
\title{File-based Database Engine for StreamFind}
\description{
The \link{EngineDataBase} R6 class provides file-based storage
for StreamFind Engine data using DuckDB. Data is accessed by reading from the database
instead of keeping it in memory.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{Metadata}}{A \link{Metadata} object loaded from database}

\item{\code{Configuration}}{Engine configuration loaded from database}

\item{\code{Workflow}}{A \link{Workflow} object loaded from database}

\item{\code{Analyses}}{Simple dummy analyses object}

\item{\code{Results}}{Simple dummy results object}

\item{\code{AuditTrail}}{Audit trail from database (read-only)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EngineDataBase-new}{\code{EngineDataBase$new()}}
\item \href{#method-EngineDataBase-get_metadata}{\code{EngineDataBase$get_metadata()}}
\item \href{#method-EngineDataBase-set_metadata}{\code{EngineDataBase$set_metadata()}}
\item \href{#method-EngineDataBase-get_configuration}{\code{EngineDataBase$get_configuration()}}
\item \href{#method-EngineDataBase-set_configuration}{\code{EngineDataBase$set_configuration()}}
\item \href{#method-EngineDataBase-get_workflow}{\code{EngineDataBase$get_workflow()}}
\item \href{#method-EngineDataBase-set_workflow}{\code{EngineDataBase$set_workflow()}}
\item \href{#method-EngineDataBase-get_analyses}{\code{EngineDataBase$get_analyses()}}
\item \href{#method-EngineDataBase-set_analyses}{\code{EngineDataBase$set_analyses()}}
\item \href{#method-EngineDataBase-get_results}{\code{EngineDataBase$get_results()}}
\item \href{#method-EngineDataBase-set_results}{\code{EngineDataBase$set_results()}}
\item \href{#method-EngineDataBase-get_audit_trail}{\code{EngineDataBase$get_audit_trail()}}
\item \href{#method-EngineDataBase-add_audit_entry}{\code{EngineDataBase$add_audit_entry()}}
\item \href{#method-EngineDataBase-query_database}{\code{EngineDataBase$query_database()}}
\item \href{#method-EngineDataBase-get_engine_info}{\code{EngineDataBase$get_engine_info()}}
\item \href{#method-EngineDataBase-list_tables}{\code{EngineDataBase$list_tables()}}
\item \href{#method-EngineDataBase-get_table_info}{\code{EngineDataBase$get_table_info()}}
\item \href{#method-EngineDataBase-clone}{\code{EngineDataBase$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EngineDataBase-new"></a>}}
\if{latex}{\out{\hypertarget{method-EngineDataBase-new}{}}}
\subsection{Method \code{new()}}{
Cleanup when object is destroyed


Initialize EngineDataBase
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineDataBase$new(db, id = NULL, data_type = "Generic")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{db}}{Path to DuckDB database file}

\item{\code{id}}{Unique engine identifier}

\item{\code{data_type}}{Type of engine data (e.g., "MassSpec", "Raman")}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EngineDataBase-get_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-EngineDataBase-get_metadata}{}}}
\subsection{Method \code{get_metadata()}}{
Get metadata from database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineDataBase$get_metadata()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A Metadata object or NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EngineDataBase-set_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-EngineDataBase-set_metadata}{}}}
\subsection{Method \code{set_metadata()}}{
Set metadata in database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineDataBase$set_metadata(metadata)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metadata}}{A Metadata object or named list}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EngineDataBase-get_configuration"></a>}}
\if{latex}{\out{\hypertarget{method-EngineDataBase-get_configuration}{}}}
\subsection{Method \code{get_configuration()}}{
Get configuration from database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineDataBase$get_configuration()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A configuration list or NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EngineDataBase-set_configuration"></a>}}
\if{latex}{\out{\hypertarget{method-EngineDataBase-set_configuration}{}}}
\subsection{Method \code{set_configuration()}}{
Set configuration in database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineDataBase$set_configuration(configuration)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{configuration}}{A configuration list}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EngineDataBase-get_workflow"></a>}}
\if{latex}{\out{\hypertarget{method-EngineDataBase-get_workflow}{}}}
\subsection{Method \code{get_workflow()}}{
Get workflow from database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineDataBase$get_workflow()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A Workflow object or NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EngineDataBase-set_workflow"></a>}}
\if{latex}{\out{\hypertarget{method-EngineDataBase-set_workflow}{}}}
\subsection{Method \code{set_workflow()}}{
Set workflow in database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineDataBase$set_workflow(workflow)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{workflow}}{A Workflow object or list with data_type and methods}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EngineDataBase-get_analyses"></a>}}
\if{latex}{\out{\hypertarget{method-EngineDataBase-get_analyses}{}}}
\subsection{Method \code{get_analyses()}}{
Get dummy analyses object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineDataBase$get_analyses()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A simple analyses list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EngineDataBase-set_analyses"></a>}}
\if{latex}{\out{\hypertarget{method-EngineDataBase-set_analyses}{}}}
\subsection{Method \code{set_analyses()}}{
Set dummy analyses object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineDataBase$set_analyses(analyses)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{Analyses object (ignored in dummy implementation)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EngineDataBase-get_results"></a>}}
\if{latex}{\out{\hypertarget{method-EngineDataBase-get_results}{}}}
\subsection{Method \code{get_results()}}{
Get dummy results object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineDataBase$get_results()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A simple results list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EngineDataBase-set_results"></a>}}
\if{latex}{\out{\hypertarget{method-EngineDataBase-set_results}{}}}
\subsection{Method \code{set_results()}}{
Set dummy results object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineDataBase$set_results(results)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{results}}{Results object (ignored in dummy implementation)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EngineDataBase-get_audit_trail"></a>}}
\if{latex}{\out{\hypertarget{method-EngineDataBase-get_audit_trail}{}}}
\subsection{Method \code{get_audit_trail()}}{
Get audit trail from database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineDataBase$get_audit_trail()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data.frame with audit trail entries
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EngineDataBase-add_audit_entry"></a>}}
\if{latex}{\out{\hypertarget{method-EngineDataBase-add_audit_entry}{}}}
\subsection{Method \code{add_audit_entry()}}{
Add entry to audit trail
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineDataBase$add_audit_entry(
  operation_type,
  object_type,
  object_id = NULL,
  details = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{operation_type}}{Type of operation}

\item{\code{object_type}}{Type of object}

\item{\code{object_id}}{Object identifier}

\item{\code{details}}{Additional details}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EngineDataBase-query_database"></a>}}
\if{latex}{\out{\hypertarget{method-EngineDataBase-query_database}{}}}
\subsection{Method \code{query_database()}}{
Execute SQL query on the database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineDataBase$query_database(sql, params = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sql}}{SQL query string}

\item{\code{params}}{Query parameters}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Query results
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EngineDataBase-get_engine_info"></a>}}
\if{latex}{\out{\hypertarget{method-EngineDataBase-get_engine_info}{}}}
\subsection{Method \code{get_engine_info()}}{
Get basic engine information
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineDataBase$get_engine_info()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List with engine information
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EngineDataBase-list_tables"></a>}}
\if{latex}{\out{\hypertarget{method-EngineDataBase-list_tables}{}}}
\subsection{Method \code{list_tables()}}{
List all tables in the database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineDataBase$list_tables()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Character vector of table names
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EngineDataBase-get_table_info"></a>}}
\if{latex}{\out{\hypertarget{method-EngineDataBase-get_table_info}{}}}
\subsection{Method \code{get_table_info()}}{
Get information about a specific table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineDataBase$get_table_info(table_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{table_name}}{Name of the table}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data frame with table information
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EngineDataBase-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EngineDataBase-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EngineDataBase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
