% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_007_filterFeatures.R
\name{excludeAdductsFeatures}
\alias{excludeAdductsFeatures}
\title{excludeAdductsFeatures}
\usage{
excludeAdductsFeatures(obj, value = TRUE)
}
\arguments{
\item{obj}{An \linkS4class{msFeatures} object.}

\item{value}{Logical, when \code{TRUE}, adducts with protonated/deprotonated
ions present are filtered.}
}
\description{
excludeAdductsFeatures
}
