% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_filterFeatures.R
\name{excludeIsotopesFeatures}
\alias{excludeIsotopesFeatures}
\title{excludeIsotopesFeatures}
\usage{
excludeIsotopesFeatures(obj, value = TRUE)
}
\arguments{
\item{obj}{An \linkS4class{msFeatures} object.}

\item{value}{Logical, when \code{TRUE} isotopes are filtered.}
}
\description{
excludeIsotopesFeatures
}
