% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ms_data_access.R
\name{extractEICs}
\alias{extractEICs}
\title{extractEICs}
\usage{
extractEICs(
  object = NULL,
  analyses = NULL,
  mz = NULL,
  ppm = 20,
  rt = NULL,
  sec = 60,
  id = NULL
)
}
\arguments{
\item{object}{An \linkS4class{msData} object.}

\item{analyses}{A numeric or character vector with the index or analysis name/s.
When \code{NULL}, all the analyses are used.}

\item{mz}{A vector with target \emph{m/z} values or
a two columns data.table or data.frame with minimum and maximum \emph{m/z} values.
Alternatively, \emph{m/z} and retention time values can be given as one data.table/data.frame
and the deviations given as \code{ppm} and \code{sec} are used to calculate the ranges.
The same also works for min and max values of \emph{m/z} and retention time targets.
Note that when mass/time ranges are given, \code{ppm} and \code{sec} are not used.}

\item{ppm}{A numeric vector of length one with the mass deviation, in ppm.}

\item{rt}{A vector with target retention time values or
a two columns data.table/data.frame with minimum and maximum retention time values.}

\item{sec}{A numeric vector of length one with the time deviation, in seconds.}

\item{id}{An id vector with target identifiers. When not given
is added as a combination of the \emph{m/z} and retention time ranges or values.}
}
\value{
A \code{data.table} with the columns
\code{analysis}, \code{replicate}, \code{id}, \code{rt}, and \code{intensity}
representing the analysis name, the analysis replicate name,
the target id, the retention time and the sum of the intensities
for the collected \emph{m/z} in each spectrum (i.e., MS scan), respectively.
}
\description{
Extracts MS1 level ion chromatograms (EICs) from raw data
of specified \emph{m/z} and retention time pair/s. The function uses the
\link[patRoon]{getEICs} function from \pkg{patRoon}.
}
\references{
\insertRef{patroon01}{streamFind}

\insertRef{proteo01}{streamFind}
}
