% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ms_data_access.R
\name{extractMSn}
\alias{extractMSn}
\title{extractMSn}
\usage{
extractMSn(
  object = NULL,
  analyses = NULL,
  level = 2,
  mz = NULL,
  ppm = 20,
  rt = NULL,
  sec = 60,
  id = NULL,
  settings = NULL
)
}
\arguments{
\item{object}{An \linkS4class{msData} or \linkS4class{msAnalysis} object.}

\item{analyses}{A numeric or character vector with the index or
analysis name/s. When \code{NULL}, all the analyses are used.}

\item{level}{A numeric vector with length 1 to defined the MS level.
Currently, only level 2, corresponding to MS/MS, is possible.}

\item{mz}{A vector with target \emph{m/z} values or a two columns
\linkS4class{data.table} or data.frame with minimum and maximum
\emph{m/z} values. Alternatively, \emph{m/z} and retention time values
can be given as one \linkS4class{data.table}/data.frame and the deviations
given as \code{ppm} and \code{sec} are used to calculate the ranges.
The same also works for min and max values of \emph{m/z} and retention
time targets. Note that when mass/time ranges are given, \code{ppm} and
\code{sec} are not used.}

\item{ppm}{A numeric vector of length one with the mass deviation, in ppm.}

\item{rt}{A vector with target retention time values or
a two columns \linkS4class{data.table}/data.frame with minimum
and maximum retention time values.}

\item{sec}{A numeric vector of length one with the time deviation, in seconds.}

\item{id}{An id vector with target identifiers. When not given
is added as a combination of the \emph{m/z} and retention time ranges or values.}

\item{settings}{A \linkS4class{settings} object with the parameter settings
for extracting and averaging MSn spectra.}
}
\value{
A \code{data.table} with the columns
\code{analysis}, \code{replicate}, \code{id}, \code{ce}, \code{preMZ}, \code{mz}, \code{intensity} and \code{precursor}
representing the analysis name (i.e., file), the analysis replicate name,
the isolation target id, the collision energy applied, the \emph{m/z} of the precursor and the MSn trace, the intensity and the presence of the precursor
(i.e., \emph{m/z} matching the isolation target), respectively.
}
\description{
Extracts MSn spectra from defined isolated targets
defined by \emph{m/z} and retention time, including the respective deviations.
}
