% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_000_generics.R
\name{features}
\alias{features}
\title{features}
\usage{
features(object, ...)
}
\arguments{
\item{object}{A method specific class object.}

\item{...}{Other method specific arguments.}
}
\description{
Generic for getting features.
}
