% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_003_peakFilling.R
\name{fillingSettingsDefaultXCMS}
\alias{fillingSettingsDefaultXCMS}
\title{fillingSettingsDefaultXCMS}
\usage{
fillingSettingsDefaultXCMS()
}
\value{
A \linkS4class{settings} object containing parameters for recursive
integration of peaks from analyses not represented in a given feature.
}
\description{
fillingSettingsDefaultXCMS
}
