% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_S3.R
\name{get_analysis_classes}
\alias{get_analysis_classes}
\title{S3 generic \code{get_analysis_classes}}
\usage{
get_analysis_classes(x, ...)
}
\arguments{
\item{x}{An object from which the analysis classes will be retrieved.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S3 generic to get analysis classes from \code{x}.
}
