% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_S3.R
\name{get_analysis_names}
\alias{get_analysis_names}
\title{S3 generic \code{get_analysis_names}}
\usage{
get_analysis_names(x, ...)
}
\arguments{
\item{x}{An object from which the names will be retrieved.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S3 generic to get names from \code{x}.
}
