% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_S3.R
\name{get_blanks}
\alias{get_blanks}
\title{S3 generic \code{get_blanks}}
\usage{
get_blanks(x, ...)
}
\arguments{
\item{x}{An object from which the blanks will be retrieved.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S3 generic to get blanks from \code{x}.
}
