% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_cache_info.R
\name{get_cache_info.character}
\alias{get_cache_info.character}
\title{get_cache_info}
\usage{
\method{get_cache_info}{character}(x = "cache")
}
\arguments{
\item{x}{A character string specifying the cache file or path to use. Default is a relative
"cache" directory of the working directory.}
}
\value{
A data.table with the categories saved in the cache.
}
\description{
Gets information about the saved cache categories.
}
\references{
\insertRef{patroon01}{StreamFind}

\insertRef{patroon02}{StreamFind}
}
