% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_S3.R
\name{get_cache_info}
\alias{get_cache_info}
\title{S3 generic \code{get_cache_info}}
\usage{
get_cache_info(x, ...)
}
\arguments{
\item{x}{An object from which the cache information will be retrieved.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S3 generic to get information about the cache of \code{x}.
}
