% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_S3.R
\name{get_chromatograms}
\alias{get_chromatograms}
\title{S3 generic \code{get_chromatograms}}
\usage{
get_chromatograms(x, ...)
}
\arguments{
\item{x}{An object from which the chromatograms will be retrieved.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S3 generic to get chromatograms from \code{x}.
}
