% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ms.R
\name{get_colors}
\alias{get_colors}
\title{get_colors}
\usage{
get_colors(obj)
}
\arguments{
\item{obj}{A character vector to associate with the colors.}
}
\value{
A vector of colors. The vector is named according the \code{obj}.
}
\description{
Function to produce colors for a character vector.
}
