% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_compounds}
\alias{get_compounds}
\title{Generic \code{get_compounds}}
\usage{
get_compounds(x, ...)
}
\arguments{
\item{x}{An object from which the compounds will be retrieved.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S7 generic to get compounds from \code{x}.
}
