% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_S3.R
\name{get_concentrations}
\alias{get_concentrations}
\title{S3 generic \code{get_concentrations}}
\usage{
get_concentrations(x, ...)
}
\arguments{
\item{x}{An object from which the concentrations will be retrieved.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S3 generic to get concentrations from \code{x}.
}
