% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_install_external_tools.R
\name{get_conda_dir}
\alias{get_conda_dir}
\title{Locate the Miniforge / Conda installation}
\usage{
get_conda_dir()
}
\value{
Path to conda root, or \code{NA} if not found.
}
\description{
Checks \code{~/.streamfind/external/conda/}.
}
