% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_005_default_ProcessingSettings.R
\name{get_default_ProcessingSettings}
\alias{get_default_ProcessingSettings}
\title{get_default_ProcessingSettings}
\usage{
get_default_ProcessingSettings(call = NA_character_, algorithm = NA_character_)
}
\arguments{
\item{call}{X.}

\item{algorithm}{X.}
}
\value{
A ProcessingSettings S3 class object.
}
\description{
X.
}
