% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_features_count}
\alias{get_features_count}
\title{Generic \code{get_features_count}}
\usage{
get_features_count(x, ...)
}
\arguments{
\item{x}{An object from which the features count will be retrieved.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S7 generic to get features count from \code{x}.
}
