% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_features_eic}
\alias{get_features_eic}
\title{Generic \code{get_features_eic}}
\usage{
get_features_eic(x, ...)
}
\arguments{
\item{x}{An object from which the features EIC will be retrieved.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S7 generic to get features EIC from \code{x}.
}
