% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_features_ms1}
\alias{get_features_ms1}
\title{Generic \code{get_features_ms1}}
\usage{
get_features_ms1(x, ...)
}
\arguments{
\item{x}{An object from which the features MS1 will be retrieved.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S7 generic to get features MS1 from \code{x}.
}
