% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_fold_change}
\alias{get_fold_change}
\title{Generic \code{get_fold_change}}
\usage{
get_fold_change(x, ...)
}
\arguments{
\item{x}{An object from which the fold change will be retrieved.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S7 generic to get fold change from \code{x}.
}
