% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_groups}
\alias{get_groups}
\title{Generic \code{get_groups}}
\usage{
get_groups(x, ...)
}
\arguments{
\item{x}{An object from which the groups will be retrieved.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S7 generic to get groups from \code{x}.
}
