% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_groups_ms1}
\alias{get_groups_ms1}
\title{Generic \code{get_groups_ms1}}
\usage{
get_groups_ms1(x, ...)
}
\arguments{
\item{x}{An object from which the groups MS1 will be retrieved.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S7 generic to get groups MS1 from \code{x}.
}
