% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_internal_standards}
\alias{get_internal_standards}
\title{Generic \code{get_internal_standards}}
\usage{
get_internal_standards(x, ...)
}
\arguments{
\item{x}{An object from which the internal standards will be retrieved.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S7 generic to get internal standards from \code{x}.
}
