% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_install_external_tools.R
\name{get_lpsolve_dir}
\alias{get_lpsolve_dir}
\title{Locate the LPSolve installation directory}
\usage{
get_lpsolve_dir()
}
\value{
Path to LPSolve directory, or \code{NA} if not found.
}
\description{
Checks \code{~/.streamfind/external/lpsolve/5.5.2.11/}.
}
