% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_matrix_suppression}
\alias{get_matrix_suppression}
\title{Generic \code{get_matrix_suppression}}
\usage{
get_matrix_suppression(x, ...)
}
\arguments{
\item{x}{An object from which the matrix suppression will be retrieved.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S7 generic to get matrix suppression from \code{x}.
}
