% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_S3.R
\name{get_methods}
\alias{get_methods}
\title{S3 generic \code{get_methods}}
\usage{
get_methods(x, ...)
}
\arguments{
\item{x}{An object from which the methods will be retrieved.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S3 generic to get methods from \code{x}.
}
