% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_S3.R
\name{get_model_prediction}
\alias{get_model_prediction}
\title{S3 generic \code{get_model_prediction}}
\usage{
get_model_prediction(x, ...)
}
\arguments{
\item{x}{An object from which the model prediction will be retrieved.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S3 generic to get model prediction from \code{x}.
}
