% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_install_external_tools.R
\name{get_msvc_dir}
\alias{get_msvc_dir}
\title{Locate MSVC Build Tools installation}
\usage{
get_msvc_dir()
}
\value{
Path to VS BuildTools root, or \code{NA} if not found.
}
\description{
Searches standard Visual Studio installation paths for \code{cl.exe}.
}
