% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_S3.R
\name{get_patRoon_compounds}
\alias{get_patRoon_compounds}
\title{S3 generic \code{get_patRoon_compounds}}
\usage{
get_patRoon_compounds(x, ...)
}
\arguments{
\item{x}{An object from which the patRoon compounds will be retrieved.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S3 generic to get \href{https://github.com/rickhelmus/patRoon}{patRoon} compounds class from \code{x}.
}
