% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_S3.R
\name{get_patRoon_features}
\alias{get_patRoon_features}
\title{Generic \code{get_patRoon_features}}
\usage{
get_patRoon_features(x, ...)
}
\arguments{
\item{x}{An object from which the patRoon features will be retrieved.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S3 generic to get \href{https://github.com/rickhelmus/patRoon}{patRoon} features class
from \code{x}.
}
