% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_patRoon_MSPeakLists}
\alias{get_patRoon_MSPeakLists}
\title{Generic \code{get_patRoon_MSPeakLists}}
\usage{
get_patRoon_MSPeakLists(x, ...)
}
\arguments{
\item{x}{An object from which the patRoon MSPeakLists will be retrieved.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S7 generic to get \href{https://github.com/rickhelmus/patRoon}{patRoon} MSPeakLists class from \code{x}.
}
