% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_000_generics.R
\name{get_peaks}
\alias{get_peaks}
\title{get_peaks}
\usage{
get_peaks(object, ...)
}
\arguments{
\item{object}{A method specific class object.}

\item{...}{Other method specific arguments.}
}
\description{
Generic for getting peaks.
}
