% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_S3.R
\name{get_raw_chromatograms}
\alias{get_raw_chromatograms}
\title{Generic \code{get_raw_chromatograms}}
\usage{
get_raw_chromatograms(x, ...)
}
\arguments{
\item{x}{An object from which the raw chromatograms will be retrieved.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S3 generic to get raw chromatograms from \code{x}.
}
