% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_S3.R
\name{get_replicate_names}
\alias{get_replicate_names}
\title{Generic \code{get_replicate_names}}
\usage{
get_replicate_names(x, ...)
}
\arguments{
\item{x}{An object from which the replicates will be retrieved.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S3 generic to get replicates from \code{x}.
}
