% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_S3.R
\name{get_spectra_bpc}
\alias{get_spectra_bpc}
\title{Generic \code{get_spectra_bpc}}
\usage{
get_spectra_bpc(x, ...)
}
\arguments{
\item{x}{An object from which the spectra BPC will be retrieved.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S3 generic to get spectra BPC from \code{x}.
}
