% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_spectra_eic}
\alias{get_spectra_eic}
\title{Generic \code{get_spectra_eic}}
\usage{
get_spectra_eic(x, ...)
}
\arguments{
\item{x}{An object from which the spectra EIC will be retrieved.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S7 generic to get spectra EIC from \code{x}.
}
