% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_S3.R
\name{get_spectra_headers}
\alias{get_spectra_headers}
\title{Generic \code{get_spectra_headers}}
\usage{
get_spectra_headers(x, ...)
}
\arguments{
\item{x}{An object from which the spectra headers will be retrieved.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S3 generic to get spectra headers from \code{x}.
}
