% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_spectra_matrix}
\alias{get_spectra_matrix}
\title{Generic \code{get_spectra_matrix}}
\usage{
get_spectra_matrix(x, ...)
}
\arguments{
\item{x}{An object from which the spectra matrix will be retrieved.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S7 generic to get spectra matrix from \code{x}.
}
