% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_spectra_ms1}
\alias{get_spectra_ms1}
\title{Generic \code{get_spectra_ms1}}
\usage{
get_spectra_ms1(x, ...)
}
\arguments{
\item{x}{An object from which the spectra MS1 will be retrieved.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S7 generic to get spectra MS1 from \code{x}.
}
