% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_spectra_peaks}
\alias{get_spectra_peaks}
\title{Generic \code{get_spectra_peaks}}
\usage{
get_spectra_peaks(x, ...)
}
\arguments{
\item{x}{An object from which the spectra peaks will be retrieved.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S7 generic to get spectra peaks from \code{x}.
}
