% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_S3.R
\name{get_spectra_tic}
\alias{get_spectra_tic}
\title{Generic \code{get_spectra_tic}}
\usage{
get_spectra_tic(x, ...)
}
\arguments{
\item{x}{An object from which the spectra TIC will be retrieved.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S3 generic to get spectra TIC from \code{x}.
}
