% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_install_external_tools.R
\name{get_streamfind_dir}
\alias{get_streamfind_dir}
\title{Path to the StreamFind home directory}
\usage{
get_streamfind_dir()
}
\value{
The normalized path to \code{~/.streamfind/}.
}
\description{
Returns (and creates if needed) \code{~/.streamfind/}.
}
