% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_suspects}
\alias{get_suspects}
\title{Generic \code{get_suspects}}
\usage{
get_suspects(x, ...)
}
\arguments{
\item{x}{An object from which the suspects will be retrieved.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S7 generic to get suspects from \code{x}.
}
