% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_install_external_tools.R
\name{get_vcpkg_dir}
\alias{get_vcpkg_dir}
\title{Locate the vcpkg installation directory}
\usage{
get_vcpkg_dir()
}
\value{
Path to vcpkg root, or \code{NA} if not found.
}
\description{
Checks \code{~/.streamfind/external/vcpkg/}.
}
