% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_000_generics.R
\name{getSpectra}
\alias{getSpectra}
\title{getSpectra}
\usage{
getSpectra(object, ...)
}
\arguments{
\item{object}{A method specific class object.}

\item{...}{Other method specific arguments.}
}
\description{
Generic for getting spectra.
}
