% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_000_generics.R
\name{hasAdjustedRetentionTime}
\alias{hasAdjustedRetentionTime}
\title{hasAdjustedRetentionTime}
\usage{
hasAdjustedRetentionTime(object)
}
\arguments{
\item{object}{A method specific class object.}
}
\description{
Generic for checking is has adjusted retention time.
}
