% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_000_generics.R
\name{hasFeatures}
\alias{hasFeatures}
\title{hasFeatures}
\usage{
hasFeatures(object)
}
\arguments{
\item{object}{A method specific class object.}
}
\description{
Generic for check the presence of features.
}
