% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_S3_headers.R
\name{headers}
\alias{headers}
\title{Headers S3 Class Constructor}
\usage{
headers(name = NA_character_, path = NA_character_, date = NULL, ...)
}
\arguments{
\item{call}{Character of length one with the name of the method where the
settings are to be applied.}

\item{algorithm}{Character of length one with the name of the algorithm to
be used.}

\item{parameters}{A list with parameter settings specific for the \code{call}
method and used \code{algorithm}.}
}
\value{
A settings S3 class
}
\description{
Creates a headers S3 class object.
}
\details{
See the method documentation for more information about possible
algorithms and parameters to be used.
}
