% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_R6_MassSpecData.R
\name{import_MassSpecData}
\alias{import_MassSpecData}
\title{Function to import a MassSpecData class object from a \emph{json} or \emph{rds} file}
\usage{
import_MassSpecData(file)
}
\arguments{
\item{file}{A \emph{json} or \emph{rds} file as obtained by the MassSpecData
method \code{save()}.}
}
\value{
A \code{MassSpecData} class object.
}
\description{
Function to import a \code{MassSpecData} class object from a saved
\emph{json} or \emph{rds} file.
}
