% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_R6_MassSpecEngine.R
\name{import_MassSpecEngine}
\alias{import_MassSpecEngine}
\title{Function to import a MassSpecEngine class object from a \emph{json} or \emph{rds} file.}
\usage{
import_MassSpecEngine(file)
}
\arguments{
\item{file}{A \emph{json} or \emph{rds} file to be imported.}
}
\value{
A \code{MassSpecEngine} class object.
}
\description{
Function to import a \code{MassSpecEngine} class object from a saved
\emph{json} or \emph{rds} file.
}
