% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_R6_msData.R
\name{import_msData}
\alias{import_msData}
\title{import_msData}
\usage{
import_msData(file)
}
\arguments{
\item{file}{X.}
}
\value{
An \code{msData} class object.
}
\description{
Function to import an \code{msData} object from a saved file.
}
