% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_class_000_general.R
\name{importSettings}
\alias{importSettings}
\title{importSettings}
\usage{
importSettings(file)
}
\arguments{
\item{file}{A character string with the full file path.}
}
\description{
Imports settings as either JSON or rds.
}
