% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_S3.R
\name{info}
\alias{info}
\title{Generic \code{info}}
\usage{
info(x, ...)
}
\arguments{
\item{x}{An object from which the information will be retrieved.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S3 generic to get information from \code{x}.
}
