% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{load_spectra}
\alias{load_spectra}
\title{Generic \code{load_spectra}}
\usage{
load_spectra(x, ...)
}
\arguments{
\item{x}{An object from which the spectra will be loaded.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S7 generic to load spectra from \code{x}.
}
