% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_002_parse_data_from_files.R
\name{make_ms_analyses}
\alias{make_ms_analyses}
\title{make_ms_analyses}
\usage{
make_ms_analyses(files = NULL, runParallel = FALSE)
}
\arguments{
\item{files}{A character vector with mzML or mzXML file path/s.
Alternatively, a data.frame with the column/s file, replicate and blank
with the full file path/s, the replicate group name/s (string) and the
associated blank replicate group name (string).}

\item{runParallel}{Logical, set to \code{TRUE} for parsing the data from
files in parallel.}
}
\value{
A named list with a list object for each file in \code{files}.
}
\description{
Reads given mzML/mzXML file/s and makes a named list with a list object for
each file. The file name is used to name the list.
}
\details{
Describe the entries for an object list reflecting a file.
}
