% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ms.R
\name{make_ms_targets}
\alias{make_ms_targets}
\title{Function to make targets for parsing data within MassSpecData class methods}
\usage{
make_ms_targets(mz = NULL, rt = NULL, ppm = 20, sec = 60, id = NULL)
}
\arguments{
\item{mz}{A vector with target \emph{m/z} values (in Da) or a two columns
data.table or data.frame with minimum and maximum \emph{m/z} values (in Da).
Alternatively, \emph{m/z} and retention time values (in seconds) can be given
as one data.table or data.frame and the deviations given in the \code{ppm}
and \code{sec} arguments are used to calculate the ranges. Also works with a
data.table or data.frame with minimum and maximum values of \emph{m/z} and
retention time targets. Note that when mass/time ranges are given, the
\code{ppm} and \code{sec} arguments are not used.}

\item{rt}{A vector with target retention time values (in seconds) or
a two columns data.table or data.frame with minimum and maximum retention
time values (in seconds).}

\item{ppm}{Numeric of length one with the mass deviation, in ppm.}

\item{sec}{Numeric of length one with the time deviation, in seconds.}

\item{id}{Character with the same length as \emph{m/z} and retention time
targets to be used as identifiers. When not given, the id is built as a
combination of the \emph{m/z} and retention time ranges or values.}
}
\value{
A data.frame with columns: \emph{id}, \emph{mz}, \emph{rt}, \emph{mzmin}, \emph{mzmax},
\emph{rtmin}, \emph{rtmax}.
}
\description{
Helper function to build \emph{m/z} and retention time
target pairs for searching data. Each target is composed of an
id and \emph{m/z} (Da) and time (seconds) ranges. When mass is defined
without time, the time range return 0 and vice versa.
}
