% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ms_makeTargets.R
\name{makeTargets}
\alias{makeTargets}
\title{makeTargets}
\usage{
makeTargets(mz = NULL, rt = NULL, ppm = 20, sec = 60, id = NULL)
}
\arguments{
\item{mz}{A vector with target \emph{m/z} values or a two columns
\linkS4class{data.table} or data.frame with minimum and maximum
\emph{m/z} values. Alternatively, \emph{m/z} and retention time values
can be given as one \linkS4class{data.table}/data.frame and the deviations
given as \code{ppm} and \code{sec} are used to calculate the ranges.
The same also works for min and max values of \emph{m/z} and retention
time targets. Note that when mass/time ranges are given, \code{ppm} and
\code{sec} are not used.}

\item{rt}{A vector with target retention time values or
a two columns \linkS4class{data.table}/data.frame with minimum
and maximum retention time values.}

\item{ppm}{A numeric vector of length one with the mass deviation, in ppm.}

\item{sec}{A numeric vector of length one with the time deviation, in seconds.}

\item{id}{An id vector with target identifiers. When not given
is added as a combination of the \emph{m/z} and retention time ranges or values.}
}
\value{
A data.table with \emph{m/z} and retention time target pairs identified by an id.
}
\description{
Helper function to build \emph{m/z} and retention time
target pairs for searching data. Each target is composed of an
id and \emph{m/z} (Da) and time (seconds) ranges. When mass is defined without
time, the time range return NA and vice versa.
}
