% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{map_components}
\alias{map_components}
\title{Generic \code{map_components}}
\usage{
map_components(x, ...)
}
\arguments{
\item{x}{An object from which the components will be mapped.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S7 generic to map components from \code{x}.
}
