% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{map_features}
\alias{map_features}
\title{Generic \code{map_features}}
\usage{
map_features(x, ...)
}
\arguments{
\item{x}{An object from which the features will be mapped.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S7 generic to map features from \code{x}.
}
