% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{map_features_intensity}
\alias{map_features_intensity}
\title{Generic \code{map_features_intensity}}
\usage{
map_features_intensity(x, ...)
}
\arguments{
\item{x}{An object from which the features intensity will be mapped.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S7 generic to map features intensity from \code{x}.
}
