% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ms_plot_peaks.R
\name{mapPeaksInteractive}
\alias{mapPeaksInteractive}
\title{mapPeaksInteractive}
\usage{
mapPeaksInteractive(peaks, xlim = 60, ylim = 5, title, colorBy = "targets")
}
\arguments{
\item{peaks}{A data table with the individual peak details to plot.}

\item{xlim}{A length one or two numeric vector for setting the \emph{x} limits (in seconds) of the plot.}

\item{ylim}{A length one or two numeric vector for setting the \emph{m/z} limits of the plot.}

\item{title}{An optional character vector to be used as title.}

\item{colorBy}{Possible values are \code{"targets"} (the default),
\code{"analyses"} or \code{replicates},
for coloring by target peaks, analyses or replicates, respectively.}
}
\value{
A peak/s map plot produced through \pkg{plotly}.
}
\description{
Function for plotting peak spaces.
}
