% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_S3_MassSpecAnalysis.R
\name{MassSpecAnalysis}
\alias{MassSpecAnalysis}
\alias{validate.MassSpecAnalysis}
\alias{print.MassSpecAnalysis}
\alias{asJSON.MassSpecAnalysis}
\alias{as.MassSpecAnalysis}
\alias{parse_MassSpecAnalysis}
\title{\strong{MassSpecAnalysis} S3 class constructor, methods and functions}
\usage{
MassSpecAnalysis(
  name = NA_character_,
  replicate = NA_character_,
  blank = NA_character_,
  file = NA_character_,
  format = NA_character_,
  type = NA_character_,
  spectra_number = NA_integer_,
  spectra_mode = NA_character_,
  spectra_levels = NA_integer_,
  spectra_headers = data.table(),
  spectra = data.table(),
  chromatograms_number = NA_integer_,
  chromatograms_headers = data.table(),
  chromatograms = data.table(),
  metadata = list()
)

\method{validate}{MassSpecAnalysis}(x = NULL)

\method{print}{MassSpecAnalysis}(x, ...)

\method{asJSON}{MassSpecAnalysis}(x)

as.MassSpecAnalysis(value)

parse_MassSpecAnalysis(files = NULL, runParallel = FALSE)
}
\arguments{
\item{name}{\emph{mzML} or \emph{mzXML} file name without extension.}

\item{replicate}{Character with length one, representing the analysis replicate group name.}

\item{blank}{Character with length one, representing the associated blank replicate group name.}

\item{file}{\emph{mzML} or \emph{mzXML} full file path (with extension).}

\item{format}{Character with length one. One of \emph{mzML} or \emph{mzXML}.}

\item{type}{Character with length one defining the type of analysis.}

\item{spectra_number}{Integer with the number of spectra in the file.}

\item{spectra_headers}{data.table run information for each spectrum.}

\item{spectra}{data.table with the raw spectra (only present if loaded).}

\item{chromatograms_number}{Integer with the number of chromatograms in the file.}

\item{chromatograms_headers}{data.table headers information for each chromatogram.}

\item{chromatograms}{data.table with the raw chromatograms (only present if loaded).}

\item{metadata}{List with flexible storage for experimental metadata (e.g., concentration, location, etc.).}

\item{x}{A \emph{MassSpecAnalysis} S3 class object.}

\item{...}{Not used.}

\item{value}{A list to be checked and/or converted to \emph{MassSpecAnalysis} S3 class.}

\item{files}{A character vector with \emph{mzML} or \emph{mzXML} full file path/s.
Alternatively, a data.frame with the column/s file, replicate and blank
with the full file path/s, the replicate group name/s (string) and the
associated blank replicate group name (string).}

\item{runParallel}{Logical of length one. Set to \code{TRUE} for parallel
processing. Note that parallel processing is not always faster due to
overhead.}
}
\value{
An \emph{MassSpecAnalysis} S3 class object.
}
\description{
Creates a \emph{MassSpecAnalysis} S3 class object.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate(MassSpecAnalysis)}: S3 method to validate a \emph{MassSpecAnalysis} S3 class object, returning a logical value of length one.

\item \code{print(MassSpecAnalysis)}: S3 method to print the \emph{MassSpecAnalysis} S3 class object in the console.

\item \code{asJSON(MassSpecAnalysis)}: S3 method to converts a \emph{MassSpecAnalysis} S3 class object into a JSON string.

}}
\section{Functions}{
\itemize{
\item \code{as.MassSpecAnalysis()}: S3 method to convert the argument files in a \emph{MassSpecAnalysis} S3 class object.

\item \code{parse_MassSpecAnalysis()}: Parses information from \emph{mzML} or \emph{mzXML} file/s and returns a list with
\emph{MassSpecAnalysis} S3 class object/s. On error, returns \code{NULL}.

}}
