% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecEngine.R
\name{MassSpecEngine}
\alias{MassSpecEngine}
\title{Engine dedicated to Mass Spectrometry (MS) data}
\description{
R6 child of Engine for MassSpec data that uses MassSpecAnalyses for on-disk storage.
}
\section{Super class}{
\code{\link[StreamFind:Engine]{StreamFind::Engine}} -> \code{MassSpecEngine}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{Analyses}}{A MassSpecAnalyses object backed by DuckDB.}

\item{\code{NonTargetAnalysis}}{A MassSpecResults_NonTargetAnalysis object backed by DuckDB.}

\item{\code{Chromatograms}}{A MassSpecResults_Chromatograms object backed by DuckDB.}

\item{\code{Spectra}}{A MassSpecResults_Spectra object backed by DuckDB.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MassSpecEngine-new}{\code{MassSpecEngine$new()}}
\item \href{#method-MassSpecEngine-clone}{\code{MassSpecEngine$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Engine" data-id="add_audit_entry"><a href='../../StreamFind/html/Engine.html#method-Engine-add_audit_entry'><code>StreamFind::Engine$add_audit_entry()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Engine" data-id="add_metadata"><a href='../../StreamFind/html/Engine.html#method-Engine-add_metadata'><code>StreamFind::Engine$add_metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Engine" data-id="add_workflow"><a href='../../StreamFind/html/Engine.html#method-Engine-add_workflow'><code>StreamFind::Engine$add_workflow()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Engine" data-id="clear_cache"><a href='../../StreamFind/html/Engine.html#method-Engine-clear_cache'><code>StreamFind::Engine$clear_cache()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Engine" data-id="clear_result_databases"><a href='../../StreamFind/html/Engine.html#method-Engine-clear_result_databases'><code>StreamFind::Engine$clear_result_databases()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Engine" data-id="get_audit_trail"><a href='../../StreamFind/html/Engine.html#method-Engine-get_audit_trail'><code>StreamFind::Engine$get_audit_trail()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Engine" data-id="get_cache_info"><a href='../../StreamFind/html/Engine.html#method-Engine-get_cache_info'><code>StreamFind::Engine$get_cache_info()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Engine" data-id="get_cache_size"><a href='../../StreamFind/html/Engine.html#method-Engine-get_cache_size'><code>StreamFind::Engine$get_cache_size()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Engine" data-id="get_db_table_info"><a href='../../StreamFind/html/Engine.html#method-Engine-get_db_table_info'><code>StreamFind::Engine$get_db_table_info()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Engine" data-id="get_engine_info"><a href='../../StreamFind/html/Engine.html#method-Engine-get_engine_info'><code>StreamFind::Engine$get_engine_info()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Engine" data-id="get_metadata"><a href='../../StreamFind/html/Engine.html#method-Engine-get_metadata'><code>StreamFind::Engine$get_metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Engine" data-id="get_project_path"><a href='../../StreamFind/html/Engine.html#method-Engine-get_project_path'><code>StreamFind::Engine$get_project_path()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Engine" data-id="get_workflow"><a href='../../StreamFind/html/Engine.html#method-Engine-get_workflow'><code>StreamFind::Engine$get_workflow()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Engine" data-id="list_db_tables"><a href='../../StreamFind/html/Engine.html#method-Engine-list_db_tables'><code>StreamFind::Engine$list_db_tables()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Engine" data-id="print"><a href='../../StreamFind/html/Engine.html#method-Engine-print'><code>StreamFind::Engine$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Engine" data-id="query_db"><a href='../../StreamFind/html/Engine.html#method-Engine-query_db'><code>StreamFind::Engine$query_db()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Engine" data-id="report_quarto"><a href='../../StreamFind/html/Engine.html#method-Engine-report_quarto'><code>StreamFind::Engine$report_quarto()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Engine" data-id="run"><a href='../../StreamFind/html/Engine.html#method-Engine-run'><code>StreamFind::Engine$run()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Engine" data-id="run_app"><a href='../../StreamFind/html/Engine.html#method-Engine-run_app'><code>StreamFind::Engine$run_app()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Engine" data-id="run_workflow"><a href='../../StreamFind/html/Engine.html#method-Engine-run_workflow'><code>StreamFind::Engine$run_workflow()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MassSpecEngine-new"></a>}}
\if{latex}{\out{\hypertarget{method-MassSpecEngine-new}{}}}
\subsection{Method \code{new()}}{
Initialize MassSpecEngine
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MassSpecEngine$new(
  projectPath = "data",
  metadata = NULL,
  workflow = NULL,
  files = NULL,
  centroid = FALSE,
  levels = c(1, 2)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{projectPath}}{Path to StreamFind project folder (.sf). DB_Engine will place main.duckdb inside it. If reusing an existing projectPath, DB_Engine will connect to the existing Engine.duckdb file and preserve previous data. If the projectPath does not exist, DB_Engine will create the directory and initialize a new Engine.duckdb file.}

\item{\code{metadata}}{A named list or a \code{Metadata} object. Each list element must have length one.
There is no type restrictions except for name (must be type character), author (must be type
character) and date (must be class POSIXct or POSIXt). A file entry must be a character string
with a full path to a file.}

\item{\code{workflow}}{A \code{Workflow} object, a list of \code{ProcessingStep} objects or a character string
with a full path to a \emph{rds} or \emph{json} file containing a \code{Workflow}.}

\item{\code{files}}{A character vector with full file path/s of \emph{mzML}
or \emph{mzXML} file/s or a data.frame with the columns \emph{file}, \emph{replicate}
and \emph{blank} with the full file path of \emph{mzML} or \emph{mzXML} file/s
(character), the replicate group name (character) and the associated blank
replicate group name (character). Note the file paths must have the extension
\emph{.mzML} or \emph{.mzXML}.}

\item{\code{centroid}}{Logical (length 1). Set to \code{TRUE} to centroid data when converting from vendor formats to mzML.}

\item{\code{levels}}{Integer vector with the levels of MS data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MassSpecEngine-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MassSpecEngine-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MassSpecEngine$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
