% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_EngineDB_MassSpecEngineDB.R
\name{MassSpecEngineDB}
\alias{MassSpecEngineDB}
\title{EngineDB dedicated to Mass Spectrometry (MS) data}
\description{
R6 child of EngineDB for MassSpec data that uses MassSpecAnalysesDB for on-disk storage.
}
\section{Super class}{
\code{\link[StreamFind:EngineDB]{StreamFind::EngineDB}} -> \code{MassSpecEngineDB}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{Analyses}}{A MassSpecAnalysesDB object backed by DuckDB.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MassSpecEngineDB-new}{\code{MassSpecEngineDB$new()}}
\item \href{#method-MassSpecEngineDB-add_analyses}{\code{MassSpecEngineDB$add_analyses()}}
\item \href{#method-MassSpecEngineDB-get_analysis_names}{\code{MassSpecEngineDB$get_analysis_names()}}
\item \href{#method-MassSpecEngineDB-get_replicate_names}{\code{MassSpecEngineDB$get_replicate_names()}}
\item \href{#method-MassSpecEngineDB-get_blank_names}{\code{MassSpecEngineDB$get_blank_names()}}
\item \href{#method-MassSpecEngineDB-get_concentrations}{\code{MassSpecEngineDB$get_concentrations()}}
\item \href{#method-MassSpecEngineDB-set_replicate_names}{\code{MassSpecEngineDB$set_replicate_names()}}
\item \href{#method-MassSpecEngineDB-set_blank_names}{\code{MassSpecEngineDB$set_blank_names()}}
\item \href{#method-MassSpecEngineDB-set_concentrations}{\code{MassSpecEngineDB$set_concentrations()}}
\item \href{#method-MassSpecEngineDB-info_analyses}{\code{MassSpecEngineDB$info_analyses()}}
\item \href{#method-MassSpecEngineDB-get_spectra_headers}{\code{MassSpecEngineDB$get_spectra_headers()}}
\item \href{#method-MassSpecEngineDB-get_chromatograms_headers}{\code{MassSpecEngineDB$get_chromatograms_headers()}}
\item \href{#method-MassSpecEngineDB-list_db_tables}{\code{MassSpecEngineDB$list_db_tables()}}
\item \href{#method-MassSpecEngineDB-get_db_table_info}{\code{MassSpecEngineDB$get_db_table_info()}}
\item \href{#method-MassSpecEngineDB-query_db}{\code{MassSpecEngineDB$query_db()}}
\item \href{#method-MassSpecEngineDB-clone}{\code{MassSpecEngineDB$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="EngineDB" data-id="add_audit_entry"><a href='../../StreamFind/html/EngineDB.html#method-EngineDB-add_audit_entry'><code>StreamFind::EngineDB$add_audit_entry()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="EngineDB" data-id="get_analyses"><a href='../../StreamFind/html/EngineDB.html#method-EngineDB-get_analyses'><code>StreamFind::EngineDB$get_analyses()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="EngineDB" data-id="get_audit_trail"><a href='../../StreamFind/html/EngineDB.html#method-EngineDB-get_audit_trail'><code>StreamFind::EngineDB$get_audit_trail()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="EngineDB" data-id="get_configuration"><a href='../../StreamFind/html/EngineDB.html#method-EngineDB-get_configuration'><code>StreamFind::EngineDB$get_configuration()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="EngineDB" data-id="get_engine_info"><a href='../../StreamFind/html/EngineDB.html#method-EngineDB-get_engine_info'><code>StreamFind::EngineDB$get_engine_info()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="EngineDB" data-id="get_metadata"><a href='../../StreamFind/html/EngineDB.html#method-EngineDB-get_metadata'><code>StreamFind::EngineDB$get_metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="EngineDB" data-id="get_results"><a href='../../StreamFind/html/EngineDB.html#method-EngineDB-get_results'><code>StreamFind::EngineDB$get_results()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="EngineDB" data-id="get_workflow"><a href='../../StreamFind/html/EngineDB.html#method-EngineDB-get_workflow'><code>StreamFind::EngineDB$get_workflow()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="EngineDB" data-id="set_analyses"><a href='../../StreamFind/html/EngineDB.html#method-EngineDB-set_analyses'><code>StreamFind::EngineDB$set_analyses()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="EngineDB" data-id="set_configuration"><a href='../../StreamFind/html/EngineDB.html#method-EngineDB-set_configuration'><code>StreamFind::EngineDB$set_configuration()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="EngineDB" data-id="set_metadata"><a href='../../StreamFind/html/EngineDB.html#method-EngineDB-set_metadata'><code>StreamFind::EngineDB$set_metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="EngineDB" data-id="set_results"><a href='../../StreamFind/html/EngineDB.html#method-EngineDB-set_results'><code>StreamFind::EngineDB$set_results()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="EngineDB" data-id="set_workflow"><a href='../../StreamFind/html/EngineDB.html#method-EngineDB-set_workflow'><code>StreamFind::EngineDB$set_workflow()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MassSpecEngineDB-new"></a>}}
\if{latex}{\out{\hypertarget{method-MassSpecEngineDB-new}{}}}
\subsection{Method \code{new()}}{
Initialize MassSpecEngineDB
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MassSpecEngineDB$new(
  project_dir = "data.sf",
  metadata = NULL,
  workflow = NULL,
  analyses = NULL,
  configuration = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project_dir}}{Path to StreamFind project folder (.sf). EngineDB will place main.duckdb inside it.}

\item{\code{metadata}}{Optional metadata to persist.}

\item{\code{workflow}}{Optional workflow to persist.}

\item{\code{analyses}}{Optional MS files/data.frame to load into analyses DB.}

\item{\code{configuration}}{Optional configuration list to persist.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MassSpecEngineDB-add_analyses"></a>}}
\if{latex}{\out{\hypertarget{method-MassSpecEngineDB-add_analyses}{}}}
\subsection{Method \code{add_analyses()}}{
Add analyses (append; overwrites duplicates)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MassSpecEngineDB$add_analyses(files, centroid = FALSE, levels = c(1, 2))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{files}}{Files/data.frame accepted by \code{.get_MassSpecAnalysis_from_files}.}

\item{\code{centroid}}{Logical, passed to parser.}

\item{\code{levels}}{Integer vector, MS levels to keep.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MassSpecEngineDB-get_analysis_names"></a>}}
\if{latex}{\out{\hypertarget{method-MassSpecEngineDB-get_analysis_names}{}}}
\subsection{Method \code{get_analysis_names()}}{
Get analysis names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MassSpecEngineDB$get_analysis_names()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MassSpecEngineDB-get_replicate_names"></a>}}
\if{latex}{\out{\hypertarget{method-MassSpecEngineDB-get_replicate_names}{}}}
\subsection{Method \code{get_replicate_names()}}{
Get replicate names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MassSpecEngineDB$get_replicate_names()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MassSpecEngineDB-get_blank_names"></a>}}
\if{latex}{\out{\hypertarget{method-MassSpecEngineDB-get_blank_names}{}}}
\subsection{Method \code{get_blank_names()}}{
Get blank names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MassSpecEngineDB$get_blank_names()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MassSpecEngineDB-get_concentrations"></a>}}
\if{latex}{\out{\hypertarget{method-MassSpecEngineDB-get_concentrations}{}}}
\subsection{Method \code{get_concentrations()}}{
Get concentrations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MassSpecEngineDB$get_concentrations()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MassSpecEngineDB-set_replicate_names"></a>}}
\if{latex}{\out{\hypertarget{method-MassSpecEngineDB-set_replicate_names}{}}}
\subsection{Method \code{set_replicate_names()}}{
Set replicate names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MassSpecEngineDB$set_replicate_names(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Character vector of replicate names matching analyses.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MassSpecEngineDB-set_blank_names"></a>}}
\if{latex}{\out{\hypertarget{method-MassSpecEngineDB-set_blank_names}{}}}
\subsection{Method \code{set_blank_names()}}{
Set blank names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MassSpecEngineDB$set_blank_names(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Character vector of blank names matching analyses.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MassSpecEngineDB-set_concentrations"></a>}}
\if{latex}{\out{\hypertarget{method-MassSpecEngineDB-set_concentrations}{}}}
\subsection{Method \code{set_concentrations()}}{
Set concentrations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MassSpecEngineDB$set_concentrations(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Numeric vector of concentrations matching analyses.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MassSpecEngineDB-info_analyses"></a>}}
\if{latex}{\out{\hypertarget{method-MassSpecEngineDB-info_analyses}{}}}
\subsection{Method \code{info_analyses()}}{
Get info summary
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MassSpecEngineDB$info_analyses()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MassSpecEngineDB-get_spectra_headers"></a>}}
\if{latex}{\out{\hypertarget{method-MassSpecEngineDB-get_spectra_headers}{}}}
\subsection{Method \code{get_spectra_headers()}}{
Get spectra headers for an analysis
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MassSpecEngineDB$get_spectra_headers(analysis = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analysis}}{Analysis name.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MassSpecEngineDB-get_chromatograms_headers"></a>}}
\if{latex}{\out{\hypertarget{method-MassSpecEngineDB-get_chromatograms_headers}{}}}
\subsection{Method \code{get_chromatograms_headers()}}{
Get chromatograms headers for an analysis
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MassSpecEngineDB$get_chromatograms_headers(analysis = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analysis}}{Analysis name.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MassSpecEngineDB-list_db_tables"></a>}}
\if{latex}{\out{\hypertarget{method-MassSpecEngineDB-list_db_tables}{}}}
\subsection{Method \code{list_db_tables()}}{
List tables in analyses DB
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MassSpecEngineDB$list_db_tables()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MassSpecEngineDB-get_db_table_info"></a>}}
\if{latex}{\out{\hypertarget{method-MassSpecEngineDB-get_db_table_info}{}}}
\subsection{Method \code{get_db_table_info()}}{
Get table info in analyses DB
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MassSpecEngineDB$get_db_table_info(table_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{table_name}}{Table name.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MassSpecEngineDB-query_db"></a>}}
\if{latex}{\out{\hypertarget{method-MassSpecEngineDB-query_db}{}}}
\subsection{Method \code{query_db()}}{
Run query on analyses DB
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MassSpecEngineDB$query_db(sql, params = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sql}}{SQL string.}

\item{\code{params}}{Optional params.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MassSpecEngineDB-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MassSpecEngineDB-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MassSpecEngineDB$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
