% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecMethod_AnnotateFeatures.R
\name{MassSpecMethod_AnnotateFeatures_StreamFind}
\alias{MassSpecMethod_AnnotateFeatures_StreamFind}
\title{Mass Spectrometry Method for Annotating Features in NTS (StreamFind algorithm)}
\usage{
MassSpecMethod_AnnotateFeatures_StreamFind(
  rtWindowAlignment = 0.3,
  maxIsotopes = 8,
  maxCharge = 1,
  maxGaps = 1
)
}
\arguments{
\item{rtWindowAlignment}{Numeric (length 1) with the proportion of the monoisotopic feature time
window to be used for retrieving isotopic candidates.}

\item{maxIsotopes}{Numeric (length 1) with the maximum number of isotopic steps.}

\item{maxCharge}{Numeric (length 1) with the maximum charge that ions can be ionized to find
isotopes.}

\item{maxGaps}{Numeric (length 1) with the maximum of allowed gaps in isotopic chains.}
}
\value{
A \code{MassSpecMethod_AnnotateFeatures_StreamFind} object.
}
\description{
Method for annotation of isotopic and adduct features. The method uses the
\code{maxIsotopes} to define the maximum length of the isotopic chain. The list of candidate features
is build with the \code{rtWindowAlignment} and the maximum mass increment to match the maximum chain
length. Then, the mass difference  of the natural isotopes and a given monoisotopic ion (i.e.,
feature) are targeted. Each candidate is then evaluated according to the mass error and the
expected relative intensity. Adducts are annotated by the mass difference of the monoisotopic
ion and the adduct ion. Elements considered for annotation of isotopes are C, H, N, O, S, Cl, Br
and Si, which are the most relevent for organic molecules. The adducts considered are Na+, K+,
NH4+, CH3OH+, DMSO+, CH3CN+ for positive ionization and Cl-,Br-, -2H+Na, -2H+K, CHO2- and
CH3COO- for negative ionization. In-source fragments are not yet annotated in the current
version but planned.
}
