% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecMethod_AverageSpectra.R
\name{MassSpecMethod_AverageSpectra_StreamFind}
\alias{MassSpecMethod_AverageSpectra_StreamFind}
\alias{validate_object.MassSpecMethod_AverageSpectra_StreamFind}
\title{MassSpecMethod_AverageSpectra_StreamFind Class}
\usage{
MassSpecMethod_AverageSpectra_StreamFind(
  by = "replicates",
  weightedAveraged = TRUE
)

\method{validate_object}{MassSpecMethod_AverageSpectra_StreamFind}(x)
}
\arguments{
\item{by}{Character (length 1) with the grouping variable for averaging. Possible variables are
\code{replicates}, \code{chrom_peaks}, \code{rt}, \code{replicates+chrom_peaks}, \code{replicates+rt}, \code{chrom_peaks+rt},
\code{replicates+chrom_peaks+rt}.}

\item{weightedAveraged}{Logical (length 1) for weighted averaging.}

\item{x}{A MassSpecMethod_AverageSpectra_StreamFind object.}
}
\value{
A MassSpecMethod_AverageSpectra_StreamFind object.
}
\description{
Averages spectra based on variables.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate_object(MassSpecMethod_AverageSpectra_StreamFind)}: Validate the MassSpecMethod_AverageSpectra_StreamFind object, returning NULL if valid.

}}
