% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecMethod_CalculateSpectraCharges.R
\name{MassSpecMethod_CalculateSpectraCharges_native}
\alias{MassSpecMethod_CalculateSpectraCharges_native}
\title{MassSpecMethod_CalculateSpectraCharges_native class}
\usage{
MassSpecMethod_CalculateSpectraCharges_native(
  roundVal = 0.05,
  relLowCut = 0.05,
  absLowCut = 500,
  topCharges = 3L
)
}
\arguments{
\item{roundVal}{m/z rounding tolerance for isotope matching.}

\item{relLowCut}{Relative intensity cut-off (fraction of base peak).}

\item{absLowCut}{Absolute intensity cut-off.}

\item{topCharges}{Maximum number of charge states to return per ion.}
}
\description{
Estimate charge states of ions from MS spectra using isotope spacing.
}
