% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecMethod_CalculateSpectraCharges.R
\name{MassSpecMethod_CalculateSpectraCharges_StreamFind}
\alias{MassSpecMethod_CalculateSpectraCharges_StreamFind}
\title{\strong{MassSpecMethod_CalculateSpectraCharges_StreamFind}}
\usage{
MassSpecMethod_CalculateSpectraCharges_StreamFind(
  onlyTopScans = FALSE,
  topScans = 1,
  roundVal = 35,
  relLowCut = 0.2,
  absLowCut = 300,
  top = 5
)
}
\arguments{
\item{onlyTopScans}{Logical (length 1) to consider only the \emph{n} \code{TopScans} with the highest
intensity.}

\item{topScans}{Numeric (length 1) with the number of top scans to be considered.}

\item{roundVal}{Numeric (length 1) with the rounding value for the m/z values before applying
charge clustering.}

\item{relLowCut}{Numeric (length 1) with the relative low cut for the charge clustering.}

\item{absLowCut}{Numeric (length 1) with the absolute low cut for the charge clustering.}

\item{top}{Numeric (length 1) with the number of top charges to be considered.}
}
\value{
A MassSpecMethod_CalculateSpectraCharges_StreamFind object.
}
\description{
Calculates spectral charges from multi-charged compounds (e.g. proteins and
monoclonal antibodies) for mass deconvolution.
}
