% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/class_MassSpecMethod_CalculateTransformationProductsSimilarity.R
\name{MassSpecMethod_CalculateTransformationProductsSimilarity_native}
\alias{MassSpecMethod_CalculateTransformationProductsSimilarity_native}
\title{MassSpecMethod_CalculateTransformationProductsSimilarity_native Class}
\usage{
MassSpecMethod_CalculateTransformationProductsSimilarity_native(
  mzr = 0.008,
  minNumberSharedFragments = 1,
  minCosine = 0.2,
  filter = FALSE
)
}
\arguments{
\item{mzr}{The m/z tolerance for fragment matching. Default is 0.008.}

\item{minNumberSharedFragments}{Minimum number of shared fragments required for filtering. Default is 1.}

\item{minCosine}{Minimum cosine similarity required for filtering. Default is 0.2.}

\item{filter}{Logical, if TRUE applies filtering based on minNumberSharedFragments and minCosine. Default is FALSE.}
}
\value{
A \code{MassSpecMethod_CalculateTransformationProductsSimilarity_native} object.
}
\description{
Processing method for calculating MS2 spectral similarity between transformation products and their parent compounds using cosine similarity. This method updates the MassSpecResults_TransformationProducts object with parent similarity scores and maximum shared fragments information.
}
