% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecMethod_ClusterSpectra.R
\name{MassSpecMethod_ClusterSpectra_StreamFind}
\alias{MassSpecMethod_ClusterSpectra_StreamFind}
\title{MassSpecMethod_ClusterSpectra_StreamFind Class}
\usage{
MassSpecMethod_ClusterSpectra_StreamFind(
  val = "mz",
  clustVal = 0.001,
  presence = 0.1
)
}
\arguments{
\item{val}{Character (length 1) with the variable to be used for clustering.}

\item{clustVal}{Numeric (length 1) with the clustering value.}

\item{presence}{Numeric (length 1) with the minimum presence of traces in a cluster to be
considered.}
}
\value{
A MassSpecMethod_ClusterSpectra_StreamFind object.
}
\description{
Clusters spectra based on a variable (i.e. column name).
}
