% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/class_MassSpecMethod_CorrectChromatogramsBaseline.R
\name{MassSpecMethod_CorrectChromatogramsBaseline_baseline_als}
\alias{MassSpecMethod_CorrectChromatogramsBaseline_baseline_als}
\title{MassSpecMethod_CorrectChromatogramsBaseline_baseline_als class}
\usage{
MassSpecMethod_CorrectChromatogramsBaseline_baseline_als(
  lambda = 1e+05,
  p = 0.001,
  maxit = 10L
)
}
\arguments{
\item{lambda}{Smoothing penalty.}

\item{p}{Asymmetry parameter (0 < p < 0.5; typically 0.001–0.01).}

\item{maxit}{Maximum number of iterations.}
}
\description{
Estimate and subtract chromatogram baselines using the Asymmetric Least Squares (ALS) algorithm (Eilers & Boelens 2005).
}
