% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecMethod_DeconvoluteSpectra.R
\name{MassSpecMethod_DeconvoluteSpectra_native}
\alias{MassSpecMethod_DeconvoluteSpectra_native}
\title{MassSpecMethod_DeconvoluteSpectra_native class}
\usage{
MassSpecMethod_DeconvoluteSpectra_native(clustVal = 0.001, window = 1.5)
}
\arguments{
\item{clustVal}{m/z clustering tolerance for grouping charge series.}

\item{window}{Mass window for merging redundant deconvoluted masses.}
}
\description{
Deconvolute charge-state envelopes into neutral masses.
}
