% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecMethod_FeatureBlankSubtraction.R
\name{MassSpecMethod_FeatureBlankSubtraction_native}
\alias{MassSpecMethod_FeatureBlankSubtraction_native}
\title{MassSpecMethod_FeatureBlankSubtraction_native class}
\usage{
MassSpecMethod_FeatureBlankSubtraction_native(
  blankThreshold = 5,
  rtExpand = 10,
  mzExpand = 0.005
)
}
\arguments{
\item{blankThreshold}{numeric(1) threshold multiplier for blank feature intensities to consider a feature as present in blanks. To be present a features must be at least blankThreshold times more intense in samples than in blanks.}

\item{rtExpand}{numeric(1) retention time expansion window in seconds around feature RT for EIC extraction.}

\item{mzExpand}{numeric(1) m/z expansion window in Da around feature m/z for EIC extraction.}
}
\description{
Native StreamFind method for blank subtraction of features.
}
