% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecMethod_FilterFeatures.R
\name{MassSpecMethod_FilterFeatures_patRoon}
\alias{MassSpecMethod_FilterFeatures_patRoon}
\title{\strong{MassSpecMethod_FilterFeatures_patRoon}}
\usage{
MassSpecMethod_FilterFeatures_patRoon(
  absMinIntensity = NA_real_,
  relMinIntensity = NA_real_,
  preAbsMinIntensity = NA_real_,
  preRelMinIntensity = NA_real_,
  absMinAnalyses = NA_real_,
  relMinAnalyses = NA_real_,
  absMinReplicates = NA_real_,
  relMinReplicates = NA_real_,
  absMinFeatures = NA_real_,
  relMinFeatures = NA_real_,
  absMinReplicateAbundance = NA_real_,
  relMinReplicateAbundance = NA_real_,
  absMinConc = NA_real_,
  relMinConc = NA_real_,
  absMaxTox = NA_real_,
  relMaxTox = NA_real_,
  absMinConcTox = NA_real_,
  relMinConcTox = NA_real_,
  maxReplicateIntRSD = NA_real_,
  blankThreshold = NA_real_,
  retentionRange = NA_real_,
  mzRange = NA_real_,
  mzDefectRange = NA_real_,
  chromWidthRange = NA_real_,
  featQualityRange = NA_real_,
  groupQualityRange = NA_real_,
  rGroups = NA_real_,
  removeBlanks = FALSE,
  removeISTDs = FALSE,
  removeNA = FALSE,
  negate = FALSE
)
}
\arguments{
\item{absMinIntensity}{Numeric length one. Minimum absolute intensity for a feature.}

\item{relMinIntensity}{Numeric length one. Minimum relative intensity for a feature.}

\item{preAbsMinIntensity}{Numeric length one. Minimum absolute intensity for a feature before
grouping.}

\item{preRelMinIntensity}{Numeric length one. Minimum relative intensity for a feature before
grouping.}

\item{absMinAnalyses}{Numeric length one. Minimum number of analyses a feature must be present
in.}

\item{relMinAnalyses}{Numeric length one. Minimum relative number of analyses a feature must be
present in.}

\item{absMinReplicates}{Numeric length one. Minimum number of replicates a feature must be
present in.}

\item{relMinReplicates}{Numeric length one. Minimum relative number of replicates a feature must
be present in.}

\item{absMinFeatures}{Numeric length one. Minimum number of features a feature group must contain.}

\item{relMinFeatures}{Numeric length one. Minimum relative number of features a feature group
must contain.}

\item{absMinReplicateAbundance}{Numeric length one. Minimum absolute abundance of a replicate.}

\item{relMinReplicateAbundance}{Numeric length one. Minimum relative abundance of a replicate.}

\item{absMinConc}{Numeric length one. Minimum absolute concentration of a feature.}

\item{relMinConc}{Numeric length one. Minimum relative concentration of a feature.}

\item{absMaxTox}{Numeric length one. Maximum absolute toxicity of a feature.}

\item{relMaxTox}{Numeric length one. Maximum relative toxicity of a feature.}

\item{absMinConcTox}{Numeric length one. Minimum absolute concentration of a feature to be
considered toxic.}

\item{relMinConcTox}{Numeric length one. Minimum relative concentration of a feature to be
considered toxic.}

\item{maxReplicateIntRSD}{Numeric length one. Maximum relative standard deviation of intensities
within a replicate.}

\item{blankThreshold}{Numeric length one. Maximum intensity of a feature to be considered a
blank.}

\item{retentionRange}{Numeric length two. Retention time range (in seconds) for a feature.}

\item{mzRange}{Numeric length two. m/z range (in Da) for a feature.}

\item{mzDefectRange}{Numeric length two. m/z defect range (in Da) for a feature.}

\item{chromWidthRange}{Numeric length two. Chromatographic width range (in seconds) for a
feature.}

\item{featQualityRange}{Numeric length two. Feature quality range for a feature.}

\item{groupQualityRange}{Numeric length two. Group quality range for a feature group.}

\item{rGroups}{Character with the replicate group names to filter.}

\item{removeBlanks}{Logical length one. Remove blank samples.}

\item{removeISTDs}{Logical length one. Remove internal standards.}

\item{removeNA}{Logical length one. Remove NA values.}

\item{negate}{Logical length one. Negate the filter.}
}
\value{
A MassSpecMethod_FilterFeatures_patRoon object.
}
\description{
Settings for filtering of features and feature groups. A full description of the
filtering parameters is in \code{\link[patRoon]{replicateGroupSubtract}} from patRoon package.
}
\details{
Note that when filters are applied to features or feature groups
these require specific results from processing modules. For instance,
subtracting the blank can only be done after grouping features. Also, some
filters require. Thus, not all filters can be applied to features.
See \code{\link[patRoon]{features-class}} and \code{\link[patRoon]{replicateGroupSubtract}}
for further information.
}
