% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecMethod_FilterFeatures.R
\name{MassSpecMethod_FilterFeatures_StreamFind}
\alias{MassSpecMethod_FilterFeatures_StreamFind}
\alias{validate_object.MassSpecMethod_FilterFeatures_StreamFind}
\title{MassSpecMethod_FilterFeatures_StreamFind S7 class}
\usage{
MassSpecMethod_FilterFeatures_StreamFind(
  correctIntensity = TRUE,
  minSnRatio = NA_real_,
  minGaussianFit = NA_real_,
  excludeIsotopes = FALSE,
  excludeAdducts = FALSE,
  minIntensity = NA_real_,
  maxDeviationInReplicate = NA_real_,
  minAbundanceInReplicate = NA_real_,
  blankThreshold = NA_real_,
  conservative = TRUE,
  onlyWithMS2 = FALSE
)

\method{validate_object}{MassSpecMethod_FilterFeatures_StreamFind}(x)
}
\arguments{
\item{correctIntensity}{Logical (length 1) with \code{TRUE} for correcting the intensity for matrix
suppression before applying intensity based filters. Note that suppression factor as obtained by
the processing method \code{CorrectMatrixSuppression} must be present. When not available the
correction is ignored even if set to \code{TRUE}.}

\item{minSnRatio}{Numeric (length 1) with the minimum signal-to-noise ratio.}

\item{minGaussianFit}{Numeric (length 1) with the minimum Gaussian fit.}

\item{excludeIsotopes}{Logical (length 1) with \code{TRUE} for filtering annotated isotopes
(only prevails the monoisotopic features).}

\item{excludeAdducts}{Logical (length 1) with \code{TRUE} for filtering annotated adducts.}

\item{minIntensity}{Numeric (length 1) with the minimum intensity threshold.}

\item{maxDeviationInReplicate}{Numeric (length 1) with the maximum standard deviation of the
intensity in each replicate group. The value must be a percentage from 0 to 100. The filter is
applied if the minimum standard deviation from all replicates is above the value.}

\item{minAbundanceInReplicate}{Numeric (length 1) with the minimum abundance in each replicate
group. The filter is applied if the maximum abundance from all replicates is below the value.}

\item{blankThreshold}{Numeric (length 1) with the intensity multiplier to set the blank
threshold. The filter is applied if the maximum mean intensity from all non-blank replicates
is below the correspondent blank intensity times the multiplier.}

\item{conservative}{Logical (length 1) with \code{TRUE} for conservative filtering. The filters are
applied on a feature group basis. If \code{FALSE}, the filters are applied on a feature basis. This
means that when \code{TRUE} a feature is only filtered if all features within a feature group comply
with the filter.}

\item{onlyWithMS2}{Logical (length 1) with \code{TRUE} for filtering features without MS2 spectra.}

\item{x}{A \code{MassSpecMethod_FilterFeatures_StreamFind} object.}
}
\value{
A \code{MassSpecMethod_FilterFeatures_StreamFind} object.
}
\description{
Settings for filtering of features and feature groups.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate_object(MassSpecMethod_FilterFeatures_StreamFind)}: Validate the MassSpecMethod_FilterFeatures_StreamFind object, returning NULL if valid.

}}
