% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecMethod_FilterFeaturesMS2.R
\name{MassSpecMethod_FilterFeaturesMS2_native}
\alias{MassSpecMethod_FilterFeaturesMS2_native}
\title{\strong{MassSpecMethod_FilterFeaturesMS2_native}}
\usage{
MassSpecMethod_FilterFeaturesMS2_native(
  top = NULL,
  minIntensity = NULL,
  relMinIntensity = NULL
)
}
\arguments{
\item{top}{An integer specifying the number of top features to be kept.}

\item{minIntensity}{A numeric value specifying the minimum intensity.}

\item{relMinIntensity}{A numeric value specifying the relative minimum intensity.}
}
\value{
A \code{MassSpecMethod_FilterFeaturesMS2_native} object.
}
\description{
Settings for filtering (i.e., cleaning) non-relevant traces in features MS2
spectrum.
}
