% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecMethod_FilterSuspects.R
\name{MassSpecMethod_FilterSuspects_StreamFind}
\alias{MassSpecMethod_FilterSuspects_StreamFind}
\title{MassSpecMethod_FilterSuspects_StreamFind Class}
\usage{
MassSpecMethod_FilterSuspects_StreamFind(
  minSharedFragments = NA_real_,
  minCuiness = NA_real_,
  maxIdLevel = NA_real_
)
}
\arguments{
\item{minSharedFragments}{Numeric (length 1) with the minimum number of shared fragments.
Suspects with shared_fragments below this value will be filtered out.}

\item{minCuiness}{Numeric (length 1) with the minimum cuiness score.
Suspects with cusiness below this value will be filtered out.}

\item{maxIdLevel}{Numeric (length 1) with the maximum identification level.
Suspects with id_level above this value will be filtered out.}
}
\value{
A \code{MassSpecMethod_FilterSuspects_StreamFind} object.
}
\description{
Settings for filtering of suspects based on screening quality metrics.
}
