% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/class_MassSpecMethod_FilterTransformationProducts.R
\name{MassSpecMethod_FilterTransformationProducts_StreamFind}
\alias{MassSpecMethod_FilterTransformationProducts_StreamFind}
\title{MassSpecMethod_FilterTransformationProducts_StreamFind Class}
\usage{
MassSpecMethod_FilterTransformationProducts_StreamFind(
  minParentSimilarity = 0,
  minSharedParentFragments = 0
)
}
\arguments{
\item{minParentSimilarity}{Numeric (length 1) with the minimum parent similarity threshold.
Transformation products with parent_similarity below this value will be filtered out.}

\item{minSharedParentFragments}{Numeric (length 1) with the minimum number of shared parent fragments.
Transformation products with number_shared_fragments below this value will be filtered out.}
}
\value{
A \code{MassSpecMethod_FilterTransformationProducts_StreamFind} object.
}
\description{
Settings for filtering of transformation products.
}
