% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecMethod_FindChromPeaks.R
\name{MassSpecMethod_FindChromPeaks_LocalMaxima}
\alias{MassSpecMethod_FindChromPeaks_LocalMaxima}
\title{MassSpecMethod_FindChromPeaks_LocalMaxima class}
\usage{
MassSpecMethod_FindChromPeaks_LocalMaxima(
  minHeight = 0,
  minWidth = 0,
  maxWidth = 0
)
}
\arguments{
\item{minHeight}{Minimum peak height (0 = no filter).}

\item{minWidth}{Minimum peak width in rt units (0 = no filter).}

\item{maxWidth}{Maximum peak width in rt units (0 = no filter).}
}
\description{
Find chromatographic peaks by local-maxima expansion (DAD / UV chromatograms).
}
