% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecMethod_FindFeatures.R
\name{MassSpecMethod_FindFeatures_openms}
\alias{MassSpecMethod_FindFeatures_openms}
\title{MassSpecMethod_FindFeatures_openms S7 class}
\usage{
MassSpecMethod_FindFeatures_openms(
  noiseThrInt = 1000,
  chromSNR = 3,
  chromFWHM = 7,
  mzPPM = 15,
  reEstimateMTSD = TRUE,
  traceTermCriterion = "sample_rate",
  traceTermOutliers = 5,
  minSampleRate = 1,
  minTraceLength = 4,
  maxTraceLength = 70,
  widthFiltering = "fixed",
  minFWHM = 4,
  maxFWHM = 35,
  traceSNRFiltering = TRUE,
  localRTRange = 0,
  localMZRange = 0,
  isotopeFilteringModel = "none",
  MZScoring13C = FALSE,
  useSmoothedInts = FALSE,
  intSearchRTWindow = 3,
  useFFMIntensities = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{noiseThrInt}{Intensity threshold below which peaks are regarded as noise.}

\item{chromSNR}{Minimum signal-to-noise a mass trace should have.}

\item{chromFWHM}{Expected chromatographic peak width (in seconds).}

\item{mzPPM}{Allowed mass deviation (in ppm).}

\item{reEstimateMTSD}{Enables dynamic re-estimation of m/z variance during
mass trace collection stage.}

\item{traceTermCriterion}{Termination criterion for the extension of mass
traces. In 'outlier' mode, trace extension cancels if a predefined number of
consecutive outliers are found (see trace_termination_outliers parameter).
In 'sample_rate' mode, trace extension in both directions stops if ratio of
found peaks versus visited spectra falls below the 'min_sample_rate' threshold.}

\item{traceTermOutliers}{Mass trace extension in one direction cancels if
this number of consecutive spectra with no detectable peaks is reached.}

\item{minSampleRate}{Minimum fraction of scans along the mass trace that must
contain a peak.}

\item{minTraceLength}{Minimum expected length of a mass trace (in seconds).}

\item{maxTraceLength}{Maximum expected length of a mass trace (in seconds).
Set to a negative value to disable maximal length check during mass trace
detection.}

\item{widthFiltering}{Enable filtering of unlikely peak widths. The fixed
setting filters out mass traces outside the \code{min_fwhm}, \code{max_fwhm} interval
(set parameters accordingly!). The auto setting filters with the 5 and 95\%
quantiles of the peak width distribution.}

\item{minFWHM}{Minimum full-width-at-half-maximum of chromatographic peaks
(in seconds). Ignored if parameter width_filtering is off or auto.}

\item{maxFWHM}{Maximum full-width-at-half-maximum of chromatographic peaks
(in seconds). Ignored if parameter width_filtering is off or auto.}

\item{traceSNRFiltering}{Apply post-filtering by signal-to-noise ratio after
smoothing.}

\item{localRTRange}{RT range where to look for coeluting mass traces.}

\item{localMZRange}{MZ range where to look for isotopic mass traces.}

\item{isotopeFilteringModel}{Remove/score candidate assemblies based on
isotope intensities. SVM isotope models for metabolites were trained with
either 2\% or 5\% RMS error. For peptides, an averagine cosine scoring is used.
Select the appropriate noise model according to the quality of measurement
or MS device.}

\item{MZScoring13C}{Use the 13C isotope peak position (~1.003355 Da) as the
expected shift in m/z for isotope mass traces (highly recommended for
lipidomics!). Disable for general metabolites
(as described in Kenar et al. 2014, MCP.).}

\item{useSmoothedInts}{Use LOWESS intensities instead of raw intensities.}

\item{intSearchRTWindow}{Retention time window (in seconds, +/- feature
retention time) that is used to find the closest data point to the retention
time to obtain the intensity of a feature (this is needed since OpenMS does
not provide this data).}

\item{useFFMIntensities}{If TRUE then peak intensities are directly loaded
from FeatureFinderMetabo output. Otherwise, intensities are loaded afterwards
from the input ‘mzML’ files, which is potentially much slower, especially
with many analyses files. However, useFFMIntensities=TRUE is still somewhat
experimental, may be less accurate and requires a recent version of OpenMS
(>=2.7).}

\item{verbose}{Logical of length one. When TRUE adds processing information to the console.}
}
\value{
A \code{MassSpecMethod_FindFeatures_openms} object.
}
\description{
Settings for finding features (i.e., chromatographic peaks) in mzML/mzXML files
using the \href{https://www.openms.org/}{OpenMS}
(\url{https://abibuilder.cs.uni-tuebingen.de/archive/openms/}) software with the algorithm
\href{https://abibuilder.cs.uni-tuebingen.de/archive/openms/Documentation/release/latest/html/TOPP_FeatureFinderMetabo.html}{FeatureFinderMetabo}.
The function uses the package \pkg{patRoon} in the background.
}
\details{
See the \link[patRoon]{findFeaturesOpenMS} function from the \pkg{patRoon} package for
more information and requirements.
}
\references{
\insertRef{patroon01}{StreamFind}

\insertRef{patroon02}{StreamFind}

\insertRef{openms01}{StreamFind}
}
