% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecMethod_FindFeaturesDB.R
\name{MassSpecMethod_FindFeaturesDB_native}
\alias{MassSpecMethod_FindFeaturesDB_native}
\title{MassSpecMethod_FindFeaturesDB_native class}
\usage{
MassSpecMethod_FindFeaturesDB_native(
  rtWindows = data.frame(rtmin = 300, rtmax = 3600),
  resolution_profile = c(35000L, 35000L, 35000L),
  noiseThreshold = 250,
  minSNR = 3,
  minTraces = 3,
  baselineWindow = 200,
  maxWidth = 100,
  base_quantile = 0.1,
  debug_mz = 0
)
}
\arguments{
\item{rtWindows}{data.frame with rtmin and rtmax columns for retention time windows for data inclusion.}

\item{resolution_profile}{integer(3) vector defining mass resolution at 100, 400, and 1000 Da for calculating m/z clustering thresholds via linear model.}

\item{noiseThreshold}{numeric(1) lowest threshold to clean data (i.e., no trace with intensity below this level is kept).}

\item{minSNR}{numeric(1) minimum signal-to-noise ratio for considering a trace and chromatographic peak.}

\item{minTraces}{numeric(1) minimum number of traces to consider a mass cluster and chromatographic peak.}

\item{baselineWindow}{numeric(1) retention time window to build a baseline in a mass cluster.}

\item{maxWidth}{numeric(1) expected maximum window for a chromatographic peak.}

\item{base_quantile}{numeric(1) quantile to estimate the baseline in a mass cluster.}
}
\description{
Native StreamFind method for finding features (i.e., chromatographic peaks) in liquid chromatography coupled to high resolution mass spectrometry files.
}
