% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecMethod_FindSpectraMaxima.R
\name{MassSpecMethod_FindSpectraMaxima_native}
\alias{MassSpecMethod_FindSpectraMaxima_native}
\title{MassSpecMethod_FindSpectraMaxima_native class}
\usage{
MassSpecMethod_FindSpectraMaxima_native(
  minWidth = 0.5,
  maxWidth = 50,
  minHeight = 500
)
}
\arguments{
\item{minWidth}{Minimum peak width in m/z units.}

\item{maxWidth}{Maximum peak width in m/z units.}

\item{minHeight}{Minimum peak height (absolute intensity).}
}
\description{
Find local intensity maxima in mass spectra (peak picking in the m/z domain).
}
