% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/class_MassSpecMethod_GenerateTransformationProducts.R
\name{MassSpecMethod_GenerateTransformationProducts_biotransformer}
\alias{MassSpecMethod_GenerateTransformationProducts_biotransformer}
\title{Mass Spectrometry Method for generating transformation products using BioTransformer via patRoon}
\usage{
MassSpecMethod_GenerateTransformationProducts_biotransformer(
  parents = data.table::data.table(name = character(), SMILES = character()),
  use_suspects = FALSE,
  use_compounds = FALSE,
  type = "env",
  generations = 2,
  maxExpGenerations = generations + 2,
  extraOpts = NULL,
  skipInvalid = TRUE,
  prefCalcChemProps = TRUE,
  neutralChemProps = FALSE,
  neutralizeTPs = TRUE,
  calcSims = FALSE,
  fpType = "extended",
  fpSimMethod = "tanimoto",
  MP = FALSE
)
}
\arguments{
\item{parents}{The parents provided as data.table with at least 'name' and 'SMILES' columns for which transformation products.}

\item{use_suspects}{Logical. If TRUE suspects in the MassSpecResults_NonTargetAnalysis will be used as parents.}

\item{use_compounds}{Logical. If TRUE compounds in the MassSpecResults_NonTargetAnalysis will be used as parents.}

\item{type}{Character. The type of prediction. Valid values are: "env", "ecbased", "cyp450", "phaseII", "hgut", "superbio", "allHuman". Sets the -b command line option. Default is "env".}

\item{generations}{Integer. The number of generations (steps) for the predictions. Sets the -s command line option. Default is 2.}

\item{maxExpGenerations}{Integer. The maximum number of generations during hierarchy expansion. Default is generations + 2.}

\item{extraOpts}{Character. Extra command line options passed to the biotransformer.jar tool. Default is NULL.}

\item{skipInvalid}{Logical. If set to TRUE then the parents will be skipped (with a warning) for which insufficient information (e.g. SMILES) is available. Default is TRUE.}

\item{prefCalcChemProps}{Logical. If TRUE then calculated chemical properties such as the formula and InChIKey are preferred over what is already present in the parent suspect list. Default is TRUE.}

\item{neutralChemProps}{Logical. If TRUE then the neutral form of the molecule is considered to calculate SMILES, formulae etc. Default is FALSE.}

\item{neutralizeTPs}{Logical. If TRUE then all resulting TP structure information is neutralized. Default is TRUE.}

\item{calcSims}{Logical. If set to TRUE then structural similarities between the parent and its TPs are calculated. Default is FALSE.}

\item{fpType}{Character. The type of structural fingerprint that should be calculated. Default is "extended".}

\item{fpSimMethod}{Character. The method for calculating similarities. Default is "tanimoto".}

\item{MP}{Logical. If TRUE then multiprocessing is enabled. Default is FALSE.}
}
\value{
A \code{MassSpecMethod_GenerateTransformationProducts_biotransformer} object.
}
\description{
Generates transformation products using the \href{https://biotransformer.ca/}{BioTransformer} tool via \href{https://github.com/rickhelmus/patRoon}{patRoon} integration. This method is a wrapper of the \code{generateTPsBioTransformer} function from the patRoon package. See more details in href{https://rickhelmus.github.io/patRoon/reference/generateTPsBioTransformer.html}{here}.
}
\references{
\insertRef{patroon01}{StreamFind}

\insertRef{patroon02}{StreamFind}

\insertRef{biotransformer01}{StreamFind}

\insertRef{biotransformer02}{StreamFind}
}
