% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/class_MassSpecMethod_GenerateTransformationProducts.R
\name{MassSpecMethod_GenerateTransformationProducts_cts}
\alias{MassSpecMethod_GenerateTransformationProducts_cts}
\title{Mass Spectrometry Method for generating transformation products using the Chemical Transformation Simulator (CTS) via patRoon.}
\usage{
MassSpecMethod_GenerateTransformationProducts_cts(
  parents = data.table::data.table(name = character(), SMILES = character()),
  use_suspects = FALSE,
  use_compounds = FALSE,
  transLibrary = character(),
  generations = 1,
  errorRetries = 3,
  skipInvalid = TRUE,
  prefCalcChemProps = TRUE,
  neutralChemProps = FALSE,
  neutralizeTPs = TRUE,
  calcLogP = "rcdk",
  calcSims = FALSE,
  fpType = "extended",
  fpSimMethod = "tanimoto",
  parallel = TRUE
)
}
\arguments{
\item{parents}{The parents provided as data.table with at least 'name' and 'SMILES' columns for which transformation products.}

\item{use_suspects}{Logical. If TRUE suspects in the MassSpecResults_NonTargetAnalysis will be used as parents.}

\item{use_compounds}{Logical. If TRUE compounds in the MassSpecResults_NonTargetAnalysis will be used as parents. If more more than one structure is available for a compound, only the first one will be used as it is assumed that this is the most relevant one (i.e., the highest score).}

\item{transLibrary}{A character specifying which transformation library should be used. Currently supported are: "hydrolysis", "abiotic_reduction", "photolysis_unranked", "photolysis_ranked", "mammalian_metabolism", "combined_abioticreduction_hydrolysis", "combined_photolysis_abiotic_hydrolysis", "pfas_environmental", "pfas_metabolism".}

\item{generations}{Integer. The number of generations (steps) for the predictions. Default is 1.}

\item{errorRetries}{Integer. The number of times to retry if an error occurs during processing of a parent. Default is 3.}

\item{skipInvalid}{Logical. If set to TRUE then the parents will be skipped (with a warning) for which insufficient information (e.g. SMILES) is available. Default is TRUE.}

\item{prefCalcChemProps}{Logical. If TRUE then calculated chemical properties such as the formula and InChIKey are preferred over what is already present in the parent suspect list. Default is TRUE.}

\item{neutralChemProps}{Logical. If TRUE then the neutral form of the molecule is considered to calculate SMILES, formulae etc. Default is FALSE.}

\item{neutralizeTPs}{Logical. If TRUE then all resulting TP structure information is neutralized. Default is TRUE.}

\item{calcLogP}{Character. Method to use for LogP calculation. Currently supported are "rcdk" (default), "obabel" or "none".}

\item{calcSims}{Logical. If set to TRUE then structural similarities between the parent and its TPs are calculated. Default is FALSE.}

\item{fpType}{Character. The type of structural fingerprint that should be calculated. Default is "extended".}

\item{fpSimMethod}{Character. The method for calculating similarities. Default is "tanimoto".}

\item{parallel}{Logical. If TRUE then parallel processing is enabled. Default is TRUE.}
}
\value{
A \code{MassSpecMethod_GenerateTransformationProducts_cts} object.
}
\description{
Generates transformation products using the \href{https://qed.epa.gov/cts/}{Chemical Transformation Simulator (CTS)} tool via \href{https://github.com/rickhelmus/patRoon}{patRoon} integration. This method is a wrapper of the \code{generateTPsCTS} function from the patRoon package. See more details in \href{https://rickhelmus.github.io/patRoon/reference/generateTPsCTS.html}{here}.
}
