% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecMethod_IntegrateChromatograms.R
\name{MassSpecMethod_IntegrateChromatograms_pracma}
\alias{MassSpecMethod_IntegrateChromatograms_pracma}
\title{MassSpecMethod_IntegrateChromatograms_pracma class}
\usage{
MassSpecMethod_IntegrateChromatograms_pracma(
  merge = TRUE,
  closeByThreshold = 45,
  minPeakHeight = 0,
  minPeakDistance = 10,
  minPeakWidth = 5,
  maxPeakWidth = 120,
  minSN = 10
)
}
\arguments{
\item{merge}{Merge nearby peaks into one?}

\item{closeByThreshold}{Merge threshold (same rt units as chromatogram).}

\item{minPeakHeight}{Minimum apex intensity.}

\item{minPeakDistance}{Minimum distance between apices (rt units).}

\item{minPeakWidth}{Minimum peak width (rt units).}

\item{maxPeakWidth}{Maximum peak width (rt units).}

\item{minSN}{Minimum signal-to-noise ratio.}
}
\description{
Find and integrate chromatographic peaks using a pracma-style algorithm.
}
