% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecMethod_LoadFeaturesMS1_MS2_EIC.R
\name{MassSpecMethod_LoadFeaturesEIC_StreamFind}
\alias{MassSpecMethod_LoadFeaturesEIC_StreamFind}
\title{MassSpecMethod_LoadFeaturesEIC_StreamFind S7 class}
\usage{
MassSpecMethod_LoadFeaturesEIC_StreamFind(
  filtered = FALSE,
  rtExpand = 120,
  mzExpand = 0,
  minIntensity = 0
)
}
\arguments{
\item{filtered}{Logical (length 1). When \code{TRUE}, filtered features/groups are considered.}

\item{rtExpand}{Numeric (length 1) with the retention time values (in seconds) for expanding left and right the
retention time range.}

\item{mzExpand}{Numeric (length 1) with the mass or mass-to-charge ratio (\emph{m/z}) (in Da) for expanding left and
right the mass range.}

\item{minIntensity}{Numeric of length one with the minimum intensity of traces to extract for
EIC.}
}
\value{
A \code{MassSpecMethod_LoadFeaturesEIC_StreamFind} object.
}
\description{
Settings for loading spectra EIC for feature groups.
}
