% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecMethod_LoadSpectra.R
\name{MassSpecMethod_LoadSpectra_chrompeaks}
\alias{MassSpecMethod_LoadSpectra_chrompeaks}
\title{MassSpecMethod_LoadSpectra_chrompeaks class}
\usage{
MassSpecMethod_LoadSpectra_chrompeaks(
  levels = 2L,
  minIntensityMS1 = 0,
  minIntensityMS2 = 0
)
}
\arguments{
\item{levels}{MS levels to load (default 2L for MS2).}

\item{minIntensityMS1}{Minimum MS1 intensity filter.}

\item{minIntensityMS2}{Minimum MS2 intensity filter.}
}
\description{
Load MS2 spectra for chromatographic peak windows identified by FindChromPeaks / IntegrateChromatograms.
}
