% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecMethod_NormalizeSpectra.R
\name{MassSpecMethod_NormalizeSpectra_meancenter}
\alias{MassSpecMethod_NormalizeSpectra_meancenter}
\title{MassSpecMethod_NormalizeSpectra_meancenter Class}
\usage{
MassSpecMethod_NormalizeSpectra_meancenter()
}
\value{
A MassSpecMethod_NormalizeSpectra_meancenter object.
}
\description{
Normalizes spectra using mean centering.
}
