% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/class_MassSpecMethod_QuantifyChromatographicPeaks.R
\name{MassSpecMethod_QuantifyChromatographicPeaks_StreamFind}
\alias{MassSpecMethod_QuantifyChromatographicPeaks_StreamFind}
\title{\strong{MassSpecMethod_QuantifyChromatographicPeaks_StreamFind}}
\usage{
MassSpecMethod_QuantifyChromatographicPeaks_StreamFind(
  calibration = 0,
  value = "area",
  model = "linear"
)
}
\arguments{
\item{calibration}{Numeric with the calibration/concentration values for each analysis. When
concentration is not know use NA_real_.}

\item{value}{Character with the value to quantify. Possible values are \emph{intensity} or \emph{area}.}

\item{model}{Character with the model to use for calibration. Possible values are \emph{linear},
\emph{quadratic}, or \emph{cubic}.}
}
\value{
A MassSpecMethod_QuantifyChromatographicPeaks_StreamFind object.
}
\description{
Quantifies chromatographic peaks based on a calibration model.
}
