% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecMethod_SmoothChromatograms.R
\name{MassSpecMethod_SmoothChromatograms_savgol}
\alias{MassSpecMethod_SmoothChromatograms_savgol}
\title{MassSpecMethod_SmoothChromatograms_savgol class}
\usage{
MassSpecMethod_SmoothChromatograms_savgol(fl = 11L, forder = 4L, dorder = 0L)
}
\arguments{
\item{fl}{Filter length (must be odd).}

\item{forder}{Polynomial order.}

\item{dorder}{Derivative order (0 = smooth, 1 = 1st derivative).}
}
\description{
Smooth chromatogram traces with a Savitzky-Golay filter.
}
