% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecMethod_SmoothSpectra.R
\name{MassSpecMethod_SmoothSpectra_movingaverage}
\alias{MassSpecMethod_SmoothSpectra_movingaverage}
\title{MassSpecMethod_SmoothSpectra_movingaverage class}
\usage{
MassSpecMethod_SmoothSpectra_movingaverage(windowSize = 5L)
}
\arguments{
\item{windowSize}{Half-window size (odd integer >= 3).}
}
\description{
Smooth spectra intensity using a moving-average filter.
}
