% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecMethod_SmoothSpectra.R
\name{MassSpecMethod_SmoothSpectra_savgol}
\alias{MassSpecMethod_SmoothSpectra_savgol}
\title{MassSpecMethod_SmoothSpectra_savgol class}
\usage{
MassSpecMethod_SmoothSpectra_savgol(fl = 5L, forder = 2L, dorder = 0L)
}
\arguments{
\item{fl}{Filter length (odd integer >= 5).}

\item{forder}{Polynomial order.}

\item{dorder}{Derivative order (0 = smoothing only).}
}
\description{
Smooth spectra intensity using a Savitzky-Golay filter.
}
