% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecMethod_SuspectScreening.R
\name{MassSpecMethod_SuspectScreening_StreamFind}
\alias{MassSpecMethod_SuspectScreening_StreamFind}
\title{MassSpecMethod_SuspectScreening_StreamFind S7 class}
\usage{
MassSpecMethod_SuspectScreening_StreamFind(
  database = data.table::data.table(name = character(), formula = character(), mass =
    numeric()),
  ppm = 5,
  sec = 10,
  ppmMS2 = 10,
  mzrMS2 = 0.008,
  minCusiness = 0.5,
  minFragments = 3,
  filtered = FALSE
)
}
\arguments{
\item{database}{A data.frame with at least the columns name and mass, indicating the name and
neutral monoisotopic mass of the suspect targets.}

\item{ppm}{Numeric of length one with the mass deviation, in ppm.}

\item{sec}{Numeric of length one with the time deviation, in seconds.}

\item{ppmMS2}{Numeric (length 1) with the mass deviation for MS2, in ppm.}

\item{mzrMS2}{Numeric length 1 with the absolute m/z range window for matching MS2 fragments with
experimental MS2 traces.}

\item{minCusiness}{Numeric length 1 with the minimum cusiness value of the experimental and
database fragments for a feature to be considered a suspect.}

\item{minFragments}{Integer (length 1) with the minimum number of fragments.}

\item{filtered}{Logical (length 1). When \code{TRUE}, filtered features/groups are considered.}
}
\value{
A \code{MassSpecMethod_SuspectScreening_StreamFind} object.
}
\description{
Settings for performing suspect screening using a data.frame with target compounds.
}
