% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecResults_Chromatograms.R
\name{MassSpecResults_Chromatograms}
\alias{MassSpecResults_Chromatograms}
\alias{validate_object.MassSpecResults_Chromatograms}
\alias{show.MassSpecResults_Chromatograms}
\alias{[.MassSpecResults_Chromatograms}
\alias{get_chromatograms.MassSpecResults_Chromatograms}
\alias{plot_chromatograms.MassSpecResults_Chromatograms}
\alias{plot_chromatograms_baseline.MassSpecResults_Chromatograms}
\alias{get_chromatograms_peaks.MassSpecResults_Chromatograms}
\alias{plot_chromatograms_peaks.MassSpecResults_Chromatograms}
\title{Constructor and methods to handle Mass Spectrometry chromatograms results}
\usage{
MassSpecResults_Chromatograms(
  chromatograms = list(),
  replicates = character(),
  is_averaged = FALSE,
  peaks = list(),
  calibration_model = list()
)

\method{validate_object}{MassSpecResults_Chromatograms}(x)

\method{show}{MassSpecResults_Chromatograms}(x)

\method{[}{MassSpecResults_Chromatograms}(x, i)

\method{get_chromatograms}{MassSpecResults_Chromatograms}(
  x,
  analyses = NULL,
  chromatograms = NULL,
  rtmin = 0,
  rtmax = 0,
  minIntensity = NULL
)

\method{plot_chromatograms}{MassSpecResults_Chromatograms}(
  x,
  analyses = NULL,
  chromatograms = NULL,
  rtmin = 0,
  rtmax = 0,
  minIntensity = NULL,
  normalized = TRUE,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  colorBy = "analyses+targets",
  legendNames = NULL,
  interactive = TRUE,
  renderEngine = "webgl"
)

\method{plot_chromatograms_baseline}{MassSpecResults_Chromatograms}(
  x,
  analyses = NULL,
  chromatograms = NULL,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  colorBy = "analyses",
  interactive = TRUE,
  renderEngine = "webgl"
)

\method{get_chromatograms_peaks}{MassSpecResults_Chromatograms}(
  x,
  analyses = NULL,
  chromatograms = NULL,
  rtmin = 0,
  rtmax = 0,
  minIntensity = NULL
)

\method{plot_chromatograms_peaks}{MassSpecResults_Chromatograms}(
  x,
  analyses = NULL,
  chromatograms = NULL,
  rtmin = 0,
  rtmax = 0,
  minIntensity = NULL,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  colorBy = "analyses+targets",
  legendNames = NULL,
  interactive = TRUE,
  renderEngine = "webgl"
)
}
\arguments{
\item{chromatograms}{A character or integer vector with the ID (i.e. name) or the index of the chromatograms.}

\item{replicates}{A character vector of replicate identifiers.}

\item{is_averaged}{A logical value indicating whether the chromatograms are averaged across replicates.}

\item{peaks}{A list of peaks, each represented as a data frame.}

\item{calibration_model}{A list representing the calibration model used for the chromatograms.}

\item{x}{An object of class \code{MassSpecResults_Chromatograms}.}

\item{i}{A character or numeric vector to perform subsetting. Check the specific method for details.}

\item{analyses}{Character or numeric vector with names or indexes of analyses in the \code{Analyses}
object.}

\item{rtmin}{Numeric (length 1) with the minimum retention time.}

\item{rtmax}{Numeric (length 1) with the maximum retention time.}

\item{minIntensity}{Numeric (length 1) with the minimum intensity.}

\item{normalized}{A logical value indicating whether the result should be normalized.}

\item{xLab}{A string with the title for the x axis.}

\item{yLab}{A string with the title for the y axis.}

\item{title}{A string with the title.}

\item{colorBy}{A string defining how to legend the plot. Possible values are \code{analyses}, \code{targets} (the default) or
\code{replicates}.}

\item{legendNames}{A character vector with the same length as the targets or \code{TRUE} or \code{FALSE} for using the name in
the added targets as legend of the plot.}

\item{interactive}{Logical (length 1). When \code{TRUE}, the data is plotted interactively using \pkg{plotly}.}

\item{renderEngine}{The engine to render the data. The default is "webgl".}
}
\value{
An object of class \code{MassSpecResults_Chromatograms}.
}
\description{
The \code{MassSpecResults_Chromatograms} class represents the results of chromatograms from mass spectrometry analyses.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate_object(MassSpecResults_Chromatograms)}: Validate the MassSpecResults_Chromatograms object, returning NULL if valid.

\item \code{show(MassSpecResults_Chromatograms)}: Show the MassSpecResults_Chromatograms object.

\item \code{[}: Subset the chromatograms object.

\item \code{get_chromatograms(MassSpecResults_Chromatograms)}: Get chromatograms from the MassSpecResults_Chromatograms object, returning a list of data.table objects for each analysis.

\item \code{plot_chromatograms(MassSpecResults_Chromatograms)}: Plot chromatograms from the MassSpecResults_Chromatograms object.

\item \code{plot_chromatograms_baseline(MassSpecResults_Chromatograms)}: Plot chromatograms with baseline from the MassSpecResults_Chromatograms object.

\item \code{get_chromatograms_peaks(MassSpecResults_Chromatograms)}: Get peaks from the chromatograms object.

\item \code{plot_chromatograms_peaks(MassSpecResults_Chromatograms)}: Plot peaks from the chromatograms object.

}}
