% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecResults_NonTargetAnalysis.R
\name{MassSpecResults_NonTargetAnalysis}
\alias{MassSpecResults_NonTargetAnalysis}
\alias{validate_object.MassSpecResults_NonTargetAnalysis}
\alias{query_db.MassSpecResults_NonTargetAnalysis}
\alias{list_db_tables.MassSpecResults_NonTargetAnalysis}
\alias{get_db_table_info.MassSpecResults_NonTargetAnalysis}
\alias{show.MassSpecResults_NonTargetAnalysis}
\alias{get_features_count.MassSpecResults_NonTargetAnalysis}
\alias{plot_features_count.MassSpecResults_NonTargetAnalysis}
\alias{get_features.MassSpecResults_NonTargetAnalysis}
\alias{get_features_profile.MassSpecResults_NonTargetAnalysis}
\alias{plot_features_profile.MassSpecResults_NonTargetAnalysis}
\alias{get_internal_standards.MassSpecResults_NonTargetAnalysis}
\alias{plot_features.MassSpecResults_NonTargetAnalysis}
\alias{map_features.MassSpecResults_NonTargetAnalysis}
\alias{plot_features_ms1.MassSpecResults_NonTargetAnalysis}
\alias{plot_features_ms2.MassSpecResults_NonTargetAnalysis}
\alias{suspect_screening.MassSpecResults_NonTargetAnalysis}
\alias{get_suspects.MassSpecResults_NonTargetAnalysis}
\alias{plot_suspects_ms2.MassSpecResults_NonTargetAnalysis}
\alias{get_fold_change.MassSpecResults_NonTargetAnalysis}
\alias{plot_fold_change.MassSpecResults_NonTargetAnalysis}
\alias{get_transformation_products.MassSpecResults_NonTargetAnalysis}
\alias{plot_transformation_products.MassSpecResults_NonTargetAnalysis}
\title{Constructor and methods to handle non-target analysis (NTA) for mass spectrometry data}
\usage{
MassSpecResults_NonTargetAnalysis(
  projectPath = ".",
  features = data.table::data.table()
)

\method{validate_object}{MassSpecResults_NonTargetAnalysis}(x)

\method{query_db}{MassSpecResults_NonTargetAnalysis}(x, sql, params = NULL)

\method{list_db_tables}{MassSpecResults_NonTargetAnalysis}(x)

\method{get_db_table_info}{MassSpecResults_NonTargetAnalysis}(x, tableName)

\method{show}{MassSpecResults_NonTargetAnalysis}(x)

\method{get_features_count}{MassSpecResults_NonTargetAnalysis}(x, analyses = NULL, filtered = FALSE)

\method{plot_features_count}{MassSpecResults_NonTargetAnalysis}(
  x,
  analyses = NULL,
  filtered = FALSE,
  yLab = NULL,
  title = NULL,
  groupBy = "analysis",
  showLegend = TRUE,
  showHoverText = TRUE
)

\method{get_features}{MassSpecResults_NonTargetAnalysis}(
  x,
  analyses = NULL,
  features = NULL,
  groups = NULL,
  components = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  filtered = FALSE
)

\method{get_features_profile}{MassSpecResults_NonTargetAnalysis}(
  x,
  analyses = NULL,
  groups = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  filtered = FALSE
)

\method{plot_features_profile}{MassSpecResults_NonTargetAnalysis}(
  x,
  analyses = NULL,
  groups = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  filtered = FALSE,
  groupBy = "analysis",
  normalized = FALSE,
  yLab = NULL,
  title = NULL,
  interactive = TRUE,
  showLegend = TRUE
)

\method{get_internal_standards}{MassSpecResults_NonTargetAnalysis}(x, analyses = NULL)

\method{plot_features}{MassSpecResults_NonTargetAnalysis}(
  x,
  analyses = NULL,
  features = NULL,
  groups = NULL,
  components = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  filtered = FALSE,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  groupBy = "feature",
  interactive = TRUE,
  showDetails = FALSE
)

\method{map_features}{MassSpecResults_NonTargetAnalysis}(
  x,
  analyses = NULL,
  features = NULL,
  groups = NULL,
  components = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  filtered = FALSE,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  groupBy = "feature",
  globalNormalization = FALSE,
  interactive = TRUE,
  showDetails = FALSE
)

\method{plot_features_ms1}{MassSpecResults_NonTargetAnalysis}(
  x,
  analyses = NULL,
  features = NULL,
  groups = NULL,
  components = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  normalized = FALSE,
  filtered = FALSE,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  groupBy = "feature",
  showText = TRUE,
  interactive = TRUE
)

\method{plot_features_ms2}{MassSpecResults_NonTargetAnalysis}(
  x,
  analyses = NULL,
  features = NULL,
  groups = NULL,
  components = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  normalized = TRUE,
  filtered = FALSE,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  groupBy = "feature",
  showText = TRUE,
  interactive = TRUE
)

\method{suspect_screening}{MassSpecResults_NonTargetAnalysis}(
  x,
  analyses = NULL,
  suspects = data.frame(),
  ppm = 5,
  sec = 10,
  ppmMS2 = 10,
  mzrMS2 = 0.008,
  minCosineSimilarity = 0.7,
  minSharedFragments = 3,
  filtered = FALSE
)

\method{get_suspects}{MassSpecResults_NonTargetAnalysis}(
  x,
  analyses = NULL,
  features = NULL,
  groups = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5
)

\method{plot_suspects_ms2}{MassSpecResults_NonTargetAnalysis}(
  x,
  analyses = NULL,
  features = NULL,
  groups = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  normalized = TRUE,
  filtered = FALSE,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  groupBy = c("feature", "name"),
  showText = TRUE,
  interactive = TRUE,
  showLegend = TRUE
)

\method{get_internal_standards}{MassSpecResults_NonTargetAnalysis}(x, analyses = NULL)

\method{get_fold_change}{MassSpecResults_NonTargetAnalysis}(
  x,
  replicatesIn = NULL,
  replicatesOut = NULL,
  groups = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 4,
  sec = 10,
  millisec = 5,
  filtered = FALSE,
  constantThreshold = 0.5,
  eliminationThreshold = 0.2,
  correctIntensity = FALSE,
  fillZerosWithLowerLimit = FALSE,
  lowerLimit = NA_real_
)

\method{plot_fold_change}{MassSpecResults_NonTargetAnalysis}(
  x,
  replicatesIn = NULL,
  replicatesOut = NULL,
  groups = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 4,
  sec = 10,
  millisec = 5,
  filtered = FALSE,
  constantThreshold = 0.5,
  eliminationThreshold = 0.2,
  correctIntensity = FALSE,
  fillZerosWithLowerLimit = FALSE,
  lowerLimit = NA_real_,
  normalized = TRUE,
  yLab = NULL,
  title = NULL,
  interactive = TRUE,
  showLegend = TRUE
)

\method{get_transformation_products}{MassSpecResults_NonTargetAnalysis}(x, parents = NULL, groups = NULL)

\method{plot_transformation_products}{MassSpecResults_NonTargetAnalysis}(
  x,
  groups = NULL,
  showMS2 = FALSE,
  showIntensityProfile = FALSE
)
}
\arguments{
\item{projectPath}{Path to the project directory where this database file will be created/used.}

\item{features}{A character vector with the feature IDs or a data.frame or data.table with a column named \code{feature}
with the feature IDs. A column named \code{analysis} can be given to specify from which analyses the features matched.}

\item{x}{A \code{MassSpecResults_NonTargetAnalysisDB} object.}

\item{sql}{SQL query string to execute against the database.}

\item{params}{Query parameters to be used with the SQL query (optional).}

\item{tableName}{Name of the table to query or retrieve information about.}

\item{analyses}{Character or numeric vector with names or indexes of analyses in the \code{Analyses}
object.}

\item{filtered}{Logical (length 1). When \code{TRUE}, filtered features/groups are considered.}

\item{yLab}{A string with the title for the y axis.}

\item{title}{A string with the title.}

\item{groupBy}{Column to use for color grouping (default: "analysis").}

\item{showLegend}{Logical (length 1). Set to \code{TRUE} to show legend.}

\item{showHoverText}{Logical (length 1), set to \code{TRUE} to show hovering text annotations.}

\item{groups}{Optional character vector of feature_group ids to seed the plot. When provided,
the network expands to include all connected precursors/products.}

\item{components}{A numeric or character vector with feature component IDs, or a data.frame/data.table with a \code{feature_component} (or \code{component}) column; an optional \code{analysis} column restricts analyses and a \code{name} column sets custom labels.}

\item{mass}{A vector with target neutral mass value/s (in Da) or a two columns data.table or data.frame named \code{min}
and \code{max} with minimum and maximum neutral mass values (in Da), respectively. Alternatively, neutral mass (in Da)
and retention time (in seconds) and/or drift time values (in milliseconds) can be given as one data.table or
data.frame with columns named \code{mass} and \code{rt} and/or \code{drift}. Then, the deviations given in the \code{ppm}, \code{sec}
and \code{millisec}  arguments are used to calculate the ranges. Also works with a data.table or data.frame with
minimum and maximum values of neutral mass, retention time and drift time targets. In this case, the column names
must be \code{min}, \code{max}, \code{rtmin}, \code{rtmax}, \code{driftmin} and \code{driftmax}. Note that when mass/time ranges are given, the
\code{ppm}, \code{sec} and \code{millisec}
arguments are not used.}

\item{mz}{A vector with target \emph{m/z} value/s (in Da) or a two columns data.table or data.frame named \code{mzmin} and
\code{mzmax} with minimum and maximum \emph{m/z} values (in Da), respectively. Alternatively, \emph{m/z} (in Da) and
retention time values (in seconds) can be given as one data.table or data.frame with columns named \code{mz} and \code{rt}
and/or \code{drift}. Then, the deviations given in the \code{ppm}, \code{sec} and \code{millisec} arguments are used to
calculate the ranges. Also works with a data.table or data.frame with minimum and maximum values of \emph{m/z},
retention time and drift time targets. In this case, the column names must be \code{mzmin}, \code{mzmax}, \code{rtmin}, \code{rtmax},
\code{driftmin} and \code{driftmax}. Note that when mass/time ranges are given, the \code{ppm}, \code{sec} and \code{millisec}
arguments are not used.}

\item{rt}{A vector with target retention time values (in seconds) or
a two columns data.table or data.frame with minimum and maximum retention
time values (in seconds).}

\item{mobility}{A vector with target drift time values (in milliseconds) or
a two columns data.table or data.frame with minimum and maximum drift
time values (in milliseconds).}

\item{ppm}{Numeric of length one with the mass deviation, in ppm.}

\item{sec}{Numeric of length one with the time deviation, in seconds.}

\item{millisec}{Numeric of length one with the drift time deviation,
in milliseconds.}

\item{normalized}{A logical value indicating whether the result should be normalized.}

\item{interactive}{Logical (length 1). When \code{TRUE}, the data is plotted interactively using \pkg{plotly}.}

\item{xLab}{A string with the title for the x axis.}

\item{showDetails}{Logical, show hover details in interactive plots.}

\item{globalNormalization}{Logical, when TRUE normalize intensities globally across all selected features.}

\item{suspects}{A data.frame with suspect information. Must contain columns: \code{name} (character) and either \code{mass}
(neutral monoisotopic mass) or \code{mz} (expected m/z). Optional columns: \code{rt} (retention time in seconds), \code{formula}
(molecular formula), \code{SMILES}, \code{fragments} or \code{fragments_mz} (MS2 fragment m/z values, semicolon-separated),
\code{fragments_int} (MS2 fragment intensities, semicolon-separated), \code{fragments_formula} (fragment formulas,
semicolon-separated).}

\item{ppmMS2}{Numeric. Mass tolerance in ppm for MS2 fragment matching. Default: 10.}

\item{mzrMS2}{Numeric. Minimum absolute m/z range for MS2 fragment matching (used when ppm range is smaller).
Default: 0.008.}

\item{minCosineSimilarity}{Numeric. Minimum cosine similarity score (0-1) for MS2 spectral matching to upgrade
identification level. Default: 0.7.}

\item{minSharedFragments}{Integer. Minimum number of shared fragments for MS2 matching to upgrade identification level.
Default: 3.}

\item{replicatesIn}{Character vector with the names of the replicates to be considered as the denominator.}

\item{replicatesOut}{Character vector with the names of the replicates to be considered as the numerator.}

\item{constantThreshold}{Numeric of length one. The threshold to consider a feature as constant.}

\item{eliminationThreshold}{Numeric of length one. The threshold to consider a feature as eliminated.}

\item{correctIntensity}{Logical (length 1) with \code{TRUE} for correcting the intensity for matrix
suppression before applying intensity based filters. Note that suppression factor as obtained by
the processing method \code{CorrectMatrixSuppression} must be present. When not available the
correction is ignored even if set to \code{TRUE}.}

\item{fillZerosWithLowerLimit}{Logical of length one. When \code{TRUE} the zero values are filled with the lower limit.}

\item{lowerLimit}{Numeric of length one. The lower limit to fill the zero values.}

\item{parents}{Optional character vector of parent compound names to filter results. When provided,
returns only transformation products where the precursor or product name matches one of the specified parents.}

\item{showMS2}{Logical. When TRUE, fetches MS2 spectra from Features table for transformation products
and precursors to display mirror plots in tooltips. Default: FALSE.}

\item{showIntensityProfile}{Logical. When TRUE, adds a normalized intensity profile plot by replicate
group (mean +/- sd) for node feature groups in the tooltip. Missing feature-group/analysis combinations
are imputed as zero before aggregation. Default: FALSE.}
}
\value{
An object of class \code{MassSpecResults_NonTargetAnalysis} (and \code{Results}) pointing to the feature DuckDB on disk and holding an \code{analyses} field with the linked \code{MassSpecAnalyses}.

A data.table with matched suspects.
Identification levels: "1" = MS1+RT+MS2 match, "2" = MS1+MS2 match, "3b" = MS1+RT match, "4" = MS1 match only.

A data.table containing transformation products with their associated metadata, including precursor information,
transformation types, and structural identifiers (SMILES).
}
\description{
Create a \code{MassSpecResults_NonTargetAnalysis} object (child of \link{Results}) that reuses an existing \code{MassSpecAnalyses} DuckDB (for analyses and spectra) and stores NTA features in its own DuckDB.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate_object(MassSpecResults_NonTargetAnalysis)}: Validates the MassSpecResults_NonTargetAnalysis object, returning NULL if valid.

\item \code{query_db(MassSpecResults_NonTargetAnalysis)}: Internal: execute a query on the DB (delegates to Results).

\item \code{list_db_tables(MassSpecResults_NonTargetAnalysis)}: Internal: list tables in the DB (delegates to Results).

\item \code{get_db_table_info(MassSpecResults_NonTargetAnalysis)}: Internal: get table info from the DB (delegates to Results).

\item \code{show(MassSpecResults_NonTargetAnalysis)}: Prints a summary of the MassSpecResults_NonTargetAnalysis object.

\item \code{get_features_count(MassSpecResults_NonTargetAnalysis)}: Returns a data table with the number of features for each analysis.

\item \code{plot_features_count(MassSpecResults_NonTargetAnalysis)}: Plots the number of features for each analysis as a bar plot.

\item \code{get_features(MassSpecResults_NonTargetAnalysis)}: Retrieves features from the MassSpecResults_NonTargetAnalysis object based on specified criteria.

\item \code{get_features_profile(MassSpecResults_NonTargetAnalysis)}: Returns a data table with feature-group profiles across analyses.

\item \code{plot_features_profile(MassSpecResults_NonTargetAnalysis)}: Plots feature-group profiles across analyses or replicates.

\item \code{get_internal_standards(MassSpecResults_NonTargetAnalysis)}: Retrieves internal standards from the InternalStandards table.

\item \code{plot_features(MassSpecResults_NonTargetAnalysis)}: Plots features from the MassSpecResults_NonTargetAnalysis object according to the specified parameters.

\item \code{map_features(MassSpecResults_NonTargetAnalysis)}: Plot RT vs m/z traces for selected features using EIC data.

\item \code{plot_features_ms1(MassSpecResults_NonTargetAnalysis)}: Plot MS1 spectra for selected features.

\item \code{plot_features_ms2(MassSpecResults_NonTargetAnalysis)}: Plot MS2 spectra for selected features.

\item \code{suspect_screening(MassSpecResults_NonTargetAnalysis)}: Performs suspect screening on features stored in the DuckDB
against a provided suspect database. Matches features based on mass or \emph{m/z}, retention time, and when available
MS2 spectral similarity.

\item \code{get_suspects(MassSpecResults_NonTargetAnalysis)}: Retrieves suspect screening results from the Suspects table in the database.

\item \code{plot_suspects_ms2(MassSpecResults_NonTargetAnalysis)}: Plot suspect MS2 spectra for selected features.

\item \code{get_internal_standards(MassSpecResults_NonTargetAnalysis)}: Retrieves internal standards from the database.

\item \code{get_fold_change(MassSpecResults_NonTargetAnalysis)}: Gets a data.table with fold-change analysis between the
\code{replicatesIn} and \code{replicatesOut}. This method is adapted from the work of
\href{https://pubs.acs.org/doi/10.1021/acs.analchem.7b03037}{Bader et al. (2017)}.

\item \code{plot_fold_change(MassSpecResults_NonTargetAnalysis)}: Plots the fold-change analysis between the \code{replicatesIn}
and \code{replicatesOut}. This method is adapted from the work of
\href{https://pubs.acs.org/doi/10.1021/acs.analchem.7b03037}{Bader et al. (2017)}.

\item \code{get_transformation_products(MassSpecResults_NonTargetAnalysis)}: Retrieves transformation products from the TransformationProducts table in the database.

\item \code{plot_transformation_products(MassSpecResults_NonTargetAnalysis)}: Plot a transformation products network.

}}
\references{
\insertRef{bader01}{StreamFind}
}
