% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecResults_NonTargetAnalysisDB.R
\name{MassSpecResults_NonTargetAnalysisDB}
\alias{MassSpecResults_NonTargetAnalysisDB}
\alias{validate_object.MassSpecResults_NonTargetAnalysisDB}
\alias{show.MassSpecResults_NonTargetAnalysisDB}
\title{Constructor and methods to handle non-target analysis results for mass spectrometry data}
\usage{
MassSpecResults_NonTargetAnalysisDB(
  db = file.path("data.sf", "MassSpecResults_NonTargetAnalysis.duckdb"),
  analyses = data.table::data.table(),
  headers = data.table::data.table(),
  features = data.table::data.table()
)

\method{validate_object}{MassSpecResults_NonTargetAnalysisDB}(x)

\method{show}{MassSpecResults_NonTargetAnalysisDB}(x)
}
\arguments{
\item{headers}{A list of data frames containing information about the spectra headers.}

\item{features}{A list of data frames containing information about the features detected by \code{rcpp_nts_find_features2()}.}

\item{x}{A \code{MassSpecResults_NonTargetAnalysisDB} object.}

\item{info}{A data frame containing information about the analyses.}
}
\value{
An object of class \code{MassSpecResults_NonTargetAnalysis2} with the following structure:
\itemize{
\item \code{type}: The type of the results, which is "MassSpec".
\item \code{name}: The name of the results, which is "MassSpecResults_NonTargetAnalysis2".
\item \code{software}: The software used for the analysis, which is "StreamFind".
\item \code{version}: The version of the software, as a character string.
\item \code{info}: A data frame containing information about the analyses.
\item \code{headers}: A list of data frames containing information about the spectra headers.
\item \code{features}: A list of data frames containing information about the features.
}
The \code{info} data.table contains the following columns: analysis, replicate, blank, polarity and file. Each \code{features} data frame contains the following columns from \code{rcpp_nts_find_features2()}:
\itemize{
\item Core feature properties: feature, group, component, adduct, rt, mz, mass, intensity, noise, sn, area
\item Chromatographic boundaries: rtmin, rtmax, width, mzmin, mzmax, ppm
\item Peak shape characterization: fwhm_rt, fwhm_mz
\item Gaussian fitting parameters: gaussian_A, gaussian_mu, gaussian_sigma, gaussian_r2
\item Processing flags: polarity, filtered, filter, filled, correction
\item Encoded profile data: eic_size, eic_rt, eic_mz, eic_intensity, eic_baseline, eic_smoothed
\item MS spectral data: ms1_size, ms1_mz, ms1_intensity, ms2_size, ms2_mz, ms2_intensity
}
}
\description{
The \code{MassSpecResults_NonTargetAnalysis2} class is a child of the \link{Results} class and is used to store results from non-target analysis (NTA) workflows for mass spectrometry data ("MassSpec"). It is specifically designed to handle the output from \code{rcpp_nts_find_features2()}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate_object(MassSpecResults_NonTargetAnalysisDB)}: Validates the MassSpecResults_NonTargetAnalysisDB object, returning NULL if valid.

\item \code{show(MassSpecResults_NonTargetAnalysisDB)}: Prints a summary of the MassSpecResults_NonTargetAnalysisDB object.

}}
