% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecResults_Spectra.R
\name{MassSpecResults_Spectra}
\alias{MassSpecResults_Spectra}
\alias{validate_object.MassSpecResults_Spectra}
\alias{query_db.MassSpecResults_Spectra}
\alias{list_db_tables.MassSpecResults_Spectra}
\alias{get_db_table_info.MassSpecResults_Spectra}
\alias{show.MassSpecResults_Spectra}
\alias{get_spectra.MassSpecResults_Spectra}
\alias{get_spectra_charges.MassSpecResults_Spectra}
\alias{get_mass_spectra.MassSpecResults_Spectra}
\alias{get_spectra_peaks.MassSpecResults_Spectra}
\title{Constructor and methods to handle Mass Spectrometry spectra results using DuckDB}
\usage{
MassSpecResults_Spectra(
  projectPath = ".",
  analyses = data.table::data.table(),
  spectra = data.table::data.table(),
  charges = data.table::data.table(),
  mass_spectra = data.table::data.table(),
  spectra_peaks = data.table::data.table()
)

\method{validate_object}{MassSpecResults_Spectra}(x)

\method{query_db}{MassSpecResults_Spectra}(x, sql, params = NULL)

\method{list_db_tables}{MassSpecResults_Spectra}(x)

\method{get_db_table_info}{MassSpecResults_Spectra}(x, tableName)

\method{show}{MassSpecResults_Spectra}(x)

\method{get_spectra}{MassSpecResults_Spectra}(x, analyses = NULL)

get_spectra_charges.MassSpecResults_Spectra(x, analyses = NULL)

get_mass_spectra.MassSpecResults_Spectra(x, analyses = NULL)

\method{get_spectra_peaks}{MassSpecResults_Spectra}(x, analyses = NULL)
}
\arguments{
\item{projectPath}{Path to the project directory where this database file will be created/used.}

\item{analyses}{Character or numeric vector with names or indexes of analyses in the \code{Analyses}
object.}

\item{spectra}{A data.table with raw spectra data.}

\item{charges}{A data.table with charge-state assignments (output of CalculateSpectraCharges).}

\item{mass_spectra}{A data.table with deconvoluted mass spectra.}

\item{spectra_peaks}{A data.table with spectral peak maxima (output of FindSpectraMaxima).}

\item{x}{An object of class \code{MassSpecResults_Spectra}.}

\item{sql}{SQL query string to execute against the database.}

\item{params}{Query parameters to be used with the SQL query (optional).}

\item{tableName}{Name of the table to query or retrieve information about.}
}
\value{
An object of class \code{MassSpecResults_Spectra}.
}
\description{
The \code{MassSpecResults_Spectra} class stores raw spectra, charge-state
assignments, deconvoluted mass spectra, and spectral peak maxima in a DuckDB backend.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate_object(MassSpecResults_Spectra)}: Validate the MassSpecResults_Spectra object.

\item \code{query_db(MassSpecResults_Spectra)}: Execute a query on the DB.

\item \code{list_db_tables(MassSpecResults_Spectra)}: List tables in the DB.

\item \code{get_db_table_info(MassSpecResults_Spectra)}: Get table info from the DB.

\item \code{show(MassSpecResults_Spectra)}: Print summary.

\item \code{get_spectra(MassSpecResults_Spectra)}: Get raw spectra data.

\item \code{get_spectra_peaks(MassSpecResults_Spectra)}: Get spectral peak maxima.

}}
\section{Functions}{
\itemize{
\item \code{get_spectra_charges.MassSpecResults_Spectra()}: Get charge-state assignments.

\item \code{get_mass_spectra.MassSpecResults_Spectra()}: Get deconvoluted mass spectra.

}}
