% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecResults_TransformationProducts.R
\name{MassSpecResults_TransformationProducts}
\alias{MassSpecResults_TransformationProducts}
\alias{validate_object.MassSpecResults_TransformationProducts}
\alias{show.MassSpecResults_TransformationProducts}
\alias{[.MassSpecResults_TransformationProducts}
\title{Constructor and methods to store transformation products results from non-target analysis}
\usage{
MassSpecResults_TransformationProducts(
  parents = data.table::data.table(),
  transformation_products = list()
)

\method{validate_object}{MassSpecResults_TransformationProducts}(x)

\method{show}{MassSpecResults_TransformationProducts}(x)

\method{[}{MassSpecResults_TransformationProducts}(x, i)
}
\arguments{
\item{parents}{A data.table containing parent compound information.}

\item{transformation_products}{A list containing transformation products for each parent compound.}

\item{x}{A \code{MassSpecResults_TransformationProducts} object.}
}
\description{
The \code{MassSpecResults_TransformationProducts} class is a child of \link{Results} and is used to store transformation products results from non-target analysis.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate_object(MassSpecResults_TransformationProducts)}: Validate the MassSpecResults_TransformationProducts object, returning NULL if valid.

\item \code{show(MassSpecResults_TransformationProducts)}: Show the MassSpecResults_TransformationProducts object.

\item \code{[}: Subset the MassSpecResults_TransformationProducts object.

}}
\seealso{
\link{Results}
}
