% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/class_lv3_S7_MassSpecSettings_AnnotateFeatures.R
\name{MassSpecSettings_AnnotateFeatures_StreamFind2}
\alias{MassSpecSettings_AnnotateFeatures_StreamFind2}
\title{\strong{MassSpecSettings_AnnotateFeatures_StreamFind2}}
\usage{
MassSpecSettings_AnnotateFeatures_StreamFind2(
  rtWindowAlignment = 0.3,
  maxIsotopes = 8,
  maxCharge = 1,
  maxGaps = 1
)
}
\arguments{
\item{rtWindowAlignment}{Numeric (length 1) with the proportion of the monoisotopic feature time window to be used
for retrieving isotopic candidates.}

\item{maxIsotopes}{Numeric (length 1) with the maximum number of isotopic steps.}

\item{maxCharge}{Numeric (length 1) with the maximum charge that ions can be ionized to find isotopes.}

\item{maxGaps}{Numeric (length 1) with the maximum of allowed gaps in isotopic chains.}

\item{elements}{Character vector with the elements to target the isotopic annotation.
Possible elements are C, H, N, O, S, Cl, Br.}

\item{mode}{Character (length 1) with the type of molecules to be targeted. For now, only "small molecules" are possible.}
}
\value{
A \code{MassSpecSettings_AnnotateFeatures_StreamFind2} object.
}
\description{
Settings for annotation of isotopic features. The method uses the \code{maxIsotopes} to define the maximum
length of the isotopic chain. The list of candidate features is build with the \code{rtWindowAlignment} and the maximum
mass increment to match the maximum chain length. Then, the mass difference  of the natural isotopes defined by
\code{elements} and a given monoisotopic ion (i.e., feature) are targeted. Each candidate is then evaluated according to
the mass error and the expected relative intensity range as defined by the \code{mode}.
}
