% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecSettings_AverageSpectra.R
\name{MassSpecSettings_AverageSpectra_StreamFind}
\alias{MassSpecSettings_AverageSpectra_StreamFind}
\title{\strong{MassSpecSettings_AverageSpectra_StreamFind}}
\usage{
MassSpecSettings_AverageSpectra_StreamFind(collapseTime = FALSE)
}
\arguments{
\item{collapseTime}{Logical (length 1). When \code{TRUE} the spectra are averaged, reducing the time variable.}
}
\value{
A MassSpecSettings_AverageSpectra_StreamFind object.
}
\description{
Averages spectra based on analysis replicate groups.
}
