% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/class_lv3_S7_MassSpecSettings_CalculateFeaturesQuality.R
\name{MassSpecSettings_CalculateFeaturesQuality_StreamFind}
\alias{MassSpecSettings_CalculateFeaturesQuality_StreamFind}
\title{\strong{MassSpecSettings_CalculateFeaturesQuality_StreamFind}}
\usage{
MassSpecSettings_CalculateFeaturesQuality_StreamFind(
  rtExpand = 120,
  mzExpand = 3e-04,
  minTraces = 6,
  filtered = FALSE
)
}
\arguments{
\item{rtExpand}{Numeric (length 1) with the retention time values (in seconds) for expanding left and right the
retention time range.}

\item{mzExpand}{Numeric (length 1) with the mass or mass-to-charge ratio (\emph{m/z}) (in Da) for expanding left and
right the mass range.}

\item{minTraces}{Numeric of length 1 with the minimum number traces for calculating feature quality.}

\item{filtered}{Logical (length 1). When \code{TRUE}, filtered features/groups are considered.}
}
\value{
A \code{MassSpecSettings_CalculateFeaturesQuality_StreamFind} object.
}
\description{
Settings for calculating quality parameters of features (e.g., signal-to-noise (sn) ratio).
}
