% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_S3_MassSpecSettings.R
\name{MassSpecSettings_CalculateQuality_StreamFind}
\alias{MassSpecSettings_CalculateQuality_StreamFind}
\title{MassSpecSettings_CalculateQuality_StreamFind}
\usage{
MassSpecSettings_CalculateQuality_StreamFind(
  rtExpand = 120,
  mzExpand = 3e-04,
  minTraces = 6,
  filtered = FALSE
)
}
\arguments{
\item{rtExpand}{Numeric (length 1) with the retention time values (in seconds) for expanding left and right the
retention time range.}

\item{mzExpand}{Numeric (length 1) with the mass or mass-to-charge ratio (\emph{m/z}) (in Da) for expanding left and
right the mass range.}

\item{minTraces}{Numeric of length 1 with the minimum number traces for calculating feature quality.}

\item{filtered}{Logical (length 1). When \code{TRUE}, filtered features/groups are considered.}
}
\value{
A ProcessingSettings S3 class object with subclass MassSpecSettings_CalculateQuality_StreamFind.
}
\description{
Settings for calculating quality parameters of features (e.g., signal-to-noise (sn) ratio).
}
