% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_S3_MassSpecSettings.R
\name{MassSpecSettings_CentroidSpectra_qCentroids}
\alias{MassSpecSettings_CentroidSpectra_qCentroids}
\title{MassSpecSettings_CentroidSpectra_qCentroids}
\usage{
MassSpecSettings_CentroidSpectra_qCentroids(maxScale = 5, mode = 1)
}
\arguments{
\item{maxScale}{Integer of length one. Maximum scale as integer (default is 5) for defining the scale limit for the peak model.}

\item{mode}{Integer of length one. \code{0} for debugging, \code{1} for silent mode (the default) and \code{2} for progress bar mode.}
}
\value{
A ProcessingSettings S3 class object with subclass MassSpecSettings_CentroidSpectra_qCentroids.
}
\description{
Centroids profile spectra using the \href{https://link.springer.com/article/10.1007/s00216-022-04224-y}{qCentroids}
algorithm, which is part of the \href{https://github.com/odea-project/qAlgorithms}{qAlgorithms} library.
}
\references{
\insertRef{qcentroids01}{StreamFind}
}
