% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_S3_MassSpecSettings.R
\name{MassSpecSettings_ClusterSpectra_StreamFind}
\alias{MassSpecSettings_ClusterSpectra_StreamFind}
\title{MassSpecSettings_ClusterSpectra_StreamFind}
\usage{
MassSpecSettings_ClusterSpectra_StreamFind(
  val = "mz",
  clustVal = 0.001,
  presence = 0.1
)
}
\arguments{
\item{val}{Character (length 1) with the variable to be used for clustering.}

\item{clustVal}{Numeric (length 1) with the clustering value.}

\item{presence}{Numeric (length 1) with the minimum presence of traces in a cluster to be considered.}
}
\value{
A ProcessingSettings S3 class object with subclass MassSpecSettings_ClusterSpectra_StreamFind.
}
\description{
Clusters spectra based on a variable (i.e. column name).
}
