% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/class_MassSpecSettings_CorrectChromatogramsBaseline.R
\name{MassSpecSettings_CorrectChromatogramsBaseline_baseline}
\alias{MassSpecSettings_CorrectChromatogramsBaseline_baseline}
\title{\strong{MassSpecSettings_CorrectChromatogramsBaseline_baseline}}
\usage{
MassSpecSettings_CorrectChromatogramsBaseline_baseline(
  method = "als",
  args = list(lambda = 5, p = 0.05, maxit = 10)
)
}
\arguments{
\item{method}{Character (length 1) with the method to be used for baseline correction. Possible values are "als",
"fillPeaks", "irls", "lowpass", "medianWindow", "modpolyfit", "peakDetection", "rfbaseline", "rollingBall", "shirley"
and "TAP".}

\item{args}{List with additional arguments for the selected method. See the documentation of the \pkg{baseline} package
for more details.}
}
\value{
A MassSpecSettings_CorrectChromatogramsBaseline_baseline object.
}
\description{
Performs baseline correction to chromatograms using the \pkg{baseline} package.
}
