% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/class_MassSpecSettings_CorrectMatrixSuppression.R
\name{MassSpecSettings_CorrectMatrixSuppression_TiChri}
\alias{MassSpecSettings_CorrectMatrixSuppression_TiChri}
\title{\strong{MassSpecSettings_CorrectMatrixSuppression_TiChri}}
\usage{
MassSpecSettings_CorrectMatrixSuppression_TiChri(
  mpRtWindow = 10,
  istdAssignment = "none",
  istdRtWindow = 5,
  istdN = 2
)
}
\arguments{
\item{mpRtWindow}{Numeric of length one with the retention time window (in seconds) for calculating the matrix profile.}

\item{istdAssignment}{Character of length one with the assignment method for internal standards. Possible values are
\code{"nearest"}, \code{"range"}, and \code{"none"}. Default is \code{"nearest"}. Setting \code{"nearest"} assigns the nearest \code{istdN} internal
standard/s, \code{"range"} assigns internal standard/s within the \code{istdRtWindow} window, and \code{"none"} does not assign internal
standards. If internal standards are assigned, the \code{tichri} value is calculated for each internal standard and used
to correct the matrix suppression for the features. If no internal standards are assigned, the correction is based
only on the TIC matrix profile, which is less accurate.}

\item{istdRtWindow}{Numeric of length one with the retention time window (in seconds) for assigning internal standards.
Default is \code{5}.}

\item{istdN}{Integer of length one with the number of internal standards to assign. Default is \code{2}.}
}
\value{
A \code{MassSpecSettings_CorrectMatrixSuppression_TiChri} object.
}
\description{
Settings for correcting matrix suppression based on the TiChri algorithm from
\href{https://pubs.acs.org/doi/10.1021/acs.analchem.1c00357}{Tisler et al. (2021)}. The algorithm calculates the
matrix profile for the total ion chromatogram (TIC) and corrects the matrix suppression for features. Internal
standards can be assigned to improve the correction. The \code{suppression_factor} is added to the feature list and can
be used to correct the features intensity. The argument/parameter \code{correctSuppression} is available in plotting
and processing methods and when \code{TRUE}, the suppression factor is used to correct the feature intensity for better
comparison across analyses with different matrix suppression.
}
\references{
\insertRef{tisler01}{StreamFind}
}
