% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_lv3_S7_MassSpecSettings_FillFeatures.R
\name{MassSpecSettings_FillFeatures_StreamFind}
\alias{MassSpecSettings_FillFeatures_StreamFind}
\title{\strong{MassSpecSettings_FillFeatures_StreamFind}}
\usage{
MassSpecSettings_FillFeatures_StreamFind(rtExpand = 1, mzExpand = 5e-04)
}
\arguments{
\item{rtExpand}{Numeric (length 1) with the retention time values (in seconds) for expanding left and right the
retention time range.}

\item{mzExpand}{Numeric (length 1) with the mass or mass-to-charge ratio (\emph{m/z}) (in Da) for expanding left and
right the mass range.}
}
\value{
A ProcessingSettings S3 class object with subclass MassSpecSettings_FillFeatures_StreamFind.
}
\description{
Settings for filling missing values in features.
}
