% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_S3_MassSpecSettings.R
\name{MassSpecSettings_FilterFeatures_StreamFind}
\alias{MassSpecSettings_FilterFeatures_StreamFind}
\title{MassSpecSettings_FilterFeatures_StreamFind}
\usage{
MassSpecSettings_FilterFeatures_StreamFind(
  minSnRatio = 3,
  excludeIsotopes = TRUE
)
}
\arguments{
\item{minSnRatio}{Numeric (length 1) with the minimum signal-to-noise ratio.}

\item{excludeIsotopes}{Logical (length 1) with \code{TRUE} for filtering annotated isotopes (only prevails the monoisotopic features).}
}
\value{
A ProcessingSettings S3 class object with subclass MassSpecSettings_FilterFeatures_StreamFind.
}
\description{
Settings for filtering of features and feature groups.
}
