% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecSettings_FindChromPeaks.R
\name{MassSpecSettings_FindChromPeaks_LocalMaxima}
\alias{MassSpecSettings_FindChromPeaks_LocalMaxima}
\title{\strong{MassSpecSettings_FindChromPeaks_LocalMaxima}}
\usage{
MassSpecSettings_FindChromPeaks_LocalMaxima(
  minWidth = 0,
  maxWidth = 0,
  minHeight = 0
)
}
\arguments{
\item{minWidth}{Numeric (length 1) with the minimum width of a peak.}

\item{maxWidth}{Numeric (length 1) with the maximum width of a peak.}

\item{minHeight}{Numeric (length 1) with the minimum height of a peak.}
}
\value{
A MassSpecSettings_FindChromPeaks_LocalMaxima object.
}
\description{
Finds peak maxima in chromatograms.
}
