% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecSettings_FindFeatures.R
\name{MassSpecSettings_FindFeatures_qalgorithms}
\alias{MassSpecSettings_FindFeatures_qalgorithms}
\title{\strong{MassSpecSettings_FindFeatures_qalgorithms}}
\usage{
MassSpecSettings_FindFeatures_qalgorithms(ppm = 5)
}
\arguments{
\item{ppm}{numeric(1) defining the maximal tolerated m/z deviation in parts per million (ppm)
only applicable for centroid data. For profile data, the ppm value is ignored.}
}
\description{
The qAlgorithms uses a comprehensive peak model developed by
\href{https://doi.org/10.1021/acs.analchem.4c00494}{Renner et al.} to
identify peaks within LC-MS data. More information can be found in the
GitHub repository of the \href{https://github.com/odea-project/qAlgorithms}{qAlgorithms} project.
The qAlgorithms is best used with profile data but centroid data is also possible. Yet, a mass
uncertainty should by supplied in parts per million (ppm) to account for the m/z deviation in
centroid data.
}
