% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_S3_MassSpecSettings.R
\name{MassSpecSettings_FindFeatures_qPeaks}
\alias{MassSpecSettings_FindFeatures_qPeaks}
\title{MassSpecSettings_FindFeatures_qPeaks}
\usage{
MassSpecSettings_FindFeatures_qPeaks()
}
\description{
Not yet implemented.
}
