% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_S3_MassSpecSettings.R
\name{MassSpecSettings_FindInternalStandards_StreamFind}
\alias{MassSpecSettings_FindInternalStandards_StreamFind}
\title{MassSpecSettings_FindInternalStandards_StreamFind}
\usage{
MassSpecSettings_FindInternalStandards_StreamFind(
  database = NULL,
  ppm = 5,
  sec = 10
)
}
\arguments{
\item{database}{A data.frame with at least the columns name, mass, and rt indicating the name, neutral monoisotopic
mass and retention time of the internal standards, respectively.}

\item{ppm}{Numeric of length one with the mass deviation, in ppm.}

\item{sec}{Numeric of length one with the time deviation, in seconds.}
}
\value{
A ProcessingSettings S3 class object with subclass MassSpecSettings_FindInternalStandards_StreamFind.
}
\description{
Settings for finding internal standards using a data.frame.
}
