% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecSettings_FindSpectraMaxima.R
\name{MassSpecSettings_FindSpectraMaxima_StreamFind}
\alias{MassSpecSettings_FindSpectraMaxima_StreamFind}
\title{\strong{MassSpecSettings_FindSpectraMaxima_StreamFind}}
\usage{
MassSpecSettings_FindSpectraMaxima_StreamFind(
  minWidth = 0,
  maxWidth = 0,
  minHeight = 0
)
}
\arguments{
\item{minWidth}{Numeric (length 1) with the minimum width of a peak.}

\item{maxWidth}{Numeric (length 1) with the maximum width of a peak.}

\item{minHeight}{Numeric (length 1) with the minimum height of a peak.}
}
\value{
A MassSpecSettings_FindSpectraMaxima_StreamFind object.
}
\description{
Finds maximum peaks in continuous spectra.
}
