% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/class_lv3_S7_MassSpecSettings_LoadFeaturesMS1_MS2_EIC.R
\name{MassSpecSettings_LoadFeaturesMS1_StreamFind}
\alias{MassSpecSettings_LoadFeaturesMS1_StreamFind}
\title{\strong{MassSpecSettings_LoadFeaturesMS1_StreamFind}}
\usage{
MassSpecSettings_LoadFeaturesMS1_StreamFind(
  rtWindow = c(-2, 2),
  mzWindow = c(-1, 6),
  mzClust = 0.005,
  presence = 0.8,
  minIntensity = 250,
  filtered = FALSE
)
}
\arguments{
\item{rtWindow}{Numeric (length 2) with the retention time values (in seconds) for expanding left and right the
retention time range. The first element expands left and the second expands right. Note that the first element
should be negative to expand to the left.}

\item{mzWindow}{Numeric (length 2) with the mass or mass-to-charge ratio (\emph{m/z}) values (in Da) for expanding
left and right the mass range. The first element expands left and the second expands right. Note that the first
element should be negative to expand to the left.}

\item{mzClust}{Numeric (length 1) with the mass deviation threshold (in Da) to cluster mass traces. Highly dependent
on the mass resolution of the MS data.}

\item{presence}{Numeric (length 1) with the required presence ratio from 0 (i.e., doesn't need to be present in any
spectra) to 1 (i.e., must be present in all spectra) for traces during clustering of spectra.}

\item{minIntensity}{Numeric (length 1) with the minimum intensity.}

\item{filtered}{Logical (length 1). When \code{TRUE}, filtered features/groups are considered.}
}
\value{
A \code{MassSpecSettings_LoadFeaturesMS1_StreamFind} object.
}
\description{
Settings for loading MS1 spectra for features.
}
