% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_S3_MassSpecSettings.R
\name{MassSpecSettings_LoadFeaturesMS2_StreamFind}
\alias{MassSpecSettings_LoadFeaturesMS2_StreamFind}
\title{MassSpecSettings_LoadFeaturesMS2_StreamFind}
\usage{
MassSpecSettings_LoadFeaturesMS2_StreamFind(
  isolationWindow = 1.3,
  mzClust = 0.005,
  presence = 0.8,
  minIntensity = 10,
  filtered = FALSE
)
}
\arguments{
\item{isolationWindow}{Numeric value with the isolation window (in Da)
applied for ion isolation and fragmentation during acquisition of tandem
data (i.e., MS2 data).}

\item{mzClust}{Numeric (length 1) with the mass deviation threshold (in Da) to cluster mass traces. Highly dependent
on the mass resolution of the MS data.}

\item{presence}{Numeric (length 1) with the required presence ratio from 0 (i.e., doesn't need to be present in any
spectra) to 1 (i.e., must be present in all spectra) for traces during clustering of spectra.}

\item{minIntensity}{Numeric (length 1) with the minimum intensity.}

\item{filtered}{Logical (length 1). When \code{TRUE}, filtered features/groups are considered.}
}
\value{
A ProcessingSettings S3 class object with subclass MassSpecSettings_LoadFeaturesMS2_StreamFind.
}
\description{
Settings for loading MS2 spectra for features.
}
