% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_lv3_S7_MassSpecSettings_LoadSpectra.R
\name{MassSpecSettings_LoadSpectra}
\alias{MassSpecSettings_LoadSpectra}
\title{\strong{MassSpecSettings_LoadSpectra}}
\usage{
MassSpecSettings_LoadSpectra(
  levels = NULL,
  mzmin = NULL,
  mzmax = NULL,
  rtmin = NULL,
  rtmax = NULL,
  mobilitymin = NULL,
  mobilitymax = NULL,
  minIntensity = 0
)
}
\arguments{
\item{levels}{A numeric vector with the levels to be used.}

\item{mzmin}{A numeric vector with the minimum m/z values to be used.}

\item{mzmax}{A numeric vector with the maximum m/z values to be used.}

\item{rtmin}{A numeric vector with the minimum retention time values to be used.}

\item{rtmax}{A numeric vector with the maximum retention time values to be used.}

\item{mobilitymin}{A numeric vector with the minimum mobility values to be used.}

\item{mobilitymax}{A numeric vector with the maximum mobility values to be used.}

\item{minIntensity}{A numeric value with the minimum intensity to be used.}
}
\value{
A \code{MassSpecSettings_LoadSpectra} object.
}
\description{
.
}
