% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecSettings_LoadSpectra.R
\name{MassSpecSettings_LoadSpectra_chrompeaks}
\alias{MassSpecSettings_LoadSpectra_chrompeaks}
\title{\strong{MassSpecSettings_LoadSpectra_chrompeaks}}
\usage{
MassSpecSettings_LoadSpectra_chrompeaks(
  levels = 1,
  mzmin = 0,
  mzmax = 0,
  minIntensity = 0
)
}
\arguments{
\item{levels}{A numeric vector with the levels to be used.}

\item{mzmin}{A numeric vector with the minimum m/z values to be used.}

\item{mzmax}{A numeric vector with the maximum m/z values to be used.}

\item{minIntensity}{A numeric value with the minimum intensity to be used.}
}
\value{
A \code{MassSpecSettings_LoadSpectra_chrompeaks} object.
}
\description{
Loads spectra based on retention time dimensions of chromatographic peaks.
}
