% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_S3_MassSpecSettings.R
\name{MassSpecSettings_NormalizeSpectra_blockweight}
\alias{MassSpecSettings_NormalizeSpectra_blockweight}
\title{MassSpecSettings_NormalizeSpectra_blockweight}
\usage{
MassSpecSettings_NormalizeSpectra_blockweight()
}
\value{
A ProcessingSettings S3 class object with subclass MassSpecSettings_NormalizeSpectra_blockweight.
}
\description{
Normalizes spectra using block weighting for downstream data evaluation.
}
