% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_S3_MassSpecSettings.R
\name{MassSpecSettings_NormalizeSpectra_meancenter}
\alias{MassSpecSettings_NormalizeSpectra_meancenter}
\title{MassSpecSettings_NormalizeSpectra_meancenter}
\usage{
MassSpecSettings_NormalizeSpectra_meancenter()
}
\value{
A ProcessingSettings S3 class object with subclass MassSpecSettings_NormalizeSpectra_meancenter.
}
\description{
Normalizes spectra using mean centering.
}
