% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_S3_MassSpecSettings.R
\name{MassSpecSettings_NormalizeSpectra_minmax}
\alias{MassSpecSettings_NormalizeSpectra_minmax}
\title{MassSpecSettings_NormalizeSpectra_minmax}
\usage{
MassSpecSettings_NormalizeSpectra_minmax()
}
\value{
A ProcessingSettings S3 class object with subclass MassSpecSettings_NormalizeSpectra_minmax.
}
\description{
Normalizes spectra using the min-max algorithm.
}
