% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_S3_MassSpecSettings.R
\name{MassSpecSettings_NormalizeSpectra_scale}
\alias{MassSpecSettings_NormalizeSpectra_scale}
\title{MassSpecSettings_NormalizeSpectra_scale}
\usage{
MassSpecSettings_NormalizeSpectra_scale()
}
\value{
A ProcessingSettings S3 class object with subclass MassSpecSettings_NormalizeSpectra_scale.
}
\description{
Normalizes spectra using scaling based on the standard deviation.
}
