% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_S3_MassSpecSettings.R
\name{MassSpecSettings_NormalizeSpectra_snv}
\alias{MassSpecSettings_NormalizeSpectra_snv}
\title{MassSpecSettings_NormalizeSpectra_snv}
\usage{
MassSpecSettings_NormalizeSpectra_snv(liftTozero = FALSE)
}
\arguments{
\item{liftTozero}{Logical (length 1) indicating if the spectra should be lifted to zero.}
}
\value{
A ProcessingSettings S3 class object with subclass MassSpecSettings_NormalizeSpectra_snv.
}
\description{
Normalizes spectra using the Standard Normal Variate (SNV) algorithm.
}
