% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_S3_MassSpecSettings.R
\name{MassSpecSettings_SmoothSpectra_movingaverage}
\alias{MassSpecSettings_SmoothSpectra_movingaverage}
\title{MassSpecSettings_SmoothSpectra_movingaverage}
\usage{
MassSpecSettings_SmoothSpectra_movingaverage(windowSize = 5)
}
\arguments{
\item{windowSize}{Numeric (length 1) with the window size for the moving average.}
}
\value{
A ProcessingSettings S3 class object with subclass MassSpecSettings_SmoothSpectra_movingaverage.
}
\description{
Smooths spectra using the moving average algorithm.
}
