% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecSettings_SubtractBlankSpectra.R
\name{MassSpecSettings_SubtractBlankSpectra_StreamFind}
\alias{MassSpecSettings_SubtractBlankSpectra_StreamFind}
\title{\strong{MassSpecSettings_SubtractBlankSpectra_StreamFind}}
\usage{
MassSpecSettings_SubtractBlankSpectra_StreamFind(negativeToZero = FALSE)
}
\arguments{
\item{negativeToZero}{Logical (length 1) indicating if negative values should be set to zero.}
}
\value{
A MassSpecSettings_SubtractBlankSpectra_StreamFind object.
}
\description{
Subtracts the blank spectra to each analysis according to the blank assignment.
}
