% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecSettings_SuspectScreening.R
\name{MassSpecSettings_SuspectScreening_StreamFind}
\alias{MassSpecSettings_SuspectScreening_StreamFind}
\title{\strong{MassSpecSettings_SuspectScreening_StreamFind}}
\usage{
MassSpecSettings_SuspectScreening_StreamFind(
  database = NULL,
  ppm = 5,
  sec = 10,
  ppmMS2 = 10,
  minFragments = 3,
  isolationWindow = 1.3,
  mzClust = 0.003,
  presence = 0.8,
  minIntensity = 0,
  filtered = FALSE
)
}
\arguments{
\item{database}{A data.frame with at least the columns name and mass, indicating the name and neutral monoisotopic
mass of the suspect targets.}

\item{ppm}{Numeric of length one with the mass deviation, in ppm.}

\item{sec}{Numeric of length one with the time deviation, in seconds.}

\item{ppmMS2}{Numeric (length 1) with the mass deviation for MS2, in ppm.}

\item{minFragments}{Integer (length 1) with the minimum number of fragments.}

\item{isolationWindow}{Numeric value with the isolation window (in Da)
applied for ion isolation and fragmentation during acquisition of tandem
data (i.e., MS2 data).}

\item{mzClust}{Numeric (length 1) with the mass deviation threshold (in Da) to cluster mass traces. Highly dependent
on the mass resolution of the MS data.}

\item{presence}{Numeric (length 1) with the required presence ratio from 0 (i.e., doesn't need to be present in any
spectra) to 1 (i.e., must be present in all spectra) for traces during clustering of spectra.}

\item{minIntensity}{Numeric (length 1) with the minimum intensity.}

\item{filtered}{Logical (length 1). When \code{TRUE}, filtered features/groups are considered.}
}
\value{
A \code{MassSpecSettings_SuspectScreening_StreamFind} object.
}
\description{
Settings for performing suspect screening using a data.frame with target compounds.
}
