% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_007_filterFeatures.R
\name{maxReplicateIntensityDeviationFeatures}
\alias{maxReplicateIntensityDeviationFeatures}
\title{maxReplicateIntensityDeviationFeatures}
\usage{
maxReplicateIntensityDeviationFeatures(obj, value = 40)
}
\arguments{
\item{obj}{An \linkS4class{msFeatures} object.}

\item{value}{A numerical value set at the desired sd percentage maximum.}
}
\description{
maxReplicateIntensityDeviationFeatures
}
