% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MethodsMassSpecChromatograms.R
\name{Method_MassSpecChromatograms_FilterChromatogramsRetentionTime}
\alias{Method_MassSpecChromatograms_FilterChromatogramsRetentionTime}
\title{Method_MassSpecChromatograms_FilterChromatogramsRetentionTime}
\usage{
Method_MassSpecChromatograms_FilterChromatogramsRetentionTime(rtmin, rtmax)
}
\arguments{
\item{rtmin}{Numeric(1) minimum retention time to keep (in seconds).}

\item{rtmax}{Numeric(1) maximum retention time to keep (in seconds).}
}
\value{
A \code{Method} object of class \code{Method_MassSpecChromatograms_FilterChromatogramsRetentionTime}.
}
\description{
Create a \code{Method} child object for filtering loaded chromatograms
by retention time. Points outside the specified RT window are removed from
the \code{MS_CHROMATOGRAMS} table.
}
