% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MethodsNonTargetAnalysis.R
\name{Method_NonTargetAnalysis_BlankSubtraction}
\alias{Method_NonTargetAnalysis_BlankSubtraction}
\title{Method_NonTargetAnalysis_BlankSubtraction}
\usage{
Method_NonTargetAnalysis_BlankSubtraction(
  blankThreshold = 5,
  rtExpand = 10,
  mzExpand = 0.005
)
}
\arguments{
\item{blankThreshold}{Numeric(1) threshold multiplier for blank feature
intensities.}

\item{rtExpand}{Numeric(1) retention-time expansion window in seconds.}

\item{mzExpand}{Numeric(1) m/z expansion window in Da.}
}
\value{
A \code{Method} object of class \code{Method_NonTargetAnalysis_BlankSubtraction}.
}
\description{
Create a \code{Method} child object for subtracting blank-derived
features.
}
