% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MethodsNonTargetAnalysis.R
\name{Method_NonTargetAnalysis_FilterFeaturesMS2}
\alias{Method_NonTargetAnalysis_FilterFeaturesMS2}
\title{Method_NonTargetAnalysis_FilterFeaturesMS2}
\usage{
Method_NonTargetAnalysis_FilterFeaturesMS2(
  top = 0L,
  minIntensity = NA_real_,
  relMinIntensity = NA_real_,
  blankClean = FALSE,
  mzClust = 0.005,
  blankPresenceThreshold = 0.8,
  globalPresenceThreshold = 0.1
)
}
\arguments{
\item{top}{Integer(1) number of top MS2 spectra to keep per feature. Use \code{0}
to disable.}

\item{minIntensity, relMinIntensity}{Optional numeric scalar thresholds. Use
\code{NA_real_} to disable.}

\item{blankClean}{Logical(1) whether to remove fragments prevalent in
blanks.}

\item{mzClust}{Numeric(1) m/z clustering tolerance.}

\item{blankPresenceThreshold}{Numeric(1) blank presence threshold.}

\item{globalPresenceThreshold}{Numeric(1) global presence threshold.}
}
\value{
A \code{Method} object of class \code{Method_NonTargetAnalysis_FilterFeaturesMS2}.
}
\description{
Create a \code{Method} child object for filtering MS2 traces loaded
onto features.
}
