% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MethodsNonTargetAnalysis.R
\name{Method_NonTargetAnalysis_FilterInternalStandards}
\alias{Method_NonTargetAnalysis_FilterInternalStandards}
\title{Method_NonTargetAnalysis_FilterInternalStandards}
\usage{
Method_NonTargetAnalysis_FilterInternalStandards(
  names = character(),
  minScore = NA_real_,
  maxErrorRT = NA_real_,
  maxErrorMass = NA_real_,
  idLevels = integer(),
  minSharedFragments = 0L,
  minCosineSimilarity = NA_real_
)
}
\arguments{
\item{names}{Optional character vector of internal-standard names to match.}

\item{minScore, maxErrorRT, maxErrorMass, minCosineSimilarity}{Optional numeric
scalar thresholds. Use \code{NA_real_} to disable a threshold.}

\item{idLevels}{Optional integer vector of identification levels to keep.}

\item{minSharedFragments}{Integer(1) minimum number of shared fragments.}
}
\value{
A \code{Method} object of class
\code{Method_NonTargetAnalysis_FilterInternalStandards}.
}
\description{
Create a \code{Method} child object for filtering internal-standard
hits.
}
