% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MethodsNonTargetAnalysis.R
\name{Method_NonTargetAnalysis_LoadFeaturesMS1}
\alias{Method_NonTargetAnalysis_LoadFeaturesMS1}
\title{Method_NonTargetAnalysis_LoadFeaturesMS1}
\usage{
Method_NonTargetAnalysis_LoadFeaturesMS1(
  filtered = FALSE,
  rtWindow = c(-2, 2),
  mzWindow = c(-1, 6),
  minTracesIntensity = 250,
  mzClust = 0.005,
  presence = 0.8
)
}
\arguments{
\item{filtered}{Logical(1) whether to include features already marked as
filtered.}

\item{rtWindow}{Numeric length-2 vector of retention-time offsets in
seconds.}

\item{mzWindow}{Numeric length-2 vector of m/z offsets in Da.}

\item{minTracesIntensity}{Numeric(1) minimum trace intensity to extract.}

\item{mzClust}{Numeric(1) m/z tolerance used when clustering traces.}

\item{presence}{Numeric(1) minimum fraction of scans required to keep a
cluster.}
}
\value{
A \code{Method} object of class \code{Method_NonTargetAnalysis_LoadFeaturesMS1}.
}
\description{
Create a \code{Method} child object for loading MS1 traces into NTA
features.
}
