% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ms_filters_features.R
\name{minIntensity}
\alias{minIntensity}
\title{minIntensity}
\usage{
minIntensity(obj, value = 5000)
}
\arguments{
\item{obj}{An \linkS4class{ntsData} object.}

\item{value}{A numerical value with the desired minimum intensity for features.}
}
\description{
minIntensity
}
