% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_007_filterFeatures.R
\name{minIntensityFeatures}
\alias{minIntensityFeatures}
\title{minIntensityFeatures}
\usage{
minIntensityFeatures(obj, value = 5000)
}
\arguments{
\item{obj}{An \linkS4class{msFeatures} object.}

\item{value}{A numerical value with the desired minimum intensity for features.}
}
\description{
minIntensityFeatures
}
