% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_007_filterFeatures.R
\name{minReplicateAbundanceFeatures}
\alias{minReplicateAbundanceFeatures}
\title{minReplicateAbundanceFeatures}
\usage{
minReplicateAbundanceFeatures(obj, value = 3)
}
\arguments{
\item{obj}{An \linkS4class{msFeatures} object.}

\item{value}{A numerical value set at the desired minimum representation in replicates.}
}
\description{
minReplicateAbundanceFeatures
}
