% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_000_generics.R
\name{MS2s}
\alias{MS2s}
\title{MS2s}
\usage{
MS2s(object, ...)
}
\arguments{
\item{object}{A method specific class object.}

\item{...}{Other method specific arguments.}
}
\description{
Generic for getting MS 2 spectra.
}
