% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_S3_msAnalysis.R
\name{msAnalysis}
\alias{msAnalysis}
\alias{as.msAnalysis}
\title{msAnalysis S3 Class Constructor}
\usage{
msAnalysis(
  name = NA_character_,
  replicate = NA_character_,
  blank = NA_character_,
  file = NA_character_,
  type = NA_character_,
  instrument = list(),
  time_stamp = NA_character_,
  spectra_number = NA_integer_,
  spectra_mode = NA_character_,
  spectra_levels = NA_integer_,
  mz_low = NA_real_,
  mz_high = NA_real_,
  rt_start = NA_real_,
  rt_end = NA_real_,
  polarity = NA_character_,
  chromatograms_number = NA_integer_,
  ion_mobility = NA,
  tic = data.table(),
  bpc = data.table(),
  spectra = data.table(),
  chromatograms = data.table(),
  features = data.table(),
  metadata = list()
)

as.msAnalysis(analysis)
}
\arguments{
\item{name}{X.}

\item{replicate}{X.}

\item{blank}{X.}

\item{file}{X.}

\item{type}{X.}

\item{instrument}{X.}

\item{time_stamp}{X.}

\item{spectra_number}{X.}

\item{spectra_mode}{X.}

\item{spectra_levels}{X.}

\item{mz_low}{X.}

\item{mz_high}{X.}

\item{rt_start}{X.}

\item{rt_end}{X.}

\item{polarity}{X.}

\item{chromatograms_number}{X.}

\item{ion_mobility}{X.}

\item{tic}{X.}

\item{bpc}{X.}

\item{spectra}{X.}

\item{chromatograms}{X.}

\item{features}{X.}

\item{metadata}{X.}
}
\value{
An msAnalysis S3 class object.
}
\description{
Creates an msAnalysis S3 class object.
}
