% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_class_004_msData.R
\docType{class}
\name{msData-class}
\alias{msData-class}
\alias{show,msData-method}
\alias{getAnalyses,msData-method}
\alias{[,msData,ANY,missing,missing-method}
\alias{polarities,msData-method}
\alias{EICs,msData-method}
\alias{EICs,msData,msData-method}
\alias{plotEICs,msData-method}
\alias{TICs,msData-method}
\alias{plotTICs,msData-method}
\alias{XICs,msData-method}
\alias{plotXICs,msData-method}
\alias{addParameters,msData-method}
\alias{addParameters,msData,msData-method}
\alias{getParameters,msData-method}
\alias{as.features,msData-method}
\alias{as.features,msData,msData-method}
\alias{peaks,msData-method}
\alias{peaks,msData,msData-method}
\alias{plotPeaks,msData-method}
\alias{plotPeaks,msData,msData-method}
\alias{mapPeaks,msData-method}
\alias{features,msData-method}
\alias{plotFeatures,msData-method}
\title{msData}
\usage{
\S4method{show}{msData}(object)

\S4method{getAnalyses}{msData}(object, analyses = NULL)

\S4method{[}{msData,ANY,missing,missing}(x, i, j, ..., drop = TRUE)

\S4method{polarities}{msData}(object)

\S4method{EICs}{msData}(
  object,
  analyses = NULL,
  mz = NULL,
  ppm = 20,
  rt = NULL,
  sec = 60,
  id = NULL
)

\S4method{plotEICs}{msData}(
  object,
  analyses = NULL,
  mz = NULL,
  ppm = 20,
  rt = NULL,
  sec = 30,
  id = NULL,
  colorBy = "targets",
  legendNames = NULL,
  title = NULL,
  interactive = FALSE
)

\S4method{TICs}{msData}(object, analyses = NULL)

\S4method{plotTICs}{msData}(
  object,
  samples = NULL,
  colorBy = "analyses",
  title = NULL,
  interactive = FALSE
)

\S4method{XICs}{msData}(
  object,
  analyses = NULL,
  mz = NULL,
  ppm = 20,
  rt = NULL,
  sec = 60,
  id = NULL
)

\S4method{plotXICs}{msData}(
  object,
  analyses = NULL,
  mz = NULL,
  ppm = 20,
  rt = NULL,
  sec = 60,
  id = NULL,
  legendNames = NULL,
  plotTargetMark = TRUE,
  targetsMark = NULL,
  ppmMark = 5,
  secMark = 10,
  numberRows = 1
)

\S4method{addParameters}{msData}(object, settings, where = "analyses", analyses = NULL)

\S4method{getParameters}{msData}(object, where = "analyses", analyses = NULL, call = NULL)

\S4method{as.features}{msData}(object)

\S4method{peaks}{msData}(
  object,
  analyses = NULL,
  targetsID = NULL,
  mz = NULL,
  ppm = 20,
  rt = NULL,
  sec = 60
)

\S4method{plotPeaks}{msData}(
  object,
  analyses = NULL,
  targetsID = NULL,
  mz = NULL,
  ppm = 20,
  rt = NULL,
  sec = 30,
  colorBy = "targets",
  legendNames = NULL,
  title = NULL,
  interactive = FALSE
)

\S4method{mapPeaks}{msData}(
  object,
  analyses = NULL,
  targetsID = NULL,
  mz = NULL,
  ppm = 20,
  rt = NULL,
  sec = 30,
  colorBy = "targets",
  legendNames = NULL,
  xlim = 30,
  ylim = 0.05,
  title = NULL
)

\S4method{features}{msData}(
  object,
  analyses = NULL,
  targetsID = NULL,
  mz = NULL,
  ppm = 20,
  rt = NULL,
  sec = 60,
  complete = FALSE,
  average = TRUE
)

\S4method{plotFeatures}{msData}(
  object,
  analyses = NULL,
  targetsID = NULL,
  mz = NULL,
  ppm = 20,
  rt = NULL,
  sec = 30,
  colorBy = "targets",
  legendNames = NULL,
  title = NULL,
  interactive = FALSE
)
}
\arguments{
\item{object}{An \linkS4class{msData} object.}

\item{analyses}{A numeric or character vector with the index or analysis name/s.
When \code{NULL}, all the analyses are used.}

\item{i}{A character or numeric vector with the indice/s or name/s of
the analysis to keep while sub-setting.}

\item{mz}{A vector with target \emph{m/z} values or
a two columns data.table or data.frame with minimum and maximum \emph{m/z} values.
Alternatively, \emph{m/z} and retention time values can be given as one data.table/data.frame
and the deviations given as \code{ppm} and \code{sec} are used to calculate the ranges.
The same also works for min and max values of \emph{m/z} and retention time targets.
Note that when mass/time ranges are given, \code{ppm} and \code{sec} are not used.}

\item{ppm}{A numeric vector of length one with the mass deviation, in ppm.}

\item{rt}{A vector with target retention time values or
a two columns data.table/data.frame with minimum and maximum retention time values.}

\item{sec}{A numeric vector of length one with the time deviation, in seconds.}

\item{id}{An id vector with target identifiers. When not given
is added as a combination of the \emph{m/z} and retention time ranges or values.}

\item{colorBy}{A length one character vector applicable to the respective method.}

\item{legendNames}{A character vector with the same length and order
as the number and order of targets to be used as plot legend.}

\item{title}{A character string with the plot tile.}

\item{interactive}{Logical value, set to \code{TRUE} to use
the \pkg{plotly} instead of \pkg{base}. The default is \code{FALSE}.}

\item{plotTargetMark}{Logical, set to \code{TRUE} to plot a target mark.}

\item{targetsMark}{A two columns \link[data.table]{data.table} or \link{data.frame} with
\emph{m/z} and time targets. The column must be named with "mz" and "rt" for
\emph{m/z} and time values, respectively.}

\item{ppmMark}{A numeric vector of length one to define the mass deviation, in ppm,
of the target mark.}

\item{secMark}{A numeric vector of length one to define the time deviation, in seconds,
of the target mark.}

\item{numberRows}{A numeric vector of length one to define
the number of rows to grid the plots.}

\item{where}{A character vector defining where to get the \linkS4class{settings}.}

\item{call}{The call name of the settings to retrieve.}

\item{targetsID}{A character vector with the ID of targets (e.g., peaks or features IDs).}

\item{xlim}{A length one or two numeric vector for setting the \emph{x} limits of a plot.}

\item{ylim}{A length one or two numeric vector for setting the \emph{y} limits of a plot.}

\item{complete}{Logical, set to \code{TRUE} for a complete version of the output.}

\item{settigns}{A \linkS4class{settings} object.}
}
\description{
An S4 class object to store and manage processing of files with MS data.
The \code{msData} object inherits the \linkS4class{streamProject} structure with type defined as
\emph{ms}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: prints the details of an \linkS4class{msData} object.

\item \code{getAnalyses}: getter for analysis names.

\item \code{[}: subset on analysis, using analysis index or name.
Note that filtering of features is lazy to make it fast.
Only average \emph{m/z}, retention time and intensity are updated.
Run updateFeatureTable() with \code{FAST} set to \code{FALSE} for a complete update of
features. Note that the update might invalitate present quality parameters and annotation.

\item \code{polarities}: getter for analyses polarity.

\item \code{EICs}: get extracted ion chromatograms (EICs)
for specified \emph{m/z} (Da) and retention time (seconds) targets
in given analyses. The arguments \code{mz}, \code{ppm}, \code{rt}
and \code{sec} are used to construct the targets.
See ?\link{makeTargets} for more information.

\item \code{plotEICs}: a method for plotting extracted ion chromatograms (EICs)
of data in an \linkS4class{msData} object.
The arguments for data collection are the same as the \link{EICs} method.
A \link[data.table]{date.table} can be used instead.
The \code{colorBy} argument can be \code{"analyses"}, \code{replicates} or \code{targets}
(the default), for coloring by analyses, replicates or EICs targets, respectively.
The \code{legendNames} is a character vector with the same length as targets for plotting and
can be used to legend the plot. Note that, by setting \code{legendNames} the \code{colorBy}
is set to "targets".

\item \code{TICs}: extracts total ion chromatograms (TICs)
for analyses in an \linkS4class{msData} object.

\item \code{plotTICs}: plots a total ion chromatogram (TIC) for abalyses in the object.
The colorBy argument can be "analyses" or "replicates"
to color the plot by analyses or by analysis replicates.

\item \code{XICs}: get three dimensional (\emph{m/z}, time and intensity)
extracted ion chromatograms (XICs) for specified \emph{m/z} and retention time pair targets
in analyses of an \linkS4class{msData} object. The arguments \code{mz}, \code{ppm}, \code{rt}, \code{sec} and \code{id}
are used to construct the targets. See ?\link{makeTargets} for more information.

\item \code{plotXICs}: plots three dimensional (\emph{m/z}, time and intensity)
extracted ion chromatograms (XICs) for specified \emph{m/z} and retention time pair targets
in analyses of an \linkS4class{msData} object. The arguments \code{mz}, \code{ppm}, \code{rt},
\code{sec} and \code{id} are used to construct the targets. See ?\link{makeTargets} for more information.
When \code{plotTargetMark} is \code{TRUE} a target is plotted representing the deviations as defined
by the arguments \code{ppmMark} and \code{secMark} in ppm and seconds, respectively.
When ranges were given to build the XIC, exact \emph{m/z} and time targets can be specified with
the argument \code{targetsMark}. \code{targetsMark} should be a two column table named mz and rt with
exact \emph{m/z} and time targets. Note that the number of rows should be the same as the number of target
in the XIC. The number of rows to plot multiple targets can be defined by the \code{numberRows} argument.

\item \code{addParameters}: adds processing parameters to analyses or features as defined
by the argument \code{where}. So where can be either "analyses" or "features".

\item \code{getParameters}: gets processing parameters from analyses or features as defined
by the argument \code{where}. So where can be either "analyses" or "features".

\item \code{as.features}: converts the \linkS4class{msData}
to a \linkS4class{features} object from the package \pkg{patRoon}.

\item \code{peaks}: getter for chromatographic peaks.
The arguments \code{targetID} and \code{mz}/\code{rt} can be used
to select specific peaks. The \emph{id} of peaks and/or features can be
given in the \code{targetsID} argument to select the respective peaks.
Also, analyses can be selected using the \code{analyses} argument.

\item \code{plotPeaks}: a method for plotting chromatographic peaks
in an \linkS4class{msData} object.
The arguments \code{targetID} and \code{mz}/\code{rt} can be used
to select specific peaks. The \emph{id} of peaks and/or features can be
given in the \code{targetsID} argument to select the respective peaks.
Also, analyses can be selected using the \code{analyses} argument.
The \code{colorBy} argument can be be \code{"analyses"}, \code{replicates} or \code{targets}
(the default), for coloring by analyses, replicates or peak targets, respectively.
The \code{legendNames} is a character vector with the same length as targets for plotting and
can be used to lengend the plot. Note that, by setting \code{legendNames} the \code{colorBy}
is set to "targets" automatically.

\item \code{mapPeaks}: a method for mapping peaks mass and time space.
The \code{colorBy} argument can be be \code{"analyses"}, \code{replicates} or \code{targets}
(the default), for coloring by analyses, replicates or peak targets, respectively.
The \code{legendNames} is a character vector with the same length as targets for plotting and
can be used to legend the plot. Note that, by setting \code{legendNames} the \code{colorBy}
is set to "targets" automatically.

\item \code{features}: getter for features (i.e., grouped peaks). When
complete is set to \code{TRUE}, additional feature metadata is also returned.

\item \code{plotFeatures}: A method for plotting peaks from given features
in an \linkS4class{msData} object.
The \code{colorBy} argument can be be \code{"analyses"}, \code{replicates} or \code{targets}
(the default), for coloring by analyses, replicates or peak targets, respectively.
The \code{legendNames} is a character vector with the same length as targets for plotting and
can be used to lengend the plot. Note that, by setting \code{legendNames} the \code{colorBy}
is set to "targets" automatically.
}}

\section{Slots}{

\describe{
\item{\code{project}}{A character string with the project title.}

\item{\code{date}}{The date of the project.}

\item{\code{path}}{A character string with the project path.}

\item{\code{analyses}}{A list with \linkS4class{msAnalysis} objects for each analysis file added to the project.}

\item{\code{features}}{A \code{list} with features grouped across samples in the project.}
}}

