% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_class_003_msFeatures.R
\docType{class}
\name{msFeatures-class}
\alias{msFeatures-class}
\alias{analysisNames,msFeatures-method}
\alias{analysisNames,msFeatures,msFeatures-method}
\alias{replicateNames,msFeatures-method}
\alias{replicateNames,msFeatures,msFeatures-method}
\alias{blankReplicateNames,msFeatures-method}
\alias{blankReplicateNames,msFeatures,msFeatures-method}
\alias{addSettings,msFeatures-method}
\alias{addSettings,msFeatures,msFeatures-method}
\alias{getSettings,msFeatures-method}
\alias{getSettings,msFeatures,msFeatures-method}
\alias{features,msFeatures-method}
\alias{features,msFeatures,msFeatures-method}
\alias{[,msFeatures,ANY,missing,missing-method}
\alias{[,msFeatures,ANY,ANY,missing-method}
\title{msFeatures-class}
\usage{
\S4method{analysisNames}{msFeatures}(object)

\S4method{replicateNames}{msFeatures}(object)

\S4method{blankReplicateNames}{msFeatures}(object)

\S4method{addSettings}{msFeatures}(object, settings)

\S4method{getSettings}{msFeatures}(object, call = NULL)

\S4method{features}{msFeatures}(
  object,
  targetsID = NULL,
  mass = NULL,
  mz = NULL,
  ppm = 20,
  rt = NULL,
  sec = 60,
  id = NULL,
  filtered = TRUE,
  complete = FALSE,
  average = TRUE
)

\S4method{[}{msFeatures,ANY,missing,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{msFeatures,ANY,ANY,missing}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{object}{An \linkS4class{msFeatures} object.}

\item{settings}{A \linkS4class{settings} object.}

\item{call}{The call name of the settings to retrieve.}

\item{targetsID}{A character vector with the ID of targets
(e.g., peaks or features IDs).}

\item{mz}{A vector with target \emph{m/z} values or a two columns
\linkS4class{data.table} or data.frame with minimum and maximum
\emph{m/z} values. Alternatively, \emph{m/z} and retention time values
can be given as one \linkS4class{data.table}/data.frame and the deviations
given as \code{ppm} and \code{sec} are used to calculate the ranges.
The same also works for min and max values of \emph{m/z} and retention
time targets. Note that when mass/time ranges are given, \code{ppm} and
\code{sec} are not used.}

\item{ppm}{A numeric vector of length one with the mass deviation, in ppm.}

\item{rt}{A vector with target retention time values or
a two columns \linkS4class{data.table}/data.frame with minimum
and maximum retention time values.}

\item{sec}{A numeric vector of length one with the time deviation, in seconds.}

\item{id}{An id vector with target identifiers. When not given
is added as a combination of the \emph{m/z} and retention time ranges or values.}

\item{filtered}{Logical. When \code{TRUE} (the default),
filtered data is returned. When \code{FALSE}, filtered data is excluded.}

\item{complete}{Logical, set to \code{TRUE} for a complete
version of the output.}

\item{average}{Logical, set to \code{TRUE} for returning the intensity of
features averaged for each replicate group.}

\item{x}{An \linkS4class{msFeatures} object.}

\item{i}{The indice/s or name/s of the analyses to keep in \code{x}.}

\item{j}{The indice/s or \emph{id}/s for of features to keep..}

\item{...}{Other arguments.}

\item{drop}{Not applicable to \linkS4class{msFeatures}.}
}
\description{
An S4 class representing an MS sample/file within the \pkg{streamFind} package.
The \code{msFeatures} is used to store and manage MS data and the respective methods can be used
for inspection, processing and evaluation.
}
\section{Methods (by generic)}{
\itemize{
\item \code{analysisNames(msFeatures)}: getter for analysis names.

\item \code{replicateNames(msFeatures)}: getter for replicate names.

\item \code{blankReplicateNames(msFeatures)}: getter for blank names.

\item \code{addSettings(msFeatures)}: adds processing settings for features.

\item \code{getSettings(msFeatures)}: gets processing settings.

\item \code{features(msFeatures)}: getter for features (i.e., grouped peaks). When
complete is set to \code{TRUE}, additional feature metadata is also returned.

\item \code{x[i}: subset on analyses, using analysis index or name.

\item \code{x[i}: subset on features, using feature index or name.

}}
\section{Slots}{

\describe{
\item{\code{analyses}}{A \linkS4class{data.table} with the information about the
analyses, analysis file paths, class and polarity and replicates and
associated replicate blanks for correspondence analysis (i.e.,
grouping of peaks accross samples).}

\item{\code{intensity}}{A \linkS4class{data.table} with the a row for each feature
with the intensity of each peak in each analyses of the set.}

\item{\code{metadata}}{A \linkS4class{data.table} with metadata for each feature.}

\item{\code{annotation}}{...}

\item{\code{parameters}}{An ordered list of \linkS4class{settings} for each
processing step.}
}}

\references{
\insertRef{patroon01}{streamFind}

\insertRef{proteo01}{streamFind}

\insertRef{proteo02}{streamFind}

\insertRef{mzr01}{streamFind}

\insertRef{mzr02}{streamFind}

\insertRef{mzr03}{streamFind}

\insertRef{mzr04}{streamFind}
}
