% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ProjectClasses.R
\name{NewProjectMassSpecSpectra}
\alias{NewProjectMassSpecSpectra}
\title{Construct a Mass Spec Spectra Project}
\usage{
NewProjectMassSpecSpectra(
  db,
  project_id,
  file_paths = character(),
  replicates = character(),
  blanks = character()
)
}
\arguments{
\item{db}{Path to the project DuckDB file.}

\item{project_id}{Active project identifier.}

\item{file_paths}{Character vector with Mass Spec file paths.}

\item{replicates}{Optional character vector with replicate names.}

\item{blanks}{Optional character vector with blank names.}
}
\value{
A \code{ProjectMassSpecSpectra} object.
}
\description{
User-facing constructor wrapper for \code{ProjectMassSpecSpectra}.
}
