% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ProjectClasses.R
\name{NewProjectNonTargetAnalysis}
\alias{NewProjectNonTargetAnalysis}
\title{Construct a Non-Target Analysis Project}
\usage{
NewProjectNonTargetAnalysis(db, project_id)
}
\arguments{
\item{db}{Path to the project DuckDB file.}

\item{project_id}{Active project identifier.}
}
\value{
A \code{ProjectNonTargetAnalysis} object.
}
\description{
User-facing constructor wrapper for \code{ProjectNonTargetAnalysis}.
}
