% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_general_000_newStreamProject.R
\name{newStreamProject}
\alias{newStreamProject}
\title{newStreamProject}
\usage{
newStreamProject(
  files = NA_character_,
  path = getwd(),
  title = NA_character_,
  date = Sys.Date(),
  replicates = NULL,
  blanks = NULL,
  makeNewProject = FALSE
)
}
\arguments{
\item{files}{A character vector with the full path of files to create the project.
Alternatively, a \link{data.frame} or \link[data.table]{data.table} with the columns
file, replicate and blank corresponding to the file full path, analysis replicate names and
respective blank analysis replicate name. When a table is given with replicates and blanks,
the arguments \emph{replicates} and  \emph{blanks} are ignored.
For files, only formats \emph{mzML} and \emph{mzXML} are currently supported.
Note that certain processing algorithms have file format limitations.}

\item{path}{A character string with the directory for the project creation.
The default is the working directory as defined by \code{getwd()}.}

\item{title}{A character string with the project title.}

\item{date}{A \code{Date} object to define project date.}

\item{replicates}{Optional character vector with the same length as the number of files
to be added with the name of the analysis replicate group for each file.
The default is \code{NULL} which attempts  to group the analysis using
the name of the files. The replicate names can always be edited using the class method.}

\item{blanks}{Optional character vector with the same length as the number of files
to be added with the name of the associate blank analysis replicate group for each file.
The default is \code{NULL}.}

\item{makeNewProject}{Logical, set to \code{TRUE} to create an R project
in the given \code{path} and open a new R session.}
}
\value{
A subclass of \linkS4class{streamProject} depending on the added
file formats. For instance, an \linkS4class{msData} is returned for
\emph{mzML} and \emph{mzXML} files.
}
\description{
Creates a new \pkg{streamFind} project.
}
\note{
The format of the files added will dictate the subclass of the \linkS4class{streamProject}.
For instance, \emph{.mzML} or \emph{.mzXML} files will lead to the subclass
\linkS4class{msData}, which will contain an \linkS4class{msAnalysis} for each file.
}
