% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Results_NonTargetAnalysisResults.R
\name{NonTargetAnalysisResults}
\alias{NonTargetAnalysisResults}
\title{Results class child for Non-target Analysis Workflows for Mass Spectrometry Data}
\usage{
NonTargetAnalysisResults(
  analyses_info = data.table::data.table(),
  spectra_headers = list(),
  feature_list = list()
)
}
\arguments{
\item{analyses_info}{A data frame containing information about the analyses.}

\item{spectra_headers}{A list of data frames containing information about the spectra headers.}

\item{feature_list}{A list of data frames containing information about the features.}
}
\description{
The \code{NonTargetAnalysisResults} class is a child of the \link{Results} class and is used to store results from non-target analysis (NTA) workflows for mass spectrometry data ("MassSpec"). It contains methods and properties specific to this type of analysis.
}
\section{Slots}{

\describe{
\item{\code{analyses_info}}{(getter/setter) A data frame containing information about the analyses.}

\item{\code{spectra_headers}}{(getter/setter) A list of data frames containing information about the spectra headers.}

\item{\code{feature_list}}{(getter/setter) A list of data frames containing information about the features.}

\item{\code{spectra_polarity}}{(getter) A character vector indicating the polarity of the spectra.}

\item{\code{replicates}}{(getter/setter) A character vector indicating the replicates for each analysis.}

\item{\code{blanks}}{(getter/setter) A character vector indicating the blanks for each analysis.}

\item{\code{concentrations}}{(getter/setter) A numeric vector indicating the concentrations for each analysis.}

\item{\code{number_analyses}}{(getter) An integer indicating the number of analyses.}

\item{\code{number_features}}{(getter) An integer indicating the number of features.}

\item{\code{has_features}}{(getter) A logical indicating if features are present.}

\item{\code{has_filtered_features}}{(getter) A logical indicating if filtered features are present.}

\item{\code{number_filtered_features}}{(getter) An integer indicating the number of filtered features.}

\item{\code{has_groups}}{(getter) A logical indicating if groups are present.}

\item{\code{number_groups}}{(getter) An integer indicating the number of groups.}

\item{\code{group_names}}{(getter) A character vector indicating the names of the groups.}

\item{\code{has_features_ms1}}{(getter) A logical indicating if MS1 features are present.}

\item{\code{has_features_ms2}}{(getter) A logical indicating if MS2 features are present.}

\item{\code{has_features_eic}}{(getter) A logical indicating if EIC features are present.}

\item{\code{has_features_suspects}}{(getter) A logical indicating if suspect features are present.}

\item{\code{info}}{(getter) A data frame containing an overview of the number of features.}
}}

