% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_open_project_classes.R
\name{open_ProjectMassSpecChromatograms}
\alias{open_ProjectMassSpecChromatograms}
\title{Construct a Mass Spec Chromatograms Project}
\usage{
open_ProjectMassSpecChromatograms(
  db,
  project_id,
  file_paths = character(),
  replicates = character(),
  blanks = character()
)
}
\arguments{
\item{db}{Path to the project DuckDB file.}

\item{project_id}{Active project identifier.}

\item{file_paths}{Character vector with Mass Spec file paths.}

\item{replicates}{Optional character vector with replicate names.}

\item{blanks}{Optional character vector with blank names.}
}
\value{
A \code{ProjectMassSpecChromatograms} object.
}
\description{
Open an existing \code{ProjectMassSpecChromatograms} or construct a new one.
This wrapper is named to reflect that a connection to an existing project
may be returned rather than always creating a fresh project.
}
