% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ProjectClasses.R
\name{OpenProjectNonTargetAnalysis}
\alias{OpenProjectNonTargetAnalysis}
\title{Open or Create a Non-Target Analysis Project}
\usage{
OpenProjectNonTargetAnalysis(db, project_id)
}
\arguments{
\item{db}{Path to the project DuckDB file.}

\item{project_id}{Active project identifier.}
}
\value{
A \code{ProjectNonTargetAnalysis} object.
}
\description{
Open an existing \code{ProjectNonTargetAnalysis} or construct a new one.
This wrapper is named to reflect that a connection to an existing project
may be returned rather than always creating a fresh project.
}
