% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_002_parse_data_from_files.R
\name{parse_ms_chromatograms}
\alias{parse_ms_chromatograms}
\title{parse_ms_chromatograms}
\usage{
parse_ms_chromatograms(files = NA_character_, runParallel = FALSE)
}
\arguments{
\item{files}{X.}

\item{runParallel}{X.}
}
\value{
X.
}
\description{
parse_ms_chromatograms
}
