% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_002_parse_data_from_files.R
\name{parse_ms_spectra}
\alias{parse_ms_spectra}
\title{parse_ms_spectra}
\usage{
parse_ms_spectra(
  files = NA_character_,
  levels = c(1, 2),
  targets = NULL,
  allTraces = TRUE,
  isolationWindow = 1.3,
  runParallel = FALSE,
  minIntensityMS1 = 0,
  minIntensityMS2 = 0
)
}
\arguments{
\item{files}{X.}

\item{levels}{X.}

\item{targets}{X.}

\item{allTraces}{X.}

\item{isolationWindow}{X.}

\item{runParallel}{X.}

\item{minIntensityMS1}{X.}

\item{minIntensityMS2}{X.}
}
\value{
A list with a spectra \code{data.frame} for each file in \code{files}.
On error, returns \code{NULL}.
}
\description{
Parses spectra from mzML or mzXML files.
}
