% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_004_peakAnnotation.R
\name{peakAnnotation}
\alias{peakAnnotation}
\title{peakAnnotation}
\usage{
peakAnnotation(object = NULL, settings = NULL)
}
\arguments{
\item{object}{An \linkS4class{msData} object with peaks/features.}

\item{settings}{A \linkS4class{settings} object.}
}
\value{
An \linkS4class{msData} object with annotated peaks/features.
}
\description{
Annotation of isotopic and adduct peaks in features.
}
