% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_003_peakFilling.R
\name{peakFilling}
\alias{peakFilling}
\title{peakFilling}
\usage{
peakFilling(object, settings = NULL)
}
\arguments{
\item{object}{An \linkS4class{msData} object containing features.}

\item{settings}{A \linkS4class{settings} object with parameter
for filling peaks.}
}
\value{
An \linkS4class{msData} object including filled missing peaks
in analyses.
}
\description{
Recursive integration for filling missing peaks
within each feature.
}
\details{
The function \code{\link[xcms]{fillChromPeaks}}
from the \pkg{xcms} package can be used,
giving the respective parameters with the argument \code{settings}.
When the algorithm is set to \emph{xcms3}, the settings are the S4 class
object \linkS4class{FillChromPeaksParam} or \linkS4class{ChromPeakAreaParam}.
See \code{\link[xcms]{fillChromPeaks}} for more information.
}
\references{
\insertRef{xcms01}{streamFind}
\insertRef{xcms02}{streamFind}
\insertRef{xcms03}{streamFind}
}
