% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_002_peakGrouping.R
\name{peakGrouping}
\alias{peakGrouping}
\title{peakGrouping}
\usage{
peakGrouping(object = NULL, algorithm = NA_character_, settings = NULL)
}
\arguments{
\item{object}{A \linkS4class{msData} object with
\linkS4class{msAnalysis} objects containing peaks.}

\item{algorithm}{The algorithm to use for peak alignment and grouping.}

\item{settings}{The respective list of parameter settings.
See \link[patRoon]{groupFeatures} for more information.}
}
\value{
An \linkS4class{msData} object
containing a features S4 class.
}
\description{
Grouping and alignment of peaks across samples.
The peak grouping uses the \pkg{patRoon} package and
the following algorithms are possible: "xcms3", "xcms", "openms".
See ?\pkg{patRoon} for further information.
}
\references{
\insertRef{patroon01}{streamFind}
}
