% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_001_peakPicking.R
\name{peakPicking}
\alias{peakPicking}
\title{peakPicking}
\usage{
peakPicking(object = NULL, settings = NULL)
}
\arguments{
\item{object}{An \linkS4class{msData} or \linkS4class{msAnalysis} object.}

\item{settigns}{A \linkS4class{settings} object.}
}
\value{
An \linkS4class{msAnalysis}/\linkS4class{msData} object
with peaks added.
}
\description{
Finds chromatographic peaks from MS data in an \linkS4class{msAnalysis}
object or in MS analyses a given \linkS4class{msData} object.
The peak picking uses the \pkg{patRoon} package for data processing.
}
\details{
See \link[patRoon]{findFeatures} for more information.
the following algorithms are available:
"xcms3", "xcms", "openms", "envipick", "sirius", "kpic2", "safd".
The parameters depend on the algorithm chosen.
See ?\pkg{patRoon} for further information.
}
\references{
\insertRef{patroon01}{streamFind}
}
\seealso{
\link[patRoon]{findFeatures}
}
