% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot_chromatograms}
\alias{plot_chromatograms}
\title{Generic \code{plot_chromatograms}}
\usage{
plot_chromatograms(x, ...)
}
\arguments{
\item{x}{An object from which the chromatograms will be plotted.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S7 generic to plot chromatograms from \code{x}.
}
