% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot_chromatograms_baseline}
\alias{plot_chromatograms_baseline}
\title{Generic \code{plot_chromatograms_baseline}}
\usage{
plot_chromatograms_baseline(x, ...)
}
\arguments{
\item{x}{An object from which the chromatograms baseline will be plotted.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S7 generic to plot chromatograms baseline from \code{x}.
}
