% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_S3.R
\name{plot_chromatograms_peaks}
\alias{plot_chromatograms_peaks}
\title{Generic \code{plot_chromatograms_peaks}}
\usage{
plot_chromatograms_peaks(x, ...)
}
\arguments{
\item{x}{An object from which the chromatograms peaks will be plotted.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S3 generic to plot chromatograms peaks from \code{x}.
}
