% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot_components}
\alias{plot_components}
\title{Generic \code{plot_components}}
\usage{
plot_components(x, ...)
}
\arguments{
\item{x}{An object from which the components will be plotted.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S7 generic to plot components from \code{x}.
}
