% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot_contributions}
\alias{plot_contributions}
\title{Generic \code{plot_contributions}}
\usage{
plot_contributions(x, ...)
}
\arguments{
\item{x}{An object from which the contributions will be plotted.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S7 generic to plot contributions from \code{x}.
}
