% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_S3.R
\name{plot_data}
\alias{plot_data}
\title{Generic \code{plot_data}}
\usage{
plot_data(x, ...)
}
\arguments{
\item{x}{An object from which the data will be plotted.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S3 generic to plot data from \code{x}.
}
