% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_S3.R
\name{plot_explained_variance}
\alias{plot_explained_variance}
\title{Generic \code{plot_explained_variance}}
\usage{
plot_explained_variance(x, ...)
}
\arguments{
\item{x}{An object from which the explained variance will be plotted.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S3 generic to plot explained variance from \code{x}.
}
