% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot_features}
\alias{plot_features}
\title{Generic \code{plot_features}}
\usage{
plot_features(x, ...)
}
\arguments{
\item{x}{An object from which the features will be plotted.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S7 generic to plot features from \code{x}.
}
