% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_S3.R
\name{plot_features_count}
\alias{plot_features_count}
\title{Generic \code{plot_features_count}}
\usage{
plot_features_count(x, ...)
}
\arguments{
\item{x}{An object from which the features count will be plotted.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S3 generic to plot features count from \code{x}.
}
