% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_S3.R
\name{plot_features_eic}
\alias{plot_features_eic}
\title{Generic \code{plot_features_eic}}
\usage{
plot_features_eic(x, ...)
}
\arguments{
\item{x}{An object from which the features EIC will be plotted.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S3 generic to plot features EIC from \code{x}.
}
