% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ms_plot_peaks.R
\name{plot_features_interactive}
\alias{plot_features_interactive}
\title{plot_features_interactive}
\usage{
plot_features_interactive(eic, fts, title, colorBy)
}
\arguments{
\item{eic}{A data table with the analysis, replicate,
id, rt, intensity and var (i.e., the plotting variables for each peak)
as columns.}

\item{title}{An optional character vector to be used as title.}

\item{colorBy}{Possible values are \code{"targets"} (the default),
\code{"analyses"} or \code{replicates},
for coloring by target peaks, analyses or replicates, respectively.}

\item{peaks}{A data table with the individual peaks to plot.}
}
\value{
A chromatographic peak plot through \pkg{plotly}.
}
\description{
Plots chromatographic peaks with the package \pkg{plotly}.
}
