% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_S3.R
\name{plot_features_ms1}
\alias{plot_features_ms1}
\title{Generic \code{plot_features_ms1}}
\usage{
plot_features_ms1(x, ...)
}
\arguments{
\item{x}{An object from which the features MS1 will be plotted.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S3 generic to plot features MS1 from \code{x}.
}
