% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_class_R6MS.R
\name{plot_group_features_aux}
\alias{plot_group_features_aux}
\title{plot_group_features_aux}
\usage{
plot_group_features_aux(fts, eic, heights, analyses)
}
\arguments{
\item{heights}{A numeric vector of length two to control the height of
the first and second plot, respectively.}
}
\description{
Plots features for each feature group.
}
