% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_S3.R
\name{plot_groups}
\alias{plot_groups}
\title{Generic \code{plot_groups}}
\usage{
plot_groups(x, ...)
}
\arguments{
\item{x}{An object from which the groups will be plotted.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S3 generic to plot groups from \code{x}.
}
