% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot_groups_ms2}
\alias{plot_groups_ms2}
\title{Generic \code{plot_groups_ms2}}
\usage{
plot_groups_ms2(x, ...)
}
\arguments{
\item{x}{An object from which the groups MS2 will be plotted.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S7 generic to plot groups MS2 from \code{x}.
}
