% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_004_plot.R
\name{plot_groups_overview_aux}
\alias{plot_groups_overview_aux}
\title{plot_groups_overview_aux}
\usage{
plot_groups_overview_aux(features, eic, heights, analyses)
}
\arguments{
\item{features}{A data table with the individual peak details to plot.}

\item{eic}{A data table with the analysis, id, rt, intensity and var
(i.e., the plotting variable for each entry) as columns.}

\item{heights}{A numeric vector of length two to control the height of
the first and second plot, respectively.}

\item{analyses}{X.}
}
\value{
plot.
}
\description{
Plots features for each feature group.
}
