% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot_groups_profile}
\alias{plot_groups_profile}
\title{Generic \code{plot_groups_profile}}
\usage{
plot_groups_profile(x, ...)
}
\arguments{
\item{x}{An object from which the groups profile will be plotted.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S7 generic to plot groups profile from \code{x}.
}
