% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ms_plot_bpc.R
\name{plot_interactive_bpc}
\alias{plot_interactive_bpc}
\title{plot_interactive_bpc}
\usage{
plot_interactive_bpc(bpc, title, colorBy)
}
\arguments{
\item{title}{A character vector to be used as title.}

\item{colorBy}{A character vector to name the legend.}

\item{eic}{A data table with the analysis, id, rt, mz, intensity and var
(i.e., the plotting variable for each entry) as columns.}
}
\value{
A BPC interactive plot.
}
\description{
Interactive plot of BPC using the \pkg{plotly} package.
}
