% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ms_plot_eic.R
\name{plot_interactive_eic}
\alias{plot_interactive_eic}
\title{plot_interactive_eic}
\usage{
plot_interactive_eic(eic, title, colorBy)
}
\arguments{
\item{eic}{A data table with the analysis, id, rt, intensity and
var (i.e., the plotting variable for each entry) as columns.}

\item{title}{A character vector to be used as title.}

\item{colorBy}{A character vector to name the legend.}
}
\value{
An EIC interactive plot.
}
\description{
Interactive plot of EIC using the \pkg{plotly} package.
}
