% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ms_plot_ms2.R
\name{plot_interactive_ms2}
\alias{plot_interactive_ms2}
\title{plot_interactive_ms2}
\usage{
plot_interactive_ms2(ms2, title)
}
\arguments{
\item{ms2}{A data.table with the analyses, id, preMZ, rt, mz, intensity,
isPre and var (i.e., the plotting variable for each entry) as columns.}

\item{title}{A character vector to be used as title.}
}
\value{
An MS2 interactive plot.
}
\description{
Interactive plot of MS2 using the \pkg{plotly} package.
}
