% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ms_plot_ms2.R
\name{plot_interactive_spectra}
\alias{plot_interactive_spectra}
\title{plot_interactive_spectra}
\usage{
plot_interactive_spectra(spectra, title = NULL)
}
\arguments{
\item{title}{A character vector to be used as title.}

\item{ms1}{A data.table with the analyses, id, mz, intensity and var
(i.e., the plotting variable for each entry) as columns.}
}
\value{
An spectra interactive plot.
}
\description{
Interactive plot of MS1 spectra using the \pkg{plotly} package.
}
