% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ms_plot_xic.R
\name{plot_interactive_xic}
\alias{plot_interactive_xic}
\title{plot_interactive_xic}
\usage{
plot_interactive_xic(
  xic,
  plotTargetMark = TRUE,
  ppmMark = 5,
  secMark = 10,
  numberRows = 1
)
}
\arguments{
\item{xic}{A \linkS4class{data.table} from the method \code{\link{XICs}}.}

\item{plotTargetMark}{Logical, set to \code{TRUE} (the default) to plot a
target mark.}

\item{ppmMark}{A numeric vector of length one to define the mass deviation,
in ppm, of the target mark. The default is 5 ppm.}

\item{secMark}{A numeric vector of length one to define the time deviation,
in seconds, of the target mark. The default is 10 ppm.}

\item{numberRows}{A numeric vector of length one to define the number of
rows to grid the plots. Note that each target is always plotted in one row
for all selected analyses.}
}
\value{
An iterative plot of the traces for the requested \emph{m/z} and
retention time pairs.
}
\description{
Plot traces or profile data for targets using expected
\emph{m/z} and retention time pairs, including deviations.
}
