% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_S3.R
\name{plot_internal_standards}
\alias{plot_internal_standards}
\title{Generic \code{plot_internal_standards}}
\usage{
plot_internal_standards(x, ...)
}
\arguments{
\item{x}{An object from which the internal standards will be plotted.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S3 generic to plot internal standards from \code{x}.
}
