% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot_loadings}
\alias{plot_loadings}
\title{Generic \code{plot_loadings}}
\usage{
plot_loadings(x, ...)
}
\arguments{
\item{x}{An object from which the loadings will be plotted.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S7 generic to plot loadings from \code{x}.
}
