% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_S3.R
\name{plot_matrix_suppression}
\alias{plot_matrix_suppression}
\title{Generic \code{plot_matrix_suppression}}
\usage{
plot_matrix_suppression(x, ...)
}
\arguments{
\item{x}{An object from which the matrix suppression will be plotted.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S3 generic to plot matrix suppression from \code{x}.
}
