% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ms_plot_functions.R
\name{plot_ms1_interactive}
\alias{plot_ms1_interactive}
\title{plot_ms1_interactive}
\usage{
plot_ms1_interactive(
  ms1 = NULL,
  legendNames = NULL,
  colorBy = "targets",
  title = NULL
)
}
\arguments{
\item{ms1}{A data.table with the analyses, id, mz, intensity and var
(i.e., the plotting variable for each entry) as columns.}

\item{legendNames}{A character vector with the same length as the unique ids
in the table given in \code{eic}.}

\item{colorBy}{A string (length 1). One of "analyses" (the default),
"levels", "targets" or "replicates". For "replicates", a column with
replicate names should be given.}

\item{title}{A character vector to be used as title.}
}
\value{
An ms1 interactive plot.
}
\description{
Interactive plot of MS1 ms1 using the \pkg{plotly} package.
}
