% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_004_plot.R
\name{plot_ms2_interactive}
\alias{plot_ms2_interactive}
\title{plot_ms2_interactive}
\usage{
plot_ms2_interactive(
  ms2 = NULL,
  legendNames = NULL,
  colorBy = "targets",
  title = NULL
)
}
\arguments{
\item{ms2}{A data.table with the id, preMZ, mz, intensity,
isPre and var (i.e., the plotting variable for each entry) as columns.}

\item{legendNames}{A character vector with the same length as the unique ids
in the table given in \code{eic}.}

\item{colorBy}{A string (length 1). One of "analyses" (the default),
"levels", "targets" or "replicates". For "replicates", a column with
replicate names should be given.}

\item{title}{A character vector to be used as title.}
}
\value{
An MS2 spectra interactive plot.
}
\description{
Interactive plot of MS2 spectra using the \pkg{plotly} package.
}
