% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_S3.R
\name{plot_overview}
\alias{plot_overview}
\title{Generic \code{plot_overview}}
\usage{
plot_overview(x, ...)
}
\arguments{
\item{x}{An object from which the overview will be plotted.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S3 generic to plot an overview from \code{x}.
}
