% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot_residual_distance}
\alias{plot_residual_distance}
\title{Generic \code{plot_residual_distance}}
\usage{
plot_residual_distance(x, ...)
}
\arguments{
\item{x}{An object from which the residual distance will be plotted.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S7 generic to plot residual distance from \code{x}.
}
