% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_S3.R
\name{plot_resolved_spectra}
\alias{plot_resolved_spectra}
\title{Generic \code{plot_resolved_spectra}}
\usage{
plot_resolved_spectra(x, ...)
}
\arguments{
\item{x}{An object from which the resolved spectra will be plotted.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S3 generic to plot resolved spectra from \code{x}.
}
