% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot_scores}
\alias{plot_scores}
\title{Generic \code{plot_scores}}
\usage{
plot_scores(x, ...)
}
\arguments{
\item{x}{An object from which the scores will be plotted.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S7 generic to plot scores from \code{x}.
}
