% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot_spectra}
\alias{plot_spectra}
\title{Generic \code{plot_spectra}}
\usage{
plot_spectra(x, ...)
}
\arguments{
\item{x}{An object from which the spectra will be plotted.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S7 generic to plot spectra from \code{x}.
}
