% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot_spectra_3d}
\alias{plot_spectra_3d}
\title{Generic \code{plot_spectra_3d}}
\usage{
plot_spectra_3d(x, ...)
}
\arguments{
\item{x}{An object from which the 3D spectra will be plotted.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S7 generic to plot 3D spectra from \code{x}.
}
