% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot_spectra_baseline}
\alias{plot_spectra_baseline}
\title{Generic \code{plot_spectra_baseline}}
\usage{
plot_spectra_baseline(x, ...)
}
\arguments{
\item{x}{An object from which the spectra baseline will be plotted.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S7 generic to plot spectra baseline from \code{x}.
}
