% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_S3.R
\name{plot_spectra_bpc}
\alias{plot_spectra_bpc}
\title{Generic \code{plot_spectra_bpc}}
\usage{
plot_spectra_bpc(x, ...)
}
\arguments{
\item{x}{An object from which the spectra BPC will be plotted.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S3 generic to plot spectra BPC from \code{x}.
}
