% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot_spectra_charges}
\alias{plot_spectra_charges}
\title{Generic \code{plot_spectra_charges}}
\usage{
plot_spectra_charges(x, ...)
}
\arguments{
\item{x}{An object from which the spectra charges will be plotted.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S7 generic to plot spectra charges from \code{x}.
}
