% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot_spectra_eic}
\alias{plot_spectra_eic}
\title{Generic \code{plot_spectra_eic}}
\usage{
plot_spectra_eic(x, ...)
}
\arguments{
\item{x}{An object from which the spectra EIC will be plotted.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S7 generic to plot spectra EIC from \code{x}.
}
