% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ms_plot_functions.R
\name{plot_spectra_interactive}
\alias{plot_spectra_interactive}
\title{plot_spectra_interactive}
\usage{
plot_spectra_interactive(spec = NULL, colorBy = "analyses")
}
\arguments{
\item{colorBy}{A string (length 1). One of "analyses" (the default),
"levels", "targets" or "replicates". For "replicates", a column with
replicate names should be given.}

\item{spectra}{A \code{data.table} with the analysis, level, rt, mz and
intensity.}
}
\value{
A 3D interactive plot.
}
\description{
3D interactive plot for spectra using the \pkg{plotly} package.
}
