% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot_spectra_ms1}
\alias{plot_spectra_ms1}
\title{Generic \code{plot_spectra_ms1}}
\usage{
plot_spectra_ms1(x, ...)
}
\arguments{
\item{x}{An object from which the spectra MS1 will be plotted.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S7 generic to plot spectra MS1 from \code{x}.
}
