% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot_spectra_peaks}
\alias{plot_spectra_peaks}
\title{Generic \code{plot_spectra_peaks}}
\usage{
plot_spectra_peaks(x, ...)
}
\arguments{
\item{x}{An object from which the spectra peaks will be plotted.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S7 generic to plot spectra peaks from \code{x}.
}
