% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_S3.R
\name{plot_spectra_tic}
\alias{plot_spectra_tic}
\title{Generic \code{plot_spectra_tic}}
\usage{
plot_spectra_tic(x, ...)
}
\arguments{
\item{x}{An object from which the spectra TIC will be plotted.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S3 generic to plot spectra TIC from \code{x}.
}
