% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot_spectra_xic}
\alias{plot_spectra_xic}
\title{Generic \code{plot_spectra_xic}}
\usage{
plot_spectra_xic(x, ...)
}
\arguments{
\item{x}{An object from which the spectra XIC will be plotted.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S7 generic to plot spectra XIC from \code{x}.
}
