% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ms_plot_eic.R
\name{plot_static_eic}
\alias{plot_static_eic}
\title{plot_static_eic}
\usage{
plot_static_eic(eic, title = NULL)
}
\arguments{
\item{eic}{A data table with the analysis, id, rt, intensity and var
(i.e., the plotting variable for each entry) as columns.}

\item{title}{A character vector to be used as title.}
}
\value{
An EIC static plot.
}
\description{
Static plot of EIC using the \pkg{base} package.
}
