% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ms_plot_ms2.R
\name{plot_static_ms2}
\alias{plot_static_ms2}
\title{plot_static_ms2}
\usage{
plot_static_ms2(ms2, title = NULL)
}
\arguments{
\item{ms2}{A data table with the id, mz, intensity, preMZ, isPre
and var (i.e., the plotting variable for each entry) as columns.}

\item{title}{A character vector to be used as title.}
}
\value{
An MSn plot.
}
\description{
Static plot of MSn spectra using the \pkg{base} package.
}
