% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ms_plot_ms2.R
\name{plot_static_spectra}
\alias{plot_static_spectra}
\title{plot_static_spectra}
\usage{
plot_static_spectra(spectra, title = NULL)
}
\arguments{
\item{spectra}{A data.table with the analyses, id, mz, intensity and var
(i.e., the plotting variable for each entry) as columns.}

\item{title}{A character vector to be used as title.}
}
\value{
An spectra plot.
}
\description{
Static plot of spectra using the \pkg{base} package.
}
