% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot_suspects}
\alias{plot_suspects}
\title{Generic \code{plot_suspects}}
\usage{
plot_suspects(x, ...)
}
\arguments{
\item{x}{An object from which the suspects will be plotted.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S7 generic to plot suspects from \code{x}.
}
