% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ms_plot_functions.R
\name{plot_xic_interactive}
\alias{plot_xic_interactive}
\title{plot_xic_interactive}
\usage{
plot_xic_interactive(
  xic,
  legendNames = NULL,
  plotTargetMark = TRUE,
  targetsMark = NULL,
  ppmMark = 5,
  secMark = 10,
  numberRows = 1
)
}
\arguments{
\item{xic}{A \code{data.table} with columns: .}

\item{legendNames}{A character vector with the same length as the unique ids
in the table given in \code{xic}.}

\item{plotTargetMark}{Logical, set to \code{TRUE} (the default) to plot a
target mark.}

\item{targetsMark}{A data.frame with columns "mz" and "rt", defining the
\emph{m/z} and retention time values of each targets (i.e., each unique id
in the \code{xic} table). Note that the number of rows should match with the
number of unique target ids.}

\item{ppmMark}{A numeric vector of length one to define the mass deviation,
in ppm, of the target mark. The default is 5 ppm.}

\item{secMark}{A numeric vector of length one to define the time deviation,
in seconds, of the target mark. The default is 10 ppm.}

\item{numberRows}{A numeric vector of length one to define the number of
rows to grid the plots. Note that each target is always plotted in one row
for all selected analyses.}
}
\value{
An iterative plot of the traces for the requested \emph{m/z} and
retention time pairs.
}
\description{
Plot traces or profile data for targets using expected
\emph{m/z} and retention time pairs, including deviations.
}
