% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ms_plot_alignment.R
\name{plotAlignment}
\alias{plotAlignment}
\title{plotAlignment}
\usage{
plotAlignment(object, interactive = TRUE)
}
\arguments{
\item{object}{An \linkS4class{msData} or \linkS4class{msAnalysis} object with adjusted retention time.}

\item{interactive}{Logical, set to \code{TRUE} for plotting with \pkg{plotly} package.}
}
\value{
A plot with the retention time alignment differences for each sample.
}
\description{
Plots the results from the retention time alignment across analyses.
}
