% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ms_plot_annotation.R
\name{plotAnnotationInteractive}
\alias{plotAnnotationInteractive}
\title{plotAnnotationInteractive}
\usage{
plotAnnotationInteractive(object, comps, colorBy = "isotopes")
}
\value{
A dotted plot with features annotation.
}
\description{
Function to plot features annotation.
}
