% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_000_generics.R
\name{plotBPCs}
\alias{plotBPCs}
\title{plotBPCs}
\usage{
plotBPCs(object, ...)
}
\arguments{
\item{object}{A method specific class object.}

\item{...}{Other method specific arguments.}
}
\description{
Generic for getting base peak chromatograms (BPCs).
}
