% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ms_plot_featurePeaks.R
\name{plotFeaturePeaks}
\alias{plotFeaturePeaks}
\title{plotFeaturePeaks}
\usage{
plotFeaturePeaks(
  object,
  analyses = NULL,
  targetsID = NULL,
  mz = NULL,
  ppm = 20,
  rt = NULL,
  sec = 30,
  id = NULL,
  heights = c(0.6, 0.4)
)
}
\arguments{
\item{object}{An \linkS4class{msData} object.}

\item{analyses}{A numeric or character vector with the index or
analysis name/s. When \code{NULL}, all the analyses are used.}

\item{targetsID}{A character vector with the ID of targets
(e.g., peaks or features IDs).}

\item{mz}{A vector with target \emph{m/z} values or a two columns
\linkS4class{data.table} or data.frame with minimum and maximum
\emph{m/z} values. Alternatively, \emph{m/z} and retention time values
can be given as one \linkS4class{data.table}/data.frame and the deviations
given as \code{ppm} and \code{sec} are used to calculate the ranges.
The same also works for min and max values of \emph{m/z} and retention
time targets. Note that when mass/time ranges are given, \code{ppm} and
\code{sec} are not used.}

\item{ppm}{A numeric vector of length one with the mass deviation, in ppm.}

\item{rt}{A vector with target retention time values or
a two columns \linkS4class{data.table}/data.frame with minimum
and maximum retention time values.}

\item{sec}{A numeric vector of length one with the time deviation, in seconds.}

\item{id}{An id vector with target identifiers. When not given
is added as a combination of the \emph{m/z} and retention time ranges or values.}

\item{heights}{A numeric vector of length two to control the height of
the first and second plot, respectively.}
}
\value{
A double plot with peak chromatograms on the top part
and feature peak groups below.
}
\description{
Plots peaks for each feature in an \linkS4class{msData} object.
}
