% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ms_plot_BPCs.R
\name{plotInteractiveBPCs}
\alias{plotInteractiveBPCs}
\title{plotInteractiveBPCs}
\usage{
plotInteractiveBPCs(bpc, title, colorBy)
}
\arguments{
\item{title}{A character vector to be used as title.}

\item{colorBy}{A character vector to name the legend.}

\item{eic}{A data table with the analysis, replicate,
id, rt, intensity and var (i.e., the plotting variable for each entry)
as columns.}
}
\value{
A EIC interactive plot.
}
\description{
Interactive plot of EICs using the \pkg{plotly} package.
}
