% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ms_plot_MSn.R
\name{plotInteractiveMSn}
\alias{plotInteractiveMSn}
\title{plotInteractiveMSn}
\usage{
plotInteractiveMSn(ms2, title)
}
\arguments{
\item{ms2}{A data table with the sample, replicate,
id, rt, intensity and var (i.e., the plotting variable for each entry)
as columns.}

\item{title}{A character vector to be used as title.}
}
\value{
A EIC interactive plot.
}
\description{
Interactive plot of EICs using the \pkg{plotly} package.
}
