% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_methods_datatable.R
\name{plotMS2s,data.table-method}
\alias{plotMS2s,data.table-method}
\title{plotMS2s-data.table}
\usage{
\S4method{plotMS2s}{data.table}(
  object = NULL,
  analyses = NULL,
  replicates = NULL,
  targets = NULL,
  legendNames = NULL,
  title = NULL,
  colorBy = "targets",
  interactive = TRUE
)
}
\arguments{
\item{object}{A \linkS4class{data.table} as produced by
the method \code{\link{MS2s}}.}

\item{analyses}{A numeric or character vector with the index or
analysis name/s. When \code{NULL}, all the analyses are used.}

\item{replicates}{A numeric or character vector with the indice/s or name/s
of replicates from the object.}

\item{targets}{A character vector with target names.}

\item{legendNames}{A character vector with the same length and order
as the number and order of targets to be used as plot legend.}

\item{title}{A character string with the plot tile.}

\item{colorBy}{A length one character vector applicable to the respective method.}

\item{interactive}{Logical value, set to \code{TRUE} to use
the \pkg{plotly} instead of \pkg{base}. The default is \code{FALSE}.}
}
\description{
Plots MS2 data for specified \emph{m/z} and retention time
(seconds) targets in a \linkS4class{data.table} as obtained by the
\code{\link{MS2s}}. The targets in the object can be filtered using the
\code{targets} argument. Also, "analyses" and "replicates" can be filtered
using the \code{analyses} and \code{replicates} arguments, respectively.
Note that the column analysis/replicate should be present.
The possible values for the \code{colorBy} argument are
"targets", "analyses", "replicates" and "ce" to color by
each target, analysis, replicate or collision energy, respectively.
}
