% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_000_generics.R
\name{plotMS2s}
\alias{plotMS2s}
\title{plotMS2s}
\usage{
plotMS2s(object, ...)
}
\arguments{
\item{object}{A method specific class object.}

\item{...}{Other method specific arguments.}
}
\description{
Generic for plotting MS 2 spectra.
}
