% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ms_plot_peaks.R
\name{plotPeaksInteractive}
\alias{plotPeaksInteractive}
\title{plotPeaksInteractive}
\usage{
plotPeaksInteractive(eic, pks, title, colorBy)
}
\arguments{
\item{eic}{A data table with the analysis, replicate,
id, rt, intensity and var (i.e., the plotting variables for each peak)
as columns.}

\item{pks}{A data table with the individual peaks to plot.}

\item{title}{An optional character vector to be used as title.}

\item{colorBy}{Possible values are \code{"targets"} (the default),
\code{"analyses"} or \code{replicates},
for coloring by target peaks, analyses or replicates, respectively.}
}
\value{
A chromatographic peak plot through \pkg{plotly}.
}
\description{
Plots chromatographic peaks from a \linkS4class{ntsData} object
with the package \pkg{platly} for an interactive user experience.
}
