% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ms_plot_EICs.R
\name{plotStaticEICs}
\alias{plotStaticEICs}
\title{plotStaticEICs}
\usage{
plotStaticEICs(eic, title = NULL)
}
\arguments{
\item{eic}{A data table with the analysis, replicate,
id, rt, intensity and var (i.e., the plotting variable for each entry)
as columns.}

\item{title}{A character vector to be used as title.}
}
\value{
A EIC plot.
}
\description{
Static plot of EICs using the \pkg{base} package.
}
