% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ms_plot_MSn.R
\name{plotStaticMSn}
\alias{plotStaticMSn}
\title{plotStaticMSn}
\usage{
plotStaticMSn(ms2, title = NULL)
}
\arguments{
\item{ms2}{A data table with the id, mz, intensity, CE, preMZ, precursor and var (i.e., the plotting variable for each entry) as columns.
Optionally, sample and replicate can be in the table.}

\item{title}{A character vector to be used as title.}
}
\value{
A MSn plot.
}
\description{
Static plot of MSn spectra using the \pkg{base} package.
}
