% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_000_generics.R
\name{plotTIC}
\alias{plotTIC}
\title{plotTIC}
\usage{
plotTIC(object, ...)
}
\arguments{
\item{object}{A method specific class object.}

\item{...}{Other method specific arguments.}
}
\description{
Generic for plotting total ion chromatogram (TIC).
}
