% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_methods_datatable.R
\name{plotTICs}
\alias{plotTICs}
\title{plotTICs}
\usage{
plotTICs(object, ...)
}
\arguments{
\item{object}{A method specific class object.}

\item{...}{Other method specific arguments.}
}
\description{
Generic for plotting total ion chromatograms (EICs).
}
