% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_methods_datatable.R
\name{plotXICs,data.table-method}
\alias{plotXICs,data.table-method}
\title{plotXICs-data.table}
\usage{
\S4method{plotXICs}{data.table}(
  object,
  analyses = NULL,
  targets = NULL,
  legendNames = NULL,
  plotTargetMark = TRUE,
  targetsMark = NULL,
  ppmMark = 5,
  secMark = 10,
  numberRows = 1
)
}
\arguments{
\item{object}{A \linkS4class{data.table} as produced by
the method \code{\link{XICs}}.}

\item{analyses}{A numeric or character vector with the index or
analysis name/s. When \code{NULL}, all the analyses are used.}

\item{targets}{A character vector with target names.}

\item{legendNames}{A character vector with the same length and order
as the number and order of targets to be used as plot legend.}

\item{plotTargetMark}{Logical, set to \code{TRUE} to plot a target mark.}

\item{targetsMark}{A two columns \link[data.table]{data.table} or \link{data.frame} with
\emph{m/z} and time targets. The column must be named with "mz" and "rt" for
\emph{m/z} and time values, respectively.}

\item{ppmMark}{A numeric vector of length one to define the mass deviation, in ppm,
of the target mark.}

\item{secMark}{A numeric vector of length one to define the time deviation, in seconds,
of the target mark.}

\item{numberRows}{A numeric vector of length one to define
the number of rows to grid the plots.}
}
\description{
Plots three dimensional (\emph{m/z}, time and intensity)
extracted ion chromatograms (XICs) for specified \emph{m/z} and
retention time pair targets in analyses of a \linkS4class{data.table}
object as produced by the \code{\link{XICs}} method.
\code{analyses} and \code{targets} can be used to filter the XIC table.
When \code{plotTargetMark} is \code{TRUE} a target is plotted representing
the deviations as defined by the arguments \code{ppmMark} and \code{secMark}
in ppm and seconds, respectively. When ranges were given to build the XIC,
exact \emph{m/z} and time targets can be specified with the argument
\code{targetsMark}. \code{targetsMark} should be a two column table
named mz and rt with exact \emph{m/z} and time targets.
Note that the number of rows should be the same as the number of target in
the XIC. The number of rows to plot multiple targets can be defined by
the \code{numberRows} argument.
}
