% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_methods_datatable.R
\name{plotXICs}
\alias{plotXICs}
\title{plotXICs}
\usage{
plotXICs(object, ...)
}
\arguments{
\item{object}{A method specific class object.}

\item{...}{Other method specific arguments.}
}
\description{
Generic for plotting three dimensional ion chromatograms (XICs).
}
