% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{predict}
\alias{predict}
\title{Generic \code{predict}}
\usage{
predict(x, ...)
}
\arguments{
\item{x}{An object from which the data will be predicted.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S7 generic to predict data from \code{x}.
}
