% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_S3_ProcessingSettings.R
\name{ProcessingSettings}
\alias{ProcessingSettings}
\alias{validate.ProcessingSettings}
\alias{as.ProcessingSettings}
\title{\strong{ProcessingSettings} S3 class constructor and methods}
\usage{
ProcessingSettings(
  call = NA_character_,
  algorithm = NA_character_,
  parameters = NULL
)

\method{validate}{ProcessingSettings}(x = NULL)

as.ProcessingSettings(value)
}
\arguments{
\item{call}{Character of length one with the name of the method where the
processing settings are to be applied.}

\item{algorithm}{Character of length one with the name of the algorithm to
be used.}

\item{parameters}{List with parameters specific for the method \code{call} and
\code{algorithm}.}

\item{x}{A ProcessingSettings S3 class object.}

\item{value}{A list to be checked and/or converted to ProcessingSettings S3 class.}
}
\value{
A ProcessingSettings S3 class
}
\description{
Creates a ProcessingSettings S3 class object.
}
\details{
See the method documentation for more information about possible
algorithms and parameters to be used.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate(ProcessingSettings)}: Validates a ProcessingSettings S3 class object, returning a logical value of
length one.

}}
\section{Functions}{
\itemize{
\item \code{as.ProcessingSettings()}: Converts the argument in a ProcessingSettings S3 class object.

}}
