% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ProjectMassSpec.R
\name{ProjectMassSpec}
\alias{ProjectMassSpec}
\title{Internal Project Mass Spec Base R6 Class}
\description{
Internal R6 child of \code{Project} exposing shared Mass Spec tables and native helpers.
}
\keyword{internal}
\section{Super class}{
\code{\link[StreamFind:Project]{StreamFind::Project}} -> \code{ProjectMassSpec}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ProjectMassSpec-new}{\code{ProjectMassSpec$new()}}
\item \href{#method-ProjectMassSpec-get_mass_spec_ptr}{\code{ProjectMassSpec$get_mass_spec_ptr()}}
\item \href{#method-ProjectMassSpec-add_analyses}{\code{ProjectMassSpec$add_analyses()}}
\item \href{#method-ProjectMassSpec-remove_analyses}{\code{ProjectMassSpec$remove_analyses()}}
\item \href{#method-ProjectMassSpec-get_analyses}{\code{ProjectMassSpec$get_analyses()}}
\item \href{#method-ProjectMassSpec-get_analysis_names}{\code{ProjectMassSpec$get_analysis_names()}}
\item \href{#method-ProjectMassSpec-get_replicate_names}{\code{ProjectMassSpec$get_replicate_names()}}
\item \href{#method-ProjectMassSpec-set_replicate_names}{\code{ProjectMassSpec$set_replicate_names()}}
\item \href{#method-ProjectMassSpec-get_blank_names}{\code{ProjectMassSpec$get_blank_names()}}
\item \href{#method-ProjectMassSpec-set_blank_names}{\code{ProjectMassSpec$set_blank_names()}}
\item \href{#method-ProjectMassSpec-get_concentrations}{\code{ProjectMassSpec$get_concentrations()}}
\item \href{#method-ProjectMassSpec-set_concentrations}{\code{ProjectMassSpec$set_concentrations()}}
\item \href{#method-ProjectMassSpec-get_spectra_headers}{\code{ProjectMassSpec$get_spectra_headers()}}
\item \href{#method-ProjectMassSpec-get_spectra_tic}{\code{ProjectMassSpec$get_spectra_tic()}}
\item \href{#method-ProjectMassSpec-get_raw_spectra}{\code{ProjectMassSpec$get_raw_spectra()}}
\item \href{#method-ProjectMassSpec-plot_spectra_tic}{\code{ProjectMassSpec$plot_spectra_tic()}}
\item \href{#method-ProjectMassSpec-plot_spectra_bpc}{\code{ProjectMassSpec$plot_spectra_bpc()}}
\item \href{#method-ProjectMassSpec-get_raw_spectra_eic}{\code{ProjectMassSpec$get_raw_spectra_eic()}}
\item \href{#method-ProjectMassSpec-plot_raw_spectra_eic}{\code{ProjectMassSpec$plot_raw_spectra_eic()}}
\item \href{#method-ProjectMassSpec-get_raw_spectra_ms1}{\code{ProjectMassSpec$get_raw_spectra_ms1()}}
\item \href{#method-ProjectMassSpec-get_raw_spectra_ms2}{\code{ProjectMassSpec$get_raw_spectra_ms2()}}
\item \href{#method-ProjectMassSpec-print}{\code{ProjectMassSpec$print()}}
\item \href{#method-ProjectMassSpec-show}{\code{ProjectMassSpec$show()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="available_processing_methods"><a href='../../StreamFind/html/Project.html#method-Project-available_processing_methods'><code>StreamFind::Project$available_processing_methods()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="copy"><a href='../../StreamFind/html/Project.html#method-Project-copy'><code>StreamFind::Project$copy()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="delete_cache"><a href='../../StreamFind/html/Project.html#method-Project-delete_cache'><code>StreamFind::Project$delete_cache()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="get_audit"><a href='../../StreamFind/html/Project.html#method-Project-get_audit'><code>StreamFind::Project$get_audit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="get_cache"><a href='../../StreamFind/html/Project.html#method-Project-get_cache'><code>StreamFind::Project$get_cache()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="get_cache_size"><a href='../../StreamFind/html/Project.html#method-Project-get_cache_size'><code>StreamFind::Project$get_cache_size()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="get_db"><a href='../../StreamFind/html/Project.html#method-Project-get_db'><code>StreamFind::Project$get_db()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="get_domain"><a href='../../StreamFind/html/Project.html#method-Project-get_domain'><code>StreamFind::Project$get_domain()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="get_metadata"><a href='../../StreamFind/html/Project.html#method-Project-get_metadata'><code>StreamFind::Project$get_metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="get_project_id"><a href='../../StreamFind/html/Project.html#method-Project-get_project_id'><code>StreamFind::Project$get_project_id()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="get_ptr"><a href='../../StreamFind/html/Project.html#method-Project-get_ptr'><code>StreamFind::Project$get_ptr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="get_workflow"><a href='../../StreamFind/html/Project.html#method-Project-get_workflow'><code>StreamFind::Project$get_workflow()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="list_tables"><a href='../../StreamFind/html/Project.html#method-Project-list_tables'><code>StreamFind::Project$list_tables()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="report_quarto"><a href='../../StreamFind/html/Project.html#method-Project-report_quarto'><code>StreamFind::Project$report_quarto()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="run_app"><a href='../../StreamFind/html/Project.html#method-Project-run_app'><code>StreamFind::Project$run_app()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="run_processing_step"><a href='../../StreamFind/html/Project.html#method-Project-run_processing_step'><code>StreamFind::Project$run_processing_step()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="run_workflow"><a href='../../StreamFind/html/Project.html#method-Project-run_workflow'><code>StreamFind::Project$run_workflow()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="set_metadata"><a href='../../StreamFind/html/Project.html#method-Project-set_metadata'><code>StreamFind::Project$set_metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="set_workflow"><a href='../../StreamFind/html/Project.html#method-Project-set_workflow'><code>StreamFind::Project$set_workflow()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="validate"><a href='../../StreamFind/html/Project.html#method-Project-validate'><code>StreamFind::Project$validate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectMassSpec-new"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectMassSpec-new}{}}}
\subsection{Method \code{new()}}{
Create a Mass Spec domain wrapper for a shared \code{Project}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectMassSpec$new(
  db,
  project_id,
  ...,
  file_paths = character(),
  replicates = character(),
  blanks = character()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{db}}{Path to DuckDB file.}

\item{\code{project_id}}{Active project identifier.}

\item{\code{...}}{Additional arguments passed to the method or ignored by the implementation.}

\item{\code{file_paths}}{Character vector with Mass Spec file paths.}

\item{\code{replicates}}{Optional character vector with replicate names.}

\item{\code{blanks}}{Optional character vector with blank names.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectMassSpec-get_mass_spec_ptr"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectMassSpec-get_mass_spec_ptr}{}}}
\subsection{Method \code{get_mass_spec_ptr()}}{
Return the native Mass Spec pointer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectMassSpec$get_mass_spec_ptr()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectMassSpec-add_analyses"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectMassSpec-add_analyses}{}}}
\subsection{Method \code{add_analyses()}}{
Add multiple Mass Spec files into the shared DB as analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectMassSpec$add_analyses(
  file_paths,
  replicates = character(),
  blanks = character()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_paths}}{Character vector with Mass Spec file paths.}

\item{\code{replicates}}{Optional character vector with replicate names.}

\item{\code{blanks}}{Optional character vector with blank names.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectMassSpec-remove_analyses"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectMassSpec-remove_analyses}{}}}
\subsection{Method \code{remove_analyses()}}{
Remove analyses from the shared DB.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectMassSpec$remove_analyses(analyses)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{Character or numeric vector with names or indexes of analyses in the \code{Analyses}
object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectMassSpec-get_analyses"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectMassSpec-get_analyses}{}}}
\subsection{Method \code{get_analyses()}}{
Return imported analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectMassSpec$get_analyses()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectMassSpec-get_analysis_names"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectMassSpec-get_analysis_names}{}}}
\subsection{Method \code{get_analysis_names()}}{
Get analysis names.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectMassSpec$get_analysis_names()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectMassSpec-get_replicate_names"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectMassSpec-get_replicate_names}{}}}
\subsection{Method \code{get_replicate_names()}}{
Get replicate names named by analysis.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectMassSpec$get_replicate_names()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectMassSpec-set_replicate_names"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectMassSpec-set_replicate_names}{}}}
\subsection{Method \code{set_replicate_names()}}{
Set replicate names by analysis order.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectMassSpec$set_replicate_names(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{An object that highly depends on the method. Check the specific method documentation for details.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectMassSpec-get_blank_names"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectMassSpec-get_blank_names}{}}}
\subsection{Method \code{get_blank_names()}}{
Get blank names named by analysis.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectMassSpec$get_blank_names()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectMassSpec-set_blank_names"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectMassSpec-set_blank_names}{}}}
\subsection{Method \code{set_blank_names()}}{
Set blank names by analysis order.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectMassSpec$set_blank_names(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{An object that highly depends on the method. Check the specific method documentation for details.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectMassSpec-get_concentrations"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectMassSpec-get_concentrations}{}}}
\subsection{Method \code{get_concentrations()}}{
Get concentrations named by analysis.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectMassSpec$get_concentrations()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectMassSpec-set_concentrations"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectMassSpec-set_concentrations}{}}}
\subsection{Method \code{set_concentrations()}}{
Set concentrations by analysis order.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectMassSpec$set_concentrations(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{An object that highly depends on the method. Check the specific method documentation for details.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectMassSpec-get_spectra_headers"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectMassSpec-get_spectra_headers}{}}}
\subsection{Method \code{get_spectra_headers()}}{
Return spectra headers for selected analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectMassSpec$get_spectra_headers(analyses = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{Character or numeric vector with names or indexes of analyses in the \code{Analyses}
object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectMassSpec-get_spectra_tic"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectMassSpec-get_spectra_tic}{}}}
\subsection{Method \code{get_spectra_tic()}}{
Return spectra TIC rows for selected analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectMassSpec$get_spectra_tic(analyses = NULL, levels = NULL, rt = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{Character or numeric vector with names or indexes of analyses in the \code{Analyses}
object.}

\item{\code{levels}}{Integer vector with the levels of MS data.}

\item{\code{rt}}{A vector with target retention time values (in seconds) or
a two columns data.table or data.frame with minimum and maximum retention
time values (in seconds).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectMassSpec-get_raw_spectra"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectMassSpec-get_raw_spectra}{}}}
\subsection{Method \code{get_raw_spectra()}}{
Get raw spectra data for selected analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectMassSpec$get_raw_spectra(
  analyses = character(),
  levels = integer(),
  mass = numeric(),
  mz = numeric(),
  rt = numeric(),
  mobility = numeric(),
  ppm = 20,
  sec = 60,
  millisec = 5,
  id = character(),
  allTraces = TRUE,
  isolationWindow = 1.3,
  minIntensityMS1 = 0,
  minIntensityMS2 = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{Character or numeric vector with names or indexes of analyses in the \code{Analyses}
object.}

\item{\code{levels}}{Integer vector with the levels of MS data.}

\item{\code{mass}}{A vector with target neutral mass value/s (in Da) or a two columns data.table or data.frame named \code{min}
and \code{max} with minimum and maximum neutral mass values (in Da), respectively. Alternatively, neutral mass (in Da)
and retention time (in seconds) and/or drift time values (in milliseconds) can be given as one data.table or
data.frame with columns named \code{mass} and \code{rt} and/or \code{drift}. Then, the deviations given in the \code{ppm}, \code{sec}
and \code{millisec}  arguments are used to calculate the ranges. Also works with a data.table or data.frame with
minimum and maximum values of neutral mass, retention time and drift time targets. In this case, the column names
must be \code{min}, \code{max}, \code{rtmin}, \code{rtmax}, \code{driftmin} and \code{driftmax}. Note that when mass/time ranges are given, the
\code{ppm}, \code{sec} and \code{millisec}
arguments are not used.}

\item{\code{mz}}{A vector with target \emph{m/z} value/s (in Da) or a two columns data.table or data.frame named \code{mzmin} and
\code{mzmax} with minimum and maximum \emph{m/z} values (in Da), respectively. Alternatively, \emph{m/z} (in Da) and
retention time values (in seconds) can be given as one data.table or data.frame with columns named \code{mz} and \code{rt}
and/or \code{drift}. Then, the deviations given in the \code{ppm}, \code{sec} and \code{millisec} arguments are used to
calculate the ranges. Also works with a data.table or data.frame with minimum and maximum values of \emph{m/z},
retention time and drift time targets. In this case, the column names must be \code{mzmin}, \code{mzmax}, \code{rtmin}, \code{rtmax},
\code{driftmin} and \code{driftmax}. Note that when mass/time ranges are given, the \code{ppm}, \code{sec} and \code{millisec}
arguments are not used.}

\item{\code{rt}}{A vector with target retention time values (in seconds) or
a two columns data.table or data.frame with minimum and maximum retention
time values (in seconds).}

\item{\code{mobility}}{A vector with target drift time values (in milliseconds) or
a two columns data.table or data.frame with minimum and maximum drift
time values (in milliseconds).}

\item{\code{ppm}}{Numeric of length one with the mass deviation, in ppm.}

\item{\code{sec}}{Numeric of length one with the time deviation, in seconds.}

\item{\code{millisec}}{Numeric of length one with the drift time deviation,
in milliseconds.}

\item{\code{id}}{Character with the same length as \emph{m/z} and retention time
targets to be used as identifier/s. When not given, the id is built as a
combination of the \emph{m/z} and retention time ranges or values.}

\item{\code{allTraces}}{Logical, when \code{TRUE} all level 2 data is returned.
When \code{FALSE} and level has 2, only the MS2 traces of MS1 targets
are returned, using the \code{isolationWindow} to calculate the mass window.}

\item{\code{isolationWindow}}{Numeric value with the isolation window (in Da)
applied for ion isolation and fragmentation during acquisition of tandem
data (i.e., MS2 data).}

\item{\code{isolationWindow}}{Numeric value with the isolation window (in Da)
applied for ion isolation and fragmentation during acquisition of tandem
data (i.e., MS2 data).}

\item{\code{minIntensityMS1}}{Numeric value with the minimum intensity of level 1
data (i.e., MS1 data).}

\item{\code{minIntensityMS2}}{Numeric value with the minimum intensity of level 2
data (i.e., MS2 data).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectMassSpec-plot_spectra_tic"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectMassSpec-plot_spectra_tic}{}}}
\subsection{Method \code{plot_spectra_tic()}}{
Plot total ion current traces for selected analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectMassSpec$plot_spectra_tic(
  analyses = NULL,
  levels = c(1, 2),
  rt = NULL,
  downsize = NULL,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  groupBy = "analysis",
  interactive = TRUE,
  colorPalette = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{Character or numeric vector with names or indexes of analyses in the \code{Analyses}
object.}

\item{\code{levels}}{Integer vector with the levels of MS data.}

\item{\code{rt}}{A vector with target retention time values (in seconds) or
a two columns data.table or data.frame with minimum and maximum retention
time values (in seconds).}

\item{\code{downsize}}{An integer of length one to downsize the TIC plot. The default is 1.}

\item{\code{xLab}}{X axis label for the plot.}

\item{\code{yLab}}{Y axis label for the plot.}

\item{\code{title}}{Plot title.}

\item{\code{groupBy}}{Column to use for color grouping (default: "analysis").}

\item{\code{interactive}}{Logical, use plotly if TRUE, ggplot2 if FALSE.}

\item{\code{colorPalette}}{Optional vector of colors for the plot; if not provided, an internal function is used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectMassSpec-plot_spectra_bpc"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectMassSpec-plot_spectra_bpc}{}}}
\subsection{Method \code{plot_spectra_bpc()}}{
Plot base peak chromatogram traces for selected analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectMassSpec$plot_spectra_bpc(
  analyses = NULL,
  levels = c(1, 2),
  rt = NULL,
  downsize = NULL,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  groupBy = "analysis",
  interactive = TRUE,
  colorPalette = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{Character or numeric vector with names or indexes of analyses in the \code{Analyses}
object.}

\item{\code{levels}}{Integer vector with the levels of MS data.}

\item{\code{rt}}{A vector with target retention time values (in seconds) or
a two columns data.table or data.frame with minimum and maximum retention
time values (in seconds).}

\item{\code{downsize}}{An integer of length one to downsize the TIC plot. The default is 1.}

\item{\code{xLab}}{X axis label for the plot.}

\item{\code{yLab}}{Y axis label for the plot.}

\item{\code{title}}{Plot title.}

\item{\code{groupBy}}{Column to use for color grouping (default: "analysis").}

\item{\code{interactive}}{Logical, use plotly if TRUE, ggplot2 if FALSE.}

\item{\code{colorPalette}}{Optional vector of colors for the plot; if not provided, an internal function is used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectMassSpec-get_raw_spectra_eic"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectMassSpec-get_raw_spectra_eic}{}}}
\subsection{Method \code{get_raw_spectra_eic()}}{
Get extracted ion chromatograms for selected analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectMassSpec$get_raw_spectra_eic(
  analyses = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  id = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{Character or numeric vector with names or indexes of analyses in the \code{Analyses}
object.}

\item{\code{mass}}{A vector with target neutral mass value/s (in Da) or a two columns data.table or data.frame named \code{min}
and \code{max} with minimum and maximum neutral mass values (in Da), respectively. Alternatively, neutral mass (in Da)
and retention time (in seconds) and/or drift time values (in milliseconds) can be given as one data.table or
data.frame with columns named \code{mass} and \code{rt} and/or \code{drift}. Then, the deviations given in the \code{ppm}, \code{sec}
and \code{millisec}  arguments are used to calculate the ranges. Also works with a data.table or data.frame with
minimum and maximum values of neutral mass, retention time and drift time targets. In this case, the column names
must be \code{min}, \code{max}, \code{rtmin}, \code{rtmax}, \code{driftmin} and \code{driftmax}. Note that when mass/time ranges are given, the
\code{ppm}, \code{sec} and \code{millisec}
arguments are not used.}

\item{\code{mz}}{A vector with target \emph{m/z} value/s (in Da) or a two columns data.table or data.frame named \code{mzmin} and
\code{mzmax} with minimum and maximum \emph{m/z} values (in Da), respectively. Alternatively, \emph{m/z} (in Da) and
retention time values (in seconds) can be given as one data.table or data.frame with columns named \code{mz} and \code{rt}
and/or \code{drift}. Then, the deviations given in the \code{ppm}, \code{sec} and \code{millisec} arguments are used to
calculate the ranges. Also works with a data.table or data.frame with minimum and maximum values of \emph{m/z},
retention time and drift time targets. In this case, the column names must be \code{mzmin}, \code{mzmax}, \code{rtmin}, \code{rtmax},
\code{driftmin} and \code{driftmax}. Note that when mass/time ranges are given, the \code{ppm}, \code{sec} and \code{millisec}
arguments are not used.}

\item{\code{rt}}{A vector with target retention time values (in seconds) or
a two columns data.table or data.frame with minimum and maximum retention
time values (in seconds).}

\item{\code{mobility}}{A vector with target drift time values (in milliseconds) or
a two columns data.table or data.frame with minimum and maximum drift
time values (in milliseconds).}

\item{\code{ppm}}{Numeric of length one with the mass deviation, in ppm.}

\item{\code{sec}}{Numeric of length one with the time deviation, in seconds.}

\item{\code{millisec}}{Numeric of length one with the drift time deviation,
in milliseconds.}

\item{\code{id}}{Character with the same length as \emph{m/z} and retention time
targets to be used as identifier/s. When not given, the id is built as a
combination of the \emph{m/z} and retention time ranges or values.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectMassSpec-plot_raw_spectra_eic"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectMassSpec-plot_raw_spectra_eic}{}}}
\subsection{Method \code{plot_raw_spectra_eic()}}{
Plot extracted ion chromatograms for selected analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectMassSpec$plot_raw_spectra_eic(
  analyses = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  id = NULL,
  downsize = NULL,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  groupBy = c("analysis", "id"),
  interactive = TRUE,
  colorPalette = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{Character or numeric vector with names or indexes of analyses in the \code{Analyses}
object.}

\item{\code{mass}}{A vector with target neutral mass value/s (in Da) or a two columns data.table or data.frame named \code{min}
and \code{max} with minimum and maximum neutral mass values (in Da), respectively. Alternatively, neutral mass (in Da)
and retention time (in seconds) and/or drift time values (in milliseconds) can be given as one data.table or
data.frame with columns named \code{mass} and \code{rt} and/or \code{drift}. Then, the deviations given in the \code{ppm}, \code{sec}
and \code{millisec}  arguments are used to calculate the ranges. Also works with a data.table or data.frame with
minimum and maximum values of neutral mass, retention time and drift time targets. In this case, the column names
must be \code{min}, \code{max}, \code{rtmin}, \code{rtmax}, \code{driftmin} and \code{driftmax}. Note that when mass/time ranges are given, the
\code{ppm}, \code{sec} and \code{millisec}
arguments are not used.}

\item{\code{mz}}{A vector with target \emph{m/z} value/s (in Da) or a two columns data.table or data.frame named \code{mzmin} and
\code{mzmax} with minimum and maximum \emph{m/z} values (in Da), respectively. Alternatively, \emph{m/z} (in Da) and
retention time values (in seconds) can be given as one data.table or data.frame with columns named \code{mz} and \code{rt}
and/or \code{drift}. Then, the deviations given in the \code{ppm}, \code{sec} and \code{millisec} arguments are used to
calculate the ranges. Also works with a data.table or data.frame with minimum and maximum values of \emph{m/z},
retention time and drift time targets. In this case, the column names must be \code{mzmin}, \code{mzmax}, \code{rtmin}, \code{rtmax},
\code{driftmin} and \code{driftmax}. Note that when mass/time ranges are given, the \code{ppm}, \code{sec} and \code{millisec}
arguments are not used.}

\item{\code{rt}}{A vector with target retention time values (in seconds) or
a two columns data.table or data.frame with minimum and maximum retention
time values (in seconds).}

\item{\code{mobility}}{A vector with target drift time values (in milliseconds) or
a two columns data.table or data.frame with minimum and maximum drift
time values (in milliseconds).}

\item{\code{ppm}}{Numeric of length one with the mass deviation, in ppm.}

\item{\code{sec}}{Numeric of length one with the time deviation, in seconds.}

\item{\code{millisec}}{Numeric of length one with the drift time deviation,
in milliseconds.}

\item{\code{id}}{Character with the same length as \emph{m/z} and retention time
targets to be used as identifier/s. When not given, the id is built as a
combination of the \emph{m/z} and retention time ranges or values.}

\item{\code{downsize}}{An integer of length one to downsize the TIC plot. The default is 1.}

\item{\code{xLab}}{X axis label for the plot.}

\item{\code{yLab}}{Y axis label for the plot.}

\item{\code{title}}{Plot title.}

\item{\code{groupBy}}{Column to use for color grouping (default: "analysis").}

\item{\code{interactive}}{Logical, use plotly if TRUE, ggplot2 if FALSE.}

\item{\code{colorPalette}}{Optional vector of colors for the plot; if not provided, an internal function is used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectMassSpec-get_raw_spectra_ms1"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectMassSpec-get_raw_spectra_ms1}{}}}
\subsection{Method \code{get_raw_spectra_ms1()}}{
Get clustered MS1 spectra for selected analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectMassSpec$get_raw_spectra_ms1(
  analyses = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  id = NULL,
  mzClust = 0.003,
  presence = 0.8,
  minIntensity = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{Character or numeric vector with names or indexes of analyses in the \code{Analyses}
object.}

\item{\code{mass}}{A vector with target neutral mass value/s (in Da) or a two columns data.table or data.frame named \code{min}
and \code{max} with minimum and maximum neutral mass values (in Da), respectively. Alternatively, neutral mass (in Da)
and retention time (in seconds) and/or drift time values (in milliseconds) can be given as one data.table or
data.frame with columns named \code{mass} and \code{rt} and/or \code{drift}. Then, the deviations given in the \code{ppm}, \code{sec}
and \code{millisec}  arguments are used to calculate the ranges. Also works with a data.table or data.frame with
minimum and maximum values of neutral mass, retention time and drift time targets. In this case, the column names
must be \code{min}, \code{max}, \code{rtmin}, \code{rtmax}, \code{driftmin} and \code{driftmax}. Note that when mass/time ranges are given, the
\code{ppm}, \code{sec} and \code{millisec}
arguments are not used.}

\item{\code{mz}}{A vector with target \emph{m/z} value/s (in Da) or a two columns data.table or data.frame named \code{mzmin} and
\code{mzmax} with minimum and maximum \emph{m/z} values (in Da), respectively. Alternatively, \emph{m/z} (in Da) and
retention time values (in seconds) can be given as one data.table or data.frame with columns named \code{mz} and \code{rt}
and/or \code{drift}. Then, the deviations given in the \code{ppm}, \code{sec} and \code{millisec} arguments are used to
calculate the ranges. Also works with a data.table or data.frame with minimum and maximum values of \emph{m/z},
retention time and drift time targets. In this case, the column names must be \code{mzmin}, \code{mzmax}, \code{rtmin}, \code{rtmax},
\code{driftmin} and \code{driftmax}. Note that when mass/time ranges are given, the \code{ppm}, \code{sec} and \code{millisec}
arguments are not used.}

\item{\code{rt}}{A vector with target retention time values (in seconds) or
a two columns data.table or data.frame with minimum and maximum retention
time values (in seconds).}

\item{\code{mobility}}{A vector with target drift time values (in milliseconds) or
a two columns data.table or data.frame with minimum and maximum drift
time values (in milliseconds).}

\item{\code{ppm}}{Numeric of length one with the mass deviation, in ppm.}

\item{\code{sec}}{Numeric of length one with the time deviation, in seconds.}

\item{\code{millisec}}{Numeric of length one with the drift time deviation,
in milliseconds.}

\item{\code{id}}{Character with the same length as \emph{m/z} and retention time
targets to be used as identifier/s. When not given, the id is built as a
combination of the \emph{m/z} and retention time ranges or values.}

\item{\code{mzClust}}{Numeric (length 1) with the mass deviation threshold (in Da) to cluster mass traces. Highly dependent
on the mass resolution of the MS data.}

\item{\code{presence}}{Numeric (length 1) with the required presence ratio from 0 (i.e., doesn't need to be present in any
spectra) to 1 (i.e., must be present in all spectra) for traces during clustering of spectra.}

\item{\code{minIntensity}}{Numeric (length 1) with the minimum intensity.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectMassSpec-get_raw_spectra_ms2"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectMassSpec-get_raw_spectra_ms2}{}}}
\subsection{Method \code{get_raw_spectra_ms2()}}{
Get clustered MS2 spectra for selected analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectMassSpec$get_raw_spectra_ms2(
  analyses = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  id = NULL,
  isolationWindow = 1.3,
  mzClust = 0.005,
  presence = 0.8,
  minIntensity = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{Character or numeric vector with names or indexes of analyses in the \code{Analyses}
object.}

\item{\code{mass}}{A vector with target neutral mass value/s (in Da) or a two columns data.table or data.frame named \code{min}
and \code{max} with minimum and maximum neutral mass values (in Da), respectively. Alternatively, neutral mass (in Da)
and retention time (in seconds) and/or drift time values (in milliseconds) can be given as one data.table or
data.frame with columns named \code{mass} and \code{rt} and/or \code{drift}. Then, the deviations given in the \code{ppm}, \code{sec}
and \code{millisec}  arguments are used to calculate the ranges. Also works with a data.table or data.frame with
minimum and maximum values of neutral mass, retention time and drift time targets. In this case, the column names
must be \code{min}, \code{max}, \code{rtmin}, \code{rtmax}, \code{driftmin} and \code{driftmax}. Note that when mass/time ranges are given, the
\code{ppm}, \code{sec} and \code{millisec}
arguments are not used.}

\item{\code{mz}}{A vector with target \emph{m/z} value/s (in Da) or a two columns data.table or data.frame named \code{mzmin} and
\code{mzmax} with minimum and maximum \emph{m/z} values (in Da), respectively. Alternatively, \emph{m/z} (in Da) and
retention time values (in seconds) can be given as one data.table or data.frame with columns named \code{mz} and \code{rt}
and/or \code{drift}. Then, the deviations given in the \code{ppm}, \code{sec} and \code{millisec} arguments are used to
calculate the ranges. Also works with a data.table or data.frame with minimum and maximum values of \emph{m/z},
retention time and drift time targets. In this case, the column names must be \code{mzmin}, \code{mzmax}, \code{rtmin}, \code{rtmax},
\code{driftmin} and \code{driftmax}. Note that when mass/time ranges are given, the \code{ppm}, \code{sec} and \code{millisec}
arguments are not used.}

\item{\code{rt}}{A vector with target retention time values (in seconds) or
a two columns data.table or data.frame with minimum and maximum retention
time values (in seconds).}

\item{\code{mobility}}{A vector with target drift time values (in milliseconds) or
a two columns data.table or data.frame with minimum and maximum drift
time values (in milliseconds).}

\item{\code{ppm}}{Numeric of length one with the mass deviation, in ppm.}

\item{\code{sec}}{Numeric of length one with the time deviation, in seconds.}

\item{\code{millisec}}{Numeric of length one with the drift time deviation,
in milliseconds.}

\item{\code{id}}{Character with the same length as \emph{m/z} and retention time
targets to be used as identifier/s. When not given, the id is built as a
combination of the \emph{m/z} and retention time ranges or values.}

\item{\code{isolationWindow}}{Numeric value with the isolation window (in Da)
applied for ion isolation and fragmentation during acquisition of tandem
data (i.e., MS2 data).}

\item{\code{isolationWindow}}{Numeric value with the isolation window (in Da)
applied for ion isolation and fragmentation during acquisition of tandem
data (i.e., MS2 data).}

\item{\code{mzClust}}{Numeric (length 1) with the mass deviation threshold (in Da) to cluster mass traces. Highly dependent
on the mass resolution of the MS data.}

\item{\code{presence}}{Numeric (length 1) with the required presence ratio from 0 (i.e., doesn't need to be present in any
spectra) to 1 (i.e., must be present in all spectra) for traces during clustering of spectra.}

\item{\code{minIntensity}}{Numeric (length 1) with the minimum intensity.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectMassSpec-print"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectMassSpec-print}{}}}
\subsection{Method \code{print()}}{
Print a short summary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectMassSpec$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments passed to the method or ignored by the implementation.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectMassSpec-show"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectMassSpec-show}{}}}
\subsection{Method \code{show()}}{
Show a short summary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectMassSpec$show(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments passed to the method or ignored by the implementation.}
}
\if{html}{\out{</div>}}
}
}
}
