% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_R6_implementation.R
\name{R6MS}
\alias{R6MS}
\title{R6 Class Representing an MS Set of Analysis Files}
\description{
A description of the R6 class.
}
\details{
The details.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{admin}}{A list with administrative information: name, author,
description, path, date.}

\item{\code{path}}{The path of the object on disk. Default to `getwd()`.}

\item{\code{date}}{The creation date and time.}

\item{\code{header}}{The header data.frame with all the analysis types, names,
replicates, associated blank replicates, polarities and full file paths.}

\item{\code{analyses}}{The named list with all the analyses. The following
entries are listed for each analysis: ...}

\item{\code{features}}{The features data.frame.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-R6MS-new}{\code{R6MS$new()}}
\item \href{#method-R6MS-print}{\code{R6MS$print()}}
\item \href{#method-R6MS-get_analyses}{\code{R6MS$get_analyses()}}
\item \href{#method-R6MS-get_analysis_spectra}{\code{R6MS$get_analysis_spectra()}}
\item \href{#method-R6MS-subset_analyses}{\code{R6MS$subset_analyses()}}
\item \href{#method-R6MS-clone}{\code{R6MS$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6MS-new"></a>}}
\if{latex}{\out{\hypertarget{method-R6MS-new}{}}}
\subsection{Method \code{new()}}{
Create a new MS set of analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6MS$new(
  files = NULL,
  admin = NULL,
  header = NULL,
  analyses = NULL,
  features = NULL,
  run_parallel = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{files}}{A character vector with the full file path of MS analyses.}

\item{\code{admin}}{A list with administrative information.}

\item{\code{header}}{A data.frame with the header of the analyses.}

\item{\code{analyses}}{A list with MS analyses information.}

\item{\code{features}}{A data.frame with features representing corresponding
peaks across MS analyses.}

\item{\code{run_parallel}}{Logical, set to \code{TRUE} for processing the data
in parallel.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `r6MS` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6MS-print"></a>}}
\if{latex}{\out{\hypertarget{method-R6MS-print}{}}}
\subsection{Method \code{print()}}{
Prints a summary of the `R6MS` object in the console.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6MS$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Console text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6MS-get_analyses"></a>}}
\if{latex}{\out{\hypertarget{method-R6MS-get_analyses}{}}}
\subsection{Method \code{get_analyses()}}{
Method to get specific analyses from the `r6MS` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6MS$get_analyses(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A numeric/character vector with the number/name of the analyses.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The list of analyses defined by `value`.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6MS-get_analysis_spectra"></a>}}
\if{latex}{\out{\hypertarget{method-R6MS-get_analysis_spectra}{}}}
\subsection{Method \code{get_analysis_spectra()}}{
Create a new MS set of analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6MS$get_analysis_spectra(
  levels = 1,
  rtr = NULL,
  preMZrange = NULL,
  minIntensityMS1 = 0,
  minIntensityMS2 = 0,
  run_parallel = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{levels}}{Name.}

\item{\code{rtr}}{Hair color.}

\item{\code{preMZrange}}{Hair color.}

\item{\code{minIntensityMS1}}{Hair color.}

\item{\code{minIntensityMS2}}{Hair color.}

\item{\code{run_parallel}}{Hair color.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame with spectra.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6MS-subset_analyses"></a>}}
\if{latex}{\out{\hypertarget{method-R6MS-subset_analyses}{}}}
\subsection{Method \code{subset_analyses()}}{
Create a new MS set of analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6MS$subset_analyses(i, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{i}}{Name.}

\item{\code{...}}{Hair color.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6MS-clone"></a>}}
\if{latex}{\out{\hypertarget{method-R6MS-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6MS$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
