% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Analyses_RamanAnalyses.R
\name{RamanAnalyses}
\alias{RamanAnalyses}
\title{RamanAnalyses Class}
\usage{
RamanAnalyses(files = NULL)
}
\arguments{
\item{files}{A \code{character} vector with full file paths to "asc", "sif", "json", "wdf", "sdf",
"csv" and/or "txt" raman files or a \code{data.frame} with colnames \code{file}, \code{replicate} and \code{blank}.
The "replicate" column is used to group the analyses and the "blank" column is used to identify
the blank samples. The "file" column is the full to the raman files.}
}
\description{
The RamanAnalyses class is used to store a list of Raman spectra.
}
\section{Slots}{

\describe{
\item{\code{analyses}}{A \code{list} of Raman spectra.}

\item{\code{replicates}}{(getter and setter) A \code{character} vector with the names of the replicates.}

\item{\code{blanks}}{(getter and setter) A \code{character} vector with the names of the blanks.}

\item{\code{concentrations}}{(getter and setter) A \code{numeric} vector with the concentrations of each
analysis.}

\item{\code{references}}{(getter and setter) A \code{character} vector with the names of the reference for
each analysis.}

\item{\code{types}}{(getter) A \code{character} vector with the names of the types of each analysis.}

\item{\code{files}}{(getter) A \code{character} vector with the names of the files of each analysis.}

\item{\code{info}}{(getter) A \code{data.frame} with the information of each analysis.}

\item{\code{has_spectra}}{(getter) A \code{logical} value indicating if the object has spectra.}

\item{\code{Spectra}}{(getter and setter) \code{RamanSpectraResults}.}
}}

