% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RamanMethod_AddShiftValues.R
\name{RamanMethod_AddShiftValues_native}
\alias{RamanMethod_AddShiftValues_native}
\alias{validate_object.RamanMethod_AddShiftValues_native}
\title{RamanMethod_AddShiftValues_native Class}
\usage{
RamanMethod_AddShiftValues_native(
  mode = "analyses",
  index = NA_integer_,
  shifts = NA_real_
)

\method{validate_object}{RamanMethod_AddShiftValues_native}(x)
}
\arguments{
\item{mode}{A character string. The mode to be used, possible values are \code{analyses} or \code{custom}
for copying from analyses or adding custom shift values, respectively.}

\item{index}{An integer vector with the index/es of the analysis/analyses to copy the shift
values from.}

\item{shifts}{A numeric vector with the shift values to be added to the spectra.}

\item{A}{RamanMethod_AddShiftValues_native object.}
}
\value{
A RamanMethod_AddShiftValues_native object.
}
\description{
Adds custom shift values to the Raman spectra or copies from one or more analyses to
the other analyses. When more than one analyses are selected, the shift values are copied from
the selected but them averaged before being added to the other analyses. If the column "rt" is
present in the spectra, the shift values are added to each retention time.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate_object(RamanMethod_AddShiftValues_native)}: Validate the RamanMethod_AddShiftValues_native object, returning NULL if valid.

}}
