% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RamanMethod_CalculateSpectraSimilary.R
\name{RamanMethod_CalculateSpectraSimilary_hqi}
\alias{RamanMethod_CalculateSpectraSimilary_hqi}
\title{RamanMethod_CalculateSpectraSimilary_hqi Class}
\usage{
RamanMethod_CalculateSpectraSimilary_hqi(
  reference = data.frame(),
  method = "pearson"
)
}
\arguments{
\item{reference}{data.frame with the reference spectra to be used for similarity. The data.frame
should have two columns with shift and intensity values. Note that the shift values should match
the shift values of the analyses.}

\item{method}{character(1) indicating the method to be used for the similarity calculation. The
available methods are "pearson", "kendall", and "spearman".}
}
\value{
A RamanMethod_CalculateSpectraSimilary_hqi object.
}
\description{
Calculate the similarity between Raman spectra based on predefined reference
spectra. The similarity is assessed by estimating the high quality index (HQI) and P-value
between a given analysis and the associated reference.
}
