% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RamanMethod_GroupPeaks.R
\name{RamanMethod_GroupPeaks_native}
\alias{RamanMethod_GroupPeaks_native}
\title{RamanMethod_GroupPeaks_native Class}
\usage{
RamanMethod_GroupPeaks_native(fillMissingPeaks = TRUE)
}
\arguments{
\item{fillMissingPeaks}{Logical (length 1) for filling missing peaks.}
}
\value{
A RamanMethod_GroupPeaks_native object.
}
\description{
Groups peaks across analyses.
}
