% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RamanMethod_InvertAxis.R
\name{RamanMethod_InvertAxis_intensity}
\alias{RamanMethod_InvertAxis_intensity}
\title{RamanMethod_InvertAxis_intensity Class}
\usage{
RamanMethod_InvertAxis_intensity()
}
\value{
A RamanMethod_InvertAxis_intensity object.
}
\description{
Inverts the intensity axis of the spectra.
}
