% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RamanMethod_NormalizeSpectra.R
\name{RamanMethod_NormalizeSpectra_meancenter}
\alias{RamanMethod_NormalizeSpectra_meancenter}
\title{\strong{RamanMethod_NormalizeSpectra_meancenter}}
\usage{
RamanMethod_NormalizeSpectra_meancenter()
}
\value{
A RamanMethod_NormalizeSpectra_meancenter object.
}
\description{
Normalizes spectra using mean centering.
}
