% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RamanMethod_NormalizeSpectra.R
\name{RamanMethod_NormalizeSpectra_scale}
\alias{RamanMethod_NormalizeSpectra_scale}
\title{\strong{RamanMethod_NormalizeSpectra_scale}}
\usage{
RamanMethod_NormalizeSpectra_scale()
}
\value{
A RamanMethod_NormalizeSpectra_scale object.
}
\description{
Normalizes spectra using scaling based on the standard deviation.
}
