% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RamanMethod_ScaleTime.R
\name{RamanMethod_ScaleTime_native}
\alias{RamanMethod_ScaleTime_native}
\title{\strong{RamanMethod_ScaleTime_native}}
\usage{
RamanMethod_ScaleTime_native(value = 0)
}
\arguments{
\item{value}{Numeric (length 1) with the multiplier to scale the time variable.}
}
\value{
A RamanMethod_ScaleTime_native object.
}
\description{
Scales the time variable of the spectra using the multiplier \code{value}.
}
