% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RamanMethod_SmoothSpectra.R
\name{RamanMethod_SmoothSpectra_movingaverage}
\alias{RamanMethod_SmoothSpectra_movingaverage}
\title{\strong{RamanMethod_SmoothSpectra_movingaverage}}
\usage{
RamanMethod_SmoothSpectra_movingaverage(windowSize = 5)
}
\arguments{
\item{windowSize}{Numeric (length 1) with the window size for the moving average.}
}
\value{
A RamanMethod_SmoothSpectra_movingaverage object.
}
\description{
Smooths spectra using the moving average algorithm.
}
