% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RamanResults_Spectra.R
\name{RamanResults_Spectra}
\alias{RamanResults_Spectra}
\alias{validate_object.RamanResults_Spectra}
\alias{show.RamanResults_Spectra}
\alias{[.RamanResults_Spectra}
\alias{[[.RamanResults_Spectra}
\title{Constructor and methods to handle Raman spectra results}
\usage{
RamanResults_Spectra(
  spectra = list(),
  is_averaged = FALSE,
  peaks = list(),
  chrom_peaks = list()
)

\method{validate_object}{RamanResults_Spectra}(x)

\method{show}{RamanResults_Spectra}(x)

\method{[}{RamanResults_Spectra}(x, i)

\method{[[}{RamanResults_Spectra}(x, i)
}
\arguments{
\item{spectra}{A list of spectra data, where each element is a \code{data.table} representing a Raman spectrum.}

\item{is_averaged}{A logical value indicating whether the spectra are averaged.}

\item{peaks}{A list of peaks, where each element is a \code{data.table} representing the peaks for a corresponding spectrum.}

\item{chrom_peaks}{A list where each element is a \code{data.table} representing the chromatographic peaks found in each spectrum, when a time dimension is present.}

\item{x}{A \code{RamanResults_Spectra} object.}

\item{i}{A character or numeric vector to perform subsetting. Check the specific method for details.}
}
\description{
The \code{RamanResults_Spectra} class is used to store Raman spectra results, including peak information and whether the spectra are averaged.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate_object(RamanResults_Spectra)}: Validate the RamanResults_Spectra object, returns NULL if valid.

\item \code{show(RamanResults_Spectra)}: Show the RamanResults_Spectra object.

\item \code{[}: Subset the \code{RamanResults_Spectra} object by analyses. The argument \code{i} can be a character vector with the names of the analyses or a numeric vector with the indices of the analyses to keep.

\item \code{[[}: Extract a single analysis from the \code{RamanResults_Spectra} object. The argument \code{i} can be a character vector with the name of the analysis or a numeric vector with the index of the analysis to extract.

}}
\seealso{
\link{Results} for the parent class.
}
