% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RamanResults_SpectraSimilarity.R
\name{RamanResults_SpectraSimilarity}
\alias{RamanResults_SpectraSimilarity}
\alias{validate_object.RamanResults_SpectraSimilarity}
\alias{show.RamanResults_SpectraSimilarity}
\title{Constructor and methods to handle Raman spectra similarity results}
\usage{
RamanResults_SpectraSimilarity(data = list())

\method{validate_object}{RamanResults_SpectraSimilarity}(x)

\method{show}{RamanResults_SpectraSimilarity}(x)
}
\arguments{
\item{data}{A list of data tables, where each table contains the results of a similarity analysis for a Raman spectrum.}

\item{x}{A \code{RamanResults_SpectraSimilarity} object.}
}
\description{
The \code{RamanResults_SpectraSimilarity} class is used to store results of similarity analysis between Raman spectra.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate_object(RamanResults_SpectraSimilarity)}: Validate the RamanResults_SpectraSimilarity object, returns NULL if valid.

\item \code{show(RamanResults_SpectraSimilarity)}: Show the RamanResults_SpectraSimilarity object.

}}
