% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RamanSettings_AverageSpectra.R
\name{RamanSettings_AverageSpectra_native}
\alias{RamanSettings_AverageSpectra_native}
\title{\strong{RamanSettings_AverageSpectra_native}}
\usage{
RamanSettings_AverageSpectra_native(by = "replicates")
}
\arguments{
\item{by}{Character (length 1) with the grouping variable for averaging. Possible variables are
\code{replicates}, \code{chrom_peaks}, \code{rt}, \code{replicates+chrom_peaks}, \code{replicates+rt}, \code{chrom_peaks+rt},
\code{replicates+chrom_peaks+rt}.}
}
\value{
A RamanSettings_AverageSpectra_native object.
}
\description{
Averages spectra based on variables.
}
