% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RamanSettings_BinSpectra.R
\name{RamanSettings_BinSpectra_StreamFind}
\alias{RamanSettings_BinSpectra_StreamFind}
\title{\strong{RamanSettings_BinSpectra_StreamFind}}
\usage{
RamanSettings_BinSpectra_StreamFind(
  unitsVal = NULL,
  unitsNumber = NULL,
  bins = NULL,
  refBinAnalysis = NULL
)
}
\arguments{
\item{unitsVal}{Character of length one with the column name of the variable to be used for binning.}

\item{unitsNumber}{Integer of length one with the number of units to be used for binning.}

\item{bins}{Named list of numeric values with the bin dimensions.}

\item{refBinAnalysis}{The analysis index to use a reference for creating the bins.}
}
\value{
A RamanSettings_BinSpectra_StreamFind object.
}
\description{
Bins spectral data according to units of a given variable (e.g., 5 retention time values) or based on
bins given as a named list of numeric values, where the names are the bin labels (i.e. the name of the column) and
the values are the bin dimensions (e.g. 5 seconds).
}
