% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RamanSettings_CorrectSpectraBaseline.R
\name{RamanSettings_CorrectSpectraBaseline_baseline}
\alias{RamanSettings_CorrectSpectraBaseline_baseline}
\title{\strong{RamanSettings_CorrectSpectraBaseline_baseline}}
\usage{
RamanSettings_CorrectSpectraBaseline_baseline(
  method = "als",
  args = list(lambda = 5, p = 0.05, maxit = 10)
)
}
\arguments{
\item{method}{Character (length 1) with the method to be used for baseline correction. Possible values are "als",
"fillPeaks", "irls", "lowpass", "medianWindow", "modpolyfit", "peakDetection", "rfbaseline", "rollingBall", "shirley"
and "TAP".}

\item{args}{List with additional arguments for the selected method. See the documentation of the \pkg{baseline} package
for more details.}
}
\value{
A RamanSettings_CorrectSpectraBaseline_baseline object.
}
\description{
Performs baseline correction to spectra using the \pkg{baseline} package.
}
