% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RamanSettings_DeleteSpectraSection.R
\name{RamanSettings_DeleteSpectraSection_native}
\alias{RamanSettings_DeleteSpectraSection_native}
\title{\strong{RamanSettings_DeleteSpectraSection_native}}
\usage{
RamanSettings_DeleteSpectraSection_native(min = NULL, max = NULL)
}
\arguments{
\item{min}{Numeric vector (length 1) with the minimum shift value to delete.}

\item{max}{Numeric vector (length 1) with the maximum shift value to delete.}
}
\value{
A RamanSettings_DeleteSpectraSection_native object.
}
\description{
Deletes a section of the spectra between \emph{shift} minimum and maximum values.
}
