% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RamanSettings_DeleteSpectraSection.R
\name{RamanSettings_DeleteSpectraSection_StreamFind}
\alias{RamanSettings_DeleteSpectraSection_StreamFind}
\title{\strong{RamanSettings_DeleteSpectraSection_StreamFind}}
\usage{
RamanSettings_DeleteSpectraSection_StreamFind(
  shiftmin = NULL,
  shiftmax = NULL,
  rtmin = NULL,
  rtmax = NULL
)
}
\arguments{
\item{shiftmin}{Numeric vector (length 1) with the minimum shift value to delete.}

\item{shiftmax}{Numeric vector (length 1) with the maximum shift value to delete.}

\item{rtmin}{Numeric vector (length 1) with the minimum retention time value to delete.}

\item{rtmax}{Numeric vector (length 1) with the maximum retention time value to delete.}
}
\value{
A RamanSettings_DeleteSpectraSection_StreamFind object.
}
\description{
Deletes a section of the spectra between \emph{shift} and/or \emph{rt} minimum and maximum values.
}
