% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RamanSettings_FilterChromPeaks.R
\name{RamanSettings_FilterChromPeaks_native}
\alias{RamanSettings_FilterChromPeaks_native}
\title{\strong{RamanSettings_FilterChromPeaks_native}}
\usage{
RamanSettings_FilterChromPeaks_native(minIntensity = 0)
}
\arguments{
\item{minIntensity}{Numeric (length 1) with the minimum intensity to keep a chromatographic peak.}
}
\value{
A RamanSettings_FilterChromPeaks_native object.
}
\description{
Settings for filtering chromatographic peaks.
}
