% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RamanSettings_FilterSpectra.R
\name{RamanSettings_FilterSpectra_StreamFind}
\alias{RamanSettings_FilterSpectra_StreamFind}
\title{\strong{RamanSettings_FilterSpectra_StreamFind}}
\usage{
RamanSettings_FilterSpectra_StreamFind(onlyChromPeaksSpectra = FALSE)
}
\arguments{
\item{onlyChromPeaksSpectra}{Logical (length 1) for keeping only spectra with chromatographic peaks.}
}
\value{
A RamanSettings_FilterSpectra_StreamFind object.
}
\description{
Various filter options for Raman spectra.
}
