% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RamanSettings_FindChromatogramsMaxima.R
\name{RamanSettings_FindChromatogramsMaxima_StreamFind}
\alias{RamanSettings_FindChromatogramsMaxima_StreamFind}
\title{\strong{RamanSettings_FindChromatogramsMaxima_StreamFind}}
\usage{
RamanSettings_FindChromatogramsMaxima_StreamFind(
  minWidth = 0,
  maxWidth = 0,
  minHeight = 0
)
}
\arguments{
\item{minWidth}{Numeric (length 1) with the minimum width of a peak.}

\item{maxWidth}{Numeric (length 1) with the maximum width of a peak.}

\item{minHeight}{Numeric (length 1) with the minimum height of a peak.}
}
\value{
A RamanSettings_FindChromatogramsMaxima_StreamFind object.
}
\description{
Finds maximum peaks in continuous spectra.
}
