% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/class_RamanSettings_FindChromPeaks_LocalMaxima.R
\name{RamanSettings_FindChromPeaks_LocalMaxima}
\alias{RamanSettings_FindChromPeaks_LocalMaxima}
\title{\strong{RamanSettings_FindChromPeaks_LocalMaxima}}
\usage{
RamanSettings_FindChromPeaks_LocalMaxima(
  minWidth = 0,
  maxWidth = 0,
  minHeight = 0
)
}
\arguments{
\item{minWidth}{Numeric (length 1) with the minimum width of a peak.}

\item{maxWidth}{Numeric (length 1) with the maximum width of a peak.}

\item{minHeight}{Numeric (length 1) with the minimum height of a peak.}
}
\value{
A RamanSettings_FindChromPeaks_LocalMaxima object.
}
\description{
Finds peak maxima in the chromatographic dimension of multiple Raman spectra.
}
