% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RamanSettings_GroupPeaks.R
\name{RamanSettings_GroupPeaks_native}
\alias{RamanSettings_GroupPeaks_native}
\title{\strong{RamanSettings_GroupPeaks_native}}
\usage{
RamanSettings_GroupPeaks_native(fillMissingPeaks = TRUE)
}
\arguments{
\item{fillMissingPeaks}{Logical (length 1) for filling missing peaks.}
}
\value{
A RamanSettings_GroupPeaks_native object.
}
\description{
Groups peaks across analyses.
}
