% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RamanSettings_InvertAxis.R
\name{RamanSettings_InvertAxis_intensity}
\alias{RamanSettings_InvertAxis_intensity}
\title{\strong{RamanSettings_InvertAxis_intensity}}
\usage{
RamanSettings_InvertAxis_intensity()
}
\value{
A RamanSettings_InvertAxis_intensity object.
}
\description{
Inverts the intensity axis of the spectra.
}
