% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_S3_RamanSettings.R
\name{RamanSettings_MergeSpectraTimeSeries_StreamFind}
\alias{RamanSettings_MergeSpectraTimeSeries_StreamFind}
\title{RamanSettings_MergeSpectraTimeSeries_StreamFind}
\usage{
RamanSettings_MergeSpectraTimeSeries_StreamFind(preCut = 2)
}
\arguments{
\item{preCut}{The number of pre Raman scans to exclude when merging.}
}
\value{
A ProcessingSettings S3 class object with subclass RamanSettings_MergeSpectraTimeSeries_StreamFind.
}
\description{
Merges Raman spectra based on time series data. It collapses data files into a single file.
}
