% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_S3_RamanSettings.R
\name{RamanSettings_NormalizeSpectra_blockweight}
\alias{RamanSettings_NormalizeSpectra_blockweight}
\title{RamanSettings_NormalizeSpectra_blockweight}
\usage{
RamanSettings_NormalizeSpectra_blockweight()
}
\value{
A ProcessingSettings S3 class object with subclass RamanSettings_NormalizeSpectra_blockweight.
}
\description{
Normalizes spectra using block weighting for downstream data evaluation.
}
