% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RamanSettings_NormalizeSpectra.R
\name{RamanSettings_NormalizeSpectra_meancenter}
\alias{RamanSettings_NormalizeSpectra_meancenter}
\title{\strong{RamanSettings_NormalizeSpectra_meancenter}}
\usage{
RamanSettings_NormalizeSpectra_meancenter()
}
\value{
A RamanSettings_NormalizeSpectra_meancenter object.
}
\description{
Normalizes spectra using mean centering.
}
