% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_S3_RamanSettings.R
\name{RamanSettings_NormalizeSpectra_minmax}
\alias{RamanSettings_NormalizeSpectra_minmax}
\title{RamanSettings_NormalizeSpectra_minmax}
\usage{
RamanSettings_NormalizeSpectra_minmax()
}
\value{
A ProcessingSettings S3 class object with subclass RamanSettings_NormalizeSpectra_minmax.
}
\description{
Normalizes spectra using the min-max algorithm.
}
