% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RamanSettings_NormalizeSpectra.R
\name{RamanSettings_NormalizeSpectra_scale}
\alias{RamanSettings_NormalizeSpectra_scale}
\title{\strong{RamanSettings_NormalizeSpectra_scale}}
\usage{
RamanSettings_NormalizeSpectra_scale()
}
\value{
A RamanSettings_NormalizeSpectra_scale object.
}
\description{
Normalizes spectra using scaling based on the standard deviation.
}
