% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RamanSettings_NormalizeSpectra.R
\name{RamanSettings_NormalizeSpectra_snv}
\alias{RamanSettings_NormalizeSpectra_snv}
\title{\strong{RamanSettings_NormalizeSpectra_snv}}
\usage{
RamanSettings_NormalizeSpectra_snv(liftTozero = FALSE)
}
\arguments{
\item{liftTozero}{Logical (length 1) indicating if the spectra should be lifted to zero.}
}
\value{
A RamanSettings_NormalizeSpectra_snv object.
}
\description{
Normalizes spectra using the Standard Normal Variate (SNV) algorithm.
}
