% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RamanSettings_RemoveCosmicRays.R
\name{RamanSettings_RemoveCosmicRays_native}
\alias{RamanSettings_RemoveCosmicRays_native}
\title{\strong{RamanSettings_RemoveCosmicRays_native}}
\usage{
RamanSettings_RemoveCosmicRays_native()
}
\value{
A RamanSettings_RemoveCosmicRays_native object.
}
\description{
Removes cosmic rays from Raman spectra.
}
