% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RamanSettings_ScaleTime.R
\name{RamanSettings_ScaleTime_native}
\alias{RamanSettings_ScaleTime_native}
\title{\strong{RamanSettings_ScaleTime_native}}
\usage{
RamanSettings_ScaleTime_native(value = 0)
}
\arguments{
\item{value}{Numeric (length 1) with the multiplier to scale the time variable.}
}
\value{
A RamanSettings_ScaleTime_native object.
}
\description{
Scales the time variable of the spectra using the multiplier \code{value}.
}
