% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RamanSettings_SmoothSpectra.R
\name{RamanSettings_SmoothSpectra_movingaverage}
\alias{RamanSettings_SmoothSpectra_movingaverage}
\title{\strong{RamanSettings_SmoothSpectra_movingaverage}}
\usage{
RamanSettings_SmoothSpectra_movingaverage(windowSize = 5)
}
\arguments{
\item{windowSize}{Numeric (length 1) with the window size for the moving average.}
}
\value{
A RamanSettings_SmoothSpectra_movingaverage object.
}
\description{
Smooths spectra using the moving average algorithm.
}
