% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_S3_RamanSettings.R
\name{RamanSettings_SmoothSpectra_savgol}
\alias{RamanSettings_SmoothSpectra_savgol}
\title{RamanSettings_SmoothSpectra_savgol}
\usage{
RamanSettings_SmoothSpectra_savgol(fl = 11, forder = 4, dorder = 0)
}
\arguments{
\item{fl}{Numeric (length 1) with the filter length (for instance fl = 51..151), has to be odd.}

\item{forder}{Numeric (length 1) with the order of the filter (2 = quadratic filter, 4 = quartic).}

\item{dorder}{Numeric (length 1) with the order of the derivative (0 = smoothing, 1 = first derivative, etc.).}
}
\value{
A ProcessingSettings S3 class object with subclass RamanSettings_SmoothSpectra_savgol.
}
\description{
Smooths spectra using the Savitzky-Golay algorithm from the \pkg{pracma} package.
}
