% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RamanSettings_SubtractBlankSpectra.R
\name{RamanSettings_SubtractBlankSpectra_StreamFind}
\alias{RamanSettings_SubtractBlankSpectra_StreamFind}
\title{\strong{RamanSettings_SubtractBlankSpectra_StreamFind}}
\usage{
RamanSettings_SubtractBlankSpectra_StreamFind(negativeToZero = FALSE)
}
\arguments{
\item{negativeToZero}{Logical (length 1) indicating if negative values should be set to zero.}
}
\value{
A RamanSettings_SubtractBlankSpectra_StreamFind object.
}
\description{
Subtracts the blank spectra to each analysis according to the blank assignment.
}
