% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RamanSettings_SubtractScansSection.R
\name{RamanSettings_SubtractScansSection_native}
\alias{RamanSettings_SubtractScansSection_native}
\title{\strong{RamanSettings_SubtractScansSection_native}}
\usage{
RamanSettings_SubtractScansSection_native(sectionWindow = c(10, 200))
}
\arguments{
\item{sectionWindow}{Numeric (length 2) with the window for the sectioning.}
}
\value{
A RamanSettings_SubtractScansSection_native object.
}
\description{
Subtracts the spectra of a section of the scans to all other spectra.
}
