% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_ms_calculate_spectra_charges}
\alias{rcpp_ms_calculate_spectra_charges}
\title{Calculate charge states for mass spectra}
\usage{
rcpp_ms_calculate_spectra_charges(
  spectra,
  round_val = 35,
  rel_low_cut = 0.2,
  abs_low_cut = 300,
  top_charges = 5L
)
}
\arguments{
\item{spectra}{data.frame with: analysis, replicate, id, polarity, rt, mz, intensity}

\item{round_val}{m/z rounding denominator for clustering (e.g. 35)}

\item{rel_low_cut}{relative intensity cut (fraction of base peak)}

\item{abs_low_cut}{absolute intensity cut}

\item{top_charges}{number of top anchor candidates}
}
\value{
data.frame with columns: analysis, replicate, id, polarity, mz, intensity,
cluster_mz, z, mass, rt
}
\description{
Groups input rows by (analysis, id), averages spectra across scans per group,
then calls the C++ charge-state algorithm.
}
