% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_ms_correct_baseline_airpls}
\alias{rcpp_ms_correct_baseline_airpls}
\title{Correct chromatogram baselines using airPLS}
\usage{
rcpp_ms_correct_baseline_airpls(
  chromatograms,
  lambda = 10,
  differences = 1L,
  itermax = 20L
)
}
\arguments{
\item{chromatograms}{data.frame with at least: analysis, id, rt, intensity}

\item{lambda}{smoothing penalty}

\item{differences}{order of difference operator (1 or 2)}

\item{itermax}{maximum iterations}
}
\value{
data.frame with updated intensity, and added/updated baseline + raw columns
}
\description{
Correct chromatogram baselines using airPLS
}
