% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_ms_correct_baseline_als}
\alias{rcpp_ms_correct_baseline_als}
\title{Correct chromatogram baselines using Asymmetric Least Squares (ALS)}
\usage{
rcpp_ms_correct_baseline_als(
  chromatograms,
  lambda = 1e+05,
  p = 0.001,
  maxit = 10L
)
}
\arguments{
\item{chromatograms}{data.frame with at least: analysis, id, rt, intensity}

\item{lambda}{smoothing penalty}

\item{p}{asymmetry parameter (0 < p < 0.5; typical 0.001–0.01)}

\item{maxit}{maximum iterations}
}
\value{
data.frame with updated intensity, and added/updated baseline + raw columns
}
\description{
Correct chromatogram baselines using Asymmetric Least Squares (ALS)
}
