% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_ms_deconvolute_spectra}
\alias{rcpp_ms_deconvolute_spectra}
\title{Deconvolute charge-state spectra to neutral-mass spectra}
\usage{
rcpp_ms_deconvolute_spectra(spectra, charges, clust_val = 0.1, window = 20)
}
\arguments{
\item{spectra}{data.frame with: analysis, replicate, id, polarity, rt, mz, intensity}

\item{charges}{data.frame from rcpp_ms_calculate_spectra_charges:
analysis, id, polarity, mz, z, cluster_mz, mass}

\item{clust_val}{mass clustering tolerance (Da)}

\item{window}{m/z window around each charge to extract (half-width in m/z units)}
}
\value{
data.frame with same schema as spectra but mz column replaced by mass
}
\description{
Deconvolute charge-state spectra to neutral-mass spectra
}
