% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_ms_find_chrom_peaks_local_maxima}
\alias{rcpp_ms_find_chrom_peaks_local_maxima}
\title{Find chromatographic peaks using local-maxima method}
\usage{
rcpp_ms_find_chrom_peaks_local_maxima(
  chromatograms,
  min_height = 0,
  min_width = 0,
  max_width = 0
)
}
\arguments{
\item{chromatograms}{data.frame with analysis, replicate, index, id, polarity,
pre_mz, pre_ce, pro_mz, rt, intensity}

\item{min_height}{minimum peak height}

\item{min_width}{minimum peak width (rt units; 0 = no filter)}

\item{max_width}{maximum peak width (rt units; 0 = no filter)}
}
\value{
data.frame of peaks with schema matching MassSpecResults_Chromatograms Peaks table
}
\description{
Find chromatographic peaks using local-maxima method
}
