% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_ms_find_spectra_maxima}
\alias{rcpp_ms_find_spectra_maxima}
\title{Find peaks/maxima in (deconvoluted) spectra}
\usage{
rcpp_ms_find_spectra_maxima(
  spectra,
  min_height = 0,
  min_width = 0,
  max_width = 0
)
}
\arguments{
\item{spectra}{data.frame with: analysis, replicate, id, polarity, rt, mass, intensity}

\item{min_height}{minimum peak height}

\item{min_width}{minimum width in mass units (0 = no filter)}

\item{max_width}{maximum width in mass units (0 = no filter)}
}
\value{
data.frame of spectral peaks
}
\description{
Operates identically to rcpp_ms_find_chrom_peaks_local_maxima but on the
mass/intensity axis of spectra.
}
