% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_ms_integrate_chromatograms}
\alias{rcpp_ms_integrate_chromatograms}
\title{Find chromatographic peaks and integrate using pracma-style algorithm}
\usage{
rcpp_ms_integrate_chromatograms(
  chromatograms,
  merge = TRUE,
  close_by_threshold = 45,
  min_peak_height = 0,
  min_peak_distance = 10,
  min_peak_width = 5,
  max_peak_width = 120,
  min_sn = 10
)
}
\arguments{
\item{chromatograms}{data.frame (same columns as above)}

\item{merge}{merge close-by peaks?}

\item{close_by_threshold}{merge distance in rt units}

\item{min_peak_height}{minimum peak height}

\item{min_peak_distance}{minimum distance between peaks (rt units)}

\item{min_peak_width}{minimum peak width (rt units)}

\item{max_peak_width}{maximum peak width (rt units)}

\item{min_sn}{minimum signal-to-noise}
}
\value{
data.frame of peaks
}
\description{
Find chromatographic peaks and integrate using pracma-style algorithm
}
