% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_ms_smooth_chromatograms}
\alias{rcpp_ms_smooth_chromatograms}
\title{Smooth chromatograms using moving-average or Savitzky-Golay filter}
\usage{
rcpp_ms_smooth_chromatograms(
  chromatograms,
  method = "movingaverage",
  window_size = 5L,
  fl = 11L,
  forder = 4L,
  dorder = 0L
)
}
\arguments{
\item{chromatograms}{data.frame with columns: analysis, replicate, index, id,
polarity, pre_mz, pre_ce, pro_mz, rt, intensity}

\item{method}{"movingaverage" or "savgol"}

\item{window_size}{window size (odd integer) for moving average}

\item{fl}{filter length (odd) for SG}

\item{forder}{polynomial order for SG}

\item{dorder}{derivative order for SG (0 = smoothing)}
}
\value{
data.frame with updated intensity column
}
\description{
Smooth chromatograms using moving-average or Savitzky-Golay filter
}
