% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{read}
\alias{read}
\title{Generic \code{read}}
\usage{
read(x, ...)
}
\arguments{
\item{x}{An object from which the data will be read.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S7 generic to read data from \code{x}.
}
