% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_S3.R
\name{remove}
\alias{remove}
\title{Generic \code{remove}}
\usage{
remove(x, ...)
}
\arguments{
\item{x}{An object from which the data will be removed.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S3 generic to remove data from \code{x}.
}
