% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_007_filterFeatures.R
\name{removeFilteredFeatures}
\alias{removeFilteredFeatures}
\title{removeFilteredFeatures}
\usage{
removeFilteredFeatures(object, which = "all")
}
\arguments{
\item{object}{An \linkS4class{msFeatures} object.}

\item{which}{A character vector with the filter tag/s of features to be removed.
The default is \emph{all} to remove all filtered features.}
}
\description{
Function to remove permanently filtered features from a
\linkS4class{msData} or \linkS4class{msFeatures} object.
}
