% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_000_generics.R
\name{replicateNames<-}
\alias{replicateNames<-}
\title{replicateNames<-}
\usage{
replicateNames(object) <- value
}
\arguments{
\item{object}{A method specific class object.}

\item{value}{A method specific vector with names.}
}
\description{
Generic for setting replicate names.
}
