% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_000_generics.R
\name{replicateNames}
\alias{replicateNames}
\title{replicateNames}
\usage{
replicateNames(object)
}
\arguments{
\item{object}{A method specific class object.}
}
\description{
Generic for getting replicate names.
}
