% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_S3.R
\name{report}
\alias{report}
\title{Generic \code{report}}
\usage{
report(x, ...)
}
\arguments{
\item{x}{An object from which the report will be generated.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S3 generic to generate a report from \code{x}.
}
