% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Results.R
\name{Results}
\alias{Results}
\alias{validate_object.Results}
\alias{query_db.Results}
\alias{list_db_tables.Results}
\alias{get_db_table_info.Results}
\title{Base class for database-backed Results}
\usage{
Results(projectPath = ".")

\method{validate_object}{Results}(x)

\method{query_db}{Results}(x, sql, params = NULL)

\method{list_db_tables}{Results}(x)

\method{get_db_table_info}{Results}(x, tableName)
}
\arguments{
\item{projectPath}{Path to the project directory where this database file will be created/used.}

\item{x}{A \code{Results} object.}

\item{sql}{SQL query string to execute against the database.}

\item{params}{Query parameters to be used with the SQL query (optional).}

\item{tableName}{Name of the table to query or retrieve information about.}
}
\value{
An object of class \code{Results}.
}
\description{
Minimal base class to represent results stored in a DuckDB file.
Used internally to expose shared S3 helpers for DB-backed results classes.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate_object(Results)}: Validate the Results object.

\item \code{query_db(Results)}: Internal: execute a query on the DB.

\item \code{list_db_tables(Results)}: Internal: list tables in the DB.

\item \code{get_db_table_info(Results)}: Internal: get table info from the DB.

}}
