% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ResultsDB.R
\name{ResultsDB}
\alias{ResultsDB}
\alias{validate_object.ResultsDB}
\alias{query_db.ResultsDB}
\alias{list_db_tables.ResultsDB}
\alias{get_db_table_info.ResultsDB}
\title{Base class for database-backed Results}
\usage{
ResultsDB(db, data_type)

\method{validate_object}{ResultsDB}(x)

\method{query_db}{ResultsDB}(x, sql, params = NULL)

\method{list_db_tables}{ResultsDB}(x)

\method{get_db_table_info}{ResultsDB}(x, tableName)
}
\arguments{
\item{db}{Path to DuckDB file.}

\item{data_type}{Data type string (e.g., "MassSpec").}

\item{x}{ResultsDB object.}

\item{sql}{SQL query string to execute against the database.}

\item{params}{Query parameters to be used with the SQL query (optional).}

\item{tableName}{Name of the table to query or retrieve information about.}
}
\value{
An object of class \code{ResultsDB}.
}
\description{
Minimal base class to represent results stored in a DuckDB file.
Requires a \code{db} path and a \code{data_type} identifier.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate_object(ResultsDB)}: Validate the ResultsDB object.

\item \code{query_db(ResultsDB)}: Internal: execute a query on the DB.

\item \code{list_db_tables(ResultsDB)}: Internal: list tables in the DB.

\item \code{get_db_table_info(ResultsDB)}: Internal: get table info from the DB.

}}
