% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{run}
\alias{run}
\title{Generic \code{run}}
\usage{
run(x, ...)
}
\arguments{
\item{x}{An object on which the process will be run.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S7 generic to run a process on \code{x}.
}
