% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{save}
\alias{save}
\title{Generic \code{save}}
\usage{
save(x, ...)
}
\arguments{
\item{x}{An object from which the data will be saved.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S7 generic to save data from \code{x}.
}
