% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_S3.R
\name{save_cache}
\alias{save_cache}
\title{Generic \code{save_cache}}
\usage{
save_cache(x, ...)
}
\arguments{
\item{x}{An object from which the cache will be saved.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S3 generic to save cache from \code{x}.
}
