% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_005_default_ProcessingSettings.R
\name{save_default_ProcessingSettings}
\alias{save_default_ProcessingSettings}
\title{save_default_ProcessingSettings}
\usage{
save_default_ProcessingSettings(
  call = NA_character_,
  software = NA_character_,
  algorithm = NA_character_,
  format = "json",
  name = "settings",
  path = getwd()
)
}
\arguments{
\item{call}{X.}

\item{software}{X.}

\item{algorithm}{X.}

\item{format}{X.}

\item{name}{X.}

\item{path}{X.}
}
\value{
Creates a json/rds files on the defined path.
}
\description{
X.
}
