% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_S3.R
\name{set_analysis_classes}
\alias{set_analysis_classes}
\title{Generic \code{set_analysis_classes}}
\usage{
set_analysis_classes(x, ...)
}
\arguments{
\item{x}{An object in which the analysis classes will be set.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S3 generic to set analysis classes in \code{x}.
}
