% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_S3.R
\name{set_blank_names}
\alias{set_blank_names}
\title{Generic \code{set_blank_names}}
\usage{
set_blank_names(x, ...)
}
\arguments{
\item{x}{An object in which the blanks will be set.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S3 generic to set blanks from \code{x}.
}
