% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_S3.R
\name{set_concentrations}
\alias{set_concentrations}
\title{Generic \code{set_concentrations}}
\usage{
set_concentrations(x, ...)
}
\arguments{
\item{x}{An object in which the concentrations will be set.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S3 generic to set concentrations in \code{x}.
}
