% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_S3.R
\name{set_replicates}
\alias{set_replicates}
\title{Generic \code{set_replicates}}
\usage{
set_replicates(x, ...)
}
\arguments{
\item{x}{An object in which the replicates will be set.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S3 generic to set replicates in \code{x}.
}
