% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_class_000_general.R
\docType{class}
\name{settings-class}
\alias{settings-class}
\alias{getAlgorithm,settings-method}
\alias{getSettings,settings-method}
\alias{getCall,settings-method}
\alias{exportSettings,settings-method}
\title{settings}
\usage{
\S4method{getAlgorithm}{settings}(object)

\S4method{getSettings}{settings}(object)

\S4method{getCall}{settings}(x)

\S4method{exportSettings}{settings}(object, name = "settings", format = "rds", path = getwd())
}
\arguments{
\item{object}{A \linkS4class{settings} object.}

\item{x}{A \linkS4class{settings} object.}

\item{name}{A character string with the file name.}

\item{format}{A character string with the format ("rds" (default) or "json").}

\item{path}{The directory path to save the file.}
}
\description{
An S4 class object with parameter settings for a given
data processing step of the basic workflow in \pkg{streamFind}.
The \code{settings} object contains the algorithm to be used for a given processing step
and the list of respective parameter settings.
}
\section{Methods (by generic)}{
\itemize{
\item \code{getAlgorithm(settings)}: getter for algorithm.

\item \code{getSettings(settings)}: getter for settings.

\item \code{getCall(settings)}: getter for function call.

\item \code{exportSettings(settings)}: exports the settings as either JSON or rds.

}}
\section{Slots}{

\describe{
\item{\code{call}}{The name of the function where the algorithm and settings are used.}

\item{\code{algorithm}}{A character string with the name of the algorithm to be used.}

\item{\code{settings}}{A list of settings dependent on the algorithm used.}
}}

