% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_S3_headers.R, R/class_S3_settings.R
\name{as.settings}
\alias{as.settings}
\alias{settings}
\title{Settings S3 Class Constructor}
\usage{
as.settings(settings)

settings(call = NA_character_, algorithm = NA_character_, parameters = NULL)

as.settings(settings)
}
\arguments{
\item{call}{Character of length one with the name of the method where the
settings are to be applied.}

\item{algorithm}{Character of length one with the name of the algorithm to
be used.}

\item{parameters}{A list with parameter settings specific for the \code{call}
method and used \code{algorithm}.}
}
\value{
A settings S3 class
}
\description{
Creates a settings S3 class.
}
\details{
See the method documentation for more information about possible
algorithms and parameters to be used.
}
