% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_005_default_ProcessingSettings.R
\name{settings_annotate_features_streamFind}
\alias{settings_annotate_features_streamFind}
\title{settings_annotate_features_streamFind}
\usage{
settings_annotate_features_streamFind(
  maxIsotopes = 5,
  maxCharge = 1,
  rtWindowAlignment = 0.2,
  maxGaps = 1,
  runParallel = FALSE
)
}
\value{
X.
}
\description{
X.
}
