% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ProcessingSettings.R
\name{Settings_average_spectra_StreamFind}
\alias{Settings_average_spectra_StreamFind}
\title{Settings_average_spectra_StreamFind}
\usage{
Settings_average_spectra_StreamFind(collapseTime = FALSE)
}
\arguments{
\item{collapseTime}{Logical (length 1). When \code{TRUE} the spectra are averaged, reducing the time variable.}
}
\value{
A ProcessingSettings S3 class object with subclass Settings_average_spectra_StreamFind.
}
\description{
Averages spectra based on analysis replicate groups.
}
