% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_default_ProcessingSettings.R
\name{Settings_bin_spectra_qBinning}
\alias{Settings_bin_spectra_qBinning}
\alias{validate.Settings_bin_spectra_qBinning}
\title{Settings_bin_spectra_qBinning}
\usage{
Settings_bin_spectra_qBinning()

\method{validate}{Settings_bin_spectra_qBinning}(x)
}
\arguments{
\item{x}{A Settings_bin_spectra_qBinning S3 class object.}
}
\value{
X.
}
\description{
Not yet implemented.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate(Settings_bin_spectra_qBinning)}: Validates the object structure, returning a logical value of length one.

}}
