% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_default_ProcessingSettings.R
\name{Settings_bin_spectra_StreamFind}
\alias{Settings_bin_spectra_StreamFind}
\alias{validate.Settings_bin_spectra_StreamFind}
\title{Settings_bin_spectra_StreamFind}
\usage{
Settings_bin_spectra_StreamFind(
  unitsVal = NULL,
  unitsNumber = NULL,
  bins = NULL
)

\method{validate}{Settings_bin_spectra_StreamFind}(x)
}
\arguments{
\item{unitsVal}{}

\item{unitsNumber}{}

\item{bins}{}

\item{x}{A Settings_bin_spectra_StreamFind S3 class object.}
}
\value{
X.
}
\description{
Not yet implemented.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate(Settings_bin_spectra_StreamFind)}: Validates the object structure, returning a logical value of length one.

}}
