% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ProcessingSettings.R
\name{Settings_cluster_spectra_StreamFind}
\alias{Settings_cluster_spectra_StreamFind}
\title{Settings_cluster_spectra_StreamFind}
\usage{
Settings_cluster_spectra_StreamFind(
  val = "mz",
  clustVal = 0.001,
  presence = 0.1
)
}
\arguments{
\item{val}{Character (length 1) with the variable to be used for clustering.}

\item{clustVal}{Numeric (length 1) with the clustering value.}

\item{presence}{Numeric (length 1) with the minimum presence of traces in a cluster to be considered.}
}
\value{
A ProcessingSettings S3 class object with subclass Settings_cluster_spectra_StreamFind.
}
\description{
Clusters spectra based on a variable (i.e. column name).
}
