% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ProcessingSettings.R
\name{Settings_correct_chromatograms_baseline_airpls}
\alias{Settings_correct_chromatograms_baseline_airpls}
\alias{validate.Settings_correct_chromatograms_baseline_airpls}
\title{Settings_correct_chromatograms_baseline_airpls}
\usage{
Settings_correct_chromatograms_baseline_airpls(
  lambda = 10,
  differences = 1,
  itermax = 20
)

\method{validate}{Settings_correct_chromatograms_baseline_airpls}(x)
}
\arguments{
\item{lambda}{}

\item{differences}{}

\item{itermax}{}

\item{x}{A Settings_correct_chromatograms_baseline_airpls S3 class object.}
}
\value{
A ProcessingSettings S3 class object with subclass Settings_correct_chromatograms_baseline_airpls.
}
\description{
Prototype.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate(Settings_correct_chromatograms_baseline_airpls)}: Validates the object structure, returning a logical value of length one.

}}
