% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_default_ProcessingSettings.R
\name{Settings_correct_chromatograms_baseline_baseline}
\alias{Settings_correct_chromatograms_baseline_baseline}
\alias{validate.Settings_correct_chromatograms_baseline_baseline}
\title{Settings_correct_chromatograms_baseline_baseline}
\usage{
Settings_correct_chromatograms_baseline_baseline(
  method = "als",
  args = list(lambda = 5, p = 0.05, maxit = 10)
)

\method{validate}{Settings_correct_chromatograms_baseline_baseline}(x)
}
\arguments{
\item{method}{}

\item{args}{}

\item{x}{A Settings_correct_chromatograms_baseline_baseline S3 class object.}
}
\value{
A ProcessingSettings S3 class object with subclass Settings_correct_chromatograms_baseline_baseline.
}
\description{
Prototype.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate(Settings_correct_chromatograms_baseline_baseline)}: Validates the object structure, returning a logical value of length one.

}}
