% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ProcessingSettings.R
\name{Settings_correct_spectra_baseline_baseline}
\alias{Settings_correct_spectra_baseline_baseline}
\title{Settings_correct_spectra_baseline_baseline}
\usage{
Settings_correct_spectra_baseline_baseline(
  method = "als",
  args = list(lambda = 5, p = 0.05, maxit = 10)
)
}
\arguments{
\item{method}{Character (length 1) with the method to be used for baseline correction. Possible values are "als",
"fillPeaks", "irls", "lowpass", "medianWindow", "modpolyfit", "peakDetection", "rfbaseline", "rollingBall", "shirley"
and "TAP".}

\item{args}{List with additional arguments for the selected method. See the documentation of the \pkg{baseline} package
for more details.}
}
\value{
A ProcessingSettings S3 class object with subclass Settings_correct_spectra_baseline_baseline.
}
\description{
Performs baseline correction to spectra using the \pkg{baseline} package.
}
