% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_default_ProcessingSettings.R
\name{Settings_correct_spectra_baseline_StreamFind}
\alias{Settings_correct_spectra_baseline_StreamFind}
\alias{validate.Settings_correct_spectra_baseline_StreamFind}
\title{Settings_correct_spectra_baseline_StreamFind}
\usage{
Settings_correct_spectra_baseline_StreamFind(
  method = "als",
  args = list(lambda = 5, p = 0.05, maxit = 10)
)

\method{validate}{Settings_correct_spectra_baseline_StreamFind}(x)
}
\arguments{
\item{method}{}

\item{args}{}

\item{x}{A Settings_correct_spectra_baseline_StreamFind S3 class object.}
}
\value{
A ProcessingSettings S3 class object with subclass Settings_correct_spectra_baseline_StreamFind.
}
\description{
Prototype.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate(Settings_correct_spectra_baseline_StreamFind)}: Validates the object structure, returning a logical value of length one.

}}
