% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_default_ProcessingSettings.R
\name{Settings_deconvolute_spectra_charges_StreamFind}
\alias{Settings_deconvolute_spectra_charges_StreamFind}
\alias{validate.Settings_deconvolute_spectra_charges_StreamFind}
\title{Settings_deconvolute_spectra_charges_StreamFind}
\usage{
Settings_deconvolute_spectra_charges_StreamFind()

\method{validate}{Settings_deconvolute_spectra_charges_StreamFind}(x)
}
\arguments{
\item{x}{A Settings_integrate_chromatograms_StreamFind S3 class object.}
}
\value{
A ProcessingSettings S3 class object with subclass Settings_deconvolute_spectra_charges_StreamFind.
}
\description{
Prototype.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate(Settings_deconvolute_spectra_charges_StreamFind)}: Validates the object structure, returning a logical value of length one.

}}
