% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ProcessingSettings.R
\name{Settings_deconvolute_spectra_StreamFind}
\alias{Settings_deconvolute_spectra_StreamFind}
\title{Settings_deconvolute_spectra_StreamFind}
\usage{
Settings_deconvolute_spectra_StreamFind(clustVal = 0.1, window = 20)
}
\arguments{
\item{clustVal}{Numeric (length 1) with the clustering value for the charge deconvolution.}

\item{window}{Optional numeric (length 1) with the window in \emph{m/z} for collecting traces of a given charge.}
}
\value{
A ProcessingSettings S3 class object with subclass Settings_deconvolute_spectra_StreamFind.
}
\description{
Deconvolutes the spectral mass-to-charge ratio (\emph{m/z}) to mass (Da) after assignment of charges.
}
