% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_default_ProcessingSettings.R
\name{Settings_delete_spectra_section_StreamFind}
\alias{Settings_delete_spectra_section_StreamFind}
\alias{validate.Settings_delete_spectra_section_StreamFind}
\title{Settings_delete_spectra_section_StreamFind}
\usage{
Settings_delete_spectra_section_StreamFind(section = list())

\method{validate}{Settings_delete_spectra_section_StreamFind}(x)
}
\arguments{
\item{section}{}

\item{x}{A Settings_delete_spectra_section_StreamFind S3 class object.}
}
\value{
A ProcessingSettings S3 class object with subclass Settings_delete_spectra_section_StreamFind.
}
\description{
Prototype.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate(Settings_delete_spectra_section_StreamFind)}: Validates the object structure, returning a logical value of length one.

}}
