% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ProcessingSettings.R
\name{Settings_fill_features_StreamFind}
\alias{Settings_fill_features_StreamFind}
\title{Settings_fill_features_StreamFind}
\usage{
Settings_fill_features_StreamFind(rtExpand = 1, mzExpand = 5e-04)
}
\arguments{
\item{rtExpand}{Numeric (length 1) with the retention time values (in seconds) for expanding left and right the
retention time range.}

\item{mzExpand}{Numeric (length 1) with the mass or mass-to-charge ratio (\emph{m/z}) (in Da) for expanding left and
right the mass range.}
}
\value{
A ProcessingSettings S3 class object with subclass Settings_fill_features_StreamFind.
}
\description{
Settings for filling missing values in features.
}
