% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ProcessingSettings.R
\name{Settings_filter_features_patRoon}
\alias{Settings_filter_features_patRoon}
\title{Settings_filter_features_patRoon}
\usage{
Settings_filter_features_patRoon(
  absMinIntensity = NULL,
  relMinIntensity = NULL,
  preAbsMinIntensity = NULL,
  preRelMinIntensity = NULL,
  absMinAnalyses = NULL,
  relMinAnalyses = NULL,
  absMinReplicates = NULL,
  relMinReplicates = NULL,
  absMinFeatures = NULL,
  relMinFeatures = NULL,
  absMinReplicateAbundance = NULL,
  relMinReplicateAbundance = NULL,
  absMinConc = NULL,
  relMinConc = NULL,
  absMaxTox = NULL,
  relMaxTox = NULL,
  absMinConcTox = NULL,
  relMinConcTox = NULL,
  maxReplicateIntRSD = NULL,
  blankThreshold = NULL,
  retentionRange = NULL,
  mzRange = NULL,
  mzDefectRange = NULL,
  chromWidthRange = NULL,
  featQualityRange = NULL,
  groupQualityRange = NULL,
  rGroups = NULL,
  results = NULL,
  removeBlanks = FALSE,
  removeISTDs = FALSE,
  checkFeaturesSession = NULL,
  removeNA = FALSE,
  negate = FALSE
)
}
\value{
A ProcessingSettings S3 class object with subclass Settings_filter_features_patRoon.
}
\description{
Settings for filtering of features and feature groups. A full description of the filtering parameters is
in \code{\link[patRoon]{replicateGroupSubtract}} from patRoon package.
}
\details{
Note that when filters are applied to features or feature groups
these require specific results from processing modules. For instance,
subtracting the blank can only be done after grouping features. Also, some
filters require. Thus, not all filters can be applied to features.
See \code{\link[patRoon]{features-class}} and \code{\link[patRoon]{replicateGroupSubtract}}
for further information.
}
