% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_005_default_ProcessingSettings.R
\name{Settings_find_features_kpic2}
\alias{Settings_find_features_kpic2}
\title{Settings_find_features_kpic2}
\usage{
Settings_find_features_kpic2()
}
\value{
X.
}
\description{
X.
}
