% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_005_default_ProcessingSettings.R
\name{settings_find_features_openms}
\alias{settings_find_features_openms}
\title{settings_find_features_openms}
\usage{
settings_find_features_openms()
}
\value{
X.
}
\description{
X.
}
