% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_005_default_ProcessingSettings.R
\name{settings_find_features_xcms3_centwave}
\alias{settings_find_features_xcms3_centwave}
\title{settings_find_features_xcms3_centwave}
\usage{
settings_find_features_xcms3_centwave()
}
\value{
X.
}
\description{
X.
}
