% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ProcessingSettings.R
\name{Settings_find_features_xcms3_matchedfilter}
\alias{Settings_find_features_xcms3_matchedfilter}
\alias{validate.Settings_find_features_xcms3_matchedfilter}
\title{Settings_find_features_xcms3_matchedfilter}
\usage{
Settings_find_features_xcms3_matchedfilter(
  binSize = 0.5,
  impute = "none",
  baseValue = 0,
  distance = 0,
  fwhm = 30,
  max = 5,
  snthresh = 20,
  steps = 2,
  mzdiff = 0.5,
  index = FALSE
)

\method{validate}{Settings_find_features_xcms3_matchedfilter}(x)
}
\arguments{
\item{binSize}{numeric(1) specifying the width of the bins/slices in m/z dimension.}

\item{impute}{Character string specifying the method to be used for missing
value imputation. Allowed values are "none" (no linear interpolation), "lin"
(linear interpolation), "linbase" (linear interpolation within a certain
bin-neighborhood) and "intlin".}

\item{baseValue}{The base value to which empty elements should be set.
This is only considered for \code{impute} as "linbase" and corresponds to the
profBinLinBase's \code{baselevel} argument.}

\item{distance}{For \code{impute} as "linbase": number of non-empty neighboring
element of an empty element that should be considered for linear
interpolation. See details section for more information.}

\item{fwhm}{numeric(1) specifying the full width at half maximum of matched
filtration gaussian model peak. Only used to calculate the actual sigma,
see below.}

\item{max}{numeric(1) representing the maximum number of peaks that are
expected/will be identified per slice.}

\item{snthresh}{numeric(1) defining the signal to noise ratio cutoff.}

\item{steps}{numeric(1) defining the number of bins to be merged before
filtration (i.e. the number of neighboring bins that will be joined to the
slice in which filtration and peak detection will be performed).}

\item{mzdiff}{numeric(1) representing the minimum difference in m/z dimension
required for peaks with overlapping retention times; can be negative to allow
overlap. During peak post-processing, peaks defined to be overlapping are
reduced to the one peak with the largest signal.}

\item{index}{logical(1) specifying whether indices should be returned instead of values for m/z and retention times.}
}
\value{
A ProcessingSettings S3 class object with subclass Settings_find_features_xcms3_matchedfilter.
}
\description{
Settings for finding features (i.e., chromatographic peaks) in mzML/mzXML files using the package
\href{https://bioconductor.org/packages/release/bioc/html/xcms.html}{xcms} (version 3) with the algorithm
\href{https://rdrr.io/bioc/xcms/man/findChromPeaks-Chromatogram-MatchedFilter.html}{MatchedFilter},
which is optimal/preferred for low resolution LC-MS data. The function uses the package \pkg{patRoon} in the background.
}
\details{
See the \link[patRoon]{findFeaturesXCMS3} function from the \pkg{patRoon} package for more information and requirements.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate(Settings_find_features_xcms3_matchedfilter)}: Validates the object structure, returning a logical value of length one.

}}
\references{
\insertRef{patroon01}{StreamFind}

\insertRef{patroon02}{StreamFind}

\insertRef{xcms01}{StreamFind}

\insertRef{xcms02}{StreamFind}

\insertRef{xcms03}{StreamFind}
}
