% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ProcessingSettings.R
\name{Settings_find_internal_standards_StreamFind}
\alias{Settings_find_internal_standards_StreamFind}
\title{Settings_find_internal_standards_StreamFind}
\usage{
Settings_find_internal_standards_StreamFind(database = NULL, ppm = 5, sec = 10)
}
\arguments{
\item{database}{A data.frame with at least the columns name, mass, and rt indicating the name, neutral monoisotopic
mass and retention time of the internal standards, respectively.}

\item{ppm}{Numeric of length one with the mass deviation, in ppm.}

\item{sec}{Numeric of length one with the time deviation, in seconds.}
}
\value{
A ProcessingSettings S3 class object with subclass Settings_find_internal_standards_StreamFind.
}
\description{
Settings for finding internal standards using a data.frame.
}
