% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_default_ProcessingSettings.R
\name{Settings_generate_compounds_MetFrag}
\alias{Settings_generate_compounds_MetFrag}
\alias{validate.Settings_generate_compounds_MetFrag}
\title{Settings_generate_compounds_MetFrag}
\usage{
Settings_generate_compounds_MetFrag()

\method{validate}{Settings_generate_compounds_MetFrag}(x)
}
\arguments{
\item{x}{A Settings_generate_compounds_MetFrag S3 class object.}
}
\value{
A ProcessingSettings S3 class object with subclass
Settings_generate_compounds_MetFrag.
}
\description{
Settings for generating compounds using
\href{https://ipb-halle.github.io/MetFrag/}{MetFrag}.
The algorithm is used via the function \link[patRoon]{generateCompounds}
from the package \pkg{patRoon}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate(Settings_generate_compounds_MetFrag)}: Validates the object structure, returning a logical value of length one.

}}
\references{
\insertRef{metfrag01}{StreamFind}

\insertRef{metfrag02}{StreamFind}

\insertRef{metfrag03}{StreamFind}

\insertRef{metfrag04}{StreamFind}

\insertRef{patroon01}{StreamFind}

\insertRef{patroon02}{StreamFind}
}
