% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_default_ProcessingSettings.R
\name{Settings_generate_formulas_GenForm}
\alias{Settings_generate_formulas_GenForm}
\alias{validate.Settings_generate_formulas_GenForm}
\title{Settings_generate_formulas_GenForm}
\usage{
Settings_generate_formulas_GenForm()

\method{validate}{Settings_generate_formulas_GenForm}(x)
}
\arguments{
\item{x}{A Settings_generate_formulas_GenForm S3 class object.}
}
\value{
A ProcessingSettings S3 class object with subclass
Settings_generate_formulas_GenForm.
}
\description{
Settings for generating formulas using the algorithm
\href{https://sourceforge.net/projects/genform/}{GenForm}.
The algorithm is used via the function \link[patRoon]{generateFormulas}
from the package \pkg{patRoon}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate(Settings_generate_formulas_GenForm)}: Validates the object structure, returning a logical value of length one.

}}
\references{
\insertRef{patroon01}{StreamFind}

\insertRef{patroon02}{StreamFind}

\insertRef{genform}{StreamFind}
}
