% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_005_default_ProcessingSettings.R
\name{settings_group_features_xcms3_peakdensity}
\alias{settings_group_features_xcms3_peakdensity}
\title{settings_group_features_xcms3_peakdensity}
\usage{
settings_group_features_xcms3_peakdensity()
}
\value{
X.
}
\description{
X.
}
