% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_default_ProcessingSettings.R
\name{Settings_group_features_xcms3_peakdensity_peakgroups}
\alias{Settings_group_features_xcms3_peakdensity_peakgroups}
\title{Settings_group_features_xcms3_peakdensity_peakgroups}
\usage{
Settings_group_features_xcms3_peakdensity_peakgroups(
  bw = 5,
  minFraction = 1,
  minSamples = 1,
  binSize = 0.008,
  pre_bw = 5,
  pre_minFraction = 1,
  pre_minSamples = 1,
  pre_binSize = 0.008,
  maxFeatures = 100,
  rtAlignMinFraction = 0.9,
  extraPeaks = 1,
  smooth = "loess",
  span = 0.2,
  family = "gaussian",
  peakGroupsMatrix = matrix(nrow = 0, ncol = 0),
  subset = integer(),
  subsetAdjust = "average"
)
}
\arguments{
\item{bw}{numeric(1) defining the bandwidth (standard deviation of the
smoothing kernel) to be used. This argument is passed to the \code{density()}
method.}

\item{minFraction}{numeric(1) defining the minimum fraction of analyses in at
least one analysis replicate group in which the features have to be present
to be considered as a feature group.}

\item{minSamples}{numeric(1) with the minimum number of analyses in at least
one analysis replicate group in which the features have to be detected to be
considered a feature group.}

\item{binSize}{numeric(1) defining the size of the overlapping slices in mz
dimension.}

\item{pre_bw}{as \code{bw} but applied before retention time alignment.}

\item{pre_minFraction}{as \code{minFraction} but applied before retention time
alignment.}

\item{pre_minSamples}{as \code{minSamples} but applied before retention time
alignment.}

\item{pre_binSize}{as \code{binSize} but applied before retention time alignment.}

\item{maxFeatures}{numeric(1) with the maximum number of feature groups to be
identified in a single mz slice.}

\item{rtAlignMinFraction}{numeric(1) between 0 and 1 defining the minimum
required fraction of samples in which peaks for the peak group were identified.
Peak groups passing this criteria will aligned across samples and retention
times of individual spectra will be adjusted based on this alignment.
For minFraction = 1 the peak group has to contain peaks in all samples of
the experiment. Note that if subset is provided, the specified fraction is
relative to the defined subset of samples and not to the total number of
samples within the experiment (i.e. a peak has to be present in the specified
proportion of subset samples).}

\item{extraPeaks}{numeric(1) defining the maximal number of additional peaks
for all samples to be assigned to a peak group (i.e. feature) for retention
time correction. For a data set with 6 samples, extraPeaks = 1 uses all peak
groups with a total peak count <= 6 + 1. The total peak count is the total
number of peaks being assigned to a peak group and considers also multiple
peaks within a sample being assigned to the group.}

\item{smooth}{character defining the function to be used, to interpolate
corrected retention times for all peak groups. Either "loess" or "linear".}

\item{span}{numeric(1) defining the degree of smoothing (if smooth = "loess").
This parameter is passed to the internal call to loess.}

\item{family}{character defining the method to be used for loess smoothing.
Allowed values are "gaussian" and "symmetric".See loess for more information.}

\item{peakGroupsMatrix}{optional matrix of (raw) retention times for the peak
groups on which the alignment should be performed. Each column represents a
sample, each row a feature/peak group. Such a matrix is for example returned
by the adjustRtimePeakGroups method.}

\item{subset}{integer with the indices of samples within the experiment on
which the alignment models should be estimated. Samples not part of the subset
are adjusted based on the closest subset sample. See description above
for more details.}

\item{subsetAdjust}{character specifying the method with which non-subset
samples should be adjusted. Supported options are "previous" and "average"
(default). See description above for more information.}
}
\value{
A ProcessingSettings S3 class object with subclass
Settings_group_features_xcms3_peakdensity_peakgroups.
}
\description{
Settings for aligning and grouping features (i.e.,
chromatographic peaks) across mzML/mzXML files using the package
\href{https://bioconductor.org/packages/release/bioc/html/xcms.html}{xcms}
(version 3) with the algorithm
\href{https://rdrr.io/bioc/xcms/man/adjustRtime-peakGroups.html}{peakGroups}
for retention time alignment and the algorithm
\href{https://rdrr.io/bioc/xcms/man/groupChromPeaks-density.html}{peakdensity}
for grouping. The function uses the package \pkg{patRoon} in the background.
}
\details{
See the \link[patRoon]{groupFeaturesXCMS3} function from the
\pkg{patRoon} package for more information and requirements.
}
\references{
\insertRef{patroon01}{StreamFind}

\insertRef{patroon02}{StreamFind}

\insertRef{xcms01}{StreamFind}

\insertRef{xcms02}{StreamFind}

\insertRef{xcms03}{StreamFind}
}
