% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_default_ProcessingSettings.R
\name{Settings_integrate_chromatograms_StreamFind}
\alias{Settings_integrate_chromatograms_StreamFind}
\alias{validate.Settings_integrate_chromatograms_StreamFind}
\title{Settings_integrate_chromatograms_StreamFind}
\usage{
Settings_integrate_chromatograms_StreamFind(
  chromatograms = NA_character_,
  smoothing = TRUE,
  windowSize = 4,
  baseline = TRUE,
  baseline_method = "als",
  baseline_args = list(lambda = 5, p = 0.05, maxit = 10),
  merge = TRUE,
  closeByThreshold = 45,
  valeyThreshold = 0.5,
  minPeakHeight = 0,
  minPeakDistance = 10,
  minPeakWidth = 5,
  maxPeakWidth = 120,
  minSN = 10
)

\method{validate}{Settings_integrate_chromatograms_StreamFind}(x)
}
\arguments{
\item{chromatograms}{}

\item{smoothing}{}

\item{windowSize}{}

\item{baseline}{}

\item{baseline_method}{}

\item{baseline_args}{}

\item{merge}{}

\item{closeByThreshold}{}

\item{valeyThreshold}{}

\item{minPeakHeight}{}

\item{minPeakDistance}{}

\item{minPeakWidth}{}

\item{maxPeakWidth}{}

\item{minSN}{}

\item{x}{A Settings_integrate_chromatograms_StreamFind S3 class object.}
}
\value{
A ProcessingSettings S3 class object with subclass Settings_integrate_chromatograms_StreamFind.
}
\description{
Prototype.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate(Settings_integrate_chromatograms_StreamFind)}: Validates the object structure, returning a logical value of length one.

}}
