% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ProcessingSettings.R
\name{Settings_load_features_eic_StreamFind}
\alias{Settings_load_features_eic_StreamFind}
\title{Settings_load_features_eic_StreamFind}
\usage{
Settings_load_features_eic_StreamFind(
  rtExpand = 120,
  mzExpand = 0,
  filtered = FALSE
)
}
\arguments{
\item{rtExpand}{Numeric (length 1) with the retention time values (in seconds) for expanding left and right the
retention time range.}

\item{mzExpand}{Numeric (length 1) with the mass or mass-to-charge ratio (\emph{m/z}) (in Da) for expanding left and
right the mass range.}

\item{filtered}{Logical (length 1). When \code{TRUE}, filtered features/groups are considered.}
}
\value{
A ProcessingSettings S3 class object with subclass Settings_load_features_eic_StreamFind.
}
\description{
Settings for loading spectra EIC for feature groups.
}
