% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_005_default_ProcessingSettings.R
\name{Settings_load_groups_ms2_streamFind}
\alias{Settings_load_groups_ms2_streamFind}
\alias{validate.Settings_load_groups_ms2_streamFind}
\title{Settings_load_groups_ms2_streamFind}
\usage{
Settings_load_groups_ms2_streamFind(
  mzClust = 0.01,
  isInAllSpectra = TRUE,
  minIntensity = 250,
  filtered = FALSE,
  runParallel = FALSE,
  verbose = FALSE
)

\method{validate}{Settings_load_groups_ms2_streamFind}(x)
}
\arguments{
\item{mzClust}{Numeric value with the mass deviation threshold (in Da) to
cluster mass traces. Highly dependent on the mass resolution of the MS data.}

\item{isInAllSpectra}{Logical with length one. When \code{TRUE}, traces are kept
if present in all spectra. For MS2 data and when applied multiple collision
energies, the presence is evaluated for traces within spectra of each
collision energy.}

\item{minIntensity}{Numeric value with the minimum intensity.}

\item{filtered}{Logical. When \code{TRUE}, filtered features/groups are considered.}

\item{runParallel}{Logical of length one. Set to \code{TRUE} for parallel
processing. Note that parallel processing is not always faster due to
overhead.}

\item{verbose}{Logical with length one. Set \code{TRUE} to output diagnostic and
result messages. When \code{FALSE}, some messages are omitted.}

\item{x}{A Settings_load_groups_ms2_streamFind S3 class object.}
}
\value{
A ProcessingSettings S3 class object with subclass
Settings_load_groups_ms2_streamFind.
}
\description{
Settings for loading MS2 spectra for feature groups.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate(Settings_load_groups_ms2_streamFind)}: Validates the Settings_load_groups_ms2_streamFind S3 class object,
returning a logical value of length one.

}}
