% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ProcessingSettings.R
\name{Settings_load_MSPeakLists_StreamFind}
\alias{Settings_load_MSPeakLists_StreamFind}
\title{Settings_load_MSPeakLists_StreamFind}
\usage{
Settings_load_MSPeakLists_StreamFind(
  clusterMzWindow = 0.005,
  topMost = 100,
  minIntensityPre = 50,
  minIntensityPost = 50,
  avgFun = mean,
  method = "distance"
)
}
\arguments{
\item{clusterMzWindow}{m/z window (in Da) used for clustering m/z values
when spectra are averaged. For method="hclust" this corresponds to the
cluster height, while for method="distance" this value is used to find
nearby masses (+/- window). Too small windows will prevent clustering
m/z values (thus erroneously treating equal masses along spectra as
different), whereas too big windows may cluster unrelated m/z values
from different or even the same spectrum together.}

\item{topMost}{Only retain this maximum number of MS peaks when generating
averaged spectra. Lowering this number may exclude more irrelevant (noisy)
MS peaks and decrease processing time, whereas higher values may avoid
excluding lower intense MS peaks that may still be of interest.}

\item{minIntensityPre}{MS peaks with intensities below this value will
be removed (applied prior to selection by \code{topMost}) before averaging.}

\item{minIntensityPost}{MS peaks with intensities below this value will
be removed after averaging.}

\item{avgFun}{Function that is used to calculate average m/z values.}

\item{method}{Method used for producing averaged MS spectra. Valid
values are "hclust", used for hierarchical clustering (using the
fastcluster package), and "distance", to use the between peak distance.
The latter method may reduces processing time and memory requirements,
at the potential cost of reduced accuracy.}
}
\value{
A ProcessingSettings S3 class object with subclass Settings_load_MSPeakLists_StreamFind.
}
\description{
Settings for converting loaded MS2 and MS1 spectra into a \code{MSPeakLists} object from patRoon.
}
