% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ProcessingSettings.R
\name{Settings_merge_spectra_time_series_StreamFind}
\alias{Settings_merge_spectra_time_series_StreamFind}
\title{Settings_merge_spectra_time_series_StreamFind}
\usage{
Settings_merge_spectra_time_series_StreamFind(preCut = 2)
}
\arguments{
\item{preCut}{The number of pre Raman scans to exclude when merging.}
}
\value{
A ProcessingSettings S3 class object with subclass Settings_merge_spectra_time_series_StreamFind.
}
\description{
Merges Raman spectra based on time series data. It collapses data files into a single file.
}
