% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ProcessingSettings.R
\name{Settings_normalize_features_patRoon_tic}
\alias{Settings_normalize_features_patRoon_tic}
\title{Settings_normalize_features_patRoon_tic}
\usage{
Settings_normalize_features_patRoon_tic()
}
\value{
A ProcessingSettings S3 class object with subclass Settings_normalize_features_patRoon_tic.
}
\description{
Settings for normalizing features using the Total Ion Current (TIC) method from the patRoon package.
}
\references{
\insertRef{patroon01}{StreamFind}

\insertRef{patroon02}{StreamFind}
}
