% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ProcessingSettings.R
\name{Settings_normalize_spectra_blockweight}
\alias{Settings_normalize_spectra_blockweight}
\title{Settings_normalize_spectra_blockweight}
\usage{
Settings_normalize_spectra_blockweight()
}
\value{
A ProcessingSettings S3 class object with subclass Settings_normalize_spectra_blockweight.
}
\description{
Normalizes spectra using block weighting for downstream data evaluation.
}
