% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ProcessingSettings.R
\name{Settings_normalize_spectra_meancenter}
\alias{Settings_normalize_spectra_meancenter}
\title{Settings_normalize_spectra_meancenter}
\usage{
Settings_normalize_spectra_meancenter()
}
\value{
A ProcessingSettings S3 class object with subclass Settings_normalize_spectra_meancenter.
}
\description{
Normalizes spectra using mean centering.
}
