% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ProcessingSettings.R
\name{Settings_normalize_spectra_minmax}
\alias{Settings_normalize_spectra_minmax}
\title{Settings_normalize_spectra_minmax}
\usage{
Settings_normalize_spectra_minmax()
}
\value{
A ProcessingSettings S3 class object with subclass Settings_normalize_spectra_minmax.
}
\description{
Normalizes spectra using the min-max algorithm.
}
