% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ProcessingSettings.R
\name{Settings_normalize_spectra_scale}
\alias{Settings_normalize_spectra_scale}
\title{Settings_normalize_spectra_scale}
\usage{
Settings_normalize_spectra_scale()
}
\value{
A ProcessingSettings S3 class object with subclass Settings_normalize_spectra_scale.
}
\description{
Normalizes spectra using scaling based on the standard deviation.
}
