% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ProcessingSettings.R
\name{Settings_normalize_spectra_snv}
\alias{Settings_normalize_spectra_snv}
\title{Settings_normalize_spectra_snv}
\usage{
Settings_normalize_spectra_snv(liftTozero = FALSE)
}
\arguments{
\item{liftTozero}{Logical (length 1) indicating if the spectra should be lifted to zero.}
}
\value{
A ProcessingSettings S3 class object with subclass Settings_normalize_spectra_snv.
}
\description{
Normalizes spectra using the Standard Normal Variate (SNV) algorithm.
}
