% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_default_ProcessingSettings.R
\name{Settings_normalize_spectra_StreamFind}
\alias{Settings_normalize_spectra_StreamFind}
\alias{validate.Settings_normalize_spectra_StreamFind}
\title{Settings_normalize_spectra_StreamFind}
\usage{
Settings_normalize_spectra_StreamFind(
  liftTozero = FALSE,
  xName = NULL,
  xVal = NULL
)

\method{validate}{Settings_normalize_spectra_StreamFind}(x)
}
\arguments{
\item{liftTozero}{}

\item{xName}{}

\item{xVal}{}

\item{x}{A Settings_normalize_spectra_StreamFind S3 class object.}
}
\value{
A ProcessingSettings S3 class object with subclass Settings_normalize_spectra_StreamFind.
}
\description{
Prototype.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate(Settings_normalize_spectra_StreamFind)}: Validates the object structure, returning a logical value of length one.

}}
