% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ProcessingSettings.R
\name{Settings_smooth_chromatograms_savgol}
\alias{Settings_smooth_chromatograms_savgol}
\alias{validate.Settings_smooth_chromatograms_savgol}
\title{Settings_smooth_chromatograms_savgol}
\usage{
Settings_smooth_chromatograms_savgol(fl = 11, forder = 4, dorder = 0)

\method{validate}{Settings_smooth_chromatograms_savgol}(x)
}
\arguments{
\item{fl}{}

\item{forder}{}

\item{dorder}{}

\item{x}{A Settings_smooth_chromatograms_savgol S3 class object.}
}
\value{
A ProcessingSettings S3 class object with subclass Settings_smooth_chromatograms_savgol.
}
\description{
Prototype.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate(Settings_smooth_chromatograms_savgol)}: Validates the object structure, returning a logical value of length one.

}}
