% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_default_ProcessingSettings.R
\name{Settings_smooth_spectra_movingaverage}
\alias{Settings_smooth_spectra_movingaverage}
\alias{validate.Settings_smooth_spectra_movingaverage}
\title{Settings_smooth_spectra_movingaverage}
\usage{
Settings_smooth_spectra_movingaverage(windowSize = 5, xValWindow = NULL)

\method{validate}{Settings_smooth_spectra_movingaverage}(x)
}
\arguments{
\item{windowSize}{}

\item{xValWindow}{}

\item{x}{A Settings_smooth_spectra_movingaverage S3 class object.}
}
\value{
A ProcessingSettings S3 class object with subclass Settings_smooth_spectra_movingaverage.
}
\description{
Prototype.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate(Settings_smooth_spectra_movingaverage)}: Validates the object structure, returning a logical value of length one.

}}
