% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ProcessingSettings.R
\name{Settings_subtract_blank_spectra_StreamFind}
\alias{Settings_subtract_blank_spectra_StreamFind}
\title{Settings_subtract_blank_spectra_StreamFind}
\usage{
Settings_subtract_blank_spectra_StreamFind()
}
\value{
A ProcessingSettings S3 class object with subclass Settings_subtract_blank_spectra_StreamFind.
}
\description{
Subtracts the blank spectra to each analysis according to the blank assignment.
}
