% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ProcessingSettings.R
\name{Settings_subtract_spectra_section_StreamFind}
\alias{Settings_subtract_spectra_section_StreamFind}
\title{Settings_subtract_spectra_section_StreamFind}
\usage{
Settings_subtract_spectra_section_StreamFind(
  sectionVal = "rt",
  sectionWindow = c(10, 200)
)
}
\arguments{
\item{sectionVal}{Character (length 1) with the variable to be used for sectioning.}

\item{sectionWindow}{Numeric (length 2) with the window for the sectioning.}
}
\value{
A ProcessingSettings S3 class object with subclass Settings_subtract_spectra_section_StreamFind.
}
\description{
Subtracts a section of the spectra based on a variable (i.e. column name).
}
