% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_000_generics.R
\name{setTitle<-}
\alias{setTitle<-}
\title{setTitle<-}
\usage{
setTitle(object) <- value
}
\arguments{
\item{object}{A method specific class object.}

\item{value}{A method specific vector with title/s.}
}
\description{
Generic for setting the title.
}
