% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{show}
\alias{show}
\title{Generic \code{show}}
\usage{
show(x, ...)
}
\arguments{
\item{x}{An object from which the data will be shown.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S7 generic to show data from \code{x}.
}
