% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_S3.R
\name{size}
\alias{size}
\title{Generic \code{size}}
\usage{
size(x, ...)
}
\arguments{
\item{x}{An object whose size will be determined.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S3 generic to get the size of \code{x}.
}
