% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ms_filters_features.R
\name{snRatio}
\alias{snRatio}
\title{snRatio}
\usage{
snRatio(obj, value = 10)
}
\arguments{
\item{obj}{An \linkS4class{msData} object.}

\item{snRatio}{A numerical value set at the desired signal-to-noise ratio for features.}
}
\description{
snRatio
}
