% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_filterFeatures.R
\name{snRatioThresholdFeatures}
\alias{snRatioThresholdFeatures}
\title{snRatioThresholdFeatures}
\usage{
snRatioThresholdFeatures(obj, value = 10)
}
\arguments{
\item{obj}{An \linkS4class{msFeatures} object.}

\item{snRatio}{A numerical value set at the desired signal-to-noise ratio for features.}
}
\description{
snRatioThresholdFeatures
}
