% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_S3_StatisticAnalysis.R
\name{StatisticAnalysis}
\alias{StatisticAnalysis}
\title{\strong{StatisticAnalysis} S3 class constructor, methods and functions}
\usage{
StatisticAnalysis(
  name = NA_character_,
  replicate = NA_character_,
  blank = NA_character_,
  data = matrix(),
  class = NA_character_
)
}
\arguments{
\item{name}{Character (length 1) with the unique identifier.}

\item{replicate}{Character (length 1), representing the analysis replicate group name.}

\item{blank}{Character (length 1), representing the associated blank replicate group name.}

\item{data}{Matrix with 1 row and columns equal to the number of variables.}

\item{classes}{Character vector with the analysis classes.}
}
\value{
An \emph{StatisticAnalysis} S3 class object.
}
\description{
Creates a \emph{StatisticAnalysis} S3 class object.
}
