% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_StatisticMethod_MakeModel.R
\name{StatisticMethod_MakeModel_mcrpure_mdatools}
\alias{StatisticMethod_MakeModel_mcrpure_mdatools}
\title{StatisticMethod_MakeModel_mcrpure_mdatools Class}
\usage{
StatisticMethod_MakeModel_mcrpure_mdatools(
  ncomp = NULL,
  purevars = NULL,
  offset = 0.05,
  exclrows = NULL,
  exclcols = NULL,
  info = ""
)
}
\arguments{
\item{ncomp}{Integer (length 1) with the number of components to be calculated.}

\item{purevars}{Integer vector with the indices of the pure variables (optional).}

\item{offset}{Numeric (length 1) offset for correcting noise in computing maximum angles (should be value within 0 and 1).}

\item{exclrows}{Integer vector with the row indices to be excluded.}

\item{exclcols}{Integer vector with the column indices to be excluded.}

\item{info}{Character (length 1) with additional information.}
}
\value{
A StatisticMethod_MakeModel_mcrpure_mdatools object.
}
\description{
Makes a Multivariate Curve Resolution (MCR) purity model based on the R package \pkg{mdatools}.
}
\references{
\insertRef{mdatools01}{StreamFind}

\insertRef{mdatools02}{StreamFind}
}
