% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_StatisticMethod_PrepareData.R
\name{StatisticMethod_PrepareData_autoscale}
\alias{StatisticMethod_PrepareData_autoscale}
\title{StatisticMethod_PrepareData_autoscale}
\usage{
StatisticMethod_PrepareData_autoscale(center = TRUE, scale = TRUE)
}
\arguments{
\item{center}{Logical (length 1) indicating if the data should be centered.}

\item{scale}{Logical (length 1) indicating if the data should be scaled.}
}
\value{
A StatisticMethod_PrepareData_autoscale object.
}
\description{
Auto scale and centers data using the \code{prep.autoscale} function from the \pkg{mdatools} package.
}
\details{
StatisticMethod_PrepareData_autoscale S7 class
}
