% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_StatisticResults_KNN_class.R
\name{StatisticResults_KNN_class}
\alias{StatisticResults_KNN_class}
\alias{validate_object.StatisticResults_KNN_class}
\alias{get_model_data.StatisticResults_KNN_class}
\alias{test.StatisticResults_KNN_class}
\alias{predict.StatisticResults_KNN_class}
\title{StatisticResults_KNN_class Class}
\usage{
StatisticResults_KNN_class(model = list())

\method{validate_object}{StatisticResults_KNN_class}(x)

\method{get_model_data}{StatisticResults_KNN_class}(x)

\method{test}{StatisticResults_KNN_class}(x, data)

\method{predict}{StatisticResults_KNN_class}(x, data)
}
\arguments{
\item{model}{A list containing the KNN model and its conditions.}

\item{x}{A \code{StatisticResults_KNN_class} object.}

\item{data}{A data frame containing the new data to test or predict.}
}
\description{
The \code{StatisticResults_KNN_class} class is used to store the results of a KNN classification model.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate_object(StatisticResults_KNN_class)}: Validate the StatisticResults_KNN_class object, returns NULL if valid.

\item \code{get_model_data(StatisticResults_KNN_class)}: Get the model data.

\item \code{test(StatisticResults_KNN_class)}: Test new data using the KNN model.

\item \code{predict(StatisticResults_KNN_class)}: Predict new data using the KNN model.

}}
