% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_StatisticResults_MCRALS_mdatools.R
\name{StatisticResults_MCRALS_mdatools}
\alias{StatisticResults_MCRALS_mdatools}
\alias{validate_object.StatisticResults_MCRALS_mdatools}
\alias{summary.StatisticResults_MCRALS_mdatools}
\alias{predict.StatisticResults_MCRALS_mdatools}
\alias{test.StatisticResults_MCRALS_mdatools}
\alias{get_model_data.StatisticResults_MCRALS_mdatools}
\alias{plot_explained_variance.StatisticResults_MCRALS_mdatools}
\alias{plot_cumulative_explained_variance.StatisticResults_MCRALS_mdatools}
\alias{plot_resolved_spectra.StatisticResults_MCRALS_mdatools}
\alias{plot_contributions.StatisticResults_MCRALS_mdatools}
\alias{plot_overview.StatisticResults_MCRALS_mdatools}
\title{StatisticResults_MCRALS_mdatools Class}
\usage{
StatisticResults_MCRALS_mdatools(model = list())

\method{validate_object}{StatisticResults_MCRALS_mdatools}(x)

\method{summary}{StatisticResults_MCRALS_mdatools}(object, ...)

\method{predict}{StatisticResults_MCRALS_mdatools}(x, data)

\method{test}{StatisticResults_MCRALS_mdatools}(x, data)

\method{get_model_data}{StatisticResults_MCRALS_mdatools}(x)

\method{plot_explained_variance}{StatisticResults_MCRALS_mdatools}(
  x,
  interactive = TRUE,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  showText = TRUE,
  showLegend = TRUE
)

\method{plot_cumulative_explained_variance}{StatisticResults_MCRALS_mdatools}(
  x,
  interactive = TRUE,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  showText = TRUE,
  showLegend = TRUE
)

\method{plot_resolved_spectra}{StatisticResults_MCRALS_mdatools}(
  x,
  interactive = TRUE,
  pcs = NULL,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  showText = TRUE,
  showLegend = TRUE
)

\method{plot_contributions}{StatisticResults_MCRALS_mdatools}(
  x,
  interactive = TRUE,
  pcs = NULL,
  xLab = NULL,
  yLab = NULL,
  title = NULL
)

\method{plot_overview}{StatisticResults_MCRALS_mdatools}(x, ...)
}
\arguments{
\item{model}{A \code{mcrals} object created with the \pkg{mdatools} package.}

\item{x}{A \code{StatisticResults_MCRALS_mdatools} object.}

\item{object}{A \code{StatisticResults_MCRALS_mdatools} object.}

\item{...}{Additional arguments passed to the plot function.}

\item{data}{A data frame containing the new data to test or predict.}

\item{interactive}{Logical (length 1). When \code{TRUE}, the data is plotted interactively using \pkg{plotly}.}

\item{xLab}{A string with the title for the x axis.}

\item{yLab}{A string with the title for the y axis.}

\item{title}{A string with the title.}

\item{showText}{Logical (length 1), set to \code{TRUE} to show the text annotations.}

\item{showLegend}{Logical (length 1). Set to \code{TRUE} to show legend.}

\item{pcs}{A vector of integers specifying the principal components to plot. If NULL, all components are plotted.}
}
\description{
Class for storing results of MCR-ALS models created with the \pkg{mdatools} package.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate_object(StatisticResults_MCRALS_mdatools)}: Validate the object, return NULL if valid.

\item \code{summary(StatisticResults_MCRALS_mdatools)}: Print a summary of the model.

\item \code{predict(StatisticResults_MCRALS_mdatools)}: Predict new data using the MCRALS model.

\item \code{test(StatisticResults_MCRALS_mdatools)}: Test new data using the MCRALS model.

\item \code{get_model_data(StatisticResults_MCRALS_mdatools)}: Get the model data as a list of data frames.

\item \code{plot_explained_variance(StatisticResults_MCRALS_mdatools)}: Plot the explained variance of the model.

\item \code{plot_cumulative_explained_variance(StatisticResults_MCRALS_mdatools)}: Plot the cumulative explained variance of the model.

\item \code{plot_resolved_spectra(StatisticResults_MCRALS_mdatools)}: Plot the resolved spectra of the MCRALS model.

\item \code{plot_contributions(StatisticResults_MCRALS_mdatools)}: Plot the contributions of the resolved spectra.

\item \code{plot_overview(StatisticResults_MCRALS_mdatools)}: Plot the explained variance of the MCRALS model.

}}
