% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_StatisticResults_Model.R
\name{StatisticResults_Model}
\alias{StatisticResults_Model}
\alias{validate_object.StatisticResults_Model}
\alias{show.StatisticResults_Model}
\title{StatisticResults_Model Class}
\usage{
StatisticResults_Model(model = list())

\method{validate_object}{StatisticResults_Model}(x)

\method{show}{StatisticResults_Model}(x)
}
\arguments{
\item{model}{A list containing the model parameters.}

\item{x}{A \code{StatisticResults_Model} object.}
}
\value{
An object of class \code{StatisticResults_Model} which inherits from \code{Results}.
}
\description{
The \code{StatisticResults_Model} class is used to store results of statistical models, including model parameters, test results, and predictions.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate_object(StatisticResults_Model)}: Validate the StatisticResults_Model object, returns NULL if valid.

\item \code{show(StatisticResults_Model)}: Show the StatisticResults_Model object.

}}
