% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_StatisticResults_PCA_mdatools.R
\name{StatisticResults_PCA_mdatools}
\alias{StatisticResults_PCA_mdatools}
\alias{validate_object.StatisticResults_PCA_mdatools}
\alias{summary.StatisticResults_PCA_mdatools}
\alias{get_model_data.StatisticResults_PCA_mdatools}
\alias{predict.StatisticResults_PCA_mdatools}
\alias{test.StatisticResults_PCA_mdatools}
\alias{plot_explained_variance.StatisticResults_PCA_mdatools}
\alias{plot_cumulative_explained_variance.StatisticResults_Model}
\alias{plot_scores.StatisticResults_PCA_mdatools}
\alias{plot_loadings.StatisticResults_PCA_mdatools}
\alias{plot_residuals.StatisticResults_PCA_mdatools}
\alias{plot_residual_distance.StatisticResults_PCA_mdatools}
\alias{plot_overview.StatisticResults_PCA_mdatools}
\title{StatisticResults_PCA_mdatools Class}
\usage{
StatisticResults_PCA_mdatools(model = list())

\method{validate_object}{StatisticResults_PCA_mdatools}(x)

\method{summary}{StatisticResults_PCA_mdatools}(object, ...)

\method{get_model_data}{StatisticResults_PCA_mdatools}(x)

\method{predict}{StatisticResults_PCA_mdatools}(x, data)

\method{test}{StatisticResults_PCA_mdatools}(x, data)

\method{plot_explained_variance}{StatisticResults_PCA_mdatools}(
  x,
  interactive = TRUE,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  showText = TRUE,
  showLegend = TRUE
)

\method{plot_cumulative_explained_variance}{StatisticResults_Model}(
  x,
  interactive = TRUE,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  showText = TRUE,
  showLegend = TRUE
)

\method{plot_scores}{StatisticResults_PCA_mdatools}(
  x,
  analyses = NULL,
  interactive = TRUE,
  pcs = 1:2,
  title = NULL,
  colorGroups = NULL,
  showText = TRUE,
  showLegend = TRUE,
  colorBy = "results"
)

\method{plot_loadings}{StatisticResults_PCA_mdatools}(
  x,
  interactive = TRUE,
  pcs = 1:2,
  colorKey = NULL,
  title = NULL,
  showText = TRUE,
  showLegend = TRUE
)

\method{plot_residuals}{StatisticResults_PCA_mdatools}(
  x,
  analyses = NULL,
  interactive = TRUE,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  colorGroups = NULL,
  showText = TRUE,
  showLegend = TRUE,
  colorBy = "results"
)

\method{plot_residual_distance}{StatisticResults_PCA_mdatools}(x, ...)

\method{plot_overview}{StatisticResults_PCA_mdatools}(x, ...)
}
\arguments{
\item{model}{An object of class \code{pca} from the \pkg{mdatools} package.}

\item{x}{A StatisticResults_PCA_mdatools object.}

\item{object}{A StatisticResults_PCA_mdatools object.}

\item{...}{Additional arguments passed to the plot function.}

\item{data}{A data frame containing the new data to test or predict.}

\item{interactive}{Logical (length 1). When \code{TRUE}, the data is plotted interactively using \pkg{plotly}.}

\item{xLab}{A string with the title for the x axis.}

\item{yLab}{A string with the title for the y axis.}

\item{title}{A string with the title.}

\item{showText}{Logical (length 1), set to \code{TRUE} to show the text annotations.}

\item{showLegend}{Logical (length 1). Set to \code{TRUE} to show legend.}

\item{analyses}{Character or numeric vector with names or indexes of analyses in the \code{Analyses}
object.}

\item{pcs}{A vector of integers specifying which principal components to plot (default is 1:2).}

\item{colorGroups}{A factor character vector with the color groups for the data.}

\item{colorBy}{A string defining how to legend the plot. Possible values are \code{analyses}, \code{targets} (the default) or
\code{replicates}.}

\item{colorKey}{A vector of colors to use for the loadings (optional).}
}
\description{
The \code{StatisticResults_PCA_mdatools} class is used to store results of PCA model created with the \pkg{mdatools} package.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate_object(StatisticResults_PCA_mdatools)}: Validate the object, returning NULL if valid.

\item \code{summary(StatisticResults_PCA_mdatools)}: Print a summary of the model.

\item \code{get_model_data(StatisticResults_PCA_mdatools)}: Get the model data as a list of data frames.

\item \code{predict(StatisticResults_PCA_mdatools)}: Predict new data using the PCA model.

\item \code{test(StatisticResults_PCA_mdatools)}: Test new data using the PCA model.

\item \code{plot_explained_variance(StatisticResults_PCA_mdatools)}: Plot the explained variance of the PCA model.

\item \code{plot_scores(StatisticResults_PCA_mdatools)}: Plot the scores of the PCA model.

\item \code{plot_loadings(StatisticResults_PCA_mdatools)}: Plot the loadings of the PCA model.

\item \code{plot_residuals(StatisticResults_PCA_mdatools)}: Plot the residuals of the PCA model.

\item \code{plot_residual_distance(StatisticResults_PCA_mdatools)}: Plot residual distances of the PCA model.

\item \code{plot_overview(StatisticResults_PCA_mdatools)}: Plot the explained variance of the PCA model.

}}
\section{Functions}{
\itemize{
\item \code{plot_cumulative_explained_variance(StatisticResults_Model)}: Plot the cumulative explained variance of the PCA model.

}}
