% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_StatisticSettings_MakeModel.R
\name{StatisticSettings_MakeModel_knn}
\alias{StatisticSettings_MakeModel_knn}
\title{\strong{StatisticSettings_MakeModel_knn}}
\usage{
StatisticSettings_MakeModel_knn(k = 3, l = 0)
}
\arguments{
\item{k}{Integer (length 1) with the number of neighbors to be used.}

\item{l}{Integer (length 1) with the minimum vote for definite decision, otherwise doubt.
(More precisely, less than k-l dissenting votes are allowed, even if k is increased by ties.)}
}
\value{
A StatisticSettings_MakeModel_knn object.
}
\description{
Makes a classification model using the k-nearest neighbors (knn) algorithm from package \pkg{class}.
}
\references{
\insertRef{class01}{StreamFind}
}
