% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_S3_StatisticSettings.R
\name{StatisticSettings_MakeModel_pca_mdatools}
\alias{StatisticSettings_MakeModel_pca_mdatools}
\title{StatisticSettings_MakeModel_pca_mdatools}
\usage{
StatisticSettings_MakeModel_pca_mdatools(
  ncomp = NULL,
  exclrows = NULL,
  exclcols = NULL,
  x.test = NULL,
  method = "svd",
  rand = NULL,
  lim.type = "ddmoments",
  alpha = 0.05,
  gamma = 0.01,
  info = ""
)
}
\arguments{
\item{ncomp}{Integer (length 1) with the number of components to be calculated.}

\item{exclrows}{Integer vector with the row indices to be excluded.}

\item{exclcols}{Integer vector with the column indices to be excluded.}

\item{x.test}{Matrix with the test data.}

\item{method}{Character (length 1) with the method to be used for PCA. Possible values are "svd" and "nipals".}

\item{rand}{Integer (length 1) with the random seed.}

\item{lim.type}{Character (length 1) with the type of limit for the PCA. Possible values are "jm" Jackson-Mudholkar
approach, "chisq" based on chi-square distribution, "ddmoments" and "ddrobust" related to data driven method. It is
highly recommended to concult the original documentation in \link[mdatools]{pca} for more information.}

\item{alpha}{Numeric (length 1) with the alpha value for the PCA.}

\item{gamma}{Numeric (length 1) with the gamma value for the PCA.}

\item{info}{Character (length 1) with additional information.}
}
\value{
A ProcessingSettings S3 class object with subclass StatisticSettings_MakeModel_pca_mdatools.
}
\description{
Makes a Principle Component Analysis (PCA) model based on the R package \pkg{mdatools}.
}
\references{
\insertRef{mdatools01}{StreamFind}
}
