% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_StatisticSettings_Quantify.R
\name{StatisticSettings_Quantify_mcrals}
\alias{StatisticSettings_Quantify_mcrals}
\title{StatisticSettings_Quantify_mcrals}
\usage{
StatisticSettings_Quantify_mcrals(concentrations = NA_real_)
}
\arguments{
\item{concentrations}{Numeric of the same length as analyses with the concentrations to be used for quantification.
When concentration is not known or not available, use \code{NA_real_}.}
}
\value{
A StatisticSettings_Quantify_mcrals object.
}
\description{
Quantify data using the mcrals statistical model and concentrations.
}
\details{
\strong{StatisticSettings_Quantify_mcrals}
}
