% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_class_001_streamProject.R
\docType{class}
\name{streamProject-class}
\alias{streamProject-class}
\alias{show,streamProject-method}
\alias{[,streamProject,ANY,missing,missing-method}
\alias{projectInfo,streamProject-method}
\alias{path,streamProject-method}
\alias{analysisInfo,streamProject-method}
\alias{files,streamProject-method}
\alias{analyses,streamProject-method}
\alias{replicates,streamProject-method}
\alias{replicates<-,streamProject-method}
\alias{blanks,streamProject-method}
\alias{blanks<-,streamProject-method}
\alias{metadata,streamProject-method}
\title{streamProject}
\usage{
\S4method{show}{streamProject}(object)

\S4method{[}{streamProject,ANY,missing,missing}(x, i, j, ..., drop = TRUE)

\S4method{projectInfo}{streamProject}(object, projectTitle = NULL, projectDate = NULL)

\S4method{path}{streamProject}(object)

\S4method{analysisInfo}{streamProject}(object)

\S4method{files}{streamProject}(object)

\S4method{analyses}{streamProject}(obj)

\S4method{replicates}{streamProject}(object)

\S4method{replicates}{streamProject}(object) <- value

\S4method{blanks}{streamProject}(object)

\S4method{blanks}{streamProject}(object) <- value

\S4method{metadata}{streamProject}(object, analyses = NULL, which = NULL)
}
\arguments{
\item{object}{An \linkS4class{streamProject} object.}

\item{i}{The indice/s or name/s of the analysis to keep in the \code{x} object.}

\item{projectTitle}{A character string to be used as project title.}

\item{projectDate}{The \link{Date} of the project.}

\item{value}{A character vector applicable to the respective method.}

\item{which}{A character vector with the entry name/s.}
}
\description{
An S4 class object containing the project title, date, path and added analyses.
The \code{streamProject} is the most basic class of the \pkg{streamFind} package.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: prints the details of an \linkS4class{streamProject} object.

\item \code{[}: subset on analysis, using analysis index or name.

\item \code{projectInfo}: setter and getter for project title, date and path.
When the \code{projectTitle} and \code{projectDate} arguments
are missing it returns a list with the project title, date and path.
If arguments are given, the project title and date are updated
based on the specified arguments, returning the original class object.

\item \code{path}: getter for project path.

\item \code{analysisInfo}: getter for analysis info as \link{data.frame} with
four columns: path, analysis, group and blank. The \link{data.frame}
can be used as analysisInfo in \pkg{patRoon}.

\item \code{files}: getter for analysis file paths.

\item \code{analyses}: getter for analysis names.

\item \code{replicates}: getter for replicate names.

\item \code{replicates<-}: setter for analysis replicate names.
The \code{value} is a character vector with the same length as
the number of analyses in the \code{object},
containing analysis replicate name for each analysis.

\item \code{blanks}: getter for blank names.

\item \code{blanks<-}: setter for associated blank replicate for each analyses.
The \code{value} is a character vector with the same length as
the number of analyses in the \code{object},
containing the associated blank replicate name of each analysis.

\item \code{metadata}: getter for analysis metadata.
Returns a \link[data.table]{date.table} with a row per analysis.
}}

\section{Slots}{

\describe{
\item{\code{project}}{A character string with the project title.}

\item{\code{date}}{The date of the project.}

\item{\code{path}}{A character string with the project path.}

\item{\code{analyses}}{A list of analyses added to the project.
The class of the analysis objects are defined according to the file formats.}
}}

