% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_class_001_streamSet.R
\docType{class}
\name{streamSet-class}
\alias{streamSet-class}
\alias{show,streamSet-method}
\alias{setTitle,streamSet-method}
\alias{setTitle,streamSet,streamSet-method}
\alias{setTitle<-,streamSet-method}
\alias{setTitle<-,streamSet,streamSet-method}
\alias{setDate,streamSet-method}
\alias{setDate,streamSet,streamSet-method}
\alias{analysisNames,streamSet-method}
\alias{analysisNames,streamSet,streamSet-method}
\alias{[,streamSet,ANY,missing,missing-method}
\alias{addAnalyses,streamSet-method}
\alias{addAnalyses,streamSet,streamSet-method}
\title{streamSet}
\usage{
\S4method{show}{streamSet}(object)

\S4method{setTitle}{streamSet}(object)

\S4method{setTitle}{streamSet}(object) <- value

\S4method{setDate}{streamSet}(object)

\S4method{analysisNames}{streamSet}(object)

\S4method{[}{streamSet,ANY,missing,missing}(x, i, j, ..., drop = TRUE)

\S4method{addAnalyses}{streamSet}(object, analysisList = NULL)
}
\arguments{
\item{object}{An \linkS4class{streamSet} object.}

\item{value}{A method specific string/vector.}

\item{x}{A \linkS4class{streamSet} object.}

\item{i}{The indice/s or name/s of the analyses to keep in \code{x}.}

\item{j}{Not applicable to \linkS4class{streamSet}.}

\item{...}{Other arguments.}

\item{drop}{Not applicable to \linkS4class{streamSet}.}

\item{analysisList}{A list of strings with analysis file full paths.}
}
\description{
An S4 class object containing the set title, date,
storage path and a list of analysis files.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show(streamSet)}: prints the details of
an \linkS4class{streamSet} object.

\item \code{setTitle(streamSet)}: getter for the title.

\item \code{setTitle(streamSet) <- value}: setter for title.
The \code{value} is a string with the title.

\item \code{setDate(streamSet)}: getter for the date.
Note that is not setter for the date.

\item \code{analysisNames(streamSet)}: getter for the analysis names.

\item \code{x[i}: subset on analyses, using analysis index or name.

\item \code{addAnalyses(streamSet)}: adds analyses to the \linkS4class{streamSet}.

}}
\section{Slots}{

\describe{
\item{\code{title}}{A character string with the stream set title.}

\item{\code{date}}{The date of the set.}

\item{\code{path}}{A character string with the set path.}

\item{\code{analyses}}{A list of analyses added to the project.
The class of the analysis objects are defined according to the file formats.
When a class is not possible to be assigned, the list contains
full path string of each file added.}
}}

