% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{test}
\alias{test}
\title{Generic \code{test}}
\usage{
test(x, ...)
}
\arguments{
\item{x}{An object from which the data will be tested.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S7 generic to test data from \code{x}.
}
