% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ms.R
\name{trim_spectra_targets}
\alias{trim_spectra_targets}
\title{trim_spectra_targets}
\usage{
trim_spectra_targets(traces, targets, preMZr)
}
\arguments{
\item{traces}{A data.frame with spectra.}

\item{targets}{A data.frame with targets (minimum and maximum).}

\item{preMZr}{A data.frame with precursor minimum and maximum values as
\code{targets} but expanded with the isolation window.}
}
\value{
The filtered \code{traces} data.frame.
}
\description{
trim_spectra_targets
}
