% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ms.R
\name{trim_vector}
\alias{trim_vector}
\title{trim_vector}
\usage{
trim_vector(v, a, b)
}
\arguments{
\item{v}{A vector to trim based on minimum and maximum value pairs.}

\item{a}{A vector with minimum values to evaluate \code{v}.}

\item{b}{A vector with maximum values to evaluate \code{v}.}
}
\value{
A logical vector with the same length as \code{v} with \code{TRUE} for
regions between \code{a} and \code{b} value pairs.
}
\description{
trim_vector
}
