% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_002_peakGrouping.R
\name{updateFeatureTable}
\alias{updateFeatureTable}
\title{updateFeatureTable}
\usage{
updateFeatureTable(object, fast = TRUE)
}
\arguments{
\item{object}{An \linkS4class{msData} object.}

\item{fast}{Logical, set to \code{TRUE} for a lazy update of features.
When \code{TRUE}, only the mz, rt, and intensity are updated.
When \code{FALSE}, the feature table is completly updated.
Note that UFIs, quality data and annotation info are lost.}
}
\value{
An \linkS4class{msData} object with updated features slot.
}
\description{
Function to update feature table from
peaks remaining in the \linkS4class{msData} object.
}
