% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_generics.R, R/class_S3_Analysis.R,
%   R/class_S3_StatisticAnalysis.R
\name{validate}
\alias{validate}
\alias{validate.Analysis}
\alias{validate.StatisticAnalysis}
\title{Validate}
\usage{
validate(x)

\method{validate}{Analysis}(x = NULL)

\method{validate}{StatisticAnalysis}(x = NULL)
}
\arguments{
\item{x}{A StatisticAnalysis S3 object.}
}
\description{
Validates an object.
}
\section{Methods (by class)}{
\itemize{
\item \code{validate(Analysis)}: description Validates an Analysis S3 object.

\item \code{validate(StatisticAnalysis)}: Validates a StatisticAnalysis S3 object

}}
