% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_R6_msData.R
\name{validate_headers}
\alias{validate_headers}
\title{validate_headers}
\usage{
validate_headers(value = NULL)
}
\arguments{
\item{value}{A headers list.}
}
\value{
A logical value of length 1.
}
\description{
Validates the headers list.
}
