% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_R6_implementation.R
\name{validate_list_ms_analysis}
\alias{validate_list_ms_analysis}
\title{validate_info_list_ms_analysis}
\usage{
validate_list_ms_analysis(value)
}
\arguments{
\item{value}{A list of information from an MS analysis mzML/mzXML file.}
}
\value{
A logical value of length 1.
}
\description{
Validates the list of information parsed from an MS analysis mzML/mzXML file.
}
