% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_R6_msData.R
\name{validate_ms_analysis}
\alias{validate_ms_analysis}
\title{validate_ms_analysis}
\usage{
validate_ms_analysis(value = NULL)
}
\arguments{
\item{value}{A list of information from an MS analysis mzML/mzXML file.}
}
\value{
A logical value of length 1.
}
\description{
Validates the list of information parsed from an MS analysis mzML/mzXML file.
}
