% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_R6_msData.R
\name{validate_ms_settings}
\alias{validate_ms_settings}
\title{validate_ms_settings}
\usage{
validate_ms_settings(value = NULL)
}
\arguments{
\item{value}{A named settings list with \emph{call}, \emph{algorithm} and
\emph{parameters} entries.}
}
\value{
A logical value of length 1.
}
\description{
Validates a settings list for data processing within the msData object.
}
