% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_S3.R
\name{validate_object}
\alias{validate_object}
\title{S3 generic \code{validate_object}}
\usage{
validate_object(x, ...)
}
\arguments{
\item{x}{An object to be validated.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S3 generic to validate_object the object \code{x}.
}
