% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_S3_settings.R
\name{validate_settings}
\alias{validate_settings}
\title{validate_settings}
\usage{
validate_settings(x = NULL)
}
\arguments{
\item{x}{A settings S3 class object.}
}
\value{
Logical of length 1.
}
\description{
Validates a settings S3 class object for data processing within the
streamFind package.
}
