% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Workflow.R
\name{Workflow}
\alias{Workflow}
\title{Workflow Metadata Container}
\usage{
Workflow(processing_steps = list())
}
\arguments{
\item{processing_steps}{A list of \code{Method} objects or compatible lists.}
}
\value{
A \code{Workflow} object.
}
\description{
Ordered list of \code{Method} metadata objects.
}
