% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{add}
\alias{add}
\title{Generic \code{add}}
\usage{
add(x, ...)
}
\arguments{
\item{x}{An object to which data will be added.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S7 generic to add data to \code{x}.
}
