% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_002_peakGrouping.R
\name{addAdjustedRetentionTime}
\alias{addAdjustedRetentionTime}
\title{addAdjustedRetentionTime}
\usage{
addAdjustedRetentionTime(object, pat)
}
\arguments{
\item{object}{An \linkS4class{msData} object.}

\item{pat}{A \linkS4class{featureGroups} object from \pkg{patRoon}.}
}
\description{
Function to add adjusted retention time information
to the spectra slot of each \linkS4class{msAnalysis} in the
\linkS4class{msData} object.
}
