% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Analyses.R
\name{Analyses}
\alias{Analyses}
\alias{validate_object.Analyses}
\title{Generic Analyses}
\usage{
Analyses(analyses = list(), results = list())

\method{validate_object}{Analyses}(x)
}
\arguments{
\item{analyses}{A list of analyses, where each element is a data entry or a connection to a raw data file.}

\item{results}{A list of results, where each element is specific \code{\link{Results}} child class.}

\item{x}{An \code{Analyses} object.}
}
\description{
The \code{Analyses} class is used harmonize the interface to data entries or links to
raw data files. \code{Analyses} child classes are used to enable a specific structure and methods
for each type of data.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate_object(Analyses)}: Validate the Analyses object, returning \code{NULL} if valid or an error if not.

}}
