% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Analyses.R
\name{Analyses}
\alias{Analyses}
\title{Generic Analyses}
\usage{
Analyses(analyses = list(), results = list())
}
\arguments{
\item{analyses}{A list of analyses, where each element is a data entry or a connection to a raw data file.}

\item{results}{A list of results, where each element is specific \code{\link{Results}} child class.}
}
\description{
The Analyses class is used harmonize the interface to data entries or links to raw data files. Analyses child classes are used to enable a specific structure and methods for each type of data.
}
\section{Slots}{

\describe{
\item{\code{data_type}}{(getter/setter) A character string indicating the type of data.}

\item{\code{possible_formats}}{(getter/setter) A character vector of possible formats for the raw data.}

\item{\code{analyses}}{(getter/setter) A list of analyses, where each element is a data entry or a connection to a raw data file.}

\item{\code{results}}{(getter/setter) A list of results, where each element is specific \code{\link{Results}} child class.}

\item{\code{info}}{(getter) A data frame containing information about the analyses.}
}}

