% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_S3_Analysis.R
\name{Analysis}
\alias{Analysis}
\title{\emph{Analysis} S3 class constructor, methods and functions}
\usage{
Analysis(
  name = NA_character_,
  replicate = NA_character_,
  blank = NA_character_
)
}
\arguments{
\item{name}{Character of length one with the name of the analysis.}

\item{replicate}{Character of length one with the analysis replicate group name.}

\item{blank}{Character of length one with the associated blank replicate group name.}
}
\value{
An \emph{Analysis} S3 class object.
}
\description{
Creates an Analysis S3 class object.
}
