% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_AuditTrail.R
\name{AuditTrail}
\alias{AuditTrail}
\alias{validate_object.AuditTrail}
\alias{add.AuditTrail}
\alias{as.data.table.AuditTrail}
\alias{show.AuditTrail}
\alias{save.AuditTrail}
\alias{read.AuditTrail}
\title{Audit Trail Register}
\usage{
AuditTrail(entries = list())

\method{validate_object}{AuditTrail}(x)

\method{add}{AuditTrail}(x, value)

\method{as.data.table}{AuditTrail}(x, ...)

\method{show}{AuditTrail}(x, ...)

\method{save}{AuditTrail}(x, file = "entries.json")

\method{read}{AuditTrail}(x, file)
}
\arguments{
\item{x}{An \code{AuditTrail} object.}

\item{value}{An object to be added to the \code{AuditTrail} as an entry.}

\item{file}{A character string representing the file path from which the \code{AuditTrail} should be read. The possible formats are JSON or RDS.}
}
\description{
The \link{AuditTrail} class is a list of \link{AuditTrailEntry} class
objects.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate_object(AuditTrail)}: Validate the AuditTrail object, returning NULL if valid.

\item \code{add(AuditTrail)}: Add an entry to the \code{AuditTrail} list,
returning the updated \code{AuditTrail}.

\item \code{as.data.table(AuditTrail)}: Convert the \code{AuditTrail} object to a data frame.

\item \code{show(AuditTrail)}: Show the \code{AuditTrail} entries in a human-readable format.

\item \code{save(AuditTrail)}: Save the \code{AuditTrail} object to a file.

\item \code{read(AuditTrail)}: Read an \code{AuditTrail} object from a file.

}}
