% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_AuditTrail.R
\name{AuditTrail}
\alias{AuditTrail}
\title{\strong{AuditTrail} S7 class constructor, slots and methods}
\usage{
AuditTrail(engine_type = NA_character_)
}
\arguments{
\item{engine_type}{A character with the engine type.}
}
\value{
A AuditTrail S7 class object.
}
\description{
Creates an AuditTrail S7 class object.
}
\section{Slots}{

\describe{
\item{\code{entries}}{\code{List} of entries with the audit trail information.}

\item{\code{length}}{Dynamic \code{Numeric} returning the number of headers.}
}}

\section{Methods}{

\itemize{
\item \code{$} retrieves the entries.
\item \code{add} adds an entry to the audit trail.
\item \code{length} returns the number of headers.
\item \code{show} prints the headers.
}
}

