% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_AuditTrail.R
\name{AuditTrail}
\alias{AuditTrail}
\title{AuditTrail Class}
\usage{
AuditTrail()
}
\description{
The AuditTrail class is used to store a list of entries, where each entry is an
instance of the \link{AuditTrailEntry} class.
}
\section{Slots}{

\describe{
\item{\code{entries}}{A list of entries, where each entry is an instance of the
\link{AuditTrailEntry} class.}
}}

