% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_class_R6MS.R
\name{build_features_table_from_patRoon}
\alias{build_features_table_from_patRoon}
\title{build_features_table_from_patRoon}
\usage{
build_features_table_from_patRoon(pat, self)
}
\arguments{
\item{pat}{A \linkS4class{features} or \linkS4class{featureGroups} object
from the package \pkg{patRoon}.}

\item{self}{An `R6MS` object. When applied within the R6, the self object.}
}
\value{
A list of with a features \linkS4class{data.table} for each analysis.
}
\description{
build_features_table_from_patRoon
}
