% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Cache.R
\name{Cache}
\alias{Cache}
\alias{save_cache.Cache}
\alias{load_cache.Cache}
\alias{get_cache_info.Cache}
\alias{clear_cache.Cache}
\alias{size.Cache}
\title{Cache S3 Class}
\usage{
Cache(db)

\method{save_cache}{Cache}(x, name, hash, description, data, ...)

\method{load_cache}{Cache}(x, name, hash, ...)

\method{get_cache_info}{Cache}(x, ...)

\method{clear_cache}{Cache}(x, value = NULL, ...)

\method{size}{Cache}(x, unit = "auto", ...)
}
\arguments{
\item{db}{Full path to the unified DuckDB file.}

\item{x}{Cache object}

\item{name}{Name of cache entry (not used, for consistency)}

\item{hash}{Hash string for cache table}

\item{description}{Description of cache entry}

\item{data}{Any R object to cache (will be serialized)}

\item{value}{Character vector of cache entry names to delete. If NULL (default), deletes entire cache database.}

\item{unit}{Optional unit for size, possible values: "bytes", "KB", "MB", "GB" or "auto" (default).}
}
\value{
An object of class \code{Cache}.

Cached R object (deserialized)

Data.frame with cache metadata

Integer
}
\description{
S3 class to manage the \code{Cache} table in a unified project DuckDB file for caching and loading data.
}
\section{Methods (by generic)}{
\itemize{
\item \code{save_cache(Cache)}: Save data to cache table.

\item \code{load_cache(Cache)}: Load data from cache table.

\item \code{get_cache_info(Cache)}: Get cache information table.

\item \code{clear_cache(Cache)}: Reset cache (delete all cache entries or specific entries by name)

\item \code{size(Cache)}: Get size of Cache.duckdb file in bytes.

}}
