% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_external_tools.R
\name{check_external_tools}
\alias{check_external_tools}
\title{Check status of external tools}
\usage{
check_external_tools(verbose = TRUE)
}
\arguments{
\item{verbose}{If \code{TRUE}, print a short status summary.}
}
\value{
A named logical vector for Java and MetFrag.
}
\description{
Check status of external tools
}
