% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_S3.R
\name{clear_cache}
\alias{clear_cache}
\title{S3 generic \code{clear_cache}}
\usage{
clear_cache(x, ...)
}
\arguments{
\item{x}{An object from which the cache will be cleared.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S3 generic to clear the cache from \code{x}.
}
