% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_clear_cache.R
\name{clear_cache}
\alias{clear_cache}
\title{clear_cache}
\usage{
clear_cache(what = NULL, file = "cache.sqlite")
}
\arguments{
\item{what}{A character string specifying the cache to remove. If NULL, a list of available
caches is shown.}

\item{file}{A character string specifying the cache file to use. Default is "cache.sqlite".}
}
\description{
Clears cached data using the approach as in the patRoon package.
}
\references{
\insertRef{patroon01}{StreamFind}

\insertRef{patroon02}{StreamFind}
}
