% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ms_data_access.R
\name{clusterMSnToPatRoon}
\alias{clusterMSnToPatRoon}
\title{clusterMSnToPatRoon}
\usage{
clusterMSnToPatRoon(
  cl_plists,
  mlists,
  targets,
  clusteringMethod,
  clusteringUnit,
  clusteringWindow,
  minIntensityPost
)
}
\arguments{
\item{cl_plists}{...}

\item{mlists}{...}

\item{targets}{...}

\item{clusteringMethod}{...}

\item{clusteringUnit}{...}

\item{clusteringWindow}{...}

\item{minIntensityPost}{...}
}
\value{
A \linkS4class{MSPeakLists} object with clustered data.
}
\description{
Function to cluster spectra and convert to
an \linkS4class{MSPeakLists} object.
}
