% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Config.R
\name{Config}
\alias{Config}
\title{Config Class}
\usage{
Config(parameters = list())
}
\arguments{
\item{parameters}{A list of \link{ConfigParameter} objects.}
}
\description{
Class representing a configuration object composed of multiple
\link{ConfigParameter} objects.
}
\section{Slots}{

\describe{
\item{\code{config_frame}}{(getter) A data frame representation of the configuration parameters.}
}}

