% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Config.R
\name{ConfigCache}
\alias{ConfigCache}
\title{Configuration Parameter for Caching}
\usage{
ConfigCache()
}
\description{
Class representing a configuration for the caching behavior.
}
\section{Slots}{

\describe{
\item{\code{value}}{Logical indicating whether to enable or disable caching.}

\item{\code{mode}}{Character indicating the caching mode (e.g., "rds" or "sqlite").}

\item{\code{folder}}{Character indicating the folder for caching (for "rds" mode).}

\item{\code{file}}{Character indicating the file for caching (for "sqlite" mode).}

\item{\code{size}}{(getter) Size of the cache.}

\item{\code{info}}{(getter) Information about the cache.}
}}

