% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Config.R
\name{ConfigDurationNotifications}
\alias{ConfigDurationNotifications}
\title{Configuration Parameter for Duration of Notifications}
\usage{
ConfigDurationNotifications(value = 10)
}
\arguments{
\item{value}{Duration in seconds for pop-up notifications.}
}
\description{
Class representing a configuration for the duration of pop-up notifications in the
app, inhiberiting from \link{ConfigParameter}.
}
