% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Config.R
\name{ConfigParameter}
\alias{ConfigParameter}
\alias{validate_object.ConfigParameter}
\title{Configuration Parameter}
\usage{
ConfigParameter(name = NA_character_, description = NA_character_)

\method{validate_object}{ConfigParameter}(x)
}
\arguments{
\item{name}{Name of the parameter.}

\item{description}{Description of the parameter.}

\item{x}{An object to validate.}
}
\description{
The \code{ConfigParameter} S3 class represents a configuration parameter in StreamFind.
The \code{ConfigParameter} is a list of at least two elements - \code{name} and \code{description}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate_object(ConfigParameter)}: Validates a \code{ConfigParameter} object, returning NULL if valid.

}}
