% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_001_convert_trim_files_.R
\name{convert_ms_files}
\alias{convert_ms_files}
\title{convert_ms_files}
\usage{
convert_ms_files(
  files = NULL,
  outputFormat = "mzML",
  outputPath = NULL,
  optList = NULL
)
}
\arguments{
\item{files}{A vector with file full path/s for conversion.}

\item{outputFormat}{A character string with the format of the output file.
Possible values are "mzML" and "mzXML".}

\item{outputPath}{The directory to place the output file. When \code{NULL}
the directory of the original file is used.}

\item{optList}{A list of further options to be passed to the respective
conversion tool. See
\url{https://proteowizard.sourceforge.io/tools/msconvert.html}
for commands available for conversion of MS files with
\href{https://proteowizard.sourceforge.io/download.html}{msConvert}.}
}
\description{
Function to convert vendor MS file formats to mzML/mzXML using
the command line of \code{MSConvert} from \emph{ProteoWizard}.
}
\note{
When giving paths make sure that single backslash is changed to
double backslash of to single forward slash.
}
\examples{

\dontrun{
# The full file path
file = E:/ms_file.d

# Options to centroid MS1 and MS2 data in file
optList <- list(filter = "peakPicking vendor msLevel=1-2")

# Convert to mzML
convert_ms_files(file, outfile = "mzML", optList = optList)
}

}
\references{
\insertRef{proteo01}{streamFind}

\insertRef{proteo02}{streamFind}
}
