% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general_settings.R
\name{createSettings}
\alias{createSettings}
\title{createSettings}
\usage{
createSettings(
  call = NA_character_,
  algorithm = NA_character_,
  settings = list()
)
}
\arguments{
\item{call}{The name of the function where the algorithm and settings are used.}

\item{algorithm}{A character string with the name of the algorithm to be used.}

\item{settings}{A list of settings dependent on the algorithm used.}
}
\value{
A \linkS4class{settings} object.
}
\description{
Creates a \linkS4class{settings} object for a given processing step.
}
