% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_methods_datatable.R
\name{plotTICs,data.table-method}
\alias{plotTICs,data.table-method}
\alias{plotTICs,data.table,data.table-method}
\title{plotTICs-data.table}
\usage{
\S4method{plotTICs}{data.table}(
  object,
  analyses = NULL,
  colorBy = "analyses",
  title = NULL,
  interactive = FALSE
)
}
\arguments{
\item{analyses}{A numeric or character vector with the index or analysis name/s.
When \code{NULL}, all the analyses are used.}

\item{colorBy}{A length one character vector applicable to the respective method.}

\item{title}{A character string with the plot tile.}

\item{interactive}{Logical value, set to \code{TRUE} to use
the \pkg{plotly} instead of \pkg{base}. The default is \code{FALSE}.}
}
\description{
Plots a total ion chromatogram (TIC) from the
\link[data.table]{data.table} obtained by the S4 method \link{TICs}.
The colorBy argument can be "analyses" or "replicates" to color
the plot by analyses or by analysis replicates.
}
