% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_DB_Analyses.R
\name{DB_Analyses}
\alias{DB_Analyses}
\alias{validate_object.DB_Analyses}
\title{Base class for database-backed Analyses}
\usage{
DB_Analyses(db, data_type)

\method{validate_object}{DB_Analyses}(x)
}
\arguments{
\item{db}{Path to DuckDB file.}

\item{data_type}{Data type string (e.g., "MassSpec").}

\item{x}{DB_Analyses object.}
}
\value{
An object of class \code{DB_Analyses}.
}
\description{
Minimal base class to represent analyses stored in a DuckDB file.
Requires a \code{db} path and a \code{type} identifier.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate_object(DB_Analyses)}: Validate the DB_Analyses object.

}}
