% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_DB_MassSpecEngine.R
\name{DB_MassSpecEngine}
\alias{DB_MassSpecEngine}
\title{DB_Engine dedicated to Mass Spectrometry (MS) data}
\description{
R6 child of DB_Engine for MassSpec data that uses DB_MassSpecAnalyses for on-disk storage.
}
\section{Super class}{
\code{\link[StreamFind:DB_Engine]{StreamFind::DB_Engine}} -> \code{DB_MassSpecEngine}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{Analyses}}{A DB_MassSpecAnalyses object backed by DuckDB.}

\item{\code{NonTargetAnalysis}}{A DB_MassSpecResults_NonTargetAnalysis object backed by DuckDB.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DB_MassSpecEngine-new}{\code{DB_MassSpecEngine$new()}}
\item \href{#method-DB_MassSpecEngine-clone}{\code{DB_MassSpecEngine$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="DB_Engine" data-id="add_audit_entry"><a href='../../StreamFind/html/DB_Engine.html#method-DB_Engine-add_audit_entry'><code>StreamFind::DB_Engine$add_audit_entry()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="DB_Engine" data-id="add_metadata"><a href='../../StreamFind/html/DB_Engine.html#method-DB_Engine-add_metadata'><code>StreamFind::DB_Engine$add_metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="DB_Engine" data-id="add_workflow"><a href='../../StreamFind/html/DB_Engine.html#method-DB_Engine-add_workflow'><code>StreamFind::DB_Engine$add_workflow()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="DB_Engine" data-id="clear_cache"><a href='../../StreamFind/html/DB_Engine.html#method-DB_Engine-clear_cache'><code>StreamFind::DB_Engine$clear_cache()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="DB_Engine" data-id="clear_result_databases"><a href='../../StreamFind/html/DB_Engine.html#method-DB_Engine-clear_result_databases'><code>StreamFind::DB_Engine$clear_result_databases()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="DB_Engine" data-id="get_audit_trail"><a href='../../StreamFind/html/DB_Engine.html#method-DB_Engine-get_audit_trail'><code>StreamFind::DB_Engine$get_audit_trail()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="DB_Engine" data-id="get_cache_info"><a href='../../StreamFind/html/DB_Engine.html#method-DB_Engine-get_cache_info'><code>StreamFind::DB_Engine$get_cache_info()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="DB_Engine" data-id="get_cache_size"><a href='../../StreamFind/html/DB_Engine.html#method-DB_Engine-get_cache_size'><code>StreamFind::DB_Engine$get_cache_size()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="DB_Engine" data-id="get_db_table_info"><a href='../../StreamFind/html/DB_Engine.html#method-DB_Engine-get_db_table_info'><code>StreamFind::DB_Engine$get_db_table_info()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="DB_Engine" data-id="get_engine_info"><a href='../../StreamFind/html/DB_Engine.html#method-DB_Engine-get_engine_info'><code>StreamFind::DB_Engine$get_engine_info()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="DB_Engine" data-id="get_metadata"><a href='../../StreamFind/html/DB_Engine.html#method-DB_Engine-get_metadata'><code>StreamFind::DB_Engine$get_metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="DB_Engine" data-id="get_workflow"><a href='../../StreamFind/html/DB_Engine.html#method-DB_Engine-get_workflow'><code>StreamFind::DB_Engine$get_workflow()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="DB_Engine" data-id="list_db_tables"><a href='../../StreamFind/html/DB_Engine.html#method-DB_Engine-list_db_tables'><code>StreamFind::DB_Engine$list_db_tables()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="DB_Engine" data-id="print"><a href='../../StreamFind/html/DB_Engine.html#method-DB_Engine-print'><code>StreamFind::DB_Engine$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="DB_Engine" data-id="project_path"><a href='../../StreamFind/html/DB_Engine.html#method-DB_Engine-project_path'><code>StreamFind::DB_Engine$project_path()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="DB_Engine" data-id="query_db"><a href='../../StreamFind/html/DB_Engine.html#method-DB_Engine-query_db'><code>StreamFind::DB_Engine$query_db()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="DB_Engine" data-id="run"><a href='../../StreamFind/html/DB_Engine.html#method-DB_Engine-run'><code>StreamFind::DB_Engine$run()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="DB_Engine" data-id="run_workflow"><a href='../../StreamFind/html/DB_Engine.html#method-DB_Engine-run_workflow'><code>StreamFind::DB_Engine$run_workflow()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DB_MassSpecEngine-new"></a>}}
\if{latex}{\out{\hypertarget{method-DB_MassSpecEngine-new}{}}}
\subsection{Method \code{new()}}{
Initialize DB_MassSpecEngine
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB_MassSpecEngine$new(
  project_path = "data",
  metadata = NULL,
  workflow = NULL,
  files = NULL,
  centroid = FALSE,
  levels = c(1, 2),
  configuration = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project_path}}{Path to StreamFind project folder (.sf). DB_Engine will place main.duckdb inside it. If reusing an existing project_path, DB_Engine will connect to the existing Engine.duckdb file and preserve previous data. If the project_path does not exist, DB_Engine will create the directory and initialize a new Engine.duckdb file.}

\item{\code{metadata}}{A named list or a \code{Metadata} object. Each list element must have length one.
There is no type restrictions except for name (must be type character), author (must be type
character) and date (must be class POSIXct or POSIXt). A file entry must be a character string
with a full path to a file.}

\item{\code{workflow}}{A \code{Workflow} object, a list of \code{ProcessingStep} objects or a character string
with a full path to a \emph{rds} or \emph{json} file containing a \code{Workflow}.}

\item{\code{files}}{A character vector with full file path/s of \emph{mzML}
or \emph{mzXML} file/s or a data.frame with the columns \emph{file}, \emph{replicate}
and \emph{blank} with the full file path of \emph{mzML} or \emph{mzXML} file/s
(character), the replicate group name (character) and the associated blank
replicate group name (character). Note the file paths must have the extension
\emph{.mzML} or \emph{.mzXML}.}

\item{\code{centroid}}{Logical (length 1). Set to \code{TRUE} to centroid data when converting from vendor formats to mzML.}

\item{\code{levels}}{Integer vector with the levels of MS data.}

\item{\code{configuration}}{A list of configuration parameters to persist or use for the engine/database.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DB_MassSpecEngine-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DB_MassSpecEngine-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB_MassSpecEngine$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
