% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Metadata.R
\name{EngineMetadata}
\alias{EngineMetadata}
\title{Engine Metadata}
\usage{
EngineMetadata(x = list(), data_type = NA_character_)
}
\arguments{
\item{x}{A named list of metadata entries. Default is an empty list.}

\item{data_type}{A character string with the data type of the engine.
Default is \code{NA_character_}.}
}
\description{
The \code{EngineMetadata} class is a subclass of the \link{Metadata} class.
It is used to store metadata of data processing engines.
}
\seealso{
\link{Metadata}
}
