% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_R6_msData.R
\name{extract_time_alignment}
\alias{extract_time_alignment}
\title{extract_time_alignment}
\usage{
extract_time_alignment(pat, self)
}
\arguments{
\item{pat}{A \linkS4class{featureGroups} object from \pkg{patRoon}.}

\item{self}{An \code{msData} object. When applied within the R6, the self object.}
}
\description{
Function to extract adjusted retention time information from
alignment results when using \code{xcms3} as algorithm for grouping and retention
time alignment.
}
