% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_S3.R
\name{get_chromatograms_peaks}
\alias{get_chromatograms_peaks}
\title{S3 generic \code{get_chromatograms_peaks}}
\usage{
get_chromatograms_peaks(x, ...)
}
\arguments{
\item{x}{An object from which the chromatograms peaks will be retrieved.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S3 generic to get chromatograms peaks from \code{x}.
}
