% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_S3.R
\name{get_components}
\alias{get_components}
\title{S3 generic \code{get_components}}
\usage{
get_components(x, ...)
}
\arguments{
\item{x}{An object from which the components will be retrieved.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S3 generic to get components from \code{x}.
}
