% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_default_ProcessingSettings.R
\name{get_default_ProcessingSettings}
\alias{get_default_ProcessingSettings}
\title{get_default_ProcessingSettings}
\usage{
get_default_ProcessingSettings(call = NA_character_, algorithm = NA_character_)
}
\arguments{
\item{call}{Character (length 1) with the method call name.}

\item{algorithm}{Character (length 1) with the algorithm name.}
}
\value{
A ProcessingSettings S3 class object with subclass as defined by
\code{call} and \code{algorithm}.
}
\description{
Saves on disk a ProcessingSettings S3 class object as the
defined \code{format}, \code{path} and \code{name}.
}
