% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_install_external_tools.R
\name{get_external_dir}
\alias{get_external_dir}
\title{Path to the external tools directory}
\usage{
get_external_dir()
}
\value{
The normalized path to \code{~/.streamfind/external/}.
}
\description{
Returns (and creates if needed) \code{~/.streamfind/external/}.
}
