% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_S3.R
\name{get_features_ms2}
\alias{get_features_ms2}
\title{S3 generic \code{get_features_ms2}}
\usage{
get_features_ms2(x, ...)
}
\arguments{
\item{x}{An object from which the features MS2 will be retrieved.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S3 generic to get features MS2 from \code{x}.
}
