% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_S3.R
\name{get_formulas}
\alias{get_formulas}
\title{S3 generic \code{get_formulas}}
\usage{
get_formulas(x, ...)
}
\arguments{
\item{x}{An object from which the formulas will be retrieved.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S3 generic to get formulas from \code{x}.
}
