% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_S3.R
\name{get_groups_ms1}
\alias{get_groups_ms1}
\title{S3 generic \code{get_groups_ms1}}
\usage{
get_groups_ms1(x, ...)
}
\arguments{
\item{x}{An object from which the groups MS1 will be retrieved.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S3 generic to get groups MS1 from \code{x}.
}
