% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_external_tools.R
\name{get_metfrag_path}
\alias{get_metfrag_path}
\title{Locate the MetFrag CL JAR}
\usage{
get_metfrag_path()
}
\value{
Path to the JAR file, or \code{NA} if not found.
}
\description{
Checks \verb{~/.streamfind/tools/metfrag/MetFragCL.jar}.
}
