% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_S3.R
\name{get_model_data}
\alias{get_model_data}
\title{S3 generic \code{get_model_data}}
\usage{
get_model_data(x, ...)
}
\arguments{
\item{x}{An object from which the model data will be retrieved.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S3 generic to get model data from \code{x}.
}
