% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_S3.R
\name{get_spectra}
\alias{get_spectra}
\title{Generic \code{get_spectra}}
\usage{
get_spectra(x, ...)
}
\arguments{
\item{x}{An object from which the spectra will be retrieved.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S3 generic to get spectra from \code{x}.
}
