% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_external_tools.R
\name{get_streamfind_dir}
\alias{get_streamfind_dir}
\title{Path to the streamfind home directory}
\usage{
get_streamfind_dir()
}
\value{
The normalized path to \verb{~/.streamfind/}.
}
\description{
Returns (and creates if needed) \verb{~/.streamfind/}.
}
