% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_S3_headers.R
\name{Headers}
\alias{Headers}
\alias{validate.Headers}
\alias{asJSON.Headers}
\alias{as.Headers}
\title{\strong{Headers} S3 class constructor, methods and functions}
\usage{
Headers(...)

\method{validate}{Headers}(x)

\method{asJSON}{Headers}(x)

as.Headers(value)
}
\arguments{
\item{...}{Arguments to be added as headers. Note that all given arguments
must be of length one. There is no type limitations. The name of the argument
becomes the name of the header entry. Alternatively, a single list argument
with the headers (all length one) can be given.}

\item{x}{A Headers S3 class object.}

\item{value}{List to be checked and converted to Headers S3 class.}
}
\value{
A Headers S3 class object.
}
\description{
Creates a Headers S3 class object.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate(Headers)}: Validates a Headers S3 class object, returning a logical value of length one.

\item \code{asJSON(Headers)}: Converts a Headers S3 class object to a JSON string.

}}
\section{Functions}{
\itemize{
\item \code{as.Headers()}: Converts the argument value into a Headers S3 class object.

}}
\note{
If an argument or element name is given, it must be type character.
If an argument or element path is given, it must be type character and exist.
If an argument or element date is given, it must be class POSIXct or POSIXt.
If given date is character, conversion to class POSIXct or POSIXt is
attempted.
}
