% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{load}
\alias{load}
\title{Generic \code{load}}
\usage{
load(x, ...)
}
\arguments{
\item{x}{An object from which the data will be loaded.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S7 generic to load data from \code{x}.
}
