% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_S3.R
\name{load_cache}
\alias{load_cache}
\title{Generic \code{load_cache}}
\usage{
load_cache(x, ...)
}
\arguments{
\item{x}{An object from which the cache will be loaded.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S3 generic to load cache from \code{x}.
}
