% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{load_chromatograms}
\alias{load_chromatograms}
\title{Generic \code{load_chromatograms}}
\usage{
load_chromatograms(x, ...)
}
\arguments{
\item{x}{An object from which the chromatograms will be loaded.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S7 generic to load chromatograms from \code{x}.
}
