% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_make_ms_targets.R
\name{make_ms_targets}
\alias{make_ms_targets}
\title{Function to make targets for parsing data within MassSpecData class methods}
\usage{
make_ms_targets(
  mz = NULL,
  rt = NULL,
  drift = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  id = NULL
)
}
\arguments{
\item{mz}{A vector with target \emph{m/z} value/s (in Da) or a two columns
data.table or data.frame named \code{mzmin} and \code{mzmax} with minimum and maximum
\emph{m/z} values (in Da), respectively. Alternatively, \emph{m/z} (in Da)
and retention time values (in seconds) can be given as one data.table or
data.frame with columns named \code{mz} and \code{rt} and/or \code{drift}. Then, the
deviations given in the \code{ppm}, \code{sec} and \code{millisec} arguments
are used to calculate the ranges. Also works with a data.table or data.frame
with minimum and maximum values of \emph{m/z}, retention time and drift time
targets. In this case, the column names must be \code{mzmin}, \code{mzmax}, \code{rtmin},
\code{rtmax}, \code{driftmin} and \code{driftmax}. Note that when mass/time ranges
are given, the \code{ppm}, \code{sec} and \code{millisec} arguments are
not used.}

\item{rt}{A vector with target retention time values (in seconds) or
a two columns data.table or data.frame with minimum and maximum retention
time values (in seconds).}

\item{drift}{A vector with target drift time values (in milliseconds) or
a two columns data.table or data.frame with minimum and maximum drift
time values (in milliseconds).}

\item{ppm}{Numeric of length one with the mass deviation, in ppm.}

\item{sec}{Numeric of length one with the time deviation, in seconds.}

\item{millisec}{Numeric of length one with the drift time deviation,
in milliseconds.}

\item{id}{Character with the same length as \emph{m/z} and retention time
targets to be used as identifier/s. When not given, the id is built as a
combination of the \emph{m/z} and retention time ranges or values.}
}
\value{
A data.frame with columns: \emph{id}, \emph{mz}, \emph{rt}, \emph{drift}, \emph{mzmin},
\emph{mzmax}, \emph{rtmin}, \emph{rtmax}, \emph{driftmin} and \emph{driftmax}.
}
\description{
Helper function to build \emph{m/z} and retention time and/or
drift time targets for searching the MS data. Each target is composed of an
id, \emph{m/z} (in Da), retention time (in seconds) and drift time (in
milliseconds) ranges. When mass is defined without time variables, the
retention and drift time ranges return 0 and vice versa.
}
