% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_S3_MassSpecAnalysis.R
\name{MassSpecAnalysis}
\alias{MassSpecAnalysis}
\alias{validate.MassSpecAnalysis}
\alias{print.MassSpecAnalysis}
\alias{asJSON.MassSpecAnalysis}
\alias{as.MassSpecAnalysis}
\alias{parse.MassSpecAnalysis}
\title{\strong{MassSpecAnalysis} S3 class constructor, methods and functions}
\usage{
MassSpecAnalysis(
  name = NA_character_,
  replicate = NA_character_,
  blank = NA_character_,
  file = NA_character_,
  format = NA_character_,
  type = NA_character_,
  time_stamp = NA_character_,
  spectra_number = NA_integer_,
  spectra_mode = NA_character_,
  spectra_levels = NA_integer_,
  mz_low = NA_real_,
  mz_high = NA_real_,
  rt_start = NA_real_,
  rt_end = NA_real_,
  polarity = NA_character_,
  has_ion_mobility = FALSE,
  chromatograms_number = NA_integer_,
  software = data.table(),
  instrument = data.table(),
  run = data.table(),
  spectra = data.table(),
  chromatograms = data.table(),
  features = data.table(),
  metadata = list()
)

\method{validate}{MassSpecAnalysis}(x = NULL)

\method{print}{MassSpecAnalysis}(x)

\method{asJSON}{MassSpecAnalysis}(x)

as.MassSpecAnalysis(value)

parse.MassSpecAnalysis(files = NULL, runParallel = FALSE)
}
\arguments{
\item{name}{\emph{mzML} or \emph{mzXML} file name without extension.}

\item{replicate}{Character with length one, representing the analysis
replicate group name.}

\item{blank}{Character with length one, representing the associated blank
replicate group name.}

\item{file}{\emph{mzML} or \emph{mzXML} full file path (with extension).}

\item{format}{Character with length one. One of \emph{mzML} or \emph{mzXML}.}

\item{type}{Character with length one. Possible values as "MS" for only MS1
spectral data, "MS/MS" for tandem spectral data (i.e., MS1 and MS2) or "SRM"
for selected reaction monitoring data (i.e., no spectra only chromatograms).}

\item{time_stamp}{Character with length one, representing the start time and
date of the data acquisition.}

\item{spectra_number}{Integer with the number of spectra in the file.}

\item{spectra_mode}{Character with length one. Possible values are \emph{centroid}
, \emph{profile} or \code{NA} for centroided data, data in profile mode or mode not
defined or absence spectra, respectively.}

\item{spectra_levels}{Integer vector with the number of levels in the
spectra (i.e., 1 or 1 and 2).}

\item{mz_low}{Numeric length one with the lowest \emph{m/z} value detected
in the spectra, in Da.}

\item{mz_high}{Numeric length one with the highest \emph{m/z} value detected
in the spectra, in Da.}

\item{rt_start}{Numeric length one with the run start time, in seconds.}

\item{rt_end}{Numeric length one with the run end time, in seconds.}

\item{polarity}{Character length one. Possible values are \emph{positive},
\emph{negative} or \emph{both} (the latter refers to polarity switching acquisition).}

\item{has_ion_mobility}{Logical length one for presence or absence of drift
separation from ion mobility.}

\item{chromatograms_number}{Integer with the number of chromatograms in the
file.}

\item{software}{data.table with software info used to acquire/process the
data (content is highly vendor dependent).}

\item{instrument}{data.table with instrument info used to acquire the
data (content is highly vendor dependent).}

\item{run}{data.table run information for each spectrum.}

\item{spectra}{data.table with the raw spectra (only present if loaded).}

\item{chromatograms}{data.table with the raw chromatograms (only present
if loaded).}

\item{features}{data.table with the features from data processing.}

\item{metadata}{List with flexible storage for experimental metadata
(e.g., concentration, location, etc.).}

\item{x}{A MassSpecAnalysis S3 class object.}

\item{value}{A list to be checked and/or converted to MassSpecAnalysis S3 class.}

\item{files}{A character vector with \emph{mzML} or \emph{mzXML} full file path/s.
Alternatively, a data.frame with the column/s file, replicate and blank
with the full file path/s, the replicate group name/s (string) and the
associated blank replicate group name (string).}

\item{runParallel}{Logical of length one. Set to \code{TRUE} for parallel
processing. Note that parallel processing is not always faster due to
overhead.}
}
\value{
An MassSpecAnalysis S3 class object.
}
\description{
Creates a MassSpecAnalysis S3 class object.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate(MassSpecAnalysis)}: Validates a MassSpecAnalysis S3 class object, returning a logical value of
length one.

\item \code{print(MassSpecAnalysis)}: Prints the MassSpecAnalysis S3 class object in the console.

\item \code{asJSON(MassSpecAnalysis)}: Converts a MassSpecAnalysis S3 class object to a JSON string.

}}
\section{Functions}{
\itemize{
\item \code{as.MassSpecAnalysis()}: Converts the argument value in a MassSpecAnalysis S3 class object.

\item \code{parse.MassSpecAnalysis()}: Parses information from \emph{mzML} or \emph{mzXML} file/s and returns a list with
MassSpecAnalysis S3 class object/s. On error, returns \code{NULL}.

}}
