% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_EngineDB_MassSpecEngineDB.R
\name{MassSpecEngineDB}
\alias{MassSpecEngineDB}
\title{EngineDB dedicated to Mass Spectrometry (MS) data}
\description{
R6 child of EngineDB for MassSpec data that uses MassSpecAnalysesDB for on-disk storage.
}
\section{Super class}{
\code{\link[StreamFind:EngineDB]{StreamFind::EngineDB}} -> \code{MassSpecEngineDB}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{Analyses}}{A MassSpecAnalysesDB object backed by DuckDB.}

\item{\code{NonTargetAnalysis}}{A MassSpecResults_NonTargetAnalysisDB object backed by DuckDB.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MassSpecEngineDB-new}{\code{MassSpecEngineDB$new()}}
\item \href{#method-MassSpecEngineDB-add_analyses}{\code{MassSpecEngineDB$add_analyses()}}
\item \href{#method-MassSpecEngineDB-get_analysis_names}{\code{MassSpecEngineDB$get_analysis_names()}}
\item \href{#method-MassSpecEngineDB-get_replicate_names}{\code{MassSpecEngineDB$get_replicate_names()}}
\item \href{#method-MassSpecEngineDB-get_blank_names}{\code{MassSpecEngineDB$get_blank_names()}}
\item \href{#method-MassSpecEngineDB-get_concentrations}{\code{MassSpecEngineDB$get_concentrations()}}
\item \href{#method-MassSpecEngineDB-set_replicate_names}{\code{MassSpecEngineDB$set_replicate_names()}}
\item \href{#method-MassSpecEngineDB-set_blank_names}{\code{MassSpecEngineDB$set_blank_names()}}
\item \href{#method-MassSpecEngineDB-set_concentrations}{\code{MassSpecEngineDB$set_concentrations()}}
\item \href{#method-MassSpecEngineDB-info_analyses}{\code{MassSpecEngineDB$info_analyses()}}
\item \href{#method-MassSpecEngineDB-get_spectra_headers}{\code{MassSpecEngineDB$get_spectra_headers()}}
\item \href{#method-MassSpecEngineDB-get_chromatograms_headers}{\code{MassSpecEngineDB$get_chromatograms_headers()}}
\item \href{#method-MassSpecEngineDB-clone}{\code{MassSpecEngineDB$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="EngineDB" data-id="add_audit_entry"><a href='../../StreamFind/html/EngineDB.html#method-EngineDB-add_audit_entry'><code>StreamFind::EngineDB$add_audit_entry()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="EngineDB" data-id="add_metadata"><a href='../../StreamFind/html/EngineDB.html#method-EngineDB-add_metadata'><code>StreamFind::EngineDB$add_metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="EngineDB" data-id="add_workflow"><a href='../../StreamFind/html/EngineDB.html#method-EngineDB-add_workflow'><code>StreamFind::EngineDB$add_workflow()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="EngineDB" data-id="get_audit_trail"><a href='../../StreamFind/html/EngineDB.html#method-EngineDB-get_audit_trail'><code>StreamFind::EngineDB$get_audit_trail()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="EngineDB" data-id="get_db_table_info"><a href='../../StreamFind/html/EngineDB.html#method-EngineDB-get_db_table_info'><code>StreamFind::EngineDB$get_db_table_info()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="EngineDB" data-id="get_engine_info"><a href='../../StreamFind/html/EngineDB.html#method-EngineDB-get_engine_info'><code>StreamFind::EngineDB$get_engine_info()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="EngineDB" data-id="get_metadata"><a href='../../StreamFind/html/EngineDB.html#method-EngineDB-get_metadata'><code>StreamFind::EngineDB$get_metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="EngineDB" data-id="get_workflow"><a href='../../StreamFind/html/EngineDB.html#method-EngineDB-get_workflow'><code>StreamFind::EngineDB$get_workflow()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="EngineDB" data-id="list_db_tables"><a href='../../StreamFind/html/EngineDB.html#method-EngineDB-list_db_tables'><code>StreamFind::EngineDB$list_db_tables()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="EngineDB" data-id="print"><a href='../../StreamFind/html/EngineDB.html#method-EngineDB-print'><code>StreamFind::EngineDB$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="EngineDB" data-id="query_db"><a href='../../StreamFind/html/EngineDB.html#method-EngineDB-query_db'><code>StreamFind::EngineDB$query_db()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="EngineDB" data-id="run"><a href='../../StreamFind/html/EngineDB.html#method-EngineDB-run'><code>StreamFind::EngineDB$run()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="EngineDB" data-id="run_workflow"><a href='../../StreamFind/html/EngineDB.html#method-EngineDB-run_workflow'><code>StreamFind::EngineDB$run_workflow()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MassSpecEngineDB-new"></a>}}
\if{latex}{\out{\hypertarget{method-MassSpecEngineDB-new}{}}}
\subsection{Method \code{new()}}{
Initialize MassSpecEngineDB
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MassSpecEngineDB$new(
  project_dir = "data.sf",
  metadata = NULL,
  workflow = NULL,
  files = NULL,
  centroid = FALSE,
  levels = c(1, 2),
  configuration = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project_dir}}{Path to StreamFind project folder (.sf). EngineDB will place main.duckdb inside it. If reusing an existing project_dir, EngineDB will connect to the existing Engine.duckdb file and preserve previous data. If the project_dir does not exist, EngineDB will create the directory and initialize a new Engine.duckdb file.}

\item{\code{metadata}}{A named list or a \code{Metadata} object. Each list element must have length one.
There is no type restrictions except for name (must be type character), author (must be type
character) and date (must be class POSIXct or POSIXt). A file entry must be a character string
with a full path to a file.}

\item{\code{workflow}}{A \code{Workflow} object, a list of \code{ProcessingStep} objects or a character string
with a full path to a \emph{rds} or \emph{json} file containing a \code{Workflow}.}

\item{\code{files}}{A character vector with full file path/s of \emph{mzML}
or \emph{mzXML} file/s or a data.frame with the columns \emph{file}, \emph{replicate}
and \emph{blank} with the full file path of \emph{mzML} or \emph{mzXML} file/s
(character), the replicate group name (character) and the associated blank
replicate group name (character). Note the file paths must have the extension
\emph{.mzML} or \emph{.mzXML}.}

\item{\code{centroid}}{Logical (length 1). Set to \code{TRUE} to centroid data when converting from vendor formats to mzML.}

\item{\code{levels}}{Integer vector with the levels of MS data.}

\item{\code{configuration}}{A list of configuration parameters to persist or use for the engine/database.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MassSpecEngineDB-add_analyses"></a>}}
\if{latex}{\out{\hypertarget{method-MassSpecEngineDB-add_analyses}{}}}
\subsection{Method \code{add_analyses()}}{
Add analyses (append; overwrites duplicates)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MassSpecEngineDB$add_analyses(files, centroid = FALSE, levels = c(1, 2))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{files}}{A character vector with full file path/s of \emph{mzML}
or \emph{mzXML} file/s or a data.frame with the columns \emph{file}, \emph{replicate}
and \emph{blank} with the full file path of \emph{mzML} or \emph{mzXML} file/s
(character), the replicate group name (character) and the associated blank
replicate group name (character). Note the file paths must have the extension
\emph{.mzML} or \emph{.mzXML}.}

\item{\code{centroid}}{Logical (length 1). Set to \code{TRUE} to centroid data when converting from vendor formats to mzML.}

\item{\code{levels}}{Integer vector with the levels of MS data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MassSpecEngineDB-get_analysis_names"></a>}}
\if{latex}{\out{\hypertarget{method-MassSpecEngineDB-get_analysis_names}{}}}
\subsection{Method \code{get_analysis_names()}}{
Get analysis names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MassSpecEngineDB$get_analysis_names()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MassSpecEngineDB-get_replicate_names"></a>}}
\if{latex}{\out{\hypertarget{method-MassSpecEngineDB-get_replicate_names}{}}}
\subsection{Method \code{get_replicate_names()}}{
Get replicate names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MassSpecEngineDB$get_replicate_names()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MassSpecEngineDB-get_blank_names"></a>}}
\if{latex}{\out{\hypertarget{method-MassSpecEngineDB-get_blank_names}{}}}
\subsection{Method \code{get_blank_names()}}{
Get blank names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MassSpecEngineDB$get_blank_names()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MassSpecEngineDB-get_concentrations"></a>}}
\if{latex}{\out{\hypertarget{method-MassSpecEngineDB-get_concentrations}{}}}
\subsection{Method \code{get_concentrations()}}{
Get concentrations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MassSpecEngineDB$get_concentrations()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MassSpecEngineDB-set_replicate_names"></a>}}
\if{latex}{\out{\hypertarget{method-MassSpecEngineDB-set_replicate_names}{}}}
\subsection{Method \code{set_replicate_names()}}{
Set replicate names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MassSpecEngineDB$set_replicate_names(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Character vector of replicate names matching analyses.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MassSpecEngineDB-set_blank_names"></a>}}
\if{latex}{\out{\hypertarget{method-MassSpecEngineDB-set_blank_names}{}}}
\subsection{Method \code{set_blank_names()}}{
Set blank names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MassSpecEngineDB$set_blank_names(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Character vector of blank names matching analyses.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MassSpecEngineDB-set_concentrations"></a>}}
\if{latex}{\out{\hypertarget{method-MassSpecEngineDB-set_concentrations}{}}}
\subsection{Method \code{set_concentrations()}}{
Set concentrations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MassSpecEngineDB$set_concentrations(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Numeric vector of concentrations matching analyses.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MassSpecEngineDB-info_analyses"></a>}}
\if{latex}{\out{\hypertarget{method-MassSpecEngineDB-info_analyses}{}}}
\subsection{Method \code{info_analyses()}}{
Get info summary
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MassSpecEngineDB$info_analyses()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MassSpecEngineDB-get_spectra_headers"></a>}}
\if{latex}{\out{\hypertarget{method-MassSpecEngineDB-get_spectra_headers}{}}}
\subsection{Method \code{get_spectra_headers()}}{
Get spectra headers for an analysis
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MassSpecEngineDB$get_spectra_headers(analyses = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{Character or numeric vector with names or indexes of analyses in the \code{Analyses}
object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MassSpecEngineDB-get_chromatograms_headers"></a>}}
\if{latex}{\out{\hypertarget{method-MassSpecEngineDB-get_chromatograms_headers}{}}}
\subsection{Method \code{get_chromatograms_headers()}}{
Get chromatograms headers for an analysis
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MassSpecEngineDB$get_chromatograms_headers(analyses = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{Character or numeric vector with names or indexes of analyses in the \code{Analyses}
object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MassSpecEngineDB-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MassSpecEngineDB-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MassSpecEngineDB$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
