% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecMethod_CalculateSpectraCharges.R
\name{MassSpecMethod_CalculateSpectraCharges_native}
\alias{MassSpecMethod_CalculateSpectraCharges_native}
\alias{validate_object.MassSpecMethod_CalculateSpectraCharges_native}
\title{MassSpecMethod_CalculateSpectraCharges_native Class}
\usage{
MassSpecMethod_CalculateSpectraCharges_native(
  onlyTopScans = FALSE,
  topScans = 1,
  roundVal = 35,
  relLowCut = 0.2,
  absLowCut = 300,
  top = 5
)

\method{validate_object}{MassSpecMethod_CalculateSpectraCharges_native}(x)
}
\arguments{
\item{onlyTopScans}{Logical (length 1) to consider only the \emph{n} \code{TopScans} with the highest
intensity.}

\item{topScans}{Numeric (length 1) with the number of top scans to be considered.}

\item{roundVal}{Numeric (length 1) with the rounding value for the m/z values before applying
charge clustering.}

\item{relLowCut}{Numeric (length 1) with the relative low cut for the charge clustering.}

\item{absLowCut}{Numeric (length 1) with the absolute low cut for the charge clustering.}

\item{top}{Numeric (length 1) with the number of top charges to be considered.}

\item{x}{A MassSpecMethod_CalculateSpectraCharges_native object.}
}
\value{
A MassSpecMethod_CalculateSpectraCharges_native object.
}
\description{
Calculates spectral charges from multi-charged compounds (e.g. proteins and
monoclonal antibodies) for mass deconvolution.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate_object(MassSpecMethod_CalculateSpectraCharges_native)}: Validate the MassSpecMethod_CalculateSpectraCharges_native object returns \code{NULL} if the object is valid.

}}
