% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/class_MassSpecMethod_CorrectChromatogramsBaseline.R
\name{MassSpecMethod_CorrectChromatogramsBaseline_airpls}
\alias{MassSpecMethod_CorrectChromatogramsBaseline_airpls}
\title{MassSpecMethod_CorrectChromatogramsBaseline_airpls S7 class}
\usage{
MassSpecMethod_CorrectChromatogramsBaseline_airpls(
  lambda = 10,
  differences = 1,
  itermax = 20
)
}
\arguments{
\item{lambda}{Numeric (length 1) with the smoothing intensity. the higher the \code{lambda} the
higher the smoothing.}

\item{differences}{Integer (length 1) indicating the order of the difference of penalties}

\item{itermax}{Integer (length 1) with the maximum number of iterations.}
}
\value{
A MassSpecMethod_CorrectChromatogramsBaseline_airpls object.
}
\description{
Performs baseline correction using adaptive iteratively reweighted Penalized Least
Squares (airPLS) based on the algorithm from Zhi-Min Zhang.
}
\references{
\insertRef{airpls01}{StreamFind}
}
