% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecMethod_FilterFeaturesMS2.R
\name{MassSpecMethod_FilterFeaturesMS2_native}
\alias{MassSpecMethod_FilterFeaturesMS2_native}
\title{MassSpecMethod_FilterFeaturesMS2_native Class}
\usage{
MassSpecMethod_FilterFeaturesMS2_native(
  top = NULL,
  minIntensity = NULL,
  relMinIntensity = NULL,
  blankClean = FALSE,
  mzClust = 0.005,
  blankPresenceThreshold = 0.8,
  globalPresenceThreshold = 0.1
)
}
\arguments{
\item{top}{An integer specifying the number of top features to be kept.}

\item{minIntensity}{A numeric value specifying the minimum intensity.}

\item{relMinIntensity}{A numeric value specifying the relative minimum intensity.}

\item{blankClean}{A logical value specifying if blank cleaning should be performed. When \code{TRUE}
MS2 traces present in the blank samples above the defined \code{blankPresenceThreshold} and above the
\code{globalPresenceThreshold} are excluded.}

\item{mzClust}{A numeric value specifying the m/z clustering threshold.}

\item{blankPresenceThreshold}{A numeric value specifying the blank presence threshold.}

\item{globalPresenceThreshold}{A numeric value specifying the global presence threshold.}
}
\value{
A \code{MassSpecMethod_FilterFeaturesMS2_native} object.
}
\description{
Settings for filtering (i.e., cleaning) non-relevant traces in features MS2
spectrum.
}
