% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecMethod_FindChromPeaks.R
\name{MassSpecMethod_FindChromPeaks_LocalMaxima}
\alias{MassSpecMethod_FindChromPeaks_LocalMaxima}
\title{MassSpecMethod_FindChromPeaks_LocalMaxima S7 class}
\usage{
MassSpecMethod_FindChromPeaks_LocalMaxima(
  minWidth = 0,
  maxWidth = 0,
  minHeight = 0
)
}
\arguments{
\item{minWidth}{Numeric (length 1) with the minimum width of a peak.}

\item{maxWidth}{Numeric (length 1) with the maximum width of a peak.}

\item{minHeight}{Numeric (length 1) with the minimum height of a peak.}
}
\value{
A MassSpecMethod_FindChromPeaks_LocalMaxima object.
}
\description{
Finds peak maxima in chromatograms.
}
