% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecMethod_FindChromPeaks.R
\name{MassSpecMethod_FindChromPeaks_pracma}
\alias{MassSpecMethod_FindChromPeaks_pracma}
\title{MassSpecMethod_FindChromPeaks_pracma S7 class}
\usage{
MassSpecMethod_FindChromPeaks_pracma(
  merge = TRUE,
  closeByThreshold = 45,
  minPeakHeight = 0,
  minPeakDistance = 10,
  minPeakWidth = 5,
  maxPeakWidth = 120,
  minSN = 10
)
}
\arguments{
\item{merge}{Logical (length 1) indicating if the nearby peaks should be merged.}

\item{closeByThreshold}{Numeric (length 1) with the maximum distance between peaks to be merged.}

\item{minPeakHeight}{Numeric (length 1) with the minimum peak height to be considered.}

\item{minPeakDistance}{Numeric (length 1) with the minimum distance between peaks.}

\item{minPeakWidth}{Numeric (length 1) with the minimum peak width.}

\item{maxPeakWidth}{Numeric (length 1) with the maximum peak width.}

\item{minSN}{Numeric (length 1) with the minimum signal-to-noise ratio.}
}
\value{
A MassSpecMethod_FindChromPeaks_pracma object.
}
\description{
Integrates chromatograms using the function \code{findpeaks} from the package
\pkg{pracma} with natively added peak exclusion and evaluation steps.
}
