% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecMethod_FindFeatures.R
\name{MassSpecMethod_FindFeatures_native}
\alias{MassSpecMethod_FindFeatures_native}
\title{MassSpecMethod_FindFeatures_native class}
\usage{
MassSpecMethod_FindFeatures_native(
  rtWindows = data.frame(rtmin = 300, rtmax = 3600),
  noiseBins = 70,
  noiseThreshold = 15,
  noiseQuantile = 0.01,
  minSNR = 3,
  mzrThreshold = 0.005,
  minTraces = 3,
  baselineWindow = 200,
  maxWidth = 100,
  minGaussFit = 0.6
)
}
\arguments{
\item{rtWindows}{data.frame with rtmin and rtmax columns for retention time windows for data inclusion.}

\item{noiseBins}{numeric(1) bin size to estimate the noise level in each scan (i.e., each rt value).}

\item{noiseThreshold}{numeric(1) lowest threshold to clean data (i.e., no trace with intensity below this level is kept).}

\item{noiseQuantile}{numeric(1) quantile threshold for dynamic estimation of noise.}

\item{minSNR}{numeric(1) minimum signal-to-noise ratio for considering a trace and chromatographic peak.}

\item{mzrThreshold}{numeric(1) mass threshold representing the mass resolution.}

\item{minTraces}{numeric(1) minimum number of traces to consider a mass cluster and chromatographic peak.}

\item{baselineWindow}{numeric(1) retention time window to build a baseline in a mass cluster.}

\item{maxWidth}{numeric(1) expected maximum window for a chromatographic peak.}

\item{minGaussFit}{numeric(1) minimum gaussian fit for a chromatographic peak.}
}
\description{
Native StreamFind method for finding features (i.e., chromatographic peaks) in liquid chromatography coupled to high resolution mass spectrometry files.
}
