% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/class_MassSpecMethod_FindTransformationProducts.R
\name{MassSpecMethod_FindTransformationProducts_native}
\alias{MassSpecMethod_FindTransformationProducts_native}
\title{MassSpecMethod_FindTransformationProducts_native Class}
\usage{
MassSpecMethod_FindTransformationProducts_native(ppm = 10)
}
\arguments{
\item{ppm}{Numeric of length one with the mass deviation, in ppm.}
}
\value{
A \code{MassSpecMethod_FindTransformationProducts_native} object.
}
\description{
Processing method for finding transformation products using a MassSpecResults_TransformationProducts object to perform suspect screening on the MassSpecResults_NonTargetAnalysis object. Returns an updated MassSpecResults_NonTargetAnalysis with the TPs added as suspects. Note that existing suspects are not replaced if they are parents.
}
