% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecMethod_IntegrateSpectra.R
\name{MassSpecMethod_IntegrateSpectra_StreamFind}
\alias{MassSpecMethod_IntegrateSpectra_StreamFind}
\title{MassSpecMethod_IntegrateSpectra_StreamFind S7 class}
\usage{
MassSpecMethod_IntegrateSpectra_StreamFind(
  merge = TRUE,
  closeByThreshold = 45,
  minPeakHeight = 0,
  minPeakDistance = 10,
  minPeakWidth = 5,
  maxPeakWidth = 120,
  minSN = 10
)
}
\arguments{
\item{merge}{Logical (length 1) indicating if the nearby peaks should be merged.}

\item{closeByThreshold}{Numeric (length 1) with the maximum distance between peaks to be merged.}

\item{minPeakHeight}{Numeric (length 1) with the minimum peak height to be considered.}

\item{minPeakDistance}{Numeric (length 1) with the minimum distance between peaks.}

\item{minPeakWidth}{Numeric (length 1) with the minimum peak width.}

\item{maxPeakWidth}{Numeric (length 1) with the maximum peak width.}

\item{minSN}{Numeric (length 1) with the minimum signal-to-noise ratio.}
}
\value{
A MassSpecMethod_IntegrateSpectra_StreamFind object.
}
\description{
Integrates Spectra using the function \code{findpeaks} from the package \pkg{pracma}
with natively added peak exclusion and evaluation steps.
}
