% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecMethod_LoadChromatograms.R
\name{MassSpecMethod_LoadChromatograms_native}
\alias{MassSpecMethod_LoadChromatograms_native}
\alias{validate_object.MassSpecMethod_LoadChromatograms_native}
\title{MassSpecMethod_LoadChromatograms_native Class}
\usage{
MassSpecMethod_LoadChromatograms_native(
  chromatograms = 1,
  rtmin = 0,
  rtmax = 0,
  minIntensity = 0
)

\method{validate_object}{MassSpecMethod_LoadChromatograms_native}(x)
}
\arguments{
\item{chromatograms}{A numeric or character vector with the chromatogram indices or names to be
used, respectively.}

\item{rtmin}{A numeric vector with the minimum retention time values to be used.}

\item{rtmax}{A numeric vector with the maximum retention time values to be used.}

\item{minIntensity}{A numeric value with the minimum intensity to be used.}

\item{x}{A MassSpecMethod_LoadChromatograms_native object.}
}
\value{
A \code{MassSpecMethod_LoadChromatograms_native} object.
}
\description{
Loads chromatograms from mass spectrometry analyses.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate_object(MassSpecMethod_LoadChromatograms_native)}: Validate the MassSpecMethod_LoadChromatograms_native object, returning NULL if valid.

}}
