% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecMethod_NormalizeSpectra.R
\name{MassSpecMethod_NormalizeSpectra_blockweight}
\alias{MassSpecMethod_NormalizeSpectra_blockweight}
\title{MassSpecMethod_NormalizeSpectra_blockweight S7 class}
\usage{
MassSpecMethod_NormalizeSpectra_blockweight()
}
\value{
A MassSpecMethod_NormalizeSpectra_blockweight object.
}
\description{
Normalizes spectra using block weighting for downstream data evaluation.
}
