% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecMethod_NormalizeSpectra.R
\name{MassSpecMethod_NormalizeSpectra_minmax}
\alias{MassSpecMethod_NormalizeSpectra_minmax}
\title{MassSpecMethod_NormalizeSpectra_minmax S7 class}
\usage{
MassSpecMethod_NormalizeSpectra_minmax()
}
\value{
A MassSpecMethod_NormalizeSpectra_minmax object.
}
\description{
Normalizes spectra using the min-max algorithm.
}
