% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecMethod_NormalizeSpectra.R
\name{MassSpecMethod_NormalizeSpectra_scale}
\alias{MassSpecMethod_NormalizeSpectra_scale}
\title{MassSpecMethod_NormalizeSpectra_scale Class}
\usage{
MassSpecMethod_NormalizeSpectra_scale()
}
\value{
A MassSpecMethod_NormalizeSpectra_scale object.
}
\description{
Normalizes spectra using scaling based on the standard deviation.
}
