% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecMethod_NormalizeSpectra.R
\name{MassSpecMethod_NormalizeSpectra_snv}
\alias{MassSpecMethod_NormalizeSpectra_snv}
\title{MassSpecMethod_NormalizeSpectra_snv S7 class}
\usage{
MassSpecMethod_NormalizeSpectra_snv(liftTozero = FALSE)
}
\arguments{
\item{liftTozero}{Logical (length 1) indicating if the spectra should be lifted to zero.}
}
\value{
A MassSpecMethod_NormalizeSpectra_snv object.
}
\description{
Normalizes spectra using the Standard Normal Variate (SNV) algorithm.
}
