% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecMethod_SmoothChromatograms.R
\name{MassSpecMethod_SmoothChromatograms_movingaverage}
\alias{MassSpecMethod_SmoothChromatograms_movingaverage}
\title{\strong{MassSpecMethod_SmoothChromatograms_movingaverage}}
\usage{
MassSpecMethod_SmoothChromatograms_movingaverage(windowSize = 5)
}
\arguments{
\item{windowSize}{Numeric (length 1) with the window size for the moving average.}
}
\value{
A MassSpecMethod_SmoothChromatograms_movingaverage object.
}
\description{
Smooths chromatograms using the moving average algorithm.
}
