% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecMethod_SmoothSpectra.R
\name{MassSpecMethod_SmoothSpectra_movingaverage}
\alias{MassSpecMethod_SmoothSpectra_movingaverage}
\title{MassSpecMethod_SmoothSpectra_movingaverage Class}
\usage{
MassSpecMethod_SmoothSpectra_movingaverage(windowSize = 5)
}
\arguments{
\item{windowSize}{Numeric (length 1) with the window size for the moving average.}
}
\value{
A MassSpecMethod_SmoothSpectra_movingaverage object.
}
\description{
Smooths spectra using the moving average algorithm.
}
