% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecMethod_SubtractBlankSpectra.R
\name{MassSpecMethod_SubtractBlankSpectra_StreamFind}
\alias{MassSpecMethod_SubtractBlankSpectra_StreamFind}
\title{MassSpecMethod_SubtractBlankSpectra_StreamFind S7 class}
\usage{
MassSpecMethod_SubtractBlankSpectra_StreamFind(negativeToZero = FALSE)
}
\arguments{
\item{negativeToZero}{Logical (length 1) indicating if negative values should be set to zero.}
}
\value{
A MassSpecMethod_SubtractBlankSpectra_StreamFind object.
}
\description{
Subtracts the blank spectra to each analysis according to the blank assignment.
}
