% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecMethod_SuspectScreening.R
\name{MassSpecMethod_SuspectScreening_patRoon}
\alias{MassSpecMethod_SuspectScreening_patRoon}
\title{\strong{MassSpecMethod_SuspectScreening_patRoon}}
\usage{
MassSpecMethod_SuspectScreening_patRoon(
  suspects = data.table::data.table(),
  rtWindow = 12,
  mzWindow = 0.005,
  filtered = FALSE
)
}
\arguments{
\item{suspects}{A data.frame with suspect information. See section Suspect list format in
\link[patRoon]{screenSuspects} for more information.}

\item{rtWindow}{The retention time window (in seconds) that will be used for matching a
suspect (+/- feature data).}

\item{mzWindow}{The m/z window that will be used for matching a suspect (+/- feature data).}

\item{filtered}{Logical (length 1). When \code{TRUE}, filtered features/groups are considered.}
}
\value{
A \code{MassSpecMethod_SuspectScreening_patRoon} object.
}
\description{
Settings for performing suspect screening using the function
\link[patRoon]{screenSuspects} from the patRoon R package.
}
\references{
\insertRef{patroon01}{StreamFind}

\insertRef{patroon02}{StreamFind}
}
