% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_S3_MassSpecSettings.R
\name{MassSpecSettings_AverageSpectra_StreamFind}
\alias{MassSpecSettings_AverageSpectra_StreamFind}
\title{MassSpecSettings_AverageSpectra_StreamFind}
\usage{
MassSpecSettings_AverageSpectra_StreamFind(collapseTime = FALSE)
}
\arguments{
\item{collapseTime}{Logical (length 1). When \code{TRUE} the spectra are averaged, reducing the time variable.}
}
\value{
A ProcessingSettings S3 class object with subclass MassSpecSettings_AverageSpectra_StreamFind.
}
\description{
Averages spectra based on analysis replicate groups.
}
