% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_S3_MassSpecSettings.R
\name{MassSpecSettings_CalculateSpectraCharges_StreamFind}
\alias{MassSpecSettings_CalculateSpectraCharges_StreamFind}
\title{MassSpecSettings_CalculateSpectraCharges_StreamFind}
\usage{
MassSpecSettings_CalculateSpectraCharges_StreamFind(
  roundVal = 35,
  relLowCut = 0.2,
  absLowCut = 300
)
}
\arguments{
\item{roundVal}{Numeric (length 1) with the rounding value for the m/z values before applying charge clustering.}

\item{relLowCut}{Numeric (length 1) with the relative low cut for the charge clustering.}

\item{absLowCut}{Numeric (length 1) with the absolute low cut for the charge clustering.}
}
\value{
A ProcessingSettings S3 class object with subclass MassSpecSettings_CalculateSpectraCharges_StreamFind.
}
\description{
Calculates spectral charges from multi-charged compounds (e.g. proteins and monoclonal antibodies) for
mass deconvolution.
}
