% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_lv3_S7_MassSpecSettings_ClusterSpectra.R
\name{MassSpecSettings_ClusterSpectra_StreamFind}
\alias{MassSpecSettings_ClusterSpectra_StreamFind}
\title{\strong{MassSpecSettings_ClusterSpectra_StreamFind}}
\usage{
MassSpecSettings_ClusterSpectra_StreamFind(
  val = "mz",
  clustVal = 0.001,
  presence = 0.1
)
}
\arguments{
\item{val}{Character (length 1) with the variable to be used for clustering.}

\item{clustVal}{Numeric (length 1) with the clustering value.}

\item{presence}{Numeric (length 1) with the minimum presence of traces in a cluster to be considered.}
}
\value{
A MassSpecSettings_ClusterSpectra_StreamFind object.
}
\description{
Clusters spectra based on a variable (i.e. column name).
}
