% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_S3_MassSpecSettings.R
\name{MassSpecSettings_DeconvoluteSpectra_StreamFind}
\alias{MassSpecSettings_DeconvoluteSpectra_StreamFind}
\title{MassSpecSettings_DeconvoluteSpectra_StreamFind}
\usage{
MassSpecSettings_DeconvoluteSpectra_StreamFind(clustVal = 0.1, window = 20)
}
\arguments{
\item{clustVal}{Numeric (length 1) with the clustering value for the charge deconvolution.}

\item{window}{Optional numeric (length 1) with the window in \emph{m/z} for collecting traces of a given charge.}
}
\value{
A ProcessingSettings S3 class object with subclass MassSpecSettings_DeconvoluteSpectra_StreamFind.
}
\description{
Deconvolutes the spectral mass-to-charge ratio (\emph{m/z}) to mass (Da) after assignment of charges.
}
