% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecSettings_FilterFeatures.R
\name{MassSpecSettings_FilterFeatures_StreamFind}
\alias{MassSpecSettings_FilterFeatures_StreamFind}
\title{\strong{MassSpecSettings_FilterFeatures_StreamFind}}
\usage{
MassSpecSettings_FilterFeatures_StreamFind(
  correctSuppression = TRUE,
  minSnRatio = NA_real_,
  excludeIsotopes = FALSE,
  excludeAdducts = FALSE,
  minIntensity = NA_real_,
  maxDeviationInReplicate = NA_real_,
  minAbundanceInReplicate = NA_real_,
  blankThreshold = NA_real_,
  blankSubtractionConservative = TRUE,
  onlyWithMS2 = FALSE
)
}
\arguments{
\item{correctSuppression}{Logical (length 1) with \code{TRUE} for correcting the intensity for matrix
suppression before applying intensity based filters. Note that suppression factor as obtained by
the processing method \code{CorrectMatrixSuppression} must be present. When not available the
correction is ignored even if set to \code{TRUE}.}

\item{minSnRatio}{Numeric (length 1) with the minimum signal-to-noise ratio.}

\item{excludeIsotopes}{Logical (length 1) with \code{TRUE} for filtering annotated isotopes
(only prevails the monoisotopic features).}

\item{excludeAdducts}{Logical (length 1) with \code{TRUE} for filtering annotated adducts.}

\item{minIntensity}{Numeric (length 1) with the minimum intensity threshold.}

\item{maxDeviationInReplicate}{Numeric (length 1) with the maximum standard deviation of the
intensity in each replicate group. The value must be a percentage from 0 to 100. The filter is
applied if the minimum standard deviation from all replicates is above the value.}

\item{minAbundanceInReplicate}{Numeric (length 1) with the minimum abundance in each replicate
group. The filter is applied if the maximum abundance from all replicates is below the value.}

\item{blankThreshold}{Numeric (length 1) with the intensity multiplier to set the blank
threshold. The filter is applied if the maximum mean intensity from all non-blank replicates
is below the correspondent blank intensity times the multiplier.}

\item{blankSubtractionConservative}{Logical (length 1) with \code{TRUE} for conservative blank
subtraction. The filter is applied if the maximum mean intensity from all non-blank replicates
is below the correspondent blank intensity times the multiplier. When set to \code{FALSE} the filter
is applied per replicate.}

\item{onlyWithMS2}{Logical (length 1) with \code{TRUE} for filtering features without MS2 spectra.}
}
\value{
A \code{MassSpecSettings_FilterFeatures_StreamFind} object.
}
\description{
Settings for filtering of features and feature groups.
}
