% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecSettings_FindInternalStandards.R
\name{MassSpecSettings_FindInternalStandards_StreamFind}
\alias{MassSpecSettings_FindInternalStandards_StreamFind}
\title{\strong{MassSpecSettings_FindInternalStandards_StreamFind}}
\usage{
MassSpecSettings_FindInternalStandards_StreamFind(
  database = data.table::data.table(name = character(), formula = character(), mass =
    numeric(), rt = numeric()),
  ppm = 5,
  sec = 10
)
}
\arguments{
\item{database}{A data.table with at least the columns name, mass, and rt indicating the name, neutral monoisotopic
mass and retention time of the internal standards, respectively.}

\item{ppm}{Numeric of length one with the mass deviation, in ppm.}

\item{sec}{Numeric of length one with the time deviation, in seconds.}
}
\value{
A \code{MassSpecSettings_FindInternalStandards_StreamFind} object.
}
\description{
Settings for finding internal standards using a data.frame.
}
