% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/class_lv3_S7_MassSpecSettings_NormalizeSpectra.R
\name{MassSpecSettings_NormalizeSpectra_blockweight}
\alias{MassSpecSettings_NormalizeSpectra_blockweight}
\title{\strong{MassSpecSettings_NormalizeSpectra_blockweight}}
\usage{
MassSpecSettings_NormalizeSpectra_blockweight()
}
\value{
A MassSpecSettings_NormalizeSpectra_blockweight object.
}
\description{
Normalizes spectra using block weighting for downstream data evaluation.
}
