% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MassSpecSettings_NormalizeSpectra.R
\name{MassSpecSettings_NormalizeSpectra_meancenter}
\alias{MassSpecSettings_NormalizeSpectra_meancenter}
\title{\strong{MassSpecSettings_NormalizeSpectra_meancenter}}
\usage{
MassSpecSettings_NormalizeSpectra_meancenter()
}
\value{
A MassSpecSettings_NormalizeSpectra_meancenter object.
}
\description{
Normalizes spectra using mean centering.
}
