% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/class_lv3_S7_MassSpecSettings_NormalizeSpectra.R
\name{MassSpecSettings_NormalizeSpectra_minmax}
\alias{MassSpecSettings_NormalizeSpectra_minmax}
\title{\strong{MassSpecSettings_NormalizeSpectra_minmax}}
\usage{
MassSpecSettings_NormalizeSpectra_minmax()
}
\value{
A MassSpecSettings_NormalizeSpectra_minmax object.
}
\description{
Normalizes spectra using the min-max algorithm.
}
