% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/class_lv3_S7_MassSpecSettings_NormalizeSpectra.R
\name{MassSpecSettings_NormalizeSpectra_scale}
\alias{MassSpecSettings_NormalizeSpectra_scale}
\title{\strong{MassSpecSettings_NormalizeSpectra_scale}}
\usage{
MassSpecSettings_NormalizeSpectra_scale()
}
\value{
A MassSpecSettings_NormalizeSpectra_scale object.
}
\description{
Normalizes spectra using scaling based on the standard deviation.
}
