% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/class_lv3_S7_MassSpecSettings_SmoothChromatograms.R
\name{MassSpecSettings_SmoothChromatograms_movingaverage}
\alias{MassSpecSettings_SmoothChromatograms_movingaverage}
\title{\strong{MassSpecSettings_SmoothChromatograms_movingaverage}}
\usage{
MassSpecSettings_SmoothChromatograms_movingaverage(windowSize = 5)
}
\arguments{
\item{windowSize}{Numeric (length 1) with the window size for the moving average.}
}
\value{
A MassSpecSettings_SmoothChromatograms_movingaverage object.
}
\description{
Smooths chromatograms using the moving average algorithm.
}
