% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_lv3_S7_MassSpecSettings_SmoothSpectra.R
\name{MassSpecSettings_SmoothSpectra_movingaverage}
\alias{MassSpecSettings_SmoothSpectra_movingaverage}
\title{\strong{MassSpecSettings_SmoothSpectra_movingaverage}}
\usage{
MassSpecSettings_SmoothSpectra_movingaverage(windowSize = 5)
}
\arguments{
\item{windowSize}{Numeric (length 1) with the window size for the moving average.}
}
\value{
A MassSpecSettings_SmoothSpectra_movingaverage object.
}
\description{
Smooths spectra using the moving average algorithm.
}
