% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Metadata.R
\name{Metadata}
\alias{Metadata}
\alias{validate_object.Metadata}
\alias{length.Metadata}
\alias{names.Metadata}
\alias{[.Metadata}
\alias{[<-.Metadata}
\alias{[[.Metadata}
\alias{[[<-.Metadata}
\alias{save.Metadata}
\alias{read.Metadata}
\alias{show.Metadata}
\title{Metadata S3 class}
\usage{
Metadata(entries = list())

\method{validate_object}{Metadata}(x)

\method{length}{Metadata}(x)

\method{names}{Metadata}(x)

\method{[}{Metadata}(x, i)

\method{[}{Metadata}(x, i) <- value

\method{[[}{Metadata}(x, i)

\method{[[}{Metadata}(x, i) <- value

\method{save}{Metadata}(x, file = "metadata.json")

\method{read}{Metadata}(x, file)

\method{show}{Metadata}(x)
}
\arguments{
\item{entries}{A named list of metadata entries as elements. Default is an empty list.
Elements must be of type character, numeric or POSIXt and must have length 1.}

\item{x}{A \code{Metadata} object.}

\item{i}{A numeric or character vector with the index or name of the entry to replace.}

\item{value}{The new value to replace the entry.}

\item{file}{A character string with the file name to read the metadata from.
The file format should be either "json" or "rds".}
}
\value{
A \code{Metadata} S3 class object.
}
\description{
The \code{Metadata} S3 class holds information, such as name, date, author
and file, as a named list with elements of length one. The \code{Metadata} is essentially
a list therefore list methods are also applicable.
}
\details{
If not given, elements name, author, date and file are set to \code{NA_character_},
\code{NA_character_}, current system time and \code{NA_character_}, respectively.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate_object(Metadata)}: Validates the \code{Metadata} object, returning \code{NULL} if valid.

\item \code{length(Metadata)}: Returns the length of the metadata entries.

\item \code{names(Metadata)}: Returns the names of the metadata entries.

\item \code{[}: Subsets the metadata entries.

\item \code{`[`(Metadata) <- value}: Replaces metadata entries.

\item \code{[[}: Extracts a single metadata entry.

\item \code{`[[`(Metadata) <- value}: Replaces a single metadata entry.

\item \code{save(Metadata)}: Saves the metadata to a file.

\item \code{read(Metadata)}: Reads metadata from a file, returning a \code{Metadata} object.

\item \code{show(Metadata)}: Prints the metadata entries.

}}
