% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Metadata.R
\name{Metadata}
\alias{Metadata}
\title{Metadata Class}
\usage{
Metadata(entries = list())
}
\arguments{
\item{entries}{Named list with metadata entries. Note that all given elements must be named and
of length one. If an element "name" is given, it must be type character. If an element "date"
is given, it must be class \code{POSIXct} or \code{POSIXt}. If given "date" is character, conversion to
class \code{POSIXct} or \code{POSIXt} is attempted.}
}
\value{
A Metadata S7 class object.
}
\description{
The Metadata S7 class holds information, such as name, date, author and file, as a
named list with elements of length one.
}
\section{Slots}{

\describe{
\item{\code{entries}}{\code{List} object with metadata entries as elements of length one.}
}}

