% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MethodsNonTargetAnalysis.R
\name{Method_NonTargetAnalysis_AnnotateComponents}
\alias{Method_NonTargetAnalysis_AnnotateComponents}
\title{Method_NonTargetAnalysis_AnnotateComponents}
\usage{
Method_NonTargetAnalysis_AnnotateComponents(
  maxIsotopes = 5L,
  maxCharge = 1L,
  maxGaps = 1L,
  ppm = 10,
  debugComponent = "",
  debugAnalysis = ""
)
}
\arguments{
\item{maxIsotopes}{Integer(1) maximum number of isotopes to consider.}

\item{maxCharge}{Integer(1) maximum charge state to consider.}

\item{maxGaps}{Integer(1) maximum number of gaps allowed in isotope
patterns.}

\item{ppm}{Numeric(1) minimum m/z tolerance in ppm.}

\item{debugComponent}{Character(1) component identifier to debug.}

\item{debugAnalysis}{Character(1) analysis name to debug.}
}
\value{
A \code{Method} object of class \code{Method_NonTargetAnalysis_AnnotateComponents}.
}
\description{
Create a \code{Method} child object for annotating isotopes,
adducts, and in-source fragments within components.
}
