% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MethodsNonTargetAnalysis.R
\name{Method_NonTargetAnalysis_AssignTransformationProducts}
\alias{Method_NonTargetAnalysis_AssignTransformationProducts}
\title{Method_NonTargetAnalysis_AssignTransformationProducts}
\usage{
Method_NonTargetAnalysis_AssignTransformationProducts(
  transformation_products = NULL,
  chromatographic_phase = c("reverse_phase", "hilic"),
  mzrMS2 = 0.008
)
}
\arguments{
\item{transformation_products}{A data.frame/data.table describing
transformation products and their precursors.}

\item{chromatographic_phase}{Character(1) chromatographic phase used for RT
plausibility checks.}

\item{mzrMS2}{Numeric(1) absolute m/z tolerance for MS2 fragment matching.}
}
\value{
A \code{Method} object of class
\code{Method_NonTargetAnalysis_AssignTransformationProducts}.
}
\description{
Create a \code{Method} child object for assigning transformation
products to suspect hits.
}
