% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MethodsNonTargetAnalysis.R
\name{Method_NonTargetAnalysis_GroupFeatures}
\alias{Method_NonTargetAnalysis_GroupFeatures}
\title{Method_NonTargetAnalysis_GroupFeatures}
\usage{
Method_NonTargetAnalysis_GroupFeatures(
  method = "internal_standards",
  rtDeviation = 5,
  ppm = 10,
  minSamples = 1L,
  binSize = 5,
  debug = FALSE,
  debugRT = 0
)
}
\arguments{
\item{method}{Character(1) alignment method, usually \code{"internal_standards"}
or \code{"obi_warp"}.}

\item{rtDeviation}{Numeric(1) retention-time tolerance in seconds.}

\item{ppm}{Numeric(1) mass tolerance in ppm.}

\item{minSamples}{Integer(1) minimum number of samples a feature must appear
in.}

\item{binSize}{Numeric(1) RT bin size in seconds.}

\item{debug}{Logical(1) whether to create a debug log.}

\item{debugRT}{Numeric(1) RT value to focus debugging on.}
}
\value{
A \code{Method} object of class \code{Method_NonTargetAnalysis_GroupFeatures}.
}
\description{
Create a \code{Method} child object for grouping aligned features
across analyses.
}
