% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MethodsNonTargetAnalysis.R
\name{Method_NonTargetAnalysis_LoadFeaturesMS2}
\alias{Method_NonTargetAnalysis_LoadFeaturesMS2}
\title{Method_NonTargetAnalysis_LoadFeaturesMS2}
\usage{
Method_NonTargetAnalysis_LoadFeaturesMS2(
  filtered = FALSE,
  minTracesIntensity = 10,
  isolationWindow = 1.3,
  mzClust = 0.005,
  presence = 0.8
)
}
\arguments{
\item{filtered}{Logical(1) whether to include features already marked as
filtered.}

\item{minTracesIntensity}{Numeric(1) minimum trace intensity to extract.}

\item{isolationWindow}{Numeric(1) isolation window around precursor m/z.}

\item{mzClust}{Numeric(1) m/z tolerance used when clustering traces.}

\item{presence}{Numeric(1) minimum fraction of scans required to keep a
cluster.}
}
\value{
A \code{Method} object of class \code{Method_NonTargetAnalysis_LoadFeaturesMS2}.
}
\description{
Create a \code{Method} child object for loading MS2 spectra into NTA
features.
}
