% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MethodsNonTargetAnalysis.R
\name{Method_NonTargetAnalysis_SuspectScreening}
\alias{Method_NonTargetAnalysis_SuspectScreening}
\title{Method_NonTargetAnalysis_SuspectScreening}
\usage{
Method_NonTargetAnalysis_SuspectScreening(
  suspects = NULL,
  analyses = character(),
  ppm = 5,
  sec = 10,
  ppmMS2 = 10,
  mzrMS2 = 0.008,
  minCosineSimilarity = 0.7,
  minSharedFragments = 3L,
  filtered = TRUE
)
}
\arguments{
\item{suspects}{A data.frame/data.table with suspect information.}

\item{analyses}{Optional character vector restricting screening to selected
analyses.}

\item{ppm, sec, ppmMS2, mzrMS2, minCosineSimilarity}{Numeric scalar thresholds
used during matching.}

\item{minSharedFragments}{Integer(1) minimum number of shared MS2 fragments.}

\item{filtered}{Logical(1) whether to include filtered features in the
search.}
}
\value{
A \code{Method} object of class \code{Method_NonTargetAnalysis_SuspectScreening}.
}
\description{
Create a \code{Method} child object for suspect screening against a
provided suspect table.
}
