% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_class_002_msAnalysis.R
\docType{class}
\name{msAnalysis-class}
\alias{msAnalysis-class}
\alias{initialize,msAnalysis-method}
\alias{show,msAnalysis-method}
\alias{files,msAnalysis-method}
\alias{files,msAnalysis,msAnalysis-method}
\alias{analysisInfo,msAnalysis-method}
\alias{analysisInfo,msAnalysis,msAnalysis-method}
\alias{analysisTable,msAnalysis-method}
\alias{analysisTable,msAnalysis,msAnalysis-method}
\alias{analyses,msAnalysis-method}
\alias{analyses,msAnalysis,msAnalysis-method}
\alias{replicates,msAnalysis-method}
\alias{replicates,msAnalysis,msAnalysis-method}
\alias{replicates<-,msAnalysis-method}
\alias{replicates<-,msAnalysis,msAnalysis-method}
\alias{blanks,msAnalysis-method}
\alias{blanks,msAnalysis,msAnalysis-method}
\alias{blanks<-,msAnalysis-method}
\alias{blanks<-,msAnalysis,msAnalysis-method}
\alias{metadata,msAnalysis-method}
\alias{metadata,msAnalysis,msAnalysis-method}
\alias{polarities,msAnalysis-method}
\alias{polarities,msAnalysis,msAnalysis-method}
\alias{EICs,msAnalysis-method}
\alias{EICs,msAnalysis,msAnalysis-method}
\alias{plotEICs,msAnalysis-method}
\alias{plotEICs,msAnalysis,msAnalysis-method}
\alias{TICs,msAnalysis-method}
\alias{TICs,msAnalysis,msAnalysis-method}
\alias{plotTICs,msAnalysis-method}
\alias{plotTICs,msAnalysis,msAnalysis-method}
\alias{XICs,msAnalysis-method}
\alias{XICs,msAnalysis,msAnalysis-method}
\alias{plotXICs,msAnalysis-method}
\alias{plotXICs,msAnalysis,msAnalysis-method}
\alias{MS2s,msAnalysis-method}
\alias{MS2s,msAnalysis,msAnalysis-method}
\alias{plotMS2s,msAnalysis-method}
\alias{plotMS2s,msAnalysis,msAnalysis-method}
\alias{loadRawData,msAnalysis-method}
\alias{loadRawData,msAnalysis,msAnalysis-method}
\alias{spectra,msAnalysis-method}
\alias{spectra,msAnalysis,msAnalysis-method}
\alias{hasAdjustedRetentionTime,msAnalysis-method}
\alias{hasAdjustedRetentionTime,msAnalysis,msAnalysis-method}
\alias{addParameters,msAnalysis-method}
\alias{addParameters,msAnalysis,msAnalysis-method}
\alias{getParameters,msAnalysis-method}
\alias{getParameters,msAnalysis,msAnalysis-method}
\alias{peaks,msAnalysis-method}
\alias{peaks,msAnalysis,msAnalysis-method}
\alias{plotPeaks,msAnalysis-method}
\alias{plotPeaks,msAnalysis,msAnalysis-method}
\alias{mapPeaks,msAnalysis-method}
\alias{mapPeaks,msAnalysis,msAnalysis-method}
\title{msAnalysis-class}
\usage{
## S4 method to initialize an "msAnalysis" object \cr
\code{new("msAnalysis", file)}

\S4method{show}{msAnalysis}(object)

\S4method{files}{msAnalysis}(object)

\S4method{analysisInfo}{msAnalysis}(obj)

\S4method{analysisTable}{msAnalysis}(object)

\S4method{analyses}{msAnalysis}(obj)

\S4method{replicates}{msAnalysis}(object)

\S4method{replicates}{msAnalysis}(object) <- value

\S4method{blanks}{msAnalysis}(object)

\S4method{blanks}{msAnalysis}(object) <- value

\S4method{metadata}{msAnalysis}(object, which = NULL)

\S4method{polarities}{msAnalysis}(object)

\S4method{EICs}{msAnalysis}(object, mz = NULL, ppm = 20, rt = NULL, sec = 60, id = NULL)

\S4method{plotEICs}{msAnalysis}(
  object,
  mz = NULL,
  ppm = 20,
  rt = NULL,
  sec = 30,
  id = NULL,
  legendNames = NULL,
  title = NULL,
  interactive = FALSE
)

\S4method{TICs}{msAnalysis}(object)

\S4method{plotTICs}{msAnalysis}(object, title = NULL, interactive = FALSE)

\S4method{XICs}{msAnalysis}(object, mz = NULL, ppm = 20, rt = NULL, sec = 60, id = NULL)

\S4method{plotXICs}{msAnalysis}(
  object,
  mz = NULL,
  ppm = 20,
  rt = NULL,
  sec = 60,
  id = NULL,
  legendNames = NULL,
  plotTargetMark = TRUE,
  targetsMark = NULL,
  ppmMark = 5,
  secMark = 10,
  numberRows = 1
)

\S4method{MS2s}{msAnalysis}(
  object = NULL,
  mz = NULL,
  ppm = 20,
  rt = NULL,
  sec = 60,
  id = NULL,
  settings = NULL
)

\S4method{plotMS2s}{msAnalysis}(
  object = NULL,
  mz = NULL,
  ppm = 20,
  rt = NULL,
  sec = 60,
  id = NULL,
  settings = NULL,
  legendNames = NULL,
  title = NULL,
  colorBy = "targets",
  interactive = FALSE
)

\S4method{loadRawData}{msAnalysis}(object)

\S4method{spectra}{msAnalysis}(object)

\S4method{hasAdjustedRetentionTime}{msAnalysis}(object)

\S4method{addParameters}{msAnalysis}(object, settings)

\S4method{getParameters}{msAnalysis}(object, call = NULL)

\S4method{peaks}{msAnalysis}(object, targetsID = NULL, mz = NULL, ppm = 20, rt = NULL, sec = 60)

\S4method{plotPeaks}{msAnalysis}(
  object,
  targetsID = NULL,
  mz = NULL,
  ppm = 20,
  rt = NULL,
  sec = 30,
  legendNames = NULL,
  title = NULL,
  interactive = FALSE
)

\S4method{mapPeaks}{msAnalysis}(
  object,
  targetsID = NULL,
  mz = NULL,
  ppm = 20,
  rt = NULL,
  sec = 30,
  legendNames = NULL,
  xlim = 30,
  ylim = 0.05,
  title = NULL
)
}
\arguments{
\item{object}{An \linkS4class{msAnalysis} object.}

\item{value}{A character string applicable to the respective method.}

\item{which}{A character vector with the entry name/s.}

\item{mz}{A vector with target \emph{m/z} values or
a two columns data.table or data.frame with minimum and maximum \emph{m/z} values.
Alternatively, \emph{m/z} and retention time values can be given as one data.table/data.frame
and the deviations given as \code{ppm} and \code{sec} are used to calculate the ranges.
The same also works for min and max values of \emph{m/z} and retention time targets.
Note that when mass/time ranges are given, \code{ppm} and \code{sec} are not used.}

\item{ppm}{A numeric vector of length one with the mass deviation, in ppm.}

\item{rt}{A vector with target retention time values or
a two columns data.table/data.frame with minimum and maximum retention time values.}

\item{sec}{A numeric vector of length one with the time deviation, in seconds.}

\item{id}{An id vector with target identifiers. When not given
is added as a combination of the \emph{m/z} and retention time ranges or values.}

\item{legendNames}{A character vector with the same length and order
as the number and order of targets to be used as plot legend.}

\item{title}{a character string to the define a title.}

\item{interactive}{Logical value, set to \code{TRUE} to use
the \pkg{plotly} instead of \pkg{base}. The default is \code{FALSE}.}

\item{plotTargetMark}{Logical, set to \code{TRUE} to plot a target mark.}

\item{targetsMark}{A two columns \link[data.table]{data.table} or \link{data.frame} with
\emph{m/z} and time targets. The column must be named with "mz" and "rt" for
\emph{m/z} and time values, respectively.}

\item{ppmMark}{A numeric vector of length one to define the mass deviation, in ppm,
of the target mark.}

\item{secMark}{A numeric vector of length one to define the time deviation, in seconds,
of the target mark.}

\item{numberRows}{A numeric vector of length one to define
the number of rows to grid the plots.}

\item{call}{The call name of the settings to retrieve.}

\item{targetsID}{A character vector with the ID of targets (e.g., peaks or features IDs).}

\item{xlim}{A length one or two numeric vector for setting the \emph{x} limits of a plot.}

\item{ylim}{A length one or two numeric vector for setting the \emph{y} limits of a plot.}

\item{file}{A character string with the complete file path.}

\item{replicate}{optional character string for analysis replicate name}

\item{blank}{optional character string for associated blank replicate name}

\item{settigns}{A \linkS4class{settings} object.}
}
\description{
An S4 class representing an MS sample/file within the \pkg{streamFind} package.
The \code{msAnalysis} is used to store and manage MS data and the respective methods can be used
for inspection, processing and evaluation.
}
\section{Methods (by generic)}{
\itemize{
\item \code{initialize}: initializes an \linkS4class{msAnalysis} object.

\item \code{show}: prints the details of an \linkS4class{msAnalysis} object.

\item \code{files}: getter for analysis file path.

\item \code{analysisInfo}: getter for analysis info as \link{data.frame} with
four columns: path, analysis, group and blank. The \link{data.frame}
can be used as analysisInfo in \pkg{patRoon}.

\item \code{analysisTable}: getter for analysis table as \link{data.table} with
four columns: file, analysis, replicate and blank.

\item \code{analyses}: getter for analysis name.

\item \code{replicates}: getter for the replicate name.

\item \code{replicates<-}: setter for analysis replicate name.
The \code{value} is a character string with analysis replicate name.

\item \code{blanks}: getter for associated blank replicate name.

\item \code{blanks<-}: setter for associated blank replicate.
The \code{value} is a character string with associated blank replicate name.

\item \code{metadata}: getter for analysis metadata.
Returns a \link[data.table]{date.table} with a column per metadata entry.

\item \code{polarities}: getter for analyses polarity.

\item \code{EICs}: get extracted ion chromatograms (EICs)
for specified \emph{m/z} (Da) and retention time (seconds) targets.
The arguments \code{mz}, \code{ppm}, \code{rt}
and \code{sec} are used to construct the targets.
See ?\link{makeTargets} for more information.

\item \code{plotEICs}: A method for plotting extracted ion chromatograms (EICs)
of data in the analysis file.
The arguments for data collection are the same as the \link{EICs} method.
A \link[data.table]{date.table} can be used instead.
The \code{legendNames} is a character vector with the same length as targets for plotting and
can be used to legend the plot. Note, the plot legends the data by target.

\item \code{TICs}: extracts the total ion chromatograms (TICs)
of analysis.

\item \code{plotTICs}: plots the total ion chromatogram (TIC) in the analysis.

\item \code{XICs}: get three dimensional (\emph{m/z}, time and intensity)
extracted ion chromatograms (XICs) for specified \emph{m/z} and retention time pair targets
in analysis. The arguments \code{mz}, \code{ppm}, \code{rt}, \code{sec} and \code{id}
are used to construct the targets. See ?\link{makeTargets} for more information.

\item \code{plotXICs}: plots three dimensional (\emph{m/z}, time and intensity)
extracted ion chromatograms (XICs) for specified \emph{m/z} and retention time pair targets
in analyses of an \linkS4class{msAnalysis} object. The arguments \code{mz}, \code{ppm}, \code{rt},
\code{sec} and \code{id} are used to construct the targets. See ?\link{makeTargets} for more information.
When \code{plotTargetMark} is \code{TRUE} a target is plotted representing the deviations as defined
by the arguments \code{ppmMark} and \code{secMark} in ppm and seconds, respectively.
When ranges were given to build the XIC, exact \emph{m/z} and time targets can be specified with
the argument \code{targetsMark}. \code{targetsMark} should be a two column table named mz and rt with
exact \emph{m/z} and time targets. Note that the number of rows should be the same as the number of target
in the XIC. The number of rows to plot multiple targets can be defined by the \code{numberRows} argument.

\item \code{MS2s}: get MS2 data for specified \emph{m/z} and retention time (seconds) targets.
The \code{clusteringUnit} defines the method used for clustering.
Possible values are \emph{euclidean} (the default) or \emph{distance}.
The mass (in Da) and time (in seconds) isolation windows to screen for the respective precursors
are defined with the arguments \code{isolationMassWindow} and \code{isolationTimeWindow}, respectively.
The \code{clusteringUnit} and \code{clusteringWindow} define
the mass deviation unit and deviation to cluster mass traces from different spectra, respectively.
For the \code{clusteringUnit}, possible values are \emph{mz} (the default) or \emph{ppm}.
The \code{minIntensityPre} and \code{minIntensityPost}
define the minimum intensity for mass traces before and after clustering, respectively.
Set \code{mergeVoltages} to \code{TRUE} for merging spectra acquired with different collision energies.
The \code{mergeBy} argument is used to merge spectra by "samples" or "replicates".
When \code{NULL}, MS2 is given per target and per sample.

\item \code{plotMS2s}: plots MS2 data for specified \emph{m/z} and retention time (seconds) targets
in analyses of an \linkS4class{msData} object. The \code{clusteringUnit} defines the method used for clustering.
Possible values are \emph{euclidean} (the default) or \emph{distance}.
The mass (in Da) and time (in seconds) isolation windows to screen for the respective precursors
are defined with the arguments \code{isolationMassWindow} and \code{isolationTimeWindow}, respectively.
The \code{clusteringUnit} and \code{clusteringWindow} define
the mass deviation unit and deviation to cluster mass traces from different spectra, respectively.
For the \code{clusteringUnit}, possible values are \emph{mz} (the default) or \emph{ppm}.
The \code{minIntensityPre} and \code{minIntensityPost}
define the minimum intensity for mass traces before and after clustering, respectively.
Set \code{mergeVoltages} to \code{TRUE} for merging spectra acquired with different collision energies.
The \code{mergeBy} argument is used to merge spectra by "samples" or "replicates".
When \code{NULL}, MS2 is given per target and per sample. The possible values for the
\code{colorBy} argument are "targets" and "voltages" to color by
each target, sample, replicate or collision energy, respectively.

\item \code{loadRawData}: adds raw spectra and chromatograms to the respective
slots of an \linkS4class{msAnalysis} object.

\item \code{spectra}: getter for slot spectra in the \linkS4class{msAnalysis}.

\item \code{hasAdjustedRetentionTime}: getter for presence of adjusted retention time
in the \linkS4class{msAnalysis}.

\item \code{addParameters}: adds processing parameters to the analysis.

\item \code{getParameters}: gets processing parameters in the analysis.

\item \code{peaks}: getter for chromatographic peaks.
The arguments \code{targetID} and \code{mz}/\code{rt} can be used
to select specific peaks. The \emph{id} of peaks and/or features can be
given in the \code{targetsID} argument to select the respective peaks.

\item \code{plotPeaks}: a method for plotting chromatographic peaks
in an \linkS4class{msAnalysis} object.
The arguments \code{targetID} and \code{mz}/\code{rt} can be used
to select specific peaks. The \emph{id} of peaks and/or features can be
given in the \code{targetsID} argument to select the respective peaks.
The \code{legendNames} is a character vector with the same length as targets for plotting and
can be used to legend the plot.

\item \code{mapPeaks}: a method for mapping peaks mass and time space.
The \code{legendNames} is a character vector with the same length as targets for plotting and
can be used to legend the plot.
}}

\section{Slots}{

\describe{
\item{\code{analysis}}{A character string with the MS file name the without extension.}

\item{\code{file}}{A character string with the complete path of the MS file.}

\item{\code{replicate}}{A character string with the respective sample replicate group name.}

\item{\code{metadata}}{A list with other information queried from the MS files.
The info corresponds to the merged output of \link[mzR]{runInfo} and \link[mzR]{instrumentInfo}
from the \pkg{mzR} package.}

\item{\code{spectra}}{A \link[data.table]{data.table} obtained by
the \link[mzR]{peaks} function of the \pkg{mzR} package.
Although the function \link[mzR]{peaks} returns a \code{list},
the output is collapsed to a \link[data.table]{data.table}.}

\item{\code{chromatograms}}{A \link[data.table]{data.table} obtained by
the \link[mzR]{chromatogram} function of the \pkg{mzR} package.
Although the function \link[mzR]{chromatogram} returns a \code{list},
the output is collapsed to a \link[data.table]{data.table}.}

\item{\code{parameters}}{A list with parameter \linkS4class{msSettings} for data processing, such as peak picking.}

\item{\code{peaks}}{A \link[data.table]{data.table} with feature details, containing the following columns:
\enumerate{
 \item \strong{id}: character, the identifier of the peak;
 \item \strong{rt}: numeric, the calculated retention time (in seconds);
 \item \strong{mz}: numeric, the calculated \emph{m/z} value;
 \item \strong{intensity}: numeric, the calculated intensity or height (in counts);
 \item \strong{area}: numeric, the calculated area of the peak;
 \item \strong{rtmin}: numeric, the minimum retention time (in seconds);
 \item \strong{rtmax}: numeric, the maximum retention time, (in seconds);
 \item \strong{mzmin}: numeric, the minimum \emph{m/z} value;
 \item \strong{mzmax}: numeric, the maximum \emph{m/z} value;
 \item \strong{adduct}: character, the annotated adduct ion representation of the peak;
 \item \strong{isotope}: numeric, the annotated isotopic number;
 \item \strong{monoIsotope}: character, the identifier of the mono isotopic ion;
 \item \strong{scans}: numeric vector, the scans used for the feature integration;
 \item \strong{feature}: character, the identifier of the peaks group (i.e., feature) after
 alignment and grouping of corresponding peaks across samples;
 \item \strong{...} other columns added from various functions.
}}
}}

\references{
\insertRef{patroon01}{streamFind}

\insertRef{proteo01}{streamFind}

\insertRef{proteo02}{streamFind}

\insertRef{mzr01}{streamFind}

\insertRef{mzr02}{streamFind}

\insertRef{mzr03}{streamFind}

\insertRef{mzr04}{streamFind}
}
