% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_class_004_msData.R
\docType{class}
\name{msData-class}
\alias{msData-class}
\alias{show,msData-method}
\alias{analysisInfo,msData-method}
\alias{analysisInfo,msData,msData-method}
\alias{analysisTable,msData-method}
\alias{analysisTable,msData,msData-method}
\alias{files,msData-method}
\alias{files,msData,msData-method}
\alias{analyses,msData-method}
\alias{analyses,msData,msData-method}
\alias{replicates,msData-method}
\alias{replicates,msData,msData-method}
\alias{replicates<-,msData-method}
\alias{replicates<-,msData,msData-method}
\alias{blanks,msData-method}
\alias{blanks,msData,msData-method}
\alias{blanks<-,msData-method}
\alias{blanks<-,msData,msData-method}
\alias{metadata,msData-method}
\alias{metadata,msData,msData-method}
\alias{addMetadata,msData-method}
\alias{addMetadata,msData,msData-method}
\alias{addAnalyses,msData-method}
\alias{addAnalyses,msData,msData-method}
\alias{getAnalyses,msData-method}
\alias{getAnalyses,msData,msData-method}
\alias{[,msData,ANY,missing,missing-method}
\alias{polarities,msData-method}
\alias{polarities,msData,msData-method}
\alias{EICs,msData-method}
\alias{EICs,msData,msData-method}
\alias{plotEICs,msData-method}
\alias{plotEICs,msData,msData-method}
\alias{TICs,msData-method}
\alias{TICs,msData,msData-method}
\alias{plotTICs,msData-method}
\alias{plotTICs,msData,msData-method}
\alias{XICs,msData-method}
\alias{XICs,msData,msData-method}
\alias{plotXICs,msData-method}
\alias{plotXICs,msData,msData-method}
\alias{MS2s,msData-method}
\alias{MS2s,msData,msData-method}
\alias{plotMS2s,msData-method}
\alias{plotMS2s,msData,msData-method}
\alias{loadRawData,msData-method}
\alias{loadRawData,msData,msData-method}
\alias{spectra,msData-method}
\alias{spectra,msData,msData-method}
\alias{hasAdjustedRetentionTime,msData-method}
\alias{hasAdjustedRetentionTime,msData,msData-method}
\alias{addParameters,msData-method}
\alias{addParameters,msData,msData-method}
\alias{getParameters,msData-method}
\alias{as.features,msData-method}
\alias{as.features,msData,msData-method}
\alias{peaks,msData-method}
\alias{peaks,msData,msData-method}
\alias{plotPeaks,msData-method}
\alias{plotPeaks,msData,msData-method}
\alias{mapPeaks,msData-method}
\alias{mapPeaks,msData,msData-method}
\alias{features,msData-method}
\alias{features,msData,msData-method}
\alias{plotFeatures,msData-method}
\alias{plotFeatures,msData,msData-method}
\alias{as.featureGroups,msData-method}
\alias{as.featureGroups,msData,msData-method}
\alias{annotation,msData-method}
\alias{annotation,msData,msData-method}
\alias{plotAnnotation,msData-method}
\alias{plotAnnotation,msData,msData-method}
\alias{[,msData,ANY,ANY,missing-method}
\alias{[,msData,ANY,ANY,ANY-method}
\title{msData}
\usage{
\S4method{show}{msData}(object)

\S4method{analysisInfo}{msData}(obj)

\S4method{analysisTable}{msData}(object)

\S4method{files}{msData}(object)

\S4method{analyses}{msData}(obj)

\S4method{replicates}{msData}(object)

\S4method{replicates}{msData}(object) <- value

\S4method{blanks}{msData}(object)

\S4method{blanks}{msData}(object) <- value

\S4method{metadata}{msData}(x, analyses = NULL, which = NULL)

\S4method{addMetadata}{msData}(object, metadata = NULL, overwrite = FALSE)

\S4method{addAnalyses}{msData}(object, analyses = NULL)

\S4method{getAnalyses}{msData}(object, analyses = NULL)

\S4method{[}{msData,ANY,missing,missing}(x, i, j, ..., drop = TRUE)

\S4method{polarities}{msData}(object)

\S4method{EICs}{msData}(
  object,
  analyses = NULL,
  mz = NULL,
  ppm = 20,
  rt = NULL,
  sec = 60,
  id = NULL
)

\S4method{plotEICs}{msData}(
  object,
  analyses = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 30,
  id = NULL,
  colorBy = "targets",
  legendNames = NULL,
  title = NULL,
  interactive = FALSE
)

\S4method{TICs}{msData}(object, analyses = NULL)

\S4method{plotTICs}{msData}(
  object,
  analyses = NULL,
  colorBy = "analyses",
  title = NULL,
  interactive = FALSE
)

\S4method{XICs}{msData}(
  object,
  analyses = NULL,
  mz = NULL,
  ppm = 20,
  rt = NULL,
  sec = 60,
  id = NULL
)

\S4method{plotXICs}{msData}(
  object,
  analyses = NULL,
  mz = NULL,
  ppm = 20,
  rt = NULL,
  sec = 60,
  id = NULL,
  legendNames = NULL,
  plotTargetMark = TRUE,
  targetsMark = NULL,
  ppmMark = 5,
  secMark = 10,
  numberRows = 1
)

\S4method{MS2s}{msData}(
  object = NULL,
  analyses = NULL,
  mz = NULL,
  ppm = 20,
  rt = NULL,
  sec = 60,
  id = NULL,
  settings = NULL
)

\S4method{plotMS2s}{msData}(
  object = NULL,
  analyses = NULL,
  mz = NULL,
  ppm = 20,
  rt = NULL,
  sec = 60,
  id = NULL,
  settings = NULL,
  legendNames = NULL,
  title = NULL,
  colorBy = "targets",
  interactive = FALSE
)

\S4method{loadRawData}{msData}(object, analyses = NULL)

\S4method{spectra}{msData}(object)

\S4method{hasAdjustedRetentionTime}{msData}(object)

\S4method{addParameters}{msData}(object, settings, where = "analyses", analyses = NULL)

\S4method{getParameters}{msData}(object, where = "analyses", analyses = NULL, call = NULL)

\S4method{as.features}{msData}(object)

\S4method{peaks}{msData}(
  object,
  analyses = NULL,
  targetsID = NULL,
  mz = NULL,
  ppm = 20,
  rt = NULL,
  sec = 60,
  filtered = TRUE
)

\S4method{plotPeaks}{msData}(
  object,
  analyses = NULL,
  targetsID = NULL,
  mz = NULL,
  ppm = 20,
  rt = NULL,
  sec = 30,
  filtered = TRUE,
  colorBy = "targets",
  legendNames = NULL,
  title = NULL,
  interactive = FALSE
)

\S4method{mapPeaks}{msData}(
  object,
  analyses = NULL,
  targetsID = NULL,
  mz = NULL,
  ppm = 20,
  rt = NULL,
  sec = 30,
  filtered = TRUE,
  colorBy = "targets",
  legendNames = NULL,
  xlim = 30,
  ylim = 0.05,
  title = NULL
)

\S4method{features}{msData}(
  object,
  targetsID = NULL,
  mz = NULL,
  ppm = 20,
  rt = NULL,
  sec = 60,
  filtered = TRUE,
  complete = FALSE,
  average = TRUE
)

\S4method{plotFeatures}{msData}(
  object,
  analyses = NULL,
  targetsID = NULL,
  mz = NULL,
  ppm = 20,
  rt = NULL,
  sec = 30,
  filtered = TRUE,
  colorBy = "targets",
  legendNames = NULL,
  title = NULL,
  interactive = FALSE
)

\S4method{as.featureGroups}{msData}(object)

\S4method{annotation}{msData}(
  object,
  targetsID = NULL,
  mz = NULL,
  ppm = 20,
  rt = NULL,
  sec = 60,
  id = NULL,
  all = FALSE
)

\S4method{plotAnnotation}{msData}(
  object,
  targetsID = NULL,
  mz = NULL,
  ppm = 20,
  rt = NULL,
  sec = 30,
  id = NULL,
  all = FALSE,
  colorBy = "isotopes"
)

\S4method{[}{msData,ANY,ANY,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{msData,ANY,ANY,ANY}(x, i, j, p)
}
\arguments{
\item{object}{An \linkS4class{msData} object.}

\item{value}{A character vector applicable to the respective method.}

\item{analyses}{A numeric or character vector with the index or analysis name/s.
When \code{NULL}, all the analyses are used.}

\item{which}{A character vector with the entry name/s.}

\item{metadata}{A list with a named vector of metadata for each analyses in the
\linkS4class{msData} object or a \code{data.frame} or \code{data.table}
with metadata added as columns and with the number of row equal to
the number of analyses in the \linkS4class{msData} object.}

\item{i}{A character or numeric vector with the indice/s or name/s of
the analysis to keep while sub-setting.}

\item{j}{The indice/s or \emph{id}/s for of analyses, features or/and peaks to keep.}

\item{mz}{A vector with target \emph{m/z} values or
a two columns data.table or data.frame with minimum and maximum \emph{m/z} values.
Alternatively, \emph{m/z} and retention time values can be given as one data.table/data.frame
and the deviations given as \code{ppm} and \code{sec} are used to calculate the ranges.
The same also works for min and max values of \emph{m/z} and retention time targets.
Note that when mass/time ranges are given, \code{ppm} and \code{sec} are not used.}

\item{ppm}{A numeric vector of length one with the mass deviation, in ppm.}

\item{rt}{A vector with target retention time values or
a two columns data.table/data.frame with minimum and maximum retention time values.}

\item{sec}{A numeric vector of length one with the time deviation, in seconds.}

\item{id}{An id vector with target identifiers. When not given
is added as a combination of the \emph{m/z} and retention time ranges or values.}

\item{colorBy}{A length one character vector applicable to the respective method.}

\item{legendNames}{A character vector with the same length and order
as the number and order of targets to be used as plot legend.}

\item{title}{A character string with the plot tile.}

\item{interactive}{Logical value, set to \code{TRUE} to use
the \pkg{plotly} instead of \pkg{base}. The default is \code{FALSE}.}

\item{plotTargetMark}{Logical, set to \code{TRUE} to plot a target mark.}

\item{targetsMark}{A two columns \link[data.table]{data.table} or \link{data.frame} with
\emph{m/z} and time targets. The column must be named with "mz" and "rt" for
\emph{m/z} and time values, respectively.}

\item{ppmMark}{A numeric vector of length one to define the mass deviation, in ppm,
of the target mark.}

\item{secMark}{A numeric vector of length one to define the time deviation, in seconds,
of the target mark.}

\item{numberRows}{A numeric vector of length one to define
the number of rows to grid the plots.}

\item{settings}{A \linkS4class{settings} S4 class object with call for \code{extractMSn}.}

\item{where}{A character vector defining where to get the \linkS4class{settings}.}

\item{call}{The call name of the settings to retrieve.}

\item{targetsID}{A character vector with the ID of targets (e.g., peaks or features IDs).}

\item{filtered}{Logical. When \code{TRUE} (the default), filtered are returned.
When \code{FALSE}, filtered data are excluded.}

\item{xlim}{A length one or two numeric vector for setting the \emph{x} limits of a plot.}

\item{ylim}{A length one or two numeric vector for setting the \emph{y} limits of a plot.}

\item{complete}{Logical, set to \code{TRUE} for a complete version of the output.}

\item{all}{Logical, set to \code{TRUE} for displaying all features/peaks.}

\item{settigns}{A \linkS4class{settings} object.}
}
\description{
An S4 class object to store and manage processing of files with MS data.
The \code{msData} object inherits the \linkS4class{streamSet} structure with type defined as
\emph{ms}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show(msData)}: prints the details of an \linkS4class{msData} object.

\item \code{analysisInfo(msData)}: getter for analysis info as \link{data.frame} with
four columns: path, analysis, group and blank. The \link{data.frame}
can be used as analysisInfo in \pkg{patRoon}.

\item \code{analysisTable(msData)}: getter for analysis table as \link{data.table} with
four columns: file, analysis, replicate and blank.

\item \code{files(msData)}: getter for analysis file paths.

\item \code{analyses(msData)}: getter for analysis names.

\item \code{replicates(msData)}: getter for replicate names.

\item \code{replicates(msData) <- value}: setter for analysis replicate names.
The \code{value} is a character vector with the same length as
the number of analyses in the \code{object},
containing analysis replicate name for each analysis.

\item \code{blanks(msData)}: getter for blank names.

\item \code{blanks(msData) <- value}: setter for associated blank replicate for each analyses.
The \code{value} is a character vector with the same length as
the number of analyses in the \code{object},
containing the associated blank replicate name of each analysis.

\item \code{metadata(msData)}: getter for analyses metadata.
Returns a \link[data.table]{date.table} with a row per analysis.

\item \code{addMetadata(msData)}: setter for analyses metadata.

\item \code{addAnalyses(msData)}: getter for analysis names.

\item \code{getAnalyses(msData)}: getter for analysis names.

\item \code{x[i}: subset on analyses, using index or analyses name.

\item \code{polarities(msData)}: getter for analyses polarity.

\item \code{EICs(msData)}: get extracted ion chromatograms (EICs)
for specified \emph{m/z} (Da) and retention time (seconds) targets
in given analyses. The arguments \code{mz}, \code{ppm}, \code{rt}
and \code{sec} are used to construct the targets.
See ?\link{makeTargets} for more information.

\item \code{plotEICs(msData)}: a method for plotting extracted ion chromatograms (EICs)
of data in an \linkS4class{msData} object.
The arguments for data collection are the same as the \link{EICs} method.
A \link[data.table]{date.table} can be used instead.
The \code{colorBy} argument can be \code{"analyses"}, \code{replicates} or \code{targets}
(the default), for coloring by analyses, replicates or EICs targets, respectively.
The \code{legendNames} is a character vector with the same length as targets for plotting and
can be used to legend the plot. Note that, by setting \code{legendNames} the \code{colorBy}
is set to "targets".

\item \code{TICs(msData)}: extracts total ion chromatograms (TICs)
for analyses in an \linkS4class{msData} object.

\item \code{plotTICs(msData)}: plots a total ion chromatogram (TIC) for abalyses in the object.
The colorBy argument can be "analyses" or "replicates"
to color the plot by analyses or by analysis replicates.

\item \code{XICs(msData)}: get three dimensional (\emph{m/z}, time and intensity)
extracted ion chromatograms (XICs) for specified \emph{m/z} and retention time pair targets
in analyses of an \linkS4class{msData} object. The arguments \code{mz}, \code{ppm}, \code{rt}, \code{sec} and \code{id}
are used to construct the targets. See ?\link{makeTargets} for more information.

\item \code{plotXICs(msData)}: plots three dimensional (\emph{m/z}, time and intensity)
extracted ion chromatograms (XICs) for specified \emph{m/z} and retention time pair targets
in analyses of an \linkS4class{msData} object. The arguments \code{mz}, \code{ppm}, \code{rt},
\code{sec} and \code{id} are used to construct the targets. See ?\link{makeTargets} for more information.
When \code{plotTargetMark} is \code{TRUE} a target is plotted representing the deviations as defined
by the arguments \code{ppmMark} and \code{secMark} in ppm and seconds, respectively.
When ranges were given to build the XIC, exact \emph{m/z} and time targets can be specified with
the argument \code{targetsMark}. \code{targetsMark} should be a two column table named mz and rt with
exact \emph{m/z} and time targets. Note that the number of rows should be the same as the number of target
in the XIC. The number of rows to plot multiple targets can be defined by the \code{numberRows} argument.

\item \code{MS2s(msData)}: get MS2 data for specified \emph{m/z} and retention time (seconds) targets
in analyses of an \linkS4class{msData} object. The \code{clusteringUnit} defines the method used for clustering.
Possible values are \emph{euclidean} (the default) or \emph{distance}.
The mass (in Da) and time (in seconds) isolation windows to screen for the respective precursors
are defined with the arguments \code{isolationMassWindow} and \code{isolationTimeWindow}, respectively.
The \code{clusteringUnit} and \code{clusteringWindow} define
the mass deviation unit and deviation to cluster mass traces from different spectra, respectively.
For the \code{clusteringUnit}, possible values are \emph{mz} (the default) or \emph{ppm}.
The \code{minIntensityPre} and \code{minIntensityPost}
define the minimum intensity for mass traces before and after clustering, respectively.
Set \code{mergeVoltages} to \code{TRUE} for merging spectra acquired with different collision energies.
The \code{mergeBy} argument is used to merge spectra by "samples" or "replicates".
When \code{NULL}, MS2 is given per target and per sample.

\item \code{plotMS2s(msData)}: plots MS2 data for specified \emph{m/z} and retention time (seconds) targets
in analyses of an \linkS4class{msData} object. The \code{clusteringUnit} defines the method used for clustering.
Possible values are \emph{euclidean} (the default) or \emph{distance}.
The mass (in Da) and time (in seconds) isolation windows to screen for the respective precursors
are defined with the arguments \code{isolationMassWindow} and \code{isolationTimeWindow}, respectively.
The \code{clusteringUnit} and \code{clusteringWindow} define
the mass deviation unit and deviation to cluster mass traces from different spectra, respectively.
For the \code{clusteringUnit}, possible values are \emph{mz} (the default) or \emph{ppm}.
The \code{minIntensityPre} and \code{minIntensityPost}
define the minimum intensity for mass traces before and after clustering, respectively.
Set \code{mergeVoltages} to \code{TRUE} for merging spectra acquired with different collision energies.
The \code{mergeBy} argument is used to merge spectra by "samples" or "replicates".
When \code{NULL}, MS2 is given per target and per sample. The possible values for the
\code{colorBy} argument are "targets", "samples", "replicates" and "voltages" to colour by
each target, sample, replicate or collision energy, respectively.

\item \code{loadRawData(msData)}: adds processing parameters to the analysis.

\item \code{spectra(msData)}: getter for slot spectra in analyses of an \linkS4class{msData} object.

\item \code{hasAdjustedRetentionTime(msData)}: getter for presence of adjusted retention time
in the analyses.

\item \code{addParameters(msData)}: adds processing parameters to analyses or features as defined
by the argument \code{where}. So where can be either "analyses" or "features".

\item \code{getParameters(msData)}: gets processing parameters from analyses or features as defined
by the argument \code{where}. So where can be either "analyses" or "features".

\item \code{as.features(msData)}: converts the \linkS4class{msData}
to a \linkS4class{features} object from the package \pkg{patRoon}.

\item \code{peaks(msData)}: getter for chromatographic peaks.
The arguments \code{targetID} and \code{mz}/\code{rt} can be used
to select specific peaks. The \emph{id} of peaks and/or features can be
given in the \code{targetsID} argument to select the respective peaks.
Also, analyses can be selected using the \code{analyses} argument.

\item \code{plotPeaks(msData)}: a method for plotting chromatographic peaks
in an \linkS4class{msData} object.
The arguments \code{targetID} and \code{mz}/\code{rt} can be used
to select specific peaks. The \emph{id} of peaks and/or features can be
given in the \code{targetsID} argument to select the respective peaks.
Also, analyses can be selected using the \code{analyses} argument.
The \code{colorBy} argument can be be \code{"analyses"}, \code{replicates} or \code{targets}
(the default), for coloring by analyses, replicates or peak targets, respectively.
The \code{legendNames} is a character vector with the same length as targets for plotting and
can be used to lengend the plot. Note that, by setting \code{legendNames} the \code{colorBy}
is set to "targets" automatically.

\item \code{mapPeaks(msData)}: a method for mapping peaks mass and time space.
The \code{colorBy} argument can be be \code{"analyses"}, \code{replicates} or \code{targets}
(the default), for coloring by analyses, replicates or peak targets, respectively.
The \code{legendNames} is a character vector with the same length as targets for plotting and
can be used to legend the plot. Note that, by setting \code{legendNames} the \code{colorBy}
is set to "targets" automatically.

\item \code{features(msData)}: getter for features (i.e., grouped peaks). When
complete is set to \code{TRUE}, additional feature metadata is also returned.

\item \code{plotFeatures(msData)}: A method for plotting peaks from given features
in an \linkS4class{msData} object.
The \code{colorBy} argument can be be \code{"analyses"}, \code{replicates} or \code{targets}
(the default), for coloring by analyses, replicates or peak targets, respectively.
The \code{legendNames} is a character vector with the same length as targets for plotting and
can be used to lengend the plot. Note that, by setting \code{legendNames} the \code{colorBy}
is set to "targets" automatically.

\item \code{as.featureGroups(msData)}: converts the \linkS4class{msData}
to a \linkS4class{featureGroups} object from the package \pkg{patRoon}.

\item \code{annotation(msData)}: getter for annotation (i.e., isotopes and adducts clusters).
When giving the argument \code{all} as \code{TRUE},
all the features within the target feature annotation cluster are included in the output.

\item \code{plotAnnotation(msData)}: plots peaks from given feature annotation targets
in the \linkS4class{msData} object. The \code{colorBy} argument can be
set to \code{"isotopes"} for coloring by monoisotopic mass instead of neutral mass of the cluster.

\item \code{x[i}: subset on analyses and features, using index or name.
Note that this method is irreversible.

\item \code{x[i}: subset on analyses, features and peaks, using index or name.
Note that this method is irreversible.

}}
\section{Slots}{

\describe{
\item{\code{title}}{A character string with the stream set title.}

\item{\code{date}}{The date of the set.}

\item{\code{path}}{A character string with the set path.}

\item{\code{analyses}}{A list with \linkS4class{msAnalysis} objects for each analysis file added to the project.}

\item{\code{features}}{A \code{list} with features grouped across samples in the project.}
}}

