% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_R6_msData.R
\name{msData}
\alias{msData}
\title{msData: R6 Class Representing an MS Set of Analysis Files}
\description{
A description of the R6 class.
}
\details{
The details.
}
\note{
Name, path and date cannot be removed.

When \code{call} is \code{NULL} all settings are removed.

Unique feature groups from the removed analyses are also removed.

The \linkS4class{settings} call must be set to "find_features".

The \linkS4class{settings} call must be set to "group_features".

When saved as json, the \code{data.table} is split by groups as
list and the \code{data.table} of each group is converted to a list.

When saved as json, the \code{data.table} is split by groups as
list and the \code{data.table} of each group is converted to a list.
}
\references{
\insertRef{patroon01}{streamFind}

\insertRef{patroon01}{streamFind}
}
\seealso{
\link[patRoon]{findFeatures}

\code{\link[patRoon]{groupFeatures}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-msData-new}{\code{msData$new()}}
\item \href{#method-msData-print}{\code{msData$print()}}
\item \href{#method-msData-get_header}{\code{msData$get_header()}}
\item \href{#method-msData-get_analyses}{\code{msData$get_analyses()}}
\item \href{#method-msData-get_number_analyses}{\code{msData$get_number_analyses()}}
\item \href{#method-msData-get_overview}{\code{msData$get_overview()}}
\item \href{#method-msData-get_analysis_names}{\code{msData$get_analysis_names()}}
\item \href{#method-msData-get_replicate_names}{\code{msData$get_replicate_names()}}
\item \href{#method-msData-get_blank_names}{\code{msData$get_blank_names()}}
\item \href{#method-msData-get_polarities}{\code{msData$get_polarities()}}
\item \href{#method-msData-get_file_paths}{\code{msData$get_file_paths()}}
\item \href{#method-msData-get_mz_high}{\code{msData$get_mz_high()}}
\item \href{#method-msData-get_rt_end}{\code{msData$get_rt_end()}}
\item \href{#method-msData-get_spectra_levels}{\code{msData$get_spectra_levels()}}
\item \href{#method-msData-get_spectra}{\code{msData$get_spectra()}}
\item \href{#method-msData-get_chromatograms}{\code{msData$get_chromatograms()}}
\item \href{#method-msData-get_tic}{\code{msData$get_tic()}}
\item \href{#method-msData-get_bpc}{\code{msData$get_bpc()}}
\item \href{#method-msData-get_eic}{\code{msData$get_eic()}}
\item \href{#method-msData-get_ms1}{\code{msData$get_ms1()}}
\item \href{#method-msData-get_ms2}{\code{msData$get_ms2()}}
\item \href{#method-msData-get_settings}{\code{msData$get_settings()}}
\item \href{#method-msData-get_features}{\code{msData$get_features()}}
\item \href{#method-msData-get_features_eic}{\code{msData$get_features_eic()}}
\item \href{#method-msData-get_features_ms1}{\code{msData$get_features_ms1()}}
\item \href{#method-msData-get_features_ms2}{\code{msData$get_features_ms2()}}
\item \href{#method-msData-get_alignment}{\code{msData$get_alignment()}}
\item \href{#method-msData-get_groups}{\code{msData$get_groups()}}
\item \href{#method-msData-get_groups_ms1}{\code{msData$get_groups_ms1()}}
\item \href{#method-msData-get_groups_ms2}{\code{msData$get_groups_ms2()}}
\item \href{#method-msData-add_header}{\code{msData$add_header()}}
\item \href{#method-msData-add_settings}{\code{msData$add_settings()}}
\item \href{#method-msData-add_analyses}{\code{msData$add_analyses()}}
\item \href{#method-msData-add_replicate_names}{\code{msData$add_replicate_names()}}
\item \href{#method-msData-add_blank_names}{\code{msData$add_blank_names()}}
\item \href{#method-msData-add_features}{\code{msData$add_features()}}
\item \href{#method-msData-add_groups}{\code{msData$add_groups()}}
\item \href{#method-msData-add_alignment}{\code{msData$add_alignment()}}
\item \href{#method-msData-remove_header}{\code{msData$remove_header()}}
\item \href{#method-msData-remove_settings}{\code{msData$remove_settings()}}
\item \href{#method-msData-remove_analyses}{\code{msData$remove_analyses()}}
\item \href{#method-msData-remove_groups}{\code{msData$remove_groups()}}
\item \href{#method-msData-remove_alignment}{\code{msData$remove_alignment()}}
\item \href{#method-msData-has_analyses}{\code{msData$has_analyses()}}
\item \href{#method-msData-has_loaded_spectra}{\code{msData$has_loaded_spectra()}}
\item \href{#method-msData-has_loaded_chromatograms}{\code{msData$has_loaded_chromatograms()}}
\item \href{#method-msData-has_settings}{\code{msData$has_settings()}}
\item \href{#method-msData-has_features}{\code{msData$has_features()}}
\item \href{#method-msData-has_alignment}{\code{msData$has_alignment()}}
\item \href{#method-msData-has_groups}{\code{msData$has_groups()}}
\item \href{#method-msData-plot_spectra}{\code{msData$plot_spectra()}}
\item \href{#method-msData-plot_xic}{\code{msData$plot_xic()}}
\item \href{#method-msData-plot_eic}{\code{msData$plot_eic()}}
\item \href{#method-msData-plot_tic}{\code{msData$plot_tic()}}
\item \href{#method-msData-plot_bpc}{\code{msData$plot_bpc()}}
\item \href{#method-msData-plot_ms2}{\code{msData$plot_ms2()}}
\item \href{#method-msData-plot_ms1}{\code{msData$plot_ms1()}}
\item \href{#method-msData-plot_features}{\code{msData$plot_features()}}
\item \href{#method-msData-map_features}{\code{msData$map_features()}}
\item \href{#method-msData-plot_features_ms1}{\code{msData$plot_features_ms1()}}
\item \href{#method-msData-plot_features_ms2}{\code{msData$plot_features_ms2()}}
\item \href{#method-msData-plot_alignment}{\code{msData$plot_alignment()}}
\item \href{#method-msData-plot_groups}{\code{msData$plot_groups()}}
\item \href{#method-msData-plot_groups_ms1}{\code{msData$plot_groups_ms1()}}
\item \href{#method-msData-plot_groups_ms2}{\code{msData$plot_groups_ms2()}}
\item \href{#method-msData-plot_groups_overview}{\code{msData$plot_groups_overview()}}
\item \href{#method-msData-load_spectra}{\code{msData$load_spectra()}}
\item \href{#method-msData-load_chromatograms}{\code{msData$load_chromatograms()}}
\item \href{#method-msData-find_features}{\code{msData$find_features()}}
\item \href{#method-msData-group_features}{\code{msData$group_features()}}
\item \href{#method-msData-as_features_patRoon}{\code{msData$as_features_patRoon()}}
\item \href{#method-msData-subset_analyses}{\code{msData$subset_analyses()}}
\item \href{#method-msData-check_analyses_argument}{\code{msData$check_analyses_argument()}}
\item \href{#method-msData-check_correspondence}{\code{msData$check_correspondence()}}
\item \href{#method-msData-save_header}{\code{msData$save_header()}}
\item \href{#method-msData-save_settings}{\code{msData$save_settings()}}
\item \href{#method-msData-save_analyses}{\code{msData$save_analyses()}}
\item \href{#method-msData-save_groups}{\code{msData$save_groups()}}
\item \href{#method-msData-save}{\code{msData$save()}}
\item \href{#method-msData-import_header}{\code{msData$import_header()}}
\item \href{#method-msData-import_settings}{\code{msData$import_settings()}}
\item \href{#method-msData-import_analyses}{\code{msData$import_analyses()}}
\item \href{#method-msData-import_groups}{\code{msData$import_groups()}}
\item \href{#method-msData-processing_function_calls}{\code{msData$processing_function_calls()}}
\item \href{#method-msData-clone}{\code{msData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-new"></a>}}
\if{latex}{\out{\hypertarget{method-msData-new}{}}}
\subsection{Method \code{new()}}{
Create a new MS set of analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$new(
  files = NULL,
  runParallel = FALSE,
  header = NULL,
  settings = NULL,
  analyses = NULL,
  groups = NULL,
  alignment = NULL,
  verbose = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{files}}{A character vector with the full file path of \emph{mzML}
or \emph{mzXML} file/s or a data.frame with the columns file, replicate
and blank with the full file path of \emph{mzML} or \emph{mzXML} file/s,
the replicate group name (string) and the associated blank replicate
group (string).}

\item{\code{runParallel}}{Logical, set to \code{TRUE} for processing the data
in parallel.}

\item{\code{header}}{A list with administrative information for the header.
When not given (i.e., \code{NULL}), a default header is generated with
name, path and date.}

\item{\code{settings}}{A list with settings.}

\item{\code{analyses}}{A list with MS analyses information.}

\item{\code{groups}}{A data.table with \code{groups} representing corresponding
features across MS analyses.}

\item{\code{alignment}}{X.}

\item{\code{verbose}}{X.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{msData} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-print"></a>}}
\if{latex}{\out{\hypertarget{method-msData-print}{}}}
\subsection{Method \code{print()}}{
Prints a summary of the \code{msData} object in the console.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Console text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-get_header"></a>}}
\if{latex}{\out{\hypertarget{method-msData-get_header}{}}}
\subsection{Method \code{get_header()}}{
Method to get the header of the \code{msData} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$get_header(value = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A character vector with the name of the entry.
Possible values are name, author, description, path and date.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The header list as defined by \code{value}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-get_analyses"></a>}}
\if{latex}{\out{\hypertarget{method-msData-get_analyses}{}}}
\subsection{Method \code{get_analyses()}}{
Method to get specific analyses from the \code{msData} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$get_analyses(analyses = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name
of the analyses.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The list of analyses defined by \code{value}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-get_number_analyses"></a>}}
\if{latex}{\out{\hypertarget{method-msData-get_number_analyses}{}}}
\subsection{Method \code{get_number_analyses()}}{
Method to get the number of analysis in the \code{msData} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$get_number_analyses()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An integer value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-get_overview"></a>}}
\if{latex}{\out{\hypertarget{method-msData-get_overview}{}}}
\subsection{Method \code{get_overview()}}{
Method to get the overview data.frame with all the analysis types,
names, replicates, associated blank replicates, polarities and full
file paths.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$get_overview()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data.frame with columns type, analysis, replicate, blank
polarity and file.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-get_analysis_names"></a>}}
\if{latex}{\out{\hypertarget{method-msData-get_analysis_names}{}}}
\subsection{Method \code{get_analysis_names()}}{
Method to get the analysis names.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$get_analysis_names(analyses = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name
of the analyses.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-get_replicate_names"></a>}}
\if{latex}{\out{\hypertarget{method-msData-get_replicate_names}{}}}
\subsection{Method \code{get_replicate_names()}}{
Method to get the analysis replicate names.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$get_replicate_names(analyses = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name
of the analyses.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-get_blank_names"></a>}}
\if{latex}{\out{\hypertarget{method-msData-get_blank_names}{}}}
\subsection{Method \code{get_blank_names()}}{
Method to get the analysis blank replicate names.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$get_blank_names(analyses = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name
of the analyses.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-get_polarities"></a>}}
\if{latex}{\out{\hypertarget{method-msData-get_polarities}{}}}
\subsection{Method \code{get_polarities()}}{
Method to get the polarity of the analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$get_polarities(analyses = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name
of the analyses.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-get_file_paths"></a>}}
\if{latex}{\out{\hypertarget{method-msData-get_file_paths}{}}}
\subsection{Method \code{get_file_paths()}}{
Method to get the polarity of the analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$get_file_paths(analyses = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name
of the analyses.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-get_mz_high"></a>}}
\if{latex}{\out{\hypertarget{method-msData-get_mz_high}{}}}
\subsection{Method \code{get_mz_high()}}{
Method to get the polarity of the analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$get_mz_high(analyses = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name
of the analyses.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-get_rt_end"></a>}}
\if{latex}{\out{\hypertarget{method-msData-get_rt_end}{}}}
\subsection{Method \code{get_rt_end()}}{
Method to get the polarity of the analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$get_rt_end(analyses = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name
of the analyses.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-get_spectra_levels"></a>}}
\if{latex}{\out{\hypertarget{method-msData-get_spectra_levels}{}}}
\subsection{Method \code{get_spectra_levels()}}{
Method to get the spectra levels of the analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$get_spectra_levels(analyses = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name
of the analyses.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list for each analysis with an integer vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-get_spectra"></a>}}
\if{latex}{\out{\hypertarget{method-msData-get_spectra}{}}}
\subsection{Method \code{get_spectra()}}{
Method to get spectra from the MS analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$get_spectra(
  analyses = NULL,
  levels = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  id = NULL,
  allTraces = TRUE,
  isolationWindow = 1.3,
  minIntensityMS1 = 0,
  minIntensityMS2 = 0,
  runParallel = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{X.}

\item{\code{levels}}{Name.}

\item{\code{mz}}{X.}

\item{\code{rt}}{X.}

\item{\code{ppm}}{X.}

\item{\code{sec}}{X.}

\item{\code{id}}{X.}

\item{\code{allTraces}}{Logical, when \code{TRUE} all level 2 data is returned.
When \code{FALSE} and level has 2, only the MS2 traces of MS1 targets
are returned, using the \code{preMZ} value and the \code{isolationWindow}.}

\item{\code{isolationWindow}}{X.}

\item{\code{minIntensityMS1}}{X.}

\item{\code{minIntensityMS2}}{X.}

\item{\code{runParallel}}{X.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame with spectra for each analyses and
targets when defined.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-get_chromatograms"></a>}}
\if{latex}{\out{\hypertarget{method-msData-get_chromatograms}{}}}
\subsection{Method \code{get_chromatograms()}}{
Method to get spectra from the MS analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$get_chromatograms(
  analyses = NULL,
  minIntensity = 0,
  runParallel = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{X.}

\item{\code{minIntensity}}{X.}

\item{\code{runParallel}}{X.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame with spectra.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-get_tic"></a>}}
\if{latex}{\out{\hypertarget{method-msData-get_tic}{}}}
\subsection{Method \code{get_tic()}}{
Method to get the total ion chromatograms (TIC) from the analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$get_tic(analyses = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name
of the analyses.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-get_bpc"></a>}}
\if{latex}{\out{\hypertarget{method-msData-get_bpc}{}}}
\subsection{Method \code{get_bpc()}}{
Method to get the base peak chromatograms (BPC) from the analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$get_bpc(analyses = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name
of the analyses.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-get_eic"></a>}}
\if{latex}{\out{\hypertarget{method-msData-get_eic}{}}}
\subsection{Method \code{get_eic()}}{
Method to get extract ion chromatograms (EIC) from the analyses based
on targets.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$get_eic(
  analyses = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  id = NULL,
  runParallel = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name
of the analyses.}

\item{\code{mz}}{X.}

\item{\code{rt}}{X.}

\item{\code{ppm}}{X.}

\item{\code{sec}}{X.}

\item{\code{id}}{X.}

\item{\code{runParallel}}{X.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-get_ms1"></a>}}
\if{latex}{\out{\hypertarget{method-msData-get_ms1}{}}}
\subsection{Method \code{get_ms1()}}{
Method to get MS1 data from the analyses based on targets.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$get_ms1(
  analyses = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  id = NULL,
  mzClust = 0.003,
  verbose = FALSE,
  minIntensity = 1000,
  runParallel = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name
of the analyses.}

\item{\code{mz}}{X.}

\item{\code{rt}}{X.}

\item{\code{ppm}}{X.}

\item{\code{sec}}{X.}

\item{\code{id}}{X.}

\item{\code{mzClust}}{X.}

\item{\code{verbose}}{X.}

\item{\code{minIntensity}}{X.}

\item{\code{runParallel}}{X.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-get_ms2"></a>}}
\if{latex}{\out{\hypertarget{method-msData-get_ms2}{}}}
\subsection{Method \code{get_ms2()}}{
Method to get MS2 data from the analyses based on targets.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$get_ms2(
  analyses = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  id = NULL,
  isolationWindow = 1.3,
  mzClust = 0.005,
  verbose = TRUE,
  minIntensity = 0,
  runParallel = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name
of the analyses.}

\item{\code{mz}}{X.}

\item{\code{rt}}{X.}

\item{\code{ppm}}{X.}

\item{\code{sec}}{X.}

\item{\code{id}}{X.}

\item{\code{isolationWindow}}{X.}

\item{\code{mzClust}}{X.}

\item{\code{verbose}}{X.}

\item{\code{minIntensity}}{X.}

\item{\code{runParallel}}{X.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-get_settings"></a>}}
\if{latex}{\out{\hypertarget{method-msData-get_settings}{}}}
\subsection{Method \code{get_settings()}}{
Method to get settings from analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$get_settings(call = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{call}}{A string with the name of function call.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-get_features"></a>}}
\if{latex}{\out{\hypertarget{method-msData-get_features}{}}}
\subsection{Method \code{get_features()}}{
Method to get features from analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$get_features(
  analyses = NULL,
  id = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  filtered = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name
of the analyses.}

\item{\code{id}}{X.}

\item{\code{mass}}{X.}

\item{\code{mz}}{X.}

\item{\code{rt}}{X.}

\item{\code{ppm}}{X.}

\item{\code{sec}}{X.}

\item{\code{filtered}}{X.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-get_features_eic"></a>}}
\if{latex}{\out{\hypertarget{method-msData-get_features_eic}{}}}
\subsection{Method \code{get_features_eic()}}{
Method to get EIC of features from analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$get_features_eic(
  analyses = NULL,
  id = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  rtExpand = 120,
  mzExpand = 0.005,
  filtered = FALSE,
  runParallel = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name
of the analyses.}

\item{\code{id}}{X.}

\item{\code{mass}}{X.}

\item{\code{mz}}{X.}

\item{\code{rt}}{X.}

\item{\code{ppm}}{X.}

\item{\code{sec}}{X.}

\item{\code{rtExpand}}{X.}

\item{\code{mzExpand}}{X.}

\item{\code{filtered}}{X.}

\item{\code{runParallel}}{X.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame/data.table.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-get_features_ms1"></a>}}
\if{latex}{\out{\hypertarget{method-msData-get_features_ms1}{}}}
\subsection{Method \code{get_features_ms1()}}{
Method to get an averaged MS1 spectrum for features in analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$get_features_ms1(
  analyses = NULL,
  id = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  rtWindow = c(-2, 2),
  mzWindow = c(-5, 100),
  mzClust = 0.003,
  minIntensity = 1000,
  verbose = TRUE,
  filtered = FALSE,
  runParallel = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name
of the analyses.}

\item{\code{id}}{X.}

\item{\code{mass}}{X.}

\item{\code{mz}}{X.}

\item{\code{rt}}{X.}

\item{\code{ppm}}{X.}

\item{\code{sec}}{X.}

\item{\code{rtWindow}}{X.}

\item{\code{mzWindow}}{X.}

\item{\code{mzClust}}{X.}

\item{\code{minIntensity}}{X.}

\item{\code{verbose}}{X.}

\item{\code{filtered}}{X.}

\item{\code{runParallel}}{X.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame/data.table.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-get_features_ms2"></a>}}
\if{latex}{\out{\hypertarget{method-msData-get_features_ms2}{}}}
\subsection{Method \code{get_features_ms2()}}{
Method to get an averaged MS2 spectrum for features in analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$get_features_ms2(
  analyses = NULL,
  id = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  isolationWindow = 1.3,
  mzClust = 0.003,
  minIntensity = 0,
  verbose = TRUE,
  filtered = FALSE,
  runParallel = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name
of the analyses.}

\item{\code{id}}{X.}

\item{\code{mass}}{X.}

\item{\code{mz}}{X.}

\item{\code{rt}}{X.}

\item{\code{ppm}}{X.}

\item{\code{sec}}{X.}

\item{\code{isolationWindow}}{X.}

\item{\code{mzClust}}{X.}

\item{\code{minIntensity}}{X.}

\item{\code{verbose}}{X.}

\item{\code{filtered}}{X.}

\item{\code{runParallel}}{X.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame/data.table.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-get_alignment"></a>}}
\if{latex}{\out{\hypertarget{method-msData-get_alignment}{}}}
\subsection{Method \code{get_alignment()}}{
Method to get alignment.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$get_alignment()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data.frame.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-get_groups"></a>}}
\if{latex}{\out{\hypertarget{method-msData-get_groups}{}}}
\subsection{Method \code{get_groups()}}{
Method to get feature groups from analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$get_groups(
  groups = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  filtered = FALSE,
  onlyIntensities = FALSE,
  average = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{groups}}{X.}

\item{\code{mass}}{X.}

\item{\code{mz}}{X.}

\item{\code{rt}}{X.}

\item{\code{ppm}}{X.}

\item{\code{sec}}{X.}

\item{\code{filtered}}{X.}

\item{\code{onlyIntensities}}{X.}

\item{\code{average}}{X.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame/data.table.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-get_groups_ms1"></a>}}
\if{latex}{\out{\hypertarget{method-msData-get_groups_ms1}{}}}
\subsection{Method \code{get_groups_ms1()}}{
Method to get an averaged MS1 spectrum for feature groups in analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$get_groups_ms1(
  groups = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  rtWindow = c(-2, 2),
  mzWindow = c(-5, 90),
  mzClustFeatures = 0.003,
  minIntensityFeatures = 1000,
  mzClustGroups = 0.003,
  minIntensityGroups = 1000,
  groupBy = "groups",
  verbose = TRUE,
  filtered = FALSE,
  runParallel = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{groups}}{X.}

\item{\code{mass}}{X.}

\item{\code{mz}}{X.}

\item{\code{rt}}{X.}

\item{\code{ppm}}{X.}

\item{\code{sec}}{X.}

\item{\code{rtWindow}}{X.}

\item{\code{mzWindow}}{X.}

\item{\code{mzClustFeatures}}{X.}

\item{\code{minIntensityFeatures}}{X.}

\item{\code{mzClustGroups}}{X.}

\item{\code{minIntensityGroups}}{X.}

\item{\code{groupBy}}{X.}

\item{\code{verbose}}{X.}

\item{\code{filtered}}{X.}

\item{\code{runParallel}}{X.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame/data.table.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-get_groups_ms2"></a>}}
\if{latex}{\out{\hypertarget{method-msData-get_groups_ms2}{}}}
\subsection{Method \code{get_groups_ms2()}}{
Method to get an averaged MS2 spectrum for feature groups in analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$get_groups_ms2(
  groups = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  isolationWindow = 1.3,
  mzClustFeatures = 0.003,
  minIntensityFeatures = 100,
  mzClustGroups = 0.003,
  minIntensityGroups = 100,
  groupBy = "groups",
  verbose = TRUE,
  filtered = FALSE,
  runParallel = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{groups}}{X.}

\item{\code{mass}}{X.}

\item{\code{mz}}{X.}

\item{\code{rt}}{X.}

\item{\code{ppm}}{X.}

\item{\code{sec}}{X.}

\item{\code{isolationWindow}}{X.}

\item{\code{mzClustFeatures}}{X.}

\item{\code{minIntensityFeatures}}{X.}

\item{\code{mzClustGroups}}{X.}

\item{\code{minIntensityGroups}}{X.}

\item{\code{groupBy}}{X.}

\item{\code{verbose}}{X.}

\item{\code{filtered}}{X.}

\item{\code{runParallel}}{X.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame/data.table.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-add_header"></a>}}
\if{latex}{\out{\hypertarget{method-msData-add_header}{}}}
\subsection{Method \code{add_header()}}{
Method to add header information to the \code{msData} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$add_header(header = NULL, verbose = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{header}}{X.}

\item{\code{verbose}}{X.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-add_settings"></a>}}
\if{latex}{\out{\hypertarget{method-msData-add_settings}{}}}
\subsection{Method \code{add_settings()}}{
Method to add processing settings to the \code{msData} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$add_settings(settings = NULL, verbose = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{settings}}{X.}

\item{\code{verbose}}{X.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-add_analyses"></a>}}
\if{latex}{\out{\hypertarget{method-msData-add_analyses}{}}}
\subsection{Method \code{add_analyses()}}{
Method to add analyses to the \code{msData} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$add_analyses(analyses, verbose = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A list of analyses.}

\item{\code{verbose}}{X.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-add_replicate_names"></a>}}
\if{latex}{\out{\hypertarget{method-msData-add_replicate_names}{}}}
\subsection{Method \code{add_replicate_names()}}{
Method to add or redefine the analysis replicate names. Changes the
\code{msData} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$add_replicate_names(value = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A character vector with the analysis replicate names.
Must be of the same length as the number of analyses.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-add_blank_names"></a>}}
\if{latex}{\out{\hypertarget{method-msData-add_blank_names}{}}}
\subsection{Method \code{add_blank_names()}}{
Method to add or redefine the analysis blank replicate names.
Changes the \code{msData} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$add_blank_names(value = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A character vector with the analysis blank replicate names.
Must be of the same length as the number of analyses.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-add_features"></a>}}
\if{latex}{\out{\hypertarget{method-msData-add_features}{}}}
\subsection{Method \code{add_features()}}{
Method to add features to each analysis in the \code{msData} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$add_features(features = NULL, verbose = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{features}}{X.}

\item{\code{verbose}}{X.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-add_groups"></a>}}
\if{latex}{\out{\hypertarget{method-msData-add_groups}{}}}
\subsection{Method \code{add_groups()}}{
Method to add feature groups in the \code{msData} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$add_groups(groups = NULL, verbose = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{groups}}{data.table.}

\item{\code{verbose}}{X.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-add_alignment"></a>}}
\if{latex}{\out{\hypertarget{method-msData-add_alignment}{}}}
\subsection{Method \code{add_alignment()}}{
Method to add time alignment results in the \code{msData} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$add_alignment(alignment = NULL, verbose = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alignment}}{list.}

\item{\code{verbose}}{X.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-remove_header"></a>}}
\if{latex}{\out{\hypertarget{method-msData-remove_header}{}}}
\subsection{Method \code{remove_header()}}{
Removes header entries from the \code{msData} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$remove_header(value = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A character vector with the names of the header entries
to be removed.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-remove_settings"></a>}}
\if{latex}{\out{\hypertarget{method-msData-remove_settings}{}}}
\subsection{Method \code{remove_settings()}}{
Removes settings from the \code{msData} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$remove_settings(call = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{call}}{A character vector with the settings call name.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-remove_analyses"></a>}}
\if{latex}{\out{\hypertarget{method-msData-remove_analyses}{}}}
\subsection{Method \code{remove_analyses()}}{
Removes analyses from the \code{msData} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$remove_analyses(analyses = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A character vector with the names or indices of the
analyses to be removed.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-remove_groups"></a>}}
\if{latex}{\out{\hypertarget{method-msData-remove_groups}{}}}
\subsection{Method \code{remove_groups()}}{
Remove feature groups from the \code{msData} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$remove_groups(groups = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{groups}}{X.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-remove_alignment"></a>}}
\if{latex}{\out{\hypertarget{method-msData-remove_alignment}{}}}
\subsection{Method \code{remove_alignment()}}{
Removes alignment results from the \code{msData} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$remove_alignment()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-has_analyses"></a>}}
\if{latex}{\out{\hypertarget{method-msData-has_analyses}{}}}
\subsection{Method \code{has_analyses()}}{
Method to check of the \code{msData} object has analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$has_analyses()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Logical value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-has_loaded_spectra"></a>}}
\if{latex}{\out{\hypertarget{method-msData-has_loaded_spectra}{}}}
\subsection{Method \code{has_loaded_spectra()}}{
Method to check for loaded spectra in given analyses names/indices.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$has_loaded_spectra(analyses = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{The analyses names/indices to check for loaded spectra.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Logical value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-has_loaded_chromatograms"></a>}}
\if{latex}{\out{\hypertarget{method-msData-has_loaded_chromatograms}{}}}
\subsection{Method \code{has_loaded_chromatograms()}}{
Method to check for loaded chromatograms in given analyses names/indices.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$has_loaded_chromatograms(analyses = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{The analyses names/indices to check for loaded
chromatograms.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Logical value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-has_settings"></a>}}
\if{latex}{\out{\hypertarget{method-msData-has_settings}{}}}
\subsection{Method \code{has_settings()}}{
Method to check if there are processing settings in the \code{msData} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$has_settings(call = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{call}}{A string with the name of function call.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Logical value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-has_features"></a>}}
\if{latex}{\out{\hypertarget{method-msData-has_features}{}}}
\subsection{Method \code{has_features()}}{
Method to check if given analyses have features.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$has_features(analyses = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{The analyses names/indices to check for loaded spectra.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Logical value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-has_alignment"></a>}}
\if{latex}{\out{\hypertarget{method-msData-has_alignment}{}}}
\subsection{Method \code{has_alignment()}}{
Method to check if there is alignment of retention time from grouping
features across analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$has_alignment()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Logical value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-has_groups"></a>}}
\if{latex}{\out{\hypertarget{method-msData-has_groups}{}}}
\subsection{Method \code{has_groups()}}{
Method to check if there are feature groups from grouping features
across analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$has_groups()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Logical value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-plot_spectra"></a>}}
\if{latex}{\out{\hypertarget{method-msData-plot_spectra}{}}}
\subsection{Method \code{plot_spectra()}}{
Plots spectra for given MS analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$plot_spectra(
  analyses = NULL,
  levels = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  id = NULL,
  allTraces = TRUE,
  isolationWindow = 1.3,
  minIntensityMS1 = 0,
  minIntensityMS2 = 0,
  runParallel = FALSE,
  colorBy = "analyses"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{X.}

\item{\code{levels}}{Name.}

\item{\code{mz}}{X.}

\item{\code{rt}}{X.}

\item{\code{ppm}}{X.}

\item{\code{sec}}{X.}

\item{\code{id}}{X.}

\item{\code{allTraces}}{Logical, when \code{TRUE} all level 2 data is returned.
When \code{FALSE} and level has 2, only the MS2 traces of MS1 targets
are returned, using the \code{preMZ} value and the \code{isolationWindow}.}

\item{\code{isolationWindow}}{X.}

\item{\code{minIntensityMS1}}{X.}

\item{\code{minIntensityMS2}}{X.}

\item{\code{runParallel}}{X.}

\item{\code{colorBy}}{X.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A 3D interactive plot.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-plot_xic"></a>}}
\if{latex}{\out{\hypertarget{method-msData-plot_xic}{}}}
\subsection{Method \code{plot_xic()}}{
Method to plot extract ion chromatograms (EIC) and \emph{m/z} vs
retention time from the analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$plot_xic(
  analyses = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  id = NULL,
  runParallel = FALSE,
  legendNames = NULL,
  plotTargetMark = TRUE,
  targetsMark = NULL,
  ppmMark = 5,
  secMark = 10,
  numberRows = 1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name
of the analyses.}

\item{\code{mz}}{X.}

\item{\code{rt}}{X.}

\item{\code{ppm}}{X.}

\item{\code{sec}}{X.}

\item{\code{id}}{X.}

\item{\code{runParallel}}{X.}

\item{\code{legendNames}}{x.}

\item{\code{plotTargetMark}}{x.}

\item{\code{targetsMark}}{x.}

\item{\code{ppmMark}}{x.}

\item{\code{secMark}}{x.}

\item{\code{numberRows}}{x.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-plot_eic"></a>}}
\if{latex}{\out{\hypertarget{method-msData-plot_eic}{}}}
\subsection{Method \code{plot_eic()}}{
Method to plot extract ion chromatograms (EIC) from the analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$plot_eic(
  analyses = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  id = NULL,
  runParallel = FALSE,
  legendNames = NULL,
  title = NULL,
  colorBy = "targets",
  interactive = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name
of the analyses.}

\item{\code{mz}}{X.}

\item{\code{rt}}{X.}

\item{\code{ppm}}{X.}

\item{\code{sec}}{X.}

\item{\code{id}}{X.}

\item{\code{runParallel}}{X.}

\item{\code{legendNames}}{x.}

\item{\code{title}}{x.}

\item{\code{colorBy}}{x.}

\item{\code{interactive}}{x.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A plot.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-plot_tic"></a>}}
\if{latex}{\out{\hypertarget{method-msData-plot_tic}{}}}
\subsection{Method \code{plot_tic()}}{
Method to plot total ion chromatograms (TIC) of the analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$plot_tic(
  analyses = NULL,
  title = NULL,
  colorBy = "analyses",
  interactive = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name
of the analyses.}

\item{\code{title}}{x.}

\item{\code{colorBy}}{x.}

\item{\code{interactive}}{x.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A plot.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-plot_bpc"></a>}}
\if{latex}{\out{\hypertarget{method-msData-plot_bpc}{}}}
\subsection{Method \code{plot_bpc()}}{
Method to plot base peak chromatograms (BPC) of the analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$plot_bpc(
  analyses = NULL,
  title = NULL,
  colorBy = "analyses",
  interactive = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name
of the analyses.}

\item{\code{title}}{x.}

\item{\code{colorBy}}{x.}

\item{\code{interactive}}{x.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A plot.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-plot_ms2"></a>}}
\if{latex}{\out{\hypertarget{method-msData-plot_ms2}{}}}
\subsection{Method \code{plot_ms2()}}{
Method to plot MS2 spectra from the analyses based on targets.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$plot_ms2(
  analyses = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  id = NULL,
  isolationWindow = 1.3,
  mzClust = 0.005,
  verbose = TRUE,
  minIntensity = 0,
  runParallel = FALSE,
  legendNames = NULL,
  title = NULL,
  colorBy = "targets",
  interactive = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name
of the analyses.}

\item{\code{mz}}{X.}

\item{\code{rt}}{X.}

\item{\code{ppm}}{X.}

\item{\code{sec}}{X.}

\item{\code{id}}{X.}

\item{\code{isolationWindow}}{X.}

\item{\code{mzClust}}{X.}

\item{\code{verbose}}{X.}

\item{\code{minIntensity}}{X.}

\item{\code{runParallel}}{X.}

\item{\code{legendNames}}{x.}

\item{\code{title}}{x.}

\item{\code{colorBy}}{x.}

\item{\code{interactive}}{x.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A plot.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-plot_ms1"></a>}}
\if{latex}{\out{\hypertarget{method-msData-plot_ms1}{}}}
\subsection{Method \code{plot_ms1()}}{
Method to plot MS1 spectra from the analyses based on targets.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$plot_ms1(
  analyses = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  id = NULL,
  mzClust = 0.001,
  verbose = FALSE,
  minIntensity = 1000,
  runParallel = FALSE,
  legendNames = NULL,
  title = NULL,
  colorBy = "targets",
  interactive = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name
of the analyses.}

\item{\code{mz}}{X.}

\item{\code{rt}}{X.}

\item{\code{ppm}}{X.}

\item{\code{sec}}{X.}

\item{\code{id}}{X.}

\item{\code{mzClust}}{X.}

\item{\code{verbose}}{X.}

\item{\code{minIntensity}}{X.}

\item{\code{runParallel}}{X.}

\item{\code{legendNames}}{x.}

\item{\code{title}}{x.}

\item{\code{colorBy}}{x.}

\item{\code{interactive}}{x.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A plot.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-plot_features"></a>}}
\if{latex}{\out{\hypertarget{method-msData-plot_features}{}}}
\subsection{Method \code{plot_features()}}{
Method to plot features from analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$plot_features(
  analyses = NULL,
  id = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  rtExpand = 120,
  mzExpand = 0.005,
  filtered = FALSE,
  runParallel = FALSE,
  legendNames = NULL,
  title = NULL,
  colorBy = "targets",
  interactive = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name
of the analyses.}

\item{\code{id}}{X.}

\item{\code{mass}}{X.}

\item{\code{mz}}{X.}

\item{\code{rt}}{X.}

\item{\code{ppm}}{X.}

\item{\code{sec}}{X.}

\item{\code{rtExpand}}{X.}

\item{\code{mzExpand}}{X.}

\item{\code{filtered}}{X.}

\item{\code{runParallel}}{X.}

\item{\code{legendNames}}{x.}

\item{\code{title}}{x.}

\item{\code{colorBy}}{x.}

\item{\code{interactive}}{x.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
plot.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-map_features"></a>}}
\if{latex}{\out{\hypertarget{method-msData-map_features}{}}}
\subsection{Method \code{map_features()}}{
Method to map retention time and \emph{m/z} of features from analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$map_features(
  analyses = NULL,
  id = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  filtered = FALSE,
  xlim = 30,
  ylim = 0.05,
  showLegend = TRUE,
  legendNames = NULL,
  title = NULL,
  colorBy = "targets",
  interactive = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name
of the analyses.}

\item{\code{id}}{X.}

\item{\code{mass}}{X.}

\item{\code{mz}}{X.}

\item{\code{rt}}{X.}

\item{\code{ppm}}{X.}

\item{\code{sec}}{X.}

\item{\code{filtered}}{X.}

\item{\code{xlim}}{X.}

\item{\code{ylim}}{X.}

\item{\code{showLegend}}{X.}

\item{\code{legendNames}}{x.}

\item{\code{title}}{x.}

\item{\code{colorBy}}{x.}

\item{\code{interactive}}{x.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A plot.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-plot_features_ms1"></a>}}
\if{latex}{\out{\hypertarget{method-msData-plot_features_ms1}{}}}
\subsection{Method \code{plot_features_ms1()}}{
Method to plot MS1 spectra from features in the analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$plot_features_ms1(
  analyses = NULL,
  id = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  rtWindow = c(-2, 2),
  mzWindow = c(-5, 100),
  mzClust = 0.003,
  minIntensity = 1000,
  verbose = TRUE,
  filtered = FALSE,
  runParallel = FALSE,
  legendNames = NULL,
  title = NULL,
  colorBy = "targets",
  interactive = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name
of the analyses.}

\item{\code{id}}{X.}

\item{\code{mass}}{X.}

\item{\code{mz}}{X.}

\item{\code{rt}}{X.}

\item{\code{ppm}}{X.}

\item{\code{sec}}{X.}

\item{\code{rtWindow}}{X.}

\item{\code{mzWindow}}{X.}

\item{\code{mzClust}}{X.}

\item{\code{minIntensity}}{X.}

\item{\code{verbose}}{X.}

\item{\code{filtered}}{X.}

\item{\code{runParallel}}{X.}

\item{\code{legendNames}}{x.}

\item{\code{title}}{x.}

\item{\code{colorBy}}{x.}

\item{\code{interactive}}{x.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A plot.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-plot_features_ms2"></a>}}
\if{latex}{\out{\hypertarget{method-msData-plot_features_ms2}{}}}
\subsection{Method \code{plot_features_ms2()}}{
Method to plot MS2 spectra from features in the analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$plot_features_ms2(
  analyses = NULL,
  id = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  isolationWindow = 1.3,
  mzClust = 0.005,
  minIntensity = 0,
  verbose = TRUE,
  filtered = FALSE,
  runParallel = FALSE,
  legendNames = NULL,
  title = NULL,
  colorBy = "targets",
  interactive = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name
of the analyses.}

\item{\code{id}}{X.}

\item{\code{mass}}{X.}

\item{\code{mz}}{X.}

\item{\code{rt}}{X.}

\item{\code{ppm}}{X.}

\item{\code{sec}}{X.}

\item{\code{isolationWindow}}{X.}

\item{\code{mzClust}}{X.}

\item{\code{minIntensity}}{X.}

\item{\code{verbose}}{X.}

\item{\code{filtered}}{X.}

\item{\code{runParallel}}{X.}

\item{\code{legendNames}}{x.}

\item{\code{title}}{x.}

\item{\code{colorBy}}{x.}

\item{\code{interactive}}{x.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A plot.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-plot_alignment"></a>}}
\if{latex}{\out{\hypertarget{method-msData-plot_alignment}{}}}
\subsection{Method \code{plot_alignment()}}{
Plots the results from the retention time alignment across analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$plot_alignment()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A plot with the retention time alignment differences
for each sample.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-plot_groups"></a>}}
\if{latex}{\out{\hypertarget{method-msData-plot_groups}{}}}
\subsection{Method \code{plot_groups()}}{
Method to plot feature groups EIC.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$plot_groups(
  groups = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  rtExpand = 120,
  mzExpand = 0.005,
  filtered = FALSE,
  runParallel = FALSE,
  legendNames = NULL,
  title = NULL,
  colorBy = "targets",
  interactive = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{groups}}{X.}

\item{\code{mass}}{X.}

\item{\code{mz}}{X.}

\item{\code{rt}}{X.}

\item{\code{ppm}}{X.}

\item{\code{sec}}{X.}

\item{\code{rtExpand}}{X.}

\item{\code{mzExpand}}{X.}

\item{\code{filtered}}{X.}

\item{\code{runParallel}}{X.}

\item{\code{legendNames}}{x.}

\item{\code{title}}{x.}

\item{\code{colorBy}}{x.}

\item{\code{interactive}}{x.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A plot.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-plot_groups_ms1"></a>}}
\if{latex}{\out{\hypertarget{method-msData-plot_groups_ms1}{}}}
\subsection{Method \code{plot_groups_ms1()}}{
Method to plot MS1 spectra from features in the analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$plot_groups_ms1(
  groups = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  rtWindow = c(-2, 2),
  mzWindow = c(-5, 90),
  mzClustFeatures = 0.005,
  minIntensityFeatures = 1000,
  mzClustGroups = 0.005,
  minIntensityGroups = 1000,
  verbose = TRUE,
  filtered = FALSE,
  runParallel = FALSE,
  legendNames = NULL,
  title = NULL,
  colorBy = "targets",
  interactive = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{groups}}{X.}

\item{\code{mass}}{X.}

\item{\code{mz}}{X.}

\item{\code{rt}}{X.}

\item{\code{ppm}}{X.}

\item{\code{sec}}{X.}

\item{\code{rtWindow}}{X.}

\item{\code{mzWindow}}{X.}

\item{\code{mzClustFeatures}}{X.}

\item{\code{minIntensityFeatures}}{X.}

\item{\code{mzClustGroups}}{X.}

\item{\code{minIntensityGroups}}{X.}

\item{\code{verbose}}{X.}

\item{\code{filtered}}{X.}

\item{\code{runParallel}}{X.}

\item{\code{legendNames}}{x.}

\item{\code{title}}{x.}

\item{\code{colorBy}}{x.}

\item{\code{interactive}}{x.}

\item{\code{groupBy}}{X.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A plot.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-plot_groups_ms2"></a>}}
\if{latex}{\out{\hypertarget{method-msData-plot_groups_ms2}{}}}
\subsection{Method \code{plot_groups_ms2()}}{
Method to plot MS1 spectra from features in the analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$plot_groups_ms2(
  groups = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  isolationWindow = 1.3,
  mzClustFeatures = 0.003,
  minIntensityFeatures = 100,
  mzClustGroups = 0.003,
  minIntensityGroups = 100,
  verbose = TRUE,
  filtered = FALSE,
  runParallel = FALSE,
  legendNames = NULL,
  title = NULL,
  colorBy = "targets",
  interactive = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{groups}}{X.}

\item{\code{mass}}{X.}

\item{\code{mz}}{X.}

\item{\code{rt}}{X.}

\item{\code{ppm}}{X.}

\item{\code{sec}}{X.}

\item{\code{isolationWindow}}{X.}

\item{\code{mzClustFeatures}}{X.}

\item{\code{minIntensityFeatures}}{X.}

\item{\code{mzClustGroups}}{X.}

\item{\code{minIntensityGroups}}{X.}

\item{\code{verbose}}{X.}

\item{\code{filtered}}{X.}

\item{\code{runParallel}}{X.}

\item{\code{legendNames}}{x.}

\item{\code{title}}{x.}

\item{\code{colorBy}}{x.}

\item{\code{interactive}}{x.}

\item{\code{groupBy}}{X.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A plot.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-plot_groups_overview"></a>}}
\if{latex}{\out{\hypertarget{method-msData-plot_groups_overview}{}}}
\subsection{Method \code{plot_groups_overview()}}{
Method to give an overview of the EIC, alignment and intensity variance
from features within target feature groups.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$plot_groups_overview(
  analyses = NULL,
  groups = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  rtExpand = 120,
  mzExpand = 0.005,
  filtered = FALSE,
  runParallel = FALSE,
  legendNames = NULL,
  title = NULL,
  heights = c(0.35, 0.5, 0.15)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{X.}

\item{\code{groups}}{X.}

\item{\code{mass}}{X.}

\item{\code{mz}}{X.}

\item{\code{rt}}{X.}

\item{\code{ppm}}{X.}

\item{\code{sec}}{X.}

\item{\code{rtExpand}}{X.}

\item{\code{mzExpand}}{X.}

\item{\code{filtered}}{X.}

\item{\code{runParallel}}{X.}

\item{\code{legendNames}}{x.}

\item{\code{title}}{x.}

\item{\code{heights}}{x.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A plot.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-load_spectra"></a>}}
\if{latex}{\out{\hypertarget{method-msData-load_spectra}{}}}
\subsection{Method \code{load_spectra()}}{
Method to load all spectra from analyses to the \code{msData} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$load_spectra(runParallel = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{runParallel}}{X.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-load_chromatograms"></a>}}
\if{latex}{\out{\hypertarget{method-msData-load_chromatograms}{}}}
\subsection{Method \code{load_chromatograms()}}{
Method to load all chromatograms from analyses to the \code{msData} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$load_chromatograms(runParallel = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{runParallel}}{X.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-find_features"></a>}}
\if{latex}{\out{\hypertarget{method-msData-find_features}{}}}
\subsection{Method \code{find_features()}}{
Finds features (i.e., chromatographic peaks) from MS data
in an \code{msData} class object. The function uses the \pkg{patRoon} package
for peak finding, enabling the use of several algorithms (see details).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$find_features(settings = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{settings}}{A \linkS4class{settings} object. When not given,
\emph{find_features} \linkS4class{settings} will be searched within
the \code{msData} object.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
See the \link[patRoon]{findFeatures} function from the
\pkg{patRoon} package or the
\href{https://rickhelmus.github.io/patRoon/reference/findFeatures.html}{reference guide} for more information. The following algorithms are
available via \pkg{patRoon}: "xcms3", "xcms", "openms", "envipick",
"sirius", "kpic2", "safd". The algorithm slot in the
\linkS4class{settings} should be one of the described. The parameters
are given as a list and should match with algorithm requirements.
Certain algorithms also require defined MS file formats and data in
profile mode.
}

\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-group_features"></a>}}
\if{latex}{\out{\hypertarget{method-msData-group_features}{}}}
\subsection{Method \code{group_features()}}{
Groups and aligns features across analyses in the \code{msData}
object. The function uses the \pkg{patRoon} package for grouping
features, enabling the use of several algorithms (see details).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$group_features(settings = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{settings}}{A \linkS4class{settings} object. When not given,
\emph{group_features} \linkS4class{settings} will be searched within
the \code{msData} object.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
See the \link[patRoon]{groupFeatures} function from the
\pkg{patRoon} package or the
\href{https://rickhelmus.github.io/patRoon/reference/groupFeatures.html}{reference guide} for more information. The following algorithms are
possible: "xcms3", "xcms", "openms" or "kpic2". The algorithm slot in the
\linkS4class{settings} should be one of the described. The parameters
are given as a list and should match with algorithm requirements.
}

\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-as_features_patRoon"></a>}}
\if{latex}{\out{\hypertarget{method-msData-as_features_patRoon}{}}}
\subsection{Method \code{as_features_patRoon()}}{
Creates a \linkS4class{features} object from the \pkg{patRoon} package
with the features in the analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$as_features_patRoon()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \linkS4class{features} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-subset_analyses"></a>}}
\if{latex}{\out{\hypertarget{method-msData-subset_analyses}{}}}
\subsection{Method \code{subset_analyses()}}{
Subsets an \code{msData} object on analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$subset_analyses(analyses = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{X.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new cloned \code{msData} object with only the analyses as defined
by the \code{analyses} argument.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-check_analyses_argument"></a>}}
\if{latex}{\out{\hypertarget{method-msData-check_analyses_argument}{}}}
\subsection{Method \code{check_analyses_argument()}}{
Checks the analyses argument as a character/integer vector to match
analyses names or indices from the \code{msData} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$check_analyses_argument(analyses = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{X.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A valid character vector with analyses names of \code{NULL}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-check_correspondence"></a>}}
\if{latex}{\out{\hypertarget{method-msData-check_correspondence}{}}}
\subsection{Method \code{check_correspondence()}}{
Checks the correspondence of features within feature groups in
the \code{msData} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$check_correspondence()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{TRUE} or \code{FALSE}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-save_header"></a>}}
\if{latex}{\out{\hypertarget{method-msData-save_header}{}}}
\subsection{Method \code{save_header()}}{
Method to save the header list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$save_header(format = "json", name = "header", path = getwd())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{format}}{X.}

\item{\code{name}}{X.}

\item{\code{path}}{X.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Saves the header list as the defined \code{format} in
\code{path} and returns invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-save_settings"></a>}}
\if{latex}{\out{\hypertarget{method-msData-save_settings}{}}}
\subsection{Method \code{save_settings()}}{
Method to save settings list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$save_settings(
  call = NULL,
  format = "json",
  name = "settings",
  path = getwd()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{call}}{x:}

\item{\code{format}}{X.}

\item{\code{name}}{X.}

\item{\code{path}}{X.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Saves the settings list as the defined \code{format} in
\code{path} and returns invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-save_analyses"></a>}}
\if{latex}{\out{\hypertarget{method-msData-save_analyses}{}}}
\subsection{Method \code{save_analyses()}}{
Method to save the list of analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$save_analyses(
  analyses = NULL,
  format = "json",
  name = "analyses",
  path = getwd()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{X.}

\item{\code{format}}{X.}

\item{\code{name}}{X.}

\item{\code{path}}{X.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Saves the list of analyses as the defined \code{format} in
\code{path} and returns invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-save_groups"></a>}}
\if{latex}{\out{\hypertarget{method-msData-save_groups}{}}}
\subsection{Method \code{save_groups()}}{
Method to save the feature groups \code{data.table}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$save_groups(format = "json", name = "groups", path = getwd())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{format}}{X.}

\item{\code{name}}{X.}

\item{\code{path}}{X.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Saves the groups \code{data.table} as the defined \code{format}
in \code{path} and returns invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-save"></a>}}
\if{latex}{\out{\hypertarget{method-msData-save}{}}}
\subsection{Method \code{save()}}{
Method to save the private fields (i.e., header, settings, analyses,
groups and alignment) of the msData object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$save(format = "json", name = "msData", path = getwd())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{format}}{X.}

\item{\code{name}}{X.}

\item{\code{path}}{X.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Saves the private fields as the defined \code{format} in
\code{path} and returns invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-import_header"></a>}}
\if{latex}{\out{\hypertarget{method-msData-import_header}{}}}
\subsection{Method \code{import_header()}}{
Method to import header to the \code{msData} object from a \emph{rds} or
\emph{json} file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$import_header(file = NA_character_, list = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{X.}

\item{\code{list}}{X.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-import_settings"></a>}}
\if{latex}{\out{\hypertarget{method-msData-import_settings}{}}}
\subsection{Method \code{import_settings()}}{
Method to import processing settings to the \code{msData} object from a
\emph{rds} or \emph{json} file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$import_settings(file = NA_character_)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{X.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-import_analyses"></a>}}
\if{latex}{\out{\hypertarget{method-msData-import_analyses}{}}}
\subsection{Method \code{import_analyses()}}{
Method to import analyses to the \code{msData} object from a \emph{rds} or
\emph{json} file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$import_analyses(file = NA_character_)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{X.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-import_groups"></a>}}
\if{latex}{\out{\hypertarget{method-msData-import_groups}{}}}
\subsection{Method \code{import_groups()}}{
Method to import feature groups to the \code{msData} object from a
\emph{rds} or \emph{json} file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$import_groups(file = NA_character_)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{X.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-processing_function_calls"></a>}}
\if{latex}{\out{\hypertarget{method-msData-processing_function_calls}{}}}
\subsection{Method \code{processing_function_calls()}}{
Possible processing function calls.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$processing_function_calls()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character vector with ordered possible function calls for data
pre and post-processing.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-msData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{msData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
