% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_class_003_msFeatures.R
\docType{class}
\name{msFeatures-class}
\alias{msFeatures-class}
\alias{analyses,msFeatures-method}
\alias{analyses,msFeatures,msFeatures-method}
\alias{replicates,msFeatures-method}
\alias{replicates,msFeatures,msFeatures-method}
\alias{blanks,msFeatures-method}
\alias{blanks,msFeatures,msFeatures-method}
\alias{addParameters,msFeatures-method}
\alias{addParameters,msFeatures,msFeatures-method}
\alias{getParameters,msFeatures-method}
\alias{getParameters,msFeatures,msFeatures-method}
\alias{features,msFeatures-method}
\alias{features,msFeatures,msFeatures-method}
\alias{[,msFeatures,ANY,missing,missing-method}
\alias{[,msFeatures,ANY,ANY,missing-method}
\title{msFeatures-class}
\usage{
\S4method{analyses}{msFeatures}(obj)

\S4method{replicates}{msFeatures}(object)

\S4method{blanks}{msFeatures}(object)

\S4method{addParameters}{msFeatures}(object, settings)

\S4method{getParameters}{msFeatures}(object, call = NULL)

\S4method{features}{msFeatures}(
  object,
  targetsID = NULL,
  mz = NULL,
  ppm = 20,
  rt = NULL,
  sec = 60,
  filtered = TRUE,
  complete = FALSE,
  average = TRUE
)

\S4method{[}{msFeatures,ANY,missing,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{msFeatures,ANY,ANY,missing}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{call}{The call name of the settings to retrieve.}

\item{complete}{Logical, set to \code{TRUE} for a complete version of the output.}

\item{i}{The indice/s or name/s of the features to keep in the \code{x} object.}

\item{settigns}{A \linkS4class{settings} object.}
}
\description{
An S4 class representing an MS sample/file within the \pkg{streamFind} package.
The \code{msFeatures} is used to store and manage MS data and the respective methods can be used
for inspection, processing and evaluation.
}
\section{Methods (by generic)}{
\itemize{
\item \code{analyses(msFeatures)}: getter for analysis names.

\item \code{replicates(msFeatures)}: getter for replicate names.

\item \code{blanks(msFeatures)}: getter for blank names.

\item \code{addParameters(msFeatures)}: adds processing parameters for features.

\item \code{getParameters(msFeatures)}: gets processing parameters.

\item \code{features(msFeatures)}: getter for features (i.e., grouped peaks). When
complete is set to \code{TRUE}, additional feature metadata is also returned.

\item \code{x[i}: subset on analyses, using analysis index or name.

\item \code{x[i}: subset on features, using feature index or name.

}}
\section{Slots}{

\describe{
\item{\code{analysis}}{A character string with the MS file name the without extension.}

\item{\code{file}}{A character string with the complete path of the MS file.}

\item{\code{replicate}}{A character string with the respective sample replicate group name.}

\item{\code{metadata}}{A list with other information queried from the MS files.
The info corresponds to the merged output of \link[mzR]{runInfo} and \link[mzR]{instrumentInfo}
from the \pkg{mzR} package.}

\item{\code{spectra}}{A \link[data.table]{data.table} obtained by
the \link[mzR]{peaks} function of the \pkg{mzR} package.
Although the function \link[mzR]{peaks} returns a \code{list},
the output is collapsed to a \link[data.table]{data.table}.}

\item{\code{chromatograms}}{A \link[data.table]{data.table} obtained by
the \link[mzR]{chromatogram} function of the \pkg{mzR} package.
Although the function \link[mzR]{chromatogram} returns a \code{list},
the output is collapsed to a \link[data.table]{data.table}.}

\item{\code{parameters}}{A list with parameter \linkS4class{msSettings} for data processing, such as peak picking.}

\item{\code{peaks}}{A \link[data.table]{data.table} with feature details, containing the following columns:
\enumerate{
 \item \strong{id}: character, the identifier of the peak;
 \item \strong{rt}: numeric, the calculated retention time (in seconds);
 \item \strong{mz}: numeric, the calculated \emph{m/z} value;
 \item \strong{intensity}: numeric, the calculated intensity or height (in counts);
 \item \strong{area}: numeric, the calculated area of the peak;
 \item \strong{rtmin}: numeric, the minimum retention time (in seconds);
 \item \strong{rtmax}: numeric, the maximum retention time, (in seconds);
 \item \strong{mzmin}: numeric, the minimum \emph{m/z} value;
 \item \strong{mzmax}: numeric, the maximum \emph{m/z} value;
 \item \strong{adduct}: character, the annotated adduct ion representation of the peak;
 \item \strong{isotope}: numeric, the annotated isotopic number;
 \item \strong{monoIsotope}: character, the identifier of the mono isotopic ion;
 \item \strong{scans}: numeric vector, the scans used for the feature integration;
 \item \strong{feature}: character, the identifier of the peaks group (i.e., feature) after
 alignment and grouping of corresponding peaks across samples;
 \item \strong{...} other columns added from various functions.
}}
}}

\references{
\insertRef{patroon01}{streamFind}

\insertRef{proteo01}{streamFind}

\insertRef{proteo02}{streamFind}

\insertRef{mzr01}{streamFind}

\insertRef{mzr02}{streamFind}

\insertRef{mzr03}{streamFind}

\insertRef{mzr04}{streamFind}
}
