% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_general_000_new_project.R
\name{newAnalysis}
\alias{newAnalysis}
\title{newAnalysis}
\usage{
newAnalysis(file = NA_character_)
}
\arguments{
\item{file}{A character vector with the full path of the file to create
the analysis object. Alternatively, a \link{data.frame} or
\link[data.table]{data.table} with the columns file, replicate and
blank corresponding to the file full path, analysis replicate name and
respective blank analysis replicate name. When a table is given with
replicate and blank, the arguments \emph{replicate} and  \emph{blank}
are ignored. For file, only formats \emph{mzML} and \emph{mzXML} are
currently supported.}

\item{replicate}{Optional character vector with the name of the
analysis replicate group. The default is \code{NULL} which attempts  to
retrieve the analysis replicate group using the name of the file.
The replicate name can always be edited using the class method.}

\item{blank}{Optional character vector with the name of the associate
blank analysis replicate group for the file. The default is \code{NULL}.}
}
\value{
An S4 class depending on the added file format.
For instance, an \linkS4class{msAnalysis} is returned for
\emph{mzML} and \emph{mzXML} files.  If more than one file is given as
\code{file}, a \linkS4class{streamSet} object is returned instead.
}
\description{
Creates a new \pkg{streamFind} analysis object.
}
\note{
The format of the file will dictate the resulting S4 class.
For instance, \emph{.mzML} or \emph{.mzXML} files will lead to the class
\linkS4class{msAnalysis}, which will contain the structure for
handling MS data.
}
