% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_002_parse_data_from_files.R
\name{parse_ms_spectra}
\alias{parse_ms_spectra}
\title{parse_ms_spectra}
\usage{
parse_ms_spectra(
  files = NA_character_,
  levels = c(1, 2),
  targets = NULL,
  preMZr = NULL,
  runParallel = FALSE
)
}
\arguments{
\item{files}{X.}

\item{levels}{X.}

\item{targets}{X.}

\item{preMZr}{X.}

\item{runParallel}{X.}
}
\description{
parse_ms_spectra
}
