% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_002_peakGrouping.R
\name{peakGrouping}
\alias{peakGrouping}
\title{peakGrouping}
\usage{
peakGrouping(object = NULL, settings = NULL)
}
\arguments{
\item{object}{A \linkS4class{msData} object with
\linkS4class{msAnalysis} objects containing peaks.}

\item{settings}{A \linkS4class{settings} object.}
}
\value{
An \linkS4class{msData} object containing a features S4 class.
}
\description{
Grouping and alignment of peaks across samples.
The peak grouping uses the \pkg{patRoon} package.
}
\details{
The following algorithms are possible: "xcms3", "xcms", "openms".
See ?\pkg{patRoon} for further information.
See \code{\link[patRoon]{groupFeatures}} for more information.
}
\note{
The call in the \linkS4class{settings} must be set to "peakGrouping".
}
\references{
\insertRef{patroon01}{streamFind}
}
\seealso{
\code{\link[patRoon]{groupFeatures}}
}
