% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_001_peakPicking.R
\name{peakPicking}
\alias{peakPicking}
\title{peakPicking}
\usage{
peakPicking(object = NULL, settings = NULL)
}
\arguments{
\item{object}{An \linkS4class{msData} or \linkS4class{msAnalysis} object.}

\item{settings}{A \linkS4class{settings} object.}
}
\value{
An \linkS4class{msAnalysis} or \linkS4class{msData} object
with peaks added to the respective slot.
}
\description{
Finds chromatographic peaks from MS data in an
\linkS4class{msAnalysis} or in MS analyses of a given \linkS4class{msData}.
The peak picking uses the \pkg{patRoon} package for data processing,
enabling the use of several algorithms (see details).
}
\details{
See the \pkg{patRoon}'s \code{\link[patRoon]{findFeatures}} function
or the \href{https://rickhelmus.github.io/patRoon/reference/findFeatures.html}
{reference guide} for more information. The following algorithms are available
 via \pkg{patRoon}: "xcms3", "xcms", "openms", "envipick", "sirius", "kpic2",
"safd". The \linkS4class{settings} and MS file format and mode depend on
the algorithm chosen.
}
\note{
The \linkS4class{settings} call must be set to "peakPicking".
}
\references{
\insertRef{patroon01}{streamFind}
}
\seealso{
\code{\link[patRoon]{findFeatures}}
}
