% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_001_peakPicking.R
\name{peakPicking}
\alias{peakPicking}
\title{peakPicking}
\usage{
peakPicking(object = NULL, algorithm = NA_character_, settings = NULL)
}
\arguments{
\item{object}{An \linkS4class{msData} or \linkS4class{msAnalysis} object.}

\item{algorithm}{A character string with the algorithm to use for peak picking.}

\item{settings}{List of parameter settings for the specified algorithm.
See \link[patRoon]{findFeatures} for more information.}
}
\value{
An \linkS4class{msData} object with peaks added per analyses.
}
\description{
Finds chromatographic peaks from centroided MS data in the
analyses of a given \linkS4class{msData} object.
The peak picking uses the \pkg{patRoon} package and
the following algorithms are available:
"xcms3", "xcms", "openms", "envipick", "sirius", "kpic2", "safd".
The parameters depend on the algorithm chosen.
See ?\pkg{patRoon} for further information.
}
\references{
\insertRef{patroon01}{streamFind}
}
