% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ms_plot_functions.R
\name{plot_features_interactive}
\alias{plot_features_interactive}
\title{plot_features_interactive}
\usage{
plot_features_interactive(
  eic = NULL,
  features = NULL,
  legendNames = NULL,
  colorBy = "targets",
  title = NULL
)
}
\arguments{
\item{eic}{A data table with the analysis, replicate,
id, rt, intensity and var (i.e., the plotting variables for each peak)
as columns.}

\item{features}{A data table with the individual peaks to plot.}

\item{legendNames}{A character vector with the same length as the unique ids
in the table given in \code{eic}.}

\item{colorBy}{A string (length 1). One of "analyses" (the default),
"levels", "targets" or "replicates". For "replicates", a column with
replicate names should be given.}

\item{title}{A character vector to be used as title.}
}
\value{
A chromatographic peak plot through \pkg{plotly}.
}
\description{
Plots chromatographic peaks with the package \pkg{plotly}.
}
