% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ms_plot_peaks.R
\name{plot_features_static}
\alias{plot_features_static}
\title{plot_features_static}
\usage{
plot_features_static(eic, peaks, title = NULL)
}
\arguments{
\item{eic}{A data table with the analysis, replicate,
id, rt, intensity and var (i.e., the plotting variables for each peak)
as columns.}

\item{peaks}{A data table with the individual peaks to plot.}

\item{title}{An optional character vector to be used as title.}
}
\value{
A plot of chromatographic peaks.
}
\description{
Static plot of chromatographic peaks using the \pkg{base} package.
}
