% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_S3.R
\name{plot_fold_change}
\alias{plot_fold_change}
\title{Generic \code{plot_fold_change}}
\usage{
plot_fold_change(x, ...)
}
\arguments{
\item{x}{An object from which the fold change will be plotted.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S3 generic to plot fold change from \code{x}.
}
