% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_S3.R
\name{plot_groups_overview}
\alias{plot_groups_overview}
\title{Generic \code{plot_groups_overview}}
\usage{
plot_groups_overview(x, ...)
}
\arguments{
\item{x}{An object from which the groups overview will be plotted.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S3 generic to plot groups overview from \code{x}.
}
