% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot_residuals}
\alias{plot_residuals}
\title{Generic \code{plot_residuals}}
\usage{
plot_residuals(x, ...)
}
\arguments{
\item{x}{An object from which the residuals will be plotted.}

\item{...}{Additional arguments passed to the method.}
}
\description{
S7 generic to plot residuals from \code{x}.
}
