% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_004_plot.R
\name{plot_spectra_interactive}
\alias{plot_spectra_interactive}
\title{plot_spectra_interactive}
\usage{
plot_spectra_interactive(spectra = NULL, colorBy = "analyses")
}
\arguments{
\item{spectra}{A \code{data.table} with the at least columns \emph{analysis}, \emph{level},
\emph{rt}, \emph{mz} and \emph{intensity}.}

\item{colorBy}{A string (length 1). One of "analyses" (the default),
"levels", "targets" or "replicates". For "replicates", a column with
replicate names should be given.}
}
\value{
A 3D interactive plot.
}
\description{
3D interactive plot for spectra using the \pkg{plotly} package.
}
