% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ms_plot_featurePeaks.R
\name{plotFeaturePeaks}
\alias{plotFeaturePeaks}
\title{plotFeaturePeaks}
\usage{
plotFeaturePeaks(
  object,
  analyses = NULL,
  targetsID = NULL,
  mz = NULL,
  ppm = 20,
  rt = NULL,
  sec = 30,
  legendNames = NULL,
  heights = c(0.6, 0.4)
)
}
\arguments{
\item{object}{An \linkS4class{msData} object.}

\item{analyses}{The index or name of the analyses.
The default is \code{NULL} and all analyses are used.}

\item{targetsID}{The identifier of the features of interest.
When not \code{NULL}, overwrites any given \code{mz} and \code{rt} value.}

\item{mz}{Optional target \emph{m/z} to find features using
the mass deviation specified by the \code{ppm} parameter.}

\item{ppm}{The mass deviation to extract the features
when \code{mz} is specified.}

\item{rt}{The retention time in minutes or seconds,
depending on the defined \code{rtUnit}, see below.
Only used when \code{mz} is specified.}

\item{rtWindow}{The time deviation to collect features.
The time unit is the defined by \code{rtUnit}.
A time interval can be given with a length 2 vector,
defining the minimum and maximum retention time.}

\item{rtUnit}{Possible entries are \code{min} or \code{sec}.
The default is \code{min}.}

\item{msLevel}{The MS level to extract the data.
For the moment, only 1 is possible.}

\item{names}{A character string with names for each feature given in \code{ID}.
Note that length should match between \code{names} and \code{ID}.
If length does not match \code{names} are not used.}
}
\value{
A double plot with peak chromatograms on the top part
and feature peak groups below.
}
\description{
Plots peaks for each feature in an \linkS4class{msData} object.
}
