% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_S3_ProcessingSettings.R
\name{ProcessingSettings}
\alias{ProcessingSettings}
\title{\strong{ProcessingSettings} S3 class constructor, methods and functions}
\usage{
ProcessingSettings(
  call = NA_character_,
  algorithm = NA_character_,
  parameters = NULL,
  version = NA_character_,
  software = NA_character_,
  developer = NA_character_,
  contact = NA_character_,
  link = NA_character_,
  doi = NA_character_
)
}
\arguments{
\item{call}{Character of length one with the name of the method where theprocessing settings are to be applied.}

\item{algorithm}{Character of length one with the name of the algorithm to be used.}

\item{parameters}{List with parameters specific for the method \code{call} and \code{algorithm}.}

\item{version}{Character of length one with the version of the processing settings.}

\item{software}{Character of length one with the name of the software or package.}

\item{developer}{Character of length one with the name of the developer/s.}

\item{contact}{Character of length one with the email of the developer.}

\item{link}{Character of length one with the a web link.}

\item{doi}{Character of length one with the DOI of algorithm.}
}
\value{
A ProcessingSettings S3 class
}
\description{
Creates a ProcessingSettings S3 class object. The ProcessingSettings are used in \pkg{StreamFind} for
processing data in a given data class method.
}
