% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_S3_ProcessingSettings.R
\name{ProcessingSettings}
\alias{ProcessingSettings}
\title{\strong{ProcessingSettings} S3 class constructor, methods and functions}
\usage{
ProcessingSettings(
  engine = class_missing,
  method = class_missing,
  algorithm = class_missing,
  parameters = class_missing,
  number_permitted = class_missing,
  version = class_missing,
  software = class_missing,
  developer = class_missing,
  contact = class_missing,
  link = class_missing,
  doi = class_missing
)
}
\arguments{
\item{engine}{Character of length one with the name of the engine where the processing settings are to be applied.}

\item{algorithm}{Character of length one with the name of the algorithm to be used.}

\item{parameters}{List with parameters specific for the method \code{call} and \code{algorithm}.}

\item{version}{Character of length one with the version of the processing settings.}

\item{software}{Character of length one with the name of the software or package.}

\item{developer}{Character of length one with the name of the developer/s.}

\item{contact}{Character of length one with the email of the developer.}

\item{link}{Character of length one with the a web link.}

\item{doi}{Character of length one with the DOI of algorithm.}

\item{call}{Character of length one with the name of the method where the processing settings are to be applied.}
}
\value{
A ProcessingSettings S3 class
}
\description{
Creates a ProcessingSettings S3 class object. The ProcessingSettings are used in \pkg{StreamFind} for
processing data in a given data class method.
}
