% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_S3_ProcessingSettings.R
\name{ProcessingSettings}
\alias{ProcessingSettings}
\alias{validate.ProcessingSettings}
\alias{asJSON.ProcessingSettings}
\alias{as.ProcessingSettings}
\title{\strong{ProcessingSettings} S3 class constructor, methods and functions}
\usage{
ProcessingSettings(
  call = NA_character_,
  algorithm = NA_character_,
  parameters = NULL
)

\method{validate}{ProcessingSettings}(x = NULL)

\method{asJSON}{ProcessingSettings}(x)

as.ProcessingSettings(value)
}
\arguments{
\item{call}{Character of length one with the name of the method where the
processing settings are to be applied.}

\item{algorithm}{Character of length one with the name of the algorithm to
be used.}

\item{parameters}{List with parameters specific for the method \code{call} and
\code{algorithm}.}

\item{x}{A ProcessingSettings S3 class object.}

\item{value}{A list to be checked and/or converted to ProcessingSettings S3
class.}
}
\value{
A ProcessingSettings S3 class
}
\description{
Creates a ProcessingSettings S3 class object. The ProcessingSettings are used
in \pkg{streamFind} for processing data in a given data class method.
}
\details{
See the documentation of the method where the processing settings
are to be applied for more information about applicable algorithms and
parameters.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate(ProcessingSettings)}: Validates a ProcessingSettings S3 class object, returning a logical value of
length one.

\item \code{asJSON(ProcessingSettings)}: Converts a ProcessingSettings S3 class object to a JSON string.

}}
\section{Functions}{
\itemize{
\item \code{as.ProcessingSettings()}: Converts the argument in a ProcessingSettings S3 class object.

}}
