% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ProcessingStep.R
\name{ProcessingStep}
\alias{ProcessingStep}
\title{Generic Processing Step}
\usage{
ProcessingStep(
  data_type = class_missing,
  method = class_missing,
  required = class_missing,
  algorithm = class_missing,
  parameters = class_missing,
  number_permitted = class_missing,
  version = class_missing,
  software = class_missing,
  developer = class_missing,
  contact = class_missing,
  link = class_missing,
  doi = class_missing
)
}
\arguments{
\item{data_type}{A character string representing the data type (e.g., "MassSpec", "Raman").}

\item{method}{A character string representing the method used (e.g., "BaselineCorrection").}

\item{required}{A character vector of required preceding methods.}

\item{algorithm}{A character string representing the algorithm used (e.g., "baseline_als").}

\item{parameters}{A list of parameters for the processing step.}

\item{number_permitted}{A numeric value indicating the number of permitted instances.}

\item{version}{A character string representing the version of the processing step.}

\item{software}{A character string representing the original software.}

\item{developer}{A character string representing the developer of the processing step.}

\item{contact}{A character string representing the contact information for the developer.}

\item{link}{A character string representing the link to the origin of the algorithm or link to
additional information.}

\item{doi}{A character string representing the DOI of the algorithm or additional information.}
}
\description{
The \code{ProcessingStep} class is used to define a processing step within a
\link{Workflow}. It contains information about the data type, method, algorithm,
parameters, and other relevant metadata for the processing step. The \code{ProcessingStep} is the
parent class of all processing methods in StreamFind.
}
\section{Slots}{

\describe{
\item{\code{call}}{(getter) A character string representing the call to the processing step.}
}}

