% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_lv1_S7_ProjectHeaders.R
\name{ProjectHeaders}
\alias{ProjectHeaders}
\title{\strong{ProjectHeaders} S7 class constructor, slots and methods}
\usage{
ProjectHeaders(...)
}
\arguments{
\item{...}{Arguments to be added as headers. Note that all given arguments must be of length one. The name of the
argument becomes the name of the header entry. Alternatively, a single list argument with the headers (all length
one) can be given. If an argument or element name is given, it must be type character. If an argument or element
date is given, it must be class \code{POSIXct} or \code{POSIXt}. If given date is character, conversion to class \code{POSIXct}
or \code{POSIXt} is attempted.}
}
\value{
A ProjectHeaders S7 class object.
}
\description{
Creates a ProjectHeaders S7 class object.
}
\section{Slots}{

\describe{
\item{\code{headers}}{\code{List} of headers.}

\item{\code{names}}{Dynamic \code{Character} returning the header names.}

\item{\code{length}}{Dynamic \code{Numeric} returning the number of headers.}
}}

\section{Methods}{

\itemize{
\item \code{[}, \code{[[}, \code{$}, \verb{[<-}, \verb{[[<-}, \verb{$<-} are implemented to get and set the elements from the headers list.
\item \code{length} returns the number of headers.
\item \code{names} returns the header names.
\item \code{show} prints the headers.
}
}

