% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ProjectMassSpecChromatograms.R
\name{ProjectMassSpecChromatograms}
\alias{ProjectMassSpecChromatograms}
\title{Project Mass Spec Chromatograms R6 Class}
\description{
R6 child of \code{ProjectMassSpec} exposing the chromatogram-focused
Mass Spec interface backed by the native C++ chromatograms facade.
}
\details{
This is a public user-facing project class.

In addition to the methods documented on this page, it inherits shared project
runtime methods such as \code{run_app()}, \code{run_workflow()}, \code{report_quarto()},
\code{metadata}, \code{workflow}, \code{get_audit()}, and \code{list_tables()}, plus shared Mass
Spec methods such as \code{import_files()}, \code{list_analyses()}, \code{get_analysis_names()},
\code{get_replicate_names()}, \code{set_replicate_names()}, \code{get_blank_names()},
\code{set_blank_names()}, and \code{get_concentrations()}.

See \code{?Project} for shared project/runtime methods and \code{?ProjectMassSpec} for
inherited shared Mass Spec methods.

Use \code{?ProjectMassSpecChromatograms} as the main entry point for the
chromatogram-specific interface on top of those base classes.
}
\section{Super classes}{
\code{\link[StreamFind:Project]{StreamFind::Project}} -> \code{\link[StreamFind:ProjectMassSpec]{StreamFind::ProjectMassSpec}} -> \code{ProjectMassSpecChromatograms}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ProjectMassSpecChromatograms-new}{\code{ProjectMassSpecChromatograms$new()}}
\item \href{#method-ProjectMassSpecChromatograms-get_mass_spec_chromatograms_ptr}{\code{ProjectMassSpecChromatograms$get_mass_spec_chromatograms_ptr()}}
\item \href{#method-ProjectMassSpecChromatograms-available_processing_methods}{\code{ProjectMassSpecChromatograms$available_processing_methods()}}
\item \href{#method-ProjectMassSpecChromatograms-get_chromatograms_headers}{\code{ProjectMassSpecChromatograms$get_chromatograms_headers()}}
\item \href{#method-ProjectMassSpecChromatograms-get_chromatograms}{\code{ProjectMassSpecChromatograms$get_chromatograms()}}
\item \href{#method-ProjectMassSpecChromatograms-plot_chromatograms}{\code{ProjectMassSpecChromatograms$plot_chromatograms()}}
\item \href{#method-ProjectMassSpecChromatograms-print}{\code{ProjectMassSpecChromatograms$print()}}
\item \href{#method-ProjectMassSpecChromatograms-show}{\code{ProjectMassSpecChromatograms$show()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="copy"><a href='../../StreamFind/html/Project.html#method-Project-copy'><code>StreamFind::Project$copy()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="get_audit"><a href='../../StreamFind/html/Project.html#method-Project-get_audit'><code>StreamFind::Project$get_audit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="get_domain"><a href='../../StreamFind/html/Project.html#method-Project-get_domain'><code>StreamFind::Project$get_domain()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="get_metadata"><a href='../../StreamFind/html/Project.html#method-Project-get_metadata'><code>StreamFind::Project$get_metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="get_ptr"><a href='../../StreamFind/html/Project.html#method-Project-get_ptr'><code>StreamFind::Project$get_ptr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="get_workflow"><a href='../../StreamFind/html/Project.html#method-Project-get_workflow'><code>StreamFind::Project$get_workflow()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="list_tables"><a href='../../StreamFind/html/Project.html#method-Project-list_tables'><code>StreamFind::Project$list_tables()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="report_quarto"><a href='../../StreamFind/html/Project.html#method-Project-report_quarto'><code>StreamFind::Project$report_quarto()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="run_app"><a href='../../StreamFind/html/Project.html#method-Project-run_app'><code>StreamFind::Project$run_app()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="run_processing_step"><a href='../../StreamFind/html/Project.html#method-Project-run_processing_step'><code>StreamFind::Project$run_processing_step()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="run_workflow"><a href='../../StreamFind/html/Project.html#method-Project-run_workflow'><code>StreamFind::Project$run_workflow()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="set_metadata"><a href='../../StreamFind/html/Project.html#method-Project-set_metadata'><code>StreamFind::Project$set_metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="set_workflow"><a href='../../StreamFind/html/Project.html#method-Project-set_workflow'><code>StreamFind::Project$set_workflow()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="validate"><a href='../../StreamFind/html/Project.html#method-Project-validate'><code>StreamFind::Project$validate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_analysis_names"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_analysis_names'><code>StreamFind::ProjectMassSpec$get_analysis_names()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_blank_names"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_blank_names'><code>StreamFind::ProjectMassSpec$get_blank_names()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_concentrations"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_concentrations'><code>StreamFind::ProjectMassSpec$get_concentrations()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_mass_spec_ptr"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_mass_spec_ptr'><code>StreamFind::ProjectMassSpec$get_mass_spec_ptr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_raw_spectra"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_raw_spectra'><code>StreamFind::ProjectMassSpec$get_raw_spectra()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_raw_spectra_bpc"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_raw_spectra_bpc'><code>StreamFind::ProjectMassSpec$get_raw_spectra_bpc()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_raw_spectra_eic"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_raw_spectra_eic'><code>StreamFind::ProjectMassSpec$get_raw_spectra_eic()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_raw_spectra_ms1"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_raw_spectra_ms1'><code>StreamFind::ProjectMassSpec$get_raw_spectra_ms1()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_raw_spectra_ms2"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_raw_spectra_ms2'><code>StreamFind::ProjectMassSpec$get_raw_spectra_ms2()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_raw_spectra_tic"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_raw_spectra_tic'><code>StreamFind::ProjectMassSpec$get_raw_spectra_tic()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_replicate_names"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_replicate_names'><code>StreamFind::ProjectMassSpec$get_replicate_names()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_spectra_headers"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_spectra_headers'><code>StreamFind::ProjectMassSpec$get_spectra_headers()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_spectra_tic"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_spectra_tic'><code>StreamFind::ProjectMassSpec$get_spectra_tic()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="import_files"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-import_files'><code>StreamFind::ProjectMassSpec$import_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="list_analyses"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-list_analyses'><code>StreamFind::ProjectMassSpec$list_analyses()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="plot_spectra_bpc"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-plot_spectra_bpc'><code>StreamFind::ProjectMassSpec$plot_spectra_bpc()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="plot_spectra_eic"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-plot_spectra_eic'><code>StreamFind::ProjectMassSpec$plot_spectra_eic()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="plot_spectra_tic"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-plot_spectra_tic'><code>StreamFind::ProjectMassSpec$plot_spectra_tic()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="remove_analysis"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-remove_analysis'><code>StreamFind::ProjectMassSpec$remove_analysis()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="set_blank_names"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-set_blank_names'><code>StreamFind::ProjectMassSpec$set_blank_names()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="set_concentrations"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-set_concentrations'><code>StreamFind::ProjectMassSpec$set_concentrations()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="set_replicate_names"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-set_replicate_names'><code>StreamFind::ProjectMassSpec$set_replicate_names()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectMassSpecChromatograms-new"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectMassSpecChromatograms-new}{}}}
\subsection{Method \code{new()}}{
Create a chromatogram-specific Mass Spec project wrapper.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectMassSpecChromatograms$new(
  db,
  project_id,
  .ptr = NULL,
  .mass_spec_ptr = NULL,
  .mass_spec_chromatograms_ptr = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{db}}{Path to the DuckDB project file.}

\item{\code{project_id}}{Active project identifier.}

\item{\code{.ptr}}{Existing native project pointer for internal use.}

\item{\code{.mass_spec_ptr}}{Existing native shared Mass Spec pointer for internal use.}

\item{\code{.mass_spec_chromatograms_ptr}}{Existing native chromatograms pointer for internal use.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectMassSpecChromatograms-get_mass_spec_chromatograms_ptr"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectMassSpecChromatograms-get_mass_spec_chromatograms_ptr}{}}}
\subsection{Method \code{get_mass_spec_chromatograms_ptr()}}{
Return the native chromatograms pointer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectMassSpecChromatograms$get_mass_spec_chromatograms_ptr()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectMassSpecChromatograms-available_processing_methods"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectMassSpecChromatograms-available_processing_methods}{}}}
\subsection{Method \code{available_processing_methods()}}{
Return chromatogram-project processing-step metadata.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectMassSpecChromatograms$available_processing_methods()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named list of \code{ProcessingStep} metadata objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectMassSpecChromatograms-get_chromatograms_headers"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectMassSpecChromatograms-get_chromatograms_headers}{}}}
\subsection{Method \code{get_chromatograms_headers()}}{
Return chromatogram headers for selected analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectMassSpecChromatograms$get_chromatograms_headers(analyses = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{Character or numeric vector with names or indexes of analyses in the \code{Analyses}
object.}

\item{\code{analyses}}{Character or numeric vector with names or indexes of analyses in the \code{Analyses}
object.}

\item{\code{analyses}}{Character or numeric vector with names or indexes of analyses in the \code{Analyses}
object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectMassSpecChromatograms-get_chromatograms"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectMassSpecChromatograms-get_chromatograms}{}}}
\subsection{Method \code{get_chromatograms()}}{
Get chromatograms for selected analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectMassSpecChromatograms$get_chromatograms(
  analyses = NULL,
  chromatograms = NULL,
  rtmin = 0,
  rtmax = 0,
  minIntensity = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{Character or numeric vector with names or indexes of analyses in the \code{Analyses}
object.}

\item{\code{analyses}}{Character or numeric vector with names or indexes of analyses in the \code{Analyses}
object.}

\item{\code{analyses}}{Character or numeric vector with names or indexes of analyses in the \code{Analyses}
object.}

\item{\code{chromatograms}}{A character or integer vector with the ID (i.e. name) or the index of the chromatograms.}

\item{\code{chromatograms}}{A character or integer vector with the ID (i.e. name) or the index of the chromatograms.}

\item{\code{rtmin}}{Numeric (length 1) with the minimum retention time.}

\item{\code{rtmin}}{Numeric (length 1) with the minimum retention time.}

\item{\code{rtmax}}{Numeric (length 1) with the maximum retention time.}

\item{\code{rtmax}}{Numeric (length 1) with the maximum retention time.}

\item{\code{minIntensity}}{Numeric (length 1) with the minimum intensity.}

\item{\code{minIntensity}}{Numeric (length 1) with the minimum intensity.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectMassSpecChromatograms-plot_chromatograms"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectMassSpecChromatograms-plot_chromatograms}{}}}
\subsection{Method \code{plot_chromatograms()}}{
Plot chromatograms for selected analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectMassSpecChromatograms$plot_chromatograms(
  analyses = NULL,
  chromatograms = NULL,
  rtmin = 0,
  rtmax = 0,
  minIntensity = NULL,
  downsize = NULL,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  groupBy = "analysis",
  interactive = TRUE,
  colorPalette = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{Character or numeric vector with names or indexes of analyses in the \code{Analyses}
object.}

\item{\code{analyses}}{Character or numeric vector with names or indexes of analyses in the \code{Analyses}
object.}

\item{\code{analyses}}{Character or numeric vector with names or indexes of analyses in the \code{Analyses}
object.}

\item{\code{chromatograms}}{A character or integer vector with the ID (i.e. name) or the index of the chromatograms.}

\item{\code{chromatograms}}{A character or integer vector with the ID (i.e. name) or the index of the chromatograms.}

\item{\code{rtmin}}{Numeric (length 1) with the minimum retention time.}

\item{\code{rtmin}}{Numeric (length 1) with the minimum retention time.}

\item{\code{rtmax}}{Numeric (length 1) with the maximum retention time.}

\item{\code{rtmax}}{Numeric (length 1) with the maximum retention time.}

\item{\code{minIntensity}}{Numeric (length 1) with the minimum intensity.}

\item{\code{minIntensity}}{Numeric (length 1) with the minimum intensity.}

\item{\code{downsize}}{An integer of length one to downsize the TIC plot. The default is 1.}

\item{\code{xLab}}{X axis label for the plot.}

\item{\code{yLab}}{Y axis label for the plot.}

\item{\code{title}}{Plot title.}

\item{\code{groupBy}}{Column to use for color grouping (default: "analysis").}

\item{\code{interactive}}{Logical, use plotly if TRUE, ggplot2 if FALSE.}

\item{\code{colorPalette}}{Optional vector of colors for the plot; if not provided, an internal function is used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectMassSpecChromatograms-print"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectMassSpecChromatograms-print}{}}}
\subsection{Method \code{print()}}{
Print a short summary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectMassSpecChromatograms$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectMassSpecChromatograms-show"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectMassSpecChromatograms-show}{}}}
\subsection{Method \code{show()}}{
Show a short summary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectMassSpecChromatograms$show(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments ignored.}
}
\if{html}{\out{</div>}}
}
}
}
