% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ProjectMassSpecChromatograms.R
\name{ProjectMassSpecChromatograms}
\alias{ProjectMassSpecChromatograms}
\title{Project Mass Spec Chromatograms R6 Class}
\description{
R6 child of \code{ProjectMassSpec} exposing the chromatogram-focused MassSpec interface.
}
\keyword{internal}
\section{Super classes}{
\code{\link[StreamFind:Project]{StreamFind::Project}} -> \code{\link[StreamFind:ProjectMassSpec]{StreamFind::ProjectMassSpec}} -> \code{ProjectMassSpecChromatograms}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ProjectMassSpecChromatograms-new}{\code{ProjectMassSpecChromatograms$new()}}
\item \href{#method-ProjectMassSpecChromatograms-get_mass_spec_chromatograms_ptr}{\code{ProjectMassSpecChromatograms$get_mass_spec_chromatograms_ptr()}}
\item \href{#method-ProjectMassSpecChromatograms-available_processing_methods}{\code{ProjectMassSpecChromatograms$available_processing_methods()}}
\item \href{#method-ProjectMassSpecChromatograms-get_chromatograms_headers}{\code{ProjectMassSpecChromatograms$get_chromatograms_headers()}}
\item \href{#method-ProjectMassSpecChromatograms-print}{\code{ProjectMassSpecChromatograms$print()}}
\item \href{#method-ProjectMassSpecChromatograms-show}{\code{ProjectMassSpecChromatograms$show()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="copy"><a href='../../StreamFind/html/Project.html#method-Project-copy'><code>StreamFind::Project$copy()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="delete_cache"><a href='../../StreamFind/html/Project.html#method-Project-delete_cache'><code>StreamFind::Project$delete_cache()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="get_audit"><a href='../../StreamFind/html/Project.html#method-Project-get_audit'><code>StreamFind::Project$get_audit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="get_cache"><a href='../../StreamFind/html/Project.html#method-Project-get_cache'><code>StreamFind::Project$get_cache()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="get_cache_size"><a href='../../StreamFind/html/Project.html#method-Project-get_cache_size'><code>StreamFind::Project$get_cache_size()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="get_db"><a href='../../StreamFind/html/Project.html#method-Project-get_db'><code>StreamFind::Project$get_db()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="get_domain"><a href='../../StreamFind/html/Project.html#method-Project-get_domain'><code>StreamFind::Project$get_domain()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="get_metadata"><a href='../../StreamFind/html/Project.html#method-Project-get_metadata'><code>StreamFind::Project$get_metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="get_project_id"><a href='../../StreamFind/html/Project.html#method-Project-get_project_id'><code>StreamFind::Project$get_project_id()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="get_ptr"><a href='../../StreamFind/html/Project.html#method-Project-get_ptr'><code>StreamFind::Project$get_ptr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="get_workflow"><a href='../../StreamFind/html/Project.html#method-Project-get_workflow'><code>StreamFind::Project$get_workflow()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="list_tables"><a href='../../StreamFind/html/Project.html#method-Project-list_tables'><code>StreamFind::Project$list_tables()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="report_quarto"><a href='../../StreamFind/html/Project.html#method-Project-report_quarto'><code>StreamFind::Project$report_quarto()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="run_app"><a href='../../StreamFind/html/Project.html#method-Project-run_app'><code>StreamFind::Project$run_app()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="run_method"><a href='../../StreamFind/html/Project.html#method-Project-run_method'><code>StreamFind::Project$run_method()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="run_workflow"><a href='../../StreamFind/html/Project.html#method-Project-run_workflow'><code>StreamFind::Project$run_workflow()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="set_metadata"><a href='../../StreamFind/html/Project.html#method-Project-set_metadata'><code>StreamFind::Project$set_metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="set_workflow"><a href='../../StreamFind/html/Project.html#method-Project-set_workflow'><code>StreamFind::Project$set_workflow()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="validate"><a href='../../StreamFind/html/Project.html#method-Project-validate'><code>StreamFind::Project$validate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="add_analyses"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-add_analyses'><code>StreamFind::ProjectMassSpec$add_analyses()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_analyses"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_analyses'><code>StreamFind::ProjectMassSpec$get_analyses()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_analysis_names"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_analysis_names'><code>StreamFind::ProjectMassSpec$get_analysis_names()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_blank_names"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_blank_names'><code>StreamFind::ProjectMassSpec$get_blank_names()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_concentrations"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_concentrations'><code>StreamFind::ProjectMassSpec$get_concentrations()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_mass_spec_ptr"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_mass_spec_ptr'><code>StreamFind::ProjectMassSpec$get_mass_spec_ptr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_raw_chromatograms"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_raw_chromatograms'><code>StreamFind::ProjectMassSpec$get_raw_chromatograms()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_raw_spectra"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_raw_spectra'><code>StreamFind::ProjectMassSpec$get_raw_spectra()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_raw_spectra_eic"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_raw_spectra_eic'><code>StreamFind::ProjectMassSpec$get_raw_spectra_eic()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_raw_spectra_ms1"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_raw_spectra_ms1'><code>StreamFind::ProjectMassSpec$get_raw_spectra_ms1()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_raw_spectra_ms2"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_raw_spectra_ms2'><code>StreamFind::ProjectMassSpec$get_raw_spectra_ms2()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_replicate_names"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_replicate_names'><code>StreamFind::ProjectMassSpec$get_replicate_names()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_spectra_headers"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_spectra_headers'><code>StreamFind::ProjectMassSpec$get_spectra_headers()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_spectra_tic"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_spectra_tic'><code>StreamFind::ProjectMassSpec$get_spectra_tic()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="plot_raw_chromatograms"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-plot_raw_chromatograms'><code>StreamFind::ProjectMassSpec$plot_raw_chromatograms()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="plot_raw_spectra_eic"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-plot_raw_spectra_eic'><code>StreamFind::ProjectMassSpec$plot_raw_spectra_eic()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="plot_raw_spectra_ms1"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-plot_raw_spectra_ms1'><code>StreamFind::ProjectMassSpec$plot_raw_spectra_ms1()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="plot_raw_spectra_ms2"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-plot_raw_spectra_ms2'><code>StreamFind::ProjectMassSpec$plot_raw_spectra_ms2()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="plot_spectra_bpc"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-plot_spectra_bpc'><code>StreamFind::ProjectMassSpec$plot_spectra_bpc()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="plot_spectra_tic"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-plot_spectra_tic'><code>StreamFind::ProjectMassSpec$plot_spectra_tic()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="remove_analyses"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-remove_analyses'><code>StreamFind::ProjectMassSpec$remove_analyses()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="set_blank_names"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-set_blank_names'><code>StreamFind::ProjectMassSpec$set_blank_names()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="set_concentrations"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-set_concentrations'><code>StreamFind::ProjectMassSpec$set_concentrations()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="set_replicate_names"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-set_replicate_names'><code>StreamFind::ProjectMassSpec$set_replicate_names()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectMassSpecChromatograms-new"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectMassSpecChromatograms-new}{}}}
\subsection{Method \code{new()}}{
Create a chromatogram-specific Mass Spec project wrapper.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectMassSpecChromatograms$new(
  db,
  project_id,
  ...,
  file_paths = character(),
  replicates = character(),
  blanks = character()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{db}}{Path to DuckDB file.}

\item{\code{project_id}}{Active project identifier.}

\item{\code{...}}{Additional arguments passed to the method or ignored by the implementation.}

\item{\code{file_paths}}{Character vector with Mass Spec file paths.}

\item{\code{replicates}}{Optional character vector with replicate names.}

\item{\code{blanks}}{Optional character vector with blank names.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectMassSpecChromatograms-get_mass_spec_chromatograms_ptr"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectMassSpecChromatograms-get_mass_spec_chromatograms_ptr}{}}}
\subsection{Method \code{get_mass_spec_chromatograms_ptr()}}{
Return the native chromatograms pointer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectMassSpecChromatograms$get_mass_spec_chromatograms_ptr()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectMassSpecChromatograms-available_processing_methods"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectMassSpecChromatograms-available_processing_methods}{}}}
\subsection{Method \code{available_processing_methods()}}{
Return chromatogram-project method metadata.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectMassSpecChromatograms$available_processing_methods()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectMassSpecChromatograms-get_chromatograms_headers"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectMassSpecChromatograms-get_chromatograms_headers}{}}}
\subsection{Method \code{get_chromatograms_headers()}}{
Return chromatogram headers for selected analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectMassSpecChromatograms$get_chromatograms_headers(analyses = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{Character or numeric vector with names or indexes of analyses in the \code{Analyses}
object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectMassSpecChromatograms-print"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectMassSpecChromatograms-print}{}}}
\subsection{Method \code{print()}}{
Print a short summary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectMassSpecChromatograms$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments passed to the method or ignored by the implementation.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectMassSpecChromatograms-show"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectMassSpecChromatograms-show}{}}}
\subsection{Method \code{show()}}{
Show a short summary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectMassSpecChromatograms$show(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments passed to the method or ignored by the implementation.}
}
\if{html}{\out{</div>}}
}
}
}
