% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ProjectMassSpec.R
\name{ProjectMassSpecS3}
\alias{ProjectMassSpecS3}
\alias{get_raw_spectra_tic.ProjectMassSpec}
\alias{get_raw_spectra_bpc.ProjectMassSpec}
\alias{plot_spectra_tic.ProjectMassSpec}
\alias{plot_spectra_bpc.ProjectMassSpec}
\alias{get_spectra_headers.ProjectMassSpec}
\alias{get_analysis_names.ProjectMassSpec}
\alias{get_replicate_names.ProjectMassSpec}
\alias{set_replicate_names.ProjectMassSpec}
\alias{get_blank_names.ProjectMassSpec}
\alias{set_blank_names.ProjectMassSpec}
\alias{get_concentrations.ProjectMassSpec}
\alias{set_concentrations.ProjectMassSpec}
\alias{get_raw_spectra.ProjectMassSpec}
\alias{get_raw_spectra_eic.ProjectMassSpec}
\alias{plot_spectra_eic.ProjectMassSpec}
\alias{get_raw_spectra_ms1.ProjectMassSpec}
\alias{get_raw_spectra_ms2.ProjectMassSpec}
\alias{get_chromatograms_headers.ProjectMassSpec}
\alias{get_chromatograms.ProjectMassSpec}
\alias{plot_chromatograms.ProjectMassSpec}
\title{ProjectMassSpec S3 Methods}
\usage{
\method{get_raw_spectra_tic}{ProjectMassSpec}(x, analyses = NULL, levels = c(1, 2), rt = NULL)

\method{get_raw_spectra_bpc}{ProjectMassSpec}(x, analyses = NULL, levels = c(1, 2), rt = NULL)

\method{plot_spectra_tic}{ProjectMassSpec}(
  x,
  analyses = NULL,
  levels = c(1, 2),
  rt = NULL,
  downsize = NULL,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  groupBy = "analysis",
  interactive = TRUE,
  colorPalette = NULL
)

\method{plot_spectra_bpc}{ProjectMassSpec}(
  x,
  analyses = NULL,
  levels = c(1, 2),
  rt = NULL,
  downsize = NULL,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  groupBy = "analysis",
  interactive = TRUE,
  colorPalette = NULL
)

\method{get_spectra_headers}{ProjectMassSpec}(x, analyses = NULL)

\method{get_analysis_names}{ProjectMassSpec}(x)

\method{get_replicate_names}{ProjectMassSpec}(x)

\method{set_replicate_names}{ProjectMassSpec}(x, value)

\method{get_blank_names}{ProjectMassSpec}(x)

\method{set_blank_names}{ProjectMassSpec}(x, value)

\method{get_concentrations}{ProjectMassSpec}(x)

\method{set_concentrations}{ProjectMassSpec}(x, value)

\method{get_raw_spectra}{ProjectMassSpec}(
  x,
  analyses = NULL,
  levels = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  id = NULL,
  allTraces = TRUE,
  isolationWindow = 1.3,
  minIntensityMS1 = 0,
  minIntensityMS2 = 0
)

\method{get_raw_spectra_eic}{ProjectMassSpec}(
  x,
  analyses = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  id = NULL
)

\method{plot_spectra_eic}{ProjectMassSpec}(
  x,
  analyses = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  id = NULL,
  downsize = NULL,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  groupBy = c("analysis", "id"),
  interactive = TRUE,
  colorPalette = NULL
)

\method{get_raw_spectra_ms1}{ProjectMassSpec}(
  x,
  analyses = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  id = NULL,
  mzClust = 0.003,
  presence = 0.8,
  minIntensity = 1000
)

\method{get_raw_spectra_ms2}{ProjectMassSpec}(
  x,
  analyses = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  id = NULL,
  isolationWindow = 1.3,
  mzClust = 0.005,
  presence = 0.8,
  minIntensity = 0
)

\method{get_chromatograms_headers}{ProjectMassSpec}(x, analyses = NULL)

\method{get_chromatograms}{ProjectMassSpec}(
  x,
  analyses = NULL,
  chromatograms = NULL,
  rtmin = 0,
  rtmax = 0,
  minIntensity = NULL
)

\method{plot_chromatograms}{ProjectMassSpec}(
  x,
  analyses = NULL,
  chromatograms = NULL,
  rtmin = 0,
  rtmax = 0,
  minIntensity = NULL,
  downsize = NULL,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  groupBy = "analysis",
  interactive = TRUE,
  colorPalette = NULL
)
}
\arguments{
\item{x}{A \code{ProjectMassSpec} object.}

\item{analyses}{Character or numeric vector with names or indexes of analyses in the \code{Analyses}
object.}

\item{levels}{Integer vector with the levels of MS data.}

\item{rt}{A vector with target retention time values (in seconds) or
a two columns data.table or data.frame with minimum and maximum retention
time values (in seconds).}

\item{downsize}{An integer of length one to downsize the TIC plot. The default is 1.}

\item{xLab}{X axis label for the plot.}

\item{yLab}{Y axis label for the plot.}

\item{title}{Plot title.}

\item{groupBy}{Column to use for color grouping (default: "analysis").}

\item{interactive}{Logical, use plotly if TRUE, ggplot2 if FALSE.}

\item{colorPalette}{Optional vector of colors for the plot; if not provided, an internal function is used.}

\item{value}{Numeric vector of concentrations matching the number of analyses.}

\item{mass}{A vector with target neutral mass value/s (in Da) or a two columns data.table or data.frame named \code{min}
and \code{max} with minimum and maximum neutral mass values (in Da), respectively. Alternatively, neutral mass (in Da)
and retention time (in seconds) and/or drift time values (in milliseconds) can be given as one data.table or
data.frame with columns named \code{mass} and \code{rt} and/or \code{drift}. Then, the deviations given in the \code{ppm}, \code{sec}
and \code{millisec}  arguments are used to calculate the ranges. Also works with a data.table or data.frame with
minimum and maximum values of neutral mass, retention time and drift time targets. In this case, the column names
must be \code{min}, \code{max}, \code{rtmin}, \code{rtmax}, \code{driftmin} and \code{driftmax}. Note that when mass/time ranges are given, the
\code{ppm}, \code{sec} and \code{millisec}
arguments are not used.}

\item{mz}{A vector with target \emph{m/z} value/s (in Da) or a two columns data.table or data.frame named \code{mzmin} and
\code{mzmax} with minimum and maximum \emph{m/z} values (in Da), respectively. Alternatively, \emph{m/z} (in Da) and
retention time values (in seconds) can be given as one data.table or data.frame with columns named \code{mz} and \code{rt}
and/or \code{drift}. Then, the deviations given in the \code{ppm}, \code{sec} and \code{millisec} arguments are used to
calculate the ranges. Also works with a data.table or data.frame with minimum and maximum values of \emph{m/z},
retention time and drift time targets. In this case, the column names must be \code{mzmin}, \code{mzmax}, \code{rtmin}, \code{rtmax},
\code{driftmin} and \code{driftmax}. Note that when mass/time ranges are given, the \code{ppm}, \code{sec} and \code{millisec}
arguments are not used.}

\item{mobility}{A vector with target drift time values (in milliseconds) or
a two columns data.table or data.frame with minimum and maximum drift
time values (in milliseconds).}

\item{ppm}{Numeric of length one with the mass deviation, in ppm.}

\item{sec}{Numeric of length one with the time deviation, in seconds.}

\item{millisec}{Numeric of length one with the drift time deviation,
in milliseconds.}

\item{id}{Character with the same length as \emph{m/z} and retention time
targets to be used as identifier/s. When not given, the id is built as a
combination of the \emph{m/z} and retention time ranges or values.}

\item{allTraces}{Logical, when \code{TRUE} all level 2 data is returned.
When \code{FALSE} and level has 2, only the MS2 traces of MS1 targets
are returned, using the \code{isolationWindow} to calculate the mass window.}

\item{isolationWindow}{Numeric value with the isolation window (in Da)
applied for ion isolation and fragmentation during acquisition of tandem
data (i.e., MS2 data).}

\item{minIntensityMS1}{Numeric value with the minimum intensity of level 1
data (i.e., MS1 data).}

\item{minIntensityMS2}{Numeric value with the minimum intensity of level 2
data (i.e., MS2 data).}

\item{mzClust}{Numeric (length 1) with the mass deviation threshold (in Da) to cluster mass traces. Highly dependent
on the mass resolution of the MS data.}

\item{presence}{Numeric (length 1) with the required presence ratio from 0 (i.e., doesn't need to be present in any
spectra) to 1 (i.e., must be present in all spectra) for traces during clustering of spectra.}

\item{minIntensity}{Numeric (length 1) with the minimum intensity.}

\item{chromatograms}{A character or integer vector with the ID (i.e. name) or the index of the chromatograms.}

\item{rtmin}{Numeric (length 1) with the minimum retention time.}

\item{rtmax}{Numeric (length 1) with the maximum retention time.}
}
\description{
S3 interface methods for \code{ProjectMassSpec}.
These methods are thin wrappers over the \code{ProjectMassSpec} R6 methods and expose
the shared-project Mass Spec API through the package generics.
}
\details{
Available methods are:

Metadata methods:
\itemize{
\item \code{get_analysis_names()}: Get analysis names.
\item \code{get_replicate_names()}: Get replicate names.
\item \code{set_replicate_names()}: Set replicate names.
\item \code{get_blank_names()}: Get blank names.
\item \code{set_blank_names()}: Set blank names.
\item \code{get_concentrations()}: Get concentrations.
\item \code{set_concentrations()}: Set concentrations.
}

Header methods:
\itemize{
\item \code{get_spectra_headers()}: Fetch spectra headers for the specified analyses.
\item \code{get_chromatograms_headers()}: Fetch chromatograms headers for the specified analyses.
\item \code{get_spectra_tic()}: Fetch spectra TIC rows for the specified analyses.
}

Spectra trace methods:
\itemize{
\item \code{get_raw_spectra_tic()}: Get the total ion current (TIC) spectra for the specified analyses.
\item \code{get_raw_spectra_bpc()}: Get the base peak chromatograms (BPC) spectra for the specified analyses.
\item \code{plot_spectra_tic()}: Plot total ion current (TIC) spectra for the specified analyses.
\item \code{plot_spectra_bpc()}: Plot base peak chromatogram (BPC) spectra for the specified analyses.
}

Spectra extraction methods:
\itemize{
\item \code{get_raw_spectra()}: Get raw spectra data from specified analyses, returning a \code{data.table} with the spectra data.
\item \code{get_raw_spectra_eic()}: Get extracted ion chromatograms (EIC) for the specified analyses and targets.
\item \code{plot_spectra_eic()}: Plot extracted ion chromatograms (EIC) for the specified analyses and targets.
\item \code{get_raw_spectra_ms1()}: Get MS1 spectra for the specified analyses and targets.
\item \code{get_raw_spectra_ms2()}: Get MS2 spectra for the specified analyses and targets.
}

Chromatogram methods:
\itemize{
\item \code{get_chromatograms()}: Get chromatograms for the specified analyses and chromatogram IDs/indices.
\item \code{plot_chromatograms()}: Plot chromatograms for the specified analyses and chromatogram IDs/indices.
}
}
