% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ProjectMassSpec.R
\name{ProjectMassSpecS3}
\alias{ProjectMassSpecS3}
\alias{info.ProjectMassSpec}
\alias{print.ProjectMassSpec}
\alias{show.ProjectMassSpec}
\alias{get_analyses.ProjectMassSpec}
\alias{add_analyses.ProjectMassSpec}
\alias{remove_analyses.ProjectMassSpec}
\alias{get_analysis_names.ProjectMassSpec}
\alias{get_replicate_names.ProjectMassSpec}
\alias{set_replicate_names.ProjectMassSpec}
\alias{get_blank_names.ProjectMassSpec}
\alias{set_blank_names.ProjectMassSpec}
\alias{get_concentrations.ProjectMassSpec}
\alias{set_concentrations.ProjectMassSpec}
\alias{get_spectra_headers.ProjectMassSpec}
\alias{get_spectra_tic.ProjectMassSpec}
\alias{plot_spectra_tic.ProjectMassSpec}
\alias{plot_spectra_bpc.ProjectMassSpec}
\alias{get_raw_spectra.ProjectMassSpec}
\alias{get_raw_spectra_eic.ProjectMassSpec}
\alias{plot_raw_spectra_eic.ProjectMassSpec}
\alias{plot_raw_spectra_ms1.ProjectMassSpec}
\alias{plot_raw_spectra_ms2.ProjectMassSpec}
\alias{get_raw_spectra_ms1.ProjectMassSpec}
\alias{get_raw_spectra_ms2.ProjectMassSpec}
\alias{get_chromatograms_headers.ProjectMassSpec}
\alias{get_raw_chromatograms.ProjectMassSpec}
\alias{plot_raw_chromatograms.ProjectMassSpec}
\title{ProjectMassSpec S3 Methods}
\usage{
\method{info}{ProjectMassSpec}(x)

\method{print}{ProjectMassSpec}(x, ...)

\method{show}{ProjectMassSpec}(x, ...)

\method{get_analyses}{ProjectMassSpec}(x)

\method{add_analyses}{ProjectMassSpec}(
  x,
  file_paths = character(),
  replicates = character(),
  blanks = character()
)

\method{remove_analyses}{ProjectMassSpec}(x, analyses = character())

\method{get_analysis_names}{ProjectMassSpec}(x)

\method{get_replicate_names}{ProjectMassSpec}(x)

\method{set_replicate_names}{ProjectMassSpec}(x, value)

\method{get_blank_names}{ProjectMassSpec}(x)

\method{set_blank_names}{ProjectMassSpec}(x, value)

\method{get_concentrations}{ProjectMassSpec}(x)

\method{set_concentrations}{ProjectMassSpec}(x, value)

\method{get_spectra_headers}{ProjectMassSpec}(x, analyses = character())

\method{get_spectra_tic}{ProjectMassSpec}(
  x,
  analyses = character(),
  levels = integer(),
  rtmin = numeric(),
  rtmax = numeric()
)

\method{plot_spectra_tic}{ProjectMassSpec}(
  x,
  analyses = character(),
  levels = c(1, 2),
  rtmin = numeric(),
  rtmax = numeric(),
  downsize = NULL,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  groupBy = "analysis",
  interactive = TRUE,
  colorPalette = NULL
)

\method{plot_spectra_bpc}{ProjectMassSpec}(
  x,
  analyses = character(),
  levels = c(1, 2),
  rtmin = numeric(),
  rtmax = numeric(),
  downsize = NULL,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  groupBy = "analysis",
  interactive = TRUE,
  colorPalette = NULL
)

\method{get_raw_spectra}{ProjectMassSpec}(
  x,
  analyses = character(),
  levels = integer(),
  mass = numeric(),
  mz = numeric(),
  rt = numeric(),
  mobility = numeric(),
  ppm = 20,
  sec = 60,
  millisec = 5,
  id = character(),
  allTraces = TRUE,
  isolationWindow = 1.3,
  minIntensityMS1 = 0,
  minIntensityMS2 = 0
)

\method{get_raw_spectra_eic}{ProjectMassSpec}(
  x,
  analyses = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  id = NULL
)

\method{plot_raw_spectra_eic}{ProjectMassSpec}(
  x,
  analyses = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  id = NULL,
  downsize = NULL,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  groupBy = c("analysis", "id"),
  interactive = TRUE,
  colorPalette = NULL
)

\method{plot_raw_spectra_ms1}{ProjectMassSpec}(
  x,
  analyses = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  id = NULL,
  mzClust = 0.003,
  presence = 0.8,
  minIntensity = 1000,
  normalized = FALSE,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  groupBy = "id",
  showText = TRUE,
  interactive = TRUE
)

\method{plot_raw_spectra_ms2}{ProjectMassSpec}(
  x,
  analyses = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  id = NULL,
  isolationWindow = 1.3,
  mzClust = 0.005,
  presence = 0,
  minIntensity = 0,
  normalized = TRUE,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  groupBy = "id",
  showText = TRUE,
  interactive = TRUE
)

\method{get_raw_spectra_ms1}{ProjectMassSpec}(
  x,
  analyses = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  id = NULL,
  mzClust = 0.003,
  presence = 0.8,
  minIntensity = 1000
)

\method{get_raw_spectra_ms2}{ProjectMassSpec}(
  x,
  analyses = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  id = NULL,
  isolationWindow = 1.3,
  mzClust = 0.005,
  presence = 0,
  minIntensity = 0
)

\method{get_chromatograms_headers}{ProjectMassSpec}(x, analyses = NULL)

\method{get_raw_chromatograms}{ProjectMassSpec}(
  x,
  analyses = NULL,
  chromatograms = NULL,
  rtmin = 0,
  rtmax = 0,
  minIntensity = NULL
)

\method{plot_raw_chromatograms}{ProjectMassSpec}(
  x,
  analyses = NULL,
  chromatograms = NULL,
  rtmin = 0,
  rtmax = 0,
  minIntensity = NULL,
  downsize = NULL,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  groupBy = "analysis",
  interactive = TRUE,
  colorPalette = NULL
)
}
\arguments{
\item{x}{A \code{ProjectMassSpec} object.}

\item{...}{Additional arguments passed to the method or ignored by the implementation.}

\item{file_paths}{Character vector with Mass Spec file paths.}

\item{replicates}{Optional character vector with replicate names.}

\item{blanks}{Optional character vector with blank names.}

\item{analyses}{Character or numeric vector with names or indexes of analyses in the \code{Analyses}
object.}

\item{value}{An object that highly depends on the method. Check the specific method documentation for details.}

\item{levels}{Integer vector with the levels of MS data.}

\item{rtmin}{Numeric (length 1) with the minimum retention time.}

\item{rtmax}{Numeric (length 1) with the maximum retention time.}

\item{downsize}{An integer of length one to downsize the TIC plot. The default is 1.}

\item{xLab}{X axis label for the plot.}

\item{yLab}{Y axis label for the plot.}

\item{title}{Plot title.}

\item{groupBy}{Column to use for color grouping (default: "analysis").}

\item{interactive}{Logical, use plotly if TRUE, ggplot2 if FALSE.}

\item{colorPalette}{Optional vector of colors for the plot; if not provided, an internal function is used.}

\item{mass}{A vector with target neutral mass value/s (in Da) or a two columns data.table or data.frame named \code{min}
and \code{max} with minimum and maximum neutral mass values (in Da), respectively. Alternatively, neutral mass (in Da)
and retention time (in seconds) and/or drift time values (in milliseconds) can be given as one data.table or
data.frame with columns named \code{mass} and \code{rt} and/or \code{drift}. Then, the deviations given in the \code{ppm}, \code{sec}
and \code{millisec}  arguments are used to calculate the ranges. Also works with a data.table or data.frame with
minimum and maximum values of neutral mass, retention time and drift time targets. In this case, the column names
must be \code{min}, \code{max}, \code{rtmin}, \code{rtmax}, \code{driftmin} and \code{driftmax}. Note that when mass/time ranges are given, the
\code{ppm}, \code{sec} and \code{millisec}
arguments are not used.}

\item{mz}{A vector with target \emph{m/z} value/s (in Da) or a two columns data.table or data.frame named \code{mzmin} and
\code{mzmax} with minimum and maximum \emph{m/z} values (in Da), respectively. Alternatively, \emph{m/z} (in Da) and
retention time values (in seconds) can be given as one data.table or data.frame with columns named \code{mz} and \code{rt}
and/or \code{drift}. Then, the deviations given in the \code{ppm}, \code{sec} and \code{millisec} arguments are used to
calculate the ranges. Also works with a data.table or data.frame with minimum and maximum values of \emph{m/z},
retention time and drift time targets. In this case, the column names must be \code{mzmin}, \code{mzmax}, \code{rtmin}, \code{rtmax},
\code{driftmin} and \code{driftmax}. Note that when mass/time ranges are given, the \code{ppm}, \code{sec} and \code{millisec}
arguments are not used.}

\item{rt}{A vector with target retention time values (in seconds) or
a two columns data.table or data.frame with minimum and maximum retention
time values (in seconds).}

\item{mobility}{A vector with target drift time values (in milliseconds) or
a two columns data.table or data.frame with minimum and maximum drift
time values (in milliseconds).}

\item{ppm}{Numeric of length one with the mass deviation, in ppm.}

\item{sec}{Numeric of length one with the time deviation, in seconds.}

\item{millisec}{Numeric of length one with the drift time deviation,
in milliseconds.}

\item{id}{Character with the same length as \emph{m/z} and retention time
targets to be used as identifier/s. When not given, the id is built as a
combination of the \emph{m/z} and retention time ranges or values.}

\item{allTraces}{Logical, when \code{TRUE} all level 2 data is returned.
When \code{FALSE} and level has 2, only the MS2 traces of MS1 targets
are returned, using the \code{isolationWindow} to calculate the mass window.}

\item{isolationWindow}{Numeric value with the isolation window (in Da)
applied for ion isolation and fragmentation during acquisition of tandem
data (i.e., MS2 data).}

\item{minIntensityMS1}{Numeric value with the minimum intensity of level 1
data (i.e., MS1 data).}

\item{minIntensityMS2}{Numeric value with the minimum intensity of level 2
data (i.e., MS2 data).}

\item{mzClust}{Numeric (length 1) with the mass deviation threshold (in Da) to cluster mass traces. Highly dependent
on the mass resolution of the MS data.}

\item{presence}{Numeric (length 1) with the required presence ratio from 0 (i.e., doesn't need to be present in any
spectra) to 1 (i.e., must be present in all spectra) for traces during clustering of spectra.}

\item{minIntensity}{Numeric (length 1) with the minimum intensity.}

\item{normalized}{A logical value indicating whether the result should be normalized.}

\item{showText}{Logical (length 1), set to \code{TRUE} to show the text annotations.}

\item{chromatograms}{A character or integer vector with the ID (i.e. name) or the index of the chromatograms.}

\item{db}{Path to DuckDB file.}

\item{project_id}{Active project identifier.}

\item{file}{A \emph{json} or \emph{rds} file to be imported.}

\item{analysis}{Character scalar with the analysis name.}
}
\description{
S3 interface methods for \code{ProjectMassSpec}.
}
\section{Functions}{
\itemize{
\item \code{info(ProjectMassSpec)}: Print a short information table for the project analyses.

\item \code{print(ProjectMassSpec)}: Print a short summary.

\item \code{show(ProjectMassSpec)}: Show a short summary.

\item \code{get_analyses(ProjectMassSpec)}: Return imported analyses.

\item \code{add_analyses(ProjectMassSpec)}: Import one or more data files into the project as analyses.

\item \code{remove_analyses(ProjectMassSpec)}: Remove analyses from the project by name or index.

\item \code{get_analysis_names(ProjectMassSpec)}: Return analysis names for the project.

\item \code{get_replicate_names(ProjectMassSpec)}: Return replicate names for the project.

\item \code{set_replicate_names(ProjectMassSpec)}: Set replicate names for the project analyses.

\item \code{get_blank_names(ProjectMassSpec)}: Return blank names for the project.

\item \code{set_blank_names(ProjectMassSpec)}: Set blank names for the project analyses.

\item \code{get_concentrations(ProjectMassSpec)}: Return concentrations vector for the project.

\item \code{set_concentrations(ProjectMassSpec)}: Set concentrations for the project analyses.

\item \code{get_spectra_headers(ProjectMassSpec)}: Fetch spectra headers for specified analyses.

\item \code{get_spectra_tic(ProjectMassSpec)}: Return spectra TIC rows for selected analyses.

\item \code{plot_spectra_tic(ProjectMassSpec)}: Plot total ion current (TIC) spectra for specified analyses.

\item \code{plot_spectra_bpc(ProjectMassSpec)}: Plot base peak chromatogram (BPC) spectra for specified analyses.

\item \code{get_raw_spectra(ProjectMassSpec)}: Get raw spectra data for specified analyses; returns a data.table with spectra rows.

\item \code{get_raw_spectra_eic(ProjectMassSpec)}: Get extracted ion chromatograms (EIC) for specified analyses.

\item \code{plot_raw_spectra_eic(ProjectMassSpec)}: Plot extracted ion chromatograms (EIC) for specified analyses and targets.

\item \code{plot_raw_spectra_ms1(ProjectMassSpec)}: Plot clustered MS1 spectra for specified analyses and targets.

\item \code{plot_raw_spectra_ms2(ProjectMassSpec)}: Plot clustered MS2 spectra for specified analyses and targets.

\item \code{get_raw_spectra_ms1(ProjectMassSpec)}: Get clustered MS1 spectra for specified analyses and targets.

\item \code{get_raw_spectra_ms2(ProjectMassSpec)}: Get clustered MS2 spectra for specified analyses and targets.

\item \code{get_chromatograms_headers(ProjectMassSpec)}: Return chromatogram headers for selected analyses.

\item \code{get_raw_chromatograms(ProjectMassSpec)}: Get raw chromatograms for selected analyses.

\item \code{plot_raw_chromatograms(ProjectMassSpec)}: Plot raw chromatograms for selected analyses.

}}
