% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ProjectMassSpecSpectra.R
\name{ProjectMassSpecSpectra}
\alias{ProjectMassSpecSpectra}
\title{Project Mass Spec Spectra R6 Class}
\description{
R6 child of \code{ProjectMassSpec} exposing the spectra-focused Mass
Spec interface backed by the native C++ spectra facade. Shared raw spectra
access remains on \code{ProjectMassSpec} so both \code{ProjectMassSpecSpectra} and
\code{ProjectNonTargetAnalysis} inherit the same base spectra helpers.
}
\details{
This is a public user-facing project class.

In addition to the methods documented on this page, it inherits shared project
runtime methods such as \code{run_app()}, \code{run_workflow()}, \code{report_quarto()},
\code{metadata}, \code{workflow}, \code{get_audit()}, and \code{list_tables()}, plus shared Mass
Spec methods such as \code{import_files()}, \code{list_analyses()}, \code{get_spectra_headers()},
\code{get_raw_spectra()}, \code{get_raw_spectra_tic()}, \code{get_raw_spectra_bpc()},
\code{get_raw_spectra_eic()}, \code{get_raw_spectra_ms1()}, \code{get_raw_spectra_ms2()},
\code{plot_spectra_tic()}, \code{plot_spectra_bpc()}, and \code{plot_spectra_eic()}.

See \code{?Project} for shared project/runtime methods and \code{?ProjectMassSpec} for
inherited shared Mass Spec methods.

Use \code{?ProjectMassSpecSpectra} as the main entry point for the spectra-specific
interface on top of those base classes.
}
\section{Super classes}{
\code{\link[StreamFind:Project]{StreamFind::Project}} -> \code{\link[StreamFind:ProjectMassSpec]{StreamFind::ProjectMassSpec}} -> \code{ProjectMassSpecSpectra}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ProjectMassSpecSpectra-new}{\code{ProjectMassSpecSpectra$new()}}
\item \href{#method-ProjectMassSpecSpectra-get_mass_spec_spectra_ptr}{\code{ProjectMassSpecSpectra$get_mass_spec_spectra_ptr()}}
\item \href{#method-ProjectMassSpecSpectra-available_processing_methods}{\code{ProjectMassSpecSpectra$available_processing_methods()}}
\item \href{#method-ProjectMassSpecSpectra-print}{\code{ProjectMassSpecSpectra$print()}}
\item \href{#method-ProjectMassSpecSpectra-show}{\code{ProjectMassSpecSpectra$show()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="copy"><a href='../../StreamFind/html/Project.html#method-Project-copy'><code>StreamFind::Project$copy()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="get_audit"><a href='../../StreamFind/html/Project.html#method-Project-get_audit'><code>StreamFind::Project$get_audit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="get_domain"><a href='../../StreamFind/html/Project.html#method-Project-get_domain'><code>StreamFind::Project$get_domain()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="get_metadata"><a href='../../StreamFind/html/Project.html#method-Project-get_metadata'><code>StreamFind::Project$get_metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="get_ptr"><a href='../../StreamFind/html/Project.html#method-Project-get_ptr'><code>StreamFind::Project$get_ptr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="get_workflow"><a href='../../StreamFind/html/Project.html#method-Project-get_workflow'><code>StreamFind::Project$get_workflow()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="list_tables"><a href='../../StreamFind/html/Project.html#method-Project-list_tables'><code>StreamFind::Project$list_tables()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="report_quarto"><a href='../../StreamFind/html/Project.html#method-Project-report_quarto'><code>StreamFind::Project$report_quarto()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="run_app"><a href='../../StreamFind/html/Project.html#method-Project-run_app'><code>StreamFind::Project$run_app()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="run_processing_step"><a href='../../StreamFind/html/Project.html#method-Project-run_processing_step'><code>StreamFind::Project$run_processing_step()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="run_workflow"><a href='../../StreamFind/html/Project.html#method-Project-run_workflow'><code>StreamFind::Project$run_workflow()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="set_metadata"><a href='../../StreamFind/html/Project.html#method-Project-set_metadata'><code>StreamFind::Project$set_metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="set_workflow"><a href='../../StreamFind/html/Project.html#method-Project-set_workflow'><code>StreamFind::Project$set_workflow()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="validate"><a href='../../StreamFind/html/Project.html#method-Project-validate'><code>StreamFind::Project$validate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_analysis_names"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_analysis_names'><code>StreamFind::ProjectMassSpec$get_analysis_names()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_blank_names"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_blank_names'><code>StreamFind::ProjectMassSpec$get_blank_names()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_concentrations"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_concentrations'><code>StreamFind::ProjectMassSpec$get_concentrations()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_mass_spec_ptr"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_mass_spec_ptr'><code>StreamFind::ProjectMassSpec$get_mass_spec_ptr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_raw_spectra"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_raw_spectra'><code>StreamFind::ProjectMassSpec$get_raw_spectra()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_raw_spectra_bpc"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_raw_spectra_bpc'><code>StreamFind::ProjectMassSpec$get_raw_spectra_bpc()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_raw_spectra_eic"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_raw_spectra_eic'><code>StreamFind::ProjectMassSpec$get_raw_spectra_eic()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_raw_spectra_ms1"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_raw_spectra_ms1'><code>StreamFind::ProjectMassSpec$get_raw_spectra_ms1()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_raw_spectra_ms2"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_raw_spectra_ms2'><code>StreamFind::ProjectMassSpec$get_raw_spectra_ms2()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_raw_spectra_tic"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_raw_spectra_tic'><code>StreamFind::ProjectMassSpec$get_raw_spectra_tic()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_replicate_names"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_replicate_names'><code>StreamFind::ProjectMassSpec$get_replicate_names()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_spectra_headers"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_spectra_headers'><code>StreamFind::ProjectMassSpec$get_spectra_headers()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_spectra_tic"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_spectra_tic'><code>StreamFind::ProjectMassSpec$get_spectra_tic()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="import_files"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-import_files'><code>StreamFind::ProjectMassSpec$import_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="list_analyses"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-list_analyses'><code>StreamFind::ProjectMassSpec$list_analyses()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="plot_spectra_bpc"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-plot_spectra_bpc'><code>StreamFind::ProjectMassSpec$plot_spectra_bpc()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="plot_spectra_eic"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-plot_spectra_eic'><code>StreamFind::ProjectMassSpec$plot_spectra_eic()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="plot_spectra_tic"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-plot_spectra_tic'><code>StreamFind::ProjectMassSpec$plot_spectra_tic()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="remove_analysis"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-remove_analysis'><code>StreamFind::ProjectMassSpec$remove_analysis()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="set_blank_names"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-set_blank_names'><code>StreamFind::ProjectMassSpec$set_blank_names()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="set_concentrations"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-set_concentrations'><code>StreamFind::ProjectMassSpec$set_concentrations()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="set_replicate_names"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-set_replicate_names'><code>StreamFind::ProjectMassSpec$set_replicate_names()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectMassSpecSpectra-new"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectMassSpecSpectra-new}{}}}
\subsection{Method \code{new()}}{
Create a spectra-specific Mass Spec project wrapper.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectMassSpecSpectra$new(
  db,
  project_id,
  .ptr = NULL,
  .mass_spec_ptr = NULL,
  .mass_spec_spectra_ptr = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{db}}{Path to the DuckDB project file.}

\item{\code{project_id}}{Active project identifier.}

\item{\code{.ptr}}{Existing native project pointer for internal use.}

\item{\code{.mass_spec_ptr}}{Existing native shared Mass Spec pointer for internal use.}

\item{\code{.mass_spec_spectra_ptr}}{Existing native spectra pointer for internal use.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectMassSpecSpectra-get_mass_spec_spectra_ptr"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectMassSpecSpectra-get_mass_spec_spectra_ptr}{}}}
\subsection{Method \code{get_mass_spec_spectra_ptr()}}{
Return the native spectra pointer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectMassSpecSpectra$get_mass_spec_spectra_ptr()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectMassSpecSpectra-available_processing_methods"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectMassSpecSpectra-available_processing_methods}{}}}
\subsection{Method \code{available_processing_methods()}}{
Return spectra-project processing-step metadata.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectMassSpecSpectra$available_processing_methods()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named list of \code{ProcessingStep} metadata objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectMassSpecSpectra-print"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectMassSpecSpectra-print}{}}}
\subsection{Method \code{print()}}{
Print a short summary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectMassSpecSpectra$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectMassSpecSpectra-show"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectMassSpecSpectra-show}{}}}
\subsection{Method \code{show()}}{
Show a short summary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectMassSpecSpectra$show(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments ignored.}
}
\if{html}{\out{</div>}}
}
}
}
