% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ProjectNonTargetAnalysis.R
\name{ProjectNonTargetAnalysis}
\alias{ProjectNonTargetAnalysis}
\title{Project Non-Target Analysis R6 Class}
\description{
R6 child of \code{ProjectMassSpec} exposing the non-target analysis-focused NTS interface.
}
\keyword{internal}
\section{Super classes}{
\code{\link[StreamFind:Project]{StreamFind::Project}} -> \code{\link[StreamFind:ProjectMassSpec]{StreamFind::ProjectMassSpec}} -> \code{ProjectNonTargetAnalysis}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ProjectNonTargetAnalysis-new}{\code{ProjectNonTargetAnalysis$new()}}
\item \href{#method-ProjectNonTargetAnalysis-get_nts_ptr}{\code{ProjectNonTargetAnalysis$get_nts_ptr()}}
\item \href{#method-ProjectNonTargetAnalysis-available_processing_methods}{\code{ProjectNonTargetAnalysis$available_processing_methods()}}
\item \href{#method-ProjectNonTargetAnalysis-get_suspects}{\code{ProjectNonTargetAnalysis$get_suspects()}}
\item \href{#method-ProjectNonTargetAnalysis-get_internal_standards}{\code{ProjectNonTargetAnalysis$get_internal_standards()}}
\item \href{#method-ProjectNonTargetAnalysis-get_transformation_products}{\code{ProjectNonTargetAnalysis$get_transformation_products()}}
\item \href{#method-ProjectNonTargetAnalysis-get_features}{\code{ProjectNonTargetAnalysis$get_features()}}
\item \href{#method-ProjectNonTargetAnalysis-get_features_profile}{\code{ProjectNonTargetAnalysis$get_features_profile()}}
\item \href{#method-ProjectNonTargetAnalysis-get_features_count}{\code{ProjectNonTargetAnalysis$get_features_count()}}
\item \href{#method-ProjectNonTargetAnalysis-info}{\code{ProjectNonTargetAnalysis$info()}}
\item \href{#method-ProjectNonTargetAnalysis-plot_features_count}{\code{ProjectNonTargetAnalysis$plot_features_count()}}
\item \href{#method-ProjectNonTargetAnalysis-plot_features_profile}{\code{ProjectNonTargetAnalysis$plot_features_profile()}}
\item \href{#method-ProjectNonTargetAnalysis-plot_features}{\code{ProjectNonTargetAnalysis$plot_features()}}
\item \href{#method-ProjectNonTargetAnalysis-map_features}{\code{ProjectNonTargetAnalysis$map_features()}}
\item \href{#method-ProjectNonTargetAnalysis-plot_features_ms1}{\code{ProjectNonTargetAnalysis$plot_features_ms1()}}
\item \href{#method-ProjectNonTargetAnalysis-plot_features_ms2}{\code{ProjectNonTargetAnalysis$plot_features_ms2()}}
\item \href{#method-ProjectNonTargetAnalysis-plot_suspects_ms2}{\code{ProjectNonTargetAnalysis$plot_suspects_ms2()}}
\item \href{#method-ProjectNonTargetAnalysis-get_fold_change}{\code{ProjectNonTargetAnalysis$get_fold_change()}}
\item \href{#method-ProjectNonTargetAnalysis-plot_fold_change}{\code{ProjectNonTargetAnalysis$plot_fold_change()}}
\item \href{#method-ProjectNonTargetAnalysis-plot_transformation_products}{\code{ProjectNonTargetAnalysis$plot_transformation_products()}}
\item \href{#method-ProjectNonTargetAnalysis-print}{\code{ProjectNonTargetAnalysis$print()}}
\item \href{#method-ProjectNonTargetAnalysis-show}{\code{ProjectNonTargetAnalysis$show()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="copy"><a href='../../StreamFind/html/Project.html#method-Project-copy'><code>StreamFind::Project$copy()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="delete_cache"><a href='../../StreamFind/html/Project.html#method-Project-delete_cache'><code>StreamFind::Project$delete_cache()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="get_audit"><a href='../../StreamFind/html/Project.html#method-Project-get_audit'><code>StreamFind::Project$get_audit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="get_cache"><a href='../../StreamFind/html/Project.html#method-Project-get_cache'><code>StreamFind::Project$get_cache()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="get_cache_size"><a href='../../StreamFind/html/Project.html#method-Project-get_cache_size'><code>StreamFind::Project$get_cache_size()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="get_db"><a href='../../StreamFind/html/Project.html#method-Project-get_db'><code>StreamFind::Project$get_db()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="get_domain"><a href='../../StreamFind/html/Project.html#method-Project-get_domain'><code>StreamFind::Project$get_domain()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="get_metadata"><a href='../../StreamFind/html/Project.html#method-Project-get_metadata'><code>StreamFind::Project$get_metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="get_project_id"><a href='../../StreamFind/html/Project.html#method-Project-get_project_id'><code>StreamFind::Project$get_project_id()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="get_ptr"><a href='../../StreamFind/html/Project.html#method-Project-get_ptr'><code>StreamFind::Project$get_ptr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="get_workflow"><a href='../../StreamFind/html/Project.html#method-Project-get_workflow'><code>StreamFind::Project$get_workflow()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="list_tables"><a href='../../StreamFind/html/Project.html#method-Project-list_tables'><code>StreamFind::Project$list_tables()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="report_quarto"><a href='../../StreamFind/html/Project.html#method-Project-report_quarto'><code>StreamFind::Project$report_quarto()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="run_app"><a href='../../StreamFind/html/Project.html#method-Project-run_app'><code>StreamFind::Project$run_app()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="run_processing_step"><a href='../../StreamFind/html/Project.html#method-Project-run_processing_step'><code>StreamFind::Project$run_processing_step()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="run_workflow"><a href='../../StreamFind/html/Project.html#method-Project-run_workflow'><code>StreamFind::Project$run_workflow()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="set_metadata"><a href='../../StreamFind/html/Project.html#method-Project-set_metadata'><code>StreamFind::Project$set_metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="set_workflow"><a href='../../StreamFind/html/Project.html#method-Project-set_workflow'><code>StreamFind::Project$set_workflow()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="Project" data-id="validate"><a href='../../StreamFind/html/Project.html#method-Project-validate'><code>StreamFind::Project$validate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="add_analyses"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-add_analyses'><code>StreamFind::ProjectMassSpec$add_analyses()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_analyses"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_analyses'><code>StreamFind::ProjectMassSpec$get_analyses()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_analysis_names"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_analysis_names'><code>StreamFind::ProjectMassSpec$get_analysis_names()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_blank_names"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_blank_names'><code>StreamFind::ProjectMassSpec$get_blank_names()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_concentrations"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_concentrations'><code>StreamFind::ProjectMassSpec$get_concentrations()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_mass_spec_ptr"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_mass_spec_ptr'><code>StreamFind::ProjectMassSpec$get_mass_spec_ptr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_raw_spectra"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_raw_spectra'><code>StreamFind::ProjectMassSpec$get_raw_spectra()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_raw_spectra_eic"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_raw_spectra_eic'><code>StreamFind::ProjectMassSpec$get_raw_spectra_eic()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_raw_spectra_ms1"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_raw_spectra_ms1'><code>StreamFind::ProjectMassSpec$get_raw_spectra_ms1()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_raw_spectra_ms2"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_raw_spectra_ms2'><code>StreamFind::ProjectMassSpec$get_raw_spectra_ms2()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_replicate_names"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_replicate_names'><code>StreamFind::ProjectMassSpec$get_replicate_names()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_spectra_headers"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_spectra_headers'><code>StreamFind::ProjectMassSpec$get_spectra_headers()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="get_spectra_tic"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-get_spectra_tic'><code>StreamFind::ProjectMassSpec$get_spectra_tic()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="plot_raw_spectra_eic"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-plot_raw_spectra_eic'><code>StreamFind::ProjectMassSpec$plot_raw_spectra_eic()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="plot_spectra_bpc"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-plot_spectra_bpc'><code>StreamFind::ProjectMassSpec$plot_spectra_bpc()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="plot_spectra_tic"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-plot_spectra_tic'><code>StreamFind::ProjectMassSpec$plot_spectra_tic()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="remove_analyses"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-remove_analyses'><code>StreamFind::ProjectMassSpec$remove_analyses()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="set_blank_names"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-set_blank_names'><code>StreamFind::ProjectMassSpec$set_blank_names()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="set_concentrations"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-set_concentrations'><code>StreamFind::ProjectMassSpec$set_concentrations()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="ProjectMassSpec" data-id="set_replicate_names"><a href='../../StreamFind/html/ProjectMassSpec.html#method-ProjectMassSpec-set_replicate_names'><code>StreamFind::ProjectMassSpec$set_replicate_names()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectNonTargetAnalysis-new"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectNonTargetAnalysis-new}{}}}
\subsection{Method \code{new()}}{
Create an NTS domain wrapper on top of a shared Mass Spec project.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectNonTargetAnalysis$new(db, project_id, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{db}}{Path to DuckDB file.}

\item{\code{project_id}}{Active project identifier.}

\item{\code{...}}{Additional arguments passed to the method or ignored by the implementation.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectNonTargetAnalysis-get_nts_ptr"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectNonTargetAnalysis-get_nts_ptr}{}}}
\subsection{Method \code{get_nts_ptr()}}{
Return the native NTS pointer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectNonTargetAnalysis$get_nts_ptr()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectNonTargetAnalysis-available_processing_methods"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectNonTargetAnalysis-available_processing_methods}{}}}
\subsection{Method \code{available_processing_methods()}}{
Return project-owned NTS processing-step metadata.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectNonTargetAnalysis$available_processing_methods()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named list of \code{ProcessingStep} metadata objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectNonTargetAnalysis-get_suspects"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectNonTargetAnalysis-get_suspects}{}}}
\subsection{Method \code{get_suspects()}}{
Return shared \code{NTS_SUSPECTS} rows for selected analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectNonTargetAnalysis$get_suspects(
  analyses = NULL,
  features = NULL,
  groups = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{Character or numeric vector with names or indexes of analyses in the \code{Analyses}
object.}

\item{\code{features}}{A character vector with the feature IDs or a data.frame or data.table with a column named \code{feature}
with the feature IDs. A column named \code{analysis} can be given to specify from which analyses the features matched.}

\item{\code{groups}}{A numeric or character vector with feature group IDs, or a data.frame/data.table with a \code{feature_group} (or \code{group}) column; an optional \code{analysis} column restricts analyses and a \code{name} column sets custom labels.}

\item{\code{mass}}{A vector with target neutral mass value/s (in Da) or a two columns data.table or data.frame named \code{min}
and \code{max} with minimum and maximum neutral mass values (in Da), respectively. Alternatively, neutral mass (in Da)
and retention time (in seconds) and/or drift time values (in milliseconds) can be given as one data.table or
data.frame with columns named \code{mass} and \code{rt} and/or \code{drift}. Then, the deviations given in the \code{ppm}, \code{sec}
and \code{millisec}  arguments are used to calculate the ranges. Also works with a data.table or data.frame with
minimum and maximum values of neutral mass, retention time and drift time targets. In this case, the column names
must be \code{min}, \code{max}, \code{rtmin}, \code{rtmax}, \code{driftmin} and \code{driftmax}. Note that when mass/time ranges are given, the
\code{ppm}, \code{sec} and \code{millisec}
arguments are not used.}

\item{\code{mz}}{A vector with target \emph{m/z} value/s (in Da) or a two columns data.table or data.frame named \code{mzmin} and
\code{mzmax} with minimum and maximum \emph{m/z} values (in Da), respectively. Alternatively, \emph{m/z} (in Da) and
retention time values (in seconds) can be given as one data.table or data.frame with columns named \code{mz} and \code{rt}
and/or \code{drift}. Then, the deviations given in the \code{ppm}, \code{sec} and \code{millisec} arguments are used to
calculate the ranges. Also works with a data.table or data.frame with minimum and maximum values of \emph{m/z},
retention time and drift time targets. In this case, the column names must be \code{mzmin}, \code{mzmax}, \code{rtmin}, \code{rtmax},
\code{driftmin} and \code{driftmax}. Note that when mass/time ranges are given, the \code{ppm}, \code{sec} and \code{millisec}
arguments are not used.}

\item{\code{rt}}{A vector with target retention time values (in seconds) or
a two columns data.table or data.frame with minimum and maximum retention
time values (in seconds).}

\item{\code{mobility}}{A vector with target drift time values (in milliseconds) or
a two columns data.table or data.frame with minimum and maximum drift
time values (in milliseconds).}

\item{\code{ppm}}{Numeric of length one with the mass deviation, in ppm.}

\item{\code{sec}}{Numeric of length one with the time deviation, in seconds.}

\item{\code{millisec}}{Numeric of length one with the drift time deviation,
in milliseconds.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectNonTargetAnalysis-get_internal_standards"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectNonTargetAnalysis-get_internal_standards}{}}}
\subsection{Method \code{get_internal_standards()}}{
Return shared \code{NTS_INTERNAL_STANDARDS} rows for selected analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectNonTargetAnalysis$get_internal_standards(
  analyses = NULL,
  features = NULL,
  groups = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{Character or numeric vector with names or indexes of analyses in the \code{Analyses}
object.}

\item{\code{features}}{A character vector with the feature IDs or a data.frame or data.table with a column named \code{feature}
with the feature IDs. A column named \code{analysis} can be given to specify from which analyses the features matched.}

\item{\code{groups}}{A numeric or character vector with feature group IDs, or a data.frame/data.table with a \code{feature_group} (or \code{group}) column; an optional \code{analysis} column restricts analyses and a \code{name} column sets custom labels.}

\item{\code{mass}}{A vector with target neutral mass value/s (in Da) or a two columns data.table or data.frame named \code{min}
and \code{max} with minimum and maximum neutral mass values (in Da), respectively. Alternatively, neutral mass (in Da)
and retention time (in seconds) and/or drift time values (in milliseconds) can be given as one data.table or
data.frame with columns named \code{mass} and \code{rt} and/or \code{drift}. Then, the deviations given in the \code{ppm}, \code{sec}
and \code{millisec}  arguments are used to calculate the ranges. Also works with a data.table or data.frame with
minimum and maximum values of neutral mass, retention time and drift time targets. In this case, the column names
must be \code{min}, \code{max}, \code{rtmin}, \code{rtmax}, \code{driftmin} and \code{driftmax}. Note that when mass/time ranges are given, the
\code{ppm}, \code{sec} and \code{millisec}
arguments are not used.}

\item{\code{mz}}{A vector with target \emph{m/z} value/s (in Da) or a two columns data.table or data.frame named \code{mzmin} and
\code{mzmax} with minimum and maximum \emph{m/z} values (in Da), respectively. Alternatively, \emph{m/z} (in Da) and
retention time values (in seconds) can be given as one data.table or data.frame with columns named \code{mz} and \code{rt}
and/or \code{drift}. Then, the deviations given in the \code{ppm}, \code{sec} and \code{millisec} arguments are used to
calculate the ranges. Also works with a data.table or data.frame with minimum and maximum values of \emph{m/z},
retention time and drift time targets. In this case, the column names must be \code{mzmin}, \code{mzmax}, \code{rtmin}, \code{rtmax},
\code{driftmin} and \code{driftmax}. Note that when mass/time ranges are given, the \code{ppm}, \code{sec} and \code{millisec}
arguments are not used.}

\item{\code{rt}}{A vector with target retention time values (in seconds) or
a two columns data.table or data.frame with minimum and maximum retention
time values (in seconds).}

\item{\code{mobility}}{A vector with target drift time values (in milliseconds) or
a two columns data.table or data.frame with minimum and maximum drift
time values (in milliseconds).}

\item{\code{ppm}}{Numeric of length one with the mass deviation, in ppm.}

\item{\code{sec}}{Numeric of length one with the time deviation, in seconds.}

\item{\code{millisec}}{Numeric of length one with the drift time deviation,
in milliseconds.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectNonTargetAnalysis-get_transformation_products"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectNonTargetAnalysis-get_transformation_products}{}}}
\subsection{Method \code{get_transformation_products()}}{
Return shared \code{NTS_TRANSFORMATION_PRODUCTS} rows.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectNonTargetAnalysis$get_transformation_products(
  parents = NULL,
  groups = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parents}}{Optional parent names to keep.}

\item{\code{groups}}{A numeric or character vector with feature group IDs, or a data.frame/data.table with a \code{feature_group} (or \code{group}) column; an optional \code{analysis} column restricts analyses and a \code{name} column sets custom labels.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectNonTargetAnalysis-get_features"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectNonTargetAnalysis-get_features}{}}}
\subsection{Method \code{get_features()}}{
Return shared \code{NTS_FEATURES} rows for selected analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectNonTargetAnalysis$get_features(
  analyses = NULL,
  features = NULL,
  groups = NULL,
  components = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  filtered = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{Character or numeric vector with names or indexes of analyses in the \code{Analyses}
object.}

\item{\code{features}}{A character vector with the feature IDs or a data.frame or data.table with a column named \code{feature}
with the feature IDs. A column named \code{analysis} can be given to specify from which analyses the features matched.}

\item{\code{groups}}{A numeric or character vector with feature group IDs, or a data.frame/data.table with a \code{feature_group} (or \code{group}) column; an optional \code{analysis} column restricts analyses and a \code{name} column sets custom labels.}

\item{\code{components}}{A numeric or character vector with feature component IDs, or a data.frame/data.table with a \code{feature_component} (or \code{component}) column; an optional \code{analysis} column restricts analyses and a \code{name} column sets custom labels.}

\item{\code{mass}}{A vector with target neutral mass value/s (in Da) or a two columns data.table or data.frame named \code{min}
and \code{max} with minimum and maximum neutral mass values (in Da), respectively. Alternatively, neutral mass (in Da)
and retention time (in seconds) and/or drift time values (in milliseconds) can be given as one data.table or
data.frame with columns named \code{mass} and \code{rt} and/or \code{drift}. Then, the deviations given in the \code{ppm}, \code{sec}
and \code{millisec}  arguments are used to calculate the ranges. Also works with a data.table or data.frame with
minimum and maximum values of neutral mass, retention time and drift time targets. In this case, the column names
must be \code{min}, \code{max}, \code{rtmin}, \code{rtmax}, \code{driftmin} and \code{driftmax}. Note that when mass/time ranges are given, the
\code{ppm}, \code{sec} and \code{millisec}
arguments are not used.}

\item{\code{mz}}{A vector with target \emph{m/z} value/s (in Da) or a two columns data.table or data.frame named \code{mzmin} and
\code{mzmax} with minimum and maximum \emph{m/z} values (in Da), respectively. Alternatively, \emph{m/z} (in Da) and
retention time values (in seconds) can be given as one data.table or data.frame with columns named \code{mz} and \code{rt}
and/or \code{drift}. Then, the deviations given in the \code{ppm}, \code{sec} and \code{millisec} arguments are used to
calculate the ranges. Also works with a data.table or data.frame with minimum and maximum values of \emph{m/z},
retention time and drift time targets. In this case, the column names must be \code{mzmin}, \code{mzmax}, \code{rtmin}, \code{rtmax},
\code{driftmin} and \code{driftmax}. Note that when mass/time ranges are given, the \code{ppm}, \code{sec} and \code{millisec}
arguments are not used.}

\item{\code{rt}}{A vector with target retention time values (in seconds) or
a two columns data.table or data.frame with minimum and maximum retention
time values (in seconds).}

\item{\code{mobility}}{A vector with target drift time values (in milliseconds) or
a two columns data.table or data.frame with minimum and maximum drift
time values (in milliseconds).}

\item{\code{ppm}}{Numeric of length one with the mass deviation, in ppm.}

\item{\code{sec}}{Numeric of length one with the time deviation, in seconds.}

\item{\code{millisec}}{Numeric of length one with the drift time deviation,
in milliseconds.}

\item{\code{filtered}}{Logical (length 1). When \code{TRUE}, filtered features/groups are considered.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectNonTargetAnalysis-get_features_profile"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectNonTargetAnalysis-get_features_profile}{}}}
\subsection{Method \code{get_features_profile()}}{
Return feature-group profiles across analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectNonTargetAnalysis$get_features_profile(
  analyses = NULL,
  groups = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  filtered = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{Character or numeric vector with names or indexes of analyses in the \code{Analyses}
object.}

\item{\code{groups}}{A numeric or character vector with feature group IDs, or a data.frame/data.table with a \code{feature_group} (or \code{group}) column; an optional \code{analysis} column restricts analyses and a \code{name} column sets custom labels.}

\item{\code{mass}}{A vector with target neutral mass value/s (in Da) or a two columns data.table or data.frame named \code{min}
and \code{max} with minimum and maximum neutral mass values (in Da), respectively. Alternatively, neutral mass (in Da)
and retention time (in seconds) and/or drift time values (in milliseconds) can be given as one data.table or
data.frame with columns named \code{mass} and \code{rt} and/or \code{drift}. Then, the deviations given in the \code{ppm}, \code{sec}
and \code{millisec}  arguments are used to calculate the ranges. Also works with a data.table or data.frame with
minimum and maximum values of neutral mass, retention time and drift time targets. In this case, the column names
must be \code{min}, \code{max}, \code{rtmin}, \code{rtmax}, \code{driftmin} and \code{driftmax}. Note that when mass/time ranges are given, the
\code{ppm}, \code{sec} and \code{millisec}
arguments are not used.}

\item{\code{mz}}{A vector with target \emph{m/z} value/s (in Da) or a two columns data.table or data.frame named \code{mzmin} and
\code{mzmax} with minimum and maximum \emph{m/z} values (in Da), respectively. Alternatively, \emph{m/z} (in Da) and
retention time values (in seconds) can be given as one data.table or data.frame with columns named \code{mz} and \code{rt}
and/or \code{drift}. Then, the deviations given in the \code{ppm}, \code{sec} and \code{millisec} arguments are used to
calculate the ranges. Also works with a data.table or data.frame with minimum and maximum values of \emph{m/z},
retention time and drift time targets. In this case, the column names must be \code{mzmin}, \code{mzmax}, \code{rtmin}, \code{rtmax},
\code{driftmin} and \code{driftmax}. Note that when mass/time ranges are given, the \code{ppm}, \code{sec} and \code{millisec}
arguments are not used.}

\item{\code{rt}}{A vector with target retention time values (in seconds) or
a two columns data.table or data.frame with minimum and maximum retention
time values (in seconds).}

\item{\code{mobility}}{A vector with target drift time values (in milliseconds) or
a two columns data.table or data.frame with minimum and maximum drift
time values (in milliseconds).}

\item{\code{ppm}}{Numeric of length one with the mass deviation, in ppm.}

\item{\code{sec}}{Numeric of length one with the time deviation, in seconds.}

\item{\code{millisec}}{Numeric of length one with the drift time deviation,
in milliseconds.}

\item{\code{filtered}}{Logical (length 1). When \code{TRUE}, filtered features/groups are considered.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectNonTargetAnalysis-get_features_count"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectNonTargetAnalysis-get_features_count}{}}}
\subsection{Method \code{get_features_count()}}{
Return a per-analysis summary of shared \code{NTS_FEATURES} rows.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectNonTargetAnalysis$get_features_count(analyses = NULL, filtered = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{Character or numeric vector with names or indexes of analyses in the \code{Analyses}
object.}

\item{\code{filtered}}{Logical (length 1). When \code{TRUE}, filtered features/groups are considered.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectNonTargetAnalysis-info"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectNonTargetAnalysis-info}{}}}
\subsection{Method \code{info()}}{
Return a compact per-analysis feature summary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectNonTargetAnalysis$info()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectNonTargetAnalysis-plot_features_count"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectNonTargetAnalysis-plot_features_count}{}}}
\subsection{Method \code{plot_features_count()}}{
Plot the number of features for selected analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectNonTargetAnalysis$plot_features_count(
  analyses = NULL,
  filtered = FALSE,
  yLab = NULL,
  title = NULL,
  groupBy = "analysis",
  showLegend = TRUE,
  showHoverText = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{Character or numeric vector with names or indexes of analyses in the \code{Analyses}
object.}

\item{\code{filtered}}{Logical (length 1). When \code{TRUE}, filtered features/groups are considered.}

\item{\code{yLab}}{A string with the title for the y axis.}

\item{\code{yLab}}{A string with the title for the y axis.}

\item{\code{title}}{A string with the title.}

\item{\code{groupBy}}{Column to use for color grouping (default: "analysis").}

\item{\code{showLegend}}{Logical (length 1). Set to \code{TRUE} to show legend.}

\item{\code{showHoverText}}{Logical (length 1), set to \code{TRUE} to show hovering text annotations.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectNonTargetAnalysis-plot_features_profile"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectNonTargetAnalysis-plot_features_profile}{}}}
\subsection{Method \code{plot_features_profile()}}{
Plot feature-group profiles across analyses or replicates.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectNonTargetAnalysis$plot_features_profile(
  analyses = NULL,
  groups = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  filtered = FALSE,
  groupBy = "analysis",
  normalized = FALSE,
  yLab = NULL,
  title = NULL,
  interactive = TRUE,
  showLegend = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{Character or numeric vector with names or indexes of analyses in the \code{Analyses}
object.}

\item{\code{groups}}{A numeric or character vector with feature group IDs, or a data.frame/data.table with a \code{feature_group} (or \code{group}) column; an optional \code{analysis} column restricts analyses and a \code{name} column sets custom labels.}

\item{\code{mass}}{A vector with target neutral mass value/s (in Da) or a two columns data.table or data.frame named \code{min}
and \code{max} with minimum and maximum neutral mass values (in Da), respectively. Alternatively, neutral mass (in Da)
and retention time (in seconds) and/or drift time values (in milliseconds) can be given as one data.table or
data.frame with columns named \code{mass} and \code{rt} and/or \code{drift}. Then, the deviations given in the \code{ppm}, \code{sec}
and \code{millisec}  arguments are used to calculate the ranges. Also works with a data.table or data.frame with
minimum and maximum values of neutral mass, retention time and drift time targets. In this case, the column names
must be \code{min}, \code{max}, \code{rtmin}, \code{rtmax}, \code{driftmin} and \code{driftmax}. Note that when mass/time ranges are given, the
\code{ppm}, \code{sec} and \code{millisec}
arguments are not used.}

\item{\code{mz}}{A vector with target \emph{m/z} value/s (in Da) or a two columns data.table or data.frame named \code{mzmin} and
\code{mzmax} with minimum and maximum \emph{m/z} values (in Da), respectively. Alternatively, \emph{m/z} (in Da) and
retention time values (in seconds) can be given as one data.table or data.frame with columns named \code{mz} and \code{rt}
and/or \code{drift}. Then, the deviations given in the \code{ppm}, \code{sec} and \code{millisec} arguments are used to
calculate the ranges. Also works with a data.table or data.frame with minimum and maximum values of \emph{m/z},
retention time and drift time targets. In this case, the column names must be \code{mzmin}, \code{mzmax}, \code{rtmin}, \code{rtmax},
\code{driftmin} and \code{driftmax}. Note that when mass/time ranges are given, the \code{ppm}, \code{sec} and \code{millisec}
arguments are not used.}

\item{\code{rt}}{A vector with target retention time values (in seconds) or
a two columns data.table or data.frame with minimum and maximum retention
time values (in seconds).}

\item{\code{mobility}}{A vector with target drift time values (in milliseconds) or
a two columns data.table or data.frame with minimum and maximum drift
time values (in milliseconds).}

\item{\code{ppm}}{Numeric of length one with the mass deviation, in ppm.}

\item{\code{sec}}{Numeric of length one with the time deviation, in seconds.}

\item{\code{millisec}}{Numeric of length one with the drift time deviation,
in milliseconds.}

\item{\code{filtered}}{Logical (length 1). When \code{TRUE}, filtered features/groups are considered.}

\item{\code{groupBy}}{Column to use for color grouping (default: "analysis").}

\item{\code{normalized}}{A logical value indicating whether the result should be normalized.}

\item{\code{yLab}}{A string with the title for the y axis.}

\item{\code{yLab}}{A string with the title for the y axis.}

\item{\code{title}}{A string with the title.}

\item{\code{interactive}}{Logical (length 1). When \code{TRUE}, the data is plotted interactively using \pkg{plotly}.}

\item{\code{showLegend}}{Logical (length 1). Set to \code{TRUE} to show legend.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectNonTargetAnalysis-plot_features"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectNonTargetAnalysis-plot_features}{}}}
\subsection{Method \code{plot_features()}}{
Plot EIC traces for selected features.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectNonTargetAnalysis$plot_features(
  analyses = NULL,
  features = NULL,
  groups = NULL,
  components = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  filtered = FALSE,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  groupBy = "feature",
  interactive = TRUE,
  showDetails = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{Character or numeric vector with names or indexes of analyses in the \code{Analyses}
object.}

\item{\code{features}}{A character vector with the feature IDs or a data.frame or data.table with a column named \code{feature}
with the feature IDs. A column named \code{analysis} can be given to specify from which analyses the features matched.}

\item{\code{groups}}{A numeric or character vector with feature group IDs, or a data.frame/data.table with a \code{feature_group} (or \code{group}) column; an optional \code{analysis} column restricts analyses and a \code{name} column sets custom labels.}

\item{\code{components}}{A numeric or character vector with feature component IDs, or a data.frame/data.table with a \code{feature_component} (or \code{component}) column; an optional \code{analysis} column restricts analyses and a \code{name} column sets custom labels.}

\item{\code{mass}}{A vector with target neutral mass value/s (in Da) or a two columns data.table or data.frame named \code{min}
and \code{max} with minimum and maximum neutral mass values (in Da), respectively. Alternatively, neutral mass (in Da)
and retention time (in seconds) and/or drift time values (in milliseconds) can be given as one data.table or
data.frame with columns named \code{mass} and \code{rt} and/or \code{drift}. Then, the deviations given in the \code{ppm}, \code{sec}
and \code{millisec}  arguments are used to calculate the ranges. Also works with a data.table or data.frame with
minimum and maximum values of neutral mass, retention time and drift time targets. In this case, the column names
must be \code{min}, \code{max}, \code{rtmin}, \code{rtmax}, \code{driftmin} and \code{driftmax}. Note that when mass/time ranges are given, the
\code{ppm}, \code{sec} and \code{millisec}
arguments are not used.}

\item{\code{mz}}{A vector with target \emph{m/z} value/s (in Da) or a two columns data.table or data.frame named \code{mzmin} and
\code{mzmax} with minimum and maximum \emph{m/z} values (in Da), respectively. Alternatively, \emph{m/z} (in Da) and
retention time values (in seconds) can be given as one data.table or data.frame with columns named \code{mz} and \code{rt}
and/or \code{drift}. Then, the deviations given in the \code{ppm}, \code{sec} and \code{millisec} arguments are used to
calculate the ranges. Also works with a data.table or data.frame with minimum and maximum values of \emph{m/z},
retention time and drift time targets. In this case, the column names must be \code{mzmin}, \code{mzmax}, \code{rtmin}, \code{rtmax},
\code{driftmin} and \code{driftmax}. Note that when mass/time ranges are given, the \code{ppm}, \code{sec} and \code{millisec}
arguments are not used.}

\item{\code{rt}}{A vector with target retention time values (in seconds) or
a two columns data.table or data.frame with minimum and maximum retention
time values (in seconds).}

\item{\code{mobility}}{A vector with target drift time values (in milliseconds) or
a two columns data.table or data.frame with minimum and maximum drift
time values (in milliseconds).}

\item{\code{ppm}}{Numeric of length one with the mass deviation, in ppm.}

\item{\code{sec}}{Numeric of length one with the time deviation, in seconds.}

\item{\code{millisec}}{Numeric of length one with the drift time deviation,
in milliseconds.}

\item{\code{filtered}}{Logical (length 1). When \code{TRUE}, filtered features/groups are considered.}

\item{\code{xLab}}{A string with the title for the x axis.}

\item{\code{yLab}}{A string with the title for the y axis.}

\item{\code{yLab}}{A string with the title for the y axis.}

\item{\code{title}}{A string with the title.}

\item{\code{groupBy}}{Column to use for color grouping (default: "analysis").}

\item{\code{interactive}}{Logical (length 1). When \code{TRUE}, the data is plotted interactively using \pkg{plotly}.}

\item{\code{showDetails}}{Logical (length 1), show hover details in interactive plots.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectNonTargetAnalysis-map_features"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectNonTargetAnalysis-map_features}{}}}
\subsection{Method \code{map_features()}}{
Plot RT versus m/z traces for selected features.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectNonTargetAnalysis$map_features(
  analyses = NULL,
  features = NULL,
  groups = NULL,
  components = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  filtered = FALSE,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  groupBy = "feature",
  globalNormalization = FALSE,
  interactive = TRUE,
  showDetails = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{Character or numeric vector with names or indexes of analyses in the \code{Analyses}
object.}

\item{\code{features}}{A character vector with the feature IDs or a data.frame or data.table with a column named \code{feature}
with the feature IDs. A column named \code{analysis} can be given to specify from which analyses the features matched.}

\item{\code{groups}}{A numeric or character vector with feature group IDs, or a data.frame/data.table with a \code{feature_group} (or \code{group}) column; an optional \code{analysis} column restricts analyses and a \code{name} column sets custom labels.}

\item{\code{components}}{A numeric or character vector with feature component IDs, or a data.frame/data.table with a \code{feature_component} (or \code{component}) column; an optional \code{analysis} column restricts analyses and a \code{name} column sets custom labels.}

\item{\code{mass}}{A vector with target neutral mass value/s (in Da) or a two columns data.table or data.frame named \code{min}
and \code{max} with minimum and maximum neutral mass values (in Da), respectively. Alternatively, neutral mass (in Da)
and retention time (in seconds) and/or drift time values (in milliseconds) can be given as one data.table or
data.frame with columns named \code{mass} and \code{rt} and/or \code{drift}. Then, the deviations given in the \code{ppm}, \code{sec}
and \code{millisec}  arguments are used to calculate the ranges. Also works with a data.table or data.frame with
minimum and maximum values of neutral mass, retention time and drift time targets. In this case, the column names
must be \code{min}, \code{max}, \code{rtmin}, \code{rtmax}, \code{driftmin} and \code{driftmax}. Note that when mass/time ranges are given, the
\code{ppm}, \code{sec} and \code{millisec}
arguments are not used.}

\item{\code{mz}}{A vector with target \emph{m/z} value/s (in Da) or a two columns data.table or data.frame named \code{mzmin} and
\code{mzmax} with minimum and maximum \emph{m/z} values (in Da), respectively. Alternatively, \emph{m/z} (in Da) and
retention time values (in seconds) can be given as one data.table or data.frame with columns named \code{mz} and \code{rt}
and/or \code{drift}. Then, the deviations given in the \code{ppm}, \code{sec} and \code{millisec} arguments are used to
calculate the ranges. Also works with a data.table or data.frame with minimum and maximum values of \emph{m/z},
retention time and drift time targets. In this case, the column names must be \code{mzmin}, \code{mzmax}, \code{rtmin}, \code{rtmax},
\code{driftmin} and \code{driftmax}. Note that when mass/time ranges are given, the \code{ppm}, \code{sec} and \code{millisec}
arguments are not used.}

\item{\code{rt}}{A vector with target retention time values (in seconds) or
a two columns data.table or data.frame with minimum and maximum retention
time values (in seconds).}

\item{\code{mobility}}{A vector with target drift time values (in milliseconds) or
a two columns data.table or data.frame with minimum and maximum drift
time values (in milliseconds).}

\item{\code{ppm}}{Numeric of length one with the mass deviation, in ppm.}

\item{\code{sec}}{Numeric of length one with the time deviation, in seconds.}

\item{\code{millisec}}{Numeric of length one with the drift time deviation,
in milliseconds.}

\item{\code{filtered}}{Logical (length 1). When \code{TRUE}, filtered features/groups are considered.}

\item{\code{xLab}}{A string with the title for the x axis.}

\item{\code{yLab}}{A string with the title for the y axis.}

\item{\code{yLab}}{A string with the title for the y axis.}

\item{\code{title}}{A string with the title.}

\item{\code{groupBy}}{Column to use for color grouping (default: "analysis").}

\item{\code{globalNormalization}}{Logical, when TRUE normalize intensities globally across all selected features.}

\item{\code{interactive}}{Logical (length 1). When \code{TRUE}, the data is plotted interactively using \pkg{plotly}.}

\item{\code{showDetails}}{Logical (length 1), show hover details in interactive plots.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectNonTargetAnalysis-plot_features_ms1"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectNonTargetAnalysis-plot_features_ms1}{}}}
\subsection{Method \code{plot_features_ms1()}}{
Plot MS1 spectra for selected features.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectNonTargetAnalysis$plot_features_ms1(
  analyses = NULL,
  features = NULL,
  groups = NULL,
  components = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  normalized = FALSE,
  filtered = FALSE,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  groupBy = "feature",
  showText = TRUE,
  interactive = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{Character or numeric vector with names or indexes of analyses in the \code{Analyses}
object.}

\item{\code{features}}{A character vector with the feature IDs or a data.frame or data.table with a column named \code{feature}
with the feature IDs. A column named \code{analysis} can be given to specify from which analyses the features matched.}

\item{\code{groups}}{A numeric or character vector with feature group IDs, or a data.frame/data.table with a \code{feature_group} (or \code{group}) column; an optional \code{analysis} column restricts analyses and a \code{name} column sets custom labels.}

\item{\code{components}}{A numeric or character vector with feature component IDs, or a data.frame/data.table with a \code{feature_component} (or \code{component}) column; an optional \code{analysis} column restricts analyses and a \code{name} column sets custom labels.}

\item{\code{mass}}{A vector with target neutral mass value/s (in Da) or a two columns data.table or data.frame named \code{min}
and \code{max} with minimum and maximum neutral mass values (in Da), respectively. Alternatively, neutral mass (in Da)
and retention time (in seconds) and/or drift time values (in milliseconds) can be given as one data.table or
data.frame with columns named \code{mass} and \code{rt} and/or \code{drift}. Then, the deviations given in the \code{ppm}, \code{sec}
and \code{millisec}  arguments are used to calculate the ranges. Also works with a data.table or data.frame with
minimum and maximum values of neutral mass, retention time and drift time targets. In this case, the column names
must be \code{min}, \code{max}, \code{rtmin}, \code{rtmax}, \code{driftmin} and \code{driftmax}. Note that when mass/time ranges are given, the
\code{ppm}, \code{sec} and \code{millisec}
arguments are not used.}

\item{\code{mz}}{A vector with target \emph{m/z} value/s (in Da) or a two columns data.table or data.frame named \code{mzmin} and
\code{mzmax} with minimum and maximum \emph{m/z} values (in Da), respectively. Alternatively, \emph{m/z} (in Da) and
retention time values (in seconds) can be given as one data.table or data.frame with columns named \code{mz} and \code{rt}
and/or \code{drift}. Then, the deviations given in the \code{ppm}, \code{sec} and \code{millisec} arguments are used to
calculate the ranges. Also works with a data.table or data.frame with minimum and maximum values of \emph{m/z},
retention time and drift time targets. In this case, the column names must be \code{mzmin}, \code{mzmax}, \code{rtmin}, \code{rtmax},
\code{driftmin} and \code{driftmax}. Note that when mass/time ranges are given, the \code{ppm}, \code{sec} and \code{millisec}
arguments are not used.}

\item{\code{rt}}{A vector with target retention time values (in seconds) or
a two columns data.table or data.frame with minimum and maximum retention
time values (in seconds).}

\item{\code{mobility}}{A vector with target drift time values (in milliseconds) or
a two columns data.table or data.frame with minimum and maximum drift
time values (in milliseconds).}

\item{\code{ppm}}{Numeric of length one with the mass deviation, in ppm.}

\item{\code{sec}}{Numeric of length one with the time deviation, in seconds.}

\item{\code{millisec}}{Numeric of length one with the drift time deviation,
in milliseconds.}

\item{\code{normalized}}{A logical value indicating whether the result should be normalized.}

\item{\code{filtered}}{Logical (length 1). When \code{TRUE}, filtered features/groups are considered.}

\item{\code{xLab}}{A string with the title for the x axis.}

\item{\code{yLab}}{A string with the title for the y axis.}

\item{\code{yLab}}{A string with the title for the y axis.}

\item{\code{title}}{A string with the title.}

\item{\code{groupBy}}{Column to use for color grouping (default: "analysis").}

\item{\code{showText}}{Logical (length 1), set to \code{TRUE} to show the text annotations.}

\item{\code{interactive}}{Logical (length 1). When \code{TRUE}, the data is plotted interactively using \pkg{plotly}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectNonTargetAnalysis-plot_features_ms2"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectNonTargetAnalysis-plot_features_ms2}{}}}
\subsection{Method \code{plot_features_ms2()}}{
Plot MS2 spectra for selected features.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectNonTargetAnalysis$plot_features_ms2(
  analyses = NULL,
  features = NULL,
  groups = NULL,
  components = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  normalized = TRUE,
  filtered = FALSE,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  groupBy = "feature",
  showText = TRUE,
  interactive = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{Character or numeric vector with names or indexes of analyses in the \code{Analyses}
object.}

\item{\code{features}}{A character vector with the feature IDs or a data.frame or data.table with a column named \code{feature}
with the feature IDs. A column named \code{analysis} can be given to specify from which analyses the features matched.}

\item{\code{groups}}{A numeric or character vector with feature group IDs, or a data.frame/data.table with a \code{feature_group} (or \code{group}) column; an optional \code{analysis} column restricts analyses and a \code{name} column sets custom labels.}

\item{\code{components}}{A numeric or character vector with feature component IDs, or a data.frame/data.table with a \code{feature_component} (or \code{component}) column; an optional \code{analysis} column restricts analyses and a \code{name} column sets custom labels.}

\item{\code{mass}}{A vector with target neutral mass value/s (in Da) or a two columns data.table or data.frame named \code{min}
and \code{max} with minimum and maximum neutral mass values (in Da), respectively. Alternatively, neutral mass (in Da)
and retention time (in seconds) and/or drift time values (in milliseconds) can be given as one data.table or
data.frame with columns named \code{mass} and \code{rt} and/or \code{drift}. Then, the deviations given in the \code{ppm}, \code{sec}
and \code{millisec}  arguments are used to calculate the ranges. Also works with a data.table or data.frame with
minimum and maximum values of neutral mass, retention time and drift time targets. In this case, the column names
must be \code{min}, \code{max}, \code{rtmin}, \code{rtmax}, \code{driftmin} and \code{driftmax}. Note that when mass/time ranges are given, the
\code{ppm}, \code{sec} and \code{millisec}
arguments are not used.}

\item{\code{mz}}{A vector with target \emph{m/z} value/s (in Da) or a two columns data.table or data.frame named \code{mzmin} and
\code{mzmax} with minimum and maximum \emph{m/z} values (in Da), respectively. Alternatively, \emph{m/z} (in Da) and
retention time values (in seconds) can be given as one data.table or data.frame with columns named \code{mz} and \code{rt}
and/or \code{drift}. Then, the deviations given in the \code{ppm}, \code{sec} and \code{millisec} arguments are used to
calculate the ranges. Also works with a data.table or data.frame with minimum and maximum values of \emph{m/z},
retention time and drift time targets. In this case, the column names must be \code{mzmin}, \code{mzmax}, \code{rtmin}, \code{rtmax},
\code{driftmin} and \code{driftmax}. Note that when mass/time ranges are given, the \code{ppm}, \code{sec} and \code{millisec}
arguments are not used.}

\item{\code{rt}}{A vector with target retention time values (in seconds) or
a two columns data.table or data.frame with minimum and maximum retention
time values (in seconds).}

\item{\code{mobility}}{A vector with target drift time values (in milliseconds) or
a two columns data.table or data.frame with minimum and maximum drift
time values (in milliseconds).}

\item{\code{ppm}}{Numeric of length one with the mass deviation, in ppm.}

\item{\code{sec}}{Numeric of length one with the time deviation, in seconds.}

\item{\code{millisec}}{Numeric of length one with the drift time deviation,
in milliseconds.}

\item{\code{normalized}}{A logical value indicating whether the result should be normalized.}

\item{\code{filtered}}{Logical (length 1). When \code{TRUE}, filtered features/groups are considered.}

\item{\code{xLab}}{A string with the title for the x axis.}

\item{\code{yLab}}{A string with the title for the y axis.}

\item{\code{yLab}}{A string with the title for the y axis.}

\item{\code{title}}{A string with the title.}

\item{\code{groupBy}}{Column to use for color grouping (default: "analysis").}

\item{\code{showText}}{Logical (length 1), set to \code{TRUE} to show the text annotations.}

\item{\code{interactive}}{Logical (length 1). When \code{TRUE}, the data is plotted interactively using \pkg{plotly}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectNonTargetAnalysis-plot_suspects_ms2"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectNonTargetAnalysis-plot_suspects_ms2}{}}}
\subsection{Method \code{plot_suspects_ms2()}}{
Plot suspect MS2 spectra for selected features.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectNonTargetAnalysis$plot_suspects_ms2(
  analyses = NULL,
  features = NULL,
  groups = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  normalized = TRUE,
  filtered = FALSE,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  groupBy = c("feature", "name"),
  showText = TRUE,
  interactive = TRUE,
  showLegend = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{Character or numeric vector with names or indexes of analyses in the \code{Analyses}
object.}

\item{\code{features}}{A character vector with the feature IDs or a data.frame or data.table with a column named \code{feature}
with the feature IDs. A column named \code{analysis} can be given to specify from which analyses the features matched.}

\item{\code{groups}}{A numeric or character vector with feature group IDs, or a data.frame/data.table with a \code{feature_group} (or \code{group}) column; an optional \code{analysis} column restricts analyses and a \code{name} column sets custom labels.}

\item{\code{mass}}{A vector with target neutral mass value/s (in Da) or a two columns data.table or data.frame named \code{min}
and \code{max} with minimum and maximum neutral mass values (in Da), respectively. Alternatively, neutral mass (in Da)
and retention time (in seconds) and/or drift time values (in milliseconds) can be given as one data.table or
data.frame with columns named \code{mass} and \code{rt} and/or \code{drift}. Then, the deviations given in the \code{ppm}, \code{sec}
and \code{millisec}  arguments are used to calculate the ranges. Also works with a data.table or data.frame with
minimum and maximum values of neutral mass, retention time and drift time targets. In this case, the column names
must be \code{min}, \code{max}, \code{rtmin}, \code{rtmax}, \code{driftmin} and \code{driftmax}. Note that when mass/time ranges are given, the
\code{ppm}, \code{sec} and \code{millisec}
arguments are not used.}

\item{\code{mz}}{A vector with target \emph{m/z} value/s (in Da) or a two columns data.table or data.frame named \code{mzmin} and
\code{mzmax} with minimum and maximum \emph{m/z} values (in Da), respectively. Alternatively, \emph{m/z} (in Da) and
retention time values (in seconds) can be given as one data.table or data.frame with columns named \code{mz} and \code{rt}
and/or \code{drift}. Then, the deviations given in the \code{ppm}, \code{sec} and \code{millisec} arguments are used to
calculate the ranges. Also works with a data.table or data.frame with minimum and maximum values of \emph{m/z},
retention time and drift time targets. In this case, the column names must be \code{mzmin}, \code{mzmax}, \code{rtmin}, \code{rtmax},
\code{driftmin} and \code{driftmax}. Note that when mass/time ranges are given, the \code{ppm}, \code{sec} and \code{millisec}
arguments are not used.}

\item{\code{rt}}{A vector with target retention time values (in seconds) or
a two columns data.table or data.frame with minimum and maximum retention
time values (in seconds).}

\item{\code{mobility}}{A vector with target drift time values (in milliseconds) or
a two columns data.table or data.frame with minimum and maximum drift
time values (in milliseconds).}

\item{\code{ppm}}{Numeric of length one with the mass deviation, in ppm.}

\item{\code{sec}}{Numeric of length one with the time deviation, in seconds.}

\item{\code{millisec}}{Numeric of length one with the drift time deviation,
in milliseconds.}

\item{\code{normalized}}{A logical value indicating whether the result should be normalized.}

\item{\code{filtered}}{Logical (length 1). When \code{TRUE}, filtered features/groups are considered.}

\item{\code{xLab}}{A string with the title for the x axis.}

\item{\code{yLab}}{A string with the title for the y axis.}

\item{\code{yLab}}{A string with the title for the y axis.}

\item{\code{title}}{A string with the title.}

\item{\code{groupBy}}{Column to use for color grouping (default: "analysis").}

\item{\code{showText}}{Logical (length 1), set to \code{TRUE} to show the text annotations.}

\item{\code{interactive}}{Logical (length 1). When \code{TRUE}, the data is plotted interactively using \pkg{plotly}.}

\item{\code{showLegend}}{Logical (length 1). Set to \code{TRUE} to show legend.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectNonTargetAnalysis-get_fold_change"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectNonTargetAnalysis-get_fold_change}{}}}
\subsection{Method \code{get_fold_change()}}{
Return fold-change categories between replicate groups.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectNonTargetAnalysis$get_fold_change(
  replicatesIn = NULL,
  replicatesOut = NULL,
  groups = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 4,
  sec = 10,
  millisec = 5,
  filtered = FALSE,
  constantThreshold = 0.5,
  eliminationThreshold = 0.2,
  correctIntensity = FALSE,
  fillZerosWithLowerLimit = FALSE,
  lowerLimit = NA_real_
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{replicatesIn}}{Character vector with replicate names used as denominator.}

\item{\code{replicatesOut}}{Character vector with replicate names used as numerator.}

\item{\code{groups}}{A numeric or character vector with feature group IDs, or a data.frame/data.table with a \code{feature_group} (or \code{group}) column; an optional \code{analysis} column restricts analyses and a \code{name} column sets custom labels.}

\item{\code{mass}}{A vector with target neutral mass value/s (in Da) or a two columns data.table or data.frame named \code{min}
and \code{max} with minimum and maximum neutral mass values (in Da), respectively. Alternatively, neutral mass (in Da)
and retention time (in seconds) and/or drift time values (in milliseconds) can be given as one data.table or
data.frame with columns named \code{mass} and \code{rt} and/or \code{drift}. Then, the deviations given in the \code{ppm}, \code{sec}
and \code{millisec}  arguments are used to calculate the ranges. Also works with a data.table or data.frame with
minimum and maximum values of neutral mass, retention time and drift time targets. In this case, the column names
must be \code{min}, \code{max}, \code{rtmin}, \code{rtmax}, \code{driftmin} and \code{driftmax}. Note that when mass/time ranges are given, the
\code{ppm}, \code{sec} and \code{millisec}
arguments are not used.}

\item{\code{mz}}{A vector with target \emph{m/z} value/s (in Da) or a two columns data.table or data.frame named \code{mzmin} and
\code{mzmax} with minimum and maximum \emph{m/z} values (in Da), respectively. Alternatively, \emph{m/z} (in Da) and
retention time values (in seconds) can be given as one data.table or data.frame with columns named \code{mz} and \code{rt}
and/or \code{drift}. Then, the deviations given in the \code{ppm}, \code{sec} and \code{millisec} arguments are used to
calculate the ranges. Also works with a data.table or data.frame with minimum and maximum values of \emph{m/z},
retention time and drift time targets. In this case, the column names must be \code{mzmin}, \code{mzmax}, \code{rtmin}, \code{rtmax},
\code{driftmin} and \code{driftmax}. Note that when mass/time ranges are given, the \code{ppm}, \code{sec} and \code{millisec}
arguments are not used.}

\item{\code{rt}}{A vector with target retention time values (in seconds) or
a two columns data.table or data.frame with minimum and maximum retention
time values (in seconds).}

\item{\code{mobility}}{A vector with target drift time values (in milliseconds) or
a two columns data.table or data.frame with minimum and maximum drift
time values (in milliseconds).}

\item{\code{ppm}}{Numeric of length one with the mass deviation, in ppm.}

\item{\code{sec}}{Numeric of length one with the time deviation, in seconds.}

\item{\code{millisec}}{Numeric of length one with the drift time deviation,
in milliseconds.}

\item{\code{filtered}}{Logical (length 1). When \code{TRUE}, filtered features/groups are considered.}

\item{\code{constantThreshold}}{Numeric threshold used to mark features as constant.}

\item{\code{eliminationThreshold}}{Numeric threshold used to mark features as eliminated.}

\item{\code{correctIntensity}}{Logical (length 1) with \code{TRUE} for correcting the intensity for matrix
suppression before applying intensity based filters. Note that suppression factor as obtained by
the processing method \code{CorrectMatrixSuppression} must be present. When not available the
correction is ignored even if set to \code{TRUE}.}

\item{\code{fillZerosWithLowerLimit}}{Logical; replace zeros before fold-change calculation.}

\item{\code{lowerLimit}}{Optional lower limit used when filling zeros.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectNonTargetAnalysis-plot_fold_change"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectNonTargetAnalysis-plot_fold_change}{}}}
\subsection{Method \code{plot_fold_change()}}{
Plot fold-change categories between replicate groups.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectNonTargetAnalysis$plot_fold_change(
  replicatesIn = NULL,
  replicatesOut = NULL,
  groups = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 4,
  sec = 10,
  millisec = 5,
  filtered = FALSE,
  constantThreshold = 0.5,
  eliminationThreshold = 0.2,
  correctIntensity = FALSE,
  fillZerosWithLowerLimit = FALSE,
  lowerLimit = NA_real_,
  normalized = TRUE,
  yLab = NULL,
  title = NULL,
  interactive = TRUE,
  showLegend = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{replicatesIn}}{Character vector with replicate names used as denominator.}

\item{\code{replicatesOut}}{Character vector with replicate names used as numerator.}

\item{\code{groups}}{A numeric or character vector with feature group IDs, or a data.frame/data.table with a \code{feature_group} (or \code{group}) column; an optional \code{analysis} column restricts analyses and a \code{name} column sets custom labels.}

\item{\code{mass}}{A vector with target neutral mass value/s (in Da) or a two columns data.table or data.frame named \code{min}
and \code{max} with minimum and maximum neutral mass values (in Da), respectively. Alternatively, neutral mass (in Da)
and retention time (in seconds) and/or drift time values (in milliseconds) can be given as one data.table or
data.frame with columns named \code{mass} and \code{rt} and/or \code{drift}. Then, the deviations given in the \code{ppm}, \code{sec}
and \code{millisec}  arguments are used to calculate the ranges. Also works with a data.table or data.frame with
minimum and maximum values of neutral mass, retention time and drift time targets. In this case, the column names
must be \code{min}, \code{max}, \code{rtmin}, \code{rtmax}, \code{driftmin} and \code{driftmax}. Note that when mass/time ranges are given, the
\code{ppm}, \code{sec} and \code{millisec}
arguments are not used.}

\item{\code{mz}}{A vector with target \emph{m/z} value/s (in Da) or a two columns data.table or data.frame named \code{mzmin} and
\code{mzmax} with minimum and maximum \emph{m/z} values (in Da), respectively. Alternatively, \emph{m/z} (in Da) and
retention time values (in seconds) can be given as one data.table or data.frame with columns named \code{mz} and \code{rt}
and/or \code{drift}. Then, the deviations given in the \code{ppm}, \code{sec} and \code{millisec} arguments are used to
calculate the ranges. Also works with a data.table or data.frame with minimum and maximum values of \emph{m/z},
retention time and drift time targets. In this case, the column names must be \code{mzmin}, \code{mzmax}, \code{rtmin}, \code{rtmax},
\code{driftmin} and \code{driftmax}. Note that when mass/time ranges are given, the \code{ppm}, \code{sec} and \code{millisec}
arguments are not used.}

\item{\code{rt}}{A vector with target retention time values (in seconds) or
a two columns data.table or data.frame with minimum and maximum retention
time values (in seconds).}

\item{\code{mobility}}{A vector with target drift time values (in milliseconds) or
a two columns data.table or data.frame with minimum and maximum drift
time values (in milliseconds).}

\item{\code{ppm}}{Numeric of length one with the mass deviation, in ppm.}

\item{\code{sec}}{Numeric of length one with the time deviation, in seconds.}

\item{\code{millisec}}{Numeric of length one with the drift time deviation,
in milliseconds.}

\item{\code{filtered}}{Logical (length 1). When \code{TRUE}, filtered features/groups are considered.}

\item{\code{constantThreshold}}{Numeric threshold used to mark features as constant.}

\item{\code{eliminationThreshold}}{Numeric threshold used to mark features as eliminated.}

\item{\code{correctIntensity}}{Logical (length 1) with \code{TRUE} for correcting the intensity for matrix
suppression before applying intensity based filters. Note that suppression factor as obtained by
the processing method \code{CorrectMatrixSuppression} must be present. When not available the
correction is ignored even if set to \code{TRUE}.}

\item{\code{fillZerosWithLowerLimit}}{Logical; replace zeros before fold-change calculation.}

\item{\code{lowerLimit}}{Optional lower limit used when filling zeros.}

\item{\code{normalized}}{A logical value indicating whether the result should be normalized.}

\item{\code{yLab}}{A string with the title for the y axis.}

\item{\code{yLab}}{A string with the title for the y axis.}

\item{\code{title}}{A string with the title.}

\item{\code{interactive}}{Logical (length 1). When \code{TRUE}, the data is plotted interactively using \pkg{plotly}.}

\item{\code{showLegend}}{Logical (length 1). Set to \code{TRUE} to show legend.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectNonTargetAnalysis-plot_transformation_products"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectNonTargetAnalysis-plot_transformation_products}{}}}
\subsection{Method \code{plot_transformation_products()}}{
Plot a transformation-products network.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectNonTargetAnalysis$plot_transformation_products(
  groups = NULL,
  showMS2 = FALSE,
  showIntensityProfile = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{groups}}{A numeric or character vector with feature group IDs, or a data.frame/data.table with a \code{feature_group} (or \code{group}) column; an optional \code{analysis} column restricts analyses and a \code{name} column sets custom labels.}

\item{\code{showMS2}}{Logical; include MS2 annotations in transformation product plots.}

\item{\code{showIntensityProfile}}{Logical; include intensity profile in transformation product plots.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectNonTargetAnalysis-print"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectNonTargetAnalysis-print}{}}}
\subsection{Method \code{print()}}{
Print a short summary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectNonTargetAnalysis$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments passed to the method or ignored by the implementation.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectNonTargetAnalysis-show"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectNonTargetAnalysis-show}{}}}
\subsection{Method \code{show()}}{
Show a short summary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectNonTargetAnalysis$show(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments passed to the method or ignored by the implementation.}
}
\if{html}{\out{</div>}}
}
}
}
