% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_class_R6MS.R
\name{R6MS}
\alias{R6MS}
\title{R6 Class Representing an MS Set of Analysis Files}
\description{
A description of the R6 class.
}
\details{
The details.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-R6MS-new}{\code{R6MS$new()}}
\item \href{#method-R6MS-print}{\code{R6MS$print()}}
\item \href{#method-R6MS-get_header}{\code{R6MS$get_header()}}
\item \href{#method-R6MS-get_analyses}{\code{R6MS$get_analyses()}}
\item \href{#method-R6MS-get_number_analyses}{\code{R6MS$get_number_analyses()}}
\item \href{#method-R6MS-get_overview}{\code{R6MS$get_overview()}}
\item \href{#method-R6MS-get_analysis_names}{\code{R6MS$get_analysis_names()}}
\item \href{#method-R6MS-get_replicate_names}{\code{R6MS$get_replicate_names()}}
\item \href{#method-R6MS-get_blank_names}{\code{R6MS$get_blank_names()}}
\item \href{#method-R6MS-get_polarities}{\code{R6MS$get_polarities()}}
\item \href{#method-R6MS-get_file_paths}{\code{R6MS$get_file_paths()}}
\item \href{#method-R6MS-get_spectra}{\code{R6MS$get_spectra()}}
\item \href{#method-R6MS-get_chromatograms}{\code{R6MS$get_chromatograms()}}
\item \href{#method-R6MS-get_tic}{\code{R6MS$get_tic()}}
\item \href{#method-R6MS-get_bpc}{\code{R6MS$get_bpc()}}
\item \href{#method-R6MS-get_eic}{\code{R6MS$get_eic()}}
\item \href{#method-R6MS-get_ms2}{\code{R6MS$get_ms2()}}
\item \href{#method-R6MS-set_replicate_names}{\code{R6MS$set_replicate_names()}}
\item \href{#method-R6MS-set_blank_names}{\code{R6MS$set_blank_names()}}
\item \href{#method-R6MS-has_loaded_spectra}{\code{R6MS$has_loaded_spectra()}}
\item \href{#method-R6MS-has_loaded_chromatograms}{\code{R6MS$has_loaded_chromatograms()}}
\item \href{#method-R6MS-add_analyses}{\code{R6MS$add_analyses()}}
\item \href{#method-R6MS-load_spectra}{\code{R6MS$load_spectra()}}
\item \href{#method-R6MS-load_chromatograms}{\code{R6MS$load_chromatograms()}}
\item \href{#method-R6MS-plot_spectra}{\code{R6MS$plot_spectra()}}
\item \href{#method-R6MS-plot_xic}{\code{R6MS$plot_xic()}}
\item \href{#method-R6MS-plot_eic}{\code{R6MS$plot_eic()}}
\item \href{#method-R6MS-plot_tic}{\code{R6MS$plot_tic()}}
\item \href{#method-R6MS-plot_bpc}{\code{R6MS$plot_bpc()}}
\item \href{#method-R6MS-plot_ms2}{\code{R6MS$plot_ms2()}}
\item \href{#method-R6MS-subset_analyses}{\code{R6MS$subset_analyses()}}
\item \href{#method-R6MS-check_analyses_argument}{\code{R6MS$check_analyses_argument()}}
\item \href{#method-R6MS-clone}{\code{R6MS$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6MS-new"></a>}}
\if{latex}{\out{\hypertarget{method-R6MS-new}{}}}
\subsection{Method \code{new()}}{
Create a new MS set of analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6MS$new(
  files = NULL,
  header = NULL,
  analyses = NULL,
  features = NULL,
  run_parallel = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{files}}{A character vector with the full file path of MS analyses
or a data.frame with the columns file, replicate and blank with the
full file path, the replicate group name (string) and the associated
blank replicate group (string).}

\item{\code{header}}{A list with administrative information.}

\item{\code{analyses}}{A list with MS analyses information.}

\item{\code{features}}{A data.frame with features representing corresponding
peaks across MS analyses.}

\item{\code{run_parallel}}{Logical, set to \code{TRUE} for processing the data
in parallel.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `r6MS` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6MS-print"></a>}}
\if{latex}{\out{\hypertarget{method-R6MS-print}{}}}
\subsection{Method \code{print()}}{
Prints a summary of the `R6MS` object in the console.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6MS$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Console text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6MS-get_header"></a>}}
\if{latex}{\out{\hypertarget{method-R6MS-get_header}{}}}
\subsection{Method \code{get_header()}}{
Method to get the header of the `R6MS` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6MS$get_header(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A character vector with the name of the entry.
Possible values are name, author, description, path and date.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The header list as defined by `value`.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6MS-get_analyses"></a>}}
\if{latex}{\out{\hypertarget{method-R6MS-get_analyses}{}}}
\subsection{Method \code{get_analyses()}}{
Method to get specific analyses from the `R6MS` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6MS$get_analyses(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A numeric/character vector with the number/name
of the analyses.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The list of analyses defined by `value`.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6MS-get_number_analyses"></a>}}
\if{latex}{\out{\hypertarget{method-R6MS-get_number_analyses}{}}}
\subsection{Method \code{get_number_analyses()}}{
Method to get the number of analysis in the `R6MS` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6MS$get_number_analyses()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An integer value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6MS-get_overview"></a>}}
\if{latex}{\out{\hypertarget{method-R6MS-get_overview}{}}}
\subsection{Method \code{get_overview()}}{
Method to get the overview data.frame with all the analysis types,
names, replicates, associated blank replicates, polarities and full
file paths.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6MS$get_overview()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data.frame with columns type, analysis, replicate, blank
polarity and file.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6MS-get_analysis_names"></a>}}
\if{latex}{\out{\hypertarget{method-R6MS-get_analysis_names}{}}}
\subsection{Method \code{get_analysis_names()}}{
Method to get the analysis names.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6MS$get_analysis_names(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A numeric/character vector with the number/name
of the analyses.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6MS-get_replicate_names"></a>}}
\if{latex}{\out{\hypertarget{method-R6MS-get_replicate_names}{}}}
\subsection{Method \code{get_replicate_names()}}{
Method to get the analysis replicate names.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6MS$get_replicate_names(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A numeric/character vector with the number/name
of the analyses.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6MS-get_blank_names"></a>}}
\if{latex}{\out{\hypertarget{method-R6MS-get_blank_names}{}}}
\subsection{Method \code{get_blank_names()}}{
Method to get the analysis blank replicate names.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6MS$get_blank_names(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A numeric/character vector with the number/name
of the analyses.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6MS-get_polarities"></a>}}
\if{latex}{\out{\hypertarget{method-R6MS-get_polarities}{}}}
\subsection{Method \code{get_polarities()}}{
Method to get the polarity of the analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6MS$get_polarities(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A numeric/character vector with the number/name
of the analyses.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6MS-get_file_paths"></a>}}
\if{latex}{\out{\hypertarget{method-R6MS-get_file_paths}{}}}
\subsection{Method \code{get_file_paths()}}{
Method to get the polarity of the analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6MS$get_file_paths(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A numeric/character vector with the number/name
of the analyses.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6MS-get_spectra"></a>}}
\if{latex}{\out{\hypertarget{method-R6MS-get_spectra}{}}}
\subsection{Method \code{get_spectra()}}{
Method to get spectra from the MS analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6MS$get_spectra(
  analyses = NULL,
  levels = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  id = NULL,
  allTraces = TRUE,
  isolationWindow = 1.3,
  minIntensityMS1 = 0,
  minIntensityMS2 = 0,
  run_parallel = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{X.}

\item{\code{levels}}{Name.}

\item{\code{mz}}{X.}

\item{\code{rt}}{X.}

\item{\code{ppm}}{X.}

\item{\code{sec}}{X.}

\item{\code{id}}{X.}

\item{\code{allTraces}}{Logical, when \code{TRUE} all level 2 data is returned.
When \code{FALSE} and level has 2, only the MS2 traces of MS1 targets
are returned, using the `preMZ` value and the `isolationWindow`.}

\item{\code{isolationWindow}}{X.}

\item{\code{minIntensityMS1}}{X.}

\item{\code{minIntensityMS2}}{X.}

\item{\code{run_parallel}}{X.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame with spectra for each analyses and
targets when defined.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6MS-get_chromatograms"></a>}}
\if{latex}{\out{\hypertarget{method-R6MS-get_chromatograms}{}}}
\subsection{Method \code{get_chromatograms()}}{
Method to get spectra from the MS analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6MS$get_chromatograms(analyses = NULL, minIntensity = 0, run_parallel = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{X.}

\item{\code{minIntensity}}{X.}

\item{\code{run_parallel}}{X.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame with spectra.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6MS-get_tic"></a>}}
\if{latex}{\out{\hypertarget{method-R6MS-get_tic}{}}}
\subsection{Method \code{get_tic()}}{
Method to get the total ion chromatograms (TIC) from the analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6MS$get_tic(analyses = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name
of the analyses.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6MS-get_bpc"></a>}}
\if{latex}{\out{\hypertarget{method-R6MS-get_bpc}{}}}
\subsection{Method \code{get_bpc()}}{
Method to get the base peak chromatograms (BPC) from the analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6MS$get_bpc(analyses = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name
of the analyses.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6MS-get_eic"></a>}}
\if{latex}{\out{\hypertarget{method-R6MS-get_eic}{}}}
\subsection{Method \code{get_eic()}}{
Method to get extract ion chromatograms (EIC) from the analyses based
on targets.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6MS$get_eic(
  analyses = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  id = NULL,
  run_parallel = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name
of the analyses.}

\item{\code{mz}}{X.}

\item{\code{rt}}{X.}

\item{\code{ppm}}{X.}

\item{\code{sec}}{X.}

\item{\code{id}}{X.}

\item{\code{run_parallel}}{X.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6MS-get_ms2"></a>}}
\if{latex}{\out{\hypertarget{method-R6MS-get_ms2}{}}}
\subsection{Method \code{get_ms2()}}{
Method to get MS2 data from the analyses based on targets.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6MS$get_ms2(
  analyses = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  id = NULL,
  isolationWindow = 1.3,
  mzClust = 0.01,
  verbose = FALSE,
  run_parallel = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name
of the analyses.}

\item{\code{mz}}{X.}

\item{\code{rt}}{X.}

\item{\code{ppm}}{X.}

\item{\code{sec}}{X.}

\item{\code{id}}{X.}

\item{\code{isolationWindow}}{X.}

\item{\code{mzClust}}{X.}

\item{\code{verbose}}{X.}

\item{\code{run_parallel}}{X.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6MS-set_replicate_names"></a>}}
\if{latex}{\out{\hypertarget{method-R6MS-set_replicate_names}{}}}
\subsection{Method \code{set_replicate_names()}}{
Method to set the analysis replicate names. Changes the `R6MS` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6MS$set_replicate_names(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A character vector with the analysis replicate names.
Must be of the same length as the number of analyses.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6MS-set_blank_names"></a>}}
\if{latex}{\out{\hypertarget{method-R6MS-set_blank_names}{}}}
\subsection{Method \code{set_blank_names()}}{
Method to set the analysis blank replicate names.
Changes the `R6MS` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6MS$set_blank_names(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A character vector with the analysis blank replicate names.
Must be of the same length as the number of analyses.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6MS-has_loaded_spectra"></a>}}
\if{latex}{\out{\hypertarget{method-R6MS-has_loaded_spectra}{}}}
\subsection{Method \code{has_loaded_spectra()}}{
Method to check for loaded spectra in given analyses names/indices.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6MS$has_loaded_spectra(analyses)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{The analyses names/indices to check for loaded spectra.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6MS-has_loaded_chromatograms"></a>}}
\if{latex}{\out{\hypertarget{method-R6MS-has_loaded_chromatograms}{}}}
\subsection{Method \code{has_loaded_chromatograms()}}{
Method to check for loaded chromatograms in given analyses names/indices.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6MS$has_loaded_chromatograms(analyses)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{The analyses names/indices to check for loaded
chromatograms.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6MS-add_analyses"></a>}}
\if{latex}{\out{\hypertarget{method-R6MS-add_analyses}{}}}
\subsection{Method \code{add_analyses()}}{
Method to add analyses to the `R6MS` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6MS$add_analyses(analyses)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A list of analyses.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6MS-load_spectra"></a>}}
\if{latex}{\out{\hypertarget{method-R6MS-load_spectra}{}}}
\subsection{Method \code{load_spectra()}}{
Method to load all spectra from analyses to the `R6MS` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6MS$load_spectra(run_parallel = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{run_parallel}}{X.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6MS-load_chromatograms"></a>}}
\if{latex}{\out{\hypertarget{method-R6MS-load_chromatograms}{}}}
\subsection{Method \code{load_chromatograms()}}{
Method to load all chromatograms from analyses to the `R6MS` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6MS$load_chromatograms(run_parallel = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{run_parallel}}{X.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6MS-plot_spectra"></a>}}
\if{latex}{\out{\hypertarget{method-R6MS-plot_spectra}{}}}
\subsection{Method \code{plot_spectra()}}{
Plots spectra for given MS analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6MS$plot_spectra(
  analyses = NULL,
  levels = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  id = NULL,
  allTraces = FALSE,
  isolationWindow = 1.3,
  minIntensityMS1 = 0,
  minIntensityMS2 = 0,
  run_parallel = FALSE,
  colorBy = "analyses"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{X.}

\item{\code{levels}}{Name.}

\item{\code{mz}}{X.}

\item{\code{rt}}{X.}

\item{\code{ppm}}{X.}

\item{\code{sec}}{X.}

\item{\code{id}}{X.}

\item{\code{allTraces}}{Logical, when \code{TRUE} all level 2 data is returned.
When \code{FALSE} and level has 2, only the MS2 traces of MS1 targets
are returned, using the `preMZ` value and the `isolationWindow`.}

\item{\code{isolationWindow}}{X.}

\item{\code{minIntensityMS1}}{X.}

\item{\code{minIntensityMS2}}{X.}

\item{\code{run_parallel}}{X.}

\item{\code{colorBy}}{X.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A 3D interactive plot.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6MS-plot_xic"></a>}}
\if{latex}{\out{\hypertarget{method-R6MS-plot_xic}{}}}
\subsection{Method \code{plot_xic()}}{
Method to plot extract ion chromatograms (EIC) and \emph{m/z} vs
retention time from the analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6MS$plot_xic(
  analyses = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  id = NULL,
  run_parallel = FALSE,
  legendNames = NULL,
  plotTargetMark = TRUE,
  targetsMark = NULL,
  ppmMark = 5,
  secMark = 10,
  numberRows = 1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name
of the analyses.}

\item{\code{mz}}{X.}

\item{\code{rt}}{X.}

\item{\code{ppm}}{X.}

\item{\code{sec}}{X.}

\item{\code{id}}{X.}

\item{\code{run_parallel}}{X.}

\item{\code{legendNames}}{x.}

\item{\code{plotTargetMark}}{x.}

\item{\code{targetsMark}}{x.}

\item{\code{ppmMark}}{x.}

\item{\code{secMark}}{x.}

\item{\code{numberRows}}{x.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6MS-plot_eic"></a>}}
\if{latex}{\out{\hypertarget{method-R6MS-plot_eic}{}}}
\subsection{Method \code{plot_eic()}}{
Method to plot extract ion chromatograms (EIC) from the analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6MS$plot_eic(
  analyses = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  id = NULL,
  run_parallel = FALSE,
  legendNames = NULL,
  title = NULL,
  colorBy = "targets",
  interactive = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name
of the analyses.}

\item{\code{mz}}{X.}

\item{\code{rt}}{X.}

\item{\code{ppm}}{X.}

\item{\code{sec}}{X.}

\item{\code{id}}{X.}

\item{\code{run_parallel}}{X.}

\item{\code{legendNames}}{x.}

\item{\code{title}}{x.}

\item{\code{colorBy}}{x.}

\item{\code{interactive}}{x.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A plot.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6MS-plot_tic"></a>}}
\if{latex}{\out{\hypertarget{method-R6MS-plot_tic}{}}}
\subsection{Method \code{plot_tic()}}{
Method to plot total ion chromatograms (TIC) of the analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6MS$plot_tic(
  analyses = NULL,
  title = NULL,
  colorBy = "analyses",
  interactive = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name
of the analyses.}

\item{\code{title}}{x.}

\item{\code{colorBy}}{x.}

\item{\code{interactive}}{x.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A plot.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6MS-plot_bpc"></a>}}
\if{latex}{\out{\hypertarget{method-R6MS-plot_bpc}{}}}
\subsection{Method \code{plot_bpc()}}{
Method to plot base peak chromatograms (BPC) of the analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6MS$plot_bpc(
  analyses = NULL,
  title = NULL,
  colorBy = "analyses",
  interactive = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name
of the analyses.}

\item{\code{title}}{x.}

\item{\code{colorBy}}{x.}

\item{\code{interactive}}{x.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A plot.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6MS-plot_ms2"></a>}}
\if{latex}{\out{\hypertarget{method-R6MS-plot_ms2}{}}}
\subsection{Method \code{plot_ms2()}}{
Method to plot MS2 data from the analyses based on targets.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6MS$plot_ms2(
  analyses = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  id = NULL,
  isolationWindow = 1.3,
  mzClust = 0.01,
  verbose = FALSE,
  run_parallel = FALSE,
  legendNames = NULL,
  title = NULL,
  colorBy = "targets",
  interactive = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name
of the analyses.}

\item{\code{mz}}{X.}

\item{\code{rt}}{X.}

\item{\code{ppm}}{X.}

\item{\code{sec}}{X.}

\item{\code{id}}{X.}

\item{\code{isolationWindow}}{X.}

\item{\code{mzClust}}{X.}

\item{\code{verbose}}{X.}

\item{\code{run_parallel}}{X.}

\item{\code{legendNames}}{x.}

\item{\code{title}}{x.}

\item{\code{colorBy}}{x.}

\item{\code{interactive}}{x.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A plot.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6MS-subset_analyses"></a>}}
\if{latex}{\out{\hypertarget{method-R6MS-subset_analyses}{}}}
\subsection{Method \code{subset_analyses()}}{
Create a new MS set of analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6MS$subset_analyses(i, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{i}}{Name.}

\item{\code{...}}{Hair color.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6MS-check_analyses_argument"></a>}}
\if{latex}{\out{\hypertarget{method-R6MS-check_analyses_argument}{}}}
\subsection{Method \code{check_analyses_argument()}}{
Check the analyses argument as a character/integer vector to match
analyses names or indices from the `R6MS` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6MS$check_analyses_argument(analyses)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{X.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A valid character vector with analyses names of `NULL`.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6MS-clone"></a>}}
\if{latex}{\out{\hypertarget{method-R6MS-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6MS$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
