% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_S3_RamanAnalysis.R
\name{RamanAnalysis}
\alias{RamanAnalysis}
\title{\strong{RamanAnalysis} S3 class constructor, methods and functions}
\usage{
RamanAnalysis(
  name = NA_character_,
  replicate = NA_character_,
  blank = NA_character_,
  file = NA_character_,
  type = NA_character_,
  metadata = list(),
  spectra = data.table()
)
}
\arguments{
\item{name}{\emph{asc} file name without extension.}

\item{replicate}{Character with length one, representing the analysis replicate group name.}

\item{blank}{Character with length one, representing the associated blank replicate group name.}

\item{file}{\emph{asc} full file path (with extension).}

\item{type}{}

\item{metadata}{List of analysis metadata taken from the \emph{asc} file or added.}

\item{spectra}{data.table with the raw spectra.}
}
\value{
An \emph{RamanAnalysis} S3 class object.
}
\description{
Creates a \emph{RamanAnalysis} S3 class object.
}
