% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_R6_ramanData.R
\name{RamanData}
\alias{RamanData}
\title{\strong{ramanData} R6 class and methods}
\description{
The RamanData R6 class is a framework with methods for parsing, processing,
visualizing and storing RAMAN data.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RamanData-new}{\code{RamanData$new()}}
\item \href{#method-RamanData-print}{\code{RamanData$print()}}
\item \href{#method-RamanData-get_headers}{\code{RamanData$get_headers()}}
\item \href{#method-RamanData-get_analyses}{\code{RamanData$get_analyses()}}
\item \href{#method-RamanData-get_overview}{\code{RamanData$get_overview()}}
\item \href{#method-RamanData-add_headers}{\code{RamanData$add_headers()}}
\item \href{#method-RamanData-clone}{\code{RamanData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamanData-new"></a>}}
\if{latex}{\out{\hypertarget{method-RamanData-new}{}}}
\subsection{Method \code{new()}}{
Creates an R6 RamanData class object. When \code{headers} are not given
(i.e., \code{NULL}), a default Headers S3 class object is generated with name
as \code{NA_character}, path as \code{get_wd()} and date as \code{Sys.time()}.
See \code{?Headers} for more information.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamanData$new(files = NULL, headers = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{files}}{Full file paths of Raman analyses.}

\item{\code{headers}}{A named list with headers or a Headers S3 class object.
Each list element must have length one. There is no type restrictions
except for name (must be type character), path (must be type
character) and date (must be class POSIXct or class POSIXt). See more
information in \code{?Headers}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamanData-print"></a>}}
\if{latex}{\out{\hypertarget{method-RamanData-print}{}}}
\subsection{Method \code{print()}}{
Prints a summary of the \code{RamanData} object in the console.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamanData$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Console text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamanData-get_headers"></a>}}
\if{latex}{\out{\hypertarget{method-RamanData-get_headers}{}}}
\subsection{Method \code{get_headers()}}{
Gets the headers.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamanData$get_headers(value = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A character vector with the name/s of the header elements.
When \code{NULL}, the entire headers list is returned.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The headers list or the header elements as defined by \code{value}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamanData-get_analyses"></a>}}
\if{latex}{\out{\hypertarget{method-RamanData-get_analyses}{}}}
\subsection{Method \code{get_analyses()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamanData$get_analyses()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamanData-get_overview"></a>}}
\if{latex}{\out{\hypertarget{method-RamanData-get_overview}{}}}
\subsection{Method \code{get_overview()}}{
Gets the overview data.frame with all the analysis types,
names, replicates, associated blank replicates and full
file paths.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamanData$get_overview()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data.frame with columns type, analysis, replicate, blank and
file.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamanData-add_headers"></a>}}
\if{latex}{\out{\hypertarget{method-RamanData-add_headers}{}}}
\subsection{Method \code{add_headers()}}{
Adds headers. If an argument or element "name" is given, it must
be type character. If an argument or element path is given, it must be
type character and exist. If an argument or element date is given, it
must be class POSIXct or POSIXt. If given date is character, conversion
to class POSIXct or POSIXt is attempted. See \code{?Headers} for more
information.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamanData$add_headers(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments to be added as headers. Note that all given arguments
must be of length one. There is no type limitations. The name of the argument
becomes the name of the header entry. Alternatively, a single list argument
with the headers (all length one) can be given.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamanData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RamanData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamanData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
