% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_R6_RamanData.R
\name{RamanData}
\alias{RamanData}
\title{\strong{RamanData} R6 class and methods}
\description{
The RamanData R6 class is a framework with methods for parsing, processing,
inspecting and storing RAMAN data.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RamanData-new}{\code{RamanData$new()}}
\item \href{#method-RamanData-print}{\code{RamanData$print()}}
\item \href{#method-RamanData-get_headers}{\code{RamanData$get_headers()}}
\item \href{#method-RamanData-get_analyses}{\code{RamanData$get_analyses()}}
\item \href{#method-RamanData-get_number_analyses}{\code{RamanData$get_number_analyses()}}
\item \href{#method-RamanData-get_analysis_names}{\code{RamanData$get_analysis_names()}}
\item \href{#method-RamanData-get_replicate_names}{\code{RamanData$get_replicate_names()}}
\item \href{#method-RamanData-get_blank_names}{\code{RamanData$get_blank_names()}}
\item \href{#method-RamanData-get_files}{\code{RamanData$get_files()}}
\item \href{#method-RamanData-get_spectra}{\code{RamanData$get_spectra()}}
\item \href{#method-RamanData-get_overview}{\code{RamanData$get_overview()}}
\item \href{#method-RamanData-add_headers}{\code{RamanData$add_headers()}}
\item \href{#method-RamanData-add_analyses}{\code{RamanData$add_analyses()}}
\item \href{#method-RamanData-add_replicate_names}{\code{RamanData$add_replicate_names()}}
\item \href{#method-RamanData-add_blank_names}{\code{RamanData$add_blank_names()}}
\item \href{#method-RamanData-remove_analyses}{\code{RamanData$remove_analyses()}}
\item \href{#method-RamanData-merge_replicates}{\code{RamanData$merge_replicates()}}
\item \href{#method-RamanData-plot_spectra}{\code{RamanData$plot_spectra()}}
\item \href{#method-RamanData-clone}{\code{RamanData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamanData-new"></a>}}
\if{latex}{\out{\hypertarget{method-RamanData-new}{}}}
\subsection{Method \code{new()}}{
Creates an R6 RamanData class object. When \code{headers} are not given
(i.e., \code{NULL}), a default Headers S3 class object is generated with name
as \code{NA_character}, path as \code{get_wd()} and date as \code{Sys.time()}.
See \code{?Headers} for more information.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamanData$new(files = NULL, headers = NULL, runParallel = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{files}}{Full file paths of Raman analyses.}

\item{\code{headers}}{A named list with headers or a ProjectHeaders S3 class object.
Each list element must have length one. There is no type restrictions
except for name (must be type character length 1), author (must be type
character length 1), path (must be type character length 1) and date (must
be class POSIXct or class POSIXt length 1). See more information in
\code{?ProjectHeaders}.}

\item{\code{runParallel}}{Logical of length one. Set to \code{TRUE} for parallel
processing. Note that parallel processing is not always faster due to
overhead.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamanData-print"></a>}}
\if{latex}{\out{\hypertarget{method-RamanData-print}{}}}
\subsection{Method \code{print()}}{
Prints a summary of the \code{RamanData} object in the console.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamanData$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Console text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamanData-get_headers"></a>}}
\if{latex}{\out{\hypertarget{method-RamanData-get_headers}{}}}
\subsection{Method \code{get_headers()}}{
Gets the headers.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamanData$get_headers(value = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A character vector with the name/s of the header elements.
When \code{NULL}, the entire headers list is returned.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The headers list or the header elements as defined by \code{value}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamanData-get_analyses"></a>}}
\if{latex}{\out{\hypertarget{method-RamanData-get_analyses}{}}}
\subsection{Method \code{get_analyses()}}{
Gets analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamanData$get_analyses(analyses = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name
of the analyses.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The list of analyses or the analyses as defined by \code{analyses}
argument.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamanData-get_number_analyses"></a>}}
\if{latex}{\out{\hypertarget{method-RamanData-get_number_analyses}{}}}
\subsection{Method \code{get_number_analyses()}}{
Gets the number of analyses present.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamanData$get_number_analyses()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An integer value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamanData-get_analysis_names"></a>}}
\if{latex}{\out{\hypertarget{method-RamanData-get_analysis_names}{}}}
\subsection{Method \code{get_analysis_names()}}{
Gets the analysis names.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamanData$get_analysis_names(analyses = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{X.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamanData-get_replicate_names"></a>}}
\if{latex}{\out{\hypertarget{method-RamanData-get_replicate_names}{}}}
\subsection{Method \code{get_replicate_names()}}{
Gets the analysis replicate names.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamanData$get_replicate_names(analyses = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name
of the analyses.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamanData-get_blank_names"></a>}}
\if{latex}{\out{\hypertarget{method-RamanData-get_blank_names}{}}}
\subsection{Method \code{get_blank_names()}}{
Gets the analysis blank replicate names.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamanData$get_blank_names(analyses = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name
of the analyses.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamanData-get_files"></a>}}
\if{latex}{\out{\hypertarget{method-RamanData-get_files}{}}}
\subsection{Method \code{get_files()}}{
Gets the full file paths of the analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamanData$get_files(analyses = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{X.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamanData-get_spectra"></a>}}
\if{latex}{\out{\hypertarget{method-RamanData-get_spectra}{}}}
\subsection{Method \code{get_spectra()}}{
Gets spectra from analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamanData$get_spectra(
  analyses = NULL,
  rt = NULL,
  shift = NULL,
  minIntensity = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name
of the analyses.}

\item{\code{rt}}{Numeric of length two with the minimum and maximum retention time,
in seconds.}

\item{\code{shift}}{Numeric of length two with the minimum and maximum shift,
in cm\if{html}{\out{<sup>}}-1\if{html}{\out{</sup>}}.}

\item{\code{minIntensity}}{Numeric value with the minimum intensity.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamanData-get_overview"></a>}}
\if{latex}{\out{\hypertarget{method-RamanData-get_overview}{}}}
\subsection{Method \code{get_overview()}}{
Gets the overview data.frame with all the analysis types,
names, replicates, associated blank replicates and full
file paths.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamanData$get_overview()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data.frame with columns type, analysis, replicate, blank and
file.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamanData-add_headers"></a>}}
\if{latex}{\out{\hypertarget{method-RamanData-add_headers}{}}}
\subsection{Method \code{add_headers()}}{
Adds headers. If an argument or element "name" is given, it must
be type character. If an argument or element path is given, it must be
type character and exist. If an argument or element date is given, it
must be class POSIXct or POSIXt. If given date is character, conversion
to class POSIXct or POSIXt is attempted. See \code{?Headers} for more
information.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamanData$add_headers(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments to be added as headers. Note that all given arguments
must be of length one. There is no type limitations. The name of the argument
becomes the name of the header entry. Alternatively, a single list argument
with the headers (all length one) can be given.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamanData-add_analyses"></a>}}
\if{latex}{\out{\hypertarget{method-RamanData-add_analyses}{}}}
\subsection{Method \code{add_analyses()}}{
Adds analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamanData$add_analyses(analyses = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A list.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamanData-add_replicate_names"></a>}}
\if{latex}{\out{\hypertarget{method-RamanData-add_replicate_names}{}}}
\subsection{Method \code{add_replicate_names()}}{
Adds or redefines the analysis replicate names.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamanData$add_replicate_names(value = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A character vector with the analysis replicate names.
Must be of the same length as the number of analyses.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamanData-add_blank_names"></a>}}
\if{latex}{\out{\hypertarget{method-RamanData-add_blank_names}{}}}
\subsection{Method \code{add_blank_names()}}{
Adds or redefines the analysis blank replicate names.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamanData$add_blank_names(value = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A character vector with the analysis blank replicate names.
Must be of the same length as the number of analyses.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamanData-remove_analyses"></a>}}
\if{latex}{\out{\hypertarget{method-RamanData-remove_analyses}{}}}
\subsection{Method \code{remove_analyses()}}{
Removes analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamanData$remove_analyses(analyses = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name
of the analyses.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamanData-merge_replicates"></a>}}
\if{latex}{\out{\hypertarget{method-RamanData-merge_replicates}{}}}
\subsection{Method \code{merge_replicates()}}{
Merges spectra for given RAMAN analyses from the same chromatographic
separation when using LC-Raman coupling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamanData$merge_replicates(preCut = 2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{preCut}}{The number of pre Raman scans to exclude when merging.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamanData-plot_spectra"></a>}}
\if{latex}{\out{\hypertarget{method-RamanData-plot_spectra}{}}}
\subsection{Method \code{plot_spectra()}}{
Plots spectra for given RAMAN analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamanData$plot_spectra(
  analyses = NULL,
  rt = NULL,
  shift = NULL,
  minIntensity = 0,
  xVal = "shift",
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  cex = 0.6,
  showLegend = TRUE,
  colorBy = "analyses",
  interactive = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name
of the analyses.}

\item{\code{rt}}{Numeric of length two with the minimum and maximum retention time,
in seconds.}

\item{\code{shift}}{Numeric of length two with the minimum and maximum shift,
in cm\if{html}{\out{<sup>}}-1\if{html}{\out{</sup>}}.}

\item{\code{minIntensity}}{Numeric value with the minimum intensity.}

\item{\code{xVal}}{Character of length one. Possible are "rt" or "shift" for
using the retention time or the shift as x axis, respectively.}

\item{\code{xLab}}{A string with the title for the x axis.}

\item{\code{yLab}}{A string with the title for the y axis.}

\item{\code{title}}{A string with the title.}

\item{\code{cex}}{A length one numeric value defining the scale of the legend for
non interactive plotting.}

\item{\code{showLegend}}{Logical. Set to \code{TRUE} to show legend.}

\item{\code{colorBy}}{A string of length 1. One of \code{analyses} (the default) or
\code{replicates}.}

\item{\code{interactive}}{Logical. When \code{TRUE}, the data is plotted interactively
using \pkg{plotly}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A plot.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamanData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RamanData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamanData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
