% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_R6_RamanEngine.R
\name{RamanEngine}
\alias{RamanEngine}
\title{\strong{RamanEngine} R6 class and methods}
\description{
The \emph{RamanEngine} R6 class is a framework for parsing, processing, inspecting and storing Raman spectroscopic data.
}
\section{Super class}{
\code{\link[StreamFind:CoreEngine]{StreamFind::CoreEngine}} -> \code{RamanEngine}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{raw_spectra}}{List of raw spectra \code{data.table} objects for each analysis.}

\item{\code{spectra}}{List of spectra \code{data.table} objects for each analysis.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RamanEngine-new}{\code{RamanEngine$new()}}
\item \href{#method-RamanEngine-get_overview}{\code{RamanEngine$get_overview()}}
\item \href{#method-RamanEngine-get_spectra}{\code{RamanEngine$get_spectra()}}
\item \href{#method-RamanEngine-subset_analyses}{\code{RamanEngine$subset_analyses()}}
\item \href{#method-RamanEngine-remove_analyses}{\code{RamanEngine$remove_analyses()}}
\item \href{#method-RamanEngine-merge_spectra_time_series}{\code{RamanEngine$merge_spectra_time_series()}}
\item \href{#method-RamanEngine-average_spectra}{\code{RamanEngine$average_spectra()}}
\item \href{#method-RamanEngine-subtract_blank_spectra}{\code{RamanEngine$subtract_blank_spectra()}}
\item \href{#method-RamanEngine-correct_spectra_baseline}{\code{RamanEngine$correct_spectra_baseline()}}
\item \href{#method-RamanEngine-bin_spectra}{\code{RamanEngine$bin_spectra()}}
\item \href{#method-RamanEngine-subtract_spectra_section}{\code{RamanEngine$subtract_spectra_section()}}
\item \href{#method-RamanEngine-delete_spectra_section}{\code{RamanEngine$delete_spectra_section()}}
\item \href{#method-RamanEngine-smooth_spectra}{\code{RamanEngine$smooth_spectra()}}
\item \href{#method-RamanEngine-normalize_spectra}{\code{RamanEngine$normalize_spectra()}}
\item \href{#method-RamanEngine-has_spectra}{\code{RamanEngine$has_spectra()}}
\item \href{#method-RamanEngine-plot_spectra}{\code{RamanEngine$plot_spectra()}}
\item \href{#method-RamanEngine-plot_spectra_baseline}{\code{RamanEngine$plot_spectra_baseline()}}
\item \href{#method-RamanEngine-plot_chromatograms}{\code{RamanEngine$plot_chromatograms()}}
\item \href{#method-RamanEngine-processing_methods}{\code{RamanEngine$processing_methods()}}
\item \href{#method-RamanEngine-clone}{\code{RamanEngine$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="add_analyses"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-add_analyses'><code>StreamFind::CoreEngine$add_analyses()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="add_blank_names"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-add_blank_names'><code>StreamFind::CoreEngine$add_blank_names()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="add_headers"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-add_headers'><code>StreamFind::CoreEngine$add_headers()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="add_metadata"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-add_metadata'><code>StreamFind::CoreEngine$add_metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="add_replicate_names"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-add_replicate_names'><code>StreamFind::CoreEngine$add_replicate_names()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="add_results"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-add_results'><code>StreamFind::CoreEngine$add_results()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="add_settings"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-add_settings'><code>StreamFind::CoreEngine$add_settings()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="get_analyses"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-get_analyses'><code>StreamFind::CoreEngine$get_analyses()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="get_analysis_names"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-get_analysis_names'><code>StreamFind::CoreEngine$get_analysis_names()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="get_blank_names"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-get_blank_names'><code>StreamFind::CoreEngine$get_blank_names()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="get_files"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-get_files'><code>StreamFind::CoreEngine$get_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="get_formats"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-get_formats'><code>StreamFind::CoreEngine$get_formats()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="get_headers"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-get_headers'><code>StreamFind::CoreEngine$get_headers()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="get_history"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-get_history'><code>StreamFind::CoreEngine$get_history()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="get_number_analyses"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-get_number_analyses'><code>StreamFind::CoreEngine$get_number_analyses()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="get_replicate_names"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-get_replicate_names'><code>StreamFind::CoreEngine$get_replicate_names()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="get_results"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-get_results'><code>StreamFind::CoreEngine$get_results()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="get_results_names"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-get_results_names'><code>StreamFind::CoreEngine$get_results_names()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="get_settings"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-get_settings'><code>StreamFind::CoreEngine$get_settings()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="get_settings_names"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-get_settings_names'><code>StreamFind::CoreEngine$get_settings_names()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="get_types"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-get_types'><code>StreamFind::CoreEngine$get_types()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="get_workflow_overview"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-get_workflow_overview'><code>StreamFind::CoreEngine$get_workflow_overview()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="has_analyses"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-has_analyses'><code>StreamFind::CoreEngine$has_analyses()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="has_results"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-has_results'><code>StreamFind::CoreEngine$has_results()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="has_settings"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-has_settings'><code>StreamFind::CoreEngine$has_settings()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="import"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-import'><code>StreamFind::CoreEngine$import()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="import_analyses"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-import_analyses'><code>StreamFind::CoreEngine$import_analyses()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="import_headers"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-import_headers'><code>StreamFind::CoreEngine$import_headers()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="import_settings"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-import_settings'><code>StreamFind::CoreEngine$import_settings()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="plot_chromatograms_peaks"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-plot_chromatograms_peaks'><code>StreamFind::CoreEngine$plot_chromatograms_peaks()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="plot_spectra_peaks"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-plot_spectra_peaks'><code>StreamFind::CoreEngine$plot_spectra_peaks()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="print"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-print'><code>StreamFind::CoreEngine$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="print_analyses"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-print_analyses'><code>StreamFind::CoreEngine$print_analyses()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="print_headers"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-print_headers'><code>StreamFind::CoreEngine$print_headers()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="print_results"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-print_results'><code>StreamFind::CoreEngine$print_results()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="print_workflow"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-print_workflow'><code>StreamFind::CoreEngine$print_workflow()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="remove_headers"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-remove_headers'><code>StreamFind::CoreEngine$remove_headers()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="remove_results"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-remove_results'><code>StreamFind::CoreEngine$remove_results()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="remove_settings"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-remove_settings'><code>StreamFind::CoreEngine$remove_settings()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="run_workflow"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-run_workflow'><code>StreamFind::CoreEngine$run_workflow()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="save"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-save'><code>StreamFind::CoreEngine$save()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="save_analyses"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-save_analyses'><code>StreamFind::CoreEngine$save_analyses()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="save_headers"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-save_headers'><code>StreamFind::CoreEngine$save_headers()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="save_settings"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-save_settings'><code>StreamFind::CoreEngine$save_settings()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamanEngine-new"></a>}}
\if{latex}{\out{\hypertarget{method-RamanEngine-new}{}}}
\subsection{Method \code{new()}}{
Creates an R6 class \emph{RamanEngine}. Child of \emph{CoreEngine} R6 class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamanEngine$new(
  files = NULL,
  headers = NULL,
  settings = NULL,
  analyses = NULL,
  results = NULL,
  runParallel = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{files}}{Vector with full paths of \strong{.asc} files from Raman analyses.}

\item{\code{headers}}{A named list with headers or a ProjectHeaders S3 class object.
Each list element must have length one. There is no type restrictions
except for name (must be type character length 1), author (must be type
character length 1), path (must be type character length 1) and date (must
be class POSIXct or class POSIXt length 1). See more information in
\code{?ProjectHeaders}.}

\item{\code{settings}}{A named list of ProcessingSettings objects or a
single ProcessingSettings object. The list names should match
the call name of each ProcessingSettings object. Alternatively, a named
list with call name, algorithm and parameters to be transformed and added
as ProcessingSettings object.}

\item{\code{settings}}{A ProcessingSettings S3 class object with the call name
applicable to the processing method. Alternatively, a named
list with \code{call}, \code{algorithm} and \code{parameters} to be transformed and used as
ProcessingSettings S3 class object. When not given, settings for the
processing method will be searched within the engine object.}

\item{\code{analyses}}{A numeric/character vector with the number/name of the analyses.}

\item{\code{results}}{A named list of objects from processing methods.}

\item{\code{runParallel}}{Logical of length one. Set to \code{TRUE} for parallel
processing. Note that parallel processing is not always faster due to
overhead.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamanEngine-get_overview"></a>}}
\if{latex}{\out{\hypertarget{method-RamanEngine-get_overview}{}}}
\subsection{Method \code{get_overview()}}{
Gets an overview data.frame of all the analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamanEngine$get_overview()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamanEngine-get_spectra"></a>}}
\if{latex}{\out{\hypertarget{method-RamanEngine-get_spectra}{}}}
\subsection{Method \code{get_spectra()}}{
Gets a \code{data.table} with spectra from analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamanEngine$get_spectra(
  analyses = NULL,
  rt = NULL,
  shift = NULL,
  minIntensity = 0,
  raw_spectra = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name of the analyses.}

\item{\code{rt}}{Numeric of length two with the minimum and maximum retention time, in seconds.}

\item{\code{shift}}{Numeric of length two with the minimum and maximum shift, in cm\if{html}{\out{<sup>}}-1\if{html}{\out{</sup>}}.}

\item{\code{minIntensity}}{Numeric value with the minimum intensity.}

\item{\code{raw_spectra}}{Logical of length one. Set to \code{TRUE} for parsing raw spectra not spectra results/processed.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamanEngine-subset_analyses"></a>}}
\if{latex}{\out{\hypertarget{method-RamanEngine-subset_analyses}{}}}
\subsection{Method \code{subset_analyses()}}{
Subsets on analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamanEngine$subset_analyses(analyses = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name of the analyses.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new cloned engine with only the analyses as defined by the analyses argument.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamanEngine-remove_analyses"></a>}}
\if{latex}{\out{\hypertarget{method-RamanEngine-remove_analyses}{}}}
\subsection{Method \code{remove_analyses()}}{
Removes analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamanEngine$remove_analyses(analyses = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name of the analyses.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamanEngine-merge_spectra_time_series"></a>}}
\if{latex}{\out{\hypertarget{method-RamanEngine-merge_spectra_time_series}{}}}
\subsection{Method \code{merge_spectra_time_series()}}{
Merges spectra for given \emph{RamanAnalyses} from the same chromatographic separation when using
LC-Raman coupling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamanEngine$merge_spectra_time_series(settings)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{settings}}{A named list of ProcessingSettings objects or a
single ProcessingSettings object. The list names should match
the call name of each ProcessingSettings object. Alternatively, a named
list with call name, algorithm and parameters to be transformed and added
as ProcessingSettings object.}

\item{\code{settings}}{A ProcessingSettings S3 class object with the call name
applicable to the processing method. Alternatively, a named
list with \code{call}, \code{algorithm} and \code{parameters} to be transformed and used as
ProcessingSettings S3 class object. When not given, settings for the
processing method will be searched within the engine object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamanEngine-average_spectra"></a>}}
\if{latex}{\out{\hypertarget{method-RamanEngine-average_spectra}{}}}
\subsection{Method \code{average_spectra()}}{
Averages spectra based on assigned analysis replicates.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamanEngine$average_spectra(settings)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{settings}}{A named list of ProcessingSettings objects or a
single ProcessingSettings object. The list names should match
the call name of each ProcessingSettings object. Alternatively, a named
list with call name, algorithm and parameters to be transformed and added
as ProcessingSettings object.}

\item{\code{settings}}{A ProcessingSettings S3 class object with the call name
applicable to the processing method. Alternatively, a named
list with \code{call}, \code{algorithm} and \code{parameters} to be transformed and used as
ProcessingSettings S3 class object. When not given, settings for the
processing method will be searched within the engine object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamanEngine-subtract_blank_spectra"></a>}}
\if{latex}{\out{\hypertarget{method-RamanEngine-subtract_blank_spectra}{}}}
\subsection{Method \code{subtract_blank_spectra()}}{
Subtracts spectra from correspondent blank analysis replicates.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamanEngine$subtract_blank_spectra(settings)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{settings}}{A named list of ProcessingSettings objects or a
single ProcessingSettings object. The list names should match
the call name of each ProcessingSettings object. Alternatively, a named
list with call name, algorithm and parameters to be transformed and added
as ProcessingSettings object.}

\item{\code{settings}}{A ProcessingSettings S3 class object with the call name
applicable to the processing method. Alternatively, a named
list with \code{call}, \code{algorithm} and \code{parameters} to be transformed and used as
ProcessingSettings S3 class object. When not given, settings for the
processing method will be searched within the engine object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamanEngine-correct_spectra_baseline"></a>}}
\if{latex}{\out{\hypertarget{method-RamanEngine-correct_spectra_baseline}{}}}
\subsection{Method \code{correct_spectra_baseline()}}{
Corrects the spectra baseline.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamanEngine$correct_spectra_baseline(settings)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{settings}}{A named list of ProcessingSettings objects or a
single ProcessingSettings object. The list names should match
the call name of each ProcessingSettings object. Alternatively, a named
list with call name, algorithm and parameters to be transformed and added
as ProcessingSettings object.}

\item{\code{settings}}{A ProcessingSettings S3 class object with the call name
applicable to the processing method. Alternatively, a named
list with \code{call}, \code{algorithm} and \code{parameters} to be transformed and used as
ProcessingSettings S3 class object. When not given, settings for the
processing method will be searched within the engine object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamanEngine-bin_spectra"></a>}}
\if{latex}{\out{\hypertarget{method-RamanEngine-bin_spectra}{}}}
\subsection{Method \code{bin_spectra()}}{
Bins spectra in units, according to a given window size.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamanEngine$bin_spectra(settings)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{settings}}{A named list of ProcessingSettings objects or a
single ProcessingSettings object. The list names should match
the call name of each ProcessingSettings object. Alternatively, a named
list with call name, algorithm and parameters to be transformed and added
as ProcessingSettings object.}

\item{\code{settings}}{A ProcessingSettings S3 class object with the call name
applicable to the processing method. Alternatively, a named
list with \code{call}, \code{algorithm} and \code{parameters} to be transformed and used as
ProcessingSettings S3 class object. When not given, settings for the
processing method will be searched within the engine object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamanEngine-subtract_spectra_section"></a>}}
\if{latex}{\out{\hypertarget{method-RamanEngine-subtract_spectra_section}{}}}
\subsection{Method \code{subtract_spectra_section()}}{
Subtracts each spectra by a spectra section in each analysis.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamanEngine$subtract_spectra_section(settings)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{settings}}{A named list of ProcessingSettings objects or a
single ProcessingSettings object. The list names should match
the call name of each ProcessingSettings object. Alternatively, a named
list with call name, algorithm and parameters to be transformed and added
as ProcessingSettings object.}

\item{\code{settings}}{A ProcessingSettings S3 class object with the call name
applicable to the processing method. Alternatively, a named
list with \code{call}, \code{algorithm} and \code{parameters} to be transformed and used as
ProcessingSettings S3 class object. When not given, settings for the
processing method will be searched within the engine object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamanEngine-delete_spectra_section"></a>}}
\if{latex}{\out{\hypertarget{method-RamanEngine-delete_spectra_section}{}}}
\subsection{Method \code{delete_spectra_section()}}{
Deletes a section of spectra in each analysis.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamanEngine$delete_spectra_section(settings)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{settings}}{A named list of ProcessingSettings objects or a
single ProcessingSettings object. The list names should match
the call name of each ProcessingSettings object. Alternatively, a named
list with call name, algorithm and parameters to be transformed and added
as ProcessingSettings object.}

\item{\code{settings}}{A ProcessingSettings S3 class object with the call name
applicable to the processing method. Alternatively, a named
list with \code{call}, \code{algorithm} and \code{parameters} to be transformed and used as
ProcessingSettings S3 class object. When not given, settings for the
processing method will be searched within the engine object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamanEngine-smooth_spectra"></a>}}
\if{latex}{\out{\hypertarget{method-RamanEngine-smooth_spectra}{}}}
\subsection{Method \code{smooth_spectra()}}{
Smooths the spectra in each analysis/replicate.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamanEngine$smooth_spectra(settings)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{settings}}{A named list of ProcessingSettings objects or a
single ProcessingSettings object. The list names should match
the call name of each ProcessingSettings object. Alternatively, a named
list with call name, algorithm and parameters to be transformed and added
as ProcessingSettings object.}

\item{\code{settings}}{A ProcessingSettings S3 class object with the call name
applicable to the processing method. Alternatively, a named
list with \code{call}, \code{algorithm} and \code{parameters} to be transformed and used as
ProcessingSettings S3 class object. When not given, settings for the
processing method will be searched within the engine object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamanEngine-normalize_spectra"></a>}}
\if{latex}{\out{\hypertarget{method-RamanEngine-normalize_spectra}{}}}
\subsection{Method \code{normalize_spectra()}}{
Normalizes spectra in each analysis/replicate.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamanEngine$normalize_spectra(settings)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{settings}}{A named list of ProcessingSettings objects or a
single ProcessingSettings object. The list names should match
the call name of each ProcessingSettings object. Alternatively, a named
list with call name, algorithm and parameters to be transformed and added
as ProcessingSettings object.}

\item{\code{settings}}{A ProcessingSettings S3 class object with the call name
applicable to the processing method. Alternatively, a named
list with \code{call}, \code{algorithm} and \code{parameters} to be transformed and used as
ProcessingSettings S3 class object. When not given, settings for the
processing method will be searched within the engine object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamanEngine-has_spectra"></a>}}
\if{latex}{\out{\hypertarget{method-RamanEngine-has_spectra}{}}}
\subsection{Method \code{has_spectra()}}{
Checks if there are spectra, returning \code{TRUE} or \code{FALSE}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamanEngine$has_spectra()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamanEngine-plot_spectra"></a>}}
\if{latex}{\out{\hypertarget{method-RamanEngine-plot_spectra}{}}}
\subsection{Method \code{plot_spectra()}}{
Plots spectra for given \emph{RamanAnalyses}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamanEngine$plot_spectra(
  analyses = NULL,
  rt = NULL,
  shift = NULL,
  minIntensity = 0,
  raw_spectra = FALSE,
  xVal = "shift",
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  cex = 0.6,
  showLegend = TRUE,
  colorBy = "analyses",
  interactive = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name of the analyses.}

\item{\code{rt}}{Numeric of length two with the minimum and maximum retention time, in seconds.}

\item{\code{shift}}{Numeric of length two with the minimum and maximum shift, in cm\if{html}{\out{<sup>}}-1\if{html}{\out{</sup>}}.}

\item{\code{minIntensity}}{Numeric value with the minimum intensity.}

\item{\code{raw_spectra}}{Logical of length one. Set to \code{TRUE} for parsing raw spectra not spectra results/processed.}

\item{\code{xVal}}{Character of length one. Possible are "rt" or "shift" for using the retention time or the shift
as x axis, respectively.}

\item{\code{xLab}}{A string with the title for the x axis.}

\item{\code{yLab}}{A string with the title for the y axis.}

\item{\code{title}}{A string with the title.}

\item{\code{cex}}{A length one numeric value defining the scale of the legend for non interactive plotting.}

\item{\code{showLegend}}{Logical. Set to \code{TRUE} to show legend.}

\item{\code{colorBy}}{A string defining how to legend the plot. Possible values are
\code{analyses}, \code{targets} (the default) or \code{replicates}. Other options are possible
for specific plotting methods.}

\item{\code{interactive}}{Logical. When \code{TRUE}, the data is plotted interactively using \pkg{plotly}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A plot.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamanEngine-plot_spectra_baseline"></a>}}
\if{latex}{\out{\hypertarget{method-RamanEngine-plot_spectra_baseline}{}}}
\subsection{Method \code{plot_spectra_baseline()}}{
Plots spectra corrected for given \emph{RamanAnalyses}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamanEngine$plot_spectra_baseline(
  analyses = NULL,
  rt = NULL,
  shift = NULL,
  minIntensity = 0,
  xVal = "shift",
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  cex = 0.6,
  showLegend = TRUE,
  colorBy = "analyses",
  interactive = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name of the analyses.}

\item{\code{rt}}{Numeric of length two with the minimum and maximum retention time, in seconds.}

\item{\code{shift}}{Numeric of length two with the minimum and maximum shift, in cm\if{html}{\out{<sup>}}-1\if{html}{\out{</sup>}}.}

\item{\code{minIntensity}}{Numeric value with the minimum intensity.}

\item{\code{xVal}}{Character of length one. Possible are "rt" or "shift" for using the retention time or the shift
as x axis, respectively.}

\item{\code{xLab}}{A string with the title for the x axis.}

\item{\code{yLab}}{A string with the title for the y axis.}

\item{\code{title}}{A string with the title.}

\item{\code{cex}}{A length one numeric value defining the scale of the legend for non interactive plotting.}

\item{\code{showLegend}}{Logical. Set to \code{TRUE} to show legend.}

\item{\code{colorBy}}{A string defining how to legend the plot. Possible values are
\code{analyses}, \code{targets} (the default) or \code{replicates}. Other options are possible
for specific plotting methods.}

\item{\code{interactive}}{Logical. When \code{TRUE}, the data is plotted interactively using \pkg{plotly}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A plot.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamanEngine-plot_chromatograms"></a>}}
\if{latex}{\out{\hypertarget{method-RamanEngine-plot_chromatograms}{}}}
\subsection{Method \code{plot_chromatograms()}}{
Plots chromatograms from analyses with spectra coupled to LC.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamanEngine$plot_chromatograms(
  analyses = NULL,
  rt = NULL,
  shift = NULL,
  minIntensity = 0,
  raw_spectra = FALSE,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  cex = 0.6,
  showLegend = TRUE,
  colorBy = "analyses",
  interactive = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name of the analyses.}

\item{\code{rt}}{Numeric of length two with the minimum and maximum retention time, in seconds.}

\item{\code{shift}}{Numeric of length two with the minimum and maximum shift, in cm\if{html}{\out{<sup>}}-1\if{html}{\out{</sup>}}.}

\item{\code{minIntensity}}{Numeric value with the minimum intensity.}

\item{\code{raw_spectra}}{Logical of length one. Set to \code{TRUE} for parsing raw spectra not spectra results/processed.}

\item{\code{xLab}}{A string with the title for the x axis.}

\item{\code{yLab}}{A string with the title for the y axis.}

\item{\code{title}}{A string with the title.}

\item{\code{cex}}{A length one numeric value defining the scale of the legend for non interactive plotting.}

\item{\code{showLegend}}{Logical. Set to \code{TRUE} to show legend.}

\item{\code{colorBy}}{A string defining how to legend the plot. Possible values are
\code{analyses}, \code{targets} (the default) or \code{replicates}. Other options are possible
for specific plotting methods.}

\item{\code{interactive}}{Logical. When \code{TRUE}, the data is plotted interactively using \pkg{plotly}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A plot.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamanEngine-processing_methods"></a>}}
\if{latex}{\out{\hypertarget{method-RamanEngine-processing_methods}{}}}
\subsection{Method \code{processing_methods()}}{
A data.table with available data processing methods.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamanEngine$processing_methods()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RamanEngine-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RamanEngine-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RamanEngine$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
