% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RamanMethod_BinScans.R
\name{RamanMethod_BinScans_native}
\alias{RamanMethod_BinScans_native}
\alias{validate_object.RamanMethod_BinScans_native}
\title{Raman Method for Binning Time Scans (native algorithm)}
\usage{
RamanMethod_BinScans_native(
  mode = c("unit"),
  value = 5,
  global = FALSE,
  refAnalysis = NA_integer_
)

\method{validate_object}{RamanMethod_BinScans_native}(x)
}
\arguments{
\item{mode}{Character (length 1) with the binning mode. Possible values are \code{scans} and \code{time}.
When \code{scans} is selected, the binning is done by the number of scans. When \code{time} is selected,
the binning is done by time unit (e.g., seconds or minutes).}

\item{value}{Numeric (length 1) with the binning size.}

\item{global}{Logical (length 1) to perform a global binning. If \code{TRUE}, the binning is performed
over all analyses. If \code{FALSE}, the binning is performed for each analysis.}

\item{refAnalysis}{Integer (length 1) with the reference analysis index to use for binning.}

\item{x}{A RamanMethod_BinScans_native object.}
}
\value{
A RamanMethod_BinScans_native object.
}
\description{
Merges scans by averaging spectra according to a given binning size.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate_object(RamanMethod_BinScans_native)}: Validate the RamanMethod_BinScans_native object, returning NULL if valid.

}}
