% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RamanMethod_DeleteScansSection.R
\name{RamanMethod_DeleteScansSection_native}
\alias{RamanMethod_DeleteScansSection_native}
\title{RamanMethod_DeleteScansSection_native Class}
\usage{
RamanMethod_DeleteScansSection_native(min = 0, max = 0)
}
\arguments{
\item{min}{Numeric vector (length 1) with the minimum scan value to delete.}

\item{max}{Numeric vector (length 1) with the maximum scan value to delete.}
}
\value{
A RamanMethod_DeleteScansSection_native object.
}
\description{
Deletes a section of the spectra based on a minimum and maximum scan number. Note
that both values are should be positive and represent the scan unit (e.g., retention time) and
not the scan number.
}
