% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RamanMethod_NormalizeSpectra.R
\name{RamanMethod_NormalizeSpectra_blockweight}
\alias{RamanMethod_NormalizeSpectra_blockweight}
\title{\strong{RamanMethod_NormalizeSpectra_blockweight}}
\usage{
RamanMethod_NormalizeSpectra_blockweight()
}
\value{
A RamanMethod_NormalizeSpectra_blockweight object.
}
\description{
Normalizes spectra using block weighting for downstream data evaluation.
}
