% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RamanMethod_NormalizeSpectra.R
\name{RamanMethod_NormalizeSpectra_minmax}
\alias{RamanMethod_NormalizeSpectra_minmax}
\title{RamanMethod_NormalizeSpectra_minmax Class}
\usage{
RamanMethod_NormalizeSpectra_minmax()
}
\value{
A RamanMethod_NormalizeSpectra_minmax object.
}
\description{
Normalizes spectra using the min-max algorithm.
}
