% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RamanMethod_NormalizeSpectra.R
\name{RamanMethod_NormalizeSpectra_snv}
\alias{RamanMethod_NormalizeSpectra_snv}
\title{RamanMethod_NormalizeSpectra_snv Class}
\usage{
RamanMethod_NormalizeSpectra_snv(liftTozero = FALSE)
}
\arguments{
\item{liftTozero}{Logical (length 1) indicating if the spectra should be lifted to zero.}
}
\value{
A RamanMethod_NormalizeSpectra_snv object.
}
\description{
Normalizes spectra using the Standard Normal Variate (SNV) algorithm.
}
