% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RamanMethod_RemoveCosmicRays.R
\name{RamanMethod_RemoveCosmicRays_native}
\alias{RamanMethod_RemoveCosmicRays_native}
\title{\strong{RamanMethod_RemoveCosmicRays_native}}
\usage{
RamanMethod_RemoveCosmicRays_native()
}
\value{
A RamanMethod_RemoveCosmicRays_native object.
}
\description{
Removes cosmic rays from Raman spectra.
}
