% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RamanMethod_SubtractBlankSpectra.R
\name{RamanMethod_SubtractBlankSpectra_StreamFind}
\alias{RamanMethod_SubtractBlankSpectra_StreamFind}
\title{\strong{RamanMethod_SubtractBlankSpectra_StreamFind}}
\usage{
RamanMethod_SubtractBlankSpectra_StreamFind(negativeToZero = FALSE)
}
\arguments{
\item{negativeToZero}{Logical (length 1) indicating if negative values should be set to zero.}
}
\value{
A RamanMethod_SubtractBlankSpectra_StreamFind object.
}
\description{
Subtracts the blank spectra to each analysis according to the blank assignment.
}
