% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RamanMethod_SubtractScansSection.R
\name{RamanMethod_SubtractScansSection_native}
\alias{RamanMethod_SubtractScansSection_native}
\title{RamanMethod_SubtractScansSection_native Class}
\usage{
RamanMethod_SubtractScansSection_native(sectionWindow = c(10, 200))
}
\arguments{
\item{sectionWindow}{Numeric (length 2) with the window for the sectioning.}
}
\value{
A RamanMethod_SubtractScansSection_native object.
}
\description{
Subtracts the spectra of a section of the scans to all other spectra.
}
