% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RamanSettings_AverageSpectra.R
\name{RamanSettings_AverageSpectra_StreamFind}
\alias{RamanSettings_AverageSpectra_StreamFind}
\title{\strong{RamanSettings_AverageSpectra_StreamFind}}
\usage{
RamanSettings_AverageSpectra_StreamFind(collapseTime = FALSE)
}
\arguments{
\item{collapseTime}{Logical (length 1). When \code{TRUE} the spectra are averaged, reducing the time variable.}
}
\value{
A RamanSettings_AverageSpectra_StreamFind object.
}
\description{
Averages spectra based on analysis replicate groups.
}
