% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_S3_RamanSettings.R
\name{RamanSettings_AverageSpectra_StreamFind}
\alias{RamanSettings_AverageSpectra_StreamFind}
\title{RamanSettings_AverageSpectra_StreamFind}
\usage{
RamanSettings_AverageSpectra_StreamFind(collapseTime = FALSE)
}
\arguments{
\item{collapseTime}{Logical (length 1). When \code{TRUE} the spectra are averaged, reducing the time variable.}
}
\value{
A ProcessingSettings S3 class object with subclass RamanSettings_AverageSpectra_StreamFind.
}
\description{
Averages spectra based on analysis replicate groups.
}
