% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RamanSettings_BinSpectra.R
\name{RamanSettings_BinSpectra_StreamFind}
\alias{RamanSettings_BinSpectra_StreamFind}
\title{\strong{RamanSettings_BinSpectra_StreamFind}}
\usage{
RamanSettings_BinSpectra_StreamFind(
  binNames = c("rt"),
  binValues = c(5),
  byUnit = FALSE,
  refBinAnalysis = NULL
)
}
\arguments{
\item{binNames}{Character with the variable names for building the bins. Possible values are \emph{rt} and \emph{shift}. Note
that the \code{binNames} must be in the spectra column names.}

\item{binValues}{Numeric with the bin values for each variable.}

\item{byUnit}{Logical of length one to bin by unit, meaning that the binning is performed by the number of units not
actual values. For instance, if the bin is 10 and \code{binName} is \emph{rt}, then the binning is performed by 10 seconds not
10 values for each bin. If byUnit is \code{FALSE}, then the binning is performed by the actual values but only the first
of binNames and binValues is used.}

\item{refBinAnalysis}{The analysis index to use a reference for creating the bins.}
}
\value{
A RamanSettings_BinSpectra_StreamFind object.
}
\description{
Bins spectral data based on variables.
}
