% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_S3_RamanSettings.R
\name{RamanSettings_DeleteSpectraSection_StreamFind}
\alias{RamanSettings_DeleteSpectraSection_StreamFind}
\title{RamanSettings_DeleteSpectraSection_StreamFind}
\usage{
RamanSettings_DeleteSpectraSection_StreamFind(section = list())
}
\arguments{
\item{section}{Named list with the variable to be used for sectioning and the window for the sectioning. The names
should match column names in the data.}
}
\value{
A ProcessingSettings S3 class object with subclass RamanSettings_DeleteSpectraSection_StreamFind.
}
\description{
Deletes a section of the spectra based on a named list of data ranges for a given variable (i.e. column name).
}
