% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RamanSettings_NormalizeSpectra.R
\name{RamanSettings_NormalizeSpectra_blockweight}
\alias{RamanSettings_NormalizeSpectra_blockweight}
\title{\strong{RamanSettings_NormalizeSpectra_blockweight}}
\usage{
RamanSettings_NormalizeSpectra_blockweight()
}
\value{
A RamanSettings_NormalizeSpectra_blockweight object.
}
\description{
Normalizes spectra using block weighting for downstream data evaluation.
}
