% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_S3_RamanSettings.R
\name{RamanSettings_NormalizeSpectra_meancenter}
\alias{RamanSettings_NormalizeSpectra_meancenter}
\title{RamanSettings_NormalizeSpectra_meancenter}
\usage{
RamanSettings_NormalizeSpectra_meancenter()
}
\value{
A ProcessingSettings S3 class object with subclass RamanSettings_NormalizeSpectra_meancenter.
}
\description{
Normalizes spectra using mean centering.
}
