% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RamanSettings_NormalizeSpectra.R
\name{RamanSettings_NormalizeSpectra_minmax}
\alias{RamanSettings_NormalizeSpectra_minmax}
\title{\strong{RamanSettings_NormalizeSpectra_minmax}}
\usage{
RamanSettings_NormalizeSpectra_minmax()
}
\value{
A RamanSettings_NormalizeSpectra_minmax object.
}
\description{
Normalizes spectra using the min-max algorithm.
}
