% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_S3_RamanSettings.R
\name{RamanSettings_NormalizeSpectra_scale}
\alias{RamanSettings_NormalizeSpectra_scale}
\title{RamanSettings_NormalizeSpectra_scale}
\usage{
RamanSettings_NormalizeSpectra_scale()
}
\value{
A ProcessingSettings S3 class object with subclass RamanSettings_NormalizeSpectra_scale.
}
\description{
Normalizes spectra using scaling based on the standard deviation.
}
