% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_S3_RamanSettings.R
\name{RamanSettings_SubtractBlankSpectra_StreamFind}
\alias{RamanSettings_SubtractBlankSpectra_StreamFind}
\title{RamanSettings_SubtractBlankSpectra_StreamFind}
\usage{
RamanSettings_SubtractBlankSpectra_StreamFind(negativeToZero = FALSE)
}
\arguments{
\item{negativeToZero}{Logical (length 1) indicating if negative values should be set to zero.}
}
\value{
A ProcessingSettings S3 class object with subclass RamanSettings_SubtractBlankSpectra_StreamFind.
}
\description{
Subtracts the blank spectra to each analysis according to the blank assignment.
}
