% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_RamanSettings_SubtractSpectraSection.R
\name{RamanSettings_SubtractSpectraSection_StreamFind}
\alias{RamanSettings_SubtractSpectraSection_StreamFind}
\title{\strong{RamanSettings_SubtractSpectraSection_StreamFind}}
\usage{
RamanSettings_SubtractSpectraSection_StreamFind(
  sectionVal = "rt",
  sectionWindow = c(10, 200)
)
}
\arguments{
\item{sectionVal}{Character (length 1) with the variable to be used for sectioning.}

\item{sectionWindow}{Numeric (length 2) with the window for the sectioning.}
}
\value{
A RamanSettings_SubtractSpectraSection_StreamFind object.
}
\description{
Subtracts a section of the spectra based on a variable (i.e. column name).
}
