% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Results.R
\name{Results}
\alias{Results}
\title{Generic Results}
\usage{
Results(
  data_type = NA_character_,
  name = NA_character_,
  software = NA_character_,
  version = NA_character_
)
}
\description{
The Results class is used to store results of data processing in
\link{Analyses} child classes. Child classes of Results are implemented for diverse types
of results for a given type of data.
}
\section{Slots}{

\describe{
\item{\code{data_type}}{A character string indicating the type of data.}

\item{\code{name}}{A character string representing the name of the results.}

\item{\code{software}}{A character string representing the name of the software used to generate the results.}

\item{\code{version}}{A character string representing the version of the software used to generate the results.}
}}

