% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Results.R
\name{Results}
\alias{Results}
\title{Generic Results}
\usage{
Results(
  name = class_missing,
  software = class_missing,
  version = class_missing
)
}
\arguments{
\item{name}{A character string representing the name of the results.}

\item{software}{A character string representing the name of the software used to generate the results.}

\item{version}{A character string representing the version of the software used to generate the results.}
}
\description{
The Results class is used to store results of data processing in
\link{Analyses} child classes. Child classes of Results are implemented for diverse types
of results.
}
