% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_default_ProcessingSettings.R
\name{save_default_ProcessingSettings}
\alias{save_default_ProcessingSettings}
\title{save_default_ProcessingSettings}
\usage{
save_default_ProcessingSettings(
  engine = NA_character_,
  call = NA_character_,
  algorithm = NA_character_,
  format = "json",
  name = "settings",
  path = getwd()
)
}
\arguments{
\item{call}{Character (length 1) with the method call name.}

\item{algorithm}{Character (length 1) with the algorithm name.}

\item{format}{Character (length 1) with the format of the saved file. Possible are "json" and "rds".}

\item{name}{Character (length 1) with the name of the file without extension.}

\item{path}{Character (length 1) with the saving path. The default is the \code{getwd()} path.}
}
\value{
Creates a json/rds files on the defined path.
}
\description{
Saves on disk a ProcessingSettings S3 class object for \code{call} and \code{algorithm} as the defined \code{format}
in the defined \code{path} and with the defined \code{name}.
}
