% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_S3_settings.R
\name{settings}
\alias{settings}
\alias{validate.settings}
\alias{as.settings}
\title{\strong{settings} S3 class constructor and methods}
\usage{
settings(call = NA_character_, algorithm = NA_character_, parameters = NULL)

\method{validate}{settings}(settings)

as.settings(value)
}
\arguments{
\item{call}{Character of length one with the name of the method where the
settings are to be applied.}

\item{algorithm}{Character of length one with the name of the algorithm to
be used.}

\item{parameters}{List with parameter settings specific for the \code{call}
method and used \code{algorithm}.}

\item{settings}{A settings S3 class object.}

\item{value}{A list to be checked and/or converted to settings S3 class.}
}
\value{
A settings S3 class
}
\description{
Creates a settings S3 class.
}
\details{
See the method documentation for more information about possible
algorithms and parameters to be used.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate(settings)}: Validates a settings S3 class object, returning a logical value of length
one.

}}
\section{Functions}{
\itemize{
\item \code{as.settings()}: Converts the argument in a settings S3 class object

}}
