% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_005_default_ProcessingSettings.R
\name{Settings_annotate_features_streamFind}
\alias{Settings_annotate_features_streamFind}
\alias{validate.Settings_annotate_features_streamFind}
\title{Settings_annotate_features_streamFind}
\usage{
Settings_annotate_features_streamFind(
  maxIsotopes = 5,
  elements = c("C", "H", "N", "O", "S", "Cl", "Br"),
  mode = "small molecules",
  maxCharge = 1,
  rtWindowAlignment = 0.3,
  maxGaps = 1,
  runParallel = FALSE
)

\method{validate}{Settings_annotate_features_streamFind}(x)
}
\arguments{
\item{maxIsotopes}{Numeric (length 1) with the maximum number of isotopic steps.}

\item{elements}{Character vector with the elements to target the isotopic annotation.
Possible elements are C, H, N, O, S, Cl, Br.}

\item{mode}{Character (length 1) with the type of molecules to be targeted.
For now, only "small molecules" are possible.}

\item{maxCharge}{Numeric (length 1) with the maximum charge that ions can be
ionized to find isotopes.}

\item{rtWindowAlignment}{Numeric (length 1) with the proportion of the
monoisotopic feature time window to be used for retrieving isotopic candidates.}

\item{maxGaps}{Numeric (length 1) with the maximum of allowed gaps in isotopic chains.}

\item{runParallel}{Logical of length one. Set to \code{TRUE} for parallel
processing. Note that parallel processing is not always faster due to
overhead.}

\item{x}{A Settings_annotate_features_streamFind S3 class object.}
}
\value{
A ProcessingSettings S3 class object with subclass
Settings_annotate_features_streamFind.
}
\description{
Settings for annotation of isotopic features. The method uses the \code{maxIsotopes}
to define the maximum length of the isotopic chain. The list of candidate features is
build with the \code{rtWindowAlignment} and the maximum mass increment to match the
maximum chain length. Then, the mass difference  of the natural isotopes defined
by \code{elements} and a given monoisotopic ion (i.e., feature) are targeted.
Each candidate is then evaluated according to the mass error and the expected
relative intensity range as defined by the \code{mode}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate(Settings_annotate_features_streamFind)}: Validates the Settings_annotate_features_streamFind S3 class object, returning a logical
value of length one.

}}
