% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_default_ProcessingSettings.R
\name{Settings_bin_spectra_qBinning}
\alias{Settings_bin_spectra_qBinning}
\title{Settings_bin_spectra_qBinning}
\usage{
Settings_bin_spectra_qBinning()
}
\value{
X.
}
\description{
X.
}
