% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ProcessingSettings.R
\name{Settings_bin_spectra_StreamFind}
\alias{Settings_bin_spectra_StreamFind}
\alias{validate.Settings_bin_spectra_StreamFind}
\title{Settings_bin_spectra_StreamFind}
\usage{
Settings_bin_spectra_StreamFind(
  unitsVal = NULL,
  unitsNumber = NULL,
  bins = NULL,
  refBinAnalysis = NULL
)

\method{validate}{Settings_bin_spectra_StreamFind}(x)
}
\arguments{
\item{unitsVal}{Character of length one with the column name of the variable to be used for binning.}

\item{unitsNumber}{Integer of length one with the number of units to be used for binning.}

\item{bins}{Named list of numeric values with the bin dimensions.}

\item{refBinAnalysis}{The analysis index to use a reference for creating the bins.}
}
\value{
A ProcessingSettings S3 class object with subclass Settings_bin_spectra_StreamFind.
}
\description{
Bins spectral data according to units of a given variable (e.g., 5 retention time values) or based on
bins given as a named list of numeric values, where the names are the bin labels (i.e. the name of the column) and
the values are the bin dimensions (e.g. 5 seconds).
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate(Settings_bin_spectra_StreamFind)}: Validates the object structure, returning a logical value of length one.

}}
