% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_default_ProcessingSettings.R
\name{Settings_calculate_quality_StreamFind}
\alias{Settings_calculate_quality_StreamFind}
\alias{validate.Settings_calculate_quality_StreamFind}
\title{Settings_calculate_quality_StreamFind}
\usage{
Settings_calculate_quality_StreamFind(
  rtExpand = 120,
  mzExpand = 3e-04,
  minTraces = 6,
  filtered = FALSE,
  runParallel = TRUE
)

\method{validate}{Settings_calculate_quality_StreamFind}(x)
}
\arguments{
\item{rtExpand}{Numeric value with the retention time values (in seconds) for
expanding left and right the mass range.}

\item{mzExpand}{Numeric value with the mass or mass-to-charge ratio
(\emph{m/z}) (in Da) for expanding left and right the mass range.}

\item{minTraces}{Numeric of length 1 with the minimum number traces for
calculating feature quality.}

\item{filtered}{Logical. When \code{TRUE}, filtered features/groups are considered.}

\item{runParallel}{Logical of length one. Set to \code{TRUE} for parallel
processing. Note that parallel processing is not always faster due to
overhead.}

\item{x}{A Settings_calculate_quality_StreamFind S3 class object.}
}
\value{
A ProcessingSettings S3 class object with subclass
Settings_calculate_quality_StreamFind.
}
\description{
Settings for calculating quality parameters of features (e.g., signal-to-noise (sn) ratio).
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate(Settings_calculate_quality_StreamFind)}: Validates the object structure, returning a logical value of length one.

}}
