% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ProcessingSettings.R
\name{Settings_calculate_quality_StreamFind}
\alias{Settings_calculate_quality_StreamFind}
\title{Settings_calculate_quality_StreamFind}
\usage{
Settings_calculate_quality_StreamFind(
  rtExpand = 120,
  mzExpand = 3e-04,
  minTraces = 6,
  filtered = FALSE,
  runParallel = TRUE
)
}
\arguments{
\item{rtExpand}{Numeric (length 1) with the retention time values (in seconds) for expanding left and right the
retention time range.}

\item{mzExpand}{Numeric (length 1) with the mass or mass-to-charge ratio (\emph{m/z}) (in Da) for expanding left and
right the mass range.}

\item{minTraces}{Numeric of length 1 with the minimum number traces for calculating feature quality.}

\item{filtered}{Logical (length 1). When \code{TRUE}, filtered features/groups are considered.}

\item{runParallel}{Logical of length one. Set to \code{TRUE} for parallel
processing. Note that parallel processing is not always faster due to
overhead.}
}
\value{
A ProcessingSettings S3 class object with subclass Settings_calculate_quality_StreamFind.
}
\description{
Settings for calculating quality parameters of features (e.g., signal-to-noise (sn) ratio).
}
