% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_default_ProcessingSettings.R
\name{Settings_calculate_spectra_charges_StreamFind}
\alias{Settings_calculate_spectra_charges_StreamFind}
\alias{validate.Settings_calculate_spectra_charges_StreamFind}
\title{Settings_calculate_spectra_charges_StreamFind}
\usage{
Settings_calculate_spectra_charges_StreamFind(
  roundVal = 35,
  relLowCut = 0.2,
  absLowCut = 300
)

\method{validate}{Settings_calculate_spectra_charges_StreamFind}(x)
}
\arguments{
\item{roundVal}{}

\item{relLowCut}{}

\item{absLowCut}{}

\item{x}{A Settings_calculate_spectra_charges_StreamFind S3 class object.}
}
\value{
A ProcessingSettings S3 class object with subclass Settings_calculate_spectra_charges_StreamFind.
}
\description{
Prototype.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate(Settings_calculate_spectra_charges_StreamFind)}: Validates the object structure, returning a logical value of length one.

}}
