% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_default_ProcessingSettings.R
\name{Settings_centroid_spectra_qCentroids}
\alias{Settings_centroid_spectra_qCentroids}
\title{Settings_centroid_spectra_qCentroids}
\usage{
Settings_centroid_spectra_qCentroids()
}
\value{
X.
}
\description{
X.
}
