% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ProcessingSettings.R
\name{Settings_cluster_spectra_StreamFind}
\alias{Settings_cluster_spectra_StreamFind}
\alias{validate.Settings_cluster_spectra_StreamFind}
\title{Settings_cluster_spectra_StreamFind}
\usage{
Settings_cluster_spectra_StreamFind(
  val = "mz",
  clustVal = 0.001,
  presence = 0.1
)

\method{validate}{Settings_cluster_spectra_StreamFind}(x)
}
\arguments{
\item{val}{}

\item{clustVal}{}

\item{presence}{}

\item{x}{A Settings_cluster_spectra_StreamFind S3 class object.}
}
\value{
A ProcessingSettings S3 class object with subclass Settings_cluster_spectra_StreamFind.
}
\description{
Prototype.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate(Settings_cluster_spectra_StreamFind)}: Validates the object structure, returning a logical value of length one.

}}
