% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ProcessingSettings.R
\name{Settings_correct_chromatograms_baseline_airpls}
\alias{Settings_correct_chromatograms_baseline_airpls}
\title{Settings_correct_chromatograms_baseline_airpls}
\usage{
Settings_correct_chromatograms_baseline_airpls(
  lambda = 10,
  differences = 1,
  itermax = 20
)
}
\arguments{
\item{lambda}{Numeric (length 1) with the smoothing intensity. the higher the \code{lambda} the higher the smoothing.}

\item{differences}{Integer (length 1) indicating the order of the difference of penalties}

\item{itermax}{Integer (length 1) with the maximum number of iterations.}
}
\value{
A ProcessingSettings S3 class object with subclass Settings_correct_chromatograms_baseline_airpls.
}
\description{
Performs baseline correction using adaptive iteratively reweighted Penalized Least Squares (airPLS)
based on the algorithm from Zhi-Min Zhang.
}
\references{
\insertRef{airpls01}{StreamFind}
}
