% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_default_ProcessingSettings.R
\name{Settings_deconvolute_spectra_charges_StreamFind}
\alias{Settings_deconvolute_spectra_charges_StreamFind}
\alias{validate.Settings_deconvolute_spectra_charges_StreamFind}
\title{Settings_deconvolute_spectra_charges_StreamFind}
\usage{
Settings_deconvolute_spectra_charges_StreamFind(
  rtmin = 315 - 2.5,
  rtmax = 315 + 2.5,
  mzmin = 2400,
  mzmax = 3800,
  presence = 0.1,
  mzClust = 0.001,
  minIntensity = 50,
  roundVal = 35,
  relLowCut = 0.2,
  absLowCut = 300,
  mzClustAverage = 0.1,
  smoothing = TRUE,
  windowSize = 25,
  baseline = TRUE,
  baseline_method = "als",
  baseline_args = list(lambda = 9, p = 0.02, maxit = 10),
  merge = TRUE,
  closeByThreshold = 45,
  valeyThreshold = 0.5,
  minPeakHeight = 100,
  minPeakDistance = 50,
  maxPeakWidth = 250,
  minPeakWidth = 50
)

\method{validate}{Settings_deconvolute_spectra_charges_StreamFind}(x)
}
\arguments{
\item{rtmin}{}

\item{rtmax}{}

\item{mzmin}{}

\item{mzmax}{}

\item{presence}{presence can be adjusted to remove noise}

\item{mzClust}{critical for resolution, in Da}

\item{minIntensity}{}

\item{roundVal}{}

\item{relLowCut}{}

\item{absLowCut}{}

\item{mzClustAverage}{}

\item{smoothing}{}

\item{windowSize}{}

\item{baseline}{}

\item{baseline_method}{}

\item{baseline_args}{}

\item{merge}{}

\item{closeByThreshold}{}

\item{valeyThreshold}{}

\item{minPeakHeight}{}

\item{minPeakDistance}{}

\item{maxPeakWidth}{}

\item{minPeakWidth}{}

\item{x}{A Settings_integrate_chromatograms_StreamFind S3 class object.}
}
\value{
A ProcessingSettings S3 class object with subclass Settings_deconvolute_spectra_charges_StreamFind.
}
\description{
Prototype.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate(Settings_deconvolute_spectra_charges_StreamFind)}: Validates the object structure, returning a logical value of length one.

}}
