% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ProcessingSettings.R
\name{Settings_deconvolute_spectra_StreamFind}
\alias{Settings_deconvolute_spectra_StreamFind}
\alias{validate.Settings_deconvolute_spectra_StreamFind}
\title{Settings_deconvolute_spectra_StreamFind}
\usage{
Settings_deconvolute_spectra_StreamFind(clustVal = 0.1, window = 20)

\method{validate}{Settings_deconvolute_spectra_StreamFind}(x)
}
\arguments{
\item{clustVal}{}

\item{window}{}

\item{x}{A Settings_deconvolute_spectra_StreamFind S3 class object.}
}
\value{
A ProcessingSettings S3 class object with subclass Settings_deconvolute_spectra_StreamFind.
}
\description{
Prototype.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate(Settings_deconvolute_spectra_StreamFind)}: Validates the object structure, returning a logical value of length one.

}}
