% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ProcessingSettings.R
\name{Settings_delete_spectra_section_StreamFind}
\alias{Settings_delete_spectra_section_StreamFind}
\title{Settings_delete_spectra_section_StreamFind}
\usage{
Settings_delete_spectra_section_StreamFind(section = list())
}
\arguments{
\item{section}{Named list with the variable to be used for sectioning and the window for the sectioning. The names
should match column names in the data.}
}
\value{
A ProcessingSettings S3 class object with subclass Settings_delete_spectra_section_StreamFind.
}
\description{
Deletes a section of the spectra based on a named list of data ranges for a given variable (i.e. column name).
}
