% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ProcessingSettings.R
\name{Settings_fill_features_StreamFind}
\alias{Settings_fill_features_StreamFind}
\title{Settings_fill_features_StreamFind}
\usage{
Settings_fill_features_StreamFind(minSnRatio = 3, excludeIsotopes = TRUE)
}
\value{
A ProcessingSettings S3 class object with subclass Settings_fill_features_StreamFind.
}
\description{
Settings for filling missing values in features.
}
