% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_005_default_ProcessingSettings.R
\name{Settings_filter_features_streamFind}
\alias{Settings_filter_features_streamFind}
\alias{validate.Settings_filter_features_streamFind}
\title{Settings_filter_features_streamFind}
\usage{
Settings_filter_features_streamFind(
  minIntensity = 5000,
  minSnRatio = 25,
  maxGroupSd = 30,
  blank = 5,
  minGroupAbundance = 3,
  excludeIsotopes = TRUE
)

\method{validate}{Settings_filter_features_streamFind}(x)
}
\arguments{
\item{x}{A Settings_filter_features_streamFind S3 class object.}
}
\value{
X.
}
\description{
X.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate(Settings_filter_features_streamFind)}: Validates the Settings_filter_features_streamFind S3 class object, returning
a logical value of length one.

}}
