% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_005_default_ProcessingSettings.R
\name{Settings_filter_features_StreamFind}
\alias{Settings_filter_features_StreamFind}
\alias{validate.Settings_filter_features_StreamFind}
\title{Settings_filter_features_StreamFind}
\usage{
Settings_filter_features_StreamFind(
  minIntensity = NULL,
  minSnRatio = NULL,
  maxGroupSd = NULL,
  blank = NULL,
  minGroupAbundance = NULL,
  excludeIsotopes = NULL,
  rtFilter = NULL,
  massFilter = NULL
)

\method{validate}{Settings_filter_features_StreamFind}(x)
}
\arguments{
\item{minIntensity}{Numeric (length 1) with the minimum intensity.}

\item{minSnRatio}{Numeric (length 1) with the minimum signal-to-noise ratio.
The filter is only applied if the features data.table contains the column "sn".}

\item{maxGroupSd}{Numeric (length 1) with the maximum intensity deviation
within each analysis replicate (in percentage).}

\item{blank}{Numeric (length 1) with the intensity threshold for blank
subtraction. All features/feature groups not higher then the \code{blank} * its
intensity are filtered.}

\item{minGroupAbundance}{Numeric (length 1) with the minimum presence of a
feature is a given analysis replicate.}

\item{excludeIsotopes}{Logical (length 1) with \code{TRUE} for filtering
annotated isotopes (only prevails the monoisotopic features).}

\item{rtFilter}{Numeric (length 2) with the min and max retention time
(in seconds) values to filter features. Features within the retention time
range are filtered out.}

\item{massFilter}{Numeric (length 2) with the min and max mass
(in Da) values to filter features. Features within the mass range are
filtered out.}

\item{x}{A Settings_filter_features_StreamFind S3 class object.}
}
\value{
A ProcessingSettings S3 class object with subclass
Settings_filter_features_StreamFind.
}
\description{
Settings for filtering of features and feature groups.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate(Settings_filter_features_StreamFind)}: Validates the Settings_filter_features_StreamFind S3 class object, returning
a logical value of length one.

}}
