% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ProcessingSettings.R
\name{Settings_filter_features_StreamFind}
\alias{Settings_filter_features_StreamFind}
\title{Settings_filter_features_StreamFind}
\usage{
Settings_filter_features_StreamFind(minSnRatio = 3, excludeIsotopes = TRUE)
}
\arguments{
\item{minSnRatio}{Numeric (length 1) with the minimum signal-to-noise ratio.}

\item{excludeIsotopes}{Logical (length 1) with \code{TRUE} for filtering annotated isotopes (only prevails the monoisotopic features).}
}
\value{
A ProcessingSettings S3 class object with subclass Settings_filter_features_StreamFind.
}
\description{
Settings for filtering of features and feature groups.
}
