% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_default_ProcessingSettings.R
\name{Settings_filter_features_StreamFind}
\alias{Settings_filter_features_StreamFind}
\alias{validate.Settings_filter_features_StreamFind}
\title{Settings_filter_features_StreamFind}
\usage{
Settings_filter_features_StreamFind(...)

\method{validate}{Settings_filter_features_StreamFind}(x)
}
\arguments{
\item{...}{Ordered filters to be applied. Possible filter arguments are:
\itemize{
\item{\strong{minIntensity}}{  Numeric (length 1) with the minimum intensity.}
\item{\strong{minSnRatio}}{  Numeric (length 1) with the minimum signal-to-noise
ratio.}
\item{\strong{maxGroupSd}}{  Numeric (length 1) with the maximum intensity
deviation within each analysis replicate (in percentage).}
\item{\strong{blank}}{  Numeric (length 1) with the intensity threshold for blank
subtraction. All features/feature groups not higher then the \code{blank} * its
intensity are filtered.}
\item{\strong{minGroupAbundance}}{  Numeric (length 1) with the minimum presence
of a feature is a given analysis replicate.}
\item{\strong{excludeIsotopes}}{  Logical (length 1) with \code{TRUE} for filtering
annotated isotopes (only prevails the monoisotopic features).}
\item{\strong{rtFilter}}{  Numeric (length 2) with the min and max retention time
(in seconds) values to filter features. Features within the retention time
range are filtered out.}
\item{\strong{massFilter}}{  Numeric (length 2) with the min and max mass
(in Da) values to filter features. Features within the mass range are
filtered out.}
\item{\strong{onlySuspects}}{  Logical (length 1) with \code{TRUE} for only keeping
features annotated with a suspect.}
}}

\item{x}{A Settings_filter_features_StreamFind S3 class object.}
}
\value{
A ProcessingSettings S3 class object with subclass
Settings_filter_features_StreamFind.
}
\description{
Settings for filtering of features and feature groups.
}
\details{
When feature groups exist the filtering is done on feature groups, meaning
that the filters are applied per feature group not individual features. For
instance, if feature groups exist the minimum intensity is applied on the
maximum intensity observed for a feature group and not the intensity of each
individual feature.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate(Settings_filter_features_StreamFind)}: Validates the object structure, returning a logical value of length one.

}}
