% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_005_default_ProcessingSettings.R
\name{settings_find_features_kpic2}
\alias{settings_find_features_kpic2}
\title{settings_find_features_kpic2}
\usage{
settings_find_features_kpic2()
}
\value{
X.
}
\description{
X.
}
