% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ProcessingSettings.R
\name{Settings_find_features_kpic2}
\alias{Settings_find_features_kpic2}
\title{Settings_find_features_kpic2}
\usage{
Settings_find_features_kpic2(
  level = 500,
  mztol = 0.01,
  gap = 2,
  width = 5,
  min_snr = 4,
  kmeans = TRUE,
  alpha = 0.3
)
}
\arguments{
\item{level}{Mass traces are only retained if their maximum values are over \code{level}.}

\item{mztol}{The initial m/z tolerance.}

\item{gap}{The number of gap points of a mass trace.}

\item{width}{The minimum length of a mass trace.}

\item{min_snr}{Minimum signal to noise ratio.}

\item{kmeans}{If \code{TRUE}, \link[KPIC]{getPIC.kmeans} is used to obtain
PICs (i.e., features). If \code{FALSE}, \link[KPIC]{getPIC} is used.}

\item{alpha}{If \code{kmeans} is \code{TRUE}, alpha is the parameter of forecasting.
If \code{kmeans} is \code{FALSE}, alpha is not used.}
}
\value{
A ProcessingSettings S3 class object with subclass Settings_find_features_kpic2.
}
\description{
Settings for finding features (i.e., chromatographic peaks) in mzML/mzXML files using the package
\href{https://github.com/hcji/KPIC2}{KPIC}. The function uses the package \pkg{patRoon} in the background.
}
\details{
See the \link[patRoon]{findFeaturesKPIC2} function from the \pkg{patRoon} package for more information and requirements.
}
\references{
\insertRef{patroon01}{StreamFind}

\insertRef{patroon02}{StreamFind}

\insertRef{kpic01}{StreamFind}
}
