% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_005_default_ProcessingSettings.R
\name{Settings_find_features_qPeaks}
\alias{Settings_find_features_qPeaks}
\title{Settings_find_features_qPeaks}
\usage{
Settings_find_features_qPeaks()
}
\value{
X.
}
\description{
X.
}
