% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_default_ProcessingSettings.R
\name{Settings_generate_compounds_metfrag}
\alias{Settings_generate_compounds_metfrag}
\alias{validate.Settings_generate_compounds_metfrag}
\title{Settings_generate_compounds_metfrag}
\usage{
Settings_generate_compounds_metfrag(
  method = "CL",
  timeout = 300,
  timeoutRetries = 5,
  errorRetries = 5,
  topMost = 5,
  dbRelMzDev = 8,
  fragRelMzDev = 10,
  fragAbsMzDev = 0.002,
  adduct = NULL,
  database = "comptox",
  extendedPubChem = "auto",
  chemSpiderToken = "",
  scoreTypes = patRoon::compoundScorings("metfrag", "comptox", onlyDefault = TRUE)$name,
  scoreWeights = 1,
  preProcessingFilters = c("UnconnectedCompoundFilter", "IsotopeFilter"),
  postProcessingFilters = c("InChIKeyFilter"),
  maxCandidatesToStop = 100,
  identifiers = NULL,
  extraOpts = NULL
)

\method{validate}{Settings_generate_compounds_metfrag}(x)
}
\arguments{
\item{method}{}

\item{timeout}{}

\item{timeoutRetries}{}

\item{errorRetries}{}

\item{topMost}{}

\item{dbRelMzDev}{}

\item{fragRelMzDev}{}

\item{fragAbsMzDev}{}

\item{adduct}{}

\item{database}{}

\item{extendedPubChem}{}

\item{chemSpiderToken}{}

\item{scoreTypes}{}

\item{scoreWeights}{}

\item{preProcessingFilters}{}

\item{postProcessingFilters}{}

\item{maxCandidatesToStop}{}

\item{identifiers}{}

\item{extraOpts}{}

\item{x}{A Settings_generate_compounds_metfrag S3 class object.}
}
\value{
A ProcessingSettings S3 class object with subclass Settings_generate_compounds_metfrag.
}
\description{
Settings for generating compounds using \href{https://ipb-halle.github.io/MetFrag/}{MetFrag}.
The algorithm is used via the function \link[patRoon]{generateCompounds} from the package \pkg{patRoon}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate(Settings_generate_compounds_metfrag)}: Validates the object structure, returning a logical value of length one.

}}
\references{
\insertRef{metfrag01}{StreamFind}

\insertRef{metfrag02}{StreamFind}

\insertRef{metfrag03}{StreamFind}

\insertRef{metfrag04}{StreamFind}

\insertRef{patroon01}{StreamFind}

\insertRef{patroon02}{StreamFind}
}
