% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_default_ProcessingSettings.R
\name{Settings_generate_formulas_genform}
\alias{Settings_generate_formulas_genform}
\alias{validate.Settings_generate_formulas_genform}
\title{Settings_generate_formulas_genform}
\usage{
Settings_generate_formulas_genform(
  relMzDev = 5,
  elements = "CHNOP",
  hetero = TRUE,
  oc = FALSE,
  thrMS = NULL,
  thrMSMS = NULL,
  thrComb = NULL,
  maxCandidates = Inf,
  extraOpts = NULL,
  calculateFeatures = TRUE,
  featThreshold = 0,
  featThresholdAnn = 0.75,
  absAlignMzDev = 0.002,
  MSMode = "both",
  isolatePrec = TRUE,
  timeout = 120,
  topMost = 50,
  batchSize = 8
)

\method{validate}{Settings_generate_formulas_genform}(x)
}
\arguments{
\item{relMzDev}{}

\item{elements}{}

\item{hetero}{}

\item{oc}{}

\item{thrMS}{}

\item{thrMSMS}{}

\item{thrComb}{}

\item{maxCandidates}{}

\item{extraOpts}{}

\item{calculateFeatures}{}

\item{featThreshold}{}

\item{featThresholdAnn}{}

\item{absAlignMzDev}{}

\item{MSMode}{}

\item{isolatePrec}{}

\item{timeout}{}

\item{topMost}{}

\item{batchSize}{}

\item{x}{A Settings_generate_formulas_genform S3 class object.}
}
\value{
A ProcessingSettings S3 class object with subclass Settings_generate_formulas_genform.
}
\description{
Settings for generating formulas using the algorithm
\href{https://sourceforge.net/projects/genform/}{GenForm}.
The algorithm is used via the function \link[patRoon]{generateFormulas}
from the package \pkg{patRoon}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate(Settings_generate_formulas_genform)}: Validates the object structure, returning a logical value of length one.

}}
\references{
\insertRef{patroon01}{StreamFind}

\insertRef{patroon02}{StreamFind}

\insertRef{genform}{StreamFind}
}
