% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_005_default_ProcessingSettings.R
\name{Settings_load_features_ms2_streamFind}
\alias{Settings_load_features_ms2_streamFind}
\alias{validate.Settings_load_features_ms2_streamFind}
\title{Settings_load_features_ms2_streamFind}
\usage{
Settings_load_features_ms2_streamFind(
  isolationWindow = 1.3,
  mzClust = 0.003,
  minIntensity = 0,
  filtered = FALSE,
  runParallel = FALSE,
  verbose = FALSE
)

\method{validate}{Settings_load_features_ms2_streamFind}(x)
}
\arguments{
\item{isolationWindow}{Numeric value with the isolation window (in Da)
applied for ion isolation and fragmentation during acquisition of tandem
data (i.e., MS2 data).}

\item{mzClust}{Numeric value with the mass deviation threshold (in Da) to
cluster mass traces. Highly dependent on the mass resolution of the MS data.}

\item{minIntensity}{Numeric value with the minimum intensity.}

\item{filtered}{Logical. When \code{TRUE}, filtered features/groups are considered.}

\item{runParallel}{Logical of length one. Set to \code{TRUE} for parallel
processing. Note that parallel processing is not always faster due to
overhead.}

\item{verbose}{Logical with length one. Set \code{TRUE} to output diagnostic and
result messages. When \code{FALSE}, some messages are omitted.}

\item{x}{A Settings_load_features_ms2_streamFind S3 class object.}
}
\value{
A ProcessingSettings S3 class object with subclass
Settings_load_features_ms2_streamFind.
}
\description{
Settings for loading MS2 spectra for features.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate(Settings_load_features_ms2_streamFind)}: Validates the Settings_load_features_ms1_streamFind S3 class object,
returning a logical value of length one.

}}
