% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_005_default_ProcessingSettings.R
\name{Settings_load_features_ms2_streamFind}
\alias{Settings_load_features_ms2_streamFind}
\alias{validate.Settings_load_features_ms2_streamFind}
\title{Settings_load_features_ms2_streamFind}
\usage{
Settings_load_features_ms2_streamFind()

\method{validate}{Settings_load_features_ms2_streamFind}(x)
}
\arguments{
\item{x}{A Settings_load_features_ms2_streamFind S3 class object.}
}
\value{
X.
}
\description{
X.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate(Settings_load_features_ms2_streamFind)}: Validates the Settings_load_features_ms1_streamFind S3 class object,
returning a logical value of length one.

}}
