% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_default_ProcessingSettings.R
\name{Settings_merge_spectra_time_series_StreamFind}
\alias{Settings_merge_spectra_time_series_StreamFind}
\alias{validate.Settings_merge_spectra_time_series_StreamFind}
\title{Settings_merge_spectra_time_series_StreamFind}
\usage{
Settings_merge_spectra_time_series_StreamFind(preCut = list())

\method{validate}{Settings_merge_spectra_time_series_StreamFind}(x)
}
\arguments{
\item{preCut}{The number of pre Raman scans to exclude when merging.}

\item{x}{A Settings_merge_spectra_time_series_StreamFind S3 class object.}
}
\value{
A ProcessingSettings S3 class object with subclass Settings_merge_spectra_time_series_StreamFind.
}
\description{
Prototype.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate(Settings_merge_spectra_time_series_StreamFind)}: Validates the object structure, returning a logical value of length one.

}}
