% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ProcessingSettings.R
\name{Settings_normalize_spectra_snv}
\alias{Settings_normalize_spectra_snv}
\alias{validate.Settings_normalize_spectra_snv}
\title{Settings_normalize_spectra_snv}
\usage{
Settings_normalize_spectra_snv(liftTozero = FALSE)

\method{validate}{Settings_normalize_spectra_snv}(x)
}
\arguments{
\item{liftTozero}{}

\item{x}{A Settings_normalize_spectra_snv S3 class object.}
}
\value{
A ProcessingSettings S3 class object with subclass Settings_normalize_spectra_snv.
}
\description{
Prototype.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate(Settings_normalize_spectra_snv)}: Validates the object structure, returning a logical value of length one.

}}
