% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ProcessingSettings.R
\name{Settings_smooth_chromatograms_movingaverage}
\alias{Settings_smooth_chromatograms_movingaverage}
\title{Settings_smooth_chromatograms_movingaverage}
\usage{
Settings_smooth_chromatograms_movingaverage(windowSize = 5, xValWindow = NULL)
}
\arguments{
\item{windowSize}{Numeric (length 1) with the window size for the moving average.}
}
\value{
A ProcessingSettings S3 class object with subclass Settings_smooth_chromatograms_movingaverage.
}
\description{
Smooths chromatograms using the moving average algorithm.
}
