% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ProcessingSettings.R
\name{Settings_subtract_blank_spectra_StreamFind}
\alias{Settings_subtract_blank_spectra_StreamFind}
\title{Settings_subtract_blank_spectra_StreamFind}
\usage{
Settings_subtract_blank_spectra_StreamFind(negativeToZero = FALSE)
}
\arguments{
\item{negativeToZero}{Logical (length 1) indicating if negative values should be set to zero.}
}
\value{
A ProcessingSettings S3 class object with subclass Settings_subtract_blank_spectra_StreamFind.
}
\description{
Subtracts the blank spectra to each analysis according to the blank assignment.
}
