% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ProcessingSettings.R
\name{Settings_subtract_spectra_section_StreamFind}
\alias{Settings_subtract_spectra_section_StreamFind}
\alias{validate.Settings_subtract_spectra_section_StreamFind}
\title{Settings_subtract_spectra_section_StreamFind}
\usage{
Settings_subtract_spectra_section_StreamFind(
  sectionVal = "rt",
  sectionWindow = c(10, 200)
)

\method{validate}{Settings_subtract_spectra_section_StreamFind}(x)
}
\arguments{
\item{sectionVal}{}

\item{sectionWindow}{}

\item{x}{A Settings_subtract_spectra_section_StreamFind S3 class object.}
}
\value{
A ProcessingSettings S3 class object with subclass Settings_subtract_spectra_section_StreamFind.
}
\description{
Prototype.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate(Settings_subtract_spectra_section_StreamFind)}: Validates the object structure, returning a logical value of length one.

}}
