% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ms_default_ProcessingSettings.R
\name{Settings_suspect_screening_StreamFind}
\alias{Settings_suspect_screening_StreamFind}
\alias{validate.Settings_suspect_screening_StreamFind}
\title{Settings_suspect_screening_StreamFind}
\usage{
Settings_suspect_screening_StreamFind(database = NULL, ppm = 4, sec = 10)

\method{validate}{Settings_suspect_screening_StreamFind}(x)
}
\arguments{
\item{database}{A data.frame with at least the columns name and mass,
indicating the name and neutral monoisotopic mass of the suspect targets.}

\item{ppm}{Numeric of length one with the mass deviation, in ppm.}

\item{sec}{Numeric of length one with the time deviation, in seconds.}

\item{x}{A Settings_suspect_screening_StreamFind S3 class object.}
}
\value{
A ProcessingSettings S3 class object with subclass
Settings_suspect_screening_StreamFind.
}
\description{
Settings for performing suspect screening using a data.frame with target
compounds.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate(Settings_suspect_screening_StreamFind)}: Validates the Settings_suspect_screening_StreamFind S3 class object,
returning a logical value of length one.

}}
