% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_R6_StatisticEngine.R
\name{StatisticEngine}
\alias{StatisticEngine}
\title{\strong{StatisticEngine} R6 class and methods}
\description{
The \emph{StatisticEngine} R6 class is a framework for performing statistical analysis on data.
}
\note{
Note that only numeric values are accepted in data and the data column names are used as variable names
and data row names are used as analyses names. The data is internally converted to \emph{StatisticAnalysis} objects.

Note that the model must be created before prediction and data must have the same number of variables as
the model. Also, any pre-processing applied to the model data should be applied to the data before prediction.
Note that only numeric values are accepted in data and the data column names are used as variable names
and data row names are used as analyses names.

Note that the classification must be prepared before using the method \code{prepare_classification} and data
must have the same number of variables as the analyses in the engine.

Note that only the model MCR is valid for quantification and concentrations must be added to analyses.
}
\section{Super class}{
\code{\link[StreamFind:CoreEngine]{StreamFind::CoreEngine}} -> \code{StatisticEngine}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{data}}{Matrix of the data, where rows represent analyses and columns variables.}

\item{\code{model}}{Statistic model.}

\item{\code{prediction_results}}{Prediction results from model.}

\item{\code{classification_results}}{Classification results.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StatisticEngine-new}{\code{StatisticEngine$new()}}
\item \href{#method-StatisticEngine-get_overview}{\code{StatisticEngine$get_overview()}}
\item \href{#method-StatisticEngine-get_classes}{\code{StatisticEngine$get_classes()}}
\item \href{#method-StatisticEngine-get_concentrations}{\code{StatisticEngine$get_concentrations()}}
\item \href{#method-StatisticEngine-get_number_variables}{\code{StatisticEngine$get_number_variables()}}
\item \href{#method-StatisticEngine-get_model_explained_variance}{\code{StatisticEngine$get_model_explained_variance()}}
\item \href{#method-StatisticEngine-get_model_scores}{\code{StatisticEngine$get_model_scores()}}
\item \href{#method-StatisticEngine-get_model_loadings}{\code{StatisticEngine$get_model_loadings()}}
\item \href{#method-StatisticEngine-get_model_residuals}{\code{StatisticEngine$get_model_residuals()}}
\item \href{#method-StatisticEngine-get_model_resolved_spectra}{\code{StatisticEngine$get_model_resolved_spectra()}}
\item \href{#method-StatisticEngine-get_model_contributions}{\code{StatisticEngine$get_model_contributions()}}
\item \href{#method-StatisticEngine-add_analyses}{\code{StatisticEngine$add_analyses()}}
\item \href{#method-StatisticEngine-add_data}{\code{StatisticEngine$add_data()}}
\item \href{#method-StatisticEngine-add_classes}{\code{StatisticEngine$add_classes()}}
\item \href{#method-StatisticEngine-add_concentrations}{\code{StatisticEngine$add_concentrations()}}
\item \href{#method-StatisticEngine-predict}{\code{StatisticEngine$predict()}}
\item \href{#method-StatisticEngine-classify}{\code{StatisticEngine$classify()}}
\item \href{#method-StatisticEngine-quantify}{\code{StatisticEngine$quantify()}}
\item \href{#method-StatisticEngine-plot_data}{\code{StatisticEngine$plot_data()}}
\item \href{#method-StatisticEngine-plot_model_explained_variance}{\code{StatisticEngine$plot_model_explained_variance()}}
\item \href{#method-StatisticEngine-plot_model_scores}{\code{StatisticEngine$plot_model_scores()}}
\item \href{#method-StatisticEngine-plot_model_residuals}{\code{StatisticEngine$plot_model_residuals()}}
\item \href{#method-StatisticEngine-plot_model_loadings}{\code{StatisticEngine$plot_model_loadings()}}
\item \href{#method-StatisticEngine-plot_model_resolved_spectra}{\code{StatisticEngine$plot_model_resolved_spectra()}}
\item \href{#method-StatisticEngine-plot_model_contributions}{\code{StatisticEngine$plot_model_contributions()}}
\item \href{#method-StatisticEngine-plot_predicted_distances}{\code{StatisticEngine$plot_predicted_distances()}}
\item \href{#method-StatisticEngine-processing_methods}{\code{StatisticEngine$processing_methods()}}
\item \href{#method-StatisticEngine-clone}{\code{StatisticEngine$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="add_blank_names"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-add_blank_names'><code>StreamFind::CoreEngine$add_blank_names()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="add_headers"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-add_headers'><code>StreamFind::CoreEngine$add_headers()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="add_metadata"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-add_metadata'><code>StreamFind::CoreEngine$add_metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="add_replicate_names"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-add_replicate_names'><code>StreamFind::CoreEngine$add_replicate_names()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="add_results"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-add_results'><code>StreamFind::CoreEngine$add_results()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="add_settings"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-add_settings'><code>StreamFind::CoreEngine$add_settings()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="export"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-export'><code>StreamFind::CoreEngine$export()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="export_analyses"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-export_analyses'><code>StreamFind::CoreEngine$export_analyses()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="export_headers"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-export_headers'><code>StreamFind::CoreEngine$export_headers()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="export_settings"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-export_settings'><code>StreamFind::CoreEngine$export_settings()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="get_analyses"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-get_analyses'><code>StreamFind::CoreEngine$get_analyses()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="get_analysis_names"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-get_analysis_names'><code>StreamFind::CoreEngine$get_analysis_names()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="get_blank_names"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-get_blank_names'><code>StreamFind::CoreEngine$get_blank_names()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="get_files"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-get_files'><code>StreamFind::CoreEngine$get_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="get_formats"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-get_formats'><code>StreamFind::CoreEngine$get_formats()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="get_headers"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-get_headers'><code>StreamFind::CoreEngine$get_headers()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="get_history"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-get_history'><code>StreamFind::CoreEngine$get_history()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="get_number_analyses"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-get_number_analyses'><code>StreamFind::CoreEngine$get_number_analyses()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="get_replicate_names"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-get_replicate_names'><code>StreamFind::CoreEngine$get_replicate_names()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="get_results"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-get_results'><code>StreamFind::CoreEngine$get_results()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="get_results_names"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-get_results_names'><code>StreamFind::CoreEngine$get_results_names()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="get_settings"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-get_settings'><code>StreamFind::CoreEngine$get_settings()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="get_settings_names"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-get_settings_names'><code>StreamFind::CoreEngine$get_settings_names()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="get_types"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-get_types'><code>StreamFind::CoreEngine$get_types()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="get_workflow_overview"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-get_workflow_overview'><code>StreamFind::CoreEngine$get_workflow_overview()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="has_analyses"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-has_analyses'><code>StreamFind::CoreEngine$has_analyses()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="has_results"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-has_results'><code>StreamFind::CoreEngine$has_results()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="has_settings"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-has_settings'><code>StreamFind::CoreEngine$has_settings()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="import"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-import'><code>StreamFind::CoreEngine$import()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="import_analyses"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-import_analyses'><code>StreamFind::CoreEngine$import_analyses()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="import_headers"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-import_headers'><code>StreamFind::CoreEngine$import_headers()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="import_settings"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-import_settings'><code>StreamFind::CoreEngine$import_settings()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="load"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-load'><code>StreamFind::CoreEngine$load()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="print"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-print'><code>StreamFind::CoreEngine$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="print_analyses"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-print_analyses'><code>StreamFind::CoreEngine$print_analyses()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="print_headers"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-print_headers'><code>StreamFind::CoreEngine$print_headers()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="print_results"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-print_results'><code>StreamFind::CoreEngine$print_results()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="print_workflow"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-print_workflow'><code>StreamFind::CoreEngine$print_workflow()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="process"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-process'><code>StreamFind::CoreEngine$process()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="remove_analyses"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-remove_analyses'><code>StreamFind::CoreEngine$remove_analyses()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="remove_headers"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-remove_headers'><code>StreamFind::CoreEngine$remove_headers()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="remove_results"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-remove_results'><code>StreamFind::CoreEngine$remove_results()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="remove_settings"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-remove_settings'><code>StreamFind::CoreEngine$remove_settings()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="run_app"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-run_app'><code>StreamFind::CoreEngine$run_app()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="run_workflow"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-run_workflow'><code>StreamFind::CoreEngine$run_workflow()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="save"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-save'><code>StreamFind::CoreEngine$save()</code></a></span></li>
<li><span class="pkg-link" data-pkg="StreamFind" data-topic="CoreEngine" data-id="subset_analyses"><a href='../../StreamFind/html/CoreEngine.html#method-CoreEngine-subset_analyses'><code>StreamFind::CoreEngine$subset_analyses()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatisticEngine-new"></a>}}
\if{latex}{\out{\hypertarget{method-StatisticEngine-new}{}}}
\subsection{Method \code{new()}}{
Creates an R6 class \emph{StatisticEngine}. Child of \emph{CoreEngine} R6 class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatisticEngine$new(
  data = NULL,
  headers = NULL,
  settings = NULL,
  analyses = NULL,
  results = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data.frame, data-table or matrix with data.}

\item{\code{headers}}{A named list with headers or a ProjectHeaders S3 class object.
Each list element must have length one. There is no type restrictions
except for name (must be type character length 1), author (must be type
character length 1), file (must be type character length 1) and date (must
be class POSIXct or class POSIXt length 1). See more information in
\code{?ProjectHeaders}.}

\item{\code{settings}}{A named list of ProcessingSettings objects or a
single ProcessingSettings object. The list names should match
the call name of each ProcessingSettings object. Alternatively, a named
list with call name, algorithm and parameters to be transformed and added
as ProcessingSettings object.}

\item{\code{analyses}}{A numeric/character vector with the number/name of the analyses.}

\item{\code{results}}{A named list of objects from processing methods.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatisticEngine-get_overview"></a>}}
\if{latex}{\out{\hypertarget{method-StatisticEngine-get_overview}{}}}
\subsection{Method \code{get_overview()}}{
Gets an overview data.frame of all the analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatisticEngine$get_overview()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatisticEngine-get_classes"></a>}}
\if{latex}{\out{\hypertarget{method-StatisticEngine-get_classes}{}}}
\subsection{Method \code{get_classes()}}{
Gets the class of each analysis.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatisticEngine$get_classes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatisticEngine-get_concentrations"></a>}}
\if{latex}{\out{\hypertarget{method-StatisticEngine-get_concentrations}{}}}
\subsection{Method \code{get_concentrations()}}{
Gets the concentration of each analysis.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatisticEngine$get_concentrations()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatisticEngine-get_number_variables"></a>}}
\if{latex}{\out{\hypertarget{method-StatisticEngine-get_number_variables}{}}}
\subsection{Method \code{get_number_variables()}}{
Gets the number of variables.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatisticEngine$get_number_variables()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatisticEngine-get_model_explained_variance"></a>}}
\if{latex}{\out{\hypertarget{method-StatisticEngine-get_model_explained_variance}{}}}
\subsection{Method \code{get_model_explained_variance()}}{
Gets model explained variance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatisticEngine$get_model_explained_variance()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatisticEngine-get_model_scores"></a>}}
\if{latex}{\out{\hypertarget{method-StatisticEngine-get_model_scores}{}}}
\subsection{Method \code{get_model_scores()}}{
Gets model scores.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatisticEngine$get_model_scores(analyses = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name of the analyses.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatisticEngine-get_model_loadings"></a>}}
\if{latex}{\out{\hypertarget{method-StatisticEngine-get_model_loadings}{}}}
\subsection{Method \code{get_model_loadings()}}{
Gets model loadings.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatisticEngine$get_model_loadings()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatisticEngine-get_model_residuals"></a>}}
\if{latex}{\out{\hypertarget{method-StatisticEngine-get_model_residuals}{}}}
\subsection{Method \code{get_model_residuals()}}{
Gets model residuals.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatisticEngine$get_model_residuals(analyses = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name of the analyses.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatisticEngine-get_model_resolved_spectra"></a>}}
\if{latex}{\out{\hypertarget{method-StatisticEngine-get_model_resolved_spectra}{}}}
\subsection{Method \code{get_model_resolved_spectra()}}{
Gets model resolved spectra.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatisticEngine$get_model_resolved_spectra(pcs = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pcs}}{Integer vector with the principle components.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatisticEngine-get_model_contributions"></a>}}
\if{latex}{\out{\hypertarget{method-StatisticEngine-get_model_contributions}{}}}
\subsection{Method \code{get_model_contributions()}}{
Gets the model contributions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatisticEngine$get_model_contributions(pcs = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pcs}}{Integer vector with the principle components.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatisticEngine-add_analyses"></a>}}
\if{latex}{\out{\hypertarget{method-StatisticEngine-add_analyses}{}}}
\subsection{Method \code{add_analyses()}}{
Adds analyses. Note that when adding new analyses, any existing results are removed.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatisticEngine$add_analyses(analyses = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A \emph{StatisticAnalysis} S3 class object or a list with \emph{StatisticAnalysis} S3 class objects as
elements (see \code{?StatisticAnalysis} for more information).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatisticEngine-add_data"></a>}}
\if{latex}{\out{\hypertarget{method-StatisticEngine-add_data}{}}}
\subsection{Method \code{add_data()}}{
Adds data to the \emph{StatisticEngine} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatisticEngine$add_data(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data.frame, data-table or matrix with data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatisticEngine-add_classes"></a>}}
\if{latex}{\out{\hypertarget{method-StatisticEngine-add_classes}{}}}
\subsection{Method \code{add_classes()}}{
Adds classes to the analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatisticEngine$add_classes(classes)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{classes}}{A character vector with the classes.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatisticEngine-add_concentrations"></a>}}
\if{latex}{\out{\hypertarget{method-StatisticEngine-add_concentrations}{}}}
\subsection{Method \code{add_concentrations()}}{
Adds concentrations to the analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatisticEngine$add_concentrations(concentrations = NA_real_)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concentrations}}{A numeric vector with the concentrations.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatisticEngine-predict"></a>}}
\if{latex}{\out{\hypertarget{method-StatisticEngine-predict}{}}}
\subsection{Method \code{predict()}}{
Predicts the data using the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatisticEngine$predict(data = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data.frame, data-table or matrix with data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatisticEngine-classify"></a>}}
\if{latex}{\out{\hypertarget{method-StatisticEngine-classify}{}}}
\subsection{Method \code{classify()}}{
Classifies the data using the classification labels of the analysis.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatisticEngine$classify(data = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data.frame, data-table or matrix with data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatisticEngine-quantify"></a>}}
\if{latex}{\out{\hypertarget{method-StatisticEngine-quantify}{}}}
\subsection{Method \code{quantify()}}{
Evaluates the model for quantification.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatisticEngine$quantify()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatisticEngine-plot_data"></a>}}
\if{latex}{\out{\hypertarget{method-StatisticEngine-plot_data}{}}}
\subsection{Method \code{plot_data()}}{
Plots the raw data in analyses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatisticEngine$plot_data(
  analyses = NULL,
  features = NULL,
  transpose = FALSE,
  interactive = TRUE,
  xLab = NULL,
  yLab = NULL,
  title = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name of the analyses.}

\item{\code{interactive}}{Logical (length 1). When \code{TRUE}, the data is plotted interactively using \pkg{plotly}.}

\item{\code{xLab}}{A string with the title for the x axis.}

\item{\code{yLab}}{A string with the title for the y axis.}

\item{\code{title}}{A string with the title.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatisticEngine-plot_model_explained_variance"></a>}}
\if{latex}{\out{\hypertarget{method-StatisticEngine-plot_model_explained_variance}{}}}
\subsection{Method \code{plot_model_explained_variance()}}{
Plots the model explained cumulative variance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatisticEngine$plot_model_explained_variance(
  interactive = TRUE,
  xLab = NULL,
  yLab = NULL,
  title = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{interactive}}{Logical (length 1). When \code{TRUE}, the data is plotted interactively using \pkg{plotly}.}

\item{\code{xLab}}{A string with the title for the x axis.}

\item{\code{yLab}}{A string with the title for the y axis.}

\item{\code{title}}{A string with the title.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatisticEngine-plot_model_scores"></a>}}
\if{latex}{\out{\hypertarget{method-StatisticEngine-plot_model_scores}{}}}
\subsection{Method \code{plot_model_scores()}}{
Plots scores of the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatisticEngine$plot_model_scores(
  analyses = NULL,
  interactive = TRUE,
  pcs = 1:2,
  title = NULL,
  colorGroups = NULL,
  showText = TRUE,
  showLegend = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name of the analyses.}

\item{\code{interactive}}{Logical (length 1). When \code{TRUE}, the data is plotted interactively using \pkg{plotly}.}

\item{\code{pcs}}{A numeric vector (length 2) with the principle components to plot.}

\item{\code{title}}{A string with the title.}

\item{\code{colorGroups}}{A factor character vector with the color groups for the scores.}

\item{\code{showText}}{Logical (length 1), set to \code{TRUE} to show the text annotations.}

\item{\code{showLegend}}{Logical (length 1). Set to \code{TRUE} to show legend.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatisticEngine-plot_model_residuals"></a>}}
\if{latex}{\out{\hypertarget{method-StatisticEngine-plot_model_residuals}{}}}
\subsection{Method \code{plot_model_residuals()}}{
Plots residuals of the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatisticEngine$plot_model_residuals(
  analyses = NULL,
  interactive = TRUE,
  xLab = NULL,
  yLab = NULL,
  title = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyses}}{A numeric/character vector with the number/name of the analyses.}

\item{\code{interactive}}{Logical (length 1). When \code{TRUE}, the data is plotted interactively using \pkg{plotly}.}

\item{\code{xLab}}{A string with the title for the x axis.}

\item{\code{yLab}}{A string with the title for the y axis.}

\item{\code{title}}{A string with the title.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatisticEngine-plot_model_loadings"></a>}}
\if{latex}{\out{\hypertarget{method-StatisticEngine-plot_model_loadings}{}}}
\subsection{Method \code{plot_model_loadings()}}{
Plots model loadings.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatisticEngine$plot_model_loadings(
  interactive = TRUE,
  pcs = 1:2,
  colorKey = NULL,
  title = NULL,
  showText = TRUE,
  showLegend = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{interactive}}{Logical (length 1). When \code{TRUE}, the data is plotted interactively using \pkg{plotly}.}

\item{\code{pcs}}{A vector with the principle components to plot.}

\item{\code{colorKey}}{A character vector with the color key for the loading variables.}

\item{\code{title}}{A string with the title.}

\item{\code{showText}}{Logical (length 1), set to \code{TRUE} to show the text annotations.}

\item{\code{showLegend}}{Logical (length 1). Set to \code{TRUE} to show legend.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatisticEngine-plot_model_resolved_spectra"></a>}}
\if{latex}{\out{\hypertarget{method-StatisticEngine-plot_model_resolved_spectra}{}}}
\subsection{Method \code{plot_model_resolved_spectra()}}{
Plots model resolved spectra.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatisticEngine$plot_model_resolved_spectra(
  interactive = TRUE,
  pcs = NULL,
  original = TRUE,
  title = NULL,
  showText = TRUE,
  showLegend = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{interactive}}{Logical (length 1). When \code{TRUE}, the data is plotted interactively using \pkg{plotly}.}

\item{\code{pcs}}{Integer vectors with the principle component to use for categorization.}

\item{\code{original}}{Logical, if TRUE the original data is plotted.}

\item{\code{title}}{A string with the title.}

\item{\code{showText}}{Logical (length 1), set to \code{TRUE} to show the text annotations.}

\item{\code{showLegend}}{Logical (length 1). Set to \code{TRUE} to show legend.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatisticEngine-plot_model_contributions"></a>}}
\if{latex}{\out{\hypertarget{method-StatisticEngine-plot_model_contributions}{}}}
\subsection{Method \code{plot_model_contributions()}}{
Plots model contributions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatisticEngine$plot_model_contributions(
  interactive = TRUE,
  pcs = NULL,
  title = NULL,
  showText = TRUE,
  showLegend = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{interactive}}{Logical (length 1). When \code{TRUE}, the data is plotted interactively using \pkg{plotly}.}

\item{\code{pcs}}{Integer vectors with the principle component to use for categorization.}

\item{\code{title}}{A string with the title.}

\item{\code{showText}}{Logical (length 1), set to \code{TRUE} to show the text annotations.}

\item{\code{showLegend}}{Logical (length 1). Set to \code{TRUE} to show legend.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatisticEngine-plot_predicted_distances"></a>}}
\if{latex}{\out{\hypertarget{method-StatisticEngine-plot_predicted_distances}{}}}
\subsection{Method \code{plot_predicted_distances()}}{
Plots scores of the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatisticEngine$plot_predicted_distances(
  pc = NULL,
  interactive = TRUE,
  title = NULL,
  showText = TRUE,
  showLegend = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pc}}{Integer (length 1) with the principle component to use for categorization.}

\item{\code{interactive}}{Logical (length 1). When \code{TRUE}, the data is plotted interactively using \pkg{plotly}.}

\item{\code{title}}{A string with the title.}

\item{\code{showText}}{Logical (length 1), set to \code{TRUE} to show the text annotations.}

\item{\code{showLegend}}{Logical (length 1). Set to \code{TRUE} to show legend.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatisticEngine-processing_methods"></a>}}
\if{latex}{\out{\hypertarget{method-StatisticEngine-processing_methods}{}}}
\subsection{Method \code{processing_methods()}}{
A data.table with available data processing methods.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatisticEngine$processing_methods()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatisticEngine-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StatisticEngine-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatisticEngine$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
