% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_StatisticMethod_Quantify.R
\name{StatisticMethod_Quantify_mcrals}
\alias{StatisticMethod_Quantify_mcrals}
\title{StatisticMethod_Quantify_mcrals}
\usage{
StatisticMethod_Quantify_mcrals(concentrations = NA_real_)
}
\arguments{
\item{concentrations}{Numeric of the same length as analyses with the concentrations to be used for quantification.
When concentration is not known or not available, use \code{NA_real_}.}
}
\value{
A StatisticMethod_Quantify_mcrals object.
}
\description{
Quantify data using the mcrals statistical model and concentrations.
}
\details{
StatisticMethod_Quantify_mcrals S7 class
}
