% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamFind-package.R
\docType{package}
\name{streamFind-package}
\alias{streamFind}
\alias{streamFind-package}
\title{streamFind: Framework and workflows for data analysis in the water cycle}
\description{
The streamFind R package is a framework for spectrometric and spectroscopic data processing with focus on environmental and quality studies of the water cycle. The streamFind R package aims to stimulate the use of advanced data analysis (e.g., non-target screening (NTS) and statistical analysis) in routine studies, promoting standardization of data processing and easing the retrospective inclusion of data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ricardobachertdacunha/streamFind}
  \item \url{https://ricardobachertdacunha.github.io/streamFind/}
  \item Report bugs at \url{https://github.com/ricardobachertdacunha/streamFind/issues}
}

}
\author{
\strong{Maintainer}: Ricardo Cunha \email{cunha@iuta.de} (\href{https://orcid.org/0000-0002-6791-0119}{ORCID})

}
\keyword{internal}
