% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamFind-package.R
\docType{package}
\name{StreamFind-package}
\alias{StreamFind}
\alias{StreamFind-package}
\title{StreamFind: Framework and workflows for data analysis in the water cycle}
\description{
The StreamFind R package is a framework for spectrometric and spectroscopic data processing with focus on environmental and quality studies of the water cycle. The StreamFind R package aims to stimulate the use of advanced data analysis (e.g., non-target screening (NTS) and statistical analysis) in routine studies, promoting standardization of data processing and easing the retrospective inclusion of data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/odea-project/StreamFind}
  \item \url{https://odea-project.github.io/StreamFind/}
  \item Report bugs at \url{https://github.com/odea-project/StreamFind/issues}
}

}
\author{
\strong{Maintainer}: Ricardo Cunha \email{cunha@iuta.de} (\href{https://orcid.org/0000-0002-6791-0119}{ORCID})

Other contributors:
\itemize{
  \item Walter Laurito \email{laurito@fzi.de} [contributor]
}

}
\keyword{internal}
