% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_class_001_streamSet.R
\docType{class}
\name{streamSet-class}
\alias{streamSet-class}
\alias{show,streamSet-method}
\alias{[,streamSet,ANY,missing,missing-method}
\alias{setInfo,streamSet-method}
\alias{setInfo,streamSet,streamSet-method}
\alias{path,streamSet-method}
\alias{path,streamSet,streamSet-method}
\alias{analysisNames,streamSet-method}
\alias{analysisNames,streamSet,streamSet-method}
\alias{addAnalyses,streamSet-method}
\alias{addAnalyses,streamSet,streamSet-method}
\title{streamSet}
\usage{
\S4method{show}{streamSet}(object)

\S4method{[}{streamSet,ANY,missing,missing}(x, i, j, ..., drop = TRUE)

\S4method{setInfo}{streamSet}(object, title = NULL, date = NULL)

\S4method{path}{streamSet}(object)

\S4method{analysisNames}{streamSet}(object)

\S4method{addAnalyses}{streamSet}(object, analysisList = NULL)
}
\arguments{
\item{object}{An \linkS4class{streamSet} object.}

\item{x}{A \linkS4class{streamSet} object.}

\item{i}{The indice/s or name/s of the analyses to keep in \code{x}.}

\item{j}{Not applicable to \linkS4class{streamSet}.}

\item{...}{Other arguments.}

\item{drop}{Not applicable to \linkS4class{streamSet}.}

\item{title}{A character string to be used as set title.}

\item{date}{The \link{Date} of the set.}

\item{analysisList}{A list of named analyses paths as character string.}
}
\description{
An S4 class object containing the set title, date,
storage path and a list of analysis files.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show(streamSet)}: prints the details of an \linkS4class{streamSet} object.

\item \code{x[i}: subset on analyses, using analysis index or name.

\item \code{setInfo(streamSet)}: setter and getter for set title, date and path.
When the \code{title} and \code{date} arguments
are missing it returns a list with the set title, date and path.
If arguments are given, the set title and date are updated
based on the specified arguments, returning the original class object.

\item \code{path(streamSet)}: getter for set path.

\item \code{analysisNames(streamSet)}: getter for the analysis names.

\item \code{addAnalyses(streamSet)}: getter for analysis names.

}}
\section{Slots}{

\describe{
\item{\code{title}}{A character string with the stream set title.}

\item{\code{date}}{The date of the set.}

\item{\code{path}}{A character string with the set path.}

\item{\code{analyses}}{A list of analyses added to the project.
The class of the analysis objects are defined according to the file formats.
When a class is not possible to be assigned, a list of file paths.}
}}

