% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ms_trimFiles.R
\name{trimSpectraFilesMZR}
\alias{trimSpectraFilesMZR}
\title{trimSpectraMZR}
\usage{
trimSpectraFilesMZR(
  files,
  MS1 = TRUE,
  MS2 = TRUE,
  rtr = NULL,
  mzr = NULL,
  mzr_ms2 = NULL,
  intensityThreshold = NULL,
  copyMetadata = TRUE,
  path = NULL,
  prefix = "trim_"
)
}
\arguments{
\item{files}{A character vector with the full path of the files to trim.}

\item{MS1}{Set to TRUE for trimming MS level one data.}

\item{MS2}{Set to TRUE for trimming MS level two data.}

\item{rtr}{A numeric vector of length 2 with the maximum and minimum
time range to trim the files, in seconds.}

\item{mzr}{A numeric vector of length 2 with the maximum and minimum
\emph{m/z} range to trim the files, in Da.}

\item{mzr_ms2}{A numeric vector of length 2 with the maximum and minimum
\emph{m/z} range to trim MS2 data files, in Da. when NULL and MS2 is TRUE
the mzr is used instead.}

\item{intensityThreshold}{A numeric vector of length one with the minimum
intensity threshold. Traces below the given intensity threshold are removed.
If a length two vector is given, the first value if applied for MS1 data
and the second to MS2 data.}

\item{copyMetadata}{Set to TRUE to copy metadata from the original file.}

\item{path}{A character string with the path to save the files.
When \code{NULL}, the file directory is used instead.}

\item{prefix}{A character string with the name prefix for the new trimmed file.}
}
\value{
Saves the trimmed \emph{mzML} or \emph{mzXML} files
in a given path, adding a pre-defined prefix to the file name.
}
\description{
Trim \emph{mzML} or \emph{mzXML} files based on time (in seconds)
and  \emph{m/z} (in Da) ranges using the \pkg{mzR} package.
}
\references{
\insertRef{mzr01}{streamFind}

\insertRef{mzr02}{streamFind}

\insertRef{mzr03}{streamFind}

\insertRef{mzr04}{streamFind}
}
