% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_S3_Analysis.R,
%   R/class_S3_MassSpecAnalysis.R, R/class_S3_ProcessingSettings.R,
%   R/class_S3_ProjectHeaders.R, R/class_S3_RamanAnalysis.R,
%   R/class_S3_StatisticAnalysis.R, R/generics.R
\name{validate.Analysis}
\alias{validate.Analysis}
\alias{validate.MassSpecAnalysis}
\alias{validate.ProcessingSettings}
\alias{validate.ProjectHeaders}
\alias{validate.RamanAnalysis}
\alias{validate.StatisticAnalysis}
\alias{validate}
\title{Validate}
\usage{
\method{validate}{Analysis}(x = NULL)

\method{validate}{MassSpecAnalysis}(x = NULL)

\method{validate}{ProcessingSettings}(x)

\method{validate}{ProjectHeaders}(x)

\method{validate}{RamanAnalysis}(x = NULL)

\method{validate}{StatisticAnalysis}(x = NULL)

validate(x)
}
\arguments{
\item{x}{An object.}
}
\description{
Validates an object.
}
\section{Methods (by class)}{
\itemize{
\item \code{validate(Analysis)}: description Validates an Analysis S3 object.

\item \code{validate(MassSpecAnalysis)}: Validates an MassSpecAnalysis S3 object.

\item \code{validate(ProcessingSettings)}: Validates a ProcessingSettings S3 object.

\item \code{validate(ProjectHeaders)}: Validates a ProjectHeaders S3 object.

\item \code{validate(RamanAnalysis)}: Validates a RamanAnalysis S3 object

\item \code{validate(StatisticAnalysis)}: Validates a StatisticAnalysis S3 object

}}
