% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_R6_msData.R
\name{validate_header}
\alias{validate_header}
\title{validate_header}
\usage{
validate_header(value = NULL)
}
\arguments{
\item{value}{A header list.}
}
\value{
A logical value of length 1.
}
\description{
Validates the header list of information.
}
