% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_S3_msAnalysis.R
\name{validate_ms_analysis}
\alias{validate_ms_analysis}
\title{validate_ms_analysis}
\usage{
validate_ms_analysis(x = NULL)
}
\arguments{
\item{x}{An msAnalysis S3 class object.}
}
\value{
Logical of length one.
}
\description{
Validates an msAnalysis S3 class object.
}
