% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Workflow.R
\name{Workflow}
\alias{Workflow}
\title{Workflow Manager}
\usage{
Workflow(processing_steps = list())
}
\arguments{
\item{processing_steps}{A list of \link{ProcessingStep} objects.}
}
\description{
The Workflow class is used to manage an ordered list of \link{ProcessingStep} objects of a specific type of data.
}
\section{Slots}{

\describe{
\item{\code{methods}}{(getter) A character vector of the methods of each processing step.}

\item{\code{overview}}{(getter) A data frame with an overview of processing steps in the workflow.}
}}

